Dentist ratings of whether caries are healthy or not based on X-rays

It consists of binary ratings, made by 5 dentists, of whether a given tooth was healthy (sound) or had caries, also known as cavities. The ratings were performed using X-ray only, which was thought to be more error-prone than visual/tactile assessment of each tooth. In total 3,689 ratings were made. This data is in 'grouped' format. Each row is one of the 'pattern' with the final columns being a tally of how many times that pattern occurs in the dataset.

caries

Format

A data.frame with 6 columns and 32 rows.

rater_1

The rating of the dentist 1

rater_2

The rating of the dentist 2

rater_3

The rating of the dentist 3

rater_4

The rating of the dentist 4

rater_5

The rating of the dentist 5

n

The number of times the rating pattern appears in the dataset

References

Espeland, Mark A., and Stanley L. Handelman. “Using Latent Class Models to Characterize and Assess Relative Error in Discrete Measurements.” Biometrics 45, no. 2 (1989): 587–99.