Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for class 'mcmc_fit'
class_probabilities(fit, ...)

# S3 method for class 'optim_fit'
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 9.1e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.91 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.289 seconds (Warm-up)
#> Chain 1:                1.394 seconds (Sampling)
#> Chain 1:                2.683 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 8.5e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.85 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.256 seconds (Warm-up)
#> Chain 2:                0.8 seconds (Sampling)
#> Chain 2:                2.056 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 8.6e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.86 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.247 seconds (Warm-up)
#> Chain 3:                1.401 seconds (Sampling)
#> Chain 3:                2.648 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 8.6e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.86 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.255 seconds (Warm-up)
#> Chain 4:                1.39 seconds (Sampling)
#> Chain 4:                2.645 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>     
#>              [,1]         [,2]         [,3]         [,4]
#>   1  9.999994e-01 1.526194e-07 3.364098e-08 3.962498e-07
#>   2  9.138528e-08 2.496639e-05 9.798572e-01 2.011773e-02
#>   3  3.817536e-01 6.176755e-01 1.267295e-04 4.441858e-04
#>   4  5.936287e-03 9.934885e-01 3.638358e-04 2.113607e-04
#>   5  2.532802e-07 9.999655e-01 3.136923e-05 2.848784e-06
#>   6  1.835774e-06 9.993498e-01 6.278928e-04 2.049991e-05
#>   7  9.993975e-01 5.929117e-04 1.027507e-06 8.563116e-06
#>   8  1.128606e-08 3.021912e-05 9.995421e-01 4.276270e-04
#>   9  1.665136e-06 9.999521e-01 4.057889e-05 5.609681e-06
#>   10 3.454960e-06 9.979762e-01 1.986600e-03 3.377409e-05
#>   11 1.387746e-08 1.868257e-08 1.297205e-04 9.998702e-01
#>   12 3.686640e-05 8.822184e-01 1.110031e-01 6.741619e-03
#>   13 9.999994e-01 1.526194e-07 3.364098e-08 3.962498e-07
#>   14 3.802047e-05 9.993862e-01 5.351609e-04 4.058582e-05
#>   15 9.999858e-01 1.239412e-05 1.683845e-07 1.603422e-06
#>   16 9.999907e-01 7.157015e-06 1.849653e-07 1.989250e-06
#>   17 9.999994e-01 1.526194e-07 3.364098e-08 3.962498e-07
#>   18 9.999994e-01 1.526194e-07 3.364098e-08 3.962498e-07
#>   19 4.283608e-05 9.999470e-01 4.478336e-06 5.661466e-06
#>   20 2.182595e-04 9.992474e-01 4.219412e-04 1.124297e-04
#>   21 5.310602e-07 9.999969e-01 1.611139e-06 9.685259e-07
#>   22 4.283608e-05 9.999470e-01 4.478336e-06 5.661466e-06
#>   23 2.532802e-07 9.999655e-01 3.136923e-05 2.848784e-06
#>   24 7.891627e-05 9.999080e-01 6.922520e-06 6.128474e-06
#>   25 9.999994e-01 1.526194e-07 3.364098e-08 3.962498e-07
#>   26 9.999994e-01 1.526194e-07 3.364098e-08 3.962498e-07
#>   27 9.723999e-07 9.999034e-01 8.968588e-05 5.894782e-06
#>   28 9.999994e-01 1.526194e-07 3.364098e-08 3.962498e-07
#>   29 9.999994e-01 1.526194e-07 3.364098e-08 3.962498e-07
#>   30 9.976891e-01 2.268401e-03 2.814194e-06 3.971433e-05
#>   31 9.999994e-01 1.526194e-07 3.364098e-08 3.962498e-07
#>   32 1.638965e-08 9.846462e-04 9.989135e-01 1.018810e-04
#>   33 9.999994e-01 1.526194e-07 3.364098e-08 3.962498e-07
#>   34 5.310602e-07 9.999969e-01 1.611139e-06 9.685259e-07
#>   35 6.738406e-07 9.976900e-01 2.292250e-03 1.710733e-05
#>   36 3.434542e-07 8.885816e-05 7.941588e-01 2.057520e-01
#>   37 1.744612e-04 9.992554e-01 5.208244e-04 4.928089e-05
#>   38 2.217921e-06 7.103346e-01 2.893790e-01 2.841957e-04
#>   39 6.627249e-08 7.817265e-04 9.975532e-01 1.664982e-03
#>   40 9.999994e-01 1.526194e-07 3.364098e-08 3.962498e-07
#>   41 9.999994e-01 1.526194e-07 3.364098e-08 3.962498e-07
#>   42 9.993975e-01 5.929117e-04 1.027507e-06 8.563116e-06
#>   43 9.723999e-07 9.999034e-01 8.968588e-05 5.894782e-06
#>   44 9.999858e-01 1.239412e-05 1.683845e-07 1.603422e-06
#>   45 5.310602e-07 9.999969e-01 1.611139e-06 9.685259e-07

# }