Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.26    0.33    0.37    0.42    0.50
#> pi[2]            0.41    0.00 0.06    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.00    0.01    0.02    0.05
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.58    0.74    0.80    0.87    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.30
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.42
#> theta[1,4,4]     0.69    0.00 0.14    0.39    0.59    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.75    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.22    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[2,2,2]     0.58    0.00 0.09    0.39    0.52    0.58    0.65    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.24    0.30    0.36    0.48
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[2,3,2]     0.08    0.00 0.08    0.00    0.03    0.06    0.11    0.29
#> theta[2,3,3]     0.62    0.00 0.14    0.35    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.12    0.04    0.12    0.20    0.29    0.48
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[2,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.76    0.90
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.10    0.00 0.06    0.01    0.05    0.09    0.13    0.24
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.88
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[3,3,1]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.29
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.14    0.24    0.41    0.50    0.61    0.76
#> theta[3,3,4]     0.23    0.00 0.12    0.05    0.15    0.22    0.31    0.49
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[3,4,3]     0.14    0.00 0.12    0.01    0.05    0.11    0.20    0.43
#> theta[3,4,4]     0.63    0.00 0.15    0.32    0.52    0.64    0.74    0.90
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.79    0.84    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.09    0.54    0.67    0.73    0.78    0.87
#> theta[4,2,3]     0.13    0.00 0.06    0.03    0.09    0.12    0.17    0.29
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.10    0.19
#> theta[4,3,1]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.28
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[4,3,3]     0.69    0.00 0.13    0.41    0.60    0.70    0.78    0.91
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.39
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.91
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.82    0.87    0.91    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[5,2,1]     0.14    0.00 0.06    0.04    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.33
#> theta[5,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.12
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.38
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.61    0.70    0.85
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.42
#> theta[5,4,4]     0.63    0.00 0.15    0.31    0.52    0.64    0.74    0.90
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_p_z[1,2]   -20.00    0.04 2.26  -24.84  -21.38  -19.89  -18.37  -16.07
#> log_p_z[1,3]   -24.85    0.07 4.22  -34.27  -27.38  -24.38  -21.90  -17.71
#> log_p_z[1,4]   -21.99    0.07 4.35  -31.81  -24.54  -21.53  -18.92  -14.94
#> log_p_z[2,1]   -29.57    0.06 4.30  -39.29  -32.14  -29.13  -26.50  -22.56
#> log_p_z[2,2]   -21.02    0.04 2.40  -26.20  -22.47  -20.87  -19.34  -16.84
#> log_p_z[2,3]    -7.67    0.02 1.27  -10.36   -8.35   -7.55   -6.83   -5.61
#> log_p_z[2,4]   -16.40    0.08 4.39  -26.57  -18.81  -15.82  -13.32   -9.34
#> log_p_z[3,1]   -11.65    0.03 1.63  -15.34  -12.65  -11.46  -10.46   -9.03
#> log_p_z[3,2]   -10.81    0.02 1.51  -14.20  -11.76  -10.64   -9.73   -8.33
#> log_p_z[3,3]   -23.36    0.05 3.36  -30.74  -25.37  -23.04  -21.06  -17.58
#> log_p_z[3,4]   -21.93    0.06 3.66  -29.86  -24.25  -21.59  -19.32  -15.75
#> log_p_z[4,1]   -14.66    0.02 1.87  -18.87  -15.81  -14.49  -13.36  -11.40
#> log_p_z[4,2]    -7.57    0.01 0.97   -9.68   -8.14   -7.49   -6.90   -5.91
#> log_p_z[4,3]   -20.59    0.07 4.22  -30.38  -22.87  -20.08  -17.64  -13.89
#> log_p_z[4,4]   -21.88    0.07 4.24  -31.59  -24.27  -21.36  -18.83  -15.17
#> log_p_z[5,1]   -21.46    0.04 2.53  -26.82  -23.07  -21.30  -19.65  -17.04
#> log_p_z[5,2]    -3.72    0.00 0.42   -4.61   -3.99   -3.69   -3.43   -2.96
#> log_p_z[5,3]   -18.55    0.07 4.07  -27.90  -20.75  -17.97  -15.77  -12.02
#> log_p_z[5,4]   -21.86    0.07 4.26  -31.80  -24.35  -21.36  -18.84  -15.00
#> log_p_z[6,1]   -21.47    0.04 2.55  -27.06  -23.07  -21.28  -19.71  -16.99
#> log_p_z[6,2]    -5.48    0.01 0.67   -6.86   -5.89   -5.45   -5.01   -4.30
#> log_p_z[6,3]   -17.34    0.07 3.92  -26.38  -19.47  -16.79  -14.61  -11.21
#> log_p_z[6,4]   -21.61    0.06 4.18  -30.98  -23.91  -21.12  -18.67  -14.92
#> log_p_z[7,1]    -5.99    0.01 0.63   -7.38   -6.38   -5.95   -5.56   -4.89
#> log_p_z[7,2]   -14.94    0.03 1.67  -18.64  -15.97  -14.79  -13.75  -12.02
#> log_p_z[7,3]   -24.29    0.06 3.47  -31.68  -26.40  -24.02  -21.83  -18.33
#> log_p_z[7,4]   -21.94    0.06 3.60  -29.91  -24.16  -21.72  -19.39  -15.82
#> log_p_z[8,1]   -29.58    0.06 4.27  -39.09  -32.09  -29.18  -26.58  -22.54
#> log_p_z[8,2]   -18.29    0.03 2.13  -22.90  -19.60  -18.13  -16.83  -14.59
#> log_p_z[8,3]    -5.74    0.01 0.85   -7.55   -6.27   -5.68   -5.16   -4.27
#> log_p_z[8,4]   -18.60    0.08 4.50  -29.17  -21.15  -18.09  -15.37  -11.58
#> log_p_z[9,1]   -19.96    0.03 2.30  -24.89  -21.37  -19.83  -18.34  -15.90
#> log_p_z[9,2]    -4.52    0.01 0.58   -5.76   -4.88   -4.48   -4.11   -3.49
#> log_p_z[9,3]   -19.53    0.07 4.17  -29.31  -21.80  -19.03  -16.63  -12.95
#> log_p_z[9,4]   -21.65    0.07 4.16  -31.48  -24.04  -21.18  -18.73  -14.82
#> log_p_z[10,1]  -22.17    0.03 2.40  -27.33  -23.71  -22.02  -20.46  -17.97
#> log_p_z[10,2]   -7.12    0.01 0.73   -8.74   -7.56   -7.07   -6.62   -5.86
#> log_p_z[10,3]  -16.76    0.05 3.09  -23.78  -18.54  -16.45  -14.55  -11.72
#> log_p_z[10,4]  -21.36    0.06 3.45  -29.23  -23.41  -21.08  -18.96  -15.49
#> log_p_z[11,1]  -29.65    0.07 4.32  -39.26  -32.17  -29.19  -26.63  -22.52
#> log_p_z[11,2]  -29.27    0.07 4.13  -38.80  -31.63  -28.75  -26.39  -22.69
#> log_p_z[11,3]  -18.30    0.07 3.60  -27.03  -20.04  -17.71  -15.88  -13.04
#> log_p_z[11,4]   -5.83    0.01 0.99   -8.06   -6.43   -5.76   -5.15   -4.10
#> log_p_z[12,1]  -22.94    0.04 2.74  -28.76  -24.66  -22.71  -20.99  -18.13
#> log_p_z[12,2]   -9.59    0.02 1.29  -12.66  -10.31   -9.43   -8.72   -7.51
#> log_p_z[12,3]  -15.00    0.07 3.92  -24.35  -17.22  -14.54  -12.25   -8.81
#> log_p_z[12,4]  -19.23    0.07 4.00  -28.57  -21.49  -18.71  -16.42  -12.90
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_p_z[13,2]  -20.00    0.04 2.26  -24.84  -21.38  -19.89  -18.37  -16.07
#> log_p_z[13,3]  -24.85    0.07 4.22  -34.27  -27.38  -24.38  -21.90  -17.71
#> log_p_z[13,4]  -21.99    0.07 4.35  -31.81  -24.54  -21.53  -18.92  -14.94
#> log_p_z[14,1]  -19.04    0.03 2.36  -24.20  -20.45  -18.89  -17.37  -15.00
#> log_p_z[14,2]   -6.24    0.01 0.84   -8.14   -6.76   -6.16   -5.66   -4.79
#> log_p_z[14,3]  -18.55    0.07 4.06  -28.00  -20.89  -18.01  -15.68  -12.05
#> log_p_z[14,4]  -21.88    0.07 4.28  -31.78  -24.30  -21.35  -18.83  -14.91
#> log_p_z[15,1]   -4.41    0.00 0.39   -5.25   -4.66   -4.38   -4.14   -3.72
#> log_p_z[15,2]  -17.24    0.03 1.83  -21.19  -18.37  -17.14  -15.92  -14.01
#> log_p_z[15,3]  -24.27    0.06 3.52  -31.98  -26.40  -23.93  -21.80  -18.30
#> log_p_z[15,4]  -21.97    0.05 3.60  -30.00  -24.21  -21.69  -19.39  -15.82
#> log_p_z[16,1]   -3.79    0.01 0.55   -5.03   -4.13   -3.75   -3.41   -2.85
#> log_p_z[16,2]  -17.70    0.03 2.14  -22.35  -19.01  -17.52  -16.20  -14.01
#> log_p_z[16,3]  -24.87    0.07 4.19  -34.20  -27.36  -24.45  -21.96  -17.98
#> log_p_z[16,4]  -21.96    0.07 4.35  -31.85  -24.54  -21.62  -18.85  -14.76
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_p_z[17,2]  -20.00    0.04 2.26  -24.84  -21.38  -19.89  -18.37  -16.07
#> log_p_z[17,3]  -24.85    0.07 4.22  -34.27  -27.38  -24.38  -21.90  -17.71
#> log_p_z[17,4]  -21.99    0.07 4.35  -31.81  -24.54  -21.53  -18.92  -14.94
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_p_z[18,2]  -20.00    0.04 2.26  -24.84  -21.38  -19.89  -18.37  -16.07
#> log_p_z[18,3]  -24.85    0.07 4.22  -34.27  -27.38  -24.38  -21.90  -17.71
#> log_p_z[18,4]  -21.99    0.07 4.35  -31.81  -24.54  -21.53  -18.92  -14.94
#> log_p_z[19,1]  -16.39    0.03 1.94  -20.52  -17.60  -16.26  -15.02  -13.05
#> log_p_z[19,2]   -4.65    0.01 0.60   -5.93   -5.02   -4.61   -4.22   -3.59
#> log_p_z[19,3]  -22.01    0.07 4.31  -31.81  -24.39  -21.54  -18.97  -15.02
#> log_p_z[19,4]  -21.87    0.07 4.21  -31.45  -24.31  -21.41  -18.90  -15.06
#> log_p_z[20,1]  -18.05    0.03 2.27  -22.98  -19.40  -17.86  -16.44  -14.14
#> log_p_z[20,2]   -7.11    0.01 0.95   -9.19   -7.69   -7.05   -6.44   -5.51
#> log_p_z[20,3]  -19.85    0.07 4.08  -29.29  -22.11  -19.28  -16.99  -13.39
#> log_p_z[20,4]  -21.64    0.06 4.15  -30.98  -23.98  -21.11  -18.75  -14.73
#> log_p_z[21,1]  -19.62    0.03 2.34  -24.48  -21.10  -19.46  -17.93  -15.62
#> log_p_z[21,2]   -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.81   -2.38
#> log_p_z[21,3]  -21.08    0.07 4.18  -30.62  -23.40  -20.63  -18.18  -14.41
#> log_p_z[21,4]  -21.87    0.07 4.23  -31.63  -24.26  -21.43  -18.84  -14.90
#> log_p_z[22,1]  -16.39    0.03 1.94  -20.52  -17.60  -16.26  -15.02  -13.05
#> log_p_z[22,2]   -4.65    0.01 0.60   -5.93   -5.02   -4.61   -4.22   -3.59
#> log_p_z[22,3]  -22.01    0.07 4.31  -31.81  -24.39  -21.54  -18.97  -15.02
#> log_p_z[22,4]  -21.87    0.07 4.21  -31.45  -24.31  -21.41  -18.90  -15.06
#> log_p_z[23,1]  -21.46    0.04 2.53  -26.82  -23.07  -21.30  -19.65  -17.04
#> log_p_z[23,2]   -3.72    0.00 0.42   -4.61   -3.99   -3.69   -3.43   -2.96
#> log_p_z[23,3]  -18.55    0.07 4.07  -27.90  -20.75  -17.97  -15.77  -12.02
#> log_p_z[23,4]  -21.86    0.07 4.26  -31.80  -24.35  -21.36  -18.84  -15.00
#> log_p_z[24,1]  -17.42    0.03 2.14  -22.04  -18.74  -17.25  -15.87  -13.80
#> log_p_z[24,2]   -5.80    0.01 0.63   -7.22   -6.18   -5.74   -5.36   -4.76
#> log_p_z[24,3]  -21.66    0.05 3.33  -29.12  -23.57  -21.33  -19.33  -16.00
#> log_p_z[24,4]  -21.89    0.06 3.57  -29.64  -24.08  -21.54  -19.37  -15.76
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_p_z[25,2]  -20.00    0.04 2.26  -24.84  -21.38  -19.89  -18.37  -16.07
#> log_p_z[25,3]  -24.85    0.07 4.22  -34.27  -27.38  -24.38  -21.90  -17.71
#> log_p_z[25,4]  -21.99    0.07 4.35  -31.81  -24.54  -21.53  -18.92  -14.94
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_p_z[26,2]  -20.00    0.04 2.26  -24.84  -21.38  -19.89  -18.37  -16.07
#> log_p_z[26,3]  -24.85    0.07 4.22  -34.27  -27.38  -24.38  -21.90  -17.71
#> log_p_z[26,4]  -21.99    0.07 4.35  -31.81  -24.54  -21.53  -18.92  -14.94
#> log_p_z[27,1]  -21.83    0.04 2.46  -27.16  -23.38  -21.66  -20.04  -17.59
#> log_p_z[27,2]   -5.65    0.01 0.53   -6.77   -5.97   -5.61   -5.28   -4.75
#> log_p_z[27,3]  -18.30    0.05 3.13  -25.26  -20.12  -17.99  -16.11  -13.23
#> log_p_z[27,4]  -21.58    0.06 3.47  -29.20  -23.65  -21.31  -19.16  -15.67
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_p_z[28,2]  -20.00    0.04 2.26  -24.84  -21.38  -19.89  -18.37  -16.07
#> log_p_z[28,3]  -24.85    0.07 4.22  -34.27  -27.38  -24.38  -21.90  -17.71
#> log_p_z[28,4]  -21.99    0.07 4.35  -31.81  -24.54  -21.53  -18.92  -14.94
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_p_z[29,2]  -20.00    0.04 2.26  -24.84  -21.38  -19.89  -18.37  -16.07
#> log_p_z[29,3]  -24.85    0.07 4.22  -34.27  -27.38  -24.38  -21.90  -17.71
#> log_p_z[29,4]  -21.99    0.07 4.35  -31.81  -24.54  -21.53  -18.92  -14.94
#> log_p_z[30,1]   -6.83    0.01 0.84   -8.72   -7.31   -6.73   -6.26   -5.44
#> log_p_z[30,2]  -14.71    0.03 1.72  -18.48  -15.81  -14.58  -13.48  -11.73
#> log_p_z[30,3]  -24.27    0.06 3.51  -32.01  -26.38  -23.89  -21.83  -18.39
#> log_p_z[30,4]  -21.95    0.05 3.62  -29.86  -24.11  -21.70  -19.40  -15.70
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_p_z[31,2]  -20.00    0.04 2.26  -24.84  -21.38  -19.89  -18.37  -16.07
#> log_p_z[31,3]  -24.85    0.07 4.22  -34.27  -27.38  -24.38  -21.90  -17.71
#> log_p_z[31,4]  -21.99    0.07 4.35  -31.81  -24.54  -21.53  -18.92  -14.94
#> log_p_z[32,1]  -29.56    0.06 4.21  -38.98  -32.09  -29.10  -26.61  -22.65
#> log_p_z[32,2]  -14.82    0.02 1.75  -18.67  -15.88  -14.68  -13.60  -11.75
#> log_p_z[32,3]   -6.07    0.01 0.92   -8.01   -6.64   -6.00   -5.42   -4.49
#> log_p_z[32,4]  -20.75    0.09 4.65  -31.30  -23.45  -20.27  -17.41  -13.28
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_p_z[33,2]  -20.00    0.04 2.26  -24.84  -21.38  -19.89  -18.37  -16.07
#> log_p_z[33,3]  -24.85    0.07 4.22  -34.27  -27.38  -24.38  -21.90  -17.71
#> log_p_z[33,4]  -21.99    0.07 4.35  -31.81  -24.54  -21.53  -18.92  -14.94
#> log_p_z[34,1]  -19.62    0.03 2.34  -24.48  -21.10  -19.46  -17.93  -15.62
#> log_p_z[34,2]   -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.81   -2.38
#> log_p_z[34,3]  -21.08    0.07 4.18  -30.62  -23.40  -20.63  -18.18  -14.41
#> log_p_z[34,4]  -21.87    0.07 4.23  -31.63  -24.26  -21.43  -18.84  -14.90
#> log_p_z[35,1]  -22.59    0.04 2.69  -28.38  -24.26  -22.45  -20.67  -17.96
#> log_p_z[35,2]   -5.53    0.01 0.69   -7.08   -5.95   -5.47   -5.04   -4.34
#> log_p_z[35,3]  -15.95    0.07 3.90  -25.15  -18.12  -15.43  -13.17   -9.85
#> log_p_z[35,4]  -21.85    0.07 4.26  -31.77  -24.31  -21.33  -18.81  -14.85
#> log_p_z[36,1]  -29.59    0.06 3.63  -37.37  -31.88  -29.26  -27.02  -23.41
#> log_p_z[36,2]  -21.88    0.04 2.34  -27.03  -23.28  -21.67  -20.21  -17.93
#> log_p_z[36,3]  -10.55    0.04 1.82  -14.88  -11.47  -10.25   -9.33   -7.90
#> log_p_z[36,4]  -14.00    0.06 3.30  -21.45  -16.02  -13.73  -11.55   -8.69
#> log_p_z[37,1]  -18.55    0.03 2.30  -23.61  -19.95  -18.37  -16.92  -14.62
#> log_p_z[37,2]   -7.61    0.01 0.81   -9.46   -8.11   -7.53   -7.04   -6.24
#> log_p_z[37,3]  -19.06    0.05 3.11  -26.14  -20.88  -18.78  -16.88  -13.87
#> log_p_z[37,4]  -21.87    0.06 3.55  -29.63  -24.02  -21.58  -19.41  -15.88
#> log_p_z[38,1]  -25.13    0.04 2.74  -31.15  -26.85  -24.96  -23.24  -20.27
#> log_p_z[38,2]   -9.61    0.01 1.00  -11.85  -10.23   -9.52   -8.91   -7.85
#> log_p_z[38,3]  -11.63    0.04 2.67  -17.87  -13.14  -11.31   -9.69   -7.47
#> log_p_z[38,4]  -21.33    0.06 3.55  -29.09  -23.58  -21.00  -18.75  -15.43
#> log_p_z[39,1]  -29.24    0.06 4.30  -38.85  -31.77  -28.78  -26.19  -22.19
#> log_p_z[39,2]  -16.82    0.03 2.03  -21.21  -18.09  -16.66  -15.41  -13.27
#> log_p_z[39,3]   -7.28    0.01 1.11   -9.67   -7.97   -7.21   -6.51   -5.31
#> log_p_z[39,4]  -18.82    0.07 4.38  -29.06  -21.32  -18.20  -15.72  -11.92
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_p_z[40,2]  -20.00    0.04 2.26  -24.84  -21.38  -19.89  -18.37  -16.07
#> log_p_z[40,3]  -24.85    0.07 4.22  -34.27  -27.38  -24.38  -21.90  -17.71
#> log_p_z[40,4]  -21.99    0.07 4.35  -31.81  -24.54  -21.53  -18.92  -14.94
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_p_z[41,2]  -20.00    0.04 2.26  -24.84  -21.38  -19.89  -18.37  -16.07
#> log_p_z[41,3]  -24.85    0.07 4.22  -34.27  -27.38  -24.38  -21.90  -17.71
#> log_p_z[41,4]  -21.99    0.07 4.35  -31.81  -24.54  -21.53  -18.92  -14.94
#> log_p_z[42,1]   -5.99    0.01 0.63   -7.38   -6.38   -5.95   -5.56   -4.89
#> log_p_z[42,2]  -14.94    0.03 1.67  -18.64  -15.97  -14.79  -13.75  -12.02
#> log_p_z[42,3]  -24.29    0.06 3.47  -31.68  -26.40  -24.02  -21.83  -18.33
#> log_p_z[42,4]  -21.94    0.06 3.60  -29.91  -24.16  -21.72  -19.39  -15.82
#> log_p_z[43,1]  -21.83    0.04 2.46  -27.16  -23.38  -21.66  -20.04  -17.59
#> log_p_z[43,2]   -5.65    0.01 0.53   -6.77   -5.97   -5.61   -5.28   -4.75
#> log_p_z[43,3]  -18.30    0.05 3.13  -25.26  -20.12  -17.99  -16.11  -13.23
#> log_p_z[43,4]  -21.58    0.06 3.47  -29.20  -23.65  -21.31  -19.16  -15.67
#> log_p_z[44,1]   -4.41    0.00 0.39   -5.25   -4.66   -4.38   -4.14   -3.72
#> log_p_z[44,2]  -17.24    0.03 1.83  -21.19  -18.37  -17.14  -15.92  -14.01
#> log_p_z[44,3]  -24.27    0.06 3.52  -31.98  -26.40  -23.93  -21.80  -18.30
#> log_p_z[44,4]  -21.97    0.05 3.60  -30.00  -24.21  -21.69  -19.39  -15.82
#> log_p_z[45,1]  -19.62    0.03 2.34  -24.48  -21.10  -19.46  -17.93  -15.62
#> log_p_z[45,2]   -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.81   -2.38
#> log_p_z[45,3]  -21.08    0.07 4.18  -30.62  -23.40  -20.63  -18.18  -14.41
#> log_p_z[45,4]  -21.87    0.07 4.23  -31.63  -24.26  -21.43  -18.84  -14.90
#> log_lik[1]      -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_lik[2]      -7.61    0.02 1.13  -10.01   -8.30   -7.53   -6.82   -5.60
#> log_lik[3]      -9.96    0.01 0.95  -11.97  -10.56   -9.92   -9.31   -8.19
#> log_lik[4]      -7.56    0.01 0.96   -9.63   -8.14   -7.49   -6.90   -5.91
#> log_lik[5]      -3.72    0.00 0.42   -4.61   -3.99   -3.69   -3.43   -2.96
#> log_lik[6]      -5.48    0.01 0.67   -6.85   -5.89   -5.45   -5.01   -4.30
#> log_lik[7]      -5.99    0.01 0.62   -7.38   -6.38   -5.95   -5.56   -4.89
#> log_lik[8]      -5.74    0.01 0.85   -7.54   -6.27   -5.68   -5.16   -4.26
#> log_lik[9]      -4.52    0.01 0.58   -5.76   -4.88   -4.48   -4.11   -3.49
#> log_lik[10]     -7.12    0.01 0.73   -8.71   -7.56   -7.07   -6.62   -5.86
#> log_lik[11]     -5.83    0.01 0.99   -8.06   -6.43   -5.76   -5.15   -4.10
#> log_lik[12]     -9.32    0.01 1.03  -11.46   -9.97   -9.25   -8.62   -7.46
#> log_lik[13]     -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_lik[14]     -6.24    0.01 0.84   -8.12   -6.76   -6.16   -5.66   -4.79
#> log_lik[15]     -4.41    0.00 0.39   -5.25   -4.66   -4.38   -4.14   -3.72
#> log_lik[16]     -3.79    0.01 0.55   -5.03   -4.13   -3.75   -3.41   -2.85
#> log_lik[17]     -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_lik[18]     -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_lik[19]     -4.65    0.01 0.60   -5.93   -5.02   -4.61   -4.22   -3.59
#> log_lik[20]     -7.11    0.01 0.95   -9.19   -7.68   -7.05   -6.44   -5.51
#> log_lik[21]     -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.81   -2.38
#> log_lik[22]     -4.65    0.01 0.60   -5.93   -5.02   -4.61   -4.22   -3.59
#> log_lik[23]     -3.72    0.00 0.42   -4.61   -3.99   -3.69   -3.43   -2.96
#> log_lik[24]     -5.80    0.01 0.63   -7.22   -6.18   -5.74   -5.36   -4.76
#> log_lik[25]     -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_lik[26]     -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_lik[27]     -5.65    0.01 0.53   -6.77   -5.97   -5.61   -5.28   -4.75
#> log_lik[28]     -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_lik[29]     -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_lik[30]     -6.83    0.01 0.84   -8.71   -7.31   -6.73   -6.26   -5.44
#> log_lik[31]     -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_lik[32]     -6.07    0.01 0.92   -8.00   -6.64   -6.00   -5.42   -4.49
#> log_lik[33]     -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_lik[34]     -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.81   -2.38
#> log_lik[35]     -5.52    0.01 0.69   -7.08   -5.94   -5.47   -5.04   -4.34
#> log_lik[36]     -9.91    0.02 1.20  -12.46  -10.62   -9.82   -9.08   -7.77
#> log_lik[37]     -7.61    0.01 0.81   -9.46   -8.11   -7.53   -7.04   -6.24
#> log_lik[38]     -9.00    0.01 0.85  -10.65   -9.55   -8.99   -8.44   -7.33
#> log_lik[39]     -7.28    0.01 1.10   -9.65   -7.96   -7.21   -6.51   -5.31
#> log_lik[40]     -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_lik[41]     -2.21    0.00 0.30   -2.84   -2.39   -2.19   -2.00   -1.66
#> log_lik[42]     -5.99    0.01 0.62   -7.38   -6.38   -5.95   -5.56   -4.89
#> log_lik[43]     -5.65    0.01 0.53   -6.77   -5.97   -5.61   -5.28   -4.75
#> log_lik[44]     -4.41    0.00 0.39   -5.25   -4.66   -4.38   -4.14   -3.72
#> log_lik[45]     -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.81   -2.38
#> lp__          -463.92    0.18 6.40 -477.43 -467.94 -463.59 -459.35 -452.63
#>               n_eff Rhat
#> pi[1]          9120    1
#> pi[2]          8373    1
#> pi[3]          6289    1
#> pi[4]          5912    1
#> theta[1,1,1]   7740    1
#> theta[1,1,2]   7321    1
#> theta[1,1,3]   4967    1
#> theta[1,1,4]   4585    1
#> theta[1,2,1]   6069    1
#> theta[1,2,2]   6595    1
#> theta[1,2,3]   7350    1
#> theta[1,2,4]   6696    1
#> theta[1,3,1]   5645    1
#> theta[1,3,2]   4519    1
#> theta[1,3,3]   5857    1
#> theta[1,3,4]   5857    1
#> theta[1,4,1]   6092    1
#> theta[1,4,2]   6049    1
#> theta[1,4,3]   4507    1
#> theta[1,4,4]   5836    1
#> theta[2,1,1]   8053    1
#> theta[2,1,2]   6191    1
#> theta[2,1,3]   4913    1
#> theta[2,1,4]   5002    1
#> theta[2,2,1]   7031    1
#> theta[2,2,2]   8788    1
#> theta[2,2,3]   8692    1
#> theta[2,2,4]   7653    1
#> theta[2,3,1]   6712    1
#> theta[2,3,2]   7203    1
#> theta[2,3,3]   7019    1
#> theta[2,3,4]   6510    1
#> theta[2,4,1]   5308    1
#> theta[2,4,2]   7063    1
#> theta[2,4,3]   7007    1
#> theta[2,4,4]   6892    1
#> theta[3,1,1]   8183    1
#> theta[3,1,2]   5802    1
#> theta[3,1,3]   4400    1
#> theta[3,1,4]   4842    1
#> theta[3,2,1]   6833    1
#> theta[3,2,2]   8306    1
#> theta[3,2,3]   6060    1
#> theta[3,2,4]   6352    1
#> theta[3,3,1]   7097    1
#> theta[3,3,2]   7076    1
#> theta[3,3,3]   7817    1
#> theta[3,3,4]   7530    1
#> theta[3,4,1]   6632    1
#> theta[3,4,2]   7125    1
#> theta[3,4,3]   5785    1
#> theta[3,4,4]   6520    1
#> theta[4,1,1]   7738    1
#> theta[4,1,2]   5831    1
#> theta[4,1,3]   4459    1
#> theta[4,1,4]   5037    1
#> theta[4,2,1]   8165    1
#> theta[4,2,2]   7805    1
#> theta[4,2,3]   7142    1
#> theta[4,2,4]   5348    1
#> theta[4,3,1]   6030    1
#> theta[4,3,2]   6334    1
#> theta[4,3,3]   6569    1
#> theta[4,3,4]   6649    1
#> theta[4,4,1]   6941    1
#> theta[4,4,2]   7661    1
#> theta[4,4,3]   6423    1
#> theta[4,4,4]   7453    1
#> theta[5,1,1]   7467    1
#> theta[5,1,2]   5187    1
#> theta[5,1,3]   5548    1
#> theta[5,1,4]   4959    1
#> theta[5,2,1]   8518    1
#> theta[5,2,2]   9356    1
#> theta[5,2,3]   8561    1
#> theta[5,2,4]   5581    1
#> theta[5,3,1]   6884    1
#> theta[5,3,2]   8854    1
#> theta[5,3,3]   7240    1
#> theta[5,3,4]   6166    1
#> theta[5,4,1]   8342    1
#> theta[5,4,2]   6641    1
#> theta[5,4,3]   6125    1
#> theta[5,4,4]   7241    1
#> log_p_z[1,1]   8491    1
#> log_p_z[1,2]   4078    1
#> log_p_z[1,3]   4063    1
#> log_p_z[1,4]   4353    1
#> log_p_z[2,1]   4942    1
#> log_p_z[2,2]   4454    1
#> log_p_z[2,3]   3057    1
#> log_p_z[2,4]   3125    1
#> log_p_z[3,1]   3826    1
#> log_p_z[3,2]   3939    1
#> log_p_z[3,3]   4109    1
#> log_p_z[3,4]   4372    1
#> log_p_z[4,1]   5664    1
#> log_p_z[4,2]   4681    1
#> log_p_z[4,3]   3593    1
#> log_p_z[4,4]   4137    1
#> log_p_z[5,1]   4888    1
#> log_p_z[5,2]   7391    1
#> log_p_z[5,3]   3502    1
#> log_p_z[5,4]   3708    1
#> log_p_z[6,1]   4444    1
#> log_p_z[6,2]   6101    1
#> log_p_z[6,3]   3437    1
#> log_p_z[6,4]   4158    1
#> log_p_z[7,1]   6415    1
#> log_p_z[7,2]   4231    1
#> log_p_z[7,3]   3773    1
#> log_p_z[7,4]   4216    1
#> log_p_z[8,1]   5072    1
#> log_p_z[8,2]   5323    1
#> log_p_z[8,3]   6488    1
#> log_p_z[8,4]   2988    1
#> log_p_z[9,1]   5394    1
#> log_p_z[9,2]   6467    1
#> log_p_z[9,3]   3560    1
#> log_p_z[9,4]   3396    1
#> log_p_z[10,1]  5111    1
#> log_p_z[10,2]  5815    1
#> log_p_z[10,3]  3549    1
#> log_p_z[10,4]  3313    1
#> log_p_z[11,1]  4263    1
#> log_p_z[11,2]  3625    1
#> log_p_z[11,3]  2338    1
#> log_p_z[11,4]  6778    1
#> log_p_z[12,1]  4630    1
#> log_p_z[12,2]  3977    1
#> log_p_z[12,3]  3368    1
#> log_p_z[12,4]  3571    1
#> log_p_z[13,1]  8491    1
#> log_p_z[13,2]  4078    1
#> log_p_z[13,3]  4063    1
#> log_p_z[13,4]  4353    1
#> log_p_z[14,1]  5234    1
#> log_p_z[14,2]  5875    1
#> log_p_z[14,3]  3304    1
#> log_p_z[14,4]  3421    1
#> log_p_z[15,1]  7423    1
#> log_p_z[15,2]  4118    1
#> log_p_z[15,3]  3851    1
#> log_p_z[15,4]  4307    1
#> log_p_z[16,1]  6175    1
#> log_p_z[16,2]  4310    1
#> log_p_z[16,3]  4117    1
#> log_p_z[16,4]  4349    1
#> log_p_z[17,1]  8491    1
#> log_p_z[17,2]  4078    1
#> log_p_z[17,3]  4063    1
#> log_p_z[17,4]  4353    1
#> log_p_z[18,1]  8491    1
#> log_p_z[18,2]  4078    1
#> log_p_z[18,3]  4063    1
#> log_p_z[18,4]  4353    1
#> log_p_z[19,1]  6033    1
#> log_p_z[19,2]  6673    1
#> log_p_z[19,3]  3547    1
#> log_p_z[19,4]  3976    1
#> log_p_z[20,1]  4366    1
#> log_p_z[20,2]  5104    1
#> log_p_z[20,3]  3573    1
#> log_p_z[20,4]  4108    1
#> log_p_z[21,1]  4860    1
#> log_p_z[21,2]  7695    1
#> log_p_z[21,3]  3545    1
#> log_p_z[21,4]  3608    1
#> log_p_z[22,1]  6033    1
#> log_p_z[22,2]  6673    1
#> log_p_z[22,3]  3547    1
#> log_p_z[22,4]  3976    1
#> log_p_z[23,1]  4888    1
#> log_p_z[23,2]  7391    1
#> log_p_z[23,3]  3502    1
#> log_p_z[23,4]  3708    1
#> log_p_z[24,1]  4558    1
#> log_p_z[24,2]  5176    1
#> log_p_z[24,3]  3739    1
#> log_p_z[24,4]  3925    1
#> log_p_z[25,1]  8491    1
#> log_p_z[25,2]  4078    1
#> log_p_z[25,3]  4063    1
#> log_p_z[25,4]  4353    1
#> log_p_z[26,1]  8491    1
#> log_p_z[26,2]  4078    1
#> log_p_z[26,3]  4063    1
#> log_p_z[26,4]  4353    1
#> log_p_z[27,1]  4750    1
#> log_p_z[27,2]  5664    1
#> log_p_z[27,3]  3558    1
#> log_p_z[27,4]  3478    1
#> log_p_z[28,1]  8491    1
#> log_p_z[28,2]  4078    1
#> log_p_z[28,3]  4063    1
#> log_p_z[28,4]  4353    1
#> log_p_z[29,1]  8491    1
#> log_p_z[29,2]  4078    1
#> log_p_z[29,3]  4063    1
#> log_p_z[29,4]  4353    1
#> log_p_z[30,1]  5036    1
#> log_p_z[30,2]  3949    1
#> log_p_z[30,3]  4011    1
#> log_p_z[30,4]  4428    1
#> log_p_z[31,1]  8491    1
#> log_p_z[31,2]  4078    1
#> log_p_z[31,3]  4063    1
#> log_p_z[31,4]  4353    1
#> log_p_z[32,1]  5022    1
#> log_p_z[32,2]  5787    1
#> log_p_z[32,3]  5480    1
#> log_p_z[32,4]  2929    1
#> log_p_z[33,1]  8491    1
#> log_p_z[33,2]  4078    1
#> log_p_z[33,3]  4063    1
#> log_p_z[33,4]  4353    1
#> log_p_z[34,1]  4860    1
#> log_p_z[34,2]  7695    1
#> log_p_z[34,3]  3545    1
#> log_p_z[34,4]  3608    1
#> log_p_z[35,1]  5265    1
#> log_p_z[35,2]  6102    1
#> log_p_z[35,3]  3514    1
#> log_p_z[35,4]  3534    1
#> log_p_z[36,1]  3341    1
#> log_p_z[36,2]  3223    1
#> log_p_z[36,3]  2219    1
#> log_p_z[36,4]  2701    1
#> log_p_z[37,1]  4922    1
#> log_p_z[37,2]  5488    1
#> log_p_z[37,3]  3842    1
#> log_p_z[37,4]  3671    1
#> log_p_z[38,1]  4357    1
#> log_p_z[38,2]  5374    1
#> log_p_z[38,3]  3544    1
#> log_p_z[38,4]  3291    1
#> log_p_z[39,1]  5118    1
#> log_p_z[39,2]  5153    1
#> log_p_z[39,3]  6300    1
#> log_p_z[39,4]  3456    1
#> log_p_z[40,1]  8491    1
#> log_p_z[40,2]  4078    1
#> log_p_z[40,3]  4063    1
#> log_p_z[40,4]  4353    1
#> log_p_z[41,1]  8491    1
#> log_p_z[41,2]  4078    1
#> log_p_z[41,3]  4063    1
#> log_p_z[41,4]  4353    1
#> log_p_z[42,1]  6415    1
#> log_p_z[42,2]  4231    1
#> log_p_z[42,3]  3773    1
#> log_p_z[42,4]  4216    1
#> log_p_z[43,1]  4750    1
#> log_p_z[43,2]  5664    1
#> log_p_z[43,3]  3558    1
#> log_p_z[43,4]  3478    1
#> log_p_z[44,1]  7423    1
#> log_p_z[44,2]  4118    1
#> log_p_z[44,3]  3851    1
#> log_p_z[44,4]  4307    1
#> log_p_z[45,1]  4860    1
#> log_p_z[45,2]  7695    1
#> log_p_z[45,3]  3545    1
#> log_p_z[45,4]  3608    1
#> log_lik[1]     8491    1
#> log_lik[2]     4300    1
#> log_lik[3]     5294    1
#> log_lik[4]     4757    1
#> log_lik[5]     7391    1
#> log_lik[6]     6101    1
#> log_lik[7]     6416    1
#> log_lik[8]     6495    1
#> log_lik[9]     6470    1
#> log_lik[10]    5831    1
#> log_lik[11]    6764    1
#> log_lik[12]    5072    1
#> log_lik[13]    8491    1
#> log_lik[14]    5880    1
#> log_lik[15]    7423    1
#> log_lik[16]    6176    1
#> log_lik[17]    8491    1
#> log_lik[18]    8491    1
#> log_lik[19]    6673    1
#> log_lik[20]    5108    1
#> log_lik[21]    7695    1
#> log_lik[22]    6673    1
#> log_lik[23]    7391    1
#> log_lik[24]    5176    1
#> log_lik[25]    8491    1
#> log_lik[26]    8491    1
#> log_lik[27]    5662    1
#> log_lik[28]    8491    1
#> log_lik[29]    8491    1
#> log_lik[30]    5051    1
#> log_lik[31]    8491    1
#> log_lik[32]    5484    1
#> log_lik[33]    8491    1
#> log_lik[34]    7695    1
#> log_lik[35]    6110    1
#> log_lik[36]    4670    1
#> log_lik[37]    5493    1
#> log_lik[38]    4586    1
#> log_lik[39]    6318    1
#> log_lik[40]    8491    1
#> log_lik[41]    8491    1
#> log_lik[42]    6416    1
#> log_lik[43]    5662    1
#> log_lik[44]    7423    1
#> log_lik[45]    7695    1
#> lp__           1292    1
#> 
#> Samples were drawn using NUTS(diag_e) at Fri Apr  3 15:24:57 2026.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }