Produce simulation data from a 'complete' rating design

Produce simulation data from a 'complete' rating design

make_complete_rating_design_sim_data(I, J, N)

Arguments

I

The number of items.

J

The number of raters.

N

The number of times each rater rates each item.

Value

Simulation data in the format required by simulate_dawid_skene_model() or simulate_hier_dawid_skene_model().

Details

A 'complete' rating design is situation where every rater rates each item the same number of times. In this function the number of times each rater rates each item is N.

Examples

make_complete_rating_design_sim_data(100, 5, 2)
#>      item rater
#> 1       1     1
#> 2       1     2
#> 3       1     3
#> 4       1     4
#> 5       1     5
#> 6       1     1
#> 7       1     2
#> 8       1     3
#> 9       1     4
#> 10      1     5
#> 11      2     1
#> 12      2     2
#> 13      2     3
#> 14      2     4
#> 15      2     5
#> 16      2     1
#> 17      2     2
#> 18      2     3
#> 19      2     4
#> 20      2     5
#> 21      3     1
#> 22      3     2
#> 23      3     3
#> 24      3     4
#> 25      3     5
#> 26      3     1
#> 27      3     2
#> 28      3     3
#> 29      3     4
#> 30      3     5
#> 31      4     1
#> 32      4     2
#> 33      4     3
#> 34      4     4
#> 35      4     5
#> 36      4     1
#> 37      4     2
#> 38      4     3
#> 39      4     4
#> 40      4     5
#> 41      5     1
#> 42      5     2
#> 43      5     3
#> 44      5     4
#> 45      5     5
#> 46      5     1
#> 47      5     2
#> 48      5     3
#> 49      5     4
#> 50      5     5
#> 51      6     1
#> 52      6     2
#> 53      6     3
#> 54      6     4
#> 55      6     5
#> 56      6     1
#> 57      6     2
#> 58      6     3
#> 59      6     4
#> 60      6     5
#> 61      7     1
#> 62      7     2
#> 63      7     3
#> 64      7     4
#> 65      7     5
#> 66      7     1
#> 67      7     2
#> 68      7     3
#> 69      7     4
#> 70      7     5
#> 71      8     1
#> 72      8     2
#> 73      8     3
#> 74      8     4
#> 75      8     5
#> 76      8     1
#> 77      8     2
#> 78      8     3
#> 79      8     4
#> 80      8     5
#> 81      9     1
#> 82      9     2
#> 83      9     3
#> 84      9     4
#> 85      9     5
#> 86      9     1
#> 87      9     2
#> 88      9     3
#> 89      9     4
#> 90      9     5
#> 91     10     1
#> 92     10     2
#> 93     10     3
#> 94     10     4
#> 95     10     5
#> 96     10     1
#> 97     10     2
#> 98     10     3
#> 99     10     4
#> 100    10     5
#> 101    11     1
#> 102    11     2
#> 103    11     3
#> 104    11     4
#> 105    11     5
#> 106    11     1
#> 107    11     2
#> 108    11     3
#> 109    11     4
#> 110    11     5
#> 111    12     1
#> 112    12     2
#> 113    12     3
#> 114    12     4
#> 115    12     5
#> 116    12     1
#> 117    12     2
#> 118    12     3
#> 119    12     4
#> 120    12     5
#> 121    13     1
#> 122    13     2
#> 123    13     3
#> 124    13     4
#> 125    13     5
#> 126    13     1
#> 127    13     2
#> 128    13     3
#> 129    13     4
#> 130    13     5
#> 131    14     1
#> 132    14     2
#> 133    14     3
#> 134    14     4
#> 135    14     5
#> 136    14     1
#> 137    14     2
#> 138    14     3
#> 139    14     4
#> 140    14     5
#> 141    15     1
#> 142    15     2
#> 143    15     3
#> 144    15     4
#> 145    15     5
#> 146    15     1
#> 147    15     2
#> 148    15     3
#> 149    15     4
#> 150    15     5
#> 151    16     1
#> 152    16     2
#> 153    16     3
#> 154    16     4
#> 155    16     5
#> 156    16     1
#> 157    16     2
#> 158    16     3
#> 159    16     4
#> 160    16     5
#> 161    17     1
#> 162    17     2
#> 163    17     3
#> 164    17     4
#> 165    17     5
#> 166    17     1
#> 167    17     2
#> 168    17     3
#> 169    17     4
#> 170    17     5
#> 171    18     1
#> 172    18     2
#> 173    18     3
#> 174    18     4
#> 175    18     5
#> 176    18     1
#> 177    18     2
#> 178    18     3
#> 179    18     4
#> 180    18     5
#> 181    19     1
#> 182    19     2
#> 183    19     3
#> 184    19     4
#> 185    19     5
#> 186    19     1
#> 187    19     2
#> 188    19     3
#> 189    19     4
#> 190    19     5
#> 191    20     1
#> 192    20     2
#> 193    20     3
#> 194    20     4
#> 195    20     5
#> 196    20     1
#> 197    20     2
#> 198    20     3
#> 199    20     4
#> 200    20     5
#> 201    21     1
#> 202    21     2
#> 203    21     3
#> 204    21     4
#> 205    21     5
#> 206    21     1
#> 207    21     2
#> 208    21     3
#> 209    21     4
#> 210    21     5
#> 211    22     1
#> 212    22     2
#> 213    22     3
#> 214    22     4
#> 215    22     5
#> 216    22     1
#> 217    22     2
#> 218    22     3
#> 219    22     4
#> 220    22     5
#> 221    23     1
#> 222    23     2
#> 223    23     3
#> 224    23     4
#> 225    23     5
#> 226    23     1
#> 227    23     2
#> 228    23     3
#> 229    23     4
#> 230    23     5
#> 231    24     1
#> 232    24     2
#> 233    24     3
#> 234    24     4
#> 235    24     5
#> 236    24     1
#> 237    24     2
#> 238    24     3
#> 239    24     4
#> 240    24     5
#> 241    25     1
#> 242    25     2
#> 243    25     3
#> 244    25     4
#> 245    25     5
#> 246    25     1
#> 247    25     2
#> 248    25     3
#> 249    25     4
#> 250    25     5
#> 251    26     1
#> 252    26     2
#> 253    26     3
#> 254    26     4
#> 255    26     5
#> 256    26     1
#> 257    26     2
#> 258    26     3
#> 259    26     4
#> 260    26     5
#> 261    27     1
#> 262    27     2
#> 263    27     3
#> 264    27     4
#> 265    27     5
#> 266    27     1
#> 267    27     2
#> 268    27     3
#> 269    27     4
#> 270    27     5
#> 271    28     1
#> 272    28     2
#> 273    28     3
#> 274    28     4
#> 275    28     5
#> 276    28     1
#> 277    28     2
#> 278    28     3
#> 279    28     4
#> 280    28     5
#> 281    29     1
#> 282    29     2
#> 283    29     3
#> 284    29     4
#> 285    29     5
#> 286    29     1
#> 287    29     2
#> 288    29     3
#> 289    29     4
#> 290    29     5
#> 291    30     1
#> 292    30     2
#> 293    30     3
#> 294    30     4
#> 295    30     5
#> 296    30     1
#> 297    30     2
#> 298    30     3
#> 299    30     4
#> 300    30     5
#> 301    31     1
#> 302    31     2
#> 303    31     3
#> 304    31     4
#> 305    31     5
#> 306    31     1
#> 307    31     2
#> 308    31     3
#> 309    31     4
#> 310    31     5
#> 311    32     1
#> 312    32     2
#> 313    32     3
#> 314    32     4
#> 315    32     5
#> 316    32     1
#> 317    32     2
#> 318    32     3
#> 319    32     4
#> 320    32     5
#> 321    33     1
#> 322    33     2
#> 323    33     3
#> 324    33     4
#> 325    33     5
#> 326    33     1
#> 327    33     2
#> 328    33     3
#> 329    33     4
#> 330    33     5
#> 331    34     1
#> 332    34     2
#> 333    34     3
#> 334    34     4
#> 335    34     5
#> 336    34     1
#> 337    34     2
#> 338    34     3
#> 339    34     4
#> 340    34     5
#> 341    35     1
#> 342    35     2
#> 343    35     3
#> 344    35     4
#> 345    35     5
#> 346    35     1
#> 347    35     2
#> 348    35     3
#> 349    35     4
#> 350    35     5
#> 351    36     1
#> 352    36     2
#> 353    36     3
#> 354    36     4
#> 355    36     5
#> 356    36     1
#> 357    36     2
#> 358    36     3
#> 359    36     4
#> 360    36     5
#> 361    37     1
#> 362    37     2
#> 363    37     3
#> 364    37     4
#> 365    37     5
#> 366    37     1
#> 367    37     2
#> 368    37     3
#> 369    37     4
#> 370    37     5
#> 371    38     1
#> 372    38     2
#> 373    38     3
#> 374    38     4
#> 375    38     5
#> 376    38     1
#> 377    38     2
#> 378    38     3
#> 379    38     4
#> 380    38     5
#> 381    39     1
#> 382    39     2
#> 383    39     3
#> 384    39     4
#> 385    39     5
#> 386    39     1
#> 387    39     2
#> 388    39     3
#> 389    39     4
#> 390    39     5
#> 391    40     1
#> 392    40     2
#> 393    40     3
#> 394    40     4
#> 395    40     5
#> 396    40     1
#> 397    40     2
#> 398    40     3
#> 399    40     4
#> 400    40     5
#> 401    41     1
#> 402    41     2
#> 403    41     3
#> 404    41     4
#> 405    41     5
#> 406    41     1
#> 407    41     2
#> 408    41     3
#> 409    41     4
#> 410    41     5
#> 411    42     1
#> 412    42     2
#> 413    42     3
#> 414    42     4
#> 415    42     5
#> 416    42     1
#> 417    42     2
#> 418    42     3
#> 419    42     4
#> 420    42     5
#> 421    43     1
#> 422    43     2
#> 423    43     3
#> 424    43     4
#> 425    43     5
#> 426    43     1
#> 427    43     2
#> 428    43     3
#> 429    43     4
#> 430    43     5
#> 431    44     1
#> 432    44     2
#> 433    44     3
#> 434    44     4
#> 435    44     5
#> 436    44     1
#> 437    44     2
#> 438    44     3
#> 439    44     4
#> 440    44     5
#> 441    45     1
#> 442    45     2
#> 443    45     3
#> 444    45     4
#> 445    45     5
#> 446    45     1
#> 447    45     2
#> 448    45     3
#> 449    45     4
#> 450    45     5
#> 451    46     1
#> 452    46     2
#> 453    46     3
#> 454    46     4
#> 455    46     5
#> 456    46     1
#> 457    46     2
#> 458    46     3
#> 459    46     4
#> 460    46     5
#> 461    47     1
#> 462    47     2
#> 463    47     3
#> 464    47     4
#> 465    47     5
#> 466    47     1
#> 467    47     2
#> 468    47     3
#> 469    47     4
#> 470    47     5
#> 471    48     1
#> 472    48     2
#> 473    48     3
#> 474    48     4
#> 475    48     5
#> 476    48     1
#> 477    48     2
#> 478    48     3
#> 479    48     4
#> 480    48     5
#> 481    49     1
#> 482    49     2
#> 483    49     3
#> 484    49     4
#> 485    49     5
#> 486    49     1
#> 487    49     2
#> 488    49     3
#> 489    49     4
#> 490    49     5
#> 491    50     1
#> 492    50     2
#> 493    50     3
#> 494    50     4
#> 495    50     5
#> 496    50     1
#> 497    50     2
#> 498    50     3
#> 499    50     4
#> 500    50     5
#> 501    51     1
#> 502    51     2
#> 503    51     3
#> 504    51     4
#> 505    51     5
#> 506    51     1
#> 507    51     2
#> 508    51     3
#> 509    51     4
#> 510    51     5
#> 511    52     1
#> 512    52     2
#> 513    52     3
#> 514    52     4
#> 515    52     5
#> 516    52     1
#> 517    52     2
#> 518    52     3
#> 519    52     4
#> 520    52     5
#> 521    53     1
#> 522    53     2
#> 523    53     3
#> 524    53     4
#> 525    53     5
#> 526    53     1
#> 527    53     2
#> 528    53     3
#> 529    53     4
#> 530    53     5
#> 531    54     1
#> 532    54     2
#> 533    54     3
#> 534    54     4
#> 535    54     5
#> 536    54     1
#> 537    54     2
#> 538    54     3
#> 539    54     4
#> 540    54     5
#> 541    55     1
#> 542    55     2
#> 543    55     3
#> 544    55     4
#> 545    55     5
#> 546    55     1
#> 547    55     2
#> 548    55     3
#> 549    55     4
#> 550    55     5
#> 551    56     1
#> 552    56     2
#> 553    56     3
#> 554    56     4
#> 555    56     5
#> 556    56     1
#> 557    56     2
#> 558    56     3
#> 559    56     4
#> 560    56     5
#> 561    57     1
#> 562    57     2
#> 563    57     3
#> 564    57     4
#> 565    57     5
#> 566    57     1
#> 567    57     2
#> 568    57     3
#> 569    57     4
#> 570    57     5
#> 571    58     1
#> 572    58     2
#> 573    58     3
#> 574    58     4
#> 575    58     5
#> 576    58     1
#> 577    58     2
#> 578    58     3
#> 579    58     4
#> 580    58     5
#> 581    59     1
#> 582    59     2
#> 583    59     3
#> 584    59     4
#> 585    59     5
#> 586    59     1
#> 587    59     2
#> 588    59     3
#> 589    59     4
#> 590    59     5
#> 591    60     1
#> 592    60     2
#> 593    60     3
#> 594    60     4
#> 595    60     5
#> 596    60     1
#> 597    60     2
#> 598    60     3
#> 599    60     4
#> 600    60     5
#> 601    61     1
#> 602    61     2
#> 603    61     3
#> 604    61     4
#> 605    61     5
#> 606    61     1
#> 607    61     2
#> 608    61     3
#> 609    61     4
#> 610    61     5
#> 611    62     1
#> 612    62     2
#> 613    62     3
#> 614    62     4
#> 615    62     5
#> 616    62     1
#> 617    62     2
#> 618    62     3
#> 619    62     4
#> 620    62     5
#> 621    63     1
#> 622    63     2
#> 623    63     3
#> 624    63     4
#> 625    63     5
#> 626    63     1
#> 627    63     2
#> 628    63     3
#> 629    63     4
#> 630    63     5
#> 631    64     1
#> 632    64     2
#> 633    64     3
#> 634    64     4
#> 635    64     5
#> 636    64     1
#> 637    64     2
#> 638    64     3
#> 639    64     4
#> 640    64     5
#> 641    65     1
#> 642    65     2
#> 643    65     3
#> 644    65     4
#> 645    65     5
#> 646    65     1
#> 647    65     2
#> 648    65     3
#> 649    65     4
#> 650    65     5
#> 651    66     1
#> 652    66     2
#> 653    66     3
#> 654    66     4
#> 655    66     5
#> 656    66     1
#> 657    66     2
#> 658    66     3
#> 659    66     4
#> 660    66     5
#> 661    67     1
#> 662    67     2
#> 663    67     3
#> 664    67     4
#> 665    67     5
#> 666    67     1
#> 667    67     2
#> 668    67     3
#> 669    67     4
#> 670    67     5
#> 671    68     1
#> 672    68     2
#> 673    68     3
#> 674    68     4
#> 675    68     5
#> 676    68     1
#> 677    68     2
#> 678    68     3
#> 679    68     4
#> 680    68     5
#> 681    69     1
#> 682    69     2
#> 683    69     3
#> 684    69     4
#> 685    69     5
#> 686    69     1
#> 687    69     2
#> 688    69     3
#> 689    69     4
#> 690    69     5
#> 691    70     1
#> 692    70     2
#> 693    70     3
#> 694    70     4
#> 695    70     5
#> 696    70     1
#> 697    70     2
#> 698    70     3
#> 699    70     4
#> 700    70     5
#> 701    71     1
#> 702    71     2
#> 703    71     3
#> 704    71     4
#> 705    71     5
#> 706    71     1
#> 707    71     2
#> 708    71     3
#> 709    71     4
#> 710    71     5
#> 711    72     1
#> 712    72     2
#> 713    72     3
#> 714    72     4
#> 715    72     5
#> 716    72     1
#> 717    72     2
#> 718    72     3
#> 719    72     4
#> 720    72     5
#> 721    73     1
#> 722    73     2
#> 723    73     3
#> 724    73     4
#> 725    73     5
#> 726    73     1
#> 727    73     2
#> 728    73     3
#> 729    73     4
#> 730    73     5
#> 731    74     1
#> 732    74     2
#> 733    74     3
#> 734    74     4
#> 735    74     5
#> 736    74     1
#> 737    74     2
#> 738    74     3
#> 739    74     4
#> 740    74     5
#> 741    75     1
#> 742    75     2
#> 743    75     3
#> 744    75     4
#> 745    75     5
#> 746    75     1
#> 747    75     2
#> 748    75     3
#> 749    75     4
#> 750    75     5
#> 751    76     1
#> 752    76     2
#> 753    76     3
#> 754    76     4
#> 755    76     5
#> 756    76     1
#> 757    76     2
#> 758    76     3
#> 759    76     4
#> 760    76     5
#> 761    77     1
#> 762    77     2
#> 763    77     3
#> 764    77     4
#> 765    77     5
#> 766    77     1
#> 767    77     2
#> 768    77     3
#> 769    77     4
#> 770    77     5
#> 771    78     1
#> 772    78     2
#> 773    78     3
#> 774    78     4
#> 775    78     5
#> 776    78     1
#> 777    78     2
#> 778    78     3
#> 779    78     4
#> 780    78     5
#> 781    79     1
#> 782    79     2
#> 783    79     3
#> 784    79     4
#> 785    79     5
#> 786    79     1
#> 787    79     2
#> 788    79     3
#> 789    79     4
#> 790    79     5
#> 791    80     1
#> 792    80     2
#> 793    80     3
#> 794    80     4
#> 795    80     5
#> 796    80     1
#> 797    80     2
#> 798    80     3
#> 799    80     4
#> 800    80     5
#> 801    81     1
#> 802    81     2
#> 803    81     3
#> 804    81     4
#> 805    81     5
#> 806    81     1
#> 807    81     2
#> 808    81     3
#> 809    81     4
#> 810    81     5
#> 811    82     1
#> 812    82     2
#> 813    82     3
#> 814    82     4
#> 815    82     5
#> 816    82     1
#> 817    82     2
#> 818    82     3
#> 819    82     4
#> 820    82     5
#> 821    83     1
#> 822    83     2
#> 823    83     3
#> 824    83     4
#> 825    83     5
#> 826    83     1
#> 827    83     2
#> 828    83     3
#> 829    83     4
#> 830    83     5
#> 831    84     1
#> 832    84     2
#> 833    84     3
#> 834    84     4
#> 835    84     5
#> 836    84     1
#> 837    84     2
#> 838    84     3
#> 839    84     4
#> 840    84     5
#> 841    85     1
#> 842    85     2
#> 843    85     3
#> 844    85     4
#> 845    85     5
#> 846    85     1
#> 847    85     2
#> 848    85     3
#> 849    85     4
#> 850    85     5
#> 851    86     1
#> 852    86     2
#> 853    86     3
#> 854    86     4
#> 855    86     5
#> 856    86     1
#> 857    86     2
#> 858    86     3
#> 859    86     4
#> 860    86     5
#> 861    87     1
#> 862    87     2
#> 863    87     3
#> 864    87     4
#> 865    87     5
#> 866    87     1
#> 867    87     2
#> 868    87     3
#> 869    87     4
#> 870    87     5
#> 871    88     1
#> 872    88     2
#> 873    88     3
#> 874    88     4
#> 875    88     5
#> 876    88     1
#> 877    88     2
#> 878    88     3
#> 879    88     4
#> 880    88     5
#> 881    89     1
#> 882    89     2
#> 883    89     3
#> 884    89     4
#> 885    89     5
#> 886    89     1
#> 887    89     2
#> 888    89     3
#> 889    89     4
#> 890    89     5
#> 891    90     1
#> 892    90     2
#> 893    90     3
#> 894    90     4
#> 895    90     5
#> 896    90     1
#> 897    90     2
#> 898    90     3
#> 899    90     4
#> 900    90     5
#> 901    91     1
#> 902    91     2
#> 903    91     3
#> 904    91     4
#> 905    91     5
#> 906    91     1
#> 907    91     2
#> 908    91     3
#> 909    91     4
#> 910    91     5
#> 911    92     1
#> 912    92     2
#> 913    92     3
#> 914    92     4
#> 915    92     5
#> 916    92     1
#> 917    92     2
#> 918    92     3
#> 919    92     4
#> 920    92     5
#> 921    93     1
#> 922    93     2
#> 923    93     3
#> 924    93     4
#> 925    93     5
#> 926    93     1
#> 927    93     2
#> 928    93     3
#> 929    93     4
#> 930    93     5
#> 931    94     1
#> 932    94     2
#> 933    94     3
#> 934    94     4
#> 935    94     5
#> 936    94     1
#> 937    94     2
#> 938    94     3
#> 939    94     4
#> 940    94     5
#> 941    95     1
#> 942    95     2
#> 943    95     3
#> 944    95     4
#> 945    95     5
#> 946    95     1
#> 947    95     2
#> 948    95     3
#> 949    95     4
#> 950    95     5
#> 951    96     1
#> 952    96     2
#> 953    96     3
#> 954    96     4
#> 955    96     5
#> 956    96     1
#> 957    96     2
#> 958    96     3
#> 959    96     4
#> 960    96     5
#> 961    97     1
#> 962    97     2
#> 963    97     3
#> 964    97     4
#> 965    97     5
#> 966    97     1
#> 967    97     2
#> 968    97     3
#> 969    97     4
#> 970    97     5
#> 971    98     1
#> 972    98     2
#> 973    98     3
#> 974    98     4
#> 975    98     5
#> 976    98     1
#> 977    98     2
#> 978    98     3
#> 979    98     4
#> 980    98     5
#> 981    99     1
#> 982    99     2
#> 983    99     3
#> 984    99     4
#> 985    99     5
#> 986    99     1
#> 987    99     2
#> 988    99     3
#> 989    99     4
#> 990    99     5
#> 991   100     1
#> 992   100     2
#> 993   100     3
#> 994   100     4
#> 995   100     5
#> 996   100     1
#> 997   100     2
#> 998   100     3
#> 999   100     4
#> 1000  100     5