Extract posterior intervals for parameters of the model

Extract posterior intervals for parameters of the model

# S3 method for class 'mcmc_fit'
posterior_interval(object, prob = 0.9, pars = c("pi", "theta"), ...)

Arguments

object

A rater mcmc_fit object.

prob

A single probability. The size of the credible interval returned. By default 0.9.

pars

The parameters to calculate the intervals for

...

Other arguments.

Value

A matrix with 2 columns. The first column is the lower bound of of the credible interval and the second is the upper bound. Each row corresponds to one individuals parameters. The rownames are the parameter names.

Details

Posterior intervals can only be calculated for models fit with MCMC. In addition, posterior intervals are not meaningful for the latent class (and indeed cannot be calculated). The full posterior distribution of the latent class can be extracted using class_probabilities

For the class conditional model the 'full' theta parameterisation (i.e. appearing to have the same number of parameters as the standard Dawid-Skene model) is calculated and returned. This is designed to allow easier comparison with the full Dawid-Skene model.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE, chains = 1)

intervals <- posterior_interval(fit)
head(intervals)
#>                        5%       95%
#> pi[1]          0.26918907 0.4915738
#> pi[2]          0.29556245 0.5289795
#> pi[3]          0.06992454 0.2288252
#> pi[4]          0.02603624 0.1354865
#> theta[1, 1, 1] 0.78022840 0.9303925
#> theta[1, 1, 2] 0.04547209 0.1785145

# }