Summarise a mcmc_fit object

Summarise a mcmc_fit object

# S3 method for mcmc_fit
summary(object, n_pars = 8, ...)

Arguments

object
An object of class mcmc_fit.

n_pars
The number of pi/theta parameters and z 'items' to display.

...
Other arguments passed to function.

Details

For the class conditional model the 'full' theta parameterisation (i.e. appearing to have the same number of parameters as the standard Dawid-Skene model) is calculated and returned. This is designed to allow easier comparison with the full Dawid-Skene model.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

summary(fit)

#> Model:
#> Bayesian Dawid and Skene Model 
#> 
#> Prior parameters:
#> 
#> alpha: default
#> beta: default
#> 
#> Fitting method: MCMC
#> 
#> pi/theta samples:
#>                mean   5%  95% Rhat ess_bulk
#> pi[1]          0.37 0.28 0.48    1  9464.96
#> pi[2]          0.41 0.30 0.51    1  9380.45
#> pi[3]          0.14 0.07 0.23    1  7246.93
#> pi[4]          0.07 0.03 0.14    1  7304.06
#> theta[1, 1, 1] 0.87 0.79 0.93    1  7422.21
#> theta[1, 1, 2] 0.10 0.05 0.17    1  7296.78
#> theta[1, 1, 3] 0.02 0.00 0.05    1  6255.92
#> theta[1, 1, 4] 0.02 0.00 0.05    1  6311.21
#> # ... with 76 more rows
#> 
#> z:
#>      MAP Pr(z = 1) Pr(z = 2) Pr(z = 3) Pr(z = 4)
#> z[1]   1      1.00      0.00      0.00      0.00
#> z[2]   3      0.00      0.00      0.97      0.03
#> z[3]   2      0.39      0.61      0.00      0.00
#> z[4]   2      0.01      0.99      0.00      0.00
#> z[5]   2      0.00      1.00      0.00      0.00
#> z[6]   2      0.00      1.00      0.00      0.00
#> z[7]   1      1.00      0.00      0.00      0.00
#> z[8]   3      0.00      0.00      1.00      0.00
#> # ... with 37 more items

# }