Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000219 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 2.19 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.927 seconds (Warm-up)
#> Chain 1:                3.331 seconds (Sampling)
#> Chain 1:                6.258 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000223 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 2.23 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.934 seconds (Warm-up)
#> Chain 2:                3.255 seconds (Sampling)
#> Chain 2:                6.189 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000206 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 2.06 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.941 seconds (Warm-up)
#> Chain 3:                3.231 seconds (Sampling)
#> Chain 3:                6.172 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000217 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 2.17 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 3.014 seconds (Warm-up)
#> Chain 4:                3.279 seconds (Sampling)
#> Chain 4:                6.293 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>     
#>              [,1]         [,2]         [,3]         [,4]
#>   1  9.999993e-01 1.407954e-07 4.135406e-08 4.727187e-07
#>   2  8.230832e-08 2.971135e-05 9.728122e-01 2.715801e-02
#>   3  3.980567e-01 6.013332e-01 1.096513e-04 5.004702e-04
#>   4  6.018433e-03 9.935215e-01 3.094741e-04 1.506196e-04
#>   5  2.726850e-07 9.999667e-01 3.095767e-05 2.093085e-06
#>   6  1.655142e-06 9.994461e-01 5.290609e-04 2.321159e-05
#>   7  9.994192e-01 5.713273e-04 6.737884e-07 8.838817e-06
#>   8  1.089116e-08 3.050502e-05 9.994889e-01 4.806192e-04
#>   9  2.079217e-06 9.999546e-01 3.686165e-05 6.426809e-06
#>   10 3.607867e-06 9.979263e-01 2.028621e-03 4.147840e-05
#>   11 1.611810e-08 1.557385e-08 1.553843e-04 9.998446e-01
#>   12 3.950814e-05 8.825494e-01 1.104029e-01 7.008173e-03
#>   13 9.999993e-01 1.407954e-07 4.135406e-08 4.727187e-07
#>   14 4.562646e-05 9.995130e-01 4.077921e-04 3.362588e-05
#>   15 9.999860e-01 1.226468e-05 1.622270e-07 1.585242e-06
#>   16 9.999906e-01 6.784886e-06 1.141193e-07 2.493189e-06
#>   17 9.999993e-01 1.407954e-07 4.135406e-08 4.727187e-07
#>   18 9.999993e-01 1.407954e-07 4.135406e-08 4.727187e-07
#>   19 4.538518e-05 9.999432e-01 4.538216e-06 6.851142e-06
#>   20 2.056623e-04 9.993011e-01 3.504586e-04 1.427905e-04
#>   21 6.053622e-07 9.999970e-01 1.484585e-06 8.985550e-07
#>   22 4.538518e-05 9.999432e-01 4.538216e-06 6.851142e-06
#>   23 2.726850e-07 9.999667e-01 3.095767e-05 2.093085e-06
#>   24 7.710960e-05 9.999088e-01 6.546968e-06 7.516097e-06
#>   25 9.999993e-01 1.407954e-07 4.135406e-08 4.727187e-07
#>   26 9.999993e-01 1.407954e-07 4.135406e-08 4.727187e-07
#>   27 1.005010e-06 9.999027e-01 9.125616e-05 5.033485e-06
#>   28 9.999993e-01 1.407954e-07 4.135406e-08 4.727187e-07
#>   29 9.999993e-01 1.407954e-07 4.135406e-08 4.727187e-07
#>   30 9.976780e-01 2.293780e-03 2.988780e-06 2.523494e-05
#>   31 9.999993e-01 1.407954e-07 4.135406e-08 4.727187e-07
#>   32 1.626718e-08 9.452432e-04 9.989617e-01 9.306268e-05
#>   33 9.999993e-01 1.407954e-07 4.135406e-08 4.727187e-07
#>   34 6.053622e-07 9.999970e-01 1.484585e-06 8.985550e-07
#>   35 8.698451e-07 9.976903e-01 2.293644e-03 1.522943e-05
#>   36 2.389531e-07 9.386536e-05 7.782265e-01 2.216794e-01
#>   37 1.942480e-04 9.994070e-01 3.481061e-04 5.066793e-05
#>   38 3.188859e-06 7.029816e-01 2.966199e-01 3.952603e-04
#>   39 8.062062e-08 6.518551e-04 9.973660e-01 1.982047e-03
#>   40 9.999993e-01 1.407954e-07 4.135406e-08 4.727187e-07
#>   41 9.999993e-01 1.407954e-07 4.135406e-08 4.727187e-07
#>   42 9.994192e-01 5.713273e-04 6.737884e-07 8.838817e-06
#>   43 1.005010e-06 9.999027e-01 9.125616e-05 5.033485e-06
#>   44 9.999860e-01 1.226468e-05 1.622270e-07 1.585242e-06
#>   45 6.053622e-07 9.999970e-01 1.484585e-06 8.985550e-07

# }