Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000156 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.56 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.71902 seconds (Warm-up)
#> Chain 1:                1.71619 seconds (Sampling)
#> Chain 1:                3.43522 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000151 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.51 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.72491 seconds (Warm-up)
#> Chain 2:                1.82293 seconds (Sampling)
#> Chain 2:                3.54784 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000142 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.42 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.65862 seconds (Warm-up)
#> Chain 3:                1.81096 seconds (Sampling)
#> Chain 3:                3.46957 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000127 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.27 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.60464 seconds (Warm-up)
#> Chain 4:                1.54547 seconds (Sampling)
#> Chain 4:                3.1501 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>        
#>                 [,1]         [,2]         [,3]         [,4]
#>    [1,] 9.999995e-01 1.290518e-07 3.112079e-08 3.629891e-07
#>    [2,] 7.382021e-08 2.119559e-05 9.781606e-01 2.181809e-02
#>    [3,] 3.941004e-01 6.053655e-01 1.278275e-04 4.062142e-04
#>    [4,] 5.613631e-03 9.938317e-01 3.885978e-04 1.660232e-04
#>    [5,] 2.604862e-07 9.999625e-01 3.501611e-05 2.197669e-06
#>    [6,] 1.710111e-06 9.992866e-01 6.876764e-04 2.397855e-05
#>    [7,] 9.994455e-01 5.455236e-04 8.321541e-07 8.138096e-06
#>    [8,] 8.787248e-09 2.646864e-05 9.994509e-01 5.226634e-04
#>    [9,] 1.544246e-06 9.999407e-01 5.022316e-05 7.511426e-06
#>   [10,] 2.586127e-06 9.977838e-01 2.165381e-03 4.827199e-05
#>   [11,] 1.150719e-08 1.971496e-08 1.202525e-04 9.998797e-01
#>   [12,] 3.311803e-05 8.784420e-01 1.143231e-01 7.201809e-03
#>   [13,] 9.999995e-01 1.290518e-07 3.112079e-08 3.629891e-07
#>   [14,] 3.681098e-05 9.994433e-01 4.867069e-04 3.320199e-05
#>   [15,] 9.999875e-01 1.085973e-05 1.444623e-07 1.526639e-06
#>   [16,] 9.999911e-01 6.544719e-06 1.961203e-07 2.161263e-06
#>   [17,] 9.999995e-01 1.290518e-07 3.112079e-08 3.629891e-07
#>   [18,] 9.999995e-01 1.290518e-07 3.112079e-08 3.629891e-07
#>   [19,] 4.025382e-05 9.999502e-01 4.601642e-06 4.993884e-06
#>   [20,] 1.927494e-04 9.993029e-01 4.158221e-04 8.849037e-05
#>   [21,] 5.122983e-07 9.999969e-01 1.702196e-06 8.932713e-07
#>   [22,] 4.025382e-05 9.999502e-01 4.601642e-06 4.993884e-06
#>   [23,] 2.604862e-07 9.999625e-01 3.501611e-05 2.197669e-06
#>   [24,] 8.398268e-05 9.999042e-01 5.468176e-06 6.398426e-06
#>   [25,] 9.999995e-01 1.290518e-07 3.112079e-08 3.629891e-07
#>   [26,] 9.999995e-01 1.290518e-07 3.112079e-08 3.629891e-07
#>   [27,] 8.499538e-07 9.999038e-01 8.975348e-05 5.642161e-06
#>   [28,] 9.999995e-01 1.290518e-07 3.112079e-08 3.629891e-07
#>   [29,] 9.999995e-01 1.290518e-07 3.112079e-08 3.629891e-07
#>   [30,] 9.978361e-01 2.123193e-03 2.743154e-06 3.798100e-05
#>   [31,] 9.999995e-01 1.290518e-07 3.112079e-08 3.629891e-07
#>   [32,] 1.234840e-08 8.757344e-04 9.990139e-01 1.103157e-04
#>   [33,] 9.999995e-01 1.290518e-07 3.112079e-08 3.629891e-07
#>   [34,] 5.122983e-07 9.999969e-01 1.702196e-06 8.932713e-07
#>   [35,] 6.516773e-07 9.976639e-01 2.319201e-03 1.622122e-05
#>   [36,] 2.827192e-07 9.742445e-05 7.840978e-01 2.158045e-01
#>   [37,] 2.057655e-04 9.993455e-01 3.928636e-04 5.586009e-05
#>   [38,] 1.819837e-06 7.125824e-01 2.870453e-01 3.704625e-04
#>   [39,] 6.121084e-08 7.301597e-04 9.973051e-01 1.964709e-03
#>   [40,] 9.999995e-01 1.290518e-07 3.112079e-08 3.629891e-07
#>   [41,] 9.999995e-01 1.290518e-07 3.112079e-08 3.629891e-07
#>   [42,] 9.994455e-01 5.455236e-04 8.321541e-07 8.138096e-06
#>   [43,] 8.499538e-07 9.999038e-01 8.975348e-05 5.642161e-06
#>   [44,] 9.999875e-01 1.085973e-05 1.444623e-07 1.526639e-06
#>   [45,] 5.122983e-07 9.999969e-01 1.702196e-06 8.932713e-07

# }