Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000136 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.36 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.48426 seconds (Warm-up)
#> Chain 1:                1.62937 seconds (Sampling)
#> Chain 1:                3.11363 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.00011 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.1 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.49742 seconds (Warm-up)
#> Chain 2:                1.65163 seconds (Sampling)
#> Chain 2:                3.14904 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000111 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.11 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.53742 seconds (Warm-up)
#> Chain 3:                1.59251 seconds (Sampling)
#> Chain 3:                3.12993 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000109 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.09 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.47425 seconds (Warm-up)
#> Chain 4:                1.52113 seconds (Sampling)
#> Chain 4:                2.99538 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>        
#>                 [,1]         [,2]         [,3]         [,4]
#>    [1,] 9.999993e-01 1.299915e-07 3.211857e-08 4.977331e-07
#>    [2,] 2.179379e-07 2.362855e-05 9.767644e-01 2.321176e-02
#>    [3,] 3.964003e-01 6.030125e-01 1.263398e-04 4.608866e-04
#>    [4,] 6.057832e-03 9.933333e-01 3.794218e-04 2.294794e-04
#>    [5,] 2.844117e-07 9.999637e-01 3.314855e-05 2.826562e-06
#>    [6,] 1.967758e-06 9.994106e-01 5.679773e-04 1.947921e-05
#>    [7,] 9.994413e-01 5.501802e-04 8.725946e-07 7.604723e-06
#>    [8,] 1.004469e-08 2.753048e-05 9.996067e-01 3.657853e-04
#>    [9,] 1.624792e-06 9.999508e-01 4.076193e-05 6.849543e-06
#>   [10,] 2.911567e-06 9.979853e-01 1.965433e-03 4.635294e-05
#>   [11,] 1.889101e-08 1.164414e-08 2.159292e-04 9.997840e-01
#>   [12,] 2.967692e-05 8.843680e-01 1.085103e-01 7.092013e-03
#>   [13,] 9.999993e-01 1.299915e-07 3.211857e-08 4.977331e-07
#>   [14,] 3.717276e-05 9.993802e-01 5.470019e-04 3.564416e-05
#>   [15,] 9.999866e-01 1.170018e-05 1.566761e-07 1.554123e-06
#>   [16,] 9.999907e-01 6.269032e-06 1.700999e-07 2.844972e-06
#>   [17,] 9.999993e-01 1.299915e-07 3.211857e-08 4.977331e-07
#>   [18,] 9.999993e-01 1.299915e-07 3.211857e-08 4.977331e-07
#>   [19,] 4.374518e-05 9.999438e-01 5.828613e-06 6.654205e-06
#>   [20,] 2.053447e-04 9.992809e-01 3.937956e-04 1.200007e-04
#>   [21,] 5.565800e-07 9.999963e-01 1.519662e-06 1.593287e-06
#>   [22,] 4.374518e-05 9.999438e-01 5.828613e-06 6.654205e-06
#>   [23,] 2.844117e-07 9.999637e-01 3.314855e-05 2.826562e-06
#>   [24,] 7.619214e-05 9.999105e-01 5.919399e-06 7.416706e-06
#>   [25,] 9.999993e-01 1.299915e-07 3.211857e-08 4.977331e-07
#>   [26,] 9.999993e-01 1.299915e-07 3.211857e-08 4.977331e-07
#>   [27,] 8.184216e-07 9.999107e-01 7.977058e-05 8.731277e-06
#>   [28,] 9.999993e-01 1.299915e-07 3.211857e-08 4.977331e-07
#>   [29,] 9.999993e-01 1.299915e-07 3.211857e-08 4.977331e-07
#>   [30,] 9.978238e-01 2.147341e-03 2.239372e-06 2.660301e-05
#>   [31,] 9.999993e-01 1.299915e-07 3.211857e-08 4.977331e-07
#>   [32,] 1.040039e-08 1.040660e-03 9.988734e-01 8.595058e-05
#>   [33,] 9.999993e-01 1.299915e-07 3.211857e-08 4.977331e-07
#>   [34,] 5.565800e-07 9.999963e-01 1.519662e-06 1.593287e-06
#>   [35,] 7.716868e-07 9.976444e-01 2.337210e-03 1.758122e-05
#>   [36,] 2.715920e-07 9.991803e-05 7.828571e-01 2.170427e-01
#>   [37,] 1.775557e-04 9.993560e-01 4.131134e-04 5.328707e-05
#>   [38,] 1.962318e-06 7.126224e-01 2.871122e-01 2.634973e-04
#>   [39,] 1.051586e-07 7.111071e-04 9.975528e-01 1.735965e-03
#>   [40,] 9.999993e-01 1.299915e-07 3.211857e-08 4.977331e-07
#>   [41,] 9.999993e-01 1.299915e-07 3.211857e-08 4.977331e-07
#>   [42,] 9.994413e-01 5.501802e-04 8.725946e-07 7.604723e-06
#>   [43,] 8.184216e-07 9.999107e-01 7.977058e-05 8.731277e-06
#>   [44,] 9.999866e-01 1.170018e-05 1.566761e-07 1.554123e-06
#>   [45,] 5.565800e-07 9.999963e-01 1.519662e-06 1.593287e-06

# }