Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.00017 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.7 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.078 seconds (Warm-up)
#> Chain 1:                2.316 seconds (Sampling)
#> Chain 1:                4.394 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.00018 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.8 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.091 seconds (Warm-up)
#> Chain 2:                1.761 seconds (Sampling)
#> Chain 2:                3.852 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000148 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.48 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.171 seconds (Warm-up)
#> Chain 3:                2.143 seconds (Sampling)
#> Chain 3:                4.314 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000147 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.47 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.151 seconds (Warm-up)
#> Chain 4:                2.038 seconds (Sampling)
#> Chain 4:                4.189 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>     
#>              [,1]         [,2]         [,3]         [,4]
#>   1  9.999993e-01 1.472863e-07 2.846919e-08 4.838794e-07
#>   2  7.373793e-08 2.185849e-05 9.772893e-01 2.268876e-02
#>   3  3.962936e-01 6.029390e-01 1.094071e-04 6.579941e-04
#>   4  4.885048e-03 9.945908e-01 3.445277e-04 1.796704e-04
#>   5  2.357576e-07 9.999653e-01 3.145203e-05 2.970788e-06
#>   6  1.500039e-06 9.993406e-01 6.245733e-04 3.334428e-05
#>   7  9.994165e-01 5.744739e-04 8.833326e-07 8.117230e-06
#>   8  1.130296e-08 2.628395e-05 9.995419e-01 4.317568e-04
#>   9  1.651722e-06 9.999477e-01 4.558540e-05 5.067839e-06
#>   10 2.873393e-06 9.976884e-01 2.262862e-03 4.583100e-05
#>   11 1.395789e-08 2.634105e-08 1.644666e-04 9.998355e-01
#>   12 3.203083e-05 8.761850e-01 1.141298e-01 9.653169e-03
#>   13 9.999993e-01 1.472863e-07 2.846919e-08 4.838794e-07
#>   14 3.681102e-05 9.993480e-01 5.685516e-04 4.667762e-05
#>   15 9.999857e-01 1.257584e-05 1.559748e-07 1.547597e-06
#>   16 9.999905e-01 6.895022e-06 1.726257e-07 2.461279e-06
#>   17 9.999993e-01 1.472863e-07 2.846919e-08 4.838794e-07
#>   18 9.999993e-01 1.472863e-07 2.846919e-08 4.838794e-07
#>   19 4.534456e-05 9.999436e-01 4.765327e-06 6.330542e-06
#>   20 1.908704e-04 9.990796e-01 5.962523e-04 1.333106e-04
#>   21 5.902607e-07 9.999966e-01 1.653353e-06 1.121595e-06
#>   22 4.534456e-05 9.999436e-01 4.765327e-06 6.330542e-06
#>   23 2.357576e-07 9.999653e-01 3.145203e-05 2.970788e-06
#>   24 7.758078e-05 9.999106e-01 4.807276e-06 6.964760e-06
#>   25 9.999993e-01 1.472863e-07 2.846919e-08 4.838794e-07
#>   26 9.999993e-01 1.472863e-07 2.846919e-08 4.838794e-07
#>   27 1.004312e-06 9.998982e-01 9.388461e-05 6.920281e-06
#>   28 9.999993e-01 1.472863e-07 2.846919e-08 4.838794e-07
#>   29 9.999993e-01 1.472863e-07 2.846919e-08 4.838794e-07
#>   30 9.980054e-01 1.962963e-03 2.566177e-06 2.905406e-05
#>   31 9.999993e-01 1.472863e-07 2.846919e-08 4.838794e-07
#>   32 2.289322e-08 8.343124e-04 9.990511e-01 1.145689e-04
#>   33 9.999993e-01 1.472863e-07 2.846919e-08 4.838794e-07
#>   34 5.902607e-07 9.999966e-01 1.653353e-06 1.121595e-06
#>   35 6.448815e-07 9.975258e-01 2.424826e-03 4.868527e-05
#>   36 3.622319e-07 1.026297e-04 7.926037e-01 2.072933e-01
#>   37 1.884741e-04 9.993939e-01 3.607190e-04 5.690951e-05
#>   38 2.046534e-06 7.036825e-01 2.959225e-01 3.929188e-04
#>   39 1.016010e-07 6.341330e-04 9.979916e-01 1.374165e-03
#>   40 9.999993e-01 1.472863e-07 2.846919e-08 4.838794e-07
#>   41 9.999993e-01 1.472863e-07 2.846919e-08 4.838794e-07
#>   42 9.994165e-01 5.744739e-04 8.833326e-07 8.117230e-06
#>   43 1.004312e-06 9.998982e-01 9.388461e-05 6.920281e-06
#>   44 9.999857e-01 1.257584e-05 1.559748e-07 1.547597e-06
#>   45 5.902607e-07 9.999966e-01 1.653353e-06 1.121595e-06

# }