Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000162 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.62 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.64812 seconds (Warm-up)
#> Chain 1:                1.68763 seconds (Sampling)
#> Chain 1:                3.33575 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000125 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.25 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.61433 seconds (Warm-up)
#> Chain 2:                1.76309 seconds (Sampling)
#> Chain 2:                3.37742 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000126 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.26 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.6258 seconds (Warm-up)
#> Chain 3:                1.74306 seconds (Sampling)
#> Chain 3:                3.36885 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000128 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.28 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.671 seconds (Warm-up)
#> Chain 4:                1.7066 seconds (Sampling)
#> Chain 4:                3.3776 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>        
#>                 [,1]         [,2]         [,3]         [,4]
#>    [1,] 9.999994e-01 1.315068e-07 2.493904e-08 4.628124e-07
#>    [2,] 8.647702e-08 2.963603e-05 9.747693e-01 2.520094e-02
#>    [3,] 3.915698e-01 6.078604e-01 1.678632e-04 4.019442e-04
#>    [4,] 6.061384e-03 9.934226e-01 3.495045e-04 1.665339e-04
#>    [5,] 2.962197e-07 9.999637e-01 3.275342e-05 3.207457e-06
#>    [6,] 1.690136e-06 9.994290e-01 5.469042e-04 2.241860e-05
#>    [7,] 9.994528e-01 5.399635e-04 7.885522e-07 6.400297e-06
#>    [8,] 1.662240e-08 2.760294e-05 9.995675e-01 4.048739e-04
#>    [9,] 1.983821e-06 9.999415e-01 5.096106e-05 5.516234e-06
#>   [10,] 3.597313e-06 9.978263e-01 2.130703e-03 3.936924e-05
#>   [11,] 1.112097e-08 2.732724e-08 1.501759e-04 9.998498e-01
#>   [12,] 3.377231e-05 8.761076e-01 1.155390e-01 8.319577e-03
#>   [13,] 9.999994e-01 1.315068e-07 2.493904e-08 4.628124e-07
#>   [14,] 3.786757e-05 9.994184e-01 5.076841e-04 3.601858e-05
#>   [15,] 9.999876e-01 1.092818e-05 1.496038e-07 1.350672e-06
#>   [16,] 9.999914e-01 6.266777e-06 1.297688e-07 2.163921e-06
#>   [17,] 9.999994e-01 1.315068e-07 2.493904e-08 4.628124e-07
#>   [18,] 9.999994e-01 1.315068e-07 2.493904e-08 4.628124e-07
#>   [19,] 4.412884e-05 9.999449e-01 5.838306e-06 5.126449e-06
#>   [20,] 1.820373e-04 9.991993e-01 5.120550e-04 1.066431e-04
#>   [21,] 5.666745e-07 9.999966e-01 1.795804e-06 1.002778e-06
#>   [22,] 4.412884e-05 9.999449e-01 5.838306e-06 5.126449e-06
#>   [23,] 2.962197e-07 9.999637e-01 3.275342e-05 3.207457e-06
#>   [24,] 8.015380e-05 9.999061e-01 7.192157e-06 6.573483e-06
#>   [25,] 9.999994e-01 1.315068e-07 2.493904e-08 4.628124e-07
#>   [26,] 9.999994e-01 1.315068e-07 2.493904e-08 4.628124e-07
#>   [27,] 1.008360e-06 9.999050e-01 8.771217e-05 6.293555e-06
#>   [28,] 9.999994e-01 1.315068e-07 2.493904e-08 4.628124e-07
#>   [29,] 9.999994e-01 1.315068e-07 2.493904e-08 4.628124e-07
#>   [30,] 9.980902e-01 1.885769e-03 2.174959e-06 2.183836e-05
#>   [31,] 9.999994e-01 1.315068e-07 2.493904e-08 4.628124e-07
#>   [32,] 2.176792e-08 9.062003e-04 9.989779e-01 1.159146e-04
#>   [33,] 9.999994e-01 1.315068e-07 2.493904e-08 4.628124e-07
#>   [34,] 5.666745e-07 9.999966e-01 1.795804e-06 1.002778e-06
#>   [35,] 8.434907e-07 9.977221e-01 2.259287e-03 1.780192e-05
#>   [36,] 3.898075e-07 1.017402e-04 7.791772e-01 2.207206e-01
#>   [37,] 2.076919e-04 9.993610e-01 3.853538e-04 4.595551e-05
#>   [38,] 3.241630e-06 7.071788e-01 2.924709e-01 3.471004e-04
#>   [39,] 1.351939e-07 6.824049e-04 9.976849e-01 1.632587e-03
#>   [40,] 9.999994e-01 1.315068e-07 2.493904e-08 4.628124e-07
#>   [41,] 9.999994e-01 1.315068e-07 2.493904e-08 4.628124e-07
#>   [42,] 9.994528e-01 5.399635e-04 7.885522e-07 6.400297e-06
#>   [43,] 1.008360e-06 9.999050e-01 8.771217e-05 6.293555e-06
#>   [44,] 9.999876e-01 1.092818e-05 1.496038e-07 1.350672e-06
#>   [45,] 5.666745e-07 9.999966e-01 1.795804e-06 1.002778e-06

# }