Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000144 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.44 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.60469 seconds (Warm-up)
#> Chain 1:                1.47207 seconds (Sampling)
#> Chain 1:                3.07676 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000122 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.22 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.62546 seconds (Warm-up)
#> Chain 2:                1.57009 seconds (Sampling)
#> Chain 2:                3.19555 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.00012 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.2 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.66972 seconds (Warm-up)
#> Chain 3:                1.24774 seconds (Sampling)
#> Chain 3:                2.91746 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000121 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.21 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.60655 seconds (Warm-up)
#> Chain 4:                1.36481 seconds (Sampling)
#> Chain 4:                2.97136 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>     
#>              [,1]         [,2]         [,3]         [,4]
#>   1  9.999994e-01 1.439972e-07 2.172163e-08 4.564853e-07
#>   2  1.065719e-07 3.257519e-05 9.730217e-01 2.694565e-02
#>   3  3.888961e-01 6.105193e-01 7.900928e-05 5.056600e-04
#>   4  5.358190e-03 9.938384e-01 3.667740e-04 4.366725e-04
#>   5  2.549729e-07 9.999646e-01 3.278487e-05 2.357088e-06
#>   6  1.659388e-06 9.994222e-01 5.558562e-04 2.023570e-05
#>   7  9.994203e-01 5.700684e-04 6.368357e-07 8.984989e-06
#>   8  7.307728e-09 2.743163e-05 9.995296e-01 4.429681e-04
#>   9  1.731619e-06 9.999428e-01 4.518187e-05 1.025552e-05
#>   10 2.948586e-06 9.978264e-01 2.112742e-03 5.786815e-05
#>   11 2.121948e-08 2.196601e-08 1.357503e-04 9.998642e-01
#>   12 4.456129e-05 8.839529e-01 1.079341e-01 8.068449e-03
#>   13 9.999994e-01 1.439972e-07 2.172163e-08 4.564853e-07
#>   14 3.410509e-05 9.993095e-01 6.050755e-04 5.129691e-05
#>   15 9.999867e-01 1.170987e-05 1.341899e-07 1.502082e-06
#>   16 9.999904e-01 6.944362e-06 1.240510e-07 2.495467e-06
#>   17 9.999994e-01 1.439972e-07 2.172163e-08 4.564853e-07
#>   18 9.999994e-01 1.439972e-07 2.172163e-08 4.564853e-07
#>   19 4.565593e-05 9.999402e-01 5.226695e-06 8.938046e-06
#>   20 2.386741e-04 9.991802e-01 3.848733e-04 1.962613e-04
#>   21 5.516289e-07 9.999969e-01 1.559433e-06 1.037157e-06
#>   22 4.565593e-05 9.999402e-01 5.226695e-06 8.938046e-06
#>   23 2.549729e-07 9.999646e-01 3.278487e-05 2.357088e-06
#>   24 8.191873e-05 9.999067e-01 4.388183e-06 7.001954e-06
#>   25 9.999994e-01 1.439972e-07 2.172163e-08 4.564853e-07
#>   26 9.999994e-01 1.439972e-07 2.172163e-08 4.564853e-07
#>   27 1.002304e-06 9.999087e-01 8.437714e-05 5.927908e-06
#>   28 9.999994e-01 1.439972e-07 2.172163e-08 4.564853e-07
#>   29 9.999994e-01 1.439972e-07 2.172163e-08 4.564853e-07
#>   30 9.975281e-01 2.437538e-03 1.728144e-06 3.261578e-05
#>   31 9.999994e-01 1.439972e-07 2.172163e-08 4.564853e-07
#>   32 1.435689e-08 9.972241e-04 9.988512e-01 1.515938e-04
#>   33 9.999994e-01 1.439972e-07 2.172163e-08 4.564853e-07
#>   34 5.516289e-07 9.999969e-01 1.559433e-06 1.037157e-06
#>   35 7.700730e-07 9.975575e-01 2.427879e-03 1.388358e-05
#>   36 2.807751e-07 1.138860e-04 7.853467e-01 2.145391e-01
#>   37 2.076272e-04 9.994309e-01 3.174629e-04 4.399321e-05
#>   38 2.243670e-06 7.171499e-01 2.825830e-01 2.649046e-04
#>   39 6.710465e-08 7.201177e-04 9.978795e-01 1.400271e-03
#>   40 9.999994e-01 1.439972e-07 2.172163e-08 4.564853e-07
#>   41 9.999994e-01 1.439972e-07 2.172163e-08 4.564853e-07
#>   42 9.994203e-01 5.700684e-04 6.368357e-07 8.984989e-06
#>   43 1.002304e-06 9.999087e-01 8.437714e-05 5.927908e-06
#>   44 9.999867e-01 1.170987e-05 1.341899e-07 1.502082e-06
#>   45 5.516289e-07 9.999969e-01 1.559433e-06 1.037157e-06

# }