Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000199 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.99 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.585 seconds (Warm-up)
#> Chain 1:                2.749 seconds (Sampling)
#> Chain 1:                5.334 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.00018 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.8 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.609 seconds (Warm-up)
#> Chain 2:                2.741 seconds (Sampling)
#> Chain 2:                5.35 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.00018 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.8 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.661 seconds (Warm-up)
#> Chain 3:                2.796 seconds (Sampling)
#> Chain 3:                5.457 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.00018 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.8 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.628 seconds (Warm-up)
#> Chain 4:                2.738 seconds (Sampling)
#> Chain 4:                5.366 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>     
#>              [,1]         [,2]         [,3]         [,4]
#>   1  9.999995e-01 1.316619e-07 2.546351e-08 3.080018e-07
#>   2  6.089989e-08 2.543976e-05 9.719656e-01 2.800892e-02
#>   3  3.988422e-01 6.005520e-01 9.556306e-05 5.102252e-04
#>   4  6.587147e-03 9.927942e-01 4.490269e-04 1.696393e-04
#>   5  2.415791e-07 9.999662e-01 3.123872e-05 2.307337e-06
#>   6  1.725525e-06 9.993540e-01 6.270954e-04 1.716686e-05
#>   7  9.994666e-01 5.263188e-04 8.055977e-07 6.302096e-06
#>   8  1.379600e-08 2.466445e-05 9.994217e-01 5.536630e-04
#>   9  1.774925e-06 9.999535e-01 3.820200e-05 6.531093e-06
#>   10 3.346588e-06 9.979509e-01 1.994681e-03 5.104572e-05
#>   11 1.880723e-08 1.912512e-08 2.011980e-04 9.997988e-01
#>   12 4.039081e-05 8.847546e-01 1.091720e-01 6.033080e-03
#>   13 9.999995e-01 1.316619e-07 2.546351e-08 3.080018e-07
#>   14 3.676015e-05 9.993046e-01 6.151808e-04 4.344303e-05
#>   15 9.999876e-01 1.116041e-05 1.396975e-07 1.071930e-06
#>   16 9.999916e-01 6.193677e-06 1.751238e-07 1.996855e-06
#>   17 9.999995e-01 1.316619e-07 2.546351e-08 3.080018e-07
#>   18 9.999995e-01 1.316619e-07 2.546351e-08 3.080018e-07
#>   19 4.735143e-05 9.999430e-01 4.113932e-06 5.521331e-06
#>   20 2.411140e-04 9.990945e-01 5.567853e-04 1.076380e-04
#>   21 5.123229e-07 9.999969e-01 1.466188e-06 1.089906e-06
#>   22 4.735143e-05 9.999430e-01 4.113932e-06 5.521331e-06
#>   23 2.415791e-07 9.999662e-01 3.123872e-05 2.307337e-06
#>   24 7.576070e-05 9.999113e-01 4.642997e-06 8.341683e-06
#>   25 9.999995e-01 1.316619e-07 2.546351e-08 3.080018e-07
#>   26 9.999995e-01 1.316619e-07 2.546351e-08 3.080018e-07
#>   27 8.951688e-07 9.999044e-01 8.698964e-05 7.731943e-06
#>   28 9.999995e-01 1.316619e-07 2.546351e-08 3.080018e-07
#>   29 9.999995e-01 1.316619e-07 2.546351e-08 3.080018e-07
#>   30 9.977228e-01 2.256217e-03 2.826021e-06 1.813131e-05
#>   31 9.999995e-01 1.316619e-07 2.546351e-08 3.080018e-07
#>   32 1.010755e-07 9.188755e-04 9.987415e-01 3.395724e-04
#>   33 9.999995e-01 1.316619e-07 2.546351e-08 3.080018e-07
#>   34 5.123229e-07 9.999969e-01 1.466188e-06 1.089906e-06
#>   35 7.307396e-07 9.975202e-01 2.464942e-03 1.416980e-05
#>   36 3.184504e-07 9.431094e-05 7.797525e-01 2.201529e-01
#>   37 2.050969e-04 9.993510e-01 3.900292e-04 5.389410e-05
#>   38 2.965692e-06 7.119307e-01 2.877129e-01 3.533852e-04
#>   39 7.772494e-08 5.897996e-04 9.973417e-01 2.068380e-03
#>   40 9.999995e-01 1.316619e-07 2.546351e-08 3.080018e-07
#>   41 9.999995e-01 1.316619e-07 2.546351e-08 3.080018e-07
#>   42 9.994666e-01 5.263188e-04 8.055977e-07 6.302096e-06
#>   43 8.951688e-07 9.999044e-01 8.698964e-05 7.731943e-06
#>   44 9.999876e-01 1.116041e-05 1.396975e-07 1.071930e-06
#>   45 5.123229e-07 9.999969e-01 1.466188e-06 1.089906e-06

# }