Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000134 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.34 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.64802 seconds (Warm-up)
#> Chain 1:                1.44311 seconds (Sampling)
#> Chain 1:                3.09112 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000129 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.29 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.64434 seconds (Warm-up)
#> Chain 2:                1.7945 seconds (Sampling)
#> Chain 2:                3.43884 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000128 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.28 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.62633 seconds (Warm-up)
#> Chain 3:                1.61165 seconds (Sampling)
#> Chain 3:                3.23798 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000117 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.17 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.6779 seconds (Warm-up)
#> Chain 4:                1.7048 seconds (Sampling)
#> Chain 4:                3.3827 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>     
#>              [,1]         [,2]         [,3]         [,4]
#>   1  9.999993e-01 1.440466e-07 2.808517e-08 5.297568e-07
#>   2  1.073519e-07 2.425047e-05 9.698751e-01 3.010057e-02
#>   3  4.020746e-01 5.971198e-01 1.304318e-04 6.751587e-04
#>   4  6.943514e-03 9.923644e-01 5.657685e-04 1.263283e-04
#>   5  2.462133e-07 9.999603e-01 3.772598e-05 1.698356e-06
#>   6  1.843365e-06 9.992648e-01 7.153759e-04 1.802718e-05
#>   7  9.994304e-01 5.603750e-04 7.329982e-07 8.497748e-06
#>   8  1.290043e-08 3.103659e-05 9.994746e-01 4.943532e-04
#>   9  1.599371e-06 9.999267e-01 6.696589e-05 4.750879e-06
#>   10 3.209595e-06 9.976741e-01 2.279878e-03 4.282254e-05
#>   11 1.272732e-08 1.537846e-08 1.308177e-04 9.998692e-01
#>   12 3.069632e-05 8.774953e-01 1.163436e-01 6.130425e-03
#>   13 9.999993e-01 1.440466e-07 2.808517e-08 5.297568e-07
#>   14 3.266115e-05 9.992574e-01 6.675246e-04 4.244174e-05
#>   15 9.999859e-01 1.246478e-05 1.508264e-07 1.490290e-06
#>   16 9.999906e-01 6.374300e-06 1.496697e-07 2.864971e-06
#>   17 9.999993e-01 1.440466e-07 2.808517e-08 5.297568e-07
#>   18 9.999993e-01 1.440466e-07 2.808517e-08 5.297568e-07
#>   19 4.770024e-05 9.999408e-01 6.410687e-06 5.051717e-06
#>   20 2.745354e-04 9.991469e-01 4.528197e-04 1.257192e-04
#>   21 5.532057e-07 9.999969e-01 1.802088e-06 7.912148e-07
#>   22 4.770024e-05 9.999408e-01 6.410687e-06 5.051717e-06
#>   23 2.462133e-07 9.999603e-01 3.772598e-05 1.698356e-06
#>   24 8.073227e-05 9.999079e-01 5.530343e-06 5.839780e-06
#>   25 9.999993e-01 1.440466e-07 2.808517e-08 5.297568e-07
#>   26 9.999993e-01 1.440466e-07 2.808517e-08 5.297568e-07
#>   27 9.894983e-07 9.999134e-01 7.913202e-05 6.505731e-06
#>   28 9.999993e-01 1.440466e-07 2.808517e-08 5.297568e-07
#>   29 9.999993e-01 1.440466e-07 2.808517e-08 5.297568e-07
#>   30 9.978497e-01 2.123825e-03 1.955024e-06 2.449600e-05
#>   31 9.999993e-01 1.440466e-07 2.808517e-08 5.297568e-07
#>   32 1.439067e-08 9.987165e-04 9.988357e-01 1.655370e-04
#>   33 9.999993e-01 1.440466e-07 2.808517e-08 5.297568e-07
#>   34 5.532057e-07 9.999969e-01 1.802088e-06 7.912148e-07
#>   35 7.250420e-07 9.970739e-01 2.911941e-03 1.348169e-05
#>   36 5.085546e-07 9.116981e-05 7.846960e-01 2.152123e-01
#>   37 1.986373e-04 9.993667e-01 3.872178e-04 4.748125e-05
#>   38 2.427372e-06 7.191131e-01 2.805653e-01 3.191921e-04
#>   39 1.010869e-07 7.787806e-04 9.972213e-01 1.999778e-03
#>   40 9.999993e-01 1.440466e-07 2.808517e-08 5.297568e-07
#>   41 9.999993e-01 1.440466e-07 2.808517e-08 5.297568e-07
#>   42 9.994304e-01 5.603750e-04 7.329982e-07 8.497748e-06
#>   43 9.894983e-07 9.999134e-01 7.913202e-05 6.505731e-06
#>   44 9.999859e-01 1.246478e-05 1.508264e-07 1.490290e-06
#>   45 5.532057e-07 9.999969e-01 1.802088e-06 7.912148e-07

# }