Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000171 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.71 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.124 seconds (Warm-up)
#> Chain 1:                2.165 seconds (Sampling)
#> Chain 1:                4.289 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000151 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.51 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.095 seconds (Warm-up)
#> Chain 2:                1.28 seconds (Sampling)
#> Chain 2:                3.375 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.00015 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.5 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.165 seconds (Warm-up)
#> Chain 3:                2.218 seconds (Sampling)
#> Chain 3:                4.383 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000146 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.46 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.035 seconds (Warm-up)
#> Chain 4:                2.128 seconds (Sampling)
#> Chain 4:                4.163 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>     
#>              [,1]         [,2]         [,3]         [,4]
#>   1  9.999993e-01 1.457001e-07 3.691652e-08 5.367580e-07
#>   2  1.260764e-07 2.636942e-05 9.748151e-01 2.515838e-02
#>   3  3.980229e-01 6.014752e-01 1.214277e-04 3.804942e-04
#>   4  5.806592e-03 9.934980e-01 4.045068e-04 2.908593e-04
#>   5  2.510513e-07 9.999688e-01 2.806918e-05 2.883563e-06
#>   6  1.719867e-06 9.994226e-01 5.559006e-04 1.975533e-05
#>   7  9.994206e-01 5.714798e-04 7.469741e-07 7.201326e-06
#>   8  1.288566e-08 2.628534e-05 9.995193e-01 4.544176e-04
#>   9  1.804099e-06 9.999563e-01 3.679608e-05 5.110613e-06
#>   10 3.413718e-06 9.982576e-01 1.702894e-03 3.611613e-05
#>   11 8.496560e-09 1.139882e-08 1.421615e-04 9.998578e-01
#>   12 3.334163e-05 8.777625e-01 1.147188e-01 7.485403e-03
#>   13 9.999993e-01 1.457001e-07 3.691652e-08 5.367580e-07
#>   14 3.450719e-05 9.994087e-01 5.232761e-04 3.346872e-05
#>   15 9.999868e-01 1.158256e-05 1.303335e-07 1.488570e-06
#>   16 9.999905e-01 7.041046e-06 3.076472e-07 2.130919e-06
#>   17 9.999993e-01 1.457001e-07 3.691652e-08 5.367580e-07
#>   18 9.999993e-01 1.457001e-07 3.691652e-08 5.367580e-07
#>   19 4.716114e-05 9.999438e-01 3.974217e-06 5.110503e-06
#>   20 2.016959e-04 9.993287e-01 3.574895e-04 1.121307e-04
#>   21 6.077088e-07 9.999968e-01 1.516858e-06 1.111634e-06
#>   22 4.716114e-05 9.999438e-01 3.974217e-06 5.110503e-06
#>   23 2.510513e-07 9.999688e-01 2.806918e-05 2.883563e-06
#>   24 8.212833e-05 9.999069e-01 5.095543e-06 5.916941e-06
#>   25 9.999993e-01 1.457001e-07 3.691652e-08 5.367580e-07
#>   26 9.999993e-01 1.457001e-07 3.691652e-08 5.367580e-07
#>   27 1.125334e-06 9.999177e-01 7.522698e-05 5.918695e-06
#>   28 9.999993e-01 1.457001e-07 3.691652e-08 5.367580e-07
#>   29 9.999993e-01 1.457001e-07 3.691652e-08 5.367580e-07
#>   30 9.979999e-01 1.978896e-03 2.685768e-06 1.854111e-05
#>   31 9.999993e-01 1.457001e-07 3.691652e-08 5.367580e-07
#>   32 1.487832e-08 8.612249e-04 9.990268e-01 1.119107e-04
#>   33 9.999993e-01 1.457001e-07 3.691652e-08 5.367580e-07
#>   34 6.077088e-07 9.999968e-01 1.516858e-06 1.111634e-06
#>   35 5.959015e-07 9.976781e-01 2.301631e-03 1.972301e-05
#>   36 4.244343e-07 9.680217e-05 7.867968e-01 2.131060e-01
#>   37 1.797080e-04 9.993880e-01 3.866584e-04 4.561369e-05
#>   38 2.735235e-06 7.130831e-01 2.866406e-01 2.735287e-04
#>   39 1.102152e-07 6.207557e-04 9.975812e-01 1.797898e-03
#>   40 9.999993e-01 1.457001e-07 3.691652e-08 5.367580e-07
#>   41 9.999993e-01 1.457001e-07 3.691652e-08 5.367580e-07
#>   42 9.994206e-01 5.714798e-04 7.469741e-07 7.201326e-06
#>   43 1.125334e-06 9.999177e-01 7.522698e-05 5.918695e-06
#>   44 9.999868e-01 1.158256e-05 1.303335e-07 1.488570e-06
#>   45 6.077088e-07 9.999968e-01 1.516858e-06 1.111634e-06

# }