Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000133 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.33 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.5149 seconds (Warm-up)
#> Chain 1:                1.62112 seconds (Sampling)
#> Chain 1:                3.13602 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000127 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.27 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.4622 seconds (Warm-up)
#> Chain 2:                1.53241 seconds (Sampling)
#> Chain 2:                2.99461 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000126 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.26 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.49131 seconds (Warm-up)
#> Chain 3:                1.64878 seconds (Sampling)
#> Chain 3:                3.14009 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.00012 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.2 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.50818 seconds (Warm-up)
#> Chain 4:                1.64755 seconds (Sampling)
#> Chain 4:                3.15573 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>        
#>                 [,1]         [,2]         [,3]         [,4]
#>    [1,] 9.999994e-01 1.330056e-07 2.877960e-08 4.628535e-07
#>    [2,] 1.660648e-07 2.159661e-05 9.778680e-01 2.211027e-02
#>    [3,] 3.947584e-01 6.045686e-01 9.212774e-05 5.808711e-04
#>    [4,] 6.133376e-03 9.932922e-01 4.551654e-04 1.192117e-04
#>    [5,] 2.737091e-07 9.999534e-01 4.409276e-05 2.250372e-06
#>    [6,] 1.760563e-06 9.990823e-01 8.938790e-04 2.202402e-05
#>    [7,] 9.994651e-01 5.248956e-04 9.096205e-07 9.083260e-06
#>    [8,] 8.666264e-09 2.830870e-05 9.995560e-01 4.157231e-04
#>    [9,] 1.809193e-06 9.999016e-01 8.505136e-05 1.153363e-05
#>   [10,] 3.345551e-06 9.974881e-01 2.470209e-03 3.833088e-05
#>   [11,] 2.226791e-08 1.285569e-08 1.476757e-04 9.998523e-01
#>   [12,] 2.872675e-05 8.750377e-01 1.175458e-01 7.387743e-03
#>   [13,] 9.999994e-01 1.330056e-07 2.877960e-08 4.628535e-07
#>   [14,] 3.938661e-05 9.992943e-01 6.266741e-04 3.961706e-05
#>   [15,] 9.999861e-01 1.169882e-05 1.629754e-07 2.031250e-06
#>   [16,] 9.999914e-01 6.055437e-06 1.434840e-07 2.430987e-06
#>   [17,] 9.999994e-01 1.330056e-07 2.877960e-08 4.628535e-07
#>   [18,] 9.999994e-01 1.330056e-07 2.877960e-08 4.628535e-07
#>   [19,] 4.287338e-05 9.999440e-01 5.449429e-06 7.700159e-06
#>   [20,] 2.116772e-04 9.991567e-01 5.031591e-04 1.284368e-04
#>   [21,] 5.774713e-07 9.999957e-01 1.854960e-06 1.915408e-06
#>   [22,] 4.287338e-05 9.999440e-01 5.449429e-06 7.700159e-06
#>   [23,] 2.737091e-07 9.999534e-01 4.409276e-05 2.250372e-06
#>   [24,] 7.928806e-05 9.999080e-01 5.892814e-06 6.790378e-06
#>   [25,] 9.999994e-01 1.330056e-07 2.877960e-08 4.628535e-07
#>   [26,] 9.999994e-01 1.330056e-07 2.877960e-08 4.628535e-07
#>   [27,] 1.061574e-06 9.999020e-01 9.018100e-05 6.766428e-06
#>   [28,] 9.999994e-01 1.330056e-07 2.877960e-08 4.628535e-07
#>   [29,] 9.999994e-01 1.330056e-07 2.877960e-08 4.628535e-07
#>   [30,] 9.979486e-01 2.023073e-03 2.103120e-06 2.621886e-05
#>   [31,] 9.999994e-01 1.330056e-07 2.877960e-08 4.628535e-07
#>   [32,] 1.860605e-08 1.166259e-03 9.987437e-01 9.006514e-05
#>   [33,] 9.999994e-01 1.330056e-07 2.877960e-08 4.628535e-07
#>   [34,] 5.774713e-07 9.999957e-01 1.854960e-06 1.915408e-06
#>   [35,] 7.032818e-07 9.968768e-01 3.107473e-03 1.506168e-05
#>   [36,] 5.652488e-07 8.628642e-05 7.961906e-01 2.037225e-01
#>   [37,] 1.870355e-04 9.992210e-01 5.465867e-04 4.539824e-05
#>   [38,] 2.177650e-06 7.130616e-01 2.866417e-01 2.945142e-04
#>   [39,] 7.377949e-08 6.231444e-04 9.975696e-01 1.807170e-03
#>   [40,] 9.999994e-01 1.330056e-07 2.877960e-08 4.628535e-07
#>   [41,] 9.999994e-01 1.330056e-07 2.877960e-08 4.628535e-07
#>   [42,] 9.994651e-01 5.248956e-04 9.096205e-07 9.083260e-06
#>   [43,] 1.061574e-06 9.999020e-01 9.018100e-05 6.766428e-06
#>   [44,] 9.999861e-01 1.169882e-05 1.629754e-07 2.031250e-06
#>   [45,] 5.774713e-07 9.999957e-01 1.854960e-06 1.915408e-06

# }