Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000177 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.77 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.28483 seconds (Warm-up)
#> Chain 1:                2.34483 seconds (Sampling)
#> Chain 1:                4.62966 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000156 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.56 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.18862 seconds (Warm-up)
#> Chain 2:                2.4185 seconds (Sampling)
#> Chain 2:                4.60712 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000144 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.44 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.17863 seconds (Warm-up)
#> Chain 3:                1.98121 seconds (Sampling)
#> Chain 3:                4.15984 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000152 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.52 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.18586 seconds (Warm-up)
#> Chain 4:                2.37643 seconds (Sampling)
#> Chain 4:                4.56229 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>        
#>                 [,1]         [,2]         [,3]         [,4]
#>    [1,] 9.999994e-01 1.422138e-07 2.966424e-08 4.253978e-07
#>    [2,] 7.383143e-08 2.097024e-05 9.725267e-01 2.745221e-02
#>    [3,] 3.934293e-01 6.059942e-01 1.197791e-04 4.567322e-04
#>    [4,] 6.810473e-03 9.926475e-01 4.012326e-04 1.407496e-04
#>    [5,] 2.688082e-07 9.999677e-01 2.896386e-05 3.037410e-06
#>    [6,] 1.788485e-06 9.994358e-01 5.463476e-04 1.607386e-05
#>    [7,] 9.993766e-01 6.131644e-04 1.014142e-06 9.193534e-06
#>    [8,] 7.782468e-09 2.784896e-05 9.994712e-01 5.009369e-04
#>    [9,] 1.524079e-06 9.999566e-01 3.532277e-05 6.541346e-06
#>   [10,] 3.156921e-06 9.979777e-01 1.978633e-03 4.054251e-05
#>   [11,] 1.985121e-08 1.847158e-08 1.671159e-04 9.998328e-01
#>   [12,] 3.342514e-05 8.796127e-01 1.133509e-01 7.003039e-03
#>   [13,] 9.999994e-01 1.422138e-07 2.966424e-08 4.253978e-07
#>   [14,] 3.420706e-05 9.994967e-01 4.311196e-04 3.802008e-05
#>   [15,] 9.999866e-01 1.178609e-05 1.761002e-07 1.471471e-06
#>   [16,] 9.999901e-01 7.432185e-06 1.484635e-07 2.286765e-06
#>   [17,] 9.999994e-01 1.422138e-07 2.966424e-08 4.253978e-07
#>   [18,] 9.999994e-01 1.422138e-07 2.966424e-08 4.253978e-07
#>   [19,] 4.334768e-05 9.999454e-01 5.575188e-06 5.645043e-06
#>   [20,] 2.346235e-04 9.993154e-01 3.530208e-04 9.699628e-05
#>   [21,] 5.227247e-07 9.999970e-01 1.483931e-06 1.015241e-06
#>   [22,] 4.334768e-05 9.999454e-01 5.575188e-06 5.645043e-06
#>   [23,] 2.688082e-07 9.999677e-01 2.896386e-05 3.037410e-06
#>   [24,] 7.448441e-05 9.999125e-01 5.625866e-06 7.372599e-06
#>   [25,] 9.999994e-01 1.422138e-07 2.966424e-08 4.253978e-07
#>   [26,] 9.999994e-01 1.422138e-07 2.966424e-08 4.253978e-07
#>   [27,] 9.975389e-07 9.999071e-01 8.573064e-05 6.132857e-06
#>   [28,] 9.999994e-01 1.422138e-07 2.966424e-08 4.253978e-07
#>   [29,] 9.999994e-01 1.422138e-07 2.966424e-08 4.253978e-07
#>   [30,] 9.979612e-01 2.011178e-03 2.417654e-06 2.524848e-05
#>   [31,] 9.999994e-01 1.422138e-07 2.966424e-08 4.253978e-07
#>   [32,] 1.067533e-08 1.101485e-03 9.987843e-01 1.142454e-04
#>   [33,] 9.999994e-01 1.422138e-07 2.966424e-08 4.253978e-07
#>   [34,] 5.227247e-07 9.999970e-01 1.483931e-06 1.015241e-06
#>   [35,] 7.931403e-07 9.977941e-01 2.191278e-03 1.382441e-05
#>   [36,] 4.365573e-07 1.014027e-04 7.874418e-01 2.124564e-01
#>   [37,] 2.507971e-04 9.992303e-01 4.768654e-04 4.202511e-05
#>   [38,] 2.048803e-06 7.143682e-01 2.852763e-01 3.534217e-04
#>   [39,] 7.770924e-08 6.276173e-04 9.976820e-01 1.690282e-03
#>   [40,] 9.999994e-01 1.422138e-07 2.966424e-08 4.253978e-07
#>   [41,] 9.999994e-01 1.422138e-07 2.966424e-08 4.253978e-07
#>   [42,] 9.993766e-01 6.131644e-04 1.014142e-06 9.193534e-06
#>   [43,] 9.975389e-07 9.999071e-01 8.573064e-05 6.132857e-06
#>   [44,] 9.999866e-01 1.178609e-05 1.761002e-07 1.471471e-06
#>   [45,] 5.227247e-07 9.999970e-01 1.483931e-06 1.015241e-06

# }