Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000138 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.38 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.45538 seconds (Warm-up)
#> Chain 1:                1.47425 seconds (Sampling)
#> Chain 1:                2.92964 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000122 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.22 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.50505 seconds (Warm-up)
#> Chain 2:                1.63628 seconds (Sampling)
#> Chain 2:                3.14132 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000119 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.19 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.51114 seconds (Warm-up)
#> Chain 3:                1.41926 seconds (Sampling)
#> Chain 3:                2.9304 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000118 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.18 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.47562 seconds (Warm-up)
#> Chain 4:                1.16296 seconds (Sampling)
#> Chain 4:                2.63858 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>     
#>              [,1]         [,2]         [,3]         [,4]
#>   1  9.999994e-01 1.472442e-07 2.739325e-08 4.678809e-07
#>   2  5.197200e-08 2.012801e-05 9.767490e-01 2.323084e-02
#>   3  3.859479e-01 6.134266e-01 1.417022e-04 4.837918e-04
#>   4  5.353115e-03 9.941744e-01 3.585706e-04 1.139359e-04
#>   5  2.459268e-07 9.999615e-01 3.628958e-05 1.967827e-06
#>   6  1.690606e-06 9.992628e-01 7.129955e-04 2.252476e-05
#>   7  9.993824e-01 5.985614e-04 1.029055e-06 1.801745e-05
#>   8  1.451980e-08 2.358985e-05 9.995488e-01 4.275570e-04
#>   9  1.554223e-06 9.999391e-01 5.328334e-05 6.042107e-06
#>   10 3.138367e-06 9.975596e-01 2.399890e-03 3.735843e-05
#>   11 2.647247e-08 1.784666e-08 1.865906e-04 9.998134e-01
#>   12 3.319322e-05 8.796605e-01 1.120493e-01 8.256960e-03
#>   13 9.999994e-01 1.472442e-07 2.739325e-08 4.678809e-07
#>   14 3.549870e-05 9.994019e-01 5.230800e-04 3.952165e-05
#>   15 9.999854e-01 1.282389e-05 1.386978e-07 1.668177e-06
#>   16 9.999889e-01 6.893869e-06 2.026888e-07 3.995450e-06
#>   17 9.999994e-01 1.472442e-07 2.739325e-08 4.678809e-07
#>   18 9.999994e-01 1.472442e-07 2.739325e-08 4.678809e-07
#>   19 4.303539e-05 9.999467e-01 5.380238e-06 4.863590e-06
#>   20 2.570486e-04 9.991371e-01 5.082772e-04 9.757547e-05
#>   21 5.288087e-07 9.999969e-01 1.682337e-06 9.274514e-07
#>   22 4.303539e-05 9.999467e-01 5.380238e-06 4.863590e-06
#>   23 2.459268e-07 9.999615e-01 3.628958e-05 1.967827e-06
#>   24 7.171008e-05 9.999179e-01 5.807829e-06 4.619910e-06
#>   25 9.999994e-01 1.472442e-07 2.739325e-08 4.678809e-07
#>   26 9.999994e-01 1.472442e-07 2.739325e-08 4.678809e-07
#>   27 1.055956e-06 9.998977e-01 9.561044e-05 5.602608e-06
#>   28 9.999994e-01 1.472442e-07 2.739325e-08 4.678809e-07
#>   29 9.999994e-01 1.472442e-07 2.739325e-08 4.678809e-07
#>   30 9.979749e-01 2.001637e-03 2.053224e-06 2.145573e-05
#>   31 9.999994e-01 1.472442e-07 2.739325e-08 4.678809e-07
#>   32 2.281262e-08 8.795133e-04 9.989823e-01 1.382033e-04
#>   33 9.999994e-01 1.472442e-07 2.739325e-08 4.678809e-07
#>   34 5.288087e-07 9.999969e-01 1.682337e-06 9.274514e-07
#>   35 7.516863e-07 9.975614e-01 2.416920e-03 2.097321e-05
#>   36 3.255512e-07 9.674629e-05 7.829526e-01 2.169503e-01
#>   37 2.033896e-04 9.993052e-01 4.526612e-04 3.874519e-05
#>   38 2.718394e-06 7.102478e-01 2.894646e-01 2.849644e-04
#>   39 1.101169e-07 7.145340e-04 9.972479e-01 2.037418e-03
#>   40 9.999994e-01 1.472442e-07 2.739325e-08 4.678809e-07
#>   41 9.999994e-01 1.472442e-07 2.739325e-08 4.678809e-07
#>   42 9.993824e-01 5.985614e-04 1.029055e-06 1.801745e-05
#>   43 1.055956e-06 9.998977e-01 9.561044e-05 5.602608e-06
#>   44 9.999854e-01 1.282389e-05 1.386978e-07 1.668177e-06
#>   45 5.288087e-07 9.999969e-01 1.682337e-06 9.274514e-07

# }