Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000131 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.31 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.43932 seconds (Warm-up)
#> Chain 1:                1.31874 seconds (Sampling)
#> Chain 1:                2.75806 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000122 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.22 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.49673 seconds (Warm-up)
#> Chain 2:                1.42226 seconds (Sampling)
#> Chain 2:                2.919 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000115 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.15 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.54552 seconds (Warm-up)
#> Chain 3:                1.596 seconds (Sampling)
#> Chain 3:                3.14152 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000112 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.12 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.48889 seconds (Warm-up)
#> Chain 4:                1.16388 seconds (Sampling)
#> Chain 4:                2.65277 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>     
#>              [,1]         [,2]         [,3]         [,4]
#>   1  9.999994e-01 1.409326e-07 3.132768e-08 4.093047e-07
#>   2  1.257272e-07 5.418331e-05 9.759648e-01 2.398089e-02
#>   3  3.988792e-01 6.004135e-01 1.392371e-04 5.680563e-04
#>   4  6.600390e-03 9.927805e-01 3.809291e-04 2.381321e-04
#>   5  2.959102e-07 9.999704e-01 2.692520e-05 2.399422e-06
#>   6  1.669349e-06 9.993209e-01 5.813965e-04 9.600727e-05
#>   7  9.994108e-01 5.781571e-04 7.711744e-07 1.027953e-05
#>   8  1.510944e-08 2.773811e-05 9.993685e-01 6.037377e-04
#>   9  1.662634e-06 9.999342e-01 5.098157e-05 1.316218e-05
#>   10 2.934228e-06 9.977129e-01 2.215179e-03 6.901596e-05
#>   11 1.195322e-08 1.730695e-08 1.452959e-04 9.998547e-01
#>   12 3.300775e-05 8.777241e-01 1.137588e-01 8.484007e-03
#>   13 9.999994e-01 1.409326e-07 3.132768e-08 4.093047e-07
#>   14 5.633460e-05 9.992960e-01 6.001731e-04 4.744973e-05
#>   15 9.999863e-01 1.199503e-05 1.527653e-07 1.507539e-06
#>   16 9.999907e-01 6.731609e-06 1.727813e-07 2.434017e-06
#>   17 9.999994e-01 1.409326e-07 3.132768e-08 4.093047e-07
#>   18 9.999994e-01 1.409326e-07 3.132768e-08 4.093047e-07
#>   19 4.656228e-05 9.999396e-01 3.702930e-06 1.017752e-05
#>   20 2.077589e-04 9.990929e-01 3.406852e-04 3.586668e-04
#>   21 6.133155e-07 9.999963e-01 1.428833e-06 1.622862e-06
#>   22 4.656228e-05 9.999396e-01 3.702930e-06 1.017752e-05
#>   23 2.959102e-07 9.999704e-01 2.692520e-05 2.399422e-06
#>   24 8.082533e-05 9.999060e-01 4.755139e-06 8.455804e-06
#>   25 9.999994e-01 1.409326e-07 3.132768e-08 4.093047e-07
#>   26 9.999994e-01 1.409326e-07 3.132768e-08 4.093047e-07
#>   27 1.007003e-06 9.999114e-01 7.957086e-05 8.023489e-06
#>   28 9.999994e-01 1.409326e-07 3.132768e-08 4.093047e-07
#>   29 9.999994e-01 1.409326e-07 3.132768e-08 4.093047e-07
#>   30 9.981081e-01 1.867635e-03 2.383263e-06 2.187490e-05
#>   31 9.999994e-01 1.409326e-07 3.132768e-08 4.093047e-07
#>   32 2.508129e-08 8.116895e-04 9.990588e-01 1.294404e-04
#>   33 9.999994e-01 1.409326e-07 3.132768e-08 4.093047e-07
#>   34 6.133155e-07 9.999963e-01 1.428833e-06 1.622862e-06
#>   35 8.217492e-07 9.978268e-01 2.155819e-03 1.653665e-05
#>   36 5.473282e-07 1.118786e-04 7.887256e-01 2.111620e-01
#>   37 1.951878e-04 9.993757e-01 3.618204e-04 6.726444e-05
#>   38 2.421860e-06 7.073414e-01 2.922436e-01 4.125780e-04
#>   39 9.358840e-08 6.104233e-04 9.974543e-01 1.935220e-03
#>   40 9.999994e-01 1.409326e-07 3.132768e-08 4.093047e-07
#>   41 9.999994e-01 1.409326e-07 3.132768e-08 4.093047e-07
#>   42 9.994108e-01 5.781571e-04 7.711744e-07 1.027953e-05
#>   43 1.007003e-06 9.999114e-01 7.957086e-05 8.023489e-06
#>   44 9.999863e-01 1.199503e-05 1.527653e-07 1.507539e-06
#>   45 6.133155e-07 9.999963e-01 1.428833e-06 1.622862e-06

# }