Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000134 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.34 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.49214 seconds (Warm-up)
#> Chain 1:                1.64118 seconds (Sampling)
#> Chain 1:                3.13332 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000117 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.17 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.50185 seconds (Warm-up)
#> Chain 2:                1.23606 seconds (Sampling)
#> Chain 2:                2.73791 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000114 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.14 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.49366 seconds (Warm-up)
#> Chain 3:                1.62325 seconds (Sampling)
#> Chain 3:                3.11691 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000113 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.13 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.54773 seconds (Warm-up)
#> Chain 4:                1.64488 seconds (Sampling)
#> Chain 4:                3.19261 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>     
#>              [,1]         [,2]         [,3]         [,4]
#>   1  9.999991e-01 1.365099e-07 3.180492e-08 7.754810e-07
#>   2  7.894780e-08 2.170983e-05 9.762595e-01 2.371869e-02
#>   3  3.865129e-01 6.127135e-01 9.768801e-05 6.759005e-04
#>   4  5.470336e-03 9.937502e-01 6.313474e-04 1.481188e-04
#>   5  2.382666e-07 9.999602e-01 3.505443e-05 4.521609e-06
#>   6  1.506196e-06 9.992111e-01 7.658118e-04 2.160721e-05
#>   7  9.994108e-01 5.781585e-04 7.927904e-07 1.027269e-05
#>   8  8.834633e-09 2.582181e-05 9.996031e-01 3.710754e-04
#>   9  1.977365e-06 9.999410e-01 4.837072e-05 8.688234e-06
#>   10 4.074597e-06 9.977917e-01 2.159840e-03 4.434236e-05
#>   11 1.265597e-08 1.412478e-08 1.236193e-04 9.998764e-01
#>   12 2.637763e-05 8.702948e-01 1.207459e-01 8.932974e-03
#>   13 9.999991e-01 1.365099e-07 3.180492e-08 7.754810e-07
#>   14 3.151666e-05 9.992276e-01 6.953160e-04 4.555436e-05
#>   15 9.999860e-01 1.189850e-05 1.475237e-07 1.953946e-06
#>   16 9.999885e-01 6.708058e-06 1.601742e-07 4.619911e-06
#>   17 9.999991e-01 1.365099e-07 3.180492e-08 7.754810e-07
#>   18 9.999991e-01 1.365099e-07 3.180492e-08 7.754810e-07
#>   19 4.413596e-05 9.999439e-01 4.766774e-06 7.217124e-06
#>   20 2.253897e-04 9.992000e-01 4.368935e-04 1.377395e-04
#>   21 5.772990e-07 9.999958e-01 1.520512e-06 2.110332e-06
#>   22 4.413596e-05 9.999439e-01 4.766774e-06 7.217124e-06
#>   23 2.382666e-07 9.999602e-01 3.505443e-05 4.521609e-06
#>   24 8.204658e-05 9.999037e-01 5.356510e-06 8.935845e-06
#>   25 9.999991e-01 1.365099e-07 3.180492e-08 7.754810e-07
#>   26 9.999991e-01 1.365099e-07 3.180492e-08 7.754810e-07
#>   27 1.180414e-06 9.999047e-01 7.946973e-05 1.469891e-05
#>   28 9.999991e-01 1.365099e-07 3.180492e-08 7.754810e-07
#>   29 9.999991e-01 1.365099e-07 3.180492e-08 7.754810e-07
#>   30 9.980054e-01 1.955254e-03 2.264211e-06 3.705389e-05
#>   31 9.999991e-01 1.365099e-07 3.180492e-08 7.754810e-07
#>   32 1.855562e-08 9.509805e-04 9.989605e-01 8.845326e-05
#>   33 9.999991e-01 1.365099e-07 3.180492e-08 7.754810e-07
#>   34 5.772990e-07 9.999958e-01 1.520512e-06 2.110332e-06
#>   35 6.861460e-07 9.974460e-01 2.536230e-03 1.710710e-05
#>   36 4.081452e-07 1.036958e-04 7.892879e-01 2.106080e-01
#>   37 2.273979e-04 9.994102e-01 3.148768e-04 4.750209e-05
#>   38 2.433027e-06 7.070434e-01 2.926409e-01 3.132893e-04
#>   39 5.809034e-08 6.077822e-04 9.977105e-01 1.681693e-03
#>   40 9.999991e-01 1.365099e-07 3.180492e-08 7.754810e-07
#>   41 9.999991e-01 1.365099e-07 3.180492e-08 7.754810e-07
#>   42 9.994108e-01 5.781585e-04 7.927904e-07 1.027269e-05
#>   43 1.180414e-06 9.999047e-01 7.946973e-05 1.469891e-05
#>   44 9.999860e-01 1.189850e-05 1.475237e-07 1.953946e-06
#>   45 5.772990e-07 9.999958e-01 1.520512e-06 2.110332e-06

# }