Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit
A rater fit object.

...
Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")

#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000317 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 3.17 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.27508 seconds (Warm-up)
#> Chain 1:                2.52212 seconds (Sampling)
#> Chain 1:                4.79721 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000159 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.59 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.29472 seconds (Warm-up)
#> Chain 2:                2.44351 seconds (Sampling)
#> Chain 2:                4.73823 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000161 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.61 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.27582 seconds (Warm-up)
#> Chain 3:                2.47764 seconds (Sampling)
#> Chain 3:                4.75346 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000161 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.61 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.35139 seconds (Warm-up)
#> Chain 4:                2.47823 seconds (Sampling)
#> Chain 4:                4.82962 seconds (Total)
#> Chain 4: 

class_probabilities(fit)

#>        
#>                 [,1]         [,2]         [,3]         [,4]
#>    [1,] 9.999994e-01 1.351267e-07 3.303493e-08 4.300377e-07
#>    [2,] 1.125225e-07 2.227309e-05 9.750261e-01 2.495151e-02
#>    [3,] 3.919821e-01 6.075114e-01 1.065362e-04 4.000471e-04
#>    [4,] 5.925647e-03 9.936282e-01 2.805647e-04 1.655477e-04
#>    [5,] 2.547814e-07 9.999722e-01 2.539779e-05 2.191791e-06
#>    [6,] 1.857638e-06 9.994937e-01 4.858176e-04 1.857864e-05
#>    [7,] 9.994240e-01 5.688069e-04 8.939876e-07 6.293674e-06
#>    [8,] 1.002739e-08 2.767696e-05 9.993662e-01 6.060721e-04
#>    [9,] 1.534765e-06 9.999426e-01 5.046771e-05 5.405998e-06
#>   [10,] 2.677471e-06 9.978748e-01 2.082108e-03 4.042864e-05
#>   [11,] 1.872431e-08 1.566321e-08 1.406805e-04 9.998593e-01
#>   [12,] 3.909727e-05 8.800238e-01 1.131422e-01 6.794929e-03
#>   [13,] 9.999994e-01 1.351267e-07 3.303493e-08 4.300377e-07
#>   [14,] 3.900057e-05 9.994562e-01 4.736460e-04 3.115280e-05
#>   [15,] 9.999864e-01 1.207875e-05 2.428251e-07 1.281038e-06
#>   [16,] 9.999909e-01 6.577026e-06 1.867671e-07 2.290537e-06
#>   [17,] 9.999994e-01 1.351267e-07 3.303493e-08 4.300377e-07
#>   [18,] 9.999994e-01 1.351267e-07 3.303493e-08 4.300377e-07
#>   [19,] 4.315840e-05 9.999473e-01 3.955276e-06 5.553070e-06
#>   [20,] 2.076500e-04 9.993255e-01 3.517796e-04 1.150902e-04
#>   [21,] 5.625113e-07 9.999969e-01 1.515215e-06 1.043711e-06
#>   [22,] 4.315840e-05 9.999473e-01 3.955276e-06 5.553070e-06
#>   [23,] 2.547814e-07 9.999722e-01 2.539779e-05 2.191791e-06
#>   [24,] 8.167570e-05 9.999069e-01 4.972304e-06 6.404288e-06
#>   [25,] 9.999994e-01 1.351267e-07 3.303493e-08 4.300377e-07
#>   [26,] 9.999994e-01 1.351267e-07 3.303493e-08 4.300377e-07
#>   [27,] 9.507733e-07 9.999117e-01 8.166159e-05 5.671055e-06
#>   [28,] 9.999994e-01 1.351267e-07 3.303493e-08 4.300377e-07
#>   [29,] 9.999994e-01 1.351267e-07 3.303493e-08 4.300377e-07
#>   [30,] 9.979030e-01 2.075841e-03 2.717376e-06 1.841885e-05
#>   [31,] 9.999994e-01 1.351267e-07 3.303493e-08 4.300377e-07
#>   [32,] 1.759009e-08 9.937405e-04 9.988407e-01 1.655706e-04
#>   [33,] 9.999994e-01 1.351267e-07 3.303493e-08 4.300377e-07
#>   [34,] 5.625113e-07 9.999969e-01 1.515215e-06 1.043711e-06
#>   [35,] 6.345702e-07 9.978470e-01 2.138232e-03 1.411650e-05
#>   [36,] 5.239812e-07 9.826500e-05 7.912504e-01 2.086508e-01
#>   [37,] 1.964917e-04 9.994311e-01 3.339060e-04 3.845947e-05
#>   [38,] 1.990586e-06 7.058739e-01 2.938541e-01 2.700154e-04
#>   [39,] 1.715981e-07 6.059928e-04 9.977115e-01 1.682327e-03
#>   [40,] 9.999994e-01 1.351267e-07 3.303493e-08 4.300377e-07
#>   [41,] 9.999994e-01 1.351267e-07 3.303493e-08 4.300377e-07
#>   [42,] 9.994240e-01 5.688069e-04 8.939876e-07 6.293674e-06
#>   [43,] 9.507733e-07 9.999117e-01 8.166159e-05 5.671055e-06
#>   [44,] 9.999864e-01 1.207875e-05 2.428251e-07 1.281038e-06
#>   [45,] 5.625113e-07 9.999969e-01 1.515215e-06 1.043711e-06

# }