Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000161 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.61 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.887 seconds (Warm-up)
#> Chain 1:                2.088 seconds (Sampling)
#> Chain 1:                3.975 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000149 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.49 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.916 seconds (Warm-up)
#> Chain 2:                2.018 seconds (Sampling)
#> Chain 2:                3.934 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.00014 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.4 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.955 seconds (Warm-up)
#> Chain 3:                2.082 seconds (Sampling)
#> Chain 3:                4.037 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.00014 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.4 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.88 seconds (Warm-up)
#> Chain 4:                2.095 seconds (Sampling)
#> Chain 4:                3.975 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>     
#>              [,1]         [,2]         [,3]         [,4]
#>   1  9.999994e-01 1.461403e-07 2.816160e-08 4.745009e-07
#>   2  7.958223e-08 3.271043e-05 9.753515e-01 2.461570e-02
#>   3  3.859536e-01 6.135209e-01 9.304595e-05 4.324784e-04
#>   4  6.331197e-03 9.931157e-01 3.520756e-04 2.010571e-04
#>   5  2.511663e-07 9.999664e-01 3.090560e-05 2.397346e-06
#>   6  1.698904e-06 9.993742e-01 6.028297e-04 2.129201e-05
#>   7  9.993811e-01 6.106828e-04 7.985160e-07 7.432167e-06
#>   8  9.542765e-09 2.750100e-05 9.995949e-01 3.775637e-04
#>   9  1.880558e-06 9.999427e-01 4.698893e-05 8.423816e-06
#>   10 3.339822e-06 9.976902e-01 2.258686e-03 4.773746e-05
#>   11 1.099189e-08 9.178537e-09 1.319817e-04 9.998680e-01
#>   12 3.573494e-05 8.691454e-01 1.236711e-01 7.147683e-03
#>   13 9.999994e-01 1.461403e-07 2.816160e-08 4.745009e-07
#>   14 3.391865e-05 9.993295e-01 5.541359e-04 8.247904e-05
#>   15 9.999860e-01 1.240663e-05 1.630801e-07 1.465859e-06
#>   16 9.999902e-01 7.232765e-06 1.394564e-07 2.444059e-06
#>   17 9.999994e-01 1.461403e-07 2.816160e-08 4.745009e-07
#>   18 9.999994e-01 1.461403e-07 2.816160e-08 4.745009e-07
#>   19 4.594036e-05 9.999425e-01 4.439580e-06 7.097365e-06
#>   20 2.092079e-04 9.992364e-01 4.068865e-04 1.475393e-04
#>   21 6.251378e-07 9.999965e-01 1.615917e-06 1.238855e-06
#>   22 4.594036e-05 9.999425e-01 4.439580e-06 7.097365e-06
#>   23 2.511663e-07 9.999664e-01 3.090560e-05 2.397346e-06
#>   24 7.843514e-05 9.999090e-01 5.824043e-06 6.782750e-06
#>   25 9.999994e-01 1.461403e-07 2.816160e-08 4.745009e-07
#>   26 9.999994e-01 1.461403e-07 2.816160e-08 4.745009e-07
#>   27 1.018085e-06 9.999007e-01 8.935687e-05 8.957916e-06
#>   28 9.999994e-01 1.461403e-07 2.816160e-08 4.745009e-07
#>   29 9.999994e-01 1.461403e-07 2.816160e-08 4.745009e-07
#>   30 9.978483e-01 2.128087e-03 2.573365e-06 2.106438e-05
#>   31 9.999994e-01 1.461403e-07 2.816160e-08 4.745009e-07
#>   32 1.301481e-08 8.743108e-04 9.990441e-01 8.153620e-05
#>   33 9.999994e-01 1.461403e-07 2.816160e-08 4.745009e-07
#>   34 6.251378e-07 9.999965e-01 1.615917e-06 1.238855e-06
#>   35 7.298278e-07 9.970861e-01 2.893584e-03 1.959877e-05
#>   36 4.597438e-07 8.404767e-05 7.845776e-01 2.153379e-01
#>   37 2.062378e-04 9.993479e-01 3.991534e-04 4.668169e-05
#>   38 2.992380e-06 6.991133e-01 3.004355e-01 4.481630e-04
#>   39 9.551301e-08 8.263450e-04 9.976452e-01 1.528359e-03
#>   40 9.999994e-01 1.461403e-07 2.816160e-08 4.745009e-07
#>   41 9.999994e-01 1.461403e-07 2.816160e-08 4.745009e-07
#>   42 9.993811e-01 6.106828e-04 7.985160e-07 7.432167e-06
#>   43 1.018085e-06 9.999007e-01 8.935687e-05 8.957916e-06
#>   44 9.999860e-01 1.240663e-05 1.630801e-07 1.465859e-06
#>   45 6.251378e-07 9.999965e-01 1.615917e-06 1.238855e-06

# }