Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000131 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.31 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.48551 seconds (Warm-up)
#> Chain 1:                1.63867 seconds (Sampling)
#> Chain 1:                3.12418 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000121 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.21 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.49969 seconds (Warm-up)
#> Chain 2:                1.63494 seconds (Sampling)
#> Chain 2:                3.13463 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000121 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.21 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.51551 seconds (Warm-up)
#> Chain 3:                1.21901 seconds (Sampling)
#> Chain 3:                2.73452 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000131 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.31 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.52401 seconds (Warm-up)
#> Chain 4:                1.60513 seconds (Sampling)
#> Chain 4:                3.12913 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>     
#>              [,1]         [,2]         [,3]         [,4]
#>   1  9.999994e-01 1.551874e-07 3.526740e-08 3.737883e-07
#>   2  9.186359e-08 2.363977e-05 9.784292e-01 2.154711e-02
#>   3  3.977916e-01 6.016742e-01 1.124049e-04 4.218074e-04
#>   4  5.684773e-03 9.936971e-01 5.078358e-04 1.103133e-04
#>   5  2.615821e-07 9.999709e-01 2.668347e-05 2.167604e-06
#>   6  1.676076e-06 9.993855e-01 5.918222e-04 2.102790e-05
#>   7  9.994053e-01 5.868815e-04 9.453636e-07 6.867257e-06
#>   8  7.873680e-09 2.590354e-05 9.994674e-01 5.066855e-04
#>   9  1.751037e-06 9.999524e-01 4.075057e-05 5.111263e-06
#>   10 3.205380e-06 9.979841e-01 1.975065e-03 3.764076e-05
#>   11 1.089847e-08 1.393195e-08 1.696569e-04 9.998303e-01
#>   12 3.004683e-05 8.819171e-01 1.104102e-01 7.642603e-03
#>   13 9.999994e-01 1.551874e-07 3.526740e-08 3.737883e-07
#>   14 3.773712e-05 9.993924e-01 5.245205e-04 4.533527e-05
#>   15 9.999863e-01 1.236826e-05 1.878106e-07 1.159521e-06
#>   16 9.999905e-01 7.048391e-06 1.630743e-07 2.238875e-06
#>   17 9.999994e-01 1.551874e-07 3.526740e-08 3.737883e-07
#>   18 9.999994e-01 1.551874e-07 3.526740e-08 3.737883e-07
#>   19 4.464849e-05 9.999453e-01 4.547948e-06 5.543530e-06
#>   20 1.989642e-04 9.991970e-01 4.729550e-04 1.310503e-04
#>   21 5.530733e-07 9.999970e-01 1.477095e-06 9.979346e-07
#>   22 4.464849e-05 9.999453e-01 4.547948e-06 5.543530e-06
#>   23 2.615821e-07 9.999709e-01 2.668347e-05 2.167604e-06
#>   24 8.121918e-05 9.999079e-01 5.220729e-06 5.655787e-06
#>   25 9.999994e-01 1.551874e-07 3.526740e-08 3.737883e-07
#>   26 9.999994e-01 1.551874e-07 3.526740e-08 3.737883e-07
#>   27 9.349018e-07 9.999136e-01 7.853910e-05 6.935077e-06
#>   28 9.999994e-01 1.551874e-07 3.526740e-08 3.737883e-07
#>   29 9.999994e-01 1.551874e-07 3.526740e-08 3.737883e-07
#>   30 9.980382e-01 1.940412e-03 2.422912e-06 1.895874e-05
#>   31 9.999994e-01 1.551874e-07 3.526740e-08 3.737883e-07
#>   32 1.462114e-08 8.696920e-04 9.990245e-01 1.058183e-04
#>   33 9.999994e-01 1.551874e-07 3.526740e-08 3.737883e-07
#>   34 5.530733e-07 9.999970e-01 1.477095e-06 9.979346e-07
#>   35 7.254015e-07 9.978329e-01 2.149407e-03 1.693904e-05
#>   36 2.118747e-07 8.296116e-05 7.759385e-01 2.239784e-01
#>   37 2.250682e-04 9.993997e-01 3.302316e-04 4.503067e-05
#>   38 2.130204e-06 7.210465e-01 2.786930e-01 2.583710e-04
#>   39 6.675602e-08 6.524385e-04 9.977645e-01 1.583042e-03
#>   40 9.999994e-01 1.551874e-07 3.526740e-08 3.737883e-07
#>   41 9.999994e-01 1.551874e-07 3.526740e-08 3.737883e-07
#>   42 9.994053e-01 5.868815e-04 9.453636e-07 6.867257e-06
#>   43 9.349018e-07 9.999136e-01 7.853910e-05 6.935077e-06
#>   44 9.999863e-01 1.236826e-05 1.878106e-07 1.159521e-06
#>   45 5.530733e-07 9.999970e-01 1.477095e-06 9.979346e-07

# }