Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit

A rater fit object.

...

Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000175 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.75 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.00783 seconds (Warm-up)
#> Chain 1:                2.1486 seconds (Sampling)
#> Chain 1:                4.15643 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000148 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.48 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.01776 seconds (Warm-up)
#> Chain 2:                2.01065 seconds (Sampling)
#> Chain 2:                4.02841 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000184 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.84 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.01256 seconds (Warm-up)
#> Chain 3:                1.80366 seconds (Sampling)
#> Chain 3:                3.81623 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.00017 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.7 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.0356 seconds (Warm-up)
#> Chain 4:                1.92969 seconds (Sampling)
#> Chain 4:                3.96529 seconds (Total)
#> Chain 4: 
class_probabilities(fit)
#>        
#>                 [,1]         [,2]         [,3]         [,4]
#>    [1,] 9.999994e-01 1.297087e-07 2.705082e-08 4.625074e-07
#>    [2,] 1.070323e-07 2.571551e-05 9.779601e-01 2.201404e-02
#>    [3,] 3.965014e-01 6.029319e-01 9.605870e-05 4.706468e-04
#>    [4,] 5.224894e-03 9.942107e-01 4.080217e-04 1.564247e-04
#>    [5,] 2.469218e-07 9.999678e-01 2.903027e-05 2.881123e-06
#>    [6,] 1.680585e-06 9.994309e-01 5.477876e-04 1.964737e-05
#>    [7,] 9.994631e-01 5.266494e-04 6.950360e-07 9.566214e-06
#>    [8,] 1.067572e-08 2.880016e-05 9.996066e-01 3.646129e-04
#>    [9,] 1.796330e-06 9.999463e-01 4.297343e-05 8.945956e-06
#>   [10,] 3.191401e-06 9.980140e-01 1.930479e-03 5.230513e-05
#>   [11,] 1.649849e-08 2.125791e-08 1.488575e-04 9.998511e-01
#>   [12,] 3.798787e-05 8.736130e-01 1.188457e-01 7.503329e-03
#>   [13,] 9.999994e-01 1.297087e-07 2.705082e-08 4.625074e-07
#>   [14,] 3.052641e-05 9.995313e-01 4.006216e-04 3.756656e-05
#>   [15,] 9.999869e-01 1.150340e-05 1.473221e-07 1.401258e-06
#>   [16,] 9.999903e-01 6.088865e-06 1.448632e-07 3.484922e-06
#>   [17,] 9.999994e-01 1.297087e-07 2.705082e-08 4.625074e-07
#>   [18,] 9.999994e-01 1.297087e-07 2.705082e-08 4.625074e-07
#>   [19,] 4.159997e-05 9.999461e-01 4.771452e-06 7.577423e-06
#>   [20,] 2.077896e-04 9.992764e-01 3.796644e-04 1.361393e-04
#>   [21,] 5.307257e-07 9.999967e-01 1.606721e-06 1.183675e-06
#>   [22,] 4.159997e-05 9.999461e-01 4.771452e-06 7.577423e-06
#>   [23,] 2.469218e-07 9.999678e-01 2.903027e-05 2.881123e-06
#>   [24,] 7.815629e-05 9.999106e-01 4.629560e-06 6.576214e-06
#>   [25,] 9.999994e-01 1.297087e-07 2.705082e-08 4.625074e-07
#>   [26,] 9.999994e-01 1.297087e-07 2.705082e-08 4.625074e-07
#>   [27,] 8.996282e-07 9.999077e-01 8.492189e-05 6.506760e-06
#>   [28,] 9.999994e-01 1.297087e-07 2.705082e-08 4.625074e-07
#>   [29,] 9.999994e-01 1.297087e-07 2.705082e-08 4.625074e-07
#>   [30,] 9.982155e-01 1.765137e-03 3.359134e-06 1.600718e-05
#>   [31,] 9.999994e-01 1.297087e-07 2.705082e-08 4.625074e-07
#>   [32,] 1.481526e-08 9.882385e-04 9.988861e-01 1.256338e-04
#>   [33,] 9.999994e-01 1.297087e-07 2.705082e-08 4.625074e-07
#>   [34,] 5.307257e-07 9.999967e-01 1.606721e-06 1.183675e-06
#>   [35,] 7.230250e-07 9.975621e-01 2.413421e-03 2.373896e-05
#>   [36,] 3.805014e-07 1.039184e-04 7.944337e-01 2.054620e-01
#>   [37,] 1.879236e-04 9.993611e-01 3.764419e-04 7.454448e-05
#>   [38,] 2.182630e-06 7.087648e-01 2.908106e-01 4.224783e-04
#>   [39,] 8.390712e-08 7.525469e-04 9.978106e-01 1.436780e-03
#>   [40,] 9.999994e-01 1.297087e-07 2.705082e-08 4.625074e-07
#>   [41,] 9.999994e-01 1.297087e-07 2.705082e-08 4.625074e-07
#>   [42,] 9.994631e-01 5.266494e-04 6.950360e-07 9.566214e-06
#>   [43,] 8.996282e-07 9.999077e-01 8.492189e-05 6.506760e-06
#>   [44,] 9.999869e-01 1.150340e-05 1.473221e-07 1.401258e-06
#>   [45,] 5.307257e-07 9.999967e-01 1.606721e-06 1.183675e-06

# }