Extract latent class probabilities from a rater fit object

Extract latent class probabilities from a rater fit object

class_probabilities(fit, ...)

# S3 method for mcmc_fit
class_probabilities(fit, ...)

# S3 method for optim_fit
class_probabilities(fit, ...)

Arguments

fit
A rater fit object.

...
Extra arguments.

Value

A I * K matrix where each element is the probably of item i being of class k. (I is the number of items and K the number of classes).

Details

The latent class probabilities are obtained by marginalising out the latent class and then calculating, for each draw of pi and theta, the conditional probability of the latent class given the other parameters and the data. Averaging these conditional probabilities gives the (unconditional) latent class probabilities retuned by this function.

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")

#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000158 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.58 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.31326 seconds (Warm-up)
#> Chain 1:                2.40169 seconds (Sampling)
#> Chain 1:                4.71494 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000156 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.56 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.16739 seconds (Warm-up)
#> Chain 2:                2.35458 seconds (Sampling)
#> Chain 2:                4.52197 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.00015 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.5 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.27806 seconds (Warm-up)
#> Chain 3:                2.45442 seconds (Sampling)
#> Chain 3:                4.73248 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.00015 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.5 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.23406 seconds (Warm-up)
#> Chain 4:                2.30249 seconds (Sampling)
#> Chain 4:                4.53656 seconds (Total)
#> Chain 4: 

class_probabilities(fit)

#>        
#>                 [,1]         [,2]         [,3]         [,4]
#>    [1,] 9.999994e-01 1.513086e-07 2.250443e-08 4.589398e-07
#>    [2,] 1.324118e-07 2.401225e-05 9.724620e-01 2.751383e-02
#>    [3,] 3.967390e-01 6.026154e-01 8.568844e-05 5.598518e-04
#>    [4,] 5.974179e-03 9.933459e-01 5.013417e-04 1.785872e-04
#>    [5,] 2.576392e-07 9.999624e-01 3.446117e-05 2.902222e-06
#>    [6,] 1.759037e-06 9.993465e-01 6.178621e-04 3.384616e-05
#>    [7,] 9.994626e-01 5.285483e-04 6.536575e-07 8.215098e-06
#>    [8,] 8.978779e-09 3.105648e-05 9.994133e-01 5.555960e-04
#>    [9,] 1.900186e-06 9.999387e-01 5.230578e-05 7.109979e-06
#>   [10,] 3.514526e-06 9.975213e-01 2.435034e-03 4.011286e-05
#>   [11,] 1.553142e-08 1.482233e-08 1.141234e-04 9.998858e-01
#>   [12,] 3.519863e-05 8.748544e-01 1.158655e-01 9.244921e-03
#>   [13,] 9.999994e-01 1.513086e-07 2.250443e-08 4.589398e-07
#>   [14,] 3.321280e-05 9.994100e-01 5.172009e-04 3.954810e-05
#>   [15,] 9.999864e-01 1.210200e-05 1.140775e-07 1.404057e-06
#>   [16,] 9.999914e-01 6.093148e-06 1.339187e-07 2.397569e-06
#>   [17,] 9.999994e-01 1.513086e-07 2.250443e-08 4.589398e-07
#>   [18,] 9.999994e-01 1.513086e-07 2.250443e-08 4.589398e-07
#>   [19,] 4.807450e-05 9.999413e-01 4.526289e-06 6.094116e-06
#>   [20,] 2.255543e-04 9.992316e-01 4.139703e-04 1.288558e-04
#>   [21,] 5.794304e-07 9.999967e-01 1.636413e-06 1.099114e-06
#>   [22,] 4.807450e-05 9.999413e-01 4.526289e-06 6.094116e-06
#>   [23,] 2.576392e-07 9.999624e-01 3.446117e-05 2.902222e-06
#>   [24,] 8.160831e-05 9.999055e-01 6.249930e-06 6.685116e-06
#>   [25,] 9.999994e-01 1.513086e-07 2.250443e-08 4.589398e-07
#>   [26,] 9.999994e-01 1.513086e-07 2.250443e-08 4.589398e-07
#>   [27,] 9.432000e-07 9.999001e-01 9.235778e-05 6.618282e-06
#>   [28,] 9.999994e-01 1.513086e-07 2.250443e-08 4.589398e-07
#>   [29,] 9.999994e-01 1.513086e-07 2.250443e-08 4.589398e-07
#>   [30,] 9.979335e-01 2.042208e-03 1.511844e-06 2.275775e-05
#>   [31,] 9.999994e-01 1.513086e-07 2.250443e-08 4.589398e-07
#>   [32,] 1.409713e-08 1.118664e-03 9.987543e-01 1.269730e-04
#>   [33,] 9.999994e-01 1.513086e-07 2.250443e-08 4.589398e-07
#>   [34,] 5.794304e-07 9.999967e-01 1.636413e-06 1.099114e-06
#>   [35,] 7.413719e-07 9.975069e-01 2.469749e-03 2.259642e-05
#>   [36,] 2.892119e-07 9.857067e-05 7.778991e-01 2.220021e-01
#>   [37,] 1.945542e-04 9.994015e-01 3.652721e-04 3.866016e-05
#>   [38,] 1.942063e-06 7.093638e-01 2.901984e-01 4.358631e-04
#>   [39,] 8.806004e-08 7.609073e-04 9.973798e-01 1.859190e-03
#>   [40,] 9.999994e-01 1.513086e-07 2.250443e-08 4.589398e-07
#>   [41,] 9.999994e-01 1.513086e-07 2.250443e-08 4.589398e-07
#>   [42,] 9.994626e-01 5.285483e-04 6.536575e-07 8.215098e-06
#>   [43,] 9.432000e-07 9.999001e-01 9.235778e-05 6.618282e-06
#>   [44,] 9.999864e-01 1.210200e-05 1.140775e-07 1.404057e-06
#>   [45,] 5.794304e-07 9.999967e-01 1.636413e-06 1.099114e-06

# }