Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.26    0.33    0.37    0.42    0.50
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.09    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.76    0.84    0.87    0.90    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.83    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.05
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.17
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.29
#> theta[1,3,3]     0.79    0.00 0.10    0.58    0.73    0.80    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.05    0.10    0.18    0.40
#> theta[1,4,4]     0.69    0.00 0.13    0.40    0.60    0.70    0.79    0.91
#> theta[2,1,1]     0.75    0.00 0.08    0.58    0.70    0.76    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.11    0.16    0.22    0.33
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.59    0.65    0.76
#> theta[2,2,3]     0.30    0.00 0.09    0.15    0.24    0.30    0.36    0.48
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.26
#> theta[2,3,3]     0.62    0.00 0.13    0.35    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.29    0.46
#> theta[2,4,1]     0.11    0.00 0.10    0.00    0.04    0.08    0.16    0.37
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.55    0.67    0.77    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.09    0.55    0.67    0.73    0.79    0.88
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.28
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[3,3,1]     0.08    0.00 0.08    0.00    0.03    0.06    0.11    0.28
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.17    0.24    0.42
#> theta[3,3,3]     0.51    0.00 0.13    0.25    0.41    0.50    0.60    0.76
#> theta[3,3,4]     0.23    0.00 0.11    0.05    0.15    0.22    0.30    0.49
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[3,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[3,4,3]     0.14    0.00 0.12    0.00    0.05    0.11    0.21    0.44
#> theta[3,4,4]     0.63    0.00 0.16    0.30    0.52    0.64    0.75    0.90
#> theta[4,1,1]     0.83    0.00 0.08    0.66    0.78    0.83    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[4,2,2]     0.72    0.00 0.09    0.54    0.67    0.73    0.78    0.88
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.08    0.12    0.17    0.29
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.10    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.26
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,3]     0.69    0.00 0.12    0.42    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.38
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.32    0.55    0.66    0.76    0.90
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.83    0.87    0.91    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.10    0.14    0.19    0.30
#> theta[5,2,2]     0.65    0.00 0.09    0.46    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.33
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.09    0.14    0.21    0.38
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.62    0.71    0.84
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.13    0.21    0.38
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[5,4,3]     0.14    0.00 0.11    0.00    0.05    0.11    0.20    0.43
#> theta[5,4,4]     0.63    0.00 0.15    0.32    0.52    0.64    0.75    0.89
#> log_p_z[1,1]    -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_p_z[1,2]   -20.05    0.04 2.34  -25.21  -21.47  -19.85  -18.42  -16.08
#> log_p_z[1,3]   -24.94    0.08 4.44  -35.06  -27.58  -24.46  -21.73  -17.69
#> log_p_z[1,4]   -21.91    0.07 4.24  -31.68  -24.29  -21.43  -18.92  -15.04
#> log_p_z[2,1]   -29.76    0.07 4.51  -40.04  -32.33  -29.20  -26.57  -22.53
#> log_p_z[2,2]   -21.08    0.04 2.43  -26.44  -22.56  -20.89  -19.40  -16.90
#> log_p_z[2,3]    -7.67    0.02 1.22  -10.36   -8.39   -7.56   -6.82   -5.57
#> log_p_z[2,4]   -16.36    0.08 4.18  -26.08  -18.68  -15.71  -13.43   -9.67
#> log_p_z[3,1]   -11.61    0.03 1.65  -15.40  -12.60  -11.40  -10.45   -8.90
#> log_p_z[3,2]   -10.83    0.03 1.52  -14.22  -11.71  -10.65   -9.76   -8.41
#> log_p_z[3,3]   -23.34    0.07 3.46  -30.92  -25.43  -23.00  -20.95  -17.44
#> log_p_z[3,4]   -21.95    0.06 3.58  -29.80  -24.11  -21.63  -19.39  -15.92
#> log_p_z[4,1]   -14.68    0.03 1.84  -18.70  -15.81  -14.53  -13.39  -11.46
#> log_p_z[4,2]    -7.57    0.01 0.93   -9.55   -8.14   -7.50   -6.92   -5.92
#> log_p_z[4,3]   -20.61    0.08 4.25  -30.20  -23.03  -20.09  -17.60  -13.85
#> log_p_z[4,4]   -21.96    0.08 4.47  -32.15  -24.45  -21.44  -18.77  -14.89
#> log_p_z[5,1]   -21.49    0.04 2.59  -27.15  -23.08  -21.33  -19.69  -16.88
#> log_p_z[5,2]    -3.72    0.01 0.44   -4.63   -4.00   -3.70   -3.41   -2.96
#> log_p_z[5,3]   -18.52    0.08 4.19  -27.81  -20.89  -17.97  -15.52  -12.07
#> log_p_z[5,4]   -21.94    0.08 4.48  -32.13  -24.40  -21.34  -18.81  -14.80
#> log_p_z[6,1]   -21.46    0.04 2.58  -26.96  -23.10  -21.21  -19.62  -17.03
#> log_p_z[6,2]    -5.49    0.01 0.68   -6.92   -5.94   -5.44   -5.01   -4.27
#> log_p_z[6,3]   -17.33    0.08 4.04  -26.81  -19.56  -16.84  -14.46  -11.14
#> log_p_z[6,4]   -21.73    0.08 4.46  -32.17  -24.23  -21.20  -18.63  -14.66
#> log_p_z[7,1]    -6.01    0.01 0.62   -7.32   -6.38   -5.95   -5.57   -4.95
#> log_p_z[7,2]   -14.97    0.03 1.72  -18.84  -16.03  -14.82  -13.75  -11.98
#> log_p_z[7,3]   -24.33    0.07 3.62  -32.34  -26.52  -23.99  -21.79  -18.21
#> log_p_z[7,4]   -21.90    0.06 3.53  -29.93  -24.00  -21.56  -19.40  -16.10
#> log_p_z[8,1]   -29.80    0.07 4.46  -39.87  -32.40  -29.35  -26.62  -22.57
#> log_p_z[8,2]   -18.34    0.03 2.15  -23.13  -19.61  -18.12  -16.86  -14.62
#> log_p_z[8,3]    -5.74    0.01 0.83   -7.53   -6.24   -5.69   -5.17   -4.27
#> log_p_z[8,4]   -18.54    0.08 4.25  -28.59  -21.03  -18.03  -15.53  -11.70
#> log_p_z[9,1]   -20.02    0.03 2.24  -24.91  -21.46  -19.85  -18.44  -16.09
#> log_p_z[9,2]    -4.52    0.01 0.60   -5.81   -4.90   -4.46   -4.10   -3.47
#> log_p_z[9,3]   -19.47    0.09 4.22  -28.83  -21.87  -18.88  -16.48  -12.92
#> log_p_z[9,4]   -21.80    0.09 4.37  -31.55  -24.25  -21.28  -18.72  -14.93
#> log_p_z[10,1]  -22.28    0.04 2.42  -27.58  -23.80  -22.08  -20.55  -18.15
#> log_p_z[10,2]   -7.13    0.01 0.73   -8.68   -7.57   -7.07   -6.61   -5.86
#> log_p_z[10,3]  -16.69    0.06 3.08  -23.59  -18.46  -16.29  -14.48  -11.72
#> log_p_z[10,4]  -21.45    0.07 3.57  -29.16  -23.61  -21.14  -18.86  -15.65
#> log_p_z[11,1]  -29.77    0.07 4.43  -39.78  -32.24  -29.31  -26.65  -22.58
#> log_p_z[11,2]  -29.41    0.08 4.47  -39.80  -31.91  -28.82  -26.28  -22.37
#> log_p_z[11,3]  -18.23    0.06 3.36  -26.12  -19.97  -17.83  -15.83  -13.12
#> log_p_z[11,4]   -5.83    0.01 0.98   -8.00   -6.43   -5.76   -5.14   -4.14
#> log_p_z[12,1]  -22.98    0.04 2.68  -28.66  -24.64  -22.78  -21.07  -18.22
#> log_p_z[12,2]   -9.62    0.02 1.28  -12.57  -10.37   -9.47   -8.73   -7.52
#> log_p_z[12,3]  -15.00    0.08 3.99  -24.00  -17.24  -14.48  -12.18   -8.91
#> log_p_z[12,4]  -19.35    0.08 4.31  -29.37  -21.77  -18.83  -16.29  -12.58
#> log_p_z[13,1]   -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_p_z[13,2]  -20.05    0.04 2.34  -25.21  -21.47  -19.85  -18.42  -16.08
#> log_p_z[13,3]  -24.94    0.08 4.44  -35.06  -27.58  -24.46  -21.73  -17.69
#> log_p_z[13,4]  -21.91    0.07 4.24  -31.68  -24.29  -21.43  -18.92  -15.04
#> log_p_z[14,1]  -19.09    0.04 2.42  -24.35  -20.56  -18.92  -17.38  -14.84
#> log_p_z[14,2]   -6.25    0.01 0.90   -8.23   -6.79   -6.17   -5.60   -4.74
#> log_p_z[14,3]  -18.57    0.08 4.18  -28.04  -20.92  -18.07  -15.59  -12.09
#> log_p_z[14,4]  -21.88    0.08 4.39  -31.51  -24.36  -21.40  -18.80  -14.95
#> log_p_z[15,1]   -4.41    0.01 0.40   -5.26   -4.65   -4.38   -4.13   -3.72
#> log_p_z[15,2]  -17.27    0.03 1.86  -21.34  -18.41  -17.16  -15.97  -14.00
#> log_p_z[15,3]  -24.35    0.06 3.62  -32.36  -26.55  -24.04  -21.72  -18.13
#> log_p_z[15,4]  -21.93    0.06 3.49  -29.85  -24.02  -21.56  -19.46  -16.12
#> log_p_z[16,1]   -3.80    0.01 0.56   -5.03   -4.13   -3.75   -3.41   -2.87
#> log_p_z[16,2]  -17.75    0.04 2.23  -22.79  -19.11  -17.55  -16.20  -13.93
#> log_p_z[16,3]  -24.93    0.09 4.45  -34.93  -27.57  -24.47  -21.80  -17.56
#> log_p_z[16,4]  -21.88    0.07 4.28  -31.59  -24.34  -21.34  -18.93  -14.95
#> log_p_z[17,1]   -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_p_z[17,2]  -20.05    0.04 2.34  -25.21  -21.47  -19.85  -18.42  -16.08
#> log_p_z[17,3]  -24.94    0.08 4.44  -35.06  -27.58  -24.46  -21.73  -17.69
#> log_p_z[17,4]  -21.91    0.07 4.24  -31.68  -24.29  -21.43  -18.92  -15.04
#> log_p_z[18,1]   -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_p_z[18,2]  -20.05    0.04 2.34  -25.21  -21.47  -19.85  -18.42  -16.08
#> log_p_z[18,3]  -24.94    0.08 4.44  -35.06  -27.58  -24.46  -21.73  -17.69
#> log_p_z[18,4]  -21.91    0.07 4.24  -31.68  -24.29  -21.43  -18.92  -15.04
#> log_p_z[19,1]  -16.42    0.03 1.94  -20.61  -17.62  -16.30  -15.08  -13.02
#> log_p_z[19,2]   -4.65    0.01 0.61   -5.96   -5.04   -4.59   -4.22   -3.59
#> log_p_z[19,3]  -21.95    0.08 4.31  -31.79  -24.43  -21.40  -18.89  -15.05
#> log_p_z[19,4]  -21.97    0.09 4.41  -32.01  -24.41  -21.43  -18.91  -14.95
#> log_p_z[20,1]  -18.01    0.03 2.26  -22.94  -19.44  -17.78  -16.43  -14.18
#> log_p_z[20,2]   -7.10    0.01 0.96   -9.26   -7.69   -7.02   -6.42   -5.43
#> log_p_z[20,3]  -19.83    0.08 4.15  -29.28  -22.15  -19.29  -16.90  -13.41
#> log_p_z[20,4]  -21.79    0.09 4.41  -31.84  -24.19  -21.31  -18.72  -14.71
#> log_p_z[21,1]  -19.63    0.04 2.37  -24.79  -21.08  -19.47  -17.94  -15.43
#> log_p_z[21,2]   -3.03    0.00 0.36   -3.78   -3.26   -3.02   -2.79   -2.34
#> log_p_z[21,3]  -21.01    0.09 4.27  -30.82  -23.45  -20.44  -17.92  -14.37
#> log_p_z[21,4]  -21.96    0.09 4.42  -32.01  -24.33  -21.33  -18.89  -15.03
#> log_p_z[22,1]  -16.42    0.03 1.94  -20.61  -17.62  -16.30  -15.08  -13.02
#> log_p_z[22,2]   -4.65    0.01 0.61   -5.96   -5.04   -4.59   -4.22   -3.59
#> log_p_z[22,3]  -21.95    0.08 4.31  -31.79  -24.43  -21.40  -18.89  -15.05
#> log_p_z[22,4]  -21.97    0.09 4.41  -32.01  -24.41  -21.43  -18.91  -14.95
#> log_p_z[23,1]  -21.49    0.04 2.59  -27.15  -23.08  -21.33  -19.69  -16.88
#> log_p_z[23,2]   -3.72    0.01 0.44   -4.63   -4.00   -3.70   -3.41   -2.96
#> log_p_z[23,3]  -18.52    0.08 4.19  -27.81  -20.89  -17.97  -15.52  -12.07
#> log_p_z[23,4]  -21.94    0.08 4.48  -32.13  -24.40  -21.34  -18.81  -14.80
#> log_p_z[24,1]  -17.42    0.03 2.16  -22.22  -18.73  -17.23  -15.87  -13.73
#> log_p_z[24,2]   -5.81    0.01 0.65   -7.29   -6.19   -5.74   -5.35   -4.75
#> log_p_z[24,3]  -21.60    0.07 3.38  -29.27  -23.60  -21.27  -19.24  -16.03
#> log_p_z[24,4]  -21.95    0.07 3.61  -29.78  -24.07  -21.48  -19.47  -15.96
#> log_p_z[25,1]   -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_p_z[25,2]  -20.05    0.04 2.34  -25.21  -21.47  -19.85  -18.42  -16.08
#> log_p_z[25,3]  -24.94    0.08 4.44  -35.06  -27.58  -24.46  -21.73  -17.69
#> log_p_z[25,4]  -21.91    0.07 4.24  -31.68  -24.29  -21.43  -18.92  -15.04
#> log_p_z[26,1]   -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_p_z[26,2]  -20.05    0.04 2.34  -25.21  -21.47  -19.85  -18.42  -16.08
#> log_p_z[26,3]  -24.94    0.08 4.44  -35.06  -27.58  -24.46  -21.73  -17.69
#> log_p_z[26,4]  -21.91    0.07 4.24  -31.68  -24.29  -21.43  -18.92  -15.04
#> log_p_z[27,1]  -21.90    0.04 2.52  -27.51  -23.46  -21.68  -20.15  -17.50
#> log_p_z[27,2]   -5.64    0.01 0.53   -6.80   -5.97   -5.61   -5.27   -4.74
#> log_p_z[27,3]  -18.23    0.06 3.16  -25.37  -20.11  -17.83  -15.97  -13.20
#> log_p_z[27,4]  -21.62    0.07 3.56  -29.47  -23.65  -21.24  -19.07  -15.74
#> log_p_z[28,1]   -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_p_z[28,2]  -20.05    0.04 2.34  -25.21  -21.47  -19.85  -18.42  -16.08
#> log_p_z[28,3]  -24.94    0.08 4.44  -35.06  -27.58  -24.46  -21.73  -17.69
#> log_p_z[28,4]  -21.91    0.07 4.24  -31.68  -24.29  -21.43  -18.92  -15.04
#> log_p_z[29,1]   -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_p_z[29,2]  -20.05    0.04 2.34  -25.21  -21.47  -19.85  -18.42  -16.08
#> log_p_z[29,3]  -24.94    0.08 4.44  -35.06  -27.58  -24.46  -21.73  -17.69
#> log_p_z[29,4]  -21.91    0.07 4.24  -31.68  -24.29  -21.43  -18.92  -15.04
#> log_p_z[30,1]   -6.81    0.01 0.83   -8.61   -7.34   -6.73   -6.22   -5.42
#> log_p_z[30,2]  -14.74    0.03 1.72  -18.59  -15.74  -14.60  -13.53  -11.85
#> log_p_z[30,3]  -24.30    0.06 3.54  -31.88  -26.40  -24.02  -21.83  -18.20
#> log_p_z[30,4]  -21.99    0.07 3.53  -29.93  -24.09  -21.62  -19.46  -16.24
#> log_p_z[31,1]   -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_p_z[31,2]  -20.05    0.04 2.34  -25.21  -21.47  -19.85  -18.42  -16.08
#> log_p_z[31,3]  -24.94    0.08 4.44  -35.06  -27.58  -24.46  -21.73  -17.69
#> log_p_z[31,4]  -21.91    0.07 4.24  -31.68  -24.29  -21.43  -18.92  -15.04
#> log_p_z[32,1]  -29.82    0.07 4.45  -40.05  -32.37  -29.35  -26.70  -22.52
#> log_p_z[32,2]  -14.87    0.03 1.73  -18.76  -15.90  -14.73  -13.67  -11.88
#> log_p_z[32,3]   -6.05    0.01 0.92   -7.97   -6.61   -5.98   -5.42   -4.42
#> log_p_z[32,4]  -20.68    0.08 4.41  -30.83  -23.22  -20.23  -17.49  -13.45
#> log_p_z[33,1]   -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_p_z[33,2]  -20.05    0.04 2.34  -25.21  -21.47  -19.85  -18.42  -16.08
#> log_p_z[33,3]  -24.94    0.08 4.44  -35.06  -27.58  -24.46  -21.73  -17.69
#> log_p_z[33,4]  -21.91    0.07 4.24  -31.68  -24.29  -21.43  -18.92  -15.04
#> log_p_z[34,1]  -19.63    0.04 2.37  -24.79  -21.08  -19.47  -17.94  -15.43
#> log_p_z[34,2]   -3.03    0.00 0.36   -3.78   -3.26   -3.02   -2.79   -2.34
#> log_p_z[34,3]  -21.01    0.09 4.27  -30.82  -23.45  -20.44  -17.92  -14.37
#> log_p_z[34,4]  -21.96    0.09 4.42  -32.01  -24.33  -21.33  -18.89  -15.03
#> log_p_z[35,1]  -22.63    0.04 2.72  -28.74  -24.33  -22.50  -20.68  -17.85
#> log_p_z[35,2]   -5.55    0.01 0.72   -7.16   -5.98   -5.48   -5.06   -4.29
#> log_p_z[35,3]  -15.93    0.08 4.05  -25.11  -18.28  -15.38  -13.01   -9.80
#> log_p_z[35,4]  -21.89    0.08 4.42  -31.88  -24.51  -21.42  -18.69  -14.92
#> log_p_z[36,1]  -29.78    0.06 3.75  -38.17  -32.00  -29.40  -27.04  -23.53
#> log_p_z[36,2]  -21.95    0.04 2.35  -27.10  -23.35  -21.73  -20.25  -17.95
#> log_p_z[36,3]  -10.50    0.03 1.69  -14.69  -11.41  -10.25   -9.32   -7.91
#> log_p_z[36,4]  -14.00    0.06 3.23  -21.16  -16.02  -13.70  -11.65   -8.80
#> log_p_z[37,1]  -18.56    0.04 2.29  -23.73  -19.98  -18.36  -16.92  -14.69
#> log_p_z[37,2]   -7.64    0.01 0.84   -9.44   -8.17   -7.58   -7.04   -6.17
#> log_p_z[37,3]  -19.01    0.06 3.21  -26.20  -20.84  -18.60  -16.74  -13.84
#> log_p_z[37,4]  -21.91    0.07 3.54  -29.97  -24.02  -21.60  -19.45  -16.01
#> log_p_z[38,1]  -25.28    0.04 2.84  -31.37  -27.12  -25.04  -23.28  -20.27
#> log_p_z[38,2]   -9.65    0.02 1.00  -11.86  -10.26   -9.55   -8.95   -7.94
#> log_p_z[38,3]  -11.61    0.05 2.74  -17.77  -13.16  -11.26   -9.59   -7.56
#> log_p_z[38,4]  -21.38    0.06 3.62  -29.37  -23.55  -21.10  -18.81  -15.31
#> log_p_z[39,1]  -29.42    0.07 4.44  -39.77  -31.96  -28.95  -26.25  -22.17
#> log_p_z[39,2]  -16.86    0.03 2.08  -21.57  -18.06  -16.64  -15.45  -13.23
#> log_p_z[39,3]   -7.28    0.02 1.08   -9.63   -7.95   -7.20   -6.50   -5.41
#> log_p_z[39,4]  -18.71    0.08 4.06  -28.55  -20.88  -18.23  -15.93  -12.23
#> log_p_z[40,1]   -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_p_z[40,2]  -20.05    0.04 2.34  -25.21  -21.47  -19.85  -18.42  -16.08
#> log_p_z[40,3]  -24.94    0.08 4.44  -35.06  -27.58  -24.46  -21.73  -17.69
#> log_p_z[40,4]  -21.91    0.07 4.24  -31.68  -24.29  -21.43  -18.92  -15.04
#> log_p_z[41,1]   -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_p_z[41,2]  -20.05    0.04 2.34  -25.21  -21.47  -19.85  -18.42  -16.08
#> log_p_z[41,3]  -24.94    0.08 4.44  -35.06  -27.58  -24.46  -21.73  -17.69
#> log_p_z[41,4]  -21.91    0.07 4.24  -31.68  -24.29  -21.43  -18.92  -15.04
#> log_p_z[42,1]   -6.01    0.01 0.62   -7.32   -6.38   -5.95   -5.57   -4.95
#> log_p_z[42,2]  -14.97    0.03 1.72  -18.84  -16.03  -14.82  -13.75  -11.98
#> log_p_z[42,3]  -24.33    0.07 3.62  -32.34  -26.52  -23.99  -21.79  -18.21
#> log_p_z[42,4]  -21.90    0.06 3.53  -29.93  -24.00  -21.56  -19.40  -16.10
#> log_p_z[43,1]  -21.90    0.04 2.52  -27.51  -23.46  -21.68  -20.15  -17.50
#> log_p_z[43,2]   -5.64    0.01 0.53   -6.80   -5.97   -5.61   -5.27   -4.74
#> log_p_z[43,3]  -18.23    0.06 3.16  -25.37  -20.11  -17.83  -15.97  -13.20
#> log_p_z[43,4]  -21.62    0.07 3.56  -29.47  -23.65  -21.24  -19.07  -15.74
#> log_p_z[44,1]   -4.41    0.01 0.40   -5.26   -4.65   -4.38   -4.13   -3.72
#> log_p_z[44,2]  -17.27    0.03 1.86  -21.34  -18.41  -17.16  -15.97  -14.00
#> log_p_z[44,3]  -24.35    0.06 3.62  -32.36  -26.55  -24.04  -21.72  -18.13
#> log_p_z[44,4]  -21.93    0.06 3.49  -29.85  -24.02  -21.56  -19.46  -16.12
#> log_p_z[45,1]  -19.63    0.04 2.37  -24.79  -21.08  -19.47  -17.94  -15.43
#> log_p_z[45,2]   -3.03    0.00 0.36   -3.78   -3.26   -3.02   -2.79   -2.34
#> log_p_z[45,3]  -21.01    0.09 4.27  -30.82  -23.45  -20.44  -17.92  -14.37
#> log_p_z[45,4]  -21.96    0.09 4.42  -32.01  -24.33  -21.33  -18.89  -15.03
#> log_lik[1]      -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_lik[2]      -7.62    0.02 1.16  -10.11   -8.35   -7.54   -6.82   -5.57
#> log_lik[3]      -9.94    0.01 0.92  -11.87  -10.52   -9.93   -9.30   -8.25
#> log_lik[4]      -7.57    0.01 0.92   -9.52   -8.13   -7.50   -6.92   -5.92
#> log_lik[5]      -3.72    0.01 0.44   -4.63   -4.00   -3.70   -3.41   -2.96
#> log_lik[6]      -5.49    0.01 0.68   -6.92   -5.94   -5.44   -5.01   -4.27
#> log_lik[7]      -6.00    0.01 0.62   -7.32   -6.38   -5.95   -5.57   -4.95
#> log_lik[8]      -5.74    0.01 0.83   -7.53   -6.24   -5.69   -5.17   -4.27
#> log_lik[9]      -4.52    0.01 0.60   -5.81   -4.90   -4.46   -4.10   -3.47
#> log_lik[10]     -7.13    0.01 0.73   -8.67   -7.57   -7.07   -6.61   -5.86
#> log_lik[11]     -5.83    0.01 0.98   -7.99   -6.43   -5.76   -5.14   -4.14
#> log_lik[12]     -9.35    0.01 1.03  -11.46  -10.01   -9.31   -8.62   -7.48
#> log_lik[13]     -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_lik[14]     -6.25    0.01 0.90   -8.22   -6.79   -6.17   -5.60   -4.74
#> log_lik[15]     -4.41    0.01 0.40   -5.26   -4.65   -4.38   -4.13   -3.72
#> log_lik[16]     -3.80    0.01 0.56   -5.03   -4.13   -3.75   -3.41   -2.87
#> log_lik[17]     -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_lik[18]     -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_lik[19]     -4.65    0.01 0.61   -5.96   -5.04   -4.59   -4.22   -3.59
#> log_lik[20]     -7.10    0.01 0.96   -9.26   -7.69   -7.02   -6.42   -5.43
#> log_lik[21]     -3.03    0.00 0.36   -3.78   -3.26   -3.02   -2.79   -2.34
#> log_lik[22]     -4.65    0.01 0.61   -5.96   -5.04   -4.59   -4.22   -3.59
#> log_lik[23]     -3.72    0.01 0.44   -4.63   -4.00   -3.70   -3.41   -2.96
#> log_lik[24]     -5.81    0.01 0.65   -7.29   -6.19   -5.74   -5.35   -4.75
#> log_lik[25]     -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_lik[26]     -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_lik[27]     -5.64    0.01 0.53   -6.80   -5.97   -5.61   -5.27   -4.74
#> log_lik[28]     -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_lik[29]     -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_lik[30]     -6.81    0.01 0.83   -8.60   -7.33   -6.73   -6.22   -5.42
#> log_lik[31]     -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_lik[32]     -6.05    0.01 0.92   -7.97   -6.61   -5.98   -5.42   -4.42
#> log_lik[33]     -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_lik[34]     -3.03    0.00 0.36   -3.78   -3.26   -3.02   -2.79   -2.34
#> log_lik[35]     -5.55    0.01 0.71   -7.13   -5.97   -5.48   -5.06   -4.29
#> log_lik[36]     -9.92    0.02 1.17  -12.39  -10.69   -9.85   -9.08   -7.86
#> log_lik[37]     -7.64    0.01 0.84   -9.43   -8.17   -7.58   -7.04   -6.17
#> log_lik[38]     -9.01    0.01 0.84  -10.72   -9.54   -9.00   -8.46   -7.33
#> log_lik[39]     -7.28    0.02 1.08   -9.61   -7.95   -7.20   -6.50   -5.41
#> log_lik[40]     -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_lik[41]     -2.20    0.00 0.30   -2.83   -2.39   -2.19   -2.00   -1.66
#> log_lik[42]     -6.00    0.01 0.62   -7.32   -6.38   -5.95   -5.57   -4.95
#> log_lik[43]     -5.64    0.01 0.53   -6.80   -5.97   -5.61   -5.27   -4.74
#> log_lik[44]     -4.41    0.01 0.40   -5.26   -4.65   -4.38   -4.13   -3.72
#> log_lik[45]     -3.03    0.00 0.36   -3.78   -3.26   -3.02   -2.79   -2.34
#> lp__          -464.12    0.18 6.41 -478.14 -468.28 -463.72 -459.58 -452.89
#>               n_eff Rhat
#> pi[1]          6558    1
#> pi[2]          6183    1
#> pi[3]          5825    1
#> pi[4]          5610    1
#> theta[1,1,1]   7066    1
#> theta[1,1,2]   6680    1
#> theta[1,1,3]   4192    1
#> theta[1,1,4]   4130    1
#> theta[1,2,1]   5568    1
#> theta[1,2,2]   5908    1
#> theta[1,2,3]   5532    1
#> theta[1,2,4]   5214    1
#> theta[1,3,1]   6487    1
#> theta[1,3,2]   3929    1
#> theta[1,3,3]   4422    1
#> theta[1,3,4]   5174    1
#> theta[1,4,1]   6996    1
#> theta[1,4,2]   5619    1
#> theta[1,4,3]   3664    1
#> theta[1,4,4]   4942    1
#> theta[2,1,1]   6122    1
#> theta[2,1,2]   5488    1
#> theta[2,1,3]   4222    1
#> theta[2,1,4]   4348    1
#> theta[2,2,1]   5567    1
#> theta[2,2,2]   6107    1
#> theta[2,2,3]   6366    1
#> theta[2,2,4]   5934    1
#> theta[2,3,1]   6051    1
#> theta[2,3,2]   5649    1
#> theta[2,3,3]   6051    1
#> theta[2,3,4]   5198    1
#> theta[2,4,1]   5623    1
#> theta[2,4,2]   7003    1
#> theta[2,4,3]   5224    1
#> theta[2,4,4]   5744    1
#> theta[3,1,1]   7005    1
#> theta[3,1,2]   5328    1
#> theta[3,1,3]   4516    1
#> theta[3,1,4]   4649    1
#> theta[3,2,1]   5830    1
#> theta[3,2,2]   6501    1
#> theta[3,2,3]   5941    1
#> theta[3,2,4]   4101    1
#> theta[3,3,1]   5933    1
#> theta[3,3,2]   6043    1
#> theta[3,3,3]   5641    1
#> theta[3,3,4]   6453    1
#> theta[3,4,1]   5799    1
#> theta[3,4,2]   6606    1
#> theta[3,4,3]   5217    1
#> theta[3,4,4]   6145    1
#> theta[4,1,1]   6686    1
#> theta[4,1,2]   5621    1
#> theta[4,1,3]   4984    1
#> theta[4,1,4]   3836    1
#> theta[4,2,1]   6294    1
#> theta[4,2,2]   6502    1
#> theta[4,2,3]   6873    1
#> theta[4,2,4]   5147    1
#> theta[4,3,1]   7020    1
#> theta[4,3,2]   4870    1
#> theta[4,3,3]   5937    1
#> theta[4,3,4]   5041    1
#> theta[4,4,1]   6291    1
#> theta[4,4,2]   5461    1
#> theta[4,4,3]   5554    1
#> theta[4,4,4]   5897    1
#> theta[5,1,1]   5326    1
#> theta[5,1,2]   4412    1
#> theta[5,1,3]   3881    1
#> theta[5,1,4]   4297    1
#> theta[5,2,1]   6802    1
#> theta[5,2,2]   6307    1
#> theta[5,2,3]   5871    1
#> theta[5,2,4]   5216    1
#> theta[5,3,1]   5739    1
#> theta[5,3,2]   7108    1
#> theta[5,3,3]   5546    1
#> theta[5,3,4]   6580    1
#> theta[5,4,1]   6157    1
#> theta[5,4,2]   5847    1
#> theta[5,4,3]   5769    1
#> theta[5,4,4]   6730    1
#> log_p_z[1,1]   7012    1
#> log_p_z[1,2]   3055    1
#> log_p_z[1,3]   2962    1
#> log_p_z[1,4]   3212    1
#> log_p_z[2,1]   4488    1
#> log_p_z[2,2]   3966    1
#> log_p_z[2,3]   3882    1
#> log_p_z[2,4]   2687    1
#> log_p_z[3,1]   3815    1
#> log_p_z[3,2]   3345    1
#> log_p_z[3,3]   2430    1
#> log_p_z[3,4]   3111    1
#> log_p_z[4,1]   4827    1
#> log_p_z[4,2]   4240    1
#> log_p_z[4,3]   2722    1
#> log_p_z[4,4]   2933    1
#> log_p_z[5,1]   4404    1
#> log_p_z[5,2]   6610    1
#> log_p_z[5,3]   2585    1
#> log_p_z[5,4]   2901    1
#> log_p_z[6,1]   5086    1
#> log_p_z[6,2]   5661    1
#> log_p_z[6,3]   2588    1
#> log_p_z[6,4]   2945    1
#> log_p_z[7,1]   4728    1
#> log_p_z[7,2]   3192    1
#> log_p_z[7,3]   2717    1
#> log_p_z[7,4]   3310    1
#> log_p_z[8,1]   4393    1
#> log_p_z[8,2]   3865    1
#> log_p_z[8,3]   4917    1
#> log_p_z[8,4]   2781    1
#> log_p_z[9,1]   4356    1
#> log_p_z[9,2]   5713    1
#> log_p_z[9,3]   2440    1
#> log_p_z[9,4]   2564    1
#> log_p_z[10,1]  4437    1
#> log_p_z[10,2]  4801    1
#> log_p_z[10,3]  2403    1
#> log_p_z[10,4]  2457    1
#> log_p_z[11,1]  4052    1
#> log_p_z[11,2]  3424    1
#> log_p_z[11,3]  2828    1
#> log_p_z[11,4]  5819    1
#> log_p_z[12,1]  4342    1
#> log_p_z[12,2]  3609    1
#> log_p_z[12,3]  2525    1
#> log_p_z[12,4]  3024    1
#> log_p_z[13,1]  7012    1
#> log_p_z[13,2]  3055    1
#> log_p_z[13,3]  2962    1
#> log_p_z[13,4]  3212    1
#> log_p_z[14,1]  4483    1
#> log_p_z[14,2]  4734    1
#> log_p_z[14,3]  2599    1
#> log_p_z[14,4]  3108    1
#> log_p_z[15,1]  5764    1
#> log_p_z[15,2]  3082    1
#> log_p_z[15,3]  3156    1
#> log_p_z[15,4]  3115    1
#> log_p_z[16,1]  5144    1
#> log_p_z[16,2]  3146    1
#> log_p_z[16,3]  2715    1
#> log_p_z[16,4]  3424    1
#> log_p_z[17,1]  7012    1
#> log_p_z[17,2]  3055    1
#> log_p_z[17,3]  2962    1
#> log_p_z[17,4]  3212    1
#> log_p_z[18,1]  7012    1
#> log_p_z[18,2]  3055    1
#> log_p_z[18,3]  2962    1
#> log_p_z[18,4]  3212    1
#> log_p_z[19,1]  4649    1
#> log_p_z[19,2]  4560    1
#> log_p_z[19,3]  2625    1
#> log_p_z[19,4]  2441    1
#> log_p_z[20,1]  5165    1
#> log_p_z[20,2]  4384    1
#> log_p_z[20,3]  2669    1
#> log_p_z[20,4]  2674    1
#> log_p_z[21,1]  4422    1
#> log_p_z[21,2]  6354    1
#> log_p_z[21,3]  2481    1
#> log_p_z[21,4]  2684    1
#> log_p_z[22,1]  4649    1
#> log_p_z[22,2]  4560    1
#> log_p_z[22,3]  2625    1
#> log_p_z[22,4]  2441    1
#> log_p_z[23,1]  4404    1
#> log_p_z[23,2]  6610    1
#> log_p_z[23,3]  2585    1
#> log_p_z[23,4]  2901    1
#> log_p_z[24,1]  4180    1
#> log_p_z[24,2]  3366    1
#> log_p_z[24,3]  2463    1
#> log_p_z[24,4]  2622    1
#> log_p_z[25,1]  7012    1
#> log_p_z[25,2]  3055    1
#> log_p_z[25,3]  2962    1
#> log_p_z[25,4]  3212    1
#> log_p_z[26,1]  7012    1
#> log_p_z[26,2]  3055    1
#> log_p_z[26,3]  2962    1
#> log_p_z[26,4]  3212    1
#> log_p_z[27,1]  4206    1
#> log_p_z[27,2]  4871    1
#> log_p_z[27,3]  2488    1
#> log_p_z[27,4]  2682    1
#> log_p_z[28,1]  7012    1
#> log_p_z[28,2]  3055    1
#> log_p_z[28,3]  2962    1
#> log_p_z[28,4]  3212    1
#> log_p_z[29,1]  7012    1
#> log_p_z[29,2]  3055    1
#> log_p_z[29,3]  2962    1
#> log_p_z[29,4]  3212    1
#> log_p_z[30,1]  5608    1
#> log_p_z[30,2]  2985    1
#> log_p_z[30,3]  3072    1
#> log_p_z[30,4]  2938    1
#> log_p_z[31,1]  7012    1
#> log_p_z[31,2]  3055    1
#> log_p_z[31,3]  2962    1
#> log_p_z[31,4]  3212    1
#> log_p_z[32,1]  4285    1
#> log_p_z[32,2]  4413    1
#> log_p_z[32,3]  3933    1
#> log_p_z[32,4]  2933    1
#> log_p_z[33,1]  7012    1
#> log_p_z[33,2]  3055    1
#> log_p_z[33,3]  2962    1
#> log_p_z[33,4]  3212    1
#> log_p_z[34,1]  4422    1
#> log_p_z[34,2]  6354    1
#> log_p_z[34,3]  2481    1
#> log_p_z[34,4]  2684    1
#> log_p_z[35,1]  3973    1
#> log_p_z[35,2]  5437    1
#> log_p_z[35,3]  2592    1
#> log_p_z[35,4]  3161    1
#> log_p_z[36,1]  3783    1
#> log_p_z[36,2]  3345    1
#> log_p_z[36,3]  2644    1
#> log_p_z[36,4]  2531    1
#> log_p_z[37,1]  3635    1
#> log_p_z[37,2]  4786    1
#> log_p_z[37,3]  2577    1
#> log_p_z[37,4]  2836    1
#> log_p_z[38,1]  4121    1
#> log_p_z[38,2]  4308    1
#> log_p_z[38,3]  2522    1
#> log_p_z[38,4]  3307    1
#> log_p_z[39,1]  4204    1
#> log_p_z[39,2]  3960    1
#> log_p_z[39,3]  4564    1
#> log_p_z[39,4]  2895    1
#> log_p_z[40,1]  7012    1
#> log_p_z[40,2]  3055    1
#> log_p_z[40,3]  2962    1
#> log_p_z[40,4]  3212    1
#> log_p_z[41,1]  7012    1
#> log_p_z[41,2]  3055    1
#> log_p_z[41,3]  2962    1
#> log_p_z[41,4]  3212    1
#> log_p_z[42,1]  4728    1
#> log_p_z[42,2]  3192    1
#> log_p_z[42,3]  2717    1
#> log_p_z[42,4]  3310    1
#> log_p_z[43,1]  4206    1
#> log_p_z[43,2]  4871    1
#> log_p_z[43,3]  2488    1
#> log_p_z[43,4]  2682    1
#> log_p_z[44,1]  5764    1
#> log_p_z[44,2]  3082    1
#> log_p_z[44,3]  3156    1
#> log_p_z[44,4]  3115    1
#> log_p_z[45,1]  4422    1
#> log_p_z[45,2]  6354    1
#> log_p_z[45,3]  2481    1
#> log_p_z[45,4]  2684    1
#> log_lik[1]     7012    1
#> log_lik[2]     4552    1
#> log_lik[3]     4366    1
#> log_lik[4]     4282    1
#> log_lik[5]     6611    1
#> log_lik[6]     5668    1
#> log_lik[7]     4731    1
#> log_lik[8]     4925    1
#> log_lik[9]     5713    1
#> log_lik[10]    4828    1
#> log_lik[11]    5823    1
#> log_lik[12]    5398    1
#> log_lik[13]    7012    1
#> log_lik[14]    4738    1
#> log_lik[15]    5764    1
#> log_lik[16]    5143    1
#> log_lik[17]    7012    1
#> log_lik[18]    7012    1
#> log_lik[19]    4560    1
#> log_lik[20]    4381    1
#> log_lik[21]    6354    1
#> log_lik[22]    4560    1
#> log_lik[23]    6611    1
#> log_lik[24]    3367    1
#> log_lik[25]    7012    1
#> log_lik[26]    7012    1
#> log_lik[27]    4871    1
#> log_lik[28]    7012    1
#> log_lik[29]    7012    1
#> log_lik[30]    5608    1
#> log_lik[31]    7012    1
#> log_lik[32]    3942    1
#> log_lik[33]    7012    1
#> log_lik[34]    6354    1
#> log_lik[35]    5478    1
#> log_lik[36]    4582    1
#> log_lik[37]    4787    1
#> log_lik[38]    4286    1
#> log_lik[39]    4563    1
#> log_lik[40]    7012    1
#> log_lik[41]    7012    1
#> log_lik[42]    4731    1
#> log_lik[43]    4871    1
#> log_lik[44]    5764    1
#> log_lik[45]    6354    1
#> lp__           1271    1
#> 
#> Samples were drawn using NUTS(diag_e) at Tue Feb  7 09:59:49 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }