Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.07    0.25    0.33    0.37    0.42    0.51
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.09    0.15
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.93
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.05    0.75    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.06    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.58    0.74    0.81    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.14    0.31
#> theta[1,4,2]     0.10    0.00 0.09    0.00    0.03    0.07    0.13    0.33
#> theta[1,4,3]     0.12    0.00 0.11    0.00    0.04    0.09    0.18    0.41
#> theta[1,4,4]     0.69    0.00 0.14    0.38    0.59    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.69    0.75    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.22    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.58    0.65    0.75
#> theta[2,2,3]     0.31    0.00 0.09    0.16    0.25    0.30    0.37    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.26
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,3]     0.62    0.00 0.13    0.34    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.22    0.00 0.11    0.04    0.13    0.20    0.29    0.47
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.74    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.10    0.00 0.06    0.01    0.05    0.09    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.78    0.87
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[3,3,2]     0.18    0.00 0.10    0.03    0.10    0.16    0.24    0.42
#> theta[3,3,3]     0.51    0.00 0.14    0.24    0.41    0.51    0.60    0.77
#> theta[3,3,4]     0.23    0.00 0.12    0.06    0.14    0.22    0.31    0.50
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[3,4,3]     0.14    0.00 0.11    0.00    0.05    0.11    0.20    0.42
#> theta[3,4,4]     0.63    0.00 0.16    0.30    0.52    0.64    0.74    0.90
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.78    0.84    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.09    0.54    0.67    0.73    0.79    0.87
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.09    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.26
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[4,3,3]     0.69    0.00 0.12    0.43    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.38
#> theta[4,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.17    0.36
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.32    0.55    0.66    0.76    0.91
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.82    0.87    0.91    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.15    0.00 0.06    0.05    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.38
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.61    0.70    0.84
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[5,4,3]     0.13    0.00 0.11    0.01    0.05    0.11    0.19    0.41
#> theta[5,4,4]     0.63    0.00 0.15    0.32    0.53    0.64    0.75    0.90
#> log_p_z[1,1]    -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_p_z[1,2]   -19.96    0.04 2.32  -25.13  -21.38  -19.78  -18.33  -16.03
#> log_p_z[1,3]   -24.83    0.08 4.31  -34.73  -27.41  -24.34  -21.78  -17.57
#> log_p_z[1,4]   -21.89    0.08 4.34  -31.76  -24.47  -21.33  -18.85  -14.74
#> log_p_z[2,1]   -29.76    0.07 4.35  -39.76  -32.23  -29.31  -26.69  -22.46
#> log_p_z[2,2]   -21.10    0.04 2.37  -26.04  -22.58  -20.87  -19.45  -17.02
#> log_p_z[2,3]    -7.62    0.02 1.19  -10.26   -8.33   -7.51   -6.78   -5.64
#> log_p_z[2,4]   -16.54    0.08 4.23  -26.16  -18.95  -16.02  -13.59   -9.55
#> log_p_z[3,1]   -11.64    0.03 1.66  -15.42  -12.61  -11.47  -10.46   -8.90
#> log_p_z[3,2]   -10.78    0.02 1.49  -14.26  -11.64  -10.59   -9.73   -8.41
#> log_p_z[3,3]   -23.39    0.06 3.44  -30.66  -25.59  -23.17  -20.90  -17.58
#> log_p_z[3,4]   -21.84    0.06 3.61  -29.72  -24.10  -21.51  -19.31  -15.69
#> log_p_z[4,1]   -14.65    0.03 1.83  -18.71  -15.75  -14.52  -13.37  -11.42
#> log_p_z[4,2]    -7.54    0.01 0.93   -9.58   -8.12   -7.48   -6.88   -5.95
#> log_p_z[4,3]   -20.72    0.08 4.43  -30.77  -23.17  -20.09  -17.59  -13.89
#> log_p_z[4,4]   -21.85    0.08 4.42  -32.08  -24.58  -21.34  -18.66  -14.56
#> log_p_z[5,1]   -21.47    0.04 2.58  -26.99  -23.08  -21.28  -19.60  -17.10
#> log_p_z[5,2]    -3.71    0.01 0.42   -4.56   -3.99   -3.69   -3.41   -2.93
#> log_p_z[5,3]   -18.70    0.08 4.31  -28.78  -21.09  -18.07  -15.66  -12.07
#> log_p_z[5,4]   -21.87    0.08 4.48  -32.13  -24.54  -21.32  -18.57  -14.52
#> log_p_z[6,1]   -21.50    0.04 2.53  -27.06  -23.03  -21.25  -19.71  -17.18
#> log_p_z[6,2]    -5.45    0.01 0.67   -6.88   -5.89   -5.41   -4.97   -4.26
#> log_p_z[6,3]   -17.49    0.08 4.17  -27.26  -19.69  -16.83  -14.56  -11.08
#> log_p_z[6,4]   -21.66    0.07 4.37  -31.57  -24.40  -21.16  -18.50  -14.55
#> log_p_z[7,1]    -5.99    0.01 0.62   -7.34   -6.36   -5.95   -5.55   -4.91
#> log_p_z[7,2]   -14.90    0.03 1.69  -18.74  -15.90  -14.74  -13.72  -11.96
#> log_p_z[7,3]   -24.31    0.06 3.53  -32.29  -26.45  -24.04  -21.85  -18.20
#> log_p_z[7,4]   -21.91    0.06 3.59  -29.81  -24.17  -21.62  -19.30  -15.85
#> log_p_z[8,1]   -29.76    0.07 4.43  -39.84  -32.39  -29.30  -26.57  -22.63
#> log_p_z[8,2]   -18.34    0.04 2.12  -23.01  -19.67  -18.16  -16.83  -14.64
#> log_p_z[8,3]    -5.72    0.01 0.83   -7.46   -6.27   -5.67   -5.15   -4.25
#> log_p_z[8,4]   -18.75    0.09 4.30  -28.35  -21.26  -18.35  -15.72  -11.70
#> log_p_z[9,1]   -19.99    0.04 2.32  -24.77  -21.47  -19.82  -18.38  -15.85
#> log_p_z[9,2]    -4.53    0.01 0.59   -5.82   -4.89   -4.48   -4.12   -3.54
#> log_p_z[9,3]   -19.65    0.08 4.36  -29.68  -22.12  -19.05  -16.59  -12.89
#> log_p_z[9,4]   -21.72    0.08 4.35  -31.50  -24.30  -21.21  -18.56  -14.65
#> log_p_z[10,1]  -22.24    0.04 2.47  -27.76  -23.72  -22.12  -20.44  -18.01
#> log_p_z[10,2]   -7.15    0.01 0.73   -8.73   -7.61   -7.10   -6.64   -5.87
#> log_p_z[10,3]  -16.84    0.06 3.18  -24.15  -18.67  -16.43  -14.58  -11.81
#> log_p_z[10,4]  -21.44    0.07 3.58  -29.32  -23.70  -21.04  -18.86  -15.48
#> log_p_z[11,1]  -29.85    0.07 4.40  -40.08  -32.39  -29.35  -26.81  -22.61
#> log_p_z[11,2]  -29.48    0.07 4.40  -39.78  -31.96  -28.93  -26.32  -22.40
#> log_p_z[11,3]  -18.17    0.07 3.38  -26.45  -19.94  -17.66  -15.95  -13.05
#> log_p_z[11,4]   -5.84    0.01 1.00   -8.03   -6.46   -5.76   -5.13   -4.12
#> log_p_z[12,1]  -22.98    0.04 2.64  -28.79  -24.63  -22.73  -21.17  -18.30
#> log_p_z[12,2]   -9.61    0.02 1.31  -12.57  -10.34   -9.48   -8.71   -7.51
#> log_p_z[12,3]  -15.11    0.08 4.13  -24.78  -17.39  -14.54  -12.16   -8.80
#> log_p_z[12,4]  -19.33    0.07 4.16  -28.75  -21.85  -18.84  -16.36  -12.72
#> log_p_z[13,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_p_z[13,2]  -19.96    0.04 2.32  -25.13  -21.38  -19.78  -18.33  -16.03
#> log_p_z[13,3]  -24.83    0.08 4.31  -34.73  -27.41  -24.34  -21.78  -17.57
#> log_p_z[13,4]  -21.89    0.08 4.34  -31.76  -24.47  -21.33  -18.85  -14.74
#> log_p_z[14,1]  -19.05    0.04 2.43  -24.31  -20.54  -18.88  -17.32  -14.91
#> log_p_z[14,2]   -6.23    0.01 0.87   -8.13   -6.75   -6.16   -5.62   -4.74
#> log_p_z[14,3]  -18.70    0.09 4.28  -28.96  -20.99  -18.12  -15.66  -12.27
#> log_p_z[14,4]  -21.93    0.08 4.41  -31.80  -24.53  -21.37  -18.88  -14.80
#> log_p_z[15,1]   -4.41    0.01 0.40   -5.27   -4.67   -4.38   -4.13   -3.72
#> log_p_z[15,2]  -17.21    0.03 1.87  -21.32  -18.40  -17.04  -15.90  -13.94
#> log_p_z[15,3]  -24.30    0.07 3.58  -32.28  -26.51  -24.00  -21.77  -18.10
#> log_p_z[15,4]  -21.89    0.07 3.57  -29.90  -24.06  -21.51  -19.49  -15.87
#> log_p_z[16,1]   -3.78    0.01 0.54   -4.97   -4.11   -3.74   -3.40   -2.84
#> log_p_z[16,2]  -17.66    0.04 2.19  -22.62  -18.91  -17.46  -16.13  -13.92
#> log_p_z[16,3]  -24.84    0.07 4.28  -34.71  -27.38  -24.41  -21.84  -17.75
#> log_p_z[16,4]  -21.90    0.08 4.38  -31.84  -24.48  -21.40  -18.81  -14.90
#> log_p_z[17,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_p_z[17,2]  -19.96    0.04 2.32  -25.13  -21.38  -19.78  -18.33  -16.03
#> log_p_z[17,3]  -24.83    0.08 4.31  -34.73  -27.41  -24.34  -21.78  -17.57
#> log_p_z[17,4]  -21.89    0.08 4.34  -31.76  -24.47  -21.33  -18.85  -14.74
#> log_p_z[18,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_p_z[18,2]  -19.96    0.04 2.32  -25.13  -21.38  -19.78  -18.33  -16.03
#> log_p_z[18,3]  -24.83    0.08 4.31  -34.73  -27.41  -24.34  -21.78  -17.57
#> log_p_z[18,4]  -21.89    0.08 4.34  -31.76  -24.47  -21.33  -18.85  -14.74
#> log_p_z[19,1]  -16.39    0.03 1.95  -20.66  -17.60  -16.24  -15.03  -12.97
#> log_p_z[19,2]   -4.65    0.01 0.59   -5.88   -5.01   -4.61   -4.24   -3.61
#> log_p_z[19,3]  -22.14    0.08 4.50  -32.36  -24.68  -21.55  -18.95  -15.11
#> log_p_z[19,4]  -21.92    0.08 4.41  -31.79  -24.55  -21.41  -18.79  -14.68
#> log_p_z[20,1]  -18.09    0.04 2.26  -23.05  -19.49  -17.87  -16.52  -14.14
#> log_p_z[20,2]   -7.09    0.01 0.98   -9.31   -7.66   -7.00   -6.41   -5.42
#> log_p_z[20,3]  -20.00    0.08 4.28  -30.23  -22.38  -19.32  -16.95  -13.37
#> log_p_z[20,4]  -21.66    0.08 4.37  -31.55  -24.29  -21.12  -18.46  -14.62
#> log_p_z[21,1]  -19.63    0.04 2.37  -24.75  -21.09  -19.44  -17.93  -15.61
#> log_p_z[21,2]   -3.05    0.00 0.35   -3.75   -3.26   -3.03   -2.81   -2.41
#> log_p_z[21,3]  -21.21    0.08 4.38  -31.28  -23.65  -20.62  -18.17  -14.31
#> log_p_z[21,4]  -21.90    0.08 4.43  -31.97  -24.51  -21.33  -18.70  -14.62
#> log_p_z[22,1]  -16.39    0.03 1.95  -20.66  -17.60  -16.24  -15.03  -12.97
#> log_p_z[22,2]   -4.65    0.01 0.59   -5.88   -5.01   -4.61   -4.24   -3.61
#> log_p_z[22,3]  -22.14    0.08 4.50  -32.36  -24.68  -21.55  -18.95  -15.11
#> log_p_z[22,4]  -21.92    0.08 4.41  -31.79  -24.55  -21.41  -18.79  -14.68
#> log_p_z[23,1]  -21.47    0.04 2.58  -26.99  -23.08  -21.28  -19.60  -17.10
#> log_p_z[23,2]   -3.71    0.01 0.42   -4.56   -3.99   -3.69   -3.41   -2.93
#> log_p_z[23,3]  -18.70    0.08 4.31  -28.78  -21.09  -18.07  -15.66  -12.07
#> log_p_z[23,4]  -21.87    0.08 4.48  -32.13  -24.54  -21.32  -18.57  -14.52
#> log_p_z[24,1]  -17.42    0.04 2.17  -22.23  -18.76  -17.21  -15.88  -13.77
#> log_p_z[24,2]   -5.80    0.01 0.64   -7.25   -6.16   -5.73   -5.35   -4.74
#> log_p_z[24,3]  -21.75    0.06 3.43  -29.40  -23.87  -21.43  -19.25  -16.04
#> log_p_z[24,4]  -21.89    0.06 3.65  -29.92  -24.17  -21.54  -19.21  -15.68
#> log_p_z[25,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_p_z[25,2]  -19.96    0.04 2.32  -25.13  -21.38  -19.78  -18.33  -16.03
#> log_p_z[25,3]  -24.83    0.08 4.31  -34.73  -27.41  -24.34  -21.78  -17.57
#> log_p_z[25,4]  -21.89    0.08 4.34  -31.76  -24.47  -21.33  -18.85  -14.74
#> log_p_z[26,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_p_z[26,2]  -19.96    0.04 2.32  -25.13  -21.38  -19.78  -18.33  -16.03
#> log_p_z[26,3]  -24.83    0.08 4.31  -34.73  -27.41  -24.34  -21.78  -17.57
#> log_p_z[26,4]  -21.89    0.08 4.34  -31.76  -24.47  -21.33  -18.85  -14.74
#> log_p_z[27,1]  -21.88    0.04 2.53  -27.61  -23.46  -21.65  -20.06  -17.63
#> log_p_z[27,2]   -5.67    0.01 0.54   -6.81   -6.01   -5.62   -5.29   -4.72
#> log_p_z[27,3]  -18.40    0.06 3.22  -25.59  -20.24  -18.06  -16.09  -13.17
#> log_p_z[27,4]  -21.63    0.06 3.59  -29.71  -23.87  -21.26  -19.04  -15.69
#> log_p_z[28,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_p_z[28,2]  -19.96    0.04 2.32  -25.13  -21.38  -19.78  -18.33  -16.03
#> log_p_z[28,3]  -24.83    0.08 4.31  -34.73  -27.41  -24.34  -21.78  -17.57
#> log_p_z[28,4]  -21.89    0.08 4.34  -31.76  -24.47  -21.33  -18.85  -14.74
#> log_p_z[29,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_p_z[29,2]  -19.96    0.04 2.32  -25.13  -21.38  -19.78  -18.33  -16.03
#> log_p_z[29,3]  -24.83    0.08 4.31  -34.73  -27.41  -24.34  -21.78  -17.57
#> log_p_z[29,4]  -21.89    0.08 4.34  -31.76  -24.47  -21.33  -18.85  -14.74
#> log_p_z[30,1]   -6.83    0.01 0.84   -8.69   -7.36   -6.78   -6.21   -5.42
#> log_p_z[30,2]  -14.69    0.03 1.73  -18.52  -15.81  -14.50  -13.45  -11.73
#> log_p_z[30,3]  -24.30    0.07 3.56  -32.00  -26.45  -24.00  -21.80  -18.12
#> log_p_z[30,4]  -21.84    0.07 3.59  -29.80  -23.96  -21.49  -19.28  -15.76
#> log_p_z[31,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_p_z[31,2]  -19.96    0.04 2.32  -25.13  -21.38  -19.78  -18.33  -16.03
#> log_p_z[31,3]  -24.83    0.08 4.31  -34.73  -27.41  -24.34  -21.78  -17.57
#> log_p_z[31,4]  -21.89    0.08 4.34  -31.76  -24.47  -21.33  -18.85  -14.74
#> log_p_z[32,1]  -29.71    0.07 4.43  -39.62  -32.43  -29.16  -26.59  -22.50
#> log_p_z[32,2]  -14.88    0.03 1.77  -18.75  -15.95  -14.70  -13.68  -11.83
#> log_p_z[32,3]   -6.06    0.01 0.93   -8.15   -6.64   -5.98   -5.40   -4.50
#> log_p_z[32,4]  -20.90    0.09 4.48  -30.94  -23.46  -20.44  -17.81  -13.42
#> log_p_z[33,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_p_z[33,2]  -19.96    0.04 2.32  -25.13  -21.38  -19.78  -18.33  -16.03
#> log_p_z[33,3]  -24.83    0.08 4.31  -34.73  -27.41  -24.34  -21.78  -17.57
#> log_p_z[33,4]  -21.89    0.08 4.34  -31.76  -24.47  -21.33  -18.85  -14.74
#> log_p_z[34,1]  -19.63    0.04 2.37  -24.75  -21.09  -19.44  -17.93  -15.61
#> log_p_z[34,2]   -3.05    0.00 0.35   -3.75   -3.26   -3.03   -2.81   -2.41
#> log_p_z[34,3]  -21.21    0.08 4.38  -31.28  -23.65  -20.62  -18.17  -14.31
#> log_p_z[34,4]  -21.90    0.08 4.43  -31.97  -24.51  -21.33  -18.70  -14.62
#> log_p_z[35,1]  -22.59    0.04 2.71  -28.54  -24.26  -22.43  -20.70  -17.87
#> log_p_z[35,2]   -5.53    0.01 0.70   -7.03   -5.95   -5.49   -5.05   -4.32
#> log_p_z[35,3]  -16.08    0.08 4.16  -26.38  -18.34  -15.49  -13.07   -9.94
#> log_p_z[35,4]  -21.90    0.07 4.44  -31.86  -24.56  -21.36  -18.70  -14.79
#> log_p_z[36,1]  -29.81    0.06 3.63  -37.81  -32.02  -29.51  -27.30  -23.46
#> log_p_z[36,2]  -21.97    0.04 2.31  -27.11  -23.38  -21.76  -20.37  -18.05
#> log_p_z[36,3]  -10.48    0.04 1.69  -14.65  -11.31  -10.21   -9.33   -8.02
#> log_p_z[36,4]  -14.13    0.07 3.21  -21.23  -16.05  -13.80  -11.80   -8.92
#> log_p_z[37,1]  -18.55    0.04 2.30  -23.50  -19.91  -18.31  -16.92  -14.56
#> log_p_z[37,2]   -7.62    0.01 0.83   -9.41   -8.14   -7.54   -7.04   -6.19
#> log_p_z[37,3]  -19.13    0.06 3.23  -26.43  -21.06  -18.79  -16.80  -13.95
#> log_p_z[37,4]  -21.92    0.06 3.62  -29.73  -24.26  -21.60  -19.30  -15.81
#> log_p_z[38,1]  -25.20    0.04 2.84  -31.20  -27.03  -25.05  -23.20  -20.25
#> log_p_z[38,2]   -9.63    0.02 1.01  -11.88  -10.23   -9.54   -8.94   -7.91
#> log_p_z[38,3]  -11.70    0.06 2.82  -18.54  -13.25  -11.31   -9.63   -7.49
#> log_p_z[38,4]  -21.45    0.07 3.59  -29.03  -23.71  -21.16  -18.89  -15.27
#> log_p_z[39,1]  -29.40    0.07 4.47  -39.72  -32.09  -28.88  -26.26  -22.17
#> log_p_z[39,2]  -16.86    0.03 2.02  -21.27  -18.14  -16.69  -15.41  -13.41
#> log_p_z[39,3]   -7.29    0.02 1.06   -9.49   -7.94   -7.21   -6.56   -5.43
#> log_p_z[39,4]  -18.94    0.08 4.19  -28.35  -21.30  -18.42  -15.96  -12.22
#> log_p_z[40,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_p_z[40,2]  -19.96    0.04 2.32  -25.13  -21.38  -19.78  -18.33  -16.03
#> log_p_z[40,3]  -24.83    0.08 4.31  -34.73  -27.41  -24.34  -21.78  -17.57
#> log_p_z[40,4]  -21.89    0.08 4.34  -31.76  -24.47  -21.33  -18.85  -14.74
#> log_p_z[41,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_p_z[41,2]  -19.96    0.04 2.32  -25.13  -21.38  -19.78  -18.33  -16.03
#> log_p_z[41,3]  -24.83    0.08 4.31  -34.73  -27.41  -24.34  -21.78  -17.57
#> log_p_z[41,4]  -21.89    0.08 4.34  -31.76  -24.47  -21.33  -18.85  -14.74
#> log_p_z[42,1]   -5.99    0.01 0.62   -7.34   -6.36   -5.95   -5.55   -4.91
#> log_p_z[42,2]  -14.90    0.03 1.69  -18.74  -15.90  -14.74  -13.72  -11.96
#> log_p_z[42,3]  -24.31    0.06 3.53  -32.29  -26.45  -24.04  -21.85  -18.20
#> log_p_z[42,4]  -21.91    0.06 3.59  -29.81  -24.17  -21.62  -19.30  -15.85
#> log_p_z[43,1]  -21.88    0.04 2.53  -27.61  -23.46  -21.65  -20.06  -17.63
#> log_p_z[43,2]   -5.67    0.01 0.54   -6.81   -6.01   -5.62   -5.29   -4.72
#> log_p_z[43,3]  -18.40    0.06 3.22  -25.59  -20.24  -18.06  -16.09  -13.17
#> log_p_z[43,4]  -21.63    0.06 3.59  -29.71  -23.87  -21.26  -19.04  -15.69
#> log_p_z[44,1]   -4.41    0.01 0.40   -5.27   -4.67   -4.38   -4.13   -3.72
#> log_p_z[44,2]  -17.21    0.03 1.87  -21.32  -18.40  -17.04  -15.90  -13.94
#> log_p_z[44,3]  -24.30    0.07 3.58  -32.28  -26.51  -24.00  -21.77  -18.10
#> log_p_z[44,4]  -21.89    0.07 3.57  -29.90  -24.06  -21.51  -19.49  -15.87
#> log_p_z[45,1]  -19.63    0.04 2.37  -24.75  -21.09  -19.44  -17.93  -15.61
#> log_p_z[45,2]   -3.05    0.00 0.35   -3.75   -3.26   -3.03   -2.81   -2.41
#> log_p_z[45,3]  -21.21    0.08 4.38  -31.28  -23.65  -20.62  -18.17  -14.31
#> log_p_z[45,4]  -21.90    0.08 4.43  -31.97  -24.51  -21.33  -18.70  -14.62
#> log_lik[1]      -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_lik[2]      -7.58    0.02 1.12  -10.05   -8.29   -7.50   -6.77   -5.64
#> log_lik[3]      -9.92    0.01 0.90  -11.75  -10.50   -9.89   -9.32   -8.22
#> log_lik[4]      -7.54    0.01 0.92   -9.53   -8.11   -7.48   -6.88   -5.95
#> log_lik[5]      -3.71    0.01 0.42   -4.56   -3.99   -3.69   -3.41   -2.93
#> log_lik[6]      -5.45    0.01 0.67   -6.87   -5.88   -5.41   -4.97   -4.26
#> log_lik[7]      -5.99    0.01 0.62   -7.34   -6.36   -5.95   -5.55   -4.91
#> log_lik[8]      -5.72    0.01 0.83   -7.46   -6.27   -5.67   -5.15   -4.25
#> log_lik[9]      -4.53    0.01 0.59   -5.82   -4.89   -4.48   -4.12   -3.54
#> log_lik[10]     -7.15    0.01 0.73   -8.72   -7.60   -7.10   -6.64   -5.87
#> log_lik[11]     -5.84    0.01 1.00   -8.03   -6.46   -5.76   -5.13   -4.12
#> log_lik[12]     -9.34    0.02 1.05  -11.54  -10.01   -9.31   -8.58   -7.45
#> log_lik[13]     -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_lik[14]     -6.22    0.01 0.87   -8.13   -6.75   -6.16   -5.62   -4.74
#> log_lik[15]     -4.41    0.01 0.40   -5.27   -4.67   -4.38   -4.13   -3.72
#> log_lik[16]     -3.78    0.01 0.54   -4.97   -4.11   -3.74   -3.40   -2.84
#> log_lik[17]     -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_lik[18]     -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_lik[19]     -4.65    0.01 0.59   -5.88   -5.01   -4.61   -4.24   -3.61
#> log_lik[20]     -7.09    0.01 0.97   -9.31   -7.66   -7.00   -6.41   -5.42
#> log_lik[21]     -3.05    0.00 0.35   -3.75   -3.26   -3.03   -2.81   -2.41
#> log_lik[22]     -4.65    0.01 0.59   -5.88   -5.01   -4.61   -4.24   -3.61
#> log_lik[23]     -3.71    0.01 0.42   -4.56   -3.99   -3.69   -3.41   -2.93
#> log_lik[24]     -5.80    0.01 0.64   -7.25   -6.16   -5.73   -5.35   -4.74
#> log_lik[25]     -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_lik[26]     -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_lik[27]     -5.67    0.01 0.54   -6.80   -6.01   -5.62   -5.29   -4.72
#> log_lik[28]     -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_lik[29]     -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_lik[30]     -6.83    0.01 0.84   -8.67   -7.35   -6.78   -6.21   -5.42
#> log_lik[31]     -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_lik[32]     -6.06    0.01 0.93   -8.14   -6.64   -5.98   -5.40   -4.50
#> log_lik[33]     -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_lik[34]     -3.05    0.00 0.35   -3.75   -3.26   -3.03   -2.81   -2.41
#> log_lik[35]     -5.52    0.01 0.69   -7.03   -5.95   -5.49   -5.04   -4.32
#> log_lik[36]     -9.95    0.02 1.15  -12.39  -10.70   -9.88   -9.14   -7.92
#> log_lik[37]     -7.62    0.01 0.83   -9.40   -8.14   -7.54   -7.04   -6.19
#> log_lik[38]     -9.01    0.01 0.86  -10.72   -9.58   -8.98   -8.44   -7.32
#> log_lik[39]     -7.28    0.02 1.05   -9.49   -7.94   -7.21   -6.56   -5.43
#> log_lik[40]     -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_lik[41]     -2.21    0.00 0.29   -2.81   -2.40   -2.19   -2.01   -1.67
#> log_lik[42]     -5.99    0.01 0.62   -7.34   -6.36   -5.95   -5.55   -4.91
#> log_lik[43]     -5.67    0.01 0.54   -6.80   -6.01   -5.62   -5.29   -4.72
#> log_lik[44]     -4.41    0.01 0.40   -5.27   -4.67   -4.38   -4.13   -3.72
#> log_lik[45]     -3.05    0.00 0.35   -3.75   -3.26   -3.03   -2.81   -2.41
#> lp__          -464.27    0.16 6.30 -477.58 -468.27 -463.87 -459.83 -453.09
#>               n_eff Rhat
#> pi[1]          7473    1
#> pi[2]          6430    1
#> pi[3]          4212    1
#> pi[4]          4849    1
#> theta[1,1,1]   5656    1
#> theta[1,1,2]   5126    1
#> theta[1,1,3]   4251    1
#> theta[1,1,4]   4250    1
#> theta[1,2,1]   5115    1
#> theta[1,2,2]   6227    1
#> theta[1,2,3]   5542    1
#> theta[1,2,4]   4336    1
#> theta[1,3,1]   5849    1
#> theta[1,3,2]   4070    1
#> theta[1,3,3]   4675    1
#> theta[1,3,4]   4902    1
#> theta[1,4,1]   5810    1
#> theta[1,4,2]   5239    1
#> theta[1,4,3]   4089    1
#> theta[1,4,4]   4665    1
#> theta[2,1,1]   6150    1
#> theta[2,1,2]   5681    1
#> theta[2,1,3]   4439    1
#> theta[2,1,4]   4265    1
#> theta[2,2,1]   6128    1
#> theta[2,2,2]   6864    1
#> theta[2,2,3]   5879    1
#> theta[2,2,4]   6611    1
#> theta[2,3,1]   5194    1
#> theta[2,3,2]   5773    1
#> theta[2,3,3]   5978    1
#> theta[2,3,4]   5430    1
#> theta[2,4,1]   4791    1
#> theta[2,4,2]   5458    1
#> theta[2,4,3]   5628    1
#> theta[2,4,4]   5183    1
#> theta[3,1,1]   5598    1
#> theta[3,1,2]   4342    1
#> theta[3,1,3]   4578    1
#> theta[3,1,4]   4449    1
#> theta[3,2,1]   5454    1
#> theta[3,2,2]   5570    1
#> theta[3,2,3]   5458    1
#> theta[3,2,4]   5119    1
#> theta[3,3,1]   6173    1
#> theta[3,3,2]   5783    1
#> theta[3,3,3]   5980    1
#> theta[3,3,4]   5730    1
#> theta[3,4,1]   5892    1
#> theta[3,4,2]   5717    1
#> theta[3,4,3]   4534    1
#> theta[3,4,4]   5581    1
#> theta[4,1,1]   5988    1
#> theta[4,1,2]   5289    1
#> theta[4,1,3]   4611    1
#> theta[4,1,4]   4274    1
#> theta[4,2,1]   7074    1
#> theta[4,2,2]   5292    1
#> theta[4,2,3]   4691    1
#> theta[4,2,4]   4805    1
#> theta[4,3,1]   6047    1
#> theta[4,3,2]   4622    1
#> theta[4,3,3]   5510    1
#> theta[4,3,4]   5233    1
#> theta[4,4,1]   5811    1
#> theta[4,4,2]   5541    1
#> theta[4,4,3]   6001    1
#> theta[4,4,4]   5514    1
#> theta[5,1,1]   6099    1
#> theta[5,1,2]   5067    1
#> theta[5,1,3]   4379    1
#> theta[5,1,4]   3992    1
#> theta[5,2,1]   6233    1
#> theta[5,2,2]   5610    1
#> theta[5,2,3]   5110    1
#> theta[5,2,4]   4620    1
#> theta[5,3,1]   6620    1
#> theta[5,3,2]   5723    1
#> theta[5,3,3]   5350    1
#> theta[5,3,4]   4845    1
#> theta[5,4,1]   5653    1
#> theta[5,4,2]   6223    1
#> theta[5,4,3]   4840    1
#> theta[5,4,4]   5973    1
#> log_p_z[1,1]   6003    1
#> log_p_z[1,2]   3796    1
#> log_p_z[1,3]   3168    1
#> log_p_z[1,4]   2833    1
#> log_p_z[2,1]   3944    1
#> log_p_z[2,2]   3427    1
#> log_p_z[2,3]   3648    1
#> log_p_z[2,4]   2583    1
#> log_p_z[3,1]   3563    1
#> log_p_z[3,2]   3758    1
#> log_p_z[3,3]   3077    1
#> log_p_z[3,4]   3502    1
#> log_p_z[4,1]   4665    1
#> log_p_z[4,2]   4551    1
#> log_p_z[4,3]   2728    1
#> log_p_z[4,4]   3309    1
#> log_p_z[5,1]   3844    1
#> log_p_z[5,2]   5634    1
#> log_p_z[5,3]   2579    1
#> log_p_z[5,4]   3539    1
#> log_p_z[6,1]   3758    1
#> log_p_z[6,2]   5912    1
#> log_p_z[6,3]   2728    1
#> log_p_z[6,4]   3434    1
#> log_p_z[7,1]   5402    1
#> log_p_z[7,2]   3759    1
#> log_p_z[7,3]   3159    1
#> log_p_z[7,4]   3096    1
#> log_p_z[8,1]   4135    1
#> log_p_z[8,2]   3669    1
#> log_p_z[8,3]   5750    1
#> log_p_z[8,4]   2431    1
#> log_p_z[9,1]   4006    1
#> log_p_z[9,2]   5151    1
#> log_p_z[9,3]   2755    1
#> log_p_z[9,4]   3327    1
#> log_p_z[10,1]  4114    1
#> log_p_z[10,2]  4463    1
#> log_p_z[10,3]  2807    1
#> log_p_z[10,4]  2935    1
#> log_p_z[11,1]  3636    1
#> log_p_z[11,2]  3520    1
#> log_p_z[11,3]  2192    1
#> log_p_z[11,4]  5650    1
#> log_p_z[12,1]  4422    1
#> log_p_z[12,2]  3096    1
#> log_p_z[12,3]  2767    1
#> log_p_z[12,4]  3306    1
#> log_p_z[13,1]  6003    1
#> log_p_z[13,2]  3796    1
#> log_p_z[13,3]  3168    1
#> log_p_z[13,4]  2833    1
#> log_p_z[14,1]  3871    1
#> log_p_z[14,2]  4749    1
#> log_p_z[14,3]  2427    1
#> log_p_z[14,4]  3329    1
#> log_p_z[15,1]  5325    1
#> log_p_z[15,2]  3872    1
#> log_p_z[15,3]  2859    1
#> log_p_z[15,4]  2700    1
#> log_p_z[16,1]  5784    1
#> log_p_z[16,2]  3721    1
#> log_p_z[16,3]  3363    1
#> log_p_z[16,4]  3266    1
#> log_p_z[17,1]  6003    1
#> log_p_z[17,2]  3796    1
#> log_p_z[17,3]  3168    1
#> log_p_z[17,4]  2833    1
#> log_p_z[18,1]  6003    1
#> log_p_z[18,2]  3796    1
#> log_p_z[18,3]  3168    1
#> log_p_z[18,4]  2833    1
#> log_p_z[19,1]  4092    1
#> log_p_z[19,2]  5216    1
#> log_p_z[19,3]  2834    1
#> log_p_z[19,4]  3145    1
#> log_p_z[20,1]  3768    1
#> log_p_z[20,2]  4306    1
#> log_p_z[20,3]  2767    1
#> log_p_z[20,4]  3245    1
#> log_p_z[21,1]  3776    1
#> log_p_z[21,2]  6269    1
#> log_p_z[21,3]  2722    1
#> log_p_z[21,4]  3424    1
#> log_p_z[22,1]  4092    1
#> log_p_z[22,2]  5216    1
#> log_p_z[22,3]  2834    1
#> log_p_z[22,4]  3145    1
#> log_p_z[23,1]  3844    1
#> log_p_z[23,2]  5634    1
#> log_p_z[23,3]  2579    1
#> log_p_z[23,4]  3539    1
#> log_p_z[24,1]  3657    1
#> log_p_z[24,2]  4191    1
#> log_p_z[24,3]  2877    1
#> log_p_z[24,4]  3511    1
#> log_p_z[25,1]  6003    1
#> log_p_z[25,2]  3796    1
#> log_p_z[25,3]  3168    1
#> log_p_z[25,4]  2833    1
#> log_p_z[26,1]  6003    1
#> log_p_z[26,2]  3796    1
#> log_p_z[26,3]  3168    1
#> log_p_z[26,4]  2833    1
#> log_p_z[27,1]  3773    1
#> log_p_z[27,2]  5766    1
#> log_p_z[27,3]  2874    1
#> log_p_z[27,4]  3253    1
#> log_p_z[28,1]  6003    1
#> log_p_z[28,2]  3796    1
#> log_p_z[28,3]  3168    1
#> log_p_z[28,4]  2833    1
#> log_p_z[29,1]  6003    1
#> log_p_z[29,2]  3796    1
#> log_p_z[29,3]  3168    1
#> log_p_z[29,4]  2833    1
#> log_p_z[30,1]  4478    1
#> log_p_z[30,2]  3670    1
#> log_p_z[30,3]  2984    1
#> log_p_z[30,4]  2753    1
#> log_p_z[31,1]  6003    1
#> log_p_z[31,2]  3796    1
#> log_p_z[31,3]  3168    1
#> log_p_z[31,4]  2833    1
#> log_p_z[32,1]  4318    1
#> log_p_z[32,2]  3858    1
#> log_p_z[32,3]  4709    1
#> log_p_z[32,4]  2440    1
#> log_p_z[33,1]  6003    1
#> log_p_z[33,2]  3796    1
#> log_p_z[33,3]  3168    1
#> log_p_z[33,4]  2833    1
#> log_p_z[34,1]  3776    1
#> log_p_z[34,2]  6269    1
#> log_p_z[34,3]  2722    1
#> log_p_z[34,4]  3424    1
#> log_p_z[35,1]  3886    1
#> log_p_z[35,2]  4574    1
#> log_p_z[35,3]  2476    1
#> log_p_z[35,4]  3587    1
#> log_p_z[36,1]  3559    1
#> log_p_z[36,2]  3215    1
#> log_p_z[36,3]  2079    1
#> log_p_z[36,4]  2353    1
#> log_p_z[37,1]  3667    1
#> log_p_z[37,2]  4326    1
#> log_p_z[37,3]  2632    1
#> log_p_z[37,4]  3487    1
#> log_p_z[38,1]  4128    1
#> log_p_z[38,2]  3491    1
#> log_p_z[38,3]  2511    1
#> log_p_z[38,4]  3003    1
#> log_p_z[39,1]  3955    1
#> log_p_z[39,2]  3862    1
#> log_p_z[39,3]  4426    1
#> log_p_z[39,4]  2673    1
#> log_p_z[40,1]  6003    1
#> log_p_z[40,2]  3796    1
#> log_p_z[40,3]  3168    1
#> log_p_z[40,4]  2833    1
#> log_p_z[41,1]  6003    1
#> log_p_z[41,2]  3796    1
#> log_p_z[41,3]  3168    1
#> log_p_z[41,4]  2833    1
#> log_p_z[42,1]  5402    1
#> log_p_z[42,2]  3759    1
#> log_p_z[42,3]  3159    1
#> log_p_z[42,4]  3096    1
#> log_p_z[43,1]  3773    1
#> log_p_z[43,2]  5766    1
#> log_p_z[43,3]  2874    1
#> log_p_z[43,4]  3253    1
#> log_p_z[44,1]  5325    1
#> log_p_z[44,2]  3872    1
#> log_p_z[44,3]  2859    1
#> log_p_z[44,4]  2700    1
#> log_p_z[45,1]  3776    1
#> log_p_z[45,2]  6269    1
#> log_p_z[45,3]  2722    1
#> log_p_z[45,4]  3424    1
#> log_lik[1]     6003    1
#> log_lik[2]     4190    1
#> log_lik[3]     4704    1
#> log_lik[4]     4601    1
#> log_lik[5]     5635    1
#> log_lik[6]     5919    1
#> log_lik[7]     5403    1
#> log_lik[8]     5768    1
#> log_lik[9]     5152    1
#> log_lik[10]    4474    1
#> log_lik[11]    5650    1
#> log_lik[12]    4762    1
#> log_lik[13]    6003    1
#> log_lik[14]    4750    1
#> log_lik[15]    5325    1
#> log_lik[16]    5784    1
#> log_lik[17]    6003    1
#> log_lik[18]    6003    1
#> log_lik[19]    5216    1
#> log_lik[20]    4307    1
#> log_lik[21]    6269    1
#> log_lik[22]    5216    1
#> log_lik[23]    5635    1
#> log_lik[24]    4192    1
#> log_lik[25]    6003    1
#> log_lik[26]    6003    1
#> log_lik[27]    5767    1
#> log_lik[28]    6003    1
#> log_lik[29]    6003    1
#> log_lik[30]    4520    1
#> log_lik[31]    6003    1
#> log_lik[32]    4711    1
#> log_lik[33]    6003    1
#> log_lik[34]    6269    1
#> log_lik[35]    4590    1
#> log_lik[36]    4395    1
#> log_lik[37]    4328    1
#> log_lik[38]    4146    1
#> log_lik[39]    4471    1
#> log_lik[40]    6003    1
#> log_lik[41]    6003    1
#> log_lik[42]    5403    1
#> log_lik[43]    5767    1
#> log_lik[44]    5325    1
#> log_lik[45]    6269    1
#> lp__           1570    1
#> 
#> Samples were drawn using NUTS(diag_e) at Sat Feb 18 00:44:31 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }