Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.26    0.33    0.37    0.42    0.51
#> pi[2]            0.41    0.00 0.07    0.27    0.36    0.41    0.45    0.55
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.08    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.90    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.05    0.75    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.29
#> theta[1,3,3]     0.79    0.00 0.09    0.58    0.74    0.81    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.09    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.42
#> theta[1,4,4]     0.68    0.00 0.14    0.39    0.59    0.69    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.09    0.57    0.69    0.76    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.08    0.05    0.11    0.16    0.22    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.59    0.65    0.76
#> theta[2,2,3]     0.30    0.00 0.08    0.16    0.24    0.30    0.36    0.48
#> theta[2,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,3]     0.63    0.00 0.13    0.35    0.54    0.64    0.72    0.85
#> theta[2,3,4]     0.21    0.00 0.12    0.03    0.13    0.20    0.28    0.47
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,3]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.55    0.67    0.77    0.90
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.24
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.68    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.28
#> theta[3,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.12
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.14    0.25    0.41    0.51    0.61    0.77
#> theta[3,3,4]     0.23    0.00 0.11    0.05    0.15    0.22    0.31    0.49
#> theta[3,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[3,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[3,4,3]     0.14    0.00 0.12    0.00    0.05    0.11    0.21    0.42
#> theta[3,4,4]     0.63    0.00 0.15    0.32    0.53    0.64    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.08    0.66    0.78    0.83    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.08    0.55    0.67    0.73    0.79    0.88
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.09    0.13    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.18
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.39
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.32    0.55    0.67    0.76    0.91
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.82    0.87    0.91    0.97
#> theta[5,1,2]     0.06    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.09    0.14    0.18    0.30
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.41
#> theta[5,3,3]     0.61    0.00 0.13    0.33    0.52    0.62    0.70    0.84
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.14    0.21    0.38
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[5,4,3]     0.14    0.00 0.11    0.00    0.05    0.11    0.20    0.42
#> theta[5,4,4]     0.63    0.00 0.15    0.31    0.52    0.64    0.75    0.90
#> log_p_z[1,1]    -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_p_z[1,2]   -20.02    0.04 2.36  -25.26  -21.48  -19.79  -18.34  -16.00
#> log_p_z[1,3]   -24.89    0.08 4.50  -35.29  -27.38  -24.32  -21.73  -17.60
#> log_p_z[1,4]   -21.91    0.08 4.21  -31.31  -24.40  -21.39  -18.98  -14.89
#> log_p_z[2,1]   -29.77    0.07 4.44  -40.08  -32.22  -29.29  -26.63  -22.60
#> log_p_z[2,2]   -21.00    0.03 2.41  -26.15  -22.49  -20.81  -19.30  -16.85
#> log_p_z[2,3]    -7.68    0.02 1.26  -10.47   -8.41   -7.56   -6.80   -5.61
#> log_p_z[2,4]   -16.34    0.09 4.31  -26.40  -18.83  -15.78  -13.26   -9.38
#> log_p_z[3,1]   -11.60    0.02 1.60  -15.11  -12.61  -11.46  -10.44   -8.98
#> log_p_z[3,2]   -10.82    0.03 1.51  -14.30  -11.68  -10.63   -9.75   -8.32
#> log_p_z[3,3]   -23.39    0.06 3.56  -31.19  -25.54  -22.98  -20.86  -17.33
#> log_p_z[3,4]   -21.89    0.06 3.49  -29.44  -24.03  -21.67  -19.42  -15.86
#> log_p_z[4,1]   -14.65    0.03 1.87  -18.67  -15.83  -14.51  -13.30  -11.33
#> log_p_z[4,2]    -7.59    0.02 1.00   -9.68   -8.15   -7.50   -6.91   -5.94
#> log_p_z[4,3]   -20.71    0.07 4.31  -30.41  -23.33  -20.15  -17.57  -13.93
#> log_p_z[4,4]   -21.89    0.07 4.29  -31.70  -24.47  -21.34  -18.80  -14.92
#> log_p_z[5,1]   -21.44    0.04 2.54  -26.85  -23.10  -21.32  -19.68  -16.94
#> log_p_z[5,2]    -3.72    0.01 0.44   -4.66   -4.00   -3.70   -3.42   -2.91
#> log_p_z[5,3]   -18.66    0.07 4.24  -28.38  -21.09  -18.14  -15.59  -12.04
#> log_p_z[5,4]   -21.89    0.07 4.34  -32.18  -24.41  -21.40  -18.85  -14.76
#> log_p_z[6,1]   -21.44    0.04 2.54  -26.92  -23.04  -21.21  -19.68  -16.92
#> log_p_z[6,2]    -5.47    0.01 0.66   -6.87   -5.90   -5.42   -5.00   -4.32
#> log_p_z[6,3]   -17.45    0.07 4.10  -27.30  -19.68  -16.93  -14.50  -11.07
#> log_p_z[6,4]   -21.66    0.07 4.29  -31.54  -24.13  -21.21  -18.63  -14.58
#> log_p_z[7,1]    -6.00    0.01 0.63   -7.35   -6.40   -5.96   -5.56   -4.93
#> log_p_z[7,2]   -14.96    0.03 1.75  -18.84  -16.04  -14.77  -13.71  -12.02
#> log_p_z[7,3]   -24.33    0.06 3.68  -32.28  -26.56  -23.98  -21.79  -17.96
#> log_p_z[7,4]   -21.94    0.06 3.49  -29.57  -24.07  -21.68  -19.44  -15.84
#> log_p_z[8,1]   -29.74    0.07 4.39  -40.18  -32.11  -29.26  -26.72  -22.50
#> log_p_z[8,2]   -18.30    0.03 2.12  -22.93  -19.62  -18.18  -16.82  -14.59
#> log_p_z[8,3]    -5.73    0.01 0.85   -7.56   -6.27   -5.69   -5.13   -4.26
#> log_p_z[8,4]   -18.55    0.09 4.31  -28.14  -21.10  -18.06  -15.47  -11.61
#> log_p_z[9,1]   -19.96    0.04 2.34  -24.94  -21.41  -19.83  -18.28  -15.93
#> log_p_z[9,2]    -4.52    0.01 0.58   -5.79   -4.87   -4.48   -4.12   -3.48
#> log_p_z[9,3]   -19.65    0.07 4.32  -29.54  -22.15  -19.14  -16.57  -12.93
#> log_p_z[9,4]   -21.70    0.07 4.30  -31.45  -24.21  -21.16  -18.66  -14.58
#> log_p_z[10,1]  -22.23    0.04 2.51  -27.87  -23.77  -22.03  -20.44  -18.02
#> log_p_z[10,2]   -7.12    0.01 0.73   -8.71   -7.57   -7.05   -6.63   -5.85
#> log_p_z[10,3]  -16.84    0.05 3.18  -24.00  -18.70  -16.49  -14.61  -11.67
#> log_p_z[10,4]  -21.37    0.06 3.56  -29.44  -23.51  -21.07  -18.85  -15.45
#> log_p_z[11,1]  -29.71    0.07 4.31  -39.49  -32.22  -29.32  -26.66  -22.55
#> log_p_z[11,2]  -29.26    0.07 4.29  -39.07  -31.74  -28.80  -26.20  -22.21
#> log_p_z[11,3]  -18.34    0.08 3.50  -26.79  -20.06  -17.76  -15.99  -13.05
#> log_p_z[11,4]   -5.83    0.01 1.00   -7.98   -6.46   -5.75   -5.11   -4.12
#> log_p_z[12,1]  -22.90    0.04 2.64  -28.60  -24.53  -22.71  -21.07  -18.26
#> log_p_z[12,2]   -9.61    0.02 1.28  -12.75  -10.30   -9.46   -8.72   -7.60
#> log_p_z[12,3]  -15.11    0.07 4.03  -24.41  -17.39  -14.59  -12.14   -8.88
#> log_p_z[12,4]  -19.30    0.07 4.24  -29.21  -21.78  -18.83  -16.29  -12.41
#> log_p_z[13,1]   -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_p_z[13,2]  -20.02    0.04 2.36  -25.26  -21.48  -19.79  -18.34  -16.00
#> log_p_z[13,3]  -24.89    0.08 4.50  -35.29  -27.38  -24.32  -21.73  -17.60
#> log_p_z[13,4]  -21.91    0.08 4.21  -31.31  -24.40  -21.39  -18.98  -14.89
#> log_p_z[14,1]  -19.05    0.04 2.38  -24.12  -20.54  -18.91  -17.35  -14.91
#> log_p_z[14,2]   -6.25    0.01 0.86   -8.21   -6.76   -6.17   -5.64   -4.80
#> log_p_z[14,3]  -18.63    0.07 4.23  -28.09  -21.18  -18.07  -15.59  -12.01
#> log_p_z[14,4]  -21.94    0.07 4.36  -32.16  -24.47  -21.41  -18.83  -14.72
#> log_p_z[15,1]   -4.42    0.01 0.41   -5.28   -4.68   -4.39   -4.12   -3.71
#> log_p_z[15,2]  -17.26    0.03 1.91  -21.55  -18.43  -17.10  -15.92  -13.96
#> log_p_z[15,3]  -24.33    0.06 3.66  -32.47  -26.56  -24.00  -21.74  -18.10
#> log_p_z[15,4]  -21.91    0.06 3.49  -29.53  -24.01  -21.61  -19.40  -15.98
#> log_p_z[16,1]   -3.80    0.01 0.55   -4.96   -4.16   -3.75   -3.40   -2.85
#> log_p_z[16,2]  -17.72    0.04 2.23  -22.80  -19.06  -17.48  -16.14  -14.04
#> log_p_z[16,3]  -24.89    0.08 4.54  -35.31  -27.49  -24.29  -21.70  -17.56
#> log_p_z[16,4]  -21.93    0.07 4.22  -31.67  -24.33  -21.52  -18.96  -14.99
#> log_p_z[17,1]   -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_p_z[17,2]  -20.02    0.04 2.36  -25.26  -21.48  -19.79  -18.34  -16.00
#> log_p_z[17,3]  -24.89    0.08 4.50  -35.29  -27.38  -24.32  -21.73  -17.60
#> log_p_z[17,4]  -21.91    0.08 4.21  -31.31  -24.40  -21.39  -18.98  -14.89
#> log_p_z[18,1]   -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_p_z[18,2]  -20.02    0.04 2.36  -25.26  -21.48  -19.79  -18.34  -16.00
#> log_p_z[18,3]  -24.89    0.08 4.50  -35.29  -27.38  -24.32  -21.73  -17.60
#> log_p_z[18,4]  -21.91    0.08 4.21  -31.31  -24.40  -21.39  -18.98  -14.89
#> log_p_z[19,1]  -16.41    0.03 1.97  -20.65  -17.67  -16.30  -15.00  -12.99
#> log_p_z[19,2]   -4.65    0.01 0.62   -5.94   -5.06   -4.62   -4.21   -3.58
#> log_p_z[19,3]  -22.14    0.07 4.43  -32.14  -24.73  -21.60  -18.98  -14.98
#> log_p_z[19,4]  -21.92    0.08 4.28  -31.82  -24.39  -21.46  -18.90  -14.82
#> log_p_z[20,1]  -18.02    0.03 2.22  -22.95  -19.39  -17.87  -16.46  -14.12
#> log_p_z[20,2]   -7.09    0.01 0.96   -9.24   -7.65   -6.99   -6.42   -5.43
#> log_p_z[20,3]  -19.98    0.07 4.22  -30.03  -22.37  -19.38  -16.98  -13.41
#> log_p_z[20,4]  -21.66    0.07 4.27  -31.29  -24.18  -21.17  -18.63  -14.64
#> log_p_z[21,1]  -19.61    0.04 2.33  -24.57  -21.06  -19.48  -17.96  -15.52
#> log_p_z[21,2]   -3.04    0.00 0.36   -3.80   -3.27   -3.02   -2.79   -2.38
#> log_p_z[21,3]  -21.18    0.07 4.35  -31.21  -23.58  -20.68  -18.06  -14.25
#> log_p_z[21,4]  -21.92    0.07 4.31  -31.78  -24.44  -21.52  -18.85  -14.75
#> log_p_z[22,1]  -16.41    0.03 1.97  -20.65  -17.67  -16.30  -15.00  -12.99
#> log_p_z[22,2]   -4.65    0.01 0.62   -5.94   -5.06   -4.62   -4.21   -3.58
#> log_p_z[22,3]  -22.14    0.07 4.43  -32.14  -24.73  -21.60  -18.98  -14.98
#> log_p_z[22,4]  -21.92    0.08 4.28  -31.82  -24.39  -21.46  -18.90  -14.82
#> log_p_z[23,1]  -21.44    0.04 2.54  -26.85  -23.10  -21.32  -19.68  -16.94
#> log_p_z[23,2]   -3.72    0.01 0.44   -4.66   -4.00   -3.70   -3.42   -2.91
#> log_p_z[23,3]  -18.66    0.07 4.24  -28.38  -21.09  -18.14  -15.59  -12.04
#> log_p_z[23,4]  -21.89    0.07 4.34  -32.18  -24.41  -21.40  -18.85  -14.76
#> log_p_z[24,1]  -17.40    0.03 2.12  -22.11  -18.71  -17.26  -15.92  -13.75
#> log_p_z[24,2]   -5.80    0.01 0.62   -7.23   -6.15   -5.74   -5.38   -4.75
#> log_p_z[24,3]  -21.74    0.06 3.47  -29.29  -23.87  -21.38  -19.32  -15.98
#> log_p_z[24,4]  -21.91    0.06 3.56  -29.52  -24.14  -21.62  -19.44  -15.73
#> log_p_z[25,1]   -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_p_z[25,2]  -20.02    0.04 2.36  -25.26  -21.48  -19.79  -18.34  -16.00
#> log_p_z[25,3]  -24.89    0.08 4.50  -35.29  -27.38  -24.32  -21.73  -17.60
#> log_p_z[25,4]  -21.91    0.08 4.21  -31.31  -24.40  -21.39  -18.98  -14.89
#> log_p_z[26,1]   -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_p_z[26,2]  -20.02    0.04 2.36  -25.26  -21.48  -19.79  -18.34  -16.00
#> log_p_z[26,3]  -24.89    0.08 4.50  -35.29  -27.38  -24.32  -21.73  -17.60
#> log_p_z[26,4]  -21.91    0.08 4.21  -31.31  -24.40  -21.39  -18.98  -14.89
#> log_p_z[27,1]  -21.89    0.04 2.48  -27.21  -23.45  -21.66  -20.14  -17.61
#> log_p_z[27,2]   -5.64    0.01 0.53   -6.80   -5.97   -5.61   -5.27   -4.72
#> log_p_z[27,3]  -18.37    0.05 3.23  -25.77  -20.28  -18.04  -16.09  -13.00
#> log_p_z[27,4]  -21.58    0.06 3.55  -29.47  -23.71  -21.35  -19.08  -15.70
#> log_p_z[28,1]   -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_p_z[28,2]  -20.02    0.04 2.36  -25.26  -21.48  -19.79  -18.34  -16.00
#> log_p_z[28,3]  -24.89    0.08 4.50  -35.29  -27.38  -24.32  -21.73  -17.60
#> log_p_z[28,4]  -21.91    0.08 4.21  -31.31  -24.40  -21.39  -18.98  -14.89
#> log_p_z[29,1]   -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_p_z[29,2]  -20.02    0.04 2.36  -25.26  -21.48  -19.79  -18.34  -16.00
#> log_p_z[29,3]  -24.89    0.08 4.50  -35.29  -27.38  -24.32  -21.73  -17.60
#> log_p_z[29,4]  -21.91    0.08 4.21  -31.31  -24.40  -21.39  -18.98  -14.89
#> log_p_z[30,1]   -6.81    0.01 0.84   -8.68   -7.31   -6.72   -6.23   -5.43
#> log_p_z[30,2]  -14.73    0.03 1.78  -18.71  -15.82  -14.59  -13.46  -11.70
#> log_p_z[30,3]  -24.36    0.06 3.61  -32.31  -26.54  -24.03  -21.83  -18.33
#> log_p_z[30,4]  -21.87    0.06 3.47  -29.48  -24.00  -21.62  -19.47  -15.95
#> log_p_z[31,1]   -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_p_z[31,2]  -20.02    0.04 2.36  -25.26  -21.48  -19.79  -18.34  -16.00
#> log_p_z[31,3]  -24.89    0.08 4.50  -35.29  -27.38  -24.32  -21.73  -17.60
#> log_p_z[31,4]  -21.91    0.08 4.21  -31.31  -24.40  -21.39  -18.98  -14.89
#> log_p_z[32,1]  -29.73    0.07 4.44  -40.31  -32.34  -29.18  -26.62  -22.51
#> log_p_z[32,2]  -14.83    0.02 1.78  -18.68  -15.95  -14.68  -13.59  -11.77
#> log_p_z[32,3]   -6.07    0.01 0.94   -8.14   -6.63   -5.99   -5.43   -4.47
#> log_p_z[32,4]  -20.70    0.09 4.46  -30.73  -23.34  -20.16  -17.56  -13.47
#> log_p_z[33,1]   -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_p_z[33,2]  -20.02    0.04 2.36  -25.26  -21.48  -19.79  -18.34  -16.00
#> log_p_z[33,3]  -24.89    0.08 4.50  -35.29  -27.38  -24.32  -21.73  -17.60
#> log_p_z[33,4]  -21.91    0.08 4.21  -31.31  -24.40  -21.39  -18.98  -14.89
#> log_p_z[34,1]  -19.61    0.04 2.33  -24.57  -21.06  -19.48  -17.96  -15.52
#> log_p_z[34,2]   -3.04    0.00 0.36   -3.80   -3.27   -3.02   -2.79   -2.38
#> log_p_z[34,3]  -21.18    0.07 4.35  -31.21  -23.58  -20.68  -18.06  -14.25
#> log_p_z[34,4]  -21.92    0.07 4.31  -31.78  -24.44  -21.52  -18.85  -14.75
#> log_p_z[35,1]  -22.57    0.04 2.72  -28.47  -24.28  -22.43  -20.63  -17.86
#> log_p_z[35,2]   -5.54    0.01 0.70   -7.08   -5.94   -5.48   -5.06   -4.33
#> log_p_z[35,3]  -16.03    0.07 4.08  -25.37  -18.47  -15.45  -13.00   -9.82
#> log_p_z[35,4]  -21.91    0.07 4.33  -32.02  -24.44  -21.46  -18.84  -14.84
#> log_p_z[36,1]  -29.75    0.06 3.63  -37.70  -31.96  -29.45  -27.14  -23.77
#> log_p_z[36,2]  -21.89    0.04 2.31  -26.93  -23.26  -21.66  -20.29  -17.95
#> log_p_z[36,3]  -10.57    0.04 1.77  -15.13  -11.45  -10.31   -9.35   -7.90
#> log_p_z[36,4]  -13.98    0.07 3.36  -21.49  -16.01  -13.63  -11.49   -8.63
#> log_p_z[37,1]  -18.53    0.03 2.32  -23.65  -19.89  -18.31  -16.90  -14.64
#> log_p_z[37,2]   -7.62    0.01 0.80   -9.39   -8.11   -7.55   -7.06   -6.23
#> log_p_z[37,3]  -19.11    0.06 3.25  -26.19  -21.06  -18.72  -16.79  -13.75
#> log_p_z[37,4]  -21.93    0.06 3.54  -29.26  -24.20  -21.67  -19.39  -15.80
#> log_p_z[38,1]  -25.19    0.04 2.87  -31.49  -26.94  -24.97  -23.17  -20.29
#> log_p_z[38,2]   -9.62    0.01 1.02  -11.85  -10.24   -9.53   -8.90   -7.85
#> log_p_z[38,3]  -11.69    0.05 2.78  -17.97  -13.35  -11.29   -9.58   -7.54
#> log_p_z[38,4]  -21.37    0.06 3.55  -29.36  -23.57  -21.07  -18.88  -15.24
#> log_p_z[39,1]  -29.39    0.07 4.36  -39.61  -31.70  -28.95  -26.44  -22.10
#> log_p_z[39,2]  -16.82    0.03 2.02  -21.20  -18.05  -16.69  -15.43  -13.31
#> log_p_z[39,3]   -7.26    0.01 1.08   -9.63   -7.94   -7.19   -6.50   -5.37
#> log_p_z[39,4]  -18.76    0.08 4.22  -28.69  -21.12  -18.26  -15.72  -12.12
#> log_p_z[40,1]   -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_p_z[40,2]  -20.02    0.04 2.36  -25.26  -21.48  -19.79  -18.34  -16.00
#> log_p_z[40,3]  -24.89    0.08 4.50  -35.29  -27.38  -24.32  -21.73  -17.60
#> log_p_z[40,4]  -21.91    0.08 4.21  -31.31  -24.40  -21.39  -18.98  -14.89
#> log_p_z[41,1]   -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_p_z[41,2]  -20.02    0.04 2.36  -25.26  -21.48  -19.79  -18.34  -16.00
#> log_p_z[41,3]  -24.89    0.08 4.50  -35.29  -27.38  -24.32  -21.73  -17.60
#> log_p_z[41,4]  -21.91    0.08 4.21  -31.31  -24.40  -21.39  -18.98  -14.89
#> log_p_z[42,1]   -6.00    0.01 0.63   -7.35   -6.40   -5.96   -5.56   -4.93
#> log_p_z[42,2]  -14.96    0.03 1.75  -18.84  -16.04  -14.77  -13.71  -12.02
#> log_p_z[42,3]  -24.33    0.06 3.68  -32.28  -26.56  -23.98  -21.79  -17.96
#> log_p_z[42,4]  -21.94    0.06 3.49  -29.57  -24.07  -21.68  -19.44  -15.84
#> log_p_z[43,1]  -21.89    0.04 2.48  -27.21  -23.45  -21.66  -20.14  -17.61
#> log_p_z[43,2]   -5.64    0.01 0.53   -6.80   -5.97   -5.61   -5.27   -4.72
#> log_p_z[43,3]  -18.37    0.05 3.23  -25.77  -20.28  -18.04  -16.09  -13.00
#> log_p_z[43,4]  -21.58    0.06 3.55  -29.47  -23.71  -21.35  -19.08  -15.70
#> log_p_z[44,1]   -4.42    0.01 0.41   -5.28   -4.68   -4.39   -4.12   -3.71
#> log_p_z[44,2]  -17.26    0.03 1.91  -21.55  -18.43  -17.10  -15.92  -13.96
#> log_p_z[44,3]  -24.33    0.06 3.66  -32.47  -26.56  -24.00  -21.74  -18.10
#> log_p_z[44,4]  -21.91    0.06 3.49  -29.53  -24.01  -21.61  -19.40  -15.98
#> log_p_z[45,1]  -19.61    0.04 2.33  -24.57  -21.06  -19.48  -17.96  -15.52
#> log_p_z[45,2]   -3.04    0.00 0.36   -3.80   -3.27   -3.02   -2.79   -2.38
#> log_p_z[45,3]  -21.18    0.07 4.35  -31.21  -23.58  -20.68  -18.06  -14.25
#> log_p_z[45,4]  -21.92    0.07 4.31  -31.78  -24.44  -21.52  -18.85  -14.75
#> log_lik[1]      -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_lik[2]      -7.62    0.02 1.16  -10.08   -8.36   -7.55   -6.79   -5.61
#> log_lik[3]      -9.95    0.01 0.92  -11.87  -10.55   -9.89   -9.32   -8.22
#> log_lik[4]      -7.58    0.02 0.97   -9.66   -8.14   -7.50   -6.91   -5.94
#> log_lik[5]      -3.72    0.01 0.44   -4.66   -4.00   -3.70   -3.42   -2.91
#> log_lik[6]      -5.47    0.01 0.66   -6.86   -5.90   -5.42   -5.00   -4.32
#> log_lik[7]      -6.00    0.01 0.62   -7.35   -6.40   -5.96   -5.56   -4.93
#> log_lik[8]      -5.73    0.01 0.85   -7.56   -6.27   -5.69   -5.13   -4.26
#> log_lik[9]      -4.52    0.01 0.58   -5.79   -4.87   -4.48   -4.12   -3.48
#> log_lik[10]     -7.12    0.01 0.72   -8.70   -7.56   -7.05   -6.63   -5.85
#> log_lik[11]     -5.83    0.01 1.00   -7.98   -6.46   -5.75   -5.11   -4.12
#> log_lik[12]     -9.34    0.01 1.03  -11.57   -9.99   -9.29   -8.62   -7.55
#> log_lik[13]     -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_lik[14]     -6.25    0.01 0.86   -8.20   -6.76   -6.17   -5.63   -4.80
#> log_lik[15]     -4.42    0.01 0.41   -5.28   -4.68   -4.39   -4.12   -3.71
#> log_lik[16]     -3.80    0.01 0.55   -4.96   -4.16   -3.75   -3.40   -2.85
#> log_lik[17]     -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_lik[18]     -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_lik[19]     -4.65    0.01 0.62   -5.94   -5.06   -4.62   -4.21   -3.58
#> log_lik[20]     -7.09    0.01 0.95   -9.23   -7.65   -6.99   -6.42   -5.43
#> log_lik[21]     -3.04    0.00 0.36   -3.80   -3.27   -3.02   -2.79   -2.38
#> log_lik[22]     -4.65    0.01 0.62   -5.94   -5.06   -4.62   -4.21   -3.58
#> log_lik[23]     -3.72    0.01 0.44   -4.66   -4.00   -3.70   -3.42   -2.91
#> log_lik[24]     -5.80    0.01 0.62   -7.23   -6.15   -5.74   -5.38   -4.75
#> log_lik[25]     -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_lik[26]     -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_lik[27]     -5.64    0.01 0.53   -6.80   -5.97   -5.61   -5.27   -4.72
#> log_lik[28]     -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_lik[29]     -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_lik[30]     -6.81    0.01 0.83   -8.67   -7.30   -6.72   -6.23   -5.43
#> log_lik[31]     -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_lik[32]     -6.07    0.01 0.94   -8.13   -6.63   -5.99   -5.43   -4.47
#> log_lik[33]     -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_lik[34]     -3.04    0.00 0.36   -3.80   -3.27   -3.02   -2.79   -2.38
#> log_lik[35]     -5.53    0.01 0.70   -7.08   -5.94   -5.48   -5.06   -4.33
#> log_lik[36]     -9.93    0.02 1.19  -12.46  -10.68   -9.86   -9.10   -7.79
#> log_lik[37]     -7.62    0.01 0.80   -9.39   -8.11   -7.55   -7.06   -6.23
#> log_lik[38]     -9.00    0.01 0.85  -10.73   -9.53   -8.97   -8.44   -7.35
#> log_lik[39]     -7.26    0.01 1.07   -9.60   -7.94   -7.19   -6.50   -5.37
#> log_lik[40]     -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_lik[41]     -2.21    0.00 0.31   -2.86   -2.40   -2.20   -1.99   -1.64
#> log_lik[42]     -6.00    0.01 0.62   -7.35   -6.40   -5.96   -5.56   -4.93
#> log_lik[43]     -5.64    0.01 0.53   -6.80   -5.97   -5.61   -5.27   -4.72
#> log_lik[44]     -4.42    0.01 0.41   -5.28   -4.68   -4.39   -4.12   -3.71
#> log_lik[45]     -3.04    0.00 0.36   -3.80   -3.27   -3.02   -2.79   -2.38
#> lp__          -464.01    0.17 6.19 -477.24 -467.91 -463.79 -459.66 -452.79
#>               n_eff Rhat
#> pi[1]          7437    1
#> pi[2]          7917    1
#> pi[3]          5920    1
#> pi[4]          6173    1
#> theta[1,1,1]   6769    1
#> theta[1,1,2]   6906    1
#> theta[1,1,3]   4673    1
#> theta[1,1,4]   5248    1
#> theta[1,2,1]   7059    1
#> theta[1,2,2]   7927    1
#> theta[1,2,3]   7023    1
#> theta[1,2,4]   5465    1
#> theta[1,3,1]   6583    1
#> theta[1,3,2]   4168    1
#> theta[1,3,3]   5339    1
#> theta[1,3,4]   5624    1
#> theta[1,4,1]   7101    1
#> theta[1,4,2]   6233    1
#> theta[1,4,3]   3695    1
#> theta[1,4,4]   4721    1
#> theta[2,1,1]   8004    1
#> theta[2,1,2]   6704    1
#> theta[2,1,3]   4707    1
#> theta[2,1,4]   5008    1
#> theta[2,2,1]   7107    1
#> theta[2,2,2]   6541    1
#> theta[2,2,3]   6689    1
#> theta[2,2,4]   6409    1
#> theta[2,3,1]   6462    1
#> theta[2,3,2]   6799    1
#> theta[2,3,3]   6088    1
#> theta[2,3,4]   5089    1
#> theta[2,4,1]   6914    1
#> theta[2,4,2]   7243    1
#> theta[2,4,3]   6107    1
#> theta[2,4,4]   5538    1
#> theta[3,1,1]   6405    1
#> theta[3,1,2]   4218    1
#> theta[3,1,3]   4242    1
#> theta[3,1,4]   5600    1
#> theta[3,2,1]   6017    1
#> theta[3,2,2]   7261    1
#> theta[3,2,3]   7018    1
#> theta[3,2,4]   5595    1
#> theta[3,3,1]   5844    1
#> theta[3,3,2]   6234    1
#> theta[3,3,3]   7107    1
#> theta[3,3,4]   7397    1
#> theta[3,4,1]   6694    1
#> theta[3,4,2]   6355    1
#> theta[3,4,3]   4400    1
#> theta[3,4,4]   5173    1
#> theta[4,1,1]   7288    1
#> theta[4,1,2]   5442    1
#> theta[4,1,3]   4762    1
#> theta[4,1,4]   4695    1
#> theta[4,2,1]   8197    1
#> theta[4,2,2]   7863    1
#> theta[4,2,3]   6669    1
#> theta[4,2,4]   6447    1
#> theta[4,3,1]   6519    1
#> theta[4,3,2]   6905    1
#> theta[4,3,3]   6606    1
#> theta[4,3,4]   5774    1
#> theta[4,4,1]   6960    1
#> theta[4,4,2]   7409    1
#> theta[4,4,3]   6733    1
#> theta[4,4,4]   7106    1
#> theta[5,1,1]   8079    1
#> theta[5,1,2]   5000    1
#> theta[5,1,3]   5084    1
#> theta[5,1,4]   5572    1
#> theta[5,2,1]   7454    1
#> theta[5,2,2]   7326    1
#> theta[5,2,3]   6246    1
#> theta[5,2,4]   4804    1
#> theta[5,3,1]   6046    1
#> theta[5,3,2]   7527    1
#> theta[5,3,3]   6916    1
#> theta[5,3,4]   6986    1
#> theta[5,4,1]   6877    1
#> theta[5,4,2]   6701    1
#> theta[5,4,3]   5087    1
#> theta[5,4,4]   5823    1
#> log_p_z[1,1]   7578    1
#> log_p_z[1,2]   3869    1
#> log_p_z[1,3]   3406    1
#> log_p_z[1,4]   3070    1
#> log_p_z[2,1]   4118    1
#> log_p_z[2,2]   4821    1
#> log_p_z[2,3]   3406    1
#> log_p_z[2,4]   2501    1
#> log_p_z[3,1]   4286    1
#> log_p_z[3,2]   3359    1
#> log_p_z[3,3]   3417    1
#> log_p_z[3,4]   3615    1
#> log_p_z[4,1]   4971    1
#> log_p_z[4,2]   2536    1
#> log_p_z[4,3]   3670    1
#> log_p_z[4,4]   3466    1
#> log_p_z[5,1]   4362    1
#> log_p_z[5,2]   7085    1
#> log_p_z[5,3]   3297    1
#> log_p_z[5,4]   3570    1
#> log_p_z[6,1]   4758    1
#> log_p_z[6,2]   6446    1
#> log_p_z[6,3]   3374    1
#> log_p_z[6,4]   3451    1
#> log_p_z[7,1]   6431    1
#> log_p_z[7,2]   3475    1
#> log_p_z[7,3]   3486    1
#> log_p_z[7,4]   3315    1
#> log_p_z[8,1]   4299    1
#> log_p_z[8,2]   5118    1
#> log_p_z[8,3]   5393    1
#> log_p_z[8,4]   2528    1
#> log_p_z[9,1]   4401    1
#> log_p_z[9,2]   5901    1
#> log_p_z[9,3]   3438    1
#> log_p_z[9,4]   3373    1
#> log_p_z[10,1]  4456    1
#> log_p_z[10,2]  5764    1
#> log_p_z[10,3]  3562    1
#> log_p_z[10,4]  3054    1
#> log_p_z[11,1]  4133    1
#> log_p_z[11,2]  4093    1
#> log_p_z[11,3]  1985    1
#> log_p_z[11,4]  5572    1
#> log_p_z[12,1]  4772    1
#> log_p_z[12,2]  4018    1
#> log_p_z[12,3]  3388    1
#> log_p_z[12,4]  3342    1
#> log_p_z[13,1]  7578    1
#> log_p_z[13,2]  3869    1
#> log_p_z[13,3]  3406    1
#> log_p_z[13,4]  3070    1
#> log_p_z[14,1]  4580    1
#> log_p_z[14,2]  5413    1
#> log_p_z[14,3]  3191    1
#> log_p_z[14,4]  3399    1
#> log_p_z[15,1]  6403    1
#> log_p_z[15,2]  3635    1
#> log_p_z[15,3]  3562    1
#> log_p_z[15,4]  3122    1
#> log_p_z[16,1]  6372    1
#> log_p_z[16,2]  3845    1
#> log_p_z[16,3]  3333    1
#> log_p_z[16,4]  3172    1
#> log_p_z[17,1]  7578    1
#> log_p_z[17,2]  3869    1
#> log_p_z[17,3]  3406    1
#> log_p_z[17,4]  3070    1
#> log_p_z[18,1]  7578    1
#> log_p_z[18,2]  3869    1
#> log_p_z[18,3]  3406    1
#> log_p_z[18,4]  3070    1
#> log_p_z[19,1]  4610    1
#> log_p_z[19,2]  5533    1
#> log_p_z[19,3]  3571    1
#> log_p_z[19,4]  3179    1
#> log_p_z[20,1]  4771    1
#> log_p_z[20,2]  5638    1
#> log_p_z[20,3]  3482    1
#> log_p_z[20,4]  3469    1
#> log_p_z[21,1]  4292    1
#> log_p_z[21,2]  9183    1
#> log_p_z[21,3]  3445    1
#> log_p_z[21,4]  3519    1
#> log_p_z[22,1]  4610    1
#> log_p_z[22,2]  5533    1
#> log_p_z[22,3]  3571    1
#> log_p_z[22,4]  3179    1
#> log_p_z[23,1]  4362    1
#> log_p_z[23,2]  7085    1
#> log_p_z[23,3]  3297    1
#> log_p_z[23,4]  3570    1
#> log_p_z[24,1]  4068    1
#> log_p_z[24,2]  5168    1
#> log_p_z[24,3]  3366    1
#> log_p_z[24,4]  3483    1
#> log_p_z[25,1]  7578    1
#> log_p_z[25,2]  3869    1
#> log_p_z[25,3]  3406    1
#> log_p_z[25,4]  3070    1
#> log_p_z[26,1]  7578    1
#> log_p_z[26,2]  3869    1
#> log_p_z[26,3]  3406    1
#> log_p_z[26,4]  3070    1
#> log_p_z[27,1]  4325    1
#> log_p_z[27,2]  6501    1
#> log_p_z[27,3]  3589    1
#> log_p_z[27,4]  3358    1
#> log_p_z[28,1]  7578    1
#> log_p_z[28,2]  3869    1
#> log_p_z[28,3]  3406    1
#> log_p_z[28,4]  3070    1
#> log_p_z[29,1]  7578    1
#> log_p_z[29,2]  3869    1
#> log_p_z[29,3]  3406    1
#> log_p_z[29,4]  3070    1
#> log_p_z[30,1]  4809    1
#> log_p_z[30,2]  3534    1
#> log_p_z[30,3]  3461    1
#> log_p_z[30,4]  3265    1
#> log_p_z[31,1]  7578    1
#> log_p_z[31,2]  3869    1
#> log_p_z[31,3]  3406    1
#> log_p_z[31,4]  3070    1
#> log_p_z[32,1]  4485    1
#> log_p_z[32,2]  5710    1
#> log_p_z[32,3]  4747    1
#> log_p_z[32,4]  2610    1
#> log_p_z[33,1]  7578    1
#> log_p_z[33,2]  3869    1
#> log_p_z[33,3]  3406    1
#> log_p_z[33,4]  3070    1
#> log_p_z[34,1]  4292    1
#> log_p_z[34,2]  9183    1
#> log_p_z[34,3]  3445    1
#> log_p_z[34,4]  3519    1
#> log_p_z[35,1]  4905    1
#> log_p_z[35,2]  5694    1
#> log_p_z[35,3]  3297    1
#> log_p_z[35,4]  3521    1
#> log_p_z[36,1]  3336    1
#> log_p_z[36,2]  3821    1
#> log_p_z[36,3]  1900    1
#> log_p_z[36,4]  2198    1
#> log_p_z[37,1]  4561    1
#> log_p_z[37,2]  5109    1
#> log_p_z[37,3]  3263    1
#> log_p_z[37,4]  3513    1
#> log_p_z[38,1]  4568    1
#> log_p_z[38,2]  4846    1
#> log_p_z[38,3]  3278    1
#> log_p_z[38,4]  3277    1
#> log_p_z[39,1]  4401    1
#> log_p_z[39,2]  5339    1
#> log_p_z[39,3]  5173    1
#> log_p_z[39,4]  2824    1
#> log_p_z[40,1]  7578    1
#> log_p_z[40,2]  3869    1
#> log_p_z[40,3]  3406    1
#> log_p_z[40,4]  3070    1
#> log_p_z[41,1]  7578    1
#> log_p_z[41,2]  3869    1
#> log_p_z[41,3]  3406    1
#> log_p_z[41,4]  3070    1
#> log_p_z[42,1]  6431    1
#> log_p_z[42,2]  3475    1
#> log_p_z[42,3]  3486    1
#> log_p_z[42,4]  3315    1
#> log_p_z[43,1]  4325    1
#> log_p_z[43,2]  6501    1
#> log_p_z[43,3]  3589    1
#> log_p_z[43,4]  3358    1
#> log_p_z[44,1]  6403    1
#> log_p_z[44,2]  3635    1
#> log_p_z[44,3]  3562    1
#> log_p_z[44,4]  3122    1
#> log_p_z[45,1]  4292    1
#> log_p_z[45,2]  9183    1
#> log_p_z[45,3]  3445    1
#> log_p_z[45,4]  3519    1
#> log_lik[1]     7578    1
#> log_lik[2]     4521    1
#> log_lik[3]     5570    1
#> log_lik[4]     3177    1
#> log_lik[5]     7084    1
#> log_lik[6]     6452    1
#> log_lik[7]     6432    1
#> log_lik[8]     5390    1
#> log_lik[9]     5901    1
#> log_lik[10]    5797    1
#> log_lik[11]    5573    1
#> log_lik[12]    5877    1
#> log_lik[13]    7578    1
#> log_lik[14]    5415    1
#> log_lik[15]    6403    1
#> log_lik[16]    6373    1
#> log_lik[17]    7578    1
#> log_lik[18]    7578    1
#> log_lik[19]    5533    1
#> log_lik[20]    5641    1
#> log_lik[21]    9183    1
#> log_lik[22]    5533    1
#> log_lik[23]    7084    1
#> log_lik[24]    5169    1
#> log_lik[25]    7578    1
#> log_lik[26]    7578    1
#> log_lik[27]    6501    1
#> log_lik[28]    7578    1
#> log_lik[29]    7578    1
#> log_lik[30]    4820    1
#> log_lik[31]    7578    1
#> log_lik[32]    4758    1
#> log_lik[33]    7578    1
#> log_lik[34]    9183    1
#> log_lik[35]    5737    1
#> log_lik[36]    4796    1
#> log_lik[37]    5113    1
#> log_lik[38]    4207    1
#> log_lik[39]    5200    1
#> log_lik[40]    7578    1
#> log_lik[41]    7578    1
#> log_lik[42]    6432    1
#> log_lik[43]    6501    1
#> log_lik[44]    6403    1
#> log_lik[45]    9183    1
#> lp__           1383    1
#> 
#> Samples were drawn using NUTS(diag_e) at Tue Mar  7 08:36:09 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }