Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.07    0.26    0.33    0.37    0.42    0.50
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.40    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.93
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.08    0.10    0.13    0.18
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.05    0.75    0.82    0.85    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.02    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.29
#> theta[1,3,3]     0.79    0.00 0.09    0.58    0.74    0.81    0.87    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.00    0.03    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.30
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.40
#> theta[1,4,4]     0.69    0.00 0.13    0.41    0.60    0.70    0.78    0.91
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.06    0.12    0.16    0.21    0.33
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.39    0.52    0.58    0.65    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.24    0.30    0.36    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.26
#> theta[2,3,3]     0.62    0.00 0.13    0.35    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.22    0.00 0.11    0.04    0.13    0.20    0.29    0.47
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.56    0.66    0.76    0.91
#> theta[3,1,1]     0.88    0.00 0.07    0.73    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.74    0.79    0.88
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[3,3,2]     0.18    0.00 0.10    0.03    0.10    0.16    0.24    0.42
#> theta[3,3,3]     0.51    0.00 0.13    0.25    0.41    0.51    0.60    0.76
#> theta[3,3,4]     0.24    0.00 0.11    0.06    0.15    0.22    0.31    0.49
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[3,4,3]     0.14    0.00 0.12    0.01    0.05    0.11    0.20    0.44
#> theta[3,4,4]     0.63    0.00 0.16    0.30    0.52    0.63    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.08    0.66    0.78    0.84    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.24
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[4,2,2]     0.72    0.00 0.09    0.54    0.67    0.73    0.79    0.88
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.08    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.10    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[4,3,3]     0.69    0.00 0.12    0.43    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.39
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.32    0.55    0.67    0.76    0.91
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.82    0.87    0.91    0.97
#> theta[5,1,2]     0.06    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.06    0.05    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.33
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.26
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.22    0.39
#> theta[5,3,3]     0.60    0.00 0.13    0.33    0.51    0.61    0.70    0.85
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,4,1]     0.11    0.00 0.10    0.00    0.04    0.08    0.16    0.37
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.42
#> theta[5,4,4]     0.63    0.00 0.15    0.31    0.53    0.64    0.75    0.90
#> log_p_z[1,1]    -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_p_z[1,2]   -20.03    0.04 2.36  -25.16  -21.45  -19.82  -18.36  -15.91
#> log_p_z[1,3]   -24.73    0.07 4.29  -34.42  -27.29  -24.22  -21.78  -17.65
#> log_p_z[1,4]   -21.82    0.07 4.18  -31.12  -24.23  -21.36  -18.86  -15.12
#> log_p_z[2,1]   -29.69    0.07 4.44  -39.70  -32.30  -29.22  -26.53  -22.49
#> log_p_z[2,2]   -21.04    0.04 2.39  -26.37  -22.53  -20.89  -19.36  -16.87
#> log_p_z[2,3]    -7.62    0.02 1.24  -10.28   -8.34   -7.48   -6.77   -5.57
#> log_p_z[2,4]   -16.38    0.08 4.25  -26.15  -18.78  -15.79  -13.33   -9.76
#> log_p_z[3,1]   -11.64    0.03 1.67  -15.51  -12.61  -11.45  -10.43   -8.92
#> log_p_z[3,2]   -10.83    0.02 1.51  -14.39  -11.70  -10.64   -9.77   -8.42
#> log_p_z[3,3]   -23.22    0.06 3.48  -31.07  -25.21  -22.87  -20.77  -17.35
#> log_p_z[3,4]   -21.80    0.06 3.51  -29.67  -23.98  -21.49  -19.30  -15.87
#> log_p_z[4,1]   -14.61    0.03 1.82  -18.57  -15.71  -14.46  -13.37  -11.36
#> log_p_z[4,2]    -7.58    0.01 0.94   -9.60   -8.14   -7.52   -6.93   -5.89
#> log_p_z[4,3]   -20.70    0.08 4.26  -30.13  -23.38  -20.28  -17.53  -13.76
#> log_p_z[4,4]   -21.86    0.08 4.33  -31.80  -24.47  -21.33  -18.81  -14.83
#> log_p_z[5,1]   -21.40    0.04 2.55  -27.13  -22.93  -21.13  -19.65  -17.11
#> log_p_z[5,2]    -3.72    0.01 0.44   -4.60   -4.01   -3.69   -3.41   -2.91
#> log_p_z[5,3]   -18.70    0.08 4.20  -28.28  -21.26  -18.26  -15.62  -11.97
#> log_p_z[5,4]   -21.85    0.08 4.29  -31.52  -24.38  -21.43  -18.82  -14.72
#> log_p_z[6,1]   -21.44    0.04 2.53  -26.89  -22.99  -21.25  -19.64  -17.10
#> log_p_z[6,2]    -5.49    0.01 0.68   -6.92   -5.93   -5.45   -5.01   -4.29
#> log_p_z[6,3]   -17.48    0.07 4.04  -26.81  -19.89  -17.04  -14.56  -11.09
#> log_p_z[6,4]   -21.64    0.07 4.25  -31.31  -24.07  -21.15  -18.69  -14.73
#> log_p_z[7,1]    -6.00    0.01 0.62   -7.36   -6.36   -5.94   -5.57   -4.92
#> log_p_z[7,2]   -14.96    0.03 1.73  -18.82  -16.01  -14.81  -13.75  -11.95
#> log_p_z[7,3]   -24.18    0.06 3.54  -32.28  -26.21  -23.86  -21.72  -18.16
#> log_p_z[7,4]   -21.80    0.06 3.45  -29.47  -23.89  -21.50  -19.38  -15.97
#> log_p_z[8,1]   -29.68    0.07 4.47  -40.04  -32.19  -29.17  -26.57  -22.45
#> log_p_z[8,2]   -18.34    0.03 2.16  -23.11  -19.62  -18.15  -16.79  -14.58
#> log_p_z[8,3]    -5.70    0.01 0.83   -7.50   -6.21   -5.65   -5.13   -4.23
#> log_p_z[8,4]   -18.56    0.08 4.33  -28.53  -21.02  -17.96  -15.47  -11.70
#> log_p_z[9,1]   -19.91    0.03 2.27  -24.77  -21.30  -19.77  -18.30  -15.98
#> log_p_z[9,2]    -4.52    0.01 0.60   -5.85   -4.88   -4.48   -4.11   -3.46
#> log_p_z[9,3]   -19.64    0.08 4.28  -29.45  -22.27  -19.16  -16.52  -12.75
#> log_p_z[9,4]   -21.57    0.08 4.23  -31.29  -23.99  -21.08  -18.55  -14.77
#> log_p_z[10,1]  -22.17    0.04 2.46  -27.62  -23.63  -21.94  -20.43  -17.89
#> log_p_z[10,2]   -7.12    0.01 0.72   -8.69   -7.54   -7.07   -6.63   -5.88
#> log_p_z[10,3]  -16.80    0.06 3.13  -23.92  -18.74  -16.51  -14.56  -11.57
#> log_p_z[10,4]  -21.25    0.06 3.50  -29.01  -23.28  -20.89  -18.77  -15.40
#> log_p_z[11,1]  -29.77    0.06 4.41  -39.60  -32.28  -29.28  -26.69  -22.56
#> log_p_z[11,2]  -29.38    0.07 4.44  -40.04  -31.88  -28.84  -26.30  -22.23
#> log_p_z[11,3]  -18.26    0.07 3.44  -26.61  -20.02  -17.72  -15.88  -13.16
#> log_p_z[11,4]   -5.82    0.01 0.96   -7.91   -6.43   -5.74   -5.14   -4.16
#> log_p_z[12,1]  -22.93    0.04 2.68  -28.86  -24.57  -22.69  -21.03  -18.24
#> log_p_z[12,2]   -9.62    0.02 1.34  -12.63  -10.28   -9.46   -8.74   -7.48
#> log_p_z[12,3]  -15.18    0.07 4.00  -24.16  -17.59  -14.76  -12.21   -9.00
#> log_p_z[12,4]  -19.21    0.07 4.07  -28.51  -21.50  -18.66  -16.43  -12.68
#> log_p_z[13,1]   -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_p_z[13,2]  -20.03    0.04 2.36  -25.16  -21.45  -19.82  -18.36  -15.91
#> log_p_z[13,3]  -24.73    0.07 4.29  -34.42  -27.29  -24.22  -21.78  -17.65
#> log_p_z[13,4]  -21.82    0.07 4.18  -31.12  -24.23  -21.36  -18.86  -15.12
#> log_p_z[14,1]  -18.99    0.04 2.39  -24.30  -20.42  -18.78  -17.33  -14.93
#> log_p_z[14,2]   -6.25    0.01 0.90   -8.23   -6.77   -6.17   -5.65   -4.67
#> log_p_z[14,3]  -18.74    0.08 4.23  -28.59  -21.31  -18.24  -15.68  -11.95
#> log_p_z[14,4]  -21.85    0.08 4.22  -31.41  -24.24  -21.50  -18.91  -14.65
#> log_p_z[15,1]   -4.41    0.00 0.40   -5.27   -4.65   -4.38   -4.14   -3.71
#> log_p_z[15,2]  -17.26    0.03 1.88  -21.26  -18.43  -17.13  -15.93  -13.92
#> log_p_z[15,3]  -24.24    0.06 3.54  -32.27  -26.33  -23.90  -21.82  -18.21
#> log_p_z[15,4]  -21.84    0.06 3.49  -29.46  -23.91  -21.49  -19.40  -15.88
#> log_p_z[16,1]   -3.80    0.01 0.52   -4.91   -4.12   -3.77   -3.45   -2.86
#> log_p_z[16,2]  -17.72    0.04 2.23  -22.80  -19.05  -17.54  -16.16  -13.89
#> log_p_z[16,3]  -24.67    0.07 4.29  -34.60  -27.06  -24.07  -21.71  -17.75
#> log_p_z[16,4]  -21.78    0.07 4.16  -31.15  -24.15  -21.38  -18.80  -14.98
#> log_p_z[17,1]   -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_p_z[17,2]  -20.03    0.04 2.36  -25.16  -21.45  -19.82  -18.36  -15.91
#> log_p_z[17,3]  -24.73    0.07 4.29  -34.42  -27.29  -24.22  -21.78  -17.65
#> log_p_z[17,4]  -21.82    0.07 4.18  -31.12  -24.23  -21.36  -18.86  -15.12
#> log_p_z[18,1]   -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_p_z[18,2]  -20.03    0.04 2.36  -25.16  -21.45  -19.82  -18.36  -15.91
#> log_p_z[18,3]  -24.73    0.07 4.29  -34.42  -27.29  -24.22  -21.78  -17.65
#> log_p_z[18,4]  -21.82    0.07 4.18  -31.12  -24.23  -21.36  -18.86  -15.12
#> log_p_z[19,1]  -16.36    0.03 1.92  -20.45  -17.50  -16.22  -15.01  -13.00
#> log_p_z[19,2]   -4.65    0.01 0.59   -5.88   -5.03   -4.62   -4.23   -3.60
#> log_p_z[19,3]  -22.10    0.08 4.39  -32.00  -24.69  -21.67  -18.87  -14.97
#> log_p_z[19,4]  -21.82    0.08 4.27  -31.49  -24.31  -21.35  -18.74  -14.84
#> log_p_z[20,1]  -18.04    0.03 2.25  -23.00  -19.42  -17.87  -16.43  -14.24
#> log_p_z[20,2]   -7.13    0.01 0.98   -9.30   -7.73   -7.06   -6.44   -5.48
#> log_p_z[20,3]  -20.01    0.07 4.16  -29.39  -22.56  -19.56  -16.98  -13.37
#> log_p_z[20,4]  -21.65    0.07 4.27  -31.69  -24.09  -21.16  -18.73  -14.66
#> log_p_z[21,1]  -19.58    0.04 2.35  -24.73  -21.04  -19.38  -17.92  -15.50
#> log_p_z[21,2]   -3.05    0.00 0.36   -3.78   -3.29   -3.03   -2.80   -2.39
#> log_p_z[21,3]  -21.17    0.08 4.30  -30.99  -23.86  -20.73  -18.05  -14.22
#> log_p_z[21,4]  -21.83    0.07 4.27  -31.54  -24.30  -21.32  -18.84  -14.85
#> log_p_z[22,1]  -16.36    0.03 1.92  -20.45  -17.50  -16.22  -15.01  -13.00
#> log_p_z[22,2]   -4.65    0.01 0.59   -5.88   -5.03   -4.62   -4.23   -3.60
#> log_p_z[22,3]  -22.10    0.08 4.39  -32.00  -24.69  -21.67  -18.87  -14.97
#> log_p_z[22,4]  -21.82    0.08 4.27  -31.49  -24.31  -21.35  -18.74  -14.84
#> log_p_z[23,1]  -21.40    0.04 2.55  -27.13  -22.93  -21.13  -19.65  -17.11
#> log_p_z[23,2]   -3.72    0.01 0.44   -4.60   -4.01   -3.69   -3.41   -2.91
#> log_p_z[23,3]  -18.70    0.08 4.20  -28.28  -21.26  -18.26  -15.62  -11.97
#> log_p_z[23,4]  -21.85    0.08 4.29  -31.52  -24.38  -21.43  -18.82  -14.72
#> log_p_z[24,1]  -17.39    0.03 2.15  -22.19  -18.69  -17.20  -15.88  -13.69
#> log_p_z[24,2]   -5.81    0.01 0.65   -7.25   -6.19   -5.75   -5.36   -4.74
#> log_p_z[24,3]  -21.66    0.06 3.41  -28.94  -23.79  -21.44  -19.15  -15.76
#> log_p_z[24,4]  -21.81    0.06 3.53  -29.50  -24.01  -21.46  -19.31  -15.93
#> log_p_z[25,1]   -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_p_z[25,2]  -20.03    0.04 2.36  -25.16  -21.45  -19.82  -18.36  -15.91
#> log_p_z[25,3]  -24.73    0.07 4.29  -34.42  -27.29  -24.22  -21.78  -17.65
#> log_p_z[25,4]  -21.82    0.07 4.18  -31.12  -24.23  -21.36  -18.86  -15.12
#> log_p_z[26,1]   -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_p_z[26,2]  -20.03    0.04 2.36  -25.16  -21.45  -19.82  -18.36  -15.91
#> log_p_z[26,3]  -24.73    0.07 4.29  -34.42  -27.29  -24.22  -21.78  -17.65
#> log_p_z[26,4]  -21.82    0.07 4.18  -31.12  -24.23  -21.36  -18.86  -15.12
#> log_p_z[27,1]  -21.84    0.04 2.47  -27.35  -23.32  -21.66  -20.06  -17.58
#> log_p_z[27,2]   -5.65    0.01 0.53   -6.80   -5.98   -5.62   -5.28   -4.69
#> log_p_z[27,3]  -18.33    0.06 3.18  -25.52  -20.30  -18.07  -16.03  -13.01
#> log_p_z[27,4]  -21.51    0.06 3.49  -29.15  -23.63  -21.14  -19.06  -15.58
#> log_p_z[28,1]   -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_p_z[28,2]  -20.03    0.04 2.36  -25.16  -21.45  -19.82  -18.36  -15.91
#> log_p_z[28,3]  -24.73    0.07 4.29  -34.42  -27.29  -24.22  -21.78  -17.65
#> log_p_z[28,4]  -21.82    0.07 4.18  -31.12  -24.23  -21.36  -18.86  -15.12
#> log_p_z[29,1]   -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_p_z[29,2]  -20.03    0.04 2.36  -25.16  -21.45  -19.82  -18.36  -15.91
#> log_p_z[29,3]  -24.73    0.07 4.29  -34.42  -27.29  -24.22  -21.78  -17.65
#> log_p_z[29,4]  -21.82    0.07 4.18  -31.12  -24.23  -21.36  -18.86  -15.12
#> log_p_z[30,1]   -6.82    0.01 0.85   -8.70   -7.33   -6.74   -6.22   -5.41
#> log_p_z[30,2]  -14.73    0.03 1.72  -18.55  -15.75  -14.56  -13.52  -11.78
#> log_p_z[30,3]  -24.21    0.06 3.54  -32.28  -26.21  -23.89  -21.73  -18.30
#> log_p_z[30,4]  -21.84    0.06 3.52  -29.58  -23.96  -21.43  -19.33  -15.83
#> log_p_z[31,1]   -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_p_z[31,2]  -20.03    0.04 2.36  -25.16  -21.45  -19.82  -18.36  -15.91
#> log_p_z[31,3]  -24.73    0.07 4.29  -34.42  -27.29  -24.22  -21.78  -17.65
#> log_p_z[31,4]  -21.82    0.07 4.18  -31.12  -24.23  -21.36  -18.86  -15.12
#> log_p_z[32,1]  -29.63    0.07 4.51  -40.12  -32.10  -29.08  -26.49  -22.50
#> log_p_z[32,2]  -14.82    0.02 1.69  -18.58  -15.84  -14.70  -13.67  -11.83
#> log_p_z[32,3]   -6.07    0.01 0.95   -8.13   -6.66   -5.99   -5.38   -4.43
#> log_p_z[32,4]  -20.68    0.08 4.48  -30.89  -23.26  -20.09  -17.49  -13.43
#> log_p_z[33,1]   -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_p_z[33,2]  -20.03    0.04 2.36  -25.16  -21.45  -19.82  -18.36  -15.91
#> log_p_z[33,3]  -24.73    0.07 4.29  -34.42  -27.29  -24.22  -21.78  -17.65
#> log_p_z[33,4]  -21.82    0.07 4.18  -31.12  -24.23  -21.36  -18.86  -15.12
#> log_p_z[34,1]  -19.58    0.04 2.35  -24.73  -21.04  -19.38  -17.92  -15.50
#> log_p_z[34,2]   -3.05    0.00 0.36   -3.78   -3.29   -3.03   -2.80   -2.39
#> log_p_z[34,3]  -21.17    0.08 4.30  -30.99  -23.86  -20.73  -18.05  -14.22
#> log_p_z[34,4]  -21.83    0.07 4.27  -31.54  -24.30  -21.32  -18.84  -14.85
#> log_p_z[35,1]  -22.54    0.04 2.68  -28.53  -24.17  -22.35  -20.64  -17.92
#> log_p_z[35,2]   -5.55    0.01 0.70   -7.10   -5.97   -5.49   -5.06   -4.34
#> log_p_z[35,3]  -16.12    0.07 4.06  -25.59  -18.60  -15.66  -13.11   -9.85
#> log_p_z[35,4]  -21.89    0.08 4.29  -31.75  -24.48  -21.48  -18.88  -14.85
#> log_p_z[36,1]  -29.69    0.06 3.68  -37.86  -31.81  -29.39  -27.11  -23.37
#> log_p_z[36,2]  -21.90    0.04 2.41  -27.33  -23.28  -21.65  -20.22  -17.86
#> log_p_z[36,3]  -10.53    0.04 1.74  -14.82  -11.40  -10.25   -9.31   -7.91
#> log_p_z[36,4]  -13.93    0.06 3.24  -21.36  -15.83  -13.62  -11.57   -8.78
#> log_p_z[37,1]  -18.54    0.03 2.31  -23.50  -19.96  -18.33  -16.88  -14.60
#> log_p_z[37,2]   -7.64    0.01 0.82   -9.38   -8.18   -7.58   -7.05   -6.19
#> log_p_z[37,3]  -19.08    0.06 3.19  -26.17  -21.04  -18.75  -16.77  -13.79
#> log_p_z[37,4]  -21.85    0.06 3.53  -29.56  -23.98  -21.55  -19.28  -15.92
#> log_p_z[38,1]  -25.13    0.04 2.83  -31.47  -26.83  -24.90  -23.13  -20.30
#> log_p_z[38,2]   -9.62    0.02 0.99  -11.80  -10.21   -9.54   -8.94   -7.88
#> log_p_z[38,3]  -11.75    0.05 2.77  -18.05  -13.40  -11.47   -9.68   -7.48
#> log_p_z[38,4]  -21.32    0.07 3.58  -29.19  -23.53  -21.03  -18.78  -15.21
#> log_p_z[39,1]  -29.35    0.07 4.41  -39.40  -31.99  -28.81  -26.22  -22.32
#> log_p_z[39,2]  -16.87    0.03 2.07  -21.32  -18.12  -16.68  -15.41  -13.31
#> log_p_z[39,3]   -7.23    0.02 1.11   -9.58   -7.90   -7.15   -6.45   -5.27
#> log_p_z[39,4]  -18.81    0.08 4.19  -28.24  -21.16  -18.32  -15.85  -12.25
#> log_p_z[40,1]   -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_p_z[40,2]  -20.03    0.04 2.36  -25.16  -21.45  -19.82  -18.36  -15.91
#> log_p_z[40,3]  -24.73    0.07 4.29  -34.42  -27.29  -24.22  -21.78  -17.65
#> log_p_z[40,4]  -21.82    0.07 4.18  -31.12  -24.23  -21.36  -18.86  -15.12
#> log_p_z[41,1]   -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_p_z[41,2]  -20.03    0.04 2.36  -25.16  -21.45  -19.82  -18.36  -15.91
#> log_p_z[41,3]  -24.73    0.07 4.29  -34.42  -27.29  -24.22  -21.78  -17.65
#> log_p_z[41,4]  -21.82    0.07 4.18  -31.12  -24.23  -21.36  -18.86  -15.12
#> log_p_z[42,1]   -6.00    0.01 0.62   -7.36   -6.36   -5.94   -5.57   -4.92
#> log_p_z[42,2]  -14.96    0.03 1.73  -18.82  -16.01  -14.81  -13.75  -11.95
#> log_p_z[42,3]  -24.18    0.06 3.54  -32.28  -26.21  -23.86  -21.72  -18.16
#> log_p_z[42,4]  -21.80    0.06 3.45  -29.47  -23.89  -21.50  -19.38  -15.97
#> log_p_z[43,1]  -21.84    0.04 2.47  -27.35  -23.32  -21.66  -20.06  -17.58
#> log_p_z[43,2]   -5.65    0.01 0.53   -6.80   -5.98   -5.62   -5.28   -4.69
#> log_p_z[43,3]  -18.33    0.06 3.18  -25.52  -20.30  -18.07  -16.03  -13.01
#> log_p_z[43,4]  -21.51    0.06 3.49  -29.15  -23.63  -21.14  -19.06  -15.58
#> log_p_z[44,1]   -4.41    0.00 0.40   -5.27   -4.65   -4.38   -4.14   -3.71
#> log_p_z[44,2]  -17.26    0.03 1.88  -21.26  -18.43  -17.13  -15.93  -13.92
#> log_p_z[44,3]  -24.24    0.06 3.54  -32.27  -26.33  -23.90  -21.82  -18.21
#> log_p_z[44,4]  -21.84    0.06 3.49  -29.46  -23.91  -21.49  -19.40  -15.88
#> log_p_z[45,1]  -19.58    0.04 2.35  -24.73  -21.04  -19.38  -17.92  -15.50
#> log_p_z[45,2]   -3.05    0.00 0.36   -3.78   -3.29   -3.03   -2.80   -2.39
#> log_p_z[45,3]  -21.17    0.08 4.30  -30.99  -23.86  -20.73  -18.05  -14.22
#> log_p_z[45,4]  -21.83    0.07 4.27  -31.54  -24.30  -21.32  -18.84  -14.85
#> log_lik[1]      -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_lik[2]      -7.57    0.02 1.14  -10.01   -8.30   -7.47   -6.76   -5.57
#> log_lik[3]      -9.95    0.01 0.92  -11.85  -10.53   -9.92   -9.30   -8.25
#> log_lik[4]      -7.57    0.01 0.93   -9.56   -8.14   -7.51   -6.93   -5.89
#> log_lik[5]      -3.72    0.01 0.44   -4.60   -4.01   -3.69   -3.41   -2.91
#> log_lik[6]      -5.49    0.01 0.68   -6.92   -5.93   -5.45   -5.01   -4.29
#> log_lik[7]      -6.00    0.01 0.62   -7.36   -6.36   -5.94   -5.57   -4.92
#> log_lik[8]      -5.70    0.01 0.83   -7.50   -6.21   -5.65   -5.13   -4.23
#> log_lik[9]      -4.52    0.01 0.60   -5.85   -4.88   -4.48   -4.11   -3.46
#> log_lik[10]     -7.12    0.01 0.72   -8.69   -7.54   -7.07   -6.63   -5.88
#> log_lik[11]     -5.82    0.01 0.96   -7.91   -6.43   -5.74   -5.14   -4.16
#> log_lik[12]     -9.34    0.01 1.03  -11.53   -9.97   -9.29   -8.65   -7.42
#> log_lik[13]     -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_lik[14]     -6.25    0.01 0.90   -8.23   -6.77   -6.17   -5.65   -4.67
#> log_lik[15]     -4.41    0.00 0.40   -5.27   -4.65   -4.38   -4.14   -3.71
#> log_lik[16]     -3.80    0.01 0.52   -4.91   -4.12   -3.77   -3.45   -2.86
#> log_lik[17]     -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_lik[18]     -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_lik[19]     -4.65    0.01 0.59   -5.88   -5.03   -4.62   -4.23   -3.60
#> log_lik[20]     -7.13    0.01 0.97   -9.30   -7.72   -7.06   -6.44   -5.48
#> log_lik[21]     -3.05    0.00 0.36   -3.78   -3.29   -3.03   -2.80   -2.39
#> log_lik[22]     -4.65    0.01 0.59   -5.88   -5.03   -4.62   -4.23   -3.60
#> log_lik[23]     -3.72    0.01 0.44   -4.60   -4.01   -3.69   -3.41   -2.91
#> log_lik[24]     -5.81    0.01 0.65   -7.25   -6.19   -5.75   -5.36   -4.74
#> log_lik[25]     -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_lik[26]     -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_lik[27]     -5.65    0.01 0.53   -6.80   -5.98   -5.62   -5.28   -4.69
#> log_lik[28]     -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_lik[29]     -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_lik[30]     -6.82    0.01 0.84   -8.69   -7.33   -6.74   -6.22   -5.41
#> log_lik[31]     -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_lik[32]     -6.07    0.01 0.95   -8.11   -6.66   -5.99   -5.38   -4.43
#> log_lik[33]     -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_lik[34]     -3.05    0.00 0.36   -3.78   -3.29   -3.03   -2.80   -2.39
#> log_lik[35]     -5.54    0.01 0.70   -7.08   -5.97   -5.49   -5.06   -4.34
#> log_lik[36]     -9.92    0.02 1.17  -12.42  -10.67   -9.86   -9.10   -7.81
#> log_lik[37]     -7.64    0.01 0.82   -9.38   -8.18   -7.58   -7.05   -6.19
#> log_lik[38]     -9.01    0.02 0.87  -10.70   -9.57   -9.01   -8.45   -7.34
#> log_lik[39]     -7.23    0.02 1.11   -9.57   -7.90   -7.15   -6.45   -5.27
#> log_lik[40]     -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_lik[41]     -2.22    0.00 0.29   -2.82   -2.41   -2.21   -2.01   -1.68
#> log_lik[42]     -6.00    0.01 0.62   -7.36   -6.36   -5.94   -5.57   -4.92
#> log_lik[43]     -5.65    0.01 0.53   -6.80   -5.98   -5.62   -5.28   -4.69
#> log_lik[44]     -4.41    0.00 0.40   -5.27   -4.65   -4.38   -4.14   -3.71
#> log_lik[45]     -3.05    0.00 0.36   -3.78   -3.29   -3.03   -2.80   -2.39
#> lp__          -463.98    0.18 6.30 -477.09 -468.09 -463.59 -459.50 -452.59
#>               n_eff Rhat
#> pi[1]          7033    1
#> pi[2]          6422    1
#> pi[3]          5338    1
#> pi[4]          4802    1
#> theta[1,1,1]   7196    1
#> theta[1,1,2]   6631    1
#> theta[1,1,3]   4938    1
#> theta[1,1,4]   4765    1
#> theta[1,2,1]   5821    1
#> theta[1,2,2]   6449    1
#> theta[1,2,3]   6352    1
#> theta[1,2,4]   5603    1
#> theta[1,3,1]   6144    1
#> theta[1,3,2]   3522    1
#> theta[1,3,3]   4661    1
#> theta[1,3,4]   5442    1
#> theta[1,4,1]   5984    1
#> theta[1,4,2]   5184    1
#> theta[1,4,3]   4448    1
#> theta[1,4,4]   5400    1
#> theta[2,1,1]   6733    1
#> theta[2,1,2]   6093    1
#> theta[2,1,3]   4418    1
#> theta[2,1,4]   4223    1
#> theta[2,2,1]   6627    1
#> theta[2,2,2]   5834    1
#> theta[2,2,3]   5123    1
#> theta[2,2,4]   5523    1
#> theta[2,3,1]   6209    1
#> theta[2,3,2]   6336    1
#> theta[2,3,3]   4753    1
#> theta[2,3,4]   5139    1
#> theta[2,4,1]   6635    1
#> theta[2,4,2]   5617    1
#> theta[2,4,3]   7510    1
#> theta[2,4,4]   6531    1
#> theta[3,1,1]   7411    1
#> theta[3,1,2]   4811    1
#> theta[3,1,3]   4899    1
#> theta[3,1,4]   4550    1
#> theta[3,2,1]   6570    1
#> theta[3,2,2]   6970    1
#> theta[3,2,3]   6432    1
#> theta[3,2,4]   5234    1
#> theta[3,3,1]   6487    1
#> theta[3,3,2]   6980    1
#> theta[3,3,3]   6836    1
#> theta[3,3,4]   6246    1
#> theta[3,4,1]   7239    1
#> theta[3,4,2]   6459    1
#> theta[3,4,3]   5008    1
#> theta[3,4,4]   5912    1
#> theta[4,1,1]   6491    1
#> theta[4,1,2]   4314    1
#> theta[4,1,3]   4351    1
#> theta[4,1,4]   4444    1
#> theta[4,2,1]   6140    1
#> theta[4,2,2]   6524    1
#> theta[4,2,3]   5563    1
#> theta[4,2,4]   5467    1
#> theta[4,3,1]   6155    1
#> theta[4,3,2]   6383    1
#> theta[4,3,3]   6119    1
#> theta[4,3,4]   6117    1
#> theta[4,4,1]   6970    1
#> theta[4,4,2]   5884    1
#> theta[4,4,3]   6227    1
#> theta[4,4,4]   5710    1
#> theta[5,1,1]   7765    1
#> theta[5,1,2]   5170    1
#> theta[5,1,3]   4559    1
#> theta[5,1,4]   5075    1
#> theta[5,2,1]   6573    1
#> theta[5,2,2]   6706    1
#> theta[5,2,3]   5988    1
#> theta[5,2,4]   5589    1
#> theta[5,3,1]   5637    1
#> theta[5,3,2]   6551    1
#> theta[5,3,3]   6450    1
#> theta[5,3,4]   6479    1
#> theta[5,4,1]   6249    1
#> theta[5,4,2]   6872    1
#> theta[5,4,3]   5282    1
#> theta[5,4,4]   6061    1
#> log_p_z[1,1]   6078    1
#> log_p_z[1,2]   3894    1
#> log_p_z[1,3]   3575    1
#> log_p_z[1,4]   3310    1
#> log_p_z[2,1]   4484    1
#> log_p_z[2,2]   3765    1
#> log_p_z[2,3]   4138    1
#> log_p_z[2,4]   2968    1
#> log_p_z[3,1]   3862    1
#> log_p_z[3,2]   3740    1
#> log_p_z[3,3]   3169    1
#> log_p_z[3,4]   3503    1
#> log_p_z[4,1]   4820    1
#> log_p_z[4,2]   5309    1
#> log_p_z[4,3]   3177    1
#> log_p_z[4,4]   3058    1
#> log_p_z[5,1]   4071    1
#> log_p_z[5,2]   5241    1
#> log_p_z[5,3]   3093    1
#> log_p_z[5,4]   3181    1
#> log_p_z[6,1]   4394    1
#> log_p_z[6,2]   5690    1
#> log_p_z[6,3]   3065    1
#> log_p_z[6,4]   3237    1
#> log_p_z[7,1]   6429    1
#> log_p_z[7,2]   3784    1
#> log_p_z[7,3]   3435    1
#> log_p_z[7,4]   3523    1
#> log_p_z[8,1]   4494    1
#> log_p_z[8,2]   3932    1
#> log_p_z[8,3]   5955    1
#> log_p_z[8,4]   3009    1
#> log_p_z[9,1]   4523    1
#> log_p_z[9,2]   4923    1
#> log_p_z[9,3]   3036    1
#> log_p_z[9,4]   3180    1
#> log_p_z[10,1]  4226    1
#> log_p_z[10,2]  4881    1
#> log_p_z[10,3]  3060    1
#> log_p_z[10,4]  3090    1
#> log_p_z[11,1]  4734    1
#> log_p_z[11,2]  3775    1
#> log_p_z[11,3]  2285    1
#> log_p_z[11,4]  5721    1
#> log_p_z[12,1]  4533    1
#> log_p_z[12,2]  3364    1
#> log_p_z[12,3]  3023    1
#> log_p_z[12,4]  3014    1
#> log_p_z[13,1]  6078    1
#> log_p_z[13,2]  3894    1
#> log_p_z[13,3]  3575    1
#> log_p_z[13,4]  3310    1
#> log_p_z[14,1]  4416    1
#> log_p_z[14,2]  4346    1
#> log_p_z[14,3]  3141    1
#> log_p_z[14,4]  3021    1
#> log_p_z[15,1]  6453    1
#> log_p_z[15,2]  3955    1
#> log_p_z[15,3]  3674    1
#> log_p_z[15,4]  3270    1
#> log_p_z[16,1]  6136    1
#> log_p_z[16,2]  3811    1
#> log_p_z[16,3]  3445    1
#> log_p_z[16,4]  3639    1
#> log_p_z[17,1]  6078    1
#> log_p_z[17,2]  3894    1
#> log_p_z[17,3]  3575    1
#> log_p_z[17,4]  3310    1
#> log_p_z[18,1]  6078    1
#> log_p_z[18,2]  3894    1
#> log_p_z[18,3]  3575    1
#> log_p_z[18,4]  3310    1
#> log_p_z[19,1]  4433    1
#> log_p_z[19,2]  5979    1
#> log_p_z[19,3]  3132    1
#> log_p_z[19,4]  3111    1
#> log_p_z[20,1]  4363    1
#> log_p_z[20,2]  5136    1
#> log_p_z[20,3]  3118    1
#> log_p_z[20,4]  3427    1
#> log_p_z[21,1]  4157    1
#> log_p_z[21,2]  5900    1
#> log_p_z[21,3]  3005    1
#> log_p_z[21,4]  3254    1
#> log_p_z[22,1]  4433    1
#> log_p_z[22,2]  5979    1
#> log_p_z[22,3]  3132    1
#> log_p_z[22,4]  3111    1
#> log_p_z[23,1]  4071    1
#> log_p_z[23,2]  5241    1
#> log_p_z[23,3]  3093    1
#> log_p_z[23,4]  3181    1
#> log_p_z[24,1]  4090    1
#> log_p_z[24,2]  4195    1
#> log_p_z[24,3]  2981    1
#> log_p_z[24,4]  3395    1
#> log_p_z[25,1]  6078    1
#> log_p_z[25,2]  3894    1
#> log_p_z[25,3]  3575    1
#> log_p_z[25,4]  3310    1
#> log_p_z[26,1]  6078    1
#> log_p_z[26,2]  3894    1
#> log_p_z[26,3]  3575    1
#> log_p_z[26,4]  3310    1
#> log_p_z[27,1]  4154    1
#> log_p_z[27,2]  6006    1
#> log_p_z[27,3]  3020    1
#> log_p_z[27,4]  3211    1
#> log_p_z[28,1]  6078    1
#> log_p_z[28,2]  3894    1
#> log_p_z[28,3]  3575    1
#> log_p_z[28,4]  3310    1
#> log_p_z[29,1]  6078    1
#> log_p_z[29,2]  3894    1
#> log_p_z[29,3]  3575    1
#> log_p_z[29,4]  3310    1
#> log_p_z[30,1]  4307    1
#> log_p_z[30,2]  3842    1
#> log_p_z[30,3]  3475    1
#> log_p_z[30,4]  3326    1
#> log_p_z[31,1]  6078    1
#> log_p_z[31,2]  3894    1
#> log_p_z[31,3]  3575    1
#> log_p_z[31,4]  3310    1
#> log_p_z[32,1]  4492    1
#> log_p_z[32,2]  4893    1
#> log_p_z[32,3]  4835    1
#> log_p_z[32,4]  3020    1
#> log_p_z[33,1]  6078    1
#> log_p_z[33,2]  3894    1
#> log_p_z[33,3]  3575    1
#> log_p_z[33,4]  3310    1
#> log_p_z[34,1]  4157    1
#> log_p_z[34,2]  5900    1
#> log_p_z[34,3]  3005    1
#> log_p_z[34,4]  3254    1
#> log_p_z[35,1]  4417    1
#> log_p_z[35,2]  4335    1
#> log_p_z[35,3]  3094    1
#> log_p_z[35,4]  2762    1
#> log_p_z[36,1]  3950    1
#> log_p_z[36,2]  3708    1
#> log_p_z[36,3]  2282    1
#> log_p_z[36,4]  2875    1
#> log_p_z[37,1]  4521    1
#> log_p_z[37,2]  4477    1
#> log_p_z[37,3]  3013    1
#> log_p_z[37,4]  3195    1
#> log_p_z[38,1]  4380    1
#> log_p_z[38,2]  3832    1
#> log_p_z[38,3]  3118    1
#> log_p_z[38,4]  2928    1
#> log_p_z[39,1]  4403    1
#> log_p_z[39,2]  3944    1
#> log_p_z[39,3]  5083    1
#> log_p_z[39,4]  3097    1
#> log_p_z[40,1]  6078    1
#> log_p_z[40,2]  3894    1
#> log_p_z[40,3]  3575    1
#> log_p_z[40,4]  3310    1
#> log_p_z[41,1]  6078    1
#> log_p_z[41,2]  3894    1
#> log_p_z[41,3]  3575    1
#> log_p_z[41,4]  3310    1
#> log_p_z[42,1]  6429    1
#> log_p_z[42,2]  3784    1
#> log_p_z[42,3]  3435    1
#> log_p_z[42,4]  3523    1
#> log_p_z[43,1]  4154    1
#> log_p_z[43,2]  6006    1
#> log_p_z[43,3]  3020    1
#> log_p_z[43,4]  3211    1
#> log_p_z[44,1]  6453    1
#> log_p_z[44,2]  3955    1
#> log_p_z[44,3]  3674    1
#> log_p_z[44,4]  3270    1
#> log_p_z[45,1]  4157    1
#> log_p_z[45,2]  5900    1
#> log_p_z[45,3]  3005    1
#> log_p_z[45,4]  3254    1
#> log_lik[1]     6078    1
#> log_lik[2]     4675    1
#> log_lik[3]     4669    1
#> log_lik[4]     5410    1
#> log_lik[5]     5242    1
#> log_lik[6]     5701    1
#> log_lik[7]     6437    1
#> log_lik[8]     5959    1
#> log_lik[9]     4926    1
#> log_lik[10]    4903    1
#> log_lik[11]    5722    1
#> log_lik[12]    5238    1
#> log_lik[13]    6078    1
#> log_lik[14]    4356    1
#> log_lik[15]    6453    1
#> log_lik[16]    6136    1
#> log_lik[17]    6078    1
#> log_lik[18]    6078    1
#> log_lik[19]    5979    1
#> log_lik[20]    5132    1
#> log_lik[21]    5900    1
#> log_lik[22]    5979    1
#> log_lik[23]    5242    1
#> log_lik[24]    4195    1
#> log_lik[25]    6078    1
#> log_lik[26]    6078    1
#> log_lik[27]    6005    1
#> log_lik[28]    6078    1
#> log_lik[29]    6078    1
#> log_lik[30]    4330    1
#> log_lik[31]    6078    1
#> log_lik[32]    4840    1
#> log_lik[33]    6078    1
#> log_lik[34]    5900    1
#> log_lik[35]    4387    1
#> log_lik[36]    4157    1
#> log_lik[37]    4478    1
#> log_lik[38]    3363    1
#> log_lik[39]    5151    1
#> log_lik[40]    6078    1
#> log_lik[41]    6078    1
#> log_lik[42]    6437    1
#> log_lik[43]    6005    1
#> log_lik[44]    6453    1
#> log_lik[45]    5900    1
#> lp__           1268    1
#> 
#> Samples were drawn using NUTS(diag_e) at Sat Mar 11 08:39:50 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }