Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.25    0.33    0.38    0.42    0.51
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.90    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.13
#> theta[1,2,2]     0.85    0.00 0.04    0.75    0.83    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.02    0.03    0.06    0.15
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.80    0.00 0.09    0.59    0.74    0.81    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.21
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.40
#> theta[1,4,4]     0.69    0.00 0.14    0.40    0.60    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.76    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.08    0.05    0.11    0.16    0.21    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[2,2,2]     0.58    0.00 0.09    0.39    0.52    0.58    0.65    0.76
#> theta[2,2,3]     0.30    0.00 0.09    0.15    0.24    0.30    0.36    0.49
#> theta[2,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.28
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[2,3,3]     0.62    0.00 0.13    0.36    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.28    0.47
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,3]     0.12    0.00 0.10    0.01    0.04    0.09    0.17    0.36
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.76    0.90
#> theta[3,1,1]     0.88    0.00 0.06    0.74    0.84    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.22
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.12
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.24    0.42
#> theta[3,3,3]     0.51    0.00 0.14    0.25    0.41    0.50    0.60    0.77
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.31    0.48
#> theta[3,4,1]     0.12    0.00 0.10    0.01    0.04    0.09    0.16    0.37
#> theta[3,4,2]     0.12    0.00 0.11    0.00    0.04    0.09    0.17    0.39
#> theta[3,4,3]     0.14    0.00 0.12    0.01    0.05    0.11    0.20    0.44
#> theta[3,4,4]     0.62    0.00 0.15    0.30    0.52    0.64    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.78    0.84    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.24
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.08    0.55    0.67    0.73    0.78    0.87
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.09    0.13    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.18
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[4,3,2]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.28
#> theta[4,3,3]     0.69    0.00 0.12    0.42    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.38
#> theta[4,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.77    0.91
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.82    0.87    0.91    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.06    0.04    0.10    0.13    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.46    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.33
#> theta[5,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.09    0.14    0.21    0.40
#> theta[5,3,3]     0.60    0.00 0.13    0.33    0.51    0.61    0.70    0.85
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.40
#> theta[5,4,4]     0.63    0.00 0.15    0.33    0.53    0.64    0.74    0.89
#> log_p_z[1,1]    -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_p_z[1,2]   -19.97    0.04 2.29  -25.08  -21.36  -19.80  -18.36  -16.03
#> log_p_z[1,3]   -24.82    0.07 4.25  -34.60  -27.27  -24.30  -21.83  -17.96
#> log_p_z[1,4]   -21.90    0.07 4.39  -32.05  -24.33  -21.38  -18.85  -14.85
#> log_p_z[2,1]   -29.74    0.06 4.34  -39.42  -32.12  -29.34  -26.77  -22.44
#> log_p_z[2,2]   -20.99    0.04 2.38  -26.18  -22.47  -20.77  -19.32  -16.82
#> log_p_z[2,3]    -7.63    0.02 1.22  -10.34   -8.32   -7.52   -6.79   -5.62
#> log_p_z[2,4]   -16.40    0.08 4.25  -26.29  -18.83  -15.82  -13.41   -9.74
#> log_p_z[3,1]   -11.63    0.03 1.66  -15.48  -12.61  -11.46  -10.42   -8.94
#> log_p_z[3,2]   -10.78    0.02 1.48  -14.20  -11.65  -10.62   -9.76   -8.41
#> log_p_z[3,3]   -23.31    0.06 3.38  -30.94  -25.41  -23.00  -20.92  -17.51
#> log_p_z[3,4]   -21.89    0.06 3.59  -29.73  -24.09  -21.52  -19.31  -15.88
#> log_p_z[4,1]   -14.68    0.03 1.92  -18.78  -15.83  -14.54  -13.39  -11.30
#> log_p_z[4,2]    -7.57    0.01 0.93   -9.57   -8.15   -7.50   -6.92   -5.94
#> log_p_z[4,3]   -20.74    0.08 4.31  -30.47  -23.16  -20.20  -17.64  -14.00
#> log_p_z[4,4]   -21.85    0.07 4.31  -32.02  -24.35  -21.34  -18.81  -14.86
#> log_p_z[5,1]   -21.45    0.04 2.54  -27.05  -23.02  -21.29  -19.65  -17.02
#> log_p_z[5,2]    -3.72    0.01 0.44   -4.66   -4.01   -3.70   -3.41   -2.95
#> log_p_z[5,3]   -18.70    0.08 4.20  -28.31  -21.09  -18.12  -15.71  -12.19
#> log_p_z[5,4]   -21.86    0.07 4.26  -31.73  -24.40  -21.36  -18.88  -14.84
#> log_p_z[6,1]   -21.48    0.04 2.59  -27.16  -23.07  -21.28  -19.65  -16.95
#> log_p_z[6,2]    -5.46    0.01 0.65   -6.85   -5.88   -5.42   -5.00   -4.30
#> log_p_z[6,3]   -17.49    0.07 4.06  -26.71  -19.74  -16.92  -14.60  -11.38
#> log_p_z[6,4]   -21.65    0.07 4.23  -31.35  -24.10  -21.25  -18.69  -14.74
#> log_p_z[7,1]    -6.01    0.01 0.63   -7.38   -6.41   -5.96   -5.55   -4.90
#> log_p_z[7,2]   -14.94    0.03 1.72  -18.76  -15.97  -14.78  -13.72  -12.03
#> log_p_z[7,3]   -24.25    0.06 3.50  -32.14  -26.40  -23.93  -21.77  -18.46
#> log_p_z[7,4]   -21.90    0.06 3.57  -29.96  -24.05  -21.63  -19.38  -15.94
#> log_p_z[8,1]   -29.74    0.06 4.38  -39.35  -32.29  -29.32  -26.72  -22.35
#> log_p_z[8,2]   -18.31    0.03 2.16  -23.16  -19.61  -18.11  -16.83  -14.55
#> log_p_z[8,3]    -5.72    0.01 0.80   -7.34   -6.24   -5.69   -5.16   -4.30
#> log_p_z[8,4]   -18.50    0.08 4.28  -28.46  -20.96  -17.89  -15.52  -11.63
#> log_p_z[9,1]   -19.95    0.03 2.24  -24.87  -21.32  -19.81  -18.36  -15.99
#> log_p_z[9,2]    -4.50    0.01 0.56   -5.70   -4.85   -4.47   -4.12   -3.50
#> log_p_z[9,3]   -19.64    0.08 4.31  -29.27  -22.19  -19.02  -16.52  -12.99
#> log_p_z[9,4]   -21.65    0.06 4.16  -31.10  -24.14  -21.16  -18.76  -14.78
#> log_p_z[10,1]  -22.20    0.03 2.35  -27.42  -23.67  -22.02  -20.58  -18.06
#> log_p_z[10,2]   -7.11    0.01 0.70   -8.66   -7.55   -7.06   -6.63   -5.90
#> log_p_z[10,3]  -16.81    0.05 3.17  -23.77  -18.67  -16.40  -14.51  -11.81
#> log_p_z[10,4]  -21.34    0.06 3.46  -29.01  -23.44  -21.06  -18.84  -15.44
#> log_p_z[11,1]  -29.81    0.06 4.44  -39.75  -32.37  -29.36  -26.71  -22.45
#> log_p_z[11,2]  -29.32    0.07 4.34  -38.92  -31.88  -28.88  -26.16  -22.30
#> log_p_z[11,3]  -18.19    0.06 3.33  -26.21  -19.88  -17.73  -15.90  -13.29
#> log_p_z[11,4]   -5.83    0.01 0.98   -7.89   -6.45   -5.74   -5.13   -4.11
#> log_p_z[12,1]  -22.99    0.04 2.73  -28.85  -24.63  -22.82  -21.07  -18.22
#> log_p_z[12,2]   -9.54    0.02 1.24  -12.40  -10.25   -9.43   -8.68   -7.46
#> log_p_z[12,3]  -15.18    0.07 3.99  -24.11  -17.41  -14.69  -12.28   -9.06
#> log_p_z[12,4]  -19.20    0.07 4.02  -28.63  -21.40  -18.80  -16.43  -12.55
#> log_p_z[13,1]   -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_p_z[13,2]  -19.97    0.04 2.29  -25.08  -21.36  -19.80  -18.36  -16.03
#> log_p_z[13,3]  -24.82    0.07 4.25  -34.60  -27.27  -24.30  -21.83  -17.96
#> log_p_z[13,4]  -21.90    0.07 4.39  -32.05  -24.33  -21.38  -18.85  -14.85
#> log_p_z[14,1]  -19.05    0.03 2.37  -24.17  -20.54  -18.90  -17.40  -14.96
#> log_p_z[14,2]   -6.26    0.01 0.88   -8.21   -6.79   -6.19   -5.65   -4.72
#> log_p_z[14,3]  -18.70    0.08 4.21  -28.38  -21.05  -18.14  -15.74  -12.21
#> log_p_z[14,4]  -21.86    0.07 4.25  -31.52  -24.43  -21.44  -18.82  -14.87
#> log_p_z[15,1]   -4.41    0.00 0.41   -5.30   -4.66   -4.38   -4.12   -3.72
#> log_p_z[15,2]  -17.22    0.03 1.84  -21.19  -18.34  -17.11  -15.92  -14.03
#> log_p_z[15,3]  -24.31    0.06 3.53  -32.25  -26.44  -23.97  -21.82  -18.30
#> log_p_z[15,4]  -21.89    0.06 3.63  -30.33  -24.05  -21.51  -19.33  -15.73
#> log_p_z[16,1]   -3.80    0.01 0.55   -5.00   -4.15   -3.76   -3.41   -2.85
#> log_p_z[16,2]  -17.69    0.03 2.20  -22.71  -18.98  -17.47  -16.13  -14.08
#> log_p_z[16,3]  -24.76    0.07 4.21  -34.24  -27.26  -24.31  -21.74  -17.98
#> log_p_z[16,4]  -21.91    0.07 4.32  -31.67  -24.36  -21.41  -18.90  -14.95
#> log_p_z[17,1]   -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_p_z[17,2]  -19.97    0.04 2.29  -25.08  -21.36  -19.80  -18.36  -16.03
#> log_p_z[17,3]  -24.82    0.07 4.25  -34.60  -27.27  -24.30  -21.83  -17.96
#> log_p_z[17,4]  -21.90    0.07 4.39  -32.05  -24.33  -21.38  -18.85  -14.85
#> log_p_z[18,1]   -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_p_z[18,2]  -19.97    0.04 2.29  -25.08  -21.36  -19.80  -18.36  -16.03
#> log_p_z[18,3]  -24.82    0.07 4.25  -34.60  -27.27  -24.30  -21.83  -17.96
#> log_p_z[18,4]  -21.90    0.07 4.39  -32.05  -24.33  -21.38  -18.85  -14.85
#> log_p_z[19,1]  -16.38    0.02 1.90  -20.39  -17.56  -16.24  -15.04  -12.97
#> log_p_z[19,2]   -4.66    0.01 0.60   -5.93   -5.06   -4.63   -4.24   -3.61
#> log_p_z[19,3]  -22.11    0.08 4.43  -32.17  -24.58  -21.56  -18.98  -15.14
#> log_p_z[19,4]  -21.91    0.07 4.30  -31.86  -24.42  -21.41  -18.84  -14.90
#> log_p_z[20,1]  -18.03    0.03 2.29  -22.97  -19.48  -17.84  -16.38  -14.07
#> log_p_z[20,2]   -7.06    0.01 0.95   -9.16   -7.64   -6.99   -6.39   -5.42
#> log_p_z[20,3]  -20.04    0.08 4.23  -29.64  -22.41  -19.50  -16.97  -13.57
#> log_p_z[20,4]  -21.68    0.07 4.24  -31.18  -24.13  -21.26  -18.65  -14.83
#> log_p_z[21,1]  -19.60    0.03 2.33  -24.67  -21.07  -19.44  -17.94  -15.49
#> log_p_z[21,2]   -3.04    0.00 0.35   -3.80   -3.26   -3.03   -2.80   -2.40
#> log_p_z[21,3]  -21.18    0.08 4.34  -30.97  -23.65  -20.57  -18.12  -14.41
#> log_p_z[21,4]  -21.90    0.07 4.26  -31.36  -24.48  -21.41  -18.90  -14.94
#> log_p_z[22,1]  -16.38    0.02 1.90  -20.39  -17.56  -16.24  -15.04  -12.97
#> log_p_z[22,2]   -4.66    0.01 0.60   -5.93   -5.06   -4.63   -4.24   -3.61
#> log_p_z[22,3]  -22.11    0.08 4.43  -32.17  -24.58  -21.56  -18.98  -15.14
#> log_p_z[22,4]  -21.91    0.07 4.30  -31.86  -24.42  -21.41  -18.84  -14.90
#> log_p_z[23,1]  -21.45    0.04 2.54  -27.05  -23.02  -21.29  -19.65  -17.02
#> log_p_z[23,2]   -3.72    0.01 0.44   -4.66   -4.01   -3.70   -3.41   -2.95
#> log_p_z[23,3]  -18.70    0.08 4.20  -28.31  -21.09  -18.12  -15.71  -12.19
#> log_p_z[23,4]  -21.86    0.07 4.26  -31.73  -24.40  -21.36  -18.88  -14.84
#> log_p_z[24,1]  -17.39    0.03 2.11  -22.07  -18.74  -17.22  -15.84  -13.74
#> log_p_z[24,2]   -5.80    0.01 0.64   -7.25   -6.16   -5.73   -5.35   -4.75
#> log_p_z[24,3]  -21.69    0.06 3.42  -28.98  -23.73  -21.32  -19.26  -16.05
#> log_p_z[24,4]  -21.91    0.06 3.57  -29.61  -24.16  -21.56  -19.35  -15.90
#> log_p_z[25,1]   -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_p_z[25,2]  -19.97    0.04 2.29  -25.08  -21.36  -19.80  -18.36  -16.03
#> log_p_z[25,3]  -24.82    0.07 4.25  -34.60  -27.27  -24.30  -21.83  -17.96
#> log_p_z[25,4]  -21.90    0.07 4.39  -32.05  -24.33  -21.38  -18.85  -14.85
#> log_p_z[26,1]   -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_p_z[26,2]  -19.97    0.04 2.29  -25.08  -21.36  -19.80  -18.36  -16.03
#> log_p_z[26,3]  -24.82    0.07 4.25  -34.60  -27.27  -24.30  -21.83  -17.96
#> log_p_z[26,4]  -21.90    0.07 4.39  -32.05  -24.33  -21.38  -18.85  -14.85
#> log_p_z[27,1]  -21.85    0.03 2.42  -27.12  -23.32  -21.66  -20.15  -17.64
#> log_p_z[27,2]   -5.65    0.01 0.52   -6.74   -5.98   -5.61   -5.28   -4.73
#> log_p_z[27,3]  -18.34    0.06 3.22  -25.47  -20.26  -17.93  -16.12  -13.12
#> log_p_z[27,4]  -21.59    0.06 3.48  -29.17  -23.72  -21.25  -19.12  -15.81
#> log_p_z[28,1]   -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_p_z[28,2]  -19.97    0.04 2.29  -25.08  -21.36  -19.80  -18.36  -16.03
#> log_p_z[28,3]  -24.82    0.07 4.25  -34.60  -27.27  -24.30  -21.83  -17.96
#> log_p_z[28,4]  -21.90    0.07 4.39  -32.05  -24.33  -21.38  -18.85  -14.85
#> log_p_z[29,1]   -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_p_z[29,2]  -19.97    0.04 2.29  -25.08  -21.36  -19.80  -18.36  -16.03
#> log_p_z[29,3]  -24.82    0.07 4.25  -34.60  -27.27  -24.30  -21.83  -17.96
#> log_p_z[29,4]  -21.90    0.07 4.39  -32.05  -24.33  -21.38  -18.85  -14.85
#> log_p_z[30,1]   -6.82    0.01 0.86   -8.70   -7.33   -6.74   -6.21   -5.37
#> log_p_z[30,2]  -14.68    0.03 1.69  -18.46  -15.73  -14.56  -13.49  -11.77
#> log_p_z[30,3]  -24.31    0.06 3.53  -32.21  -26.45  -24.00  -21.81  -18.39
#> log_p_z[30,4]  -21.90    0.06 3.68  -30.21  -24.09  -21.52  -19.29  -15.80
#> log_p_z[31,1]   -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_p_z[31,2]  -19.97    0.04 2.29  -25.08  -21.36  -19.80  -18.36  -16.03
#> log_p_z[31,3]  -24.82    0.07 4.25  -34.60  -27.27  -24.30  -21.83  -17.96
#> log_p_z[31,4]  -21.90    0.07 4.39  -32.05  -24.33  -21.38  -18.85  -14.85
#> log_p_z[32,1]  -29.71    0.06 4.33  -39.13  -32.35  -29.27  -26.69  -22.42
#> log_p_z[32,2]  -14.82    0.02 1.74  -18.77  -15.87  -14.70  -13.61  -11.77
#> log_p_z[32,3]   -6.06    0.01 0.93   -8.13   -6.65   -5.97   -5.42   -4.49
#> log_p_z[32,4]  -20.63    0.08 4.42  -30.93  -23.15  -20.05  -17.57  -13.46
#> log_p_z[33,1]   -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_p_z[33,2]  -19.97    0.04 2.29  -25.08  -21.36  -19.80  -18.36  -16.03
#> log_p_z[33,3]  -24.82    0.07 4.25  -34.60  -27.27  -24.30  -21.83  -17.96
#> log_p_z[33,4]  -21.90    0.07 4.39  -32.05  -24.33  -21.38  -18.85  -14.85
#> log_p_z[34,1]  -19.60    0.03 2.33  -24.67  -21.07  -19.44  -17.94  -15.49
#> log_p_z[34,2]   -3.04    0.00 0.35   -3.80   -3.26   -3.03   -2.80   -2.40
#> log_p_z[34,3]  -21.18    0.08 4.34  -30.97  -23.65  -20.57  -18.12  -14.41
#> log_p_z[34,4]  -21.90    0.07 4.26  -31.36  -24.48  -21.41  -18.90  -14.94
#> log_p_z[35,1]  -22.60    0.04 2.62  -28.23  -24.21  -22.44  -20.78  -17.93
#> log_p_z[35,2]   -5.54    0.01 0.72   -7.08   -5.98   -5.48   -5.03   -4.31
#> log_p_z[35,3]  -16.10    0.07 4.07  -25.47  -18.30  -15.51  -13.20  -10.04
#> log_p_z[35,4]  -21.80    0.07 4.24  -31.50  -24.26  -21.38  -18.83  -14.90
#> log_p_z[36,1]  -29.77    0.06 3.63  -37.66  -32.03  -29.43  -27.14  -23.61
#> log_p_z[36,2]  -21.85    0.04 2.28  -26.89  -23.31  -21.65  -20.21  -17.80
#> log_p_z[36,3]  -10.52    0.03 1.73  -14.68  -11.43  -10.23   -9.34   -7.96
#> log_p_z[36,4]  -13.97    0.06 3.25  -21.28  -15.86  -13.65  -11.64   -8.67
#> log_p_z[37,1]  -18.54    0.03 2.22  -23.26  -19.89  -18.43  -16.94  -14.68
#> log_p_z[37,2]   -7.61    0.01 0.81   -9.33   -8.14   -7.54   -7.03   -6.23
#> log_p_z[37,3]  -19.10    0.06 3.24  -26.29  -21.04  -18.69  -16.86  -13.85
#> log_p_z[37,4]  -21.85    0.06 3.53  -29.53  -24.05  -21.59  -19.36  -15.85
#> log_p_z[38,1]  -25.21    0.04 2.73  -31.04  -26.97  -25.00  -23.26  -20.36
#> log_p_z[38,2]   -9.61    0.01 1.01  -11.87  -10.19   -9.53   -8.90   -7.83
#> log_p_z[38,3]  -11.73    0.05 2.76  -18.06  -13.28  -11.33   -9.74   -7.54
#> log_p_z[38,4]  -21.25    0.06 3.43  -28.93  -23.22  -20.96  -18.86  -15.38
#> log_p_z[39,1]  -29.39    0.06 4.35  -39.25  -31.97  -28.95  -26.33  -22.17
#> log_p_z[39,2]  -16.85    0.03 2.06  -21.41  -18.10  -16.66  -15.44  -13.32
#> log_p_z[39,3]   -7.26    0.01 1.07   -9.59   -7.91   -7.17   -6.53   -5.41
#> log_p_z[39,4]  -18.74    0.07 4.14  -28.24  -21.09  -18.10  -15.80  -12.28
#> log_p_z[40,1]   -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_p_z[40,2]  -19.97    0.04 2.29  -25.08  -21.36  -19.80  -18.36  -16.03
#> log_p_z[40,3]  -24.82    0.07 4.25  -34.60  -27.27  -24.30  -21.83  -17.96
#> log_p_z[40,4]  -21.90    0.07 4.39  -32.05  -24.33  -21.38  -18.85  -14.85
#> log_p_z[41,1]   -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_p_z[41,2]  -19.97    0.04 2.29  -25.08  -21.36  -19.80  -18.36  -16.03
#> log_p_z[41,3]  -24.82    0.07 4.25  -34.60  -27.27  -24.30  -21.83  -17.96
#> log_p_z[41,4]  -21.90    0.07 4.39  -32.05  -24.33  -21.38  -18.85  -14.85
#> log_p_z[42,1]   -6.01    0.01 0.63   -7.38   -6.41   -5.96   -5.55   -4.90
#> log_p_z[42,2]  -14.94    0.03 1.72  -18.76  -15.97  -14.78  -13.72  -12.03
#> log_p_z[42,3]  -24.25    0.06 3.50  -32.14  -26.40  -23.93  -21.77  -18.46
#> log_p_z[42,4]  -21.90    0.06 3.57  -29.96  -24.05  -21.63  -19.38  -15.94
#> log_p_z[43,1]  -21.85    0.03 2.42  -27.12  -23.32  -21.66  -20.15  -17.64
#> log_p_z[43,2]   -5.65    0.01 0.52   -6.74   -5.98   -5.61   -5.28   -4.73
#> log_p_z[43,3]  -18.34    0.06 3.22  -25.47  -20.26  -17.93  -16.12  -13.12
#> log_p_z[43,4]  -21.59    0.06 3.48  -29.17  -23.72  -21.25  -19.12  -15.81
#> log_p_z[44,1]   -4.41    0.00 0.41   -5.30   -4.66   -4.38   -4.12   -3.72
#> log_p_z[44,2]  -17.22    0.03 1.84  -21.19  -18.34  -17.11  -15.92  -14.03
#> log_p_z[44,3]  -24.31    0.06 3.53  -32.25  -26.44  -23.97  -21.82  -18.30
#> log_p_z[44,4]  -21.89    0.06 3.63  -30.33  -24.05  -21.51  -19.33  -15.73
#> log_p_z[45,1]  -19.60    0.03 2.33  -24.67  -21.07  -19.44  -17.94  -15.49
#> log_p_z[45,2]   -3.04    0.00 0.35   -3.80   -3.26   -3.03   -2.80   -2.40
#> log_p_z[45,3]  -21.18    0.08 4.34  -30.97  -23.65  -20.57  -18.12  -14.41
#> log_p_z[45,4]  -21.90    0.07 4.26  -31.36  -24.48  -21.41  -18.90  -14.94
#> log_lik[1]      -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_lik[2]      -7.59    0.02 1.13  -10.07   -8.29   -7.51   -6.78   -5.62
#> log_lik[3]      -9.93    0.01 0.92  -11.82  -10.50   -9.88   -9.29   -8.24
#> log_lik[4]      -7.57    0.01 0.92   -9.53   -8.15   -7.50   -6.92   -5.94
#> log_lik[5]      -3.72    0.01 0.44   -4.66   -4.01   -3.70   -3.41   -2.95
#> log_lik[6]      -5.46    0.01 0.65   -6.85   -5.88   -5.42   -5.00   -4.30
#> log_lik[7]      -6.00    0.01 0.63   -7.38   -6.41   -5.96   -5.55   -4.90
#> log_lik[8]      -5.72    0.01 0.79   -7.34   -6.24   -5.69   -5.16   -4.30
#> log_lik[9]      -4.50    0.01 0.56   -5.70   -4.85   -4.47   -4.12   -3.50
#> log_lik[10]     -7.11    0.01 0.70   -8.64   -7.55   -7.06   -6.63   -5.90
#> log_lik[11]     -5.83    0.01 0.98   -7.89   -6.45   -5.74   -5.13   -4.11
#> log_lik[12]     -9.32    0.01 1.04  -11.48  -10.01   -9.28   -8.60   -7.44
#> log_lik[13]     -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_lik[14]     -6.26    0.01 0.88   -8.20   -6.79   -6.19   -5.65   -4.72
#> log_lik[15]     -4.41    0.00 0.41   -5.30   -4.66   -4.38   -4.12   -3.72
#> log_lik[16]     -3.80    0.01 0.55   -5.00   -4.15   -3.76   -3.41   -2.85
#> log_lik[17]     -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_lik[18]     -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_lik[19]     -4.66    0.01 0.60   -5.93   -5.06   -4.63   -4.24   -3.61
#> log_lik[20]     -7.06    0.01 0.95   -9.16   -7.64   -6.99   -6.39   -5.42
#> log_lik[21]     -3.04    0.00 0.35   -3.80   -3.26   -3.03   -2.80   -2.40
#> log_lik[22]     -4.66    0.01 0.60   -5.93   -5.06   -4.63   -4.24   -3.61
#> log_lik[23]     -3.72    0.01 0.44   -4.66   -4.01   -3.70   -3.41   -2.95
#> log_lik[24]     -5.80    0.01 0.64   -7.25   -6.16   -5.73   -5.35   -4.75
#> log_lik[25]     -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_lik[26]     -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_lik[27]     -5.65    0.01 0.52   -6.74   -5.98   -5.61   -5.28   -4.73
#> log_lik[28]     -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_lik[29]     -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_lik[30]     -6.81    0.01 0.85   -8.69   -7.33   -6.74   -6.21   -5.37
#> log_lik[31]     -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_lik[32]     -6.06    0.01 0.93   -8.12   -6.64   -5.97   -5.42   -4.49
#> log_lik[33]     -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_lik[34]     -3.04    0.00 0.35   -3.80   -3.26   -3.03   -2.80   -2.40
#> log_lik[35]     -5.53    0.01 0.71   -7.06   -5.97   -5.48   -5.03   -4.31
#> log_lik[36]     -9.92    0.02 1.16  -12.35  -10.64   -9.86   -9.12   -7.84
#> log_lik[37]     -7.61    0.01 0.81   -9.33   -8.14   -7.54   -7.03   -6.23
#> log_lik[38]     -9.02    0.01 0.86  -10.76   -9.57   -9.02   -8.46   -7.33
#> log_lik[39]     -7.26    0.01 1.07   -9.59   -7.91   -7.17   -6.53   -5.41
#> log_lik[40]     -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_lik[41]     -2.20    0.00 0.29   -2.82   -2.39   -2.19   -2.00   -1.67
#> log_lik[42]     -6.00    0.01 0.63   -7.38   -6.41   -5.96   -5.55   -4.90
#> log_lik[43]     -5.65    0.01 0.52   -6.74   -5.98   -5.61   -5.28   -4.73
#> log_lik[44]     -4.41    0.00 0.41   -5.30   -4.66   -4.38   -4.12   -3.72
#> log_lik[45]     -3.04    0.00 0.35   -3.80   -3.26   -3.03   -2.80   -2.40
#> lp__          -463.66    0.16 6.22 -477.14 -467.57 -463.30 -459.26 -452.60
#>               n_eff Rhat
#> pi[1]         10595    1
#> pi[2]          8114    1
#> pi[3]          6459    1
#> pi[4]          6620    1
#> theta[1,1,1]   8341    1
#> theta[1,1,2]   7198    1
#> theta[1,1,3]   4923    1
#> theta[1,1,4]   4918    1
#> theta[1,2,1]   6332    1
#> theta[1,2,2]   7254    1
#> theta[1,2,3]   6892    1
#> theta[1,2,4]   5383    1
#> theta[1,3,1]   6094    1
#> theta[1,3,2]   4312    1
#> theta[1,3,3]   5190    1
#> theta[1,3,4]   6096    1
#> theta[1,4,1]   7154    1
#> theta[1,4,2]   6020    1
#> theta[1,4,3]   4142    1
#> theta[1,4,4]   4847    1
#> theta[2,1,1]   7093    1
#> theta[2,1,2]   6871    1
#> theta[2,1,3]   4851    1
#> theta[2,1,4]   5076    1
#> theta[2,2,1]   6126    1
#> theta[2,2,2]   7904    1
#> theta[2,2,3]   7325    1
#> theta[2,2,4]   5875    1
#> theta[2,3,1]   6234    1
#> theta[2,3,2]   6531    1
#> theta[2,3,3]   5899    1
#> theta[2,3,4]   5395    1
#> theta[2,4,1]   5873    1
#> theta[2,4,2]   6299    1
#> theta[2,4,3]   6755    1
#> theta[2,4,4]   6706    1
#> theta[3,1,1]   6871    1
#> theta[3,1,2]   5658    1
#> theta[3,1,3]   4966    1
#> theta[3,1,4]   5136    1
#> theta[3,2,1]   6584    1
#> theta[3,2,2]   7083    1
#> theta[3,2,3]   6879    1
#> theta[3,2,4]   5375    1
#> theta[3,3,1]   7197    1
#> theta[3,3,2]   7590    1
#> theta[3,3,3]   7180    1
#> theta[3,3,4]   6950    1
#> theta[3,4,1]   6136    1
#> theta[3,4,2]   6375    1
#> theta[3,4,3]   4820    1
#> theta[3,4,4]   5375    1
#> theta[4,1,1]   7276    1
#> theta[4,1,2]   5680    1
#> theta[4,1,3]   4843    1
#> theta[4,1,4]   4716    1
#> theta[4,2,1]   6088    1
#> theta[4,2,2]   7886    1
#> theta[4,2,3]   7049    1
#> theta[4,2,4]   6013    1
#> theta[4,3,1]   6308    1
#> theta[4,3,2]   6602    1
#> theta[4,3,3]   6561    1
#> theta[4,3,4]   5588    1
#> theta[4,4,1]   7449    1
#> theta[4,4,2]   8229    1
#> theta[4,4,3]   6970    1
#> theta[4,4,4]   7605    1
#> theta[5,1,1]   7099    1
#> theta[5,1,2]   4213    1
#> theta[5,1,3]   5423    1
#> theta[5,1,4]   4998    1
#> theta[5,2,1]   6813    1
#> theta[5,2,2]   6984    1
#> theta[5,2,3]   5750    1
#> theta[5,2,4]   5702    1
#> theta[5,3,1]   6312    1
#> theta[5,3,2]   8288    1
#> theta[5,3,3]   6641    1
#> theta[5,3,4]   6619    1
#> theta[5,4,1]   7570    1
#> theta[5,4,2]   5498    1
#> theta[5,4,3]   6233    1
#> theta[5,4,4]   6508    1
#> log_p_z[1,1]   7811    1
#> log_p_z[1,2]   4053    1
#> log_p_z[1,3]   3695    1
#> log_p_z[1,4]   3635    1
#> log_p_z[2,1]   4775    1
#> log_p_z[2,2]   4444    1
#> log_p_z[2,3]   3732    1
#> log_p_z[2,4]   3033    1
#> log_p_z[3,1]   3865    1
#> log_p_z[3,2]   3746    1
#> log_p_z[3,3]   3680    1
#> log_p_z[3,4]   3612    1
#> log_p_z[4,1]   5039    1
#> log_p_z[4,2]   5364    1
#> log_p_z[4,3]   3166    1
#> log_p_z[4,4]   3854    1
#> log_p_z[5,1]   4770    1
#> log_p_z[5,2]   6573    1
#> log_p_z[5,3]   3082    1
#> log_p_z[5,4]   4062    1
#> log_p_z[6,1]   4433    1
#> log_p_z[6,2]   6918    1
#> log_p_z[6,3]   3023    1
#> log_p_z[6,4]   3904    1
#> log_p_z[7,1]   6638    1
#> log_p_z[7,2]   4155    1
#> log_p_z[7,3]   3659    1
#> log_p_z[7,4]   3290    1
#> log_p_z[8,1]   4771    1
#> log_p_z[8,2]   4524    1
#> log_p_z[8,3]   5614    1
#> log_p_z[8,4]   3239    1
#> log_p_z[9,1]   5353    1
#> log_p_z[9,2]   5658    1
#> log_p_z[9,3]   3184    1
#> log_p_z[9,4]   4267    1
#> log_p_z[10,1]  5506    1
#> log_p_z[10,2]  5208    1
#> log_p_z[10,3]  3362    1
#> log_p_z[10,4]  3901    1
#> log_p_z[11,1]  4966    1
#> log_p_z[11,2]  3865    1
#> log_p_z[11,3]  3070    1
#> log_p_z[11,4]  5484    1
#> log_p_z[12,1]  5464    1
#> log_p_z[12,2]  4633    1
#> log_p_z[12,3]  2987    1
#> log_p_z[12,4]  3516    1
#> log_p_z[13,1]  7811    1
#> log_p_z[13,2]  4053    1
#> log_p_z[13,3]  3695    1
#> log_p_z[13,4]  3635    1
#> log_p_z[14,1]  4824    1
#> log_p_z[14,2]  5300    1
#> log_p_z[14,3]  3091    1
#> log_p_z[14,4]  3924    1
#> log_p_z[15,1]  7023    1
#> log_p_z[15,2]  4017    1
#> log_p_z[15,3]  3558    1
#> log_p_z[15,4]  3320    1
#> log_p_z[16,1]  6964    1
#> log_p_z[16,2]  4170    1
#> log_p_z[16,3]  3708    1
#> log_p_z[16,4]  3651    1
#> log_p_z[17,1]  7811    1
#> log_p_z[17,2]  4053    1
#> log_p_z[17,3]  3695    1
#> log_p_z[17,4]  3635    1
#> log_p_z[18,1]  7811    1
#> log_p_z[18,2]  4053    1
#> log_p_z[18,3]  3695    1
#> log_p_z[18,4]  3635    1
#> log_p_z[19,1]  6453    1
#> log_p_z[19,2]  5753    1
#> log_p_z[19,3]  3148    1
#> log_p_z[19,4]  4263    1
#> log_p_z[20,1]  4603    1
#> log_p_z[20,2]  5102    1
#> log_p_z[20,3]  2868    1
#> log_p_z[20,4]  3808    1
#> log_p_z[21,1]  4757    1
#> log_p_z[21,2]  6798    1
#> log_p_z[21,3]  3167    1
#> log_p_z[21,4]  4150    1
#> log_p_z[22,1]  6453    1
#> log_p_z[22,2]  5753    1
#> log_p_z[22,3]  3148    1
#> log_p_z[22,4]  4263    1
#> log_p_z[23,1]  4770    1
#> log_p_z[23,2]  6573    1
#> log_p_z[23,3]  3082    1
#> log_p_z[23,4]  4062    1
#> log_p_z[24,1]  4467    1
#> log_p_z[24,2]  4541    1
#> log_p_z[24,3]  3366    1
#> log_p_z[24,4]  4029    1
#> log_p_z[25,1]  7811    1
#> log_p_z[25,2]  4053    1
#> log_p_z[25,3]  3695    1
#> log_p_z[25,4]  3635    1
#> log_p_z[26,1]  7811    1
#> log_p_z[26,2]  4053    1
#> log_p_z[26,3]  3695    1
#> log_p_z[26,4]  3635    1
#> log_p_z[27,1]  5139    1
#> log_p_z[27,2]  6223    1
#> log_p_z[27,3]  3342    1
#> log_p_z[27,4]  3957    1
#> log_p_z[28,1]  7811    1
#> log_p_z[28,2]  4053    1
#> log_p_z[28,3]  3695    1
#> log_p_z[28,4]  3635    1
#> log_p_z[29,1]  7811    1
#> log_p_z[29,2]  4053    1
#> log_p_z[29,3]  3695    1
#> log_p_z[29,4]  3635    1
#> log_p_z[30,1]  5542    1
#> log_p_z[30,2]  3886    1
#> log_p_z[30,3]  3534    1
#> log_p_z[30,4]  3239    1
#> log_p_z[31,1]  7811    1
#> log_p_z[31,2]  4053    1
#> log_p_z[31,3]  3695    1
#> log_p_z[31,4]  3635    1
#> log_p_z[32,1]  4852    1
#> log_p_z[32,2]  5622    1
#> log_p_z[32,3]  5504    1
#> log_p_z[32,4]  3189    1
#> log_p_z[33,1]  7811    1
#> log_p_z[33,2]  4053    1
#> log_p_z[33,3]  3695    1
#> log_p_z[33,4]  3635    1
#> log_p_z[34,1]  4757    1
#> log_p_z[34,2]  6798    1
#> log_p_z[34,3]  3167    1
#> log_p_z[34,4]  4150    1
#> log_p_z[35,1]  4476    1
#> log_p_z[35,2]  6613    1
#> log_p_z[35,3]  3044    1
#> log_p_z[35,4]  3768    1
#> log_p_z[36,1]  4316    1
#> log_p_z[36,2]  3818    1
#> log_p_z[36,3]  2633    1
#> log_p_z[36,4]  2814    1
#> log_p_z[37,1]  4313    1
#> log_p_z[37,2]  5406    1
#> log_p_z[37,3]  3323    1
#> log_p_z[37,4]  3745    1
#> log_p_z[38,1]  4886    1
#> log_p_z[38,2]  5346    1
#> log_p_z[38,3]  3072    1
#> log_p_z[38,4]  3214    1
#> log_p_z[39,1]  4693    1
#> log_p_z[39,2]  4597    1
#> log_p_z[39,3]  5507    1
#> log_p_z[39,4]  3439    1
#> log_p_z[40,1]  7811    1
#> log_p_z[40,2]  4053    1
#> log_p_z[40,3]  3695    1
#> log_p_z[40,4]  3635    1
#> log_p_z[41,1]  7811    1
#> log_p_z[41,2]  4053    1
#> log_p_z[41,3]  3695    1
#> log_p_z[41,4]  3635    1
#> log_p_z[42,1]  6638    1
#> log_p_z[42,2]  4155    1
#> log_p_z[42,3]  3659    1
#> log_p_z[42,4]  3290    1
#> log_p_z[43,1]  5139    1
#> log_p_z[43,2]  6223    1
#> log_p_z[43,3]  3342    1
#> log_p_z[43,4]  3957    1
#> log_p_z[44,1]  7023    1
#> log_p_z[44,2]  4017    1
#> log_p_z[44,3]  3558    1
#> log_p_z[44,4]  3320    1
#> log_p_z[45,1]  4757    1
#> log_p_z[45,2]  6798    1
#> log_p_z[45,3]  3167    1
#> log_p_z[45,4]  4150    1
#> log_lik[1]     7811    1
#> log_lik[2]     4597    1
#> log_lik[3]     5003    1
#> log_lik[4]     5364    1
#> log_lik[5]     6574    1
#> log_lik[6]     6923    1
#> log_lik[7]     6640    1
#> log_lik[8]     5617    1
#> log_lik[9]     5658    1
#> log_lik[10]    5214    1
#> log_lik[11]    5484    1
#> log_lik[12]    5920    1
#> log_lik[13]    7811    1
#> log_lik[14]    5304    1
#> log_lik[15]    7023    1
#> log_lik[16]    6964    1
#> log_lik[17]    7811    1
#> log_lik[18]    7811    1
#> log_lik[19]    5753    1
#> log_lik[20]    5109    1
#> log_lik[21]    6799    1
#> log_lik[22]    5753    1
#> log_lik[23]    6574    1
#> log_lik[24]    4543    1
#> log_lik[25]    7811    1
#> log_lik[26]    7811    1
#> log_lik[27]    6222    1
#> log_lik[28]    7811    1
#> log_lik[29]    7811    1
#> log_lik[30]    5544    1
#> log_lik[31]    7811    1
#> log_lik[32]    5512    1
#> log_lik[33]    7811    1
#> log_lik[34]    6799    1
#> log_lik[35]    6660    1
#> log_lik[36]    4327    1
#> log_lik[37]    5407    1
#> log_lik[38]    4487    1
#> log_lik[39]    5532    1
#> log_lik[40]    7811    1
#> log_lik[41]    7811    1
#> log_lik[42]    6640    1
#> log_lik[43]    6222    1
#> log_lik[44]    7023    1
#> log_lik[45]    6799    1
#> lp__           1495    1
#> 
#> Samples were drawn using NUTS(diag_e) at Wed Sep  6 12:22:08 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }