Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.37    0.00 0.06    0.25    0.33    0.37    0.42    0.51
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.08    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.08    0.10    0.12    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.06    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.11    0.28
#> theta[1,3,3]     0.80    0.00 0.09    0.58    0.74    0.81    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.23
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.30
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.40
#> theta[1,4,4]     0.68    0.00 0.14    0.39    0.59    0.69    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.75    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.21    0.33
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.58    0.65    0.75
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.24    0.30    0.37    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[2,3,2]     0.08    0.00 0.08    0.00    0.03    0.06    0.11    0.29
#> theta[2,3,3]     0.62    0.00 0.14    0.35    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.29    0.47
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.56    0.67    0.77    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.09    0.13    0.22
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.26
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.14    0.24    0.41    0.51    0.61    0.77
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.14    0.22    0.30    0.49
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[3,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[3,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.21    0.42
#> theta[3,4,4]     0.63    0.00 0.15    0.32    0.52    0.64    0.74    0.90
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.78    0.83    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.24
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[4,2,2]     0.72    0.00 0.09    0.55    0.67    0.73    0.79    0.87
#> theta[4,2,3]     0.13    0.00 0.07    0.04    0.08    0.12    0.17    0.29
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[4,3,3]     0.69    0.00 0.13    0.43    0.61    0.70    0.78    0.91
#> theta[4,3,4]     0.14    0.00 0.10    0.01    0.07    0.13    0.20    0.38
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.91
#> theta[5,1,1]     0.87    0.00 0.07    0.72    0.83    0.87    0.91    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.06    0.04    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.08    0.00    0.03    0.06    0.11    0.28
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.22    0.39
#> theta[5,3,3]     0.61    0.00 0.13    0.33    0.52    0.61    0.70    0.85
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.22    0.38
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.43
#> theta[5,4,4]     0.63    0.00 0.15    0.32    0.53    0.64    0.74    0.89
#> log_p_z[1,1]    -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_p_z[1,2]   -19.99    0.03 2.28  -24.96  -21.39  -19.79  -18.40  -16.03
#> log_p_z[1,3]   -24.84    0.07 4.32  -34.67  -27.43  -24.28  -21.75  -17.84
#> log_p_z[1,4]   -21.91    0.07 4.39  -32.01  -24.48  -21.33  -18.83  -14.76
#> log_p_z[2,1]   -29.60    0.06 4.31  -39.45  -32.21  -29.16  -26.48  -22.55
#> log_p_z[2,2]   -21.07    0.04 2.41  -26.33  -22.58  -20.88  -19.41  -16.83
#> log_p_z[2,3]    -7.63    0.02 1.26  -10.39   -8.33   -7.52   -6.75   -5.58
#> log_p_z[2,4]   -16.18    0.07 4.12  -25.58  -18.52  -15.70  -13.21   -9.59
#> log_p_z[3,1]   -11.60    0.03 1.61  -15.20  -12.55  -11.46  -10.42   -8.99
#> log_p_z[3,2]   -10.79    0.02 1.47  -14.17  -11.67  -10.62   -9.75   -8.41
#> log_p_z[3,3]   -23.42    0.06 3.44  -30.90  -25.49  -23.12  -20.98  -17.64
#> log_p_z[3,4]   -21.90    0.06 3.52  -29.30  -24.07  -21.56  -19.46  -15.84
#> log_p_z[4,1]   -14.66    0.03 1.91  -18.87  -15.83  -14.52  -13.35  -11.27
#> log_p_z[4,2]    -7.55    0.01 0.94   -9.62   -8.11   -7.47   -6.91   -5.95
#> log_p_z[4,3]   -20.83    0.08 4.37  -30.93  -23.28  -20.35  -17.73  -13.85
#> log_p_z[4,4]   -21.95    0.07 4.32  -31.46  -24.52  -21.43  -18.91  -14.90
#> log_p_z[5,1]   -21.45    0.04 2.55  -26.82  -23.11  -21.28  -19.66  -16.76
#> log_p_z[5,2]    -3.71    0.01 0.44   -4.61   -3.99   -3.69   -3.41   -2.91
#> log_p_z[5,3]   -18.80    0.07 4.29  -28.57  -21.20  -18.26  -15.78  -12.02
#> log_p_z[5,4]   -21.90    0.08 4.34  -31.58  -24.40  -21.37  -18.85  -15.02
#> log_p_z[6,1]   -21.49    0.04 2.58  -27.10  -23.12  -21.32  -19.71  -16.93
#> log_p_z[6,2]    -5.47    0.01 0.66   -6.88   -5.87   -5.42   -5.02   -4.32
#> log_p_z[6,3]   -17.59    0.07 4.16  -27.09  -19.86  -16.99  -14.64  -11.24
#> log_p_z[6,4]   -21.66    0.07 4.26  -31.29  -24.08  -21.14  -18.68  -14.84
#> log_p_z[7,1]    -6.00    0.01 0.60   -7.27   -6.36   -5.95   -5.59   -4.93
#> log_p_z[7,2]   -14.93    0.03 1.68  -18.69  -15.99  -14.76  -13.74  -12.05
#> log_p_z[7,3]   -24.37    0.07 3.57  -32.09  -26.55  -24.09  -21.83  -18.32
#> log_p_z[7,4]   -21.93    0.06 3.62  -29.92  -24.21  -21.55  -19.39  -15.75
#> log_p_z[8,1]   -29.66    0.07 4.32  -39.71  -32.19  -29.13  -26.65  -22.65
#> log_p_z[8,2]   -18.34    0.03 2.13  -23.05  -19.59  -18.22  -16.81  -14.67
#> log_p_z[8,3]    -5.73    0.01 0.85   -7.50   -6.27   -5.68   -5.12   -4.21
#> log_p_z[8,4]   -18.41    0.08 4.23  -28.15  -20.84  -17.93  -15.43  -11.58
#> log_p_z[9,1]   -20.02    0.03 2.31  -24.93  -21.48  -19.89  -18.40  -15.89
#> log_p_z[9,2]    -4.51    0.01 0.57   -5.77   -4.87   -4.47   -4.11   -3.51
#> log_p_z[9,3]   -19.79    0.08 4.41  -29.84  -22.25  -19.22  -16.67  -12.71
#> log_p_z[9,4]   -21.68    0.07 4.30  -31.53  -24.15  -21.22  -18.63  -14.70
#> log_p_z[10,1]  -22.23    0.03 2.40  -27.35  -23.76  -22.07  -20.57  -18.01
#> log_p_z[10,2]   -7.13    0.01 0.72   -8.70   -7.57   -7.08   -6.63   -5.87
#> log_p_z[10,3]  -16.93    0.06 3.24  -24.21  -18.80  -16.47  -14.70  -11.61
#> log_p_z[10,4]  -21.32    0.06 3.55  -29.20  -23.45  -20.98  -18.84  -15.29
#> log_p_z[11,1]  -29.62    0.07 4.39  -39.41  -32.34  -29.15  -26.50  -22.26
#> log_p_z[11,2]  -29.44    0.07 4.32  -38.97  -31.94  -28.99  -26.35  -22.31
#> log_p_z[11,3]  -18.31    0.07 3.43  -26.82  -20.08  -17.82  -15.90  -13.05
#> log_p_z[11,4]   -5.83    0.01 0.98   -8.01   -6.42   -5.74   -5.15   -4.10
#> log_p_z[12,1]  -23.07    0.05 2.75  -28.92  -24.78  -22.83  -21.17  -18.38
#> log_p_z[12,2]   -9.59    0.02 1.29  -12.52  -10.29   -9.44   -8.74   -7.49
#> log_p_z[12,3]  -15.29    0.07 4.12  -24.97  -17.53  -14.69  -12.38   -8.95
#> log_p_z[12,4]  -19.28    0.07 4.08  -28.87  -21.51  -18.78  -16.41  -12.78
#> log_p_z[13,1]   -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_p_z[13,2]  -19.99    0.03 2.28  -24.96  -21.39  -19.79  -18.40  -16.03
#> log_p_z[13,3]  -24.84    0.07 4.32  -34.67  -27.43  -24.28  -21.75  -17.84
#> log_p_z[13,4]  -21.91    0.07 4.39  -32.01  -24.48  -21.33  -18.83  -14.76
#> log_p_z[14,1]  -19.06    0.03 2.37  -24.10  -20.56  -18.87  -17.37  -14.91
#> log_p_z[14,2]   -6.24    0.01 0.90   -8.16   -6.79   -6.15   -5.61   -4.71
#> log_p_z[14,3]  -18.81    0.08 4.28  -28.54  -21.18  -18.27  -15.72  -12.10
#> log_p_z[14,4]  -21.92    0.08 4.30  -31.48  -24.37  -21.42  -18.96  -14.90
#> log_p_z[15,1]   -4.42    0.00 0.39   -5.28   -4.65   -4.39   -4.15   -3.73
#> log_p_z[15,2]  -17.23    0.03 1.83  -21.17  -18.38  -17.10  -15.97  -14.00
#> log_p_z[15,3]  -24.34    0.06 3.57  -32.20  -26.55  -24.04  -21.83  -18.14
#> log_p_z[15,4]  -21.92    0.06 3.60  -29.75  -24.10  -21.59  -19.37  -15.83
#> log_p_z[16,1]   -3.79    0.01 0.54   -5.00   -4.12   -3.75   -3.43   -2.86
#> log_p_z[16,2]  -17.69    0.03 2.17  -22.59  -19.07  -17.47  -16.14  -14.02
#> log_p_z[16,3]  -24.87    0.08 4.29  -34.67  -27.39  -24.39  -21.92  -17.84
#> log_p_z[16,4]  -21.92    0.07 4.40  -32.11  -24.49  -21.38  -18.85  -14.82
#> log_p_z[17,1]   -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_p_z[17,2]  -19.99    0.03 2.28  -24.96  -21.39  -19.79  -18.40  -16.03
#> log_p_z[17,3]  -24.84    0.07 4.32  -34.67  -27.43  -24.28  -21.75  -17.84
#> log_p_z[17,4]  -21.91    0.07 4.39  -32.01  -24.48  -21.33  -18.83  -14.76
#> log_p_z[18,1]   -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_p_z[18,2]  -19.99    0.03 2.28  -24.96  -21.39  -19.79  -18.40  -16.03
#> log_p_z[18,3]  -24.84    0.07 4.32  -34.67  -27.43  -24.28  -21.75  -17.84
#> log_p_z[18,4]  -21.91    0.07 4.39  -32.01  -24.48  -21.33  -18.83  -14.76
#> log_p_z[19,1]  -16.38    0.03 1.93  -20.48  -17.60  -16.26  -15.02  -12.99
#> log_p_z[19,2]   -4.65    0.01 0.59   -5.89   -5.03   -4.62   -4.22   -3.62
#> log_p_z[19,3]  -22.30    0.08 4.49  -32.56  -24.81  -21.73  -19.10  -15.07
#> log_p_z[19,4]  -21.90    0.07 4.34  -31.93  -24.38  -21.42  -18.80  -14.81
#> log_p_z[20,1]  -18.05    0.04 2.25  -22.95  -19.39  -17.90  -16.47  -14.11
#> log_p_z[20,2]   -7.10    0.01 0.93   -9.10   -7.64   -7.03   -6.45   -5.47
#> log_p_z[20,3]  -20.11    0.07 4.30  -29.92  -22.46  -19.55  -17.07  -13.47
#> log_p_z[20,4]  -21.63    0.07 4.23  -31.37  -23.94  -21.10  -18.63  -14.82
#> log_p_z[21,1]  -19.59    0.03 2.31  -24.44  -21.08  -19.46  -17.99  -15.47
#> log_p_z[21,2]   -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.79   -2.40
#> log_p_z[21,3]  -21.35    0.08 4.44  -31.35  -23.80  -20.76  -18.20  -14.27
#> log_p_z[21,4]  -21.89    0.07 4.33  -31.90  -24.36  -21.43  -18.80  -14.93
#> log_p_z[22,1]  -16.38    0.03 1.93  -20.48  -17.60  -16.26  -15.02  -12.99
#> log_p_z[22,2]   -4.65    0.01 0.59   -5.89   -5.03   -4.62   -4.22   -3.62
#> log_p_z[22,3]  -22.30    0.08 4.49  -32.56  -24.81  -21.73  -19.10  -15.07
#> log_p_z[22,4]  -21.90    0.07 4.34  -31.93  -24.38  -21.42  -18.80  -14.81
#> log_p_z[23,1]  -21.45    0.04 2.55  -26.82  -23.11  -21.28  -19.66  -16.76
#> log_p_z[23,2]   -3.71    0.01 0.44   -4.61   -3.99   -3.69   -3.41   -2.91
#> log_p_z[23,3]  -18.80    0.07 4.29  -28.57  -21.20  -18.26  -15.78  -12.02
#> log_p_z[23,4]  -21.90    0.08 4.34  -31.58  -24.40  -21.37  -18.85  -15.02
#> log_p_z[24,1]  -17.39    0.03 2.10  -21.89  -18.73  -17.27  -15.92  -13.67
#> log_p_z[24,2]   -5.80    0.01 0.62   -7.22   -6.16   -5.75   -5.37   -4.77
#> log_p_z[24,3]  -21.85    0.06 3.50  -29.69  -23.94  -21.44  -19.39  -16.13
#> log_p_z[24,4]  -21.88    0.06 3.53  -29.90  -23.94  -21.61  -19.28  -16.00
#> log_p_z[25,1]   -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_p_z[25,2]  -19.99    0.03 2.28  -24.96  -21.39  -19.79  -18.40  -16.03
#> log_p_z[25,3]  -24.84    0.07 4.32  -34.67  -27.43  -24.28  -21.75  -17.84
#> log_p_z[25,4]  -21.91    0.07 4.39  -32.01  -24.48  -21.33  -18.83  -14.76
#> log_p_z[26,1]   -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_p_z[26,2]  -19.99    0.03 2.28  -24.96  -21.39  -19.79  -18.40  -16.03
#> log_p_z[26,3]  -24.84    0.07 4.32  -34.67  -27.43  -24.28  -21.75  -17.84
#> log_p_z[26,4]  -21.91    0.07 4.39  -32.01  -24.48  -21.33  -18.83  -14.76
#> log_p_z[27,1]  -21.80    0.04 2.39  -26.96  -23.31  -21.64  -20.11  -17.65
#> log_p_z[27,2]   -5.66    0.01 0.54   -6.86   -5.99   -5.62   -5.28   -4.69
#> log_p_z[27,3]  -18.49    0.06 3.29  -25.98  -20.41  -18.08  -16.16  -13.15
#> log_p_z[27,4]  -21.52    0.06 3.53  -29.48  -23.59  -21.18  -18.98  -15.65
#> log_p_z[28,1]   -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_p_z[28,2]  -19.99    0.03 2.28  -24.96  -21.39  -19.79  -18.40  -16.03
#> log_p_z[28,3]  -24.84    0.07 4.32  -34.67  -27.43  -24.28  -21.75  -17.84
#> log_p_z[28,4]  -21.91    0.07 4.39  -32.01  -24.48  -21.33  -18.83  -14.76
#> log_p_z[29,1]   -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_p_z[29,2]  -19.99    0.03 2.28  -24.96  -21.39  -19.79  -18.40  -16.03
#> log_p_z[29,3]  -24.84    0.07 4.32  -34.67  -27.43  -24.28  -21.75  -17.84
#> log_p_z[29,4]  -21.91    0.07 4.39  -32.01  -24.48  -21.33  -18.83  -14.76
#> log_p_z[30,1]   -6.81    0.01 0.83   -8.73   -7.29   -6.75   -6.22   -5.43
#> log_p_z[30,2]  -14.70    0.03 1.69  -18.31  -15.73  -14.59  -13.52  -11.79
#> log_p_z[30,3]  -24.34    0.06 3.55  -32.38  -26.55  -24.03  -21.83  -18.34
#> log_p_z[30,4]  -21.90    0.06 3.55  -29.40  -24.13  -21.62  -19.34  -15.74
#> log_p_z[31,1]   -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_p_z[31,2]  -19.99    0.03 2.28  -24.96  -21.39  -19.79  -18.40  -16.03
#> log_p_z[31,3]  -24.84    0.07 4.32  -34.67  -27.43  -24.28  -21.75  -17.84
#> log_p_z[31,4]  -21.91    0.07 4.39  -32.01  -24.48  -21.33  -18.83  -14.76
#> log_p_z[32,1]  -29.66    0.07 4.24  -39.31  -32.06  -29.10  -26.79  -22.75
#> log_p_z[32,2]  -14.85    0.02 1.72  -18.65  -15.91  -14.73  -13.68  -11.85
#> log_p_z[32,3]   -6.05    0.01 0.95   -8.20   -6.62   -5.97   -5.39   -4.45
#> log_p_z[32,4]  -20.55    0.08 4.39  -30.57  -23.04  -20.11  -17.54  -13.31
#> log_p_z[33,1]   -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_p_z[33,2]  -19.99    0.03 2.28  -24.96  -21.39  -19.79  -18.40  -16.03
#> log_p_z[33,3]  -24.84    0.07 4.32  -34.67  -27.43  -24.28  -21.75  -17.84
#> log_p_z[33,4]  -21.91    0.07 4.39  -32.01  -24.48  -21.33  -18.83  -14.76
#> log_p_z[34,1]  -19.59    0.03 2.31  -24.44  -21.08  -19.46  -17.99  -15.47
#> log_p_z[34,2]   -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.79   -2.40
#> log_p_z[34,3]  -21.35    0.08 4.44  -31.35  -23.80  -20.76  -18.20  -14.27
#> log_p_z[34,4]  -21.89    0.07 4.33  -31.90  -24.36  -21.43  -18.80  -14.93
#> log_p_z[35,1]  -22.60    0.04 2.70  -28.55  -24.25  -22.40  -20.68  -17.88
#> log_p_z[35,2]   -5.52    0.01 0.69   -7.06   -5.94   -5.47   -5.05   -4.33
#> log_p_z[35,3]  -16.20    0.07 4.16  -25.85  -18.51  -15.63  -13.21   -9.87
#> log_p_z[35,4]  -21.86    0.07 4.25  -31.65  -24.20  -21.34  -18.92  -14.93
#> log_p_z[36,1]  -29.69    0.06 3.62  -37.66  -31.94  -29.40  -27.10  -23.56
#> log_p_z[36,2]  -21.95    0.04 2.27  -26.98  -23.38  -21.78  -20.34  -18.05
#> log_p_z[36,3]  -10.56    0.04 1.74  -14.94  -11.42  -10.28   -9.33   -7.98
#> log_p_z[36,4]  -13.87    0.06 3.19  -21.00  -15.83  -13.59  -11.44   -8.71
#> log_p_z[37,1]  -18.53    0.03 2.26  -23.56  -19.86  -18.37  -16.97  -14.53
#> log_p_z[37,2]   -7.62    0.01 0.79   -9.38   -8.09   -7.57   -7.06   -6.26
#> log_p_z[37,3]  -19.25    0.06 3.34  -26.69  -21.19  -18.88  -16.87  -13.87
#> log_p_z[37,4]  -21.83    0.06 3.49  -29.88  -23.97  -21.50  -19.40  -15.93
#> log_p_z[38,1]  -25.23    0.04 2.78  -31.25  -26.89  -25.03  -23.28  -20.35
#> log_p_z[38,2]   -9.61    0.01 0.97  -11.76  -10.19   -9.53   -8.94   -7.99
#> log_p_z[38,3]  -11.78    0.05 2.82  -18.28  -13.37  -11.40   -9.75   -7.44
#> log_p_z[38,4]  -21.28    0.06 3.51  -29.14  -23.36  -21.00  -18.90  -15.18
#> log_p_z[39,1]  -29.24    0.06 4.30  -39.42  -31.79  -28.68  -26.29  -22.25
#> log_p_z[39,2]  -16.87    0.03 2.07  -21.29  -18.09  -16.75  -15.41  -13.32
#> log_p_z[39,3]   -7.29    0.02 1.12   -9.66   -8.00   -7.21   -6.48   -5.32
#> log_p_z[39,4]  -18.61    0.08 4.12  -28.44  -21.05  -18.13  -15.67  -12.14
#> log_p_z[40,1]   -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_p_z[40,2]  -19.99    0.03 2.28  -24.96  -21.39  -19.79  -18.40  -16.03
#> log_p_z[40,3]  -24.84    0.07 4.32  -34.67  -27.43  -24.28  -21.75  -17.84
#> log_p_z[40,4]  -21.91    0.07 4.39  -32.01  -24.48  -21.33  -18.83  -14.76
#> log_p_z[41,1]   -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_p_z[41,2]  -19.99    0.03 2.28  -24.96  -21.39  -19.79  -18.40  -16.03
#> log_p_z[41,3]  -24.84    0.07 4.32  -34.67  -27.43  -24.28  -21.75  -17.84
#> log_p_z[41,4]  -21.91    0.07 4.39  -32.01  -24.48  -21.33  -18.83  -14.76
#> log_p_z[42,1]   -6.00    0.01 0.60   -7.27   -6.36   -5.95   -5.59   -4.93
#> log_p_z[42,2]  -14.93    0.03 1.68  -18.69  -15.99  -14.76  -13.74  -12.05
#> log_p_z[42,3]  -24.37    0.07 3.57  -32.09  -26.55  -24.09  -21.83  -18.32
#> log_p_z[42,4]  -21.93    0.06 3.62  -29.92  -24.21  -21.55  -19.39  -15.75
#> log_p_z[43,1]  -21.80    0.04 2.39  -26.96  -23.31  -21.64  -20.11  -17.65
#> log_p_z[43,2]   -5.66    0.01 0.54   -6.86   -5.99   -5.62   -5.28   -4.69
#> log_p_z[43,3]  -18.49    0.06 3.29  -25.98  -20.41  -18.08  -16.16  -13.15
#> log_p_z[43,4]  -21.52    0.06 3.53  -29.48  -23.59  -21.18  -18.98  -15.65
#> log_p_z[44,1]   -4.42    0.00 0.39   -5.28   -4.65   -4.39   -4.15   -3.73
#> log_p_z[44,2]  -17.23    0.03 1.83  -21.17  -18.38  -17.10  -15.97  -14.00
#> log_p_z[44,3]  -24.34    0.06 3.57  -32.20  -26.55  -24.04  -21.83  -18.14
#> log_p_z[44,4]  -21.92    0.06 3.60  -29.75  -24.10  -21.59  -19.37  -15.83
#> log_p_z[45,1]  -19.59    0.03 2.31  -24.44  -21.08  -19.46  -17.99  -15.47
#> log_p_z[45,2]   -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.79   -2.40
#> log_p_z[45,3]  -21.35    0.08 4.44  -31.35  -23.80  -20.76  -18.20  -14.27
#> log_p_z[45,4]  -21.89    0.07 4.33  -31.90  -24.36  -21.43  -18.80  -14.93
#> log_lik[1]      -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_lik[2]      -7.58    0.02 1.16  -10.12   -8.28   -7.50   -6.74   -5.58
#> log_lik[3]      -9.93    0.01 0.91  -11.85  -10.51   -9.90   -9.30   -8.26
#> log_lik[4]      -7.55    0.01 0.93   -9.58   -8.10   -7.47   -6.91   -5.95
#> log_lik[5]      -3.71    0.01 0.44   -4.61   -3.99   -3.69   -3.41   -2.91
#> log_lik[6]      -5.47    0.01 0.66   -6.88   -5.87   -5.42   -5.01   -4.32
#> log_lik[7]      -5.99    0.01 0.60   -7.26   -6.36   -5.95   -5.59   -4.93
#> log_lik[8]      -5.73    0.01 0.84   -7.49   -6.27   -5.68   -5.12   -4.21
#> log_lik[9]      -4.51    0.01 0.57   -5.77   -4.87   -4.47   -4.11   -3.51
#> log_lik[10]     -7.13    0.01 0.71   -8.68   -7.57   -7.08   -6.63   -5.87
#> log_lik[11]     -5.83    0.01 0.98   -8.01   -6.42   -5.74   -5.15   -4.10
#> log_lik[12]     -9.35    0.01 1.06  -11.62  -10.02   -9.28   -8.63   -7.44
#> log_lik[13]     -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_lik[14]     -6.24    0.01 0.90   -8.16   -6.79   -6.15   -5.61   -4.71
#> log_lik[15]     -4.42    0.00 0.39   -5.28   -4.65   -4.39   -4.15   -3.73
#> log_lik[16]     -3.79    0.01 0.54   -5.00   -4.12   -3.75   -3.43   -2.86
#> log_lik[17]     -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_lik[18]     -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_lik[19]     -4.65    0.01 0.59   -5.89   -5.03   -4.62   -4.22   -3.62
#> log_lik[20]     -7.10    0.01 0.93   -9.10   -7.64   -7.03   -6.45   -5.47
#> log_lik[21]     -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.79   -2.40
#> log_lik[22]     -4.65    0.01 0.59   -5.89   -5.03   -4.62   -4.22   -3.62
#> log_lik[23]     -3.71    0.01 0.44   -4.61   -3.99   -3.69   -3.41   -2.91
#> log_lik[24]     -5.80    0.01 0.62   -7.22   -6.16   -5.75   -5.37   -4.77
#> log_lik[25]     -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_lik[26]     -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_lik[27]     -5.66    0.01 0.54   -6.86   -5.99   -5.62   -5.28   -4.69
#> log_lik[28]     -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_lik[29]     -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_lik[30]     -6.81    0.01 0.83   -8.71   -7.29   -6.75   -6.22   -5.43
#> log_lik[31]     -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_lik[32]     -6.05    0.01 0.95   -8.19   -6.62   -5.97   -5.39   -4.45
#> log_lik[33]     -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_lik[34]     -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.79   -2.40
#> log_lik[35]     -5.52    0.01 0.69   -7.05   -5.94   -5.47   -5.05   -4.33
#> log_lik[36]     -9.94    0.02 1.18  -12.56  -10.66   -9.87   -9.09   -7.89
#> log_lik[37]     -7.61    0.01 0.79   -9.38   -8.09   -7.57   -7.06   -6.26
#> log_lik[38]     -9.03    0.01 0.85  -10.73   -9.59   -9.04   -8.45   -7.31
#> log_lik[39]     -7.28    0.01 1.11   -9.64   -8.00   -7.21   -6.48   -5.32
#> log_lik[40]     -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_lik[41]     -2.22    0.00 0.31   -2.88   -2.41   -2.20   -2.00   -1.64
#> log_lik[42]     -5.99    0.01 0.60   -7.26   -6.36   -5.95   -5.59   -4.93
#> log_lik[43]     -5.66    0.01 0.54   -6.86   -5.99   -5.62   -5.28   -4.69
#> log_lik[44]     -4.42    0.00 0.39   -5.28   -4.65   -4.39   -4.15   -3.73
#> log_lik[45]     -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.79   -2.40
#> lp__          -464.08    0.17 6.18 -477.46 -468.00 -463.64 -459.77 -452.95
#>               n_eff Rhat
#> pi[1]          8517    1
#> pi[2]          7372    1
#> pi[3]          6244    1
#> pi[4]          5527    1
#> theta[1,1,1]   8099    1
#> theta[1,1,2]   7426    1
#> theta[1,1,3]   4106    1
#> theta[1,1,4]   4811    1
#> theta[1,2,1]   6547    1
#> theta[1,2,2]   7620    1
#> theta[1,2,3]   6537    1
#> theta[1,2,4]   5101    1
#> theta[1,3,1]   5582    1
#> theta[1,3,2]   4087    1
#> theta[1,3,3]   4730    1
#> theta[1,3,4]   5258    1
#> theta[1,4,1]   6094    1
#> theta[1,4,2]   6548    1
#> theta[1,4,3]   3988    1
#> theta[1,4,4]   5264    1
#> theta[2,1,1]   8388    1
#> theta[2,1,2]   6967    1
#> theta[2,1,3]   5020    1
#> theta[2,1,4]   4689    1
#> theta[2,2,1]   7511    1
#> theta[2,2,2]   7327    1
#> theta[2,2,3]   7784    1
#> theta[2,2,4]   5476    1
#> theta[2,3,1]   5665    1
#> theta[2,3,2]   6114    1
#> theta[2,3,3]   6009    1
#> theta[2,3,4]   5786    1
#> theta[2,4,1]   6355    1
#> theta[2,4,2]   6590    1
#> theta[2,4,3]   6211    1
#> theta[2,4,4]   6560    1
#> theta[3,1,1]   7742    1
#> theta[3,1,2]   5374    1
#> theta[3,1,3]   4712    1
#> theta[3,1,4]   4767    1
#> theta[3,2,1]   6380    1
#> theta[3,2,2]   7737    1
#> theta[3,2,3]   6605    1
#> theta[3,2,4]   5540    1
#> theta[3,3,1]   6466    1
#> theta[3,3,2]   6448    1
#> theta[3,3,3]   6434    1
#> theta[3,3,4]   7503    1
#> theta[3,4,1]   7587    1
#> theta[3,4,2]   6712    1
#> theta[3,4,3]   4691    1
#> theta[3,4,4]   5894    1
#> theta[4,1,1]   7956    1
#> theta[4,1,2]   5761    1
#> theta[4,1,3]   4908    1
#> theta[4,1,4]   3963    1
#> theta[4,2,1]   7145    1
#> theta[4,2,2]   8115    1
#> theta[4,2,3]   7220    1
#> theta[4,2,4]   5287    1
#> theta[4,3,1]   6176    1
#> theta[4,3,2]   6906    1
#> theta[4,3,3]   6528    1
#> theta[4,3,4]   5464    1
#> theta[4,4,1]   6657    1
#> theta[4,4,2]   6390    1
#> theta[4,4,3]   6152    1
#> theta[4,4,4]   6861    1
#> theta[5,1,1]   7240    1
#> theta[5,1,2]   4860    1
#> theta[5,1,3]   4797    1
#> theta[5,1,4]   5055    1
#> theta[5,2,1]   8472    1
#> theta[5,2,2]   7121    1
#> theta[5,2,3]   6405    1
#> theta[5,2,4]   5851    1
#> theta[5,3,1]   7251    1
#> theta[5,3,2]   7359    1
#> theta[5,3,3]   6751    1
#> theta[5,3,4]   7372    1
#> theta[5,4,1]   7032    1
#> theta[5,4,2]   5695    1
#> theta[5,4,3]   5816    1
#> theta[5,4,4]   4974    1
#> log_p_z[1,1]   7512    1
#> log_p_z[1,2]   4620    1
#> log_p_z[1,3]   3426    1
#> log_p_z[1,4]   3651    1
#> log_p_z[2,1]   4581    1
#> log_p_z[2,2]   3671    1
#> log_p_z[2,3]   3110    1
#> log_p_z[2,4]   3077    1
#> log_p_z[3,1]   3749    1
#> log_p_z[3,2]   4205    1
#> log_p_z[3,3]   2923    1
#> log_p_z[3,4]   3444    1
#> log_p_z[4,1]   4720    1
#> log_p_z[4,2]   4664    1
#> log_p_z[4,3]   3183    1
#> log_p_z[4,4]   3561    1
#> log_p_z[5,1]   4725    1
#> log_p_z[5,2]   6940    1
#> log_p_z[5,3]   3445    1
#> log_p_z[5,4]   3249    1
#> log_p_z[6,1]   4004    1
#> log_p_z[6,2]   6009    1
#> log_p_z[6,3]   3420    1
#> log_p_z[6,4]   3268    1
#> log_p_z[7,1]   6161    1
#> log_p_z[7,2]   4453    1
#> log_p_z[7,3]   2989    1
#> log_p_z[7,4]   3345    1
#> log_p_z[8,1]   4257    1
#> log_p_z[8,2]   4546    1
#> log_p_z[8,3]   6970    1
#> log_p_z[8,4]   2931    1
#> log_p_z[9,1]   5135    1
#> log_p_z[9,2]   6655    1
#> log_p_z[9,3]   3407    1
#> log_p_z[9,4]   3460    1
#> log_p_z[10,1]  4738    1
#> log_p_z[10,2]  5938    1
#> log_p_z[10,3]  3445    1
#> log_p_z[10,4]  3253    1
#> log_p_z[11,1]  4417    1
#> log_p_z[11,2]  3742    1
#> log_p_z[11,3]  2500    1
#> log_p_z[11,4]  5830    1
#> log_p_z[12,1]  3664    1
#> log_p_z[12,2]  4068    1
#> log_p_z[12,3]  3333    1
#> log_p_z[12,4]  3303    1
#> log_p_z[13,1]  7512    1
#> log_p_z[13,2]  4620    1
#> log_p_z[13,3]  3426    1
#> log_p_z[13,4]  3651    1
#> log_p_z[14,1]  4859    1
#> log_p_z[14,2]  5638    1
#> log_p_z[14,3]  3048    1
#> log_p_z[14,4]  3207    1
#> log_p_z[15,1]  7791    1
#> log_p_z[15,2]  4695    1
#> log_p_z[15,3]  3418    1
#> log_p_z[15,4]  3402    1
#> log_p_z[16,1]  6310    1
#> log_p_z[16,2]  4548    1
#> log_p_z[16,3]  3170    1
#> log_p_z[16,4]  3609    1
#> log_p_z[17,1]  7512    1
#> log_p_z[17,2]  4620    1
#> log_p_z[17,3]  3426    1
#> log_p_z[17,4]  3651    1
#> log_p_z[18,1]  7512    1
#> log_p_z[18,2]  4620    1
#> log_p_z[18,3]  3426    1
#> log_p_z[18,4]  3651    1
#> log_p_z[19,1]  5484    1
#> log_p_z[19,2]  6677    1
#> log_p_z[19,3]  3270    1
#> log_p_z[19,4]  3402    1
#> log_p_z[20,1]  3926    1
#> log_p_z[20,2]  5354    1
#> log_p_z[20,3]  3462    1
#> log_p_z[20,4]  3195    1
#> log_p_z[21,1]  4616    1
#> log_p_z[21,2]  7770    1
#> log_p_z[21,3]  3383    1
#> log_p_z[21,4]  3418    1
#> log_p_z[22,1]  5484    1
#> log_p_z[22,2]  6677    1
#> log_p_z[22,3]  3270    1
#> log_p_z[22,4]  3402    1
#> log_p_z[23,1]  4725    1
#> log_p_z[23,2]  6940    1
#> log_p_z[23,3]  3445    1
#> log_p_z[23,4]  3249    1
#> log_p_z[24,1]  4556    1
#> log_p_z[24,2]  5137    1
#> log_p_z[24,3]  3376    1
#> log_p_z[24,4]  3167    1
#> log_p_z[25,1]  7512    1
#> log_p_z[25,2]  4620    1
#> log_p_z[25,3]  3426    1
#> log_p_z[25,4]  3651    1
#> log_p_z[26,1]  7512    1
#> log_p_z[26,2]  4620    1
#> log_p_z[26,3]  3426    1
#> log_p_z[26,4]  3651    1
#> log_p_z[27,1]  4579    1
#> log_p_z[27,2]  6151    1
#> log_p_z[27,3]  3428    1
#> log_p_z[27,4]  3254    1
#> log_p_z[28,1]  7512    1
#> log_p_z[28,2]  4620    1
#> log_p_z[28,3]  3426    1
#> log_p_z[28,4]  3651    1
#> log_p_z[29,1]  7512    1
#> log_p_z[29,2]  4620    1
#> log_p_z[29,3]  3426    1
#> log_p_z[29,4]  3651    1
#> log_p_z[30,1]  5359    1
#> log_p_z[30,2]  4435    1
#> log_p_z[30,3]  3249    1
#> log_p_z[30,4]  3383    1
#> log_p_z[31,1]  7512    1
#> log_p_z[31,2]  4620    1
#> log_p_z[31,3]  3426    1
#> log_p_z[31,4]  3651    1
#> log_p_z[32,1]  4114    1
#> log_p_z[32,2]  4962    1
#> log_p_z[32,3]  5223    1
#> log_p_z[32,4]  3023    1
#> log_p_z[33,1]  7512    1
#> log_p_z[33,2]  4620    1
#> log_p_z[33,3]  3426    1
#> log_p_z[33,4]  3651    1
#> log_p_z[34,1]  4616    1
#> log_p_z[34,2]  7770    1
#> log_p_z[34,3]  3383    1
#> log_p_z[34,4]  3418    1
#> log_p_z[35,1]  5217    1
#> log_p_z[35,2]  5979    1
#> log_p_z[35,3]  3186    1
#> log_p_z[35,4]  3326    1
#> log_p_z[36,1]  3117    1
#> log_p_z[36,2]  3542    1
#> log_p_z[36,3]  2010    1
#> log_p_z[36,4]  2758    1
#> log_p_z[37,1]  5031    1
#> log_p_z[37,2]  6105    1
#> log_p_z[37,3]  2814    1
#> log_p_z[37,4]  3195    1
#> log_p_z[38,1]  4477    1
#> log_p_z[38,2]  4595    1
#> log_p_z[38,3]  3351    1
#> log_p_z[38,4]  3237    1
#> log_p_z[39,1]  4388    1
#> log_p_z[39,2]  4627    1
#> log_p_z[39,3]  5542    1
#> log_p_z[39,4]  2978    1
#> log_p_z[40,1]  7512    1
#> log_p_z[40,2]  4620    1
#> log_p_z[40,3]  3426    1
#> log_p_z[40,4]  3651    1
#> log_p_z[41,1]  7512    1
#> log_p_z[41,2]  4620    1
#> log_p_z[41,3]  3426    1
#> log_p_z[41,4]  3651    1
#> log_p_z[42,1]  6161    1
#> log_p_z[42,2]  4453    1
#> log_p_z[42,3]  2989    1
#> log_p_z[42,4]  3345    1
#> log_p_z[43,1]  4579    1
#> log_p_z[43,2]  6151    1
#> log_p_z[43,3]  3428    1
#> log_p_z[43,4]  3254    1
#> log_p_z[44,1]  7791    1
#> log_p_z[44,2]  4695    1
#> log_p_z[44,3]  3418    1
#> log_p_z[44,4]  3402    1
#> log_p_z[45,1]  4616    1
#> log_p_z[45,2]  7770    1
#> log_p_z[45,3]  3383    1
#> log_p_z[45,4]  3418    1
#> log_lik[1]     7512    1
#> log_lik[2]     3979    1
#> log_lik[3]     4753    1
#> log_lik[4]     4705    1
#> log_lik[5]     6940    1
#> log_lik[6]     6014    1
#> log_lik[7]     6164    1
#> log_lik[8]     6959    1
#> log_lik[9]     6656    1
#> log_lik[10]    5936    1
#> log_lik[11]    5830    1
#> log_lik[12]    5444    1
#> log_lik[13]    7512    1
#> log_lik[14]    5642    1
#> log_lik[15]    7791    1
#> log_lik[16]    6310    1
#> log_lik[17]    7512    1
#> log_lik[18]    7512    1
#> log_lik[19]    6677    1
#> log_lik[20]    5350    1
#> log_lik[21]    7770    1
#> log_lik[22]    6677    1
#> log_lik[23]    6940    1
#> log_lik[24]    5138    1
#> log_lik[25]    7512    1
#> log_lik[26]    7512    1
#> log_lik[27]    6152    1
#> log_lik[28]    7512    1
#> log_lik[29]    7512    1
#> log_lik[30]    5383    1
#> log_lik[31]    7512    1
#> log_lik[32]    5231    1
#> log_lik[33]    7512    1
#> log_lik[34]    7770    1
#> log_lik[35]    6014    1
#> log_lik[36]    3798    1
#> log_lik[37]    6113    1
#> log_lik[38]    4436    1
#> log_lik[39]    5572    1
#> log_lik[40]    7512    1
#> log_lik[41]    7512    1
#> log_lik[42]    6164    1
#> log_lik[43]    6152    1
#> log_lik[44]    7791    1
#> log_lik[45]    7770    1
#> lp__           1340    1
#> 
#> Samples were drawn using NUTS(diag_e) at Mon Sep 11 13:50:04 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }