Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.07    0.25    0.33    0.37    0.42    0.51
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.09    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.08    0.10    0.12    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.83    0.86    0.89    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.11    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.59    0.74    0.81    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.23
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.29
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.40
#> theta[1,4,4]     0.69    0.00 0.14    0.39    0.59    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.69    0.76    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.58    0.65    0.75
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.24    0.30    0.36    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[2,3,3]     0.62    0.00 0.13    0.35    0.54    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.29    0.46
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.55    0.67    0.76    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.74    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.68    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[3,3,2]     0.18    0.00 0.10    0.03    0.10    0.16    0.24    0.41
#> theta[3,3,3]     0.51    0.00 0.13    0.24    0.41    0.51    0.60    0.76
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.30    0.50
#> theta[3,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[3,4,3]     0.14    0.00 0.12    0.01    0.05    0.11    0.20    0.44
#> theta[3,4,4]     0.63    0.00 0.15    0.31    0.52    0.64    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.08    0.65    0.78    0.84    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.24
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[4,2,2]     0.72    0.00 0.09    0.54    0.67    0.73    0.79    0.87
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.09    0.13    0.17    0.29
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.10    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.02    0.06    0.11    0.28
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.39
#> theta[4,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,2]     0.12    0.00 0.10    0.01    0.04    0.09    0.16    0.37
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.91
#> theta[5,1,1]     0.86    0.00 0.07    0.70    0.82    0.87    0.92    0.97
#> theta[5,1,2]     0.06    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.33
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.51    0.61    0.70    0.84
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.40
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[5,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.41
#> theta[5,4,4]     0.63    0.00 0.15    0.32    0.52    0.64    0.75    0.90
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_p_z[1,2]   -20.04    0.04 2.29  -25.06  -21.43  -19.84  -18.42  -16.08
#> log_p_z[1,3]   -24.73    0.07 4.31  -34.55  -27.31  -24.29  -21.62  -17.67
#> log_p_z[1,4]   -21.87    0.07 4.19  -31.77  -24.22  -21.43  -18.91  -15.01
#> log_p_z[2,1]   -29.86    0.07 4.35  -39.77  -32.39  -29.36  -26.87  -22.72
#> log_p_z[2,2]   -21.03    0.03 2.37  -26.15  -22.52  -20.92  -19.34  -16.90
#> log_p_z[2,3]    -7.65    0.02 1.22  -10.28   -8.38   -7.51   -6.82   -5.59
#> log_p_z[2,4]   -16.31    0.08 4.26  -26.37  -18.77  -15.77  -13.34   -9.64
#> log_p_z[3,1]   -11.64    0.03 1.67  -15.44  -12.67  -11.46  -10.45   -8.84
#> log_p_z[3,2]   -10.82    0.02 1.47  -14.21  -11.70  -10.62   -9.78   -8.39
#> log_p_z[3,3]   -23.37    0.05 3.40  -30.94  -25.42  -23.08  -20.97  -17.68
#> log_p_z[3,4]   -21.85    0.06 3.43  -29.53  -23.84  -21.54  -19.41  -16.15
#> log_p_z[4,1]   -14.62    0.02 1.86  -18.59  -15.79  -14.51  -13.33  -11.35
#> log_p_z[4,2]    -7.58    0.01 0.97   -9.70   -8.19   -7.50   -6.90   -5.88
#> log_p_z[4,3]   -20.85    0.08 4.34  -30.77  -23.40  -20.37  -17.70  -13.83
#> log_p_z[4,4]   -21.85    0.07 4.30  -31.83  -24.33  -21.44  -18.81  -14.81
#> log_p_z[5,1]   -21.44    0.04 2.57  -27.01  -23.01  -21.27  -19.62  -16.95
#> log_p_z[5,2]    -3.71    0.00 0.43   -4.59   -3.98   -3.69   -3.41   -2.94
#> log_p_z[5,3]   -18.85    0.08 4.29  -28.84  -21.42  -18.38  -15.72  -12.18
#> log_p_z[5,4]   -21.86    0.07 4.25  -31.68  -24.28  -21.39  -18.96  -14.83
#> log_p_z[6,1]   -21.46    0.04 2.56  -27.00  -23.05  -21.30  -19.66  -16.81
#> log_p_z[6,2]    -5.46    0.01 0.66   -6.87   -5.88   -5.42   -5.01   -4.31
#> log_p_z[6,3]   -17.63    0.07 4.16  -27.20  -19.97  -17.14  -14.66  -11.19
#> log_p_z[6,4]   -21.62    0.07 4.21  -31.42  -24.03  -21.14  -18.69  -14.74
#> log_p_z[7,1]    -6.01    0.01 0.63   -7.37   -6.41   -5.97   -5.57   -4.91
#> log_p_z[7,2]   -14.97    0.03 1.72  -18.68  -16.01  -14.78  -13.76  -12.04
#> log_p_z[7,3]   -24.26    0.06 3.48  -32.03  -26.35  -23.99  -21.78  -18.33
#> log_p_z[7,4]   -21.88    0.06 3.50  -29.88  -23.92  -21.57  -19.45  -15.86
#> log_p_z[8,1]   -29.85    0.07 4.40  -39.99  -32.39  -29.37  -26.68  -22.72
#> log_p_z[8,2]   -18.31    0.03 2.19  -23.07  -19.68  -18.16  -16.79  -14.43
#> log_p_z[8,3]    -5.73    0.01 0.85   -7.57   -6.26   -5.68   -5.13   -4.18
#> log_p_z[8,4]   -18.53    0.08 4.35  -28.69  -21.05  -18.06  -15.44  -11.61
#> log_p_z[9,1]   -19.98    0.03 2.34  -25.25  -21.35  -19.79  -18.36  -15.85
#> log_p_z[9,2]    -4.51    0.01 0.58   -5.76   -4.87   -4.48   -4.10   -3.51
#> log_p_z[9,3]   -19.82    0.08 4.40  -29.73  -22.50  -19.29  -16.62  -12.91
#> log_p_z[9,4]   -21.61    0.06 4.15  -31.33  -24.04  -21.12  -18.69  -14.89
#> log_p_z[10,1]  -22.29    0.03 2.46  -27.84  -23.80  -22.07  -20.56  -18.03
#> log_p_z[10,2]   -7.13    0.01 0.71   -8.67   -7.58   -7.06   -6.63   -5.86
#> log_p_z[10,3]  -16.96    0.06 3.24  -24.03  -18.90  -16.62  -14.63  -11.72
#> log_p_z[10,4]  -21.29    0.06 3.44  -28.91  -23.38  -20.98  -18.85  -15.56
#> log_p_z[11,1]  -29.70    0.06 4.35  -39.72  -32.18  -29.20  -26.66  -22.55
#> log_p_z[11,2]  -29.39    0.07 4.33  -38.92  -31.99  -28.90  -26.26  -22.34
#> log_p_z[11,3]  -18.24    0.06 3.45  -26.23  -20.12  -17.74  -15.85  -12.99
#> log_p_z[11,4]   -5.82    0.01 0.98   -7.97   -6.41   -5.74   -5.14   -4.11
#> log_p_z[12,1]  -22.93    0.04 2.67  -28.71  -24.57  -22.72  -21.07  -18.22
#> log_p_z[12,2]   -9.57    0.02 1.25  -12.43  -10.27   -9.44   -8.71   -7.50
#> log_p_z[12,3]  -15.28    0.07 4.11  -24.85  -17.65  -14.74  -12.32   -8.82
#> log_p_z[12,4]  -19.21    0.06 4.11  -28.57  -21.57  -18.65  -16.28  -12.75
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_p_z[13,2]  -20.04    0.04 2.29  -25.06  -21.43  -19.84  -18.42  -16.08
#> log_p_z[13,3]  -24.73    0.07 4.31  -34.55  -27.31  -24.29  -21.62  -17.67
#> log_p_z[13,4]  -21.87    0.07 4.19  -31.77  -24.22  -21.43  -18.91  -15.01
#> log_p_z[14,1]  -19.04    0.03 2.40  -24.29  -20.44  -18.86  -17.33  -14.97
#> log_p_z[14,2]   -6.24    0.01 0.88   -8.15   -6.75   -6.16   -5.63   -4.77
#> log_p_z[14,3]  -18.79    0.08 4.25  -28.50  -21.26  -18.26  -15.74  -12.15
#> log_p_z[14,4]  -21.92    0.07 4.29  -31.69  -24.44  -21.37  -18.95  -14.89
#> log_p_z[15,1]   -4.41    0.00 0.40   -5.29   -4.66   -4.39   -4.14   -3.70
#> log_p_z[15,2]  -17.27    0.03 1.84  -21.26  -18.41  -17.11  -15.96  -14.10
#> log_p_z[15,3]  -24.25    0.06 3.54  -32.04  -26.42  -23.96  -21.81  -18.23
#> log_p_z[15,4]  -21.88    0.06 3.53  -29.76  -23.98  -21.55  -19.42  -15.98
#> log_p_z[16,1]   -3.80    0.01 0.56   -5.03   -4.14   -3.76   -3.42   -2.85
#> log_p_z[16,2]  -17.74    0.03 2.20  -22.63  -19.05  -17.50  -16.20  -14.05
#> log_p_z[16,3]  -24.74    0.07 4.26  -34.13  -27.19  -24.27  -21.70  -17.68
#> log_p_z[16,4]  -21.87    0.07 4.17  -31.63  -24.13  -21.43  -18.94  -15.13
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_p_z[17,2]  -20.04    0.04 2.29  -25.06  -21.43  -19.84  -18.42  -16.08
#> log_p_z[17,3]  -24.73    0.07 4.31  -34.55  -27.31  -24.29  -21.62  -17.67
#> log_p_z[17,4]  -21.87    0.07 4.19  -31.77  -24.22  -21.43  -18.91  -15.01
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_p_z[18,2]  -20.04    0.04 2.29  -25.06  -21.43  -19.84  -18.42  -16.08
#> log_p_z[18,3]  -24.73    0.07 4.31  -34.55  -27.31  -24.29  -21.62  -17.67
#> log_p_z[18,4]  -21.87    0.07 4.19  -31.77  -24.22  -21.43  -18.91  -15.01
#> log_p_z[19,1]  -16.41    0.03 2.00  -20.88  -17.61  -16.25  -15.02  -12.94
#> log_p_z[19,2]   -4.67    0.01 0.61   -5.97   -5.03   -4.62   -4.25   -3.58
#> log_p_z[19,3]  -22.30    0.08 4.46  -32.43  -25.10  -21.84  -19.09  -15.05
#> log_p_z[19,4]  -21.84    0.06 4.17  -31.40  -24.24  -21.44  -18.88  -15.06
#> log_p_z[20,1]  -18.06    0.03 2.31  -22.92  -19.52  -17.88  -16.43  -13.98
#> log_p_z[20,2]   -7.09    0.01 0.94   -9.10   -7.67   -7.03   -6.43   -5.46
#> log_p_z[20,3]  -20.14    0.08 4.30  -29.99  -22.71  -19.70  -16.96  -13.42
#> log_p_z[20,4]  -21.62    0.07 4.20  -31.19  -24.06  -21.05  -18.68  -14.79
#> log_p_z[21,1]  -19.64    0.03 2.41  -25.01  -21.02  -19.51  -17.94  -15.47
#> log_p_z[21,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.03   -2.79   -2.38
#> log_p_z[21,3]  -21.36    0.08 4.42  -31.47  -24.04  -20.86  -18.20  -14.24
#> log_p_z[21,4]  -21.86    0.06 4.18  -31.47  -24.24  -21.44  -18.96  -15.07
#> log_p_z[22,1]  -16.41    0.03 2.00  -20.88  -17.61  -16.25  -15.02  -12.94
#> log_p_z[22,2]   -4.67    0.01 0.61   -5.97   -5.03   -4.62   -4.25   -3.58
#> log_p_z[22,3]  -22.30    0.08 4.46  -32.43  -25.10  -21.84  -19.09  -15.05
#> log_p_z[22,4]  -21.84    0.06 4.17  -31.40  -24.24  -21.44  -18.88  -15.06
#> log_p_z[23,1]  -21.44    0.04 2.57  -27.01  -23.01  -21.27  -19.62  -16.95
#> log_p_z[23,2]   -3.71    0.00 0.43   -4.59   -3.98   -3.69   -3.41   -2.94
#> log_p_z[23,3]  -18.85    0.08 4.29  -28.84  -21.42  -18.38  -15.72  -12.18
#> log_p_z[23,4]  -21.86    0.07 4.25  -31.68  -24.28  -21.39  -18.96  -14.83
#> log_p_z[24,1]  -17.44    0.03 2.20  -22.32  -18.71  -17.29  -15.89  -13.72
#> log_p_z[24,2]   -5.81    0.01 0.65   -7.28   -6.18   -5.73   -5.35   -4.73
#> log_p_z[24,3]  -21.84    0.06 3.48  -29.61  -23.96  -21.56  -19.35  -16.06
#> log_p_z[24,4]  -21.85    0.05 3.45  -29.47  -23.97  -21.56  -19.44  -16.07
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_p_z[25,2]  -20.04    0.04 2.29  -25.06  -21.43  -19.84  -18.42  -16.08
#> log_p_z[25,3]  -24.73    0.07 4.31  -34.55  -27.31  -24.29  -21.62  -17.67
#> log_p_z[25,4]  -21.87    0.07 4.19  -31.77  -24.22  -21.43  -18.91  -15.01
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_p_z[26,2]  -20.04    0.04 2.29  -25.06  -21.43  -19.84  -18.42  -16.08
#> log_p_z[26,3]  -24.73    0.07 4.31  -34.55  -27.31  -24.29  -21.62  -17.67
#> log_p_z[26,4]  -21.87    0.07 4.19  -31.77  -24.22  -21.43  -18.91  -15.01
#> log_p_z[27,1]  -21.95    0.04 2.49  -27.63  -23.42  -21.78  -20.22  -17.66
#> log_p_z[27,2]   -5.65    0.01 0.53   -6.79   -5.99   -5.62   -5.27   -4.72
#> log_p_z[27,3]  -18.50    0.06 3.29  -25.67  -20.48  -18.18  -16.14  -13.03
#> log_p_z[27,4]  -21.55    0.06 3.43  -29.28  -23.62  -21.26  -19.08  -15.75
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_p_z[28,2]  -20.04    0.04 2.29  -25.06  -21.43  -19.84  -18.42  -16.08
#> log_p_z[28,3]  -24.73    0.07 4.31  -34.55  -27.31  -24.29  -21.62  -17.67
#> log_p_z[28,4]  -21.87    0.07 4.19  -31.77  -24.22  -21.43  -18.91  -15.01
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_p_z[29,2]  -20.04    0.04 2.29  -25.06  -21.43  -19.84  -18.42  -16.08
#> log_p_z[29,3]  -24.73    0.07 4.31  -34.55  -27.31  -24.29  -21.62  -17.67
#> log_p_z[29,4]  -21.87    0.07 4.19  -31.77  -24.22  -21.43  -18.91  -15.01
#> log_p_z[30,1]   -6.81    0.01 0.87   -8.77   -7.34   -6.72   -6.19   -5.39
#> log_p_z[30,2]  -14.75    0.03 1.68  -18.47  -15.78  -14.62  -13.54  -11.88
#> log_p_z[30,3]  -24.31    0.06 3.55  -32.12  -26.54  -23.98  -21.80  -18.20
#> log_p_z[30,4]  -21.83    0.06 3.50  -29.51  -24.00  -21.46  -19.35  -15.94
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_p_z[31,2]  -20.04    0.04 2.29  -25.06  -21.43  -19.84  -18.42  -16.08
#> log_p_z[31,3]  -24.73    0.07 4.31  -34.55  -27.31  -24.29  -21.62  -17.67
#> log_p_z[31,4]  -21.87    0.07 4.19  -31.77  -24.22  -21.43  -18.91  -15.01
#> log_p_z[32,1]  -29.86    0.07 4.37  -39.47  -32.44  -29.49  -26.72  -22.58
#> log_p_z[32,2]  -14.84    0.02 1.78  -18.73  -15.95  -14.70  -13.61  -11.71
#> log_p_z[32,3]   -6.08    0.01 0.95   -8.11   -6.66   -6.02   -5.41   -4.42
#> log_p_z[32,4]  -20.66    0.08 4.49  -30.87  -23.27  -20.13  -17.44  -13.30
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_p_z[33,2]  -20.04    0.04 2.29  -25.06  -21.43  -19.84  -18.42  -16.08
#> log_p_z[33,3]  -24.73    0.07 4.31  -34.55  -27.31  -24.29  -21.62  -17.67
#> log_p_z[33,4]  -21.87    0.07 4.19  -31.77  -24.22  -21.43  -18.91  -15.01
#> log_p_z[34,1]  -19.64    0.03 2.41  -25.01  -21.02  -19.51  -17.94  -15.47
#> log_p_z[34,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.03   -2.79   -2.38
#> log_p_z[34,3]  -21.36    0.08 4.42  -31.47  -24.04  -20.86  -18.20  -14.24
#> log_p_z[34,4]  -21.86    0.06 4.18  -31.47  -24.24  -21.44  -18.96  -15.07
#> log_p_z[35,1]  -22.59    0.04 2.68  -28.49  -24.28  -22.37  -20.70  -17.91
#> log_p_z[35,2]   -5.52    0.01 0.71   -7.08   -5.96   -5.46   -5.04   -4.26
#> log_p_z[35,3]  -16.21    0.07 4.13  -25.64  -18.58  -15.69  -13.17   -9.96
#> log_p_z[35,4]  -21.87    0.07 4.29  -31.80  -24.33  -21.33  -18.86  -14.61
#> log_p_z[36,1]  -29.81    0.06 3.64  -37.91  -31.93  -29.47  -27.30  -23.57
#> log_p_z[36,2]  -21.92    0.04 2.30  -26.74  -23.34  -21.81  -20.28  -17.90
#> log_p_z[36,3]  -10.54    0.03 1.75  -14.63  -11.44  -10.27   -9.31   -7.95
#> log_p_z[36,4]  -13.92    0.06 3.24  -21.18  -15.92  -13.64  -11.47   -8.79
#> log_p_z[37,1]  -18.59    0.03 2.32  -23.61  -19.97  -18.39  -16.93  -14.59
#> log_p_z[37,2]   -7.62    0.01 0.80   -9.37   -8.10   -7.55   -7.05   -6.24
#> log_p_z[37,3]  -19.20    0.06 3.25  -26.57  -21.16  -18.88  -16.85  -13.98
#> log_p_z[37,4]  -21.85    0.06 3.47  -29.61  -24.01  -21.51  -19.41  -15.95
#> log_p_z[38,1]  -25.24    0.04 2.76  -31.17  -26.99  -25.09  -23.29  -20.45
#> log_p_z[38,2]   -9.61    0.01 1.02  -11.85  -10.24   -9.54   -8.90   -7.82
#> log_p_z[38,3]  -11.81    0.05 2.82  -18.24  -13.47  -11.47   -9.75   -7.52
#> log_p_z[38,4]  -21.30    0.06 3.56  -29.26  -23.49  -20.91  -18.81  -15.35
#> log_p_z[39,1]  -29.51    0.07 4.36  -39.52  -32.04  -29.05  -26.43  -22.48
#> log_p_z[39,2]  -16.84    0.03 2.10  -21.52  -18.14  -16.68  -15.39  -13.17
#> log_p_z[39,3]   -7.27    0.01 1.11   -9.66   -7.95   -7.19   -6.49   -5.32
#> log_p_z[39,4]  -18.79    0.08 4.25  -28.53  -21.17  -18.26  -15.72  -11.96
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_p_z[40,2]  -20.04    0.04 2.29  -25.06  -21.43  -19.84  -18.42  -16.08
#> log_p_z[40,3]  -24.73    0.07 4.31  -34.55  -27.31  -24.29  -21.62  -17.67
#> log_p_z[40,4]  -21.87    0.07 4.19  -31.77  -24.22  -21.43  -18.91  -15.01
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_p_z[41,2]  -20.04    0.04 2.29  -25.06  -21.43  -19.84  -18.42  -16.08
#> log_p_z[41,3]  -24.73    0.07 4.31  -34.55  -27.31  -24.29  -21.62  -17.67
#> log_p_z[41,4]  -21.87    0.07 4.19  -31.77  -24.22  -21.43  -18.91  -15.01
#> log_p_z[42,1]   -6.01    0.01 0.63   -7.37   -6.41   -5.97   -5.57   -4.91
#> log_p_z[42,2]  -14.97    0.03 1.72  -18.68  -16.01  -14.78  -13.76  -12.04
#> log_p_z[42,3]  -24.26    0.06 3.48  -32.03  -26.35  -23.99  -21.78  -18.33
#> log_p_z[42,4]  -21.88    0.06 3.50  -29.88  -23.92  -21.57  -19.45  -15.86
#> log_p_z[43,1]  -21.95    0.04 2.49  -27.63  -23.42  -21.78  -20.22  -17.66
#> log_p_z[43,2]   -5.65    0.01 0.53   -6.79   -5.99   -5.62   -5.27   -4.72
#> log_p_z[43,3]  -18.50    0.06 3.29  -25.67  -20.48  -18.18  -16.14  -13.03
#> log_p_z[43,4]  -21.55    0.06 3.43  -29.28  -23.62  -21.26  -19.08  -15.75
#> log_p_z[44,1]   -4.41    0.00 0.40   -5.29   -4.66   -4.39   -4.14   -3.70
#> log_p_z[44,2]  -17.27    0.03 1.84  -21.26  -18.41  -17.11  -15.96  -14.10
#> log_p_z[44,3]  -24.25    0.06 3.54  -32.04  -26.42  -23.96  -21.81  -18.23
#> log_p_z[44,4]  -21.88    0.06 3.53  -29.76  -23.98  -21.55  -19.42  -15.98
#> log_p_z[45,1]  -19.64    0.03 2.41  -25.01  -21.02  -19.51  -17.94  -15.47
#> log_p_z[45,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.03   -2.79   -2.38
#> log_p_z[45,3]  -21.36    0.08 4.42  -31.47  -24.04  -20.86  -18.20  -14.24
#> log_p_z[45,4]  -21.86    0.06 4.18  -31.47  -24.24  -21.44  -18.96  -15.07
#> log_lik[1]      -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_lik[2]      -7.60    0.02 1.13   -9.99   -8.34   -7.50   -6.81   -5.59
#> log_lik[3]      -9.94    0.01 0.91  -11.82  -10.52   -9.91   -9.29   -8.26
#> log_lik[4]      -7.57    0.01 0.97   -9.69   -8.19   -7.50   -6.90   -5.88
#> log_lik[5]      -3.71    0.00 0.43   -4.59   -3.98   -3.69   -3.41   -2.94
#> log_lik[6]      -5.46    0.01 0.65   -6.87   -5.88   -5.42   -5.01   -4.31
#> log_lik[7]      -6.01    0.01 0.63   -7.37   -6.41   -5.96   -5.57   -4.91
#> log_lik[8]      -5.73    0.01 0.84   -7.57   -6.26   -5.68   -5.13   -4.18
#> log_lik[9]      -4.51    0.01 0.58   -5.76   -4.87   -4.48   -4.10   -3.51
#> log_lik[10]     -7.12    0.01 0.71   -8.67   -7.58   -7.06   -6.63   -5.85
#> log_lik[11]     -5.82    0.01 0.98   -7.97   -6.41   -5.74   -5.14   -4.11
#> log_lik[12]     -9.32    0.01 1.02  -11.44   -9.97   -9.26   -8.60   -7.44
#> log_lik[13]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_lik[14]     -6.24    0.01 0.88   -8.15   -6.75   -6.15   -5.63   -4.77
#> log_lik[15]     -4.41    0.00 0.40   -5.29   -4.66   -4.39   -4.14   -3.70
#> log_lik[16]     -3.80    0.01 0.56   -5.03   -4.14   -3.76   -3.42   -2.85
#> log_lik[17]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_lik[18]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_lik[19]     -4.67    0.01 0.61   -5.97   -5.03   -4.62   -4.25   -3.58
#> log_lik[20]     -7.09    0.01 0.94   -9.10   -7.67   -7.03   -6.43   -5.46
#> log_lik[21]     -3.04    0.00 0.36   -3.78   -3.27   -3.03   -2.79   -2.38
#> log_lik[22]     -4.67    0.01 0.61   -5.97   -5.03   -4.62   -4.25   -3.58
#> log_lik[23]     -3.71    0.00 0.43   -4.59   -3.98   -3.69   -3.41   -2.94
#> log_lik[24]     -5.81    0.01 0.65   -7.28   -6.18   -5.73   -5.35   -4.73
#> log_lik[25]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_lik[26]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_lik[27]     -5.65    0.01 0.53   -6.79   -5.99   -5.62   -5.26   -4.72
#> log_lik[28]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_lik[29]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_lik[30]     -6.81    0.01 0.87   -8.77   -7.34   -6.72   -6.19   -5.39
#> log_lik[31]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_lik[32]     -6.08    0.01 0.95   -8.09   -6.66   -6.02   -5.41   -4.42
#> log_lik[33]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_lik[34]     -3.04    0.00 0.36   -3.78   -3.27   -3.03   -2.79   -2.38
#> log_lik[35]     -5.52    0.01 0.71   -7.05   -5.96   -5.46   -5.04   -4.26
#> log_lik[36]     -9.91    0.02 1.18  -12.45  -10.64   -9.84   -9.09   -7.82
#> log_lik[37]     -7.62    0.01 0.80   -9.36   -8.10   -7.55   -7.05   -6.23
#> log_lik[38]     -9.02    0.01 0.88  -10.74   -9.60   -9.01   -8.43   -7.32
#> log_lik[39]     -7.26    0.01 1.11   -9.64   -7.94   -7.19   -6.49   -5.32
#> log_lik[40]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_lik[41]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.66
#> log_lik[42]     -6.01    0.01 0.63   -7.37   -6.41   -5.96   -5.57   -4.91
#> log_lik[43]     -5.65    0.01 0.53   -6.79   -5.99   -5.62   -5.26   -4.72
#> log_lik[44]     -4.41    0.00 0.40   -5.29   -4.66   -4.39   -4.14   -3.70
#> log_lik[45]     -3.04    0.00 0.36   -3.78   -3.27   -3.03   -2.79   -2.38
#> lp__          -463.96    0.17 6.25 -476.94 -468.02 -463.63 -459.49 -452.69
#>               n_eff Rhat
#> pi[1]          8007    1
#> pi[2]          8039    1
#> pi[3]          6678    1
#> pi[4]          6609    1
#> theta[1,1,1]   8188    1
#> theta[1,1,2]   7640    1
#> theta[1,1,3]   5069    1
#> theta[1,1,4]   4768    1
#> theta[1,2,1]   6583    1
#> theta[1,2,2]   6867    1
#> theta[1,2,3]   6520    1
#> theta[1,2,4]   5452    1
#> theta[1,3,1]   6325    1
#> theta[1,3,2]   4571    1
#> theta[1,3,3]   5148    1
#> theta[1,3,4]   6344    1
#> theta[1,4,1]   7104    1
#> theta[1,4,2]   7243    1
#> theta[1,4,3]   4214    1
#> theta[1,4,4]   5576    1
#> theta[2,1,1]   8570    1
#> theta[2,1,2]   7362    1
#> theta[2,1,3]   4085    1
#> theta[2,1,4]   4507    1
#> theta[2,2,1]   7219    1
#> theta[2,2,2]   8058    1
#> theta[2,2,3]   6886    1
#> theta[2,2,4]   5472    1
#> theta[2,3,1]   5716    1
#> theta[2,3,2]   6052    1
#> theta[2,3,3]   6648    1
#> theta[2,3,4]   5850    1
#> theta[2,4,1]   7577    1
#> theta[2,4,2]   6954    1
#> theta[2,4,3]   7464    1
#> theta[2,4,4]   6046    1
#> theta[3,1,1]   7600    1
#> theta[3,1,2]   5172    1
#> theta[3,1,3]   5479    1
#> theta[3,1,4]   4960    1
#> theta[3,2,1]   6409    1
#> theta[3,2,2]   7022    1
#> theta[3,2,3]   7174    1
#> theta[3,2,4]   5573    1
#> theta[3,3,1]   6468    1
#> theta[3,3,2]   7210    1
#> theta[3,3,3]   7167    1
#> theta[3,3,4]   6494    1
#> theta[3,4,1]   6445    1
#> theta[3,4,2]   7103    1
#> theta[3,4,3]   5013    1
#> theta[3,4,4]   5756    1
#> theta[4,1,1]   7451    1
#> theta[4,1,2]   5855    1
#> theta[4,1,3]   4343    1
#> theta[4,1,4]   5095    1
#> theta[4,2,1]   7706    1
#> theta[4,2,2]   7671    1
#> theta[4,2,3]   7008    1
#> theta[4,2,4]   6457    1
#> theta[4,3,1]   6171    1
#> theta[4,3,2]   7099    1
#> theta[4,3,3]   5601    1
#> theta[4,3,4]   4900    1
#> theta[4,4,1]   6350    1
#> theta[4,4,2]   6455    1
#> theta[4,4,3]   7291    1
#> theta[4,4,4]   7029    1
#> theta[5,1,1]   6717    1
#> theta[5,1,2]   4919    1
#> theta[5,1,3]   4684    1
#> theta[5,1,4]   4347    1
#> theta[5,2,1]   7649    1
#> theta[5,2,2]   7969    1
#> theta[5,2,3]   6454    1
#> theta[5,2,4]   5967    1
#> theta[5,3,1]   7287    1
#> theta[5,3,2]   7541    1
#> theta[5,3,3]   6784    1
#> theta[5,3,4]   6152    1
#> theta[5,4,1]   6870    1
#> theta[5,4,2]   6880    1
#> theta[5,4,3]   5820    1
#> theta[5,4,4]   7029    1
#> log_p_z[1,1]   7935    1
#> log_p_z[1,2]   4124    1
#> log_p_z[1,3]   3816    1
#> log_p_z[1,4]   3977    1
#> log_p_z[2,1]   4039    1
#> log_p_z[2,2]   4874    1
#> log_p_z[2,3]   4038    1
#> log_p_z[2,4]   3057    1
#> log_p_z[3,1]   4396    1
#> log_p_z[3,2]   4084    1
#> log_p_z[3,3]   4046    1
#> log_p_z[3,4]   3694    1
#> log_p_z[4,1]   6018    1
#> log_p_z[4,2]   4510    1
#> log_p_z[4,3]   3013    1
#> log_p_z[4,4]   4311    1
#> log_p_z[5,1]   5109    1
#> log_p_z[5,2]   7519    1
#> log_p_z[5,3]   3089    1
#> log_p_z[5,4]   4159    1
#> log_p_z[6,1]   4718    1
#> log_p_z[6,2]   7228    1
#> log_p_z[6,3]   3119    1
#> log_p_z[6,4]   4171    1
#> log_p_z[7,1]   6471    1
#> log_p_z[7,2]   3960    1
#> log_p_z[7,3]   3520    1
#> log_p_z[7,4]   3809    1
#> log_p_z[8,1]   4289    1
#> log_p_z[8,2]   5092    1
#> log_p_z[8,3]   5799    1
#> log_p_z[8,4]   3201    1
#> log_p_z[9,1]   5937    1
#> log_p_z[9,2]   6292    1
#> log_p_z[9,3]   3075    1
#> log_p_z[9,4]   4131    1
#> log_p_z[10,1]  5034    1
#> log_p_z[10,2]  5637    1
#> log_p_z[10,3]  3084    1
#> log_p_z[10,4]  3846    1
#> log_p_z[11,1]  4846    1
#> log_p_z[11,2]  3541    1
#> log_p_z[11,3]  3195    1
#> log_p_z[11,4]  5608    1
#> log_p_z[12,1]  5010    1
#> log_p_z[12,2]  4436    1
#> log_p_z[12,3]  3050    1
#> log_p_z[12,4]  4106    1
#> log_p_z[13,1]  7935    1
#> log_p_z[13,2]  4124    1
#> log_p_z[13,3]  3816    1
#> log_p_z[13,4]  3977    1
#> log_p_z[14,1]  5057    1
#> log_p_z[14,2]  4729    1
#> log_p_z[14,3]  3136    1
#> log_p_z[14,4]  3710    1
#> log_p_z[15,1]  7194    1
#> log_p_z[15,2]  3845    1
#> log_p_z[15,3]  3564    1
#> log_p_z[15,4]  3847    1
#> log_p_z[16,1]  6875    1
#> log_p_z[16,2]  4233    1
#> log_p_z[16,3]  3868    1
#> log_p_z[16,4]  3895    1
#> log_p_z[17,1]  7935    1
#> log_p_z[17,2]  4124    1
#> log_p_z[17,3]  3816    1
#> log_p_z[17,4]  3977    1
#> log_p_z[18,1]  7935    1
#> log_p_z[18,2]  4124    1
#> log_p_z[18,3]  3816    1
#> log_p_z[18,4]  3977    1
#> log_p_z[19,1]  6115    1
#> log_p_z[19,2]  5277    1
#> log_p_z[19,3]  2985    1
#> log_p_z[19,4]  4134    1
#> log_p_z[20,1]  4809    1
#> log_p_z[20,2]  5062    1
#> log_p_z[20,3]  3173    1
#> log_p_z[20,4]  4018    1
#> log_p_z[21,1]  5232    1
#> log_p_z[21,2]  7043    1
#> log_p_z[21,3]  3092    1
#> log_p_z[21,4]  4141    1
#> log_p_z[22,1]  6115    1
#> log_p_z[22,2]  5277    1
#> log_p_z[22,3]  2985    1
#> log_p_z[22,4]  4134    1
#> log_p_z[23,1]  5109    1
#> log_p_z[23,2]  7519    1
#> log_p_z[23,3]  3089    1
#> log_p_z[23,4]  4159    1
#> log_p_z[24,1]  4806    1
#> log_p_z[24,2]  5020    1
#> log_p_z[24,3]  3368    1
#> log_p_z[24,4]  3948    1
#> log_p_z[25,1]  7935    1
#> log_p_z[25,2]  4124    1
#> log_p_z[25,3]  3816    1
#> log_p_z[25,4]  3977    1
#> log_p_z[26,1]  7935    1
#> log_p_z[26,2]  4124    1
#> log_p_z[26,3]  3816    1
#> log_p_z[26,4]  3977    1
#> log_p_z[27,1]  4652    1
#> log_p_z[27,2]  6706    1
#> log_p_z[27,3]  3183    1
#> log_p_z[27,4]  3874    1
#> log_p_z[28,1]  7935    1
#> log_p_z[28,2]  4124    1
#> log_p_z[28,3]  3816    1
#> log_p_z[28,4]  3977    1
#> log_p_z[29,1]  7935    1
#> log_p_z[29,2]  4124    1
#> log_p_z[29,3]  3816    1
#> log_p_z[29,4]  3977    1
#> log_p_z[30,1]  6123    1
#> log_p_z[30,2]  3959    1
#> log_p_z[30,3]  3580    1
#> log_p_z[30,4]  3880    1
#> log_p_z[31,1]  7935    1
#> log_p_z[31,2]  4124    1
#> log_p_z[31,3]  3816    1
#> log_p_z[31,4]  3977    1
#> log_p_z[32,1]  4352    1
#> log_p_z[32,2]  5248    1
#> log_p_z[32,3]  5140    1
#> log_p_z[32,4]  3248    1
#> log_p_z[33,1]  7935    1
#> log_p_z[33,2]  4124    1
#> log_p_z[33,3]  3816    1
#> log_p_z[33,4]  3977    1
#> log_p_z[34,1]  5232    1
#> log_p_z[34,2]  7043    1
#> log_p_z[34,3]  3092    1
#> log_p_z[34,4]  4141    1
#> log_p_z[35,1]  5004    1
#> log_p_z[35,2]  5495    1
#> log_p_z[35,3]  3093    1
#> log_p_z[35,4]  3914    1
#> log_p_z[36,1]  3725    1
#> log_p_z[36,2]  3784    1
#> log_p_z[36,3]  2866    1
#> log_p_z[36,4]  3246    1
#> log_p_z[37,1]  4874    1
#> log_p_z[37,2]  5509    1
#> log_p_z[37,3]  3354    1
#> log_p_z[37,4]  3525    1
#> log_p_z[38,1]  4567    1
#> log_p_z[38,2]  4788    1
#> log_p_z[38,3]  3120    1
#> log_p_z[38,4]  3634    1
#> log_p_z[39,1]  4334    1
#> log_p_z[39,2]  5222    1
#> log_p_z[39,3]  6196    1
#> log_p_z[39,4]  3086    1
#> log_p_z[40,1]  7935    1
#> log_p_z[40,2]  4124    1
#> log_p_z[40,3]  3816    1
#> log_p_z[40,4]  3977    1
#> log_p_z[41,1]  7935    1
#> log_p_z[41,2]  4124    1
#> log_p_z[41,3]  3816    1
#> log_p_z[41,4]  3977    1
#> log_p_z[42,1]  6471    1
#> log_p_z[42,2]  3960    1
#> log_p_z[42,3]  3520    1
#> log_p_z[42,4]  3809    1
#> log_p_z[43,1]  4652    1
#> log_p_z[43,2]  6706    1
#> log_p_z[43,3]  3183    1
#> log_p_z[43,4]  3874    1
#> log_p_z[44,1]  7194    1
#> log_p_z[44,2]  3845    1
#> log_p_z[44,3]  3564    1
#> log_p_z[44,4]  3847    1
#> log_p_z[45,1]  5232    1
#> log_p_z[45,2]  7043    1
#> log_p_z[45,3]  3092    1
#> log_p_z[45,4]  4141    1
#> log_lik[1]     7935    1
#> log_lik[2]     5047    1
#> log_lik[3]     4234    1
#> log_lik[4]     4570    1
#> log_lik[5]     7519    1
#> log_lik[6]     7232    1
#> log_lik[7]     6477    1
#> log_lik[8]     5806    1
#> log_lik[9]     6294    1
#> log_lik[10]    5657    1
#> log_lik[11]    5607    1
#> log_lik[12]    5767    1
#> log_lik[13]    7935    1
#> log_lik[14]    4730    1
#> log_lik[15]    7194    1
#> log_lik[16]    6876    1
#> log_lik[17]    7935    1
#> log_lik[18]    7935    1
#> log_lik[19]    5277    1
#> log_lik[20]    5070    1
#> log_lik[21]    7043    1
#> log_lik[22]    5277    1
#> log_lik[23]    7519    1
#> log_lik[24]    5020    1
#> log_lik[25]    7935    1
#> log_lik[26]    7935    1
#> log_lik[27]    6706    1
#> log_lik[28]    7935    1
#> log_lik[29]    7935    1
#> log_lik[30]    6138    1
#> log_lik[31]    7935    1
#> log_lik[32]    5144    1
#> log_lik[33]    7935    1
#> log_lik[34]    7043    1
#> log_lik[35]    5562    1
#> log_lik[36]    4329    1
#> log_lik[37]    5515    1
#> log_lik[38]    4529    1
#> log_lik[39]    6199    1
#> log_lik[40]    7935    1
#> log_lik[41]    7935    1
#> log_lik[42]    6477    1
#> log_lik[43]    6706    1
#> log_lik[44]    7194    1
#> log_lik[45]    7043    1
#> lp__           1313    1
#> 
#> Samples were drawn using NUTS(diag_e) at Wed Sep  6 12:36:22 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }