Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.37    0.00 0.07    0.25    0.33    0.37    0.42    0.51
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.08    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.08    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.13
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.02    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.58    0.74    0.81    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.03    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.30
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.29
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.18    0.41
#> theta[1,4,4]     0.69    0.00 0.14    0.40    0.60    0.70    0.79    0.91
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.76    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.21    0.33
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.59    0.65    0.76
#> theta[2,2,3]     0.30    0.00 0.09    0.15    0.24    0.30    0.36    0.48
#> theta[2,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,3]     0.62    0.00 0.13    0.35    0.54    0.63    0.72    0.85
#> theta[2,3,4]     0.22    0.00 0.11    0.04    0.13    0.20    0.29    0.48
#> theta[2,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.56    0.67    0.77    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.09    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.56    0.67    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.28
#> theta[3,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.13    0.25    0.41    0.51    0.60    0.76
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.31    0.48
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[3,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.43
#> theta[3,4,4]     0.63    0.00 0.15    0.31    0.52    0.64    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.78    0.84    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.87
#> theta[4,2,3]     0.13    0.00 0.06    0.03    0.08    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.61    0.70    0.78    0.91
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.39
#> theta[4,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[4,4,2]     0.12    0.00 0.11    0.00    0.04    0.09    0.17    0.40
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,4]     0.65    0.00 0.15    0.32    0.55    0.67    0.77    0.91
#> theta[5,1,1]     0.87    0.00 0.07    0.71    0.82    0.87    0.91    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.06    0.04    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.60    0.66    0.72    0.81
#> theta[5,2,3]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.33
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.28
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,3,3]     0.61    0.00 0.13    0.33    0.52    0.61    0.70    0.84
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.14    0.21    0.38
#> theta[5,4,1]     0.12    0.00 0.10    0.01    0.04    0.09    0.17    0.37
#> theta[5,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.43
#> theta[5,4,4]     0.63    0.00 0.15    0.32    0.53    0.64    0.74    0.90
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[1,2]   -20.01    0.04 2.28  -25.05  -21.38  -19.79  -18.41  -16.09
#> log_p_z[1,3]   -24.76    0.09 4.26  -34.43  -27.16  -24.36  -21.75  -17.74
#> log_p_z[1,4]   -22.07    0.08 4.44  -32.35  -24.57  -21.58  -18.96  -14.95
#> log_p_z[2,1]   -29.80    0.07 4.28  -38.99  -32.49  -29.30  -26.73  -22.88
#> log_p_z[2,2]   -21.03    0.04 2.42  -26.19  -22.57  -20.84  -19.36  -16.80
#> log_p_z[2,3]    -7.61    0.02 1.21  -10.21   -8.30   -7.51   -6.78   -5.60
#> log_p_z[2,4]   -16.32    0.08 4.14  -25.98  -18.62  -15.77  -13.36   -9.79
#> log_p_z[3,1]   -11.62    0.03 1.58  -15.20  -12.58  -11.44  -10.45   -9.06
#> log_p_z[3,2]   -10.78    0.03 1.46  -14.16  -11.60  -10.59   -9.76   -8.43
#> log_p_z[3,3]   -23.38    0.07 3.40  -30.81  -25.47  -23.08  -21.02  -17.62
#> log_p_z[3,4]   -21.99    0.07 3.70  -30.36  -24.19  -21.61  -19.41  -15.75
#> log_p_z[4,1]   -14.63    0.03 1.83  -18.46  -15.75  -14.49  -13.37  -11.35
#> log_p_z[4,2]    -7.56    0.01 0.92   -9.54   -8.13   -7.50   -6.90   -6.00
#> log_p_z[4,3]   -20.78    0.08 4.35  -30.83  -23.35  -20.29  -17.59  -13.99
#> log_p_z[4,4]   -21.91    0.08 4.24  -31.86  -24.35  -21.43  -18.87  -15.23
#> log_p_z[5,1]   -21.44    0.04 2.50  -26.80  -23.07  -21.30  -19.69  -17.00
#> log_p_z[5,2]    -3.72    0.01 0.43   -4.63   -3.99   -3.70   -3.41   -2.91
#> log_p_z[5,3]   -18.78    0.08 4.22  -28.62  -21.36  -18.31  -15.62  -12.22
#> log_p_z[5,4]   -21.89    0.08 4.29  -31.61  -24.33  -21.39  -18.87  -15.01
#> log_p_z[6,1]   -21.42    0.04 2.52  -26.81  -23.06  -21.23  -19.66  -16.97
#> log_p_z[6,2]    -5.47    0.01 0.66   -6.89   -5.89   -5.42   -5.01   -4.34
#> log_p_z[6,3]   -17.58    0.08 4.09  -27.30  -20.01  -17.07  -14.48  -11.24
#> log_p_z[6,4]   -21.65    0.08 4.21  -31.25  -24.01  -21.12  -18.66  -14.78
#> log_p_z[7,1]    -6.00    0.01 0.60   -7.32   -6.38   -5.96   -5.57   -4.93
#> log_p_z[7,2]   -14.94    0.03 1.70  -18.72  -15.96  -14.76  -13.77  -12.09
#> log_p_z[7,3]   -24.27    0.07 3.54  -31.96  -26.41  -23.98  -21.75  -18.39
#> log_p_z[7,4]   -22.00    0.07 3.69  -30.39  -24.12  -21.59  -19.44  -15.82
#> log_p_z[8,1]   -29.85    0.07 4.39  -39.85  -32.37  -29.38  -26.75  -22.58
#> log_p_z[8,2]   -18.28    0.03 2.09  -22.76  -19.59  -18.15  -16.83  -14.54
#> log_p_z[8,3]    -5.72    0.01 0.82   -7.42   -6.23   -5.65   -5.14   -4.27
#> log_p_z[8,4]   -18.49    0.08 4.27  -28.19  -21.01  -17.91  -15.36  -11.72
#> log_p_z[9,1]   -20.02    0.04 2.27  -24.77  -21.39  -19.89  -18.42  -15.94
#> log_p_z[9,2]    -4.51    0.01 0.59   -5.78   -4.87   -4.47   -4.10   -3.50
#> log_p_z[9,3]   -19.77    0.08 4.30  -29.68  -22.40  -19.19  -16.61  -13.08
#> log_p_z[9,4]   -21.67    0.08 4.30  -31.55  -24.11  -21.13  -18.71  -14.95
#> log_p_z[10,1]  -22.32    0.04 2.40  -27.44  -23.83  -22.16  -20.62  -18.05
#> log_p_z[10,2]   -7.11    0.01 0.72   -8.64   -7.58   -7.05   -6.61   -5.85
#> log_p_z[10,3]  -16.92    0.06 3.14  -23.97  -18.91  -16.51  -14.62  -11.93
#> log_p_z[10,4]  -21.34    0.07 3.55  -29.32  -23.42  -21.03  -18.83  -15.49
#> log_p_z[11,1]  -29.73    0.07 4.40  -40.07  -32.21  -29.24  -26.65  -22.57
#> log_p_z[11,2]  -29.32    0.08 4.41  -39.69  -31.87  -28.94  -26.14  -22.13
#> log_p_z[11,3]  -18.26    0.08 3.54  -26.64  -20.00  -17.72  -15.87  -12.97
#> log_p_z[11,4]   -5.82    0.01 1.01   -7.95   -6.45   -5.75   -5.12   -4.01
#> log_p_z[12,1]  -22.95    0.04 2.74  -28.83  -24.66  -22.77  -21.03  -18.19
#> log_p_z[12,2]   -9.60    0.02 1.28  -12.61  -10.31   -9.47   -8.73   -7.47
#> log_p_z[12,3]  -15.21    0.07 4.04  -24.72  -17.64  -14.69  -12.32   -8.95
#> log_p_z[12,4]  -19.21    0.08 4.05  -28.60  -21.48  -18.73  -16.35  -12.59
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[13,2]  -20.01    0.04 2.28  -25.05  -21.38  -19.79  -18.41  -16.09
#> log_p_z[13,3]  -24.76    0.09 4.26  -34.43  -27.16  -24.36  -21.75  -17.74
#> log_p_z[13,4]  -22.07    0.08 4.44  -32.35  -24.57  -21.58  -18.96  -14.95
#> log_p_z[14,1]  -19.02    0.04 2.37  -24.02  -20.52  -18.85  -17.38  -14.84
#> log_p_z[14,2]   -6.26    0.01 0.87   -8.22   -6.77   -6.17   -5.66   -4.79
#> log_p_z[14,3]  -18.76    0.08 4.22  -28.67  -21.21  -18.16  -15.69  -12.34
#> log_p_z[14,4]  -21.89    0.08 4.27  -31.26  -24.28  -21.39  -18.95  -14.97
#> log_p_z[15,1]   -4.40    0.01 0.41   -5.28   -4.67   -4.37   -4.11   -3.68
#> log_p_z[15,2]  -17.25    0.03 1.83  -21.27  -18.40  -17.10  -15.97  -14.01
#> log_p_z[15,3]  -24.27    0.07 3.52  -32.14  -26.36  -23.95  -21.79  -18.18
#> log_p_z[15,4]  -22.04    0.07 3.68  -30.27  -24.20  -21.65  -19.50  -15.88
#> log_p_z[16,1]   -3.80    0.01 0.54   -4.98   -4.13   -3.75   -3.43   -2.86
#> log_p_z[16,2]  -17.70    0.04 2.18  -22.49  -18.97  -17.45  -16.19  -14.06
#> log_p_z[16,3]  -24.77    0.09 4.30  -34.30  -27.23  -24.22  -21.74  -17.81
#> log_p_z[16,4]  -22.03    0.08 4.45  -32.42  -24.49  -21.56  -18.85  -14.96
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[17,2]  -20.01    0.04 2.28  -25.05  -21.38  -19.79  -18.41  -16.09
#> log_p_z[17,3]  -24.76    0.09 4.26  -34.43  -27.16  -24.36  -21.75  -17.74
#> log_p_z[17,4]  -22.07    0.08 4.44  -32.35  -24.57  -21.58  -18.96  -14.95
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[18,2]  -20.01    0.04 2.28  -25.05  -21.38  -19.79  -18.41  -16.09
#> log_p_z[18,3]  -24.76    0.09 4.26  -34.43  -27.16  -24.36  -21.75  -17.74
#> log_p_z[18,4]  -22.07    0.08 4.44  -32.35  -24.57  -21.58  -18.96  -14.95
#> log_p_z[19,1]  -16.42    0.03 1.91  -20.40  -17.62  -16.32  -15.03  -13.01
#> log_p_z[19,2]   -4.65    0.01 0.59   -5.92   -5.02   -4.61   -4.24   -3.62
#> log_p_z[19,3]  -22.24    0.08 4.42  -32.38  -24.93  -21.66  -18.95  -15.24
#> log_p_z[19,4]  -21.90    0.08 4.33  -31.84  -24.37  -21.42  -18.84  -15.07
#> log_p_z[20,1]  -18.01    0.04 2.23  -22.68  -19.42  -17.86  -16.39  -14.16
#> log_p_z[20,2]   -7.10    0.01 0.96   -9.21   -7.68   -7.02   -6.44   -5.42
#> log_p_z[20,3]  -20.11    0.08 4.24  -30.28  -22.52  -19.62  -16.97  -13.70
#> log_p_z[20,4]  -21.71    0.08 4.27  -31.31  -24.23  -21.22  -18.68  -14.85
#> log_p_z[21,1]  -19.62    0.04 2.27  -24.29  -21.03  -19.50  -18.03  -15.59
#> log_p_z[21,2]   -3.03    0.00 0.36   -3.77   -3.27   -3.03   -2.78   -2.38
#> log_p_z[21,3]  -21.32    0.08 4.33  -31.31  -23.96  -20.77  -18.14  -14.47
#> log_p_z[21,4]  -21.90    0.08 4.35  -31.92  -24.33  -21.36  -18.78  -15.03
#> log_p_z[22,1]  -16.42    0.03 1.91  -20.40  -17.62  -16.32  -15.03  -13.01
#> log_p_z[22,2]   -4.65    0.01 0.59   -5.92   -5.02   -4.61   -4.24   -3.62
#> log_p_z[22,3]  -22.24    0.08 4.42  -32.38  -24.93  -21.66  -18.95  -15.24
#> log_p_z[22,4]  -21.90    0.08 4.33  -31.84  -24.37  -21.42  -18.84  -15.07
#> log_p_z[23,1]  -21.44    0.04 2.50  -26.80  -23.07  -21.30  -19.69  -17.00
#> log_p_z[23,2]   -3.72    0.01 0.43   -4.63   -3.99   -3.70   -3.41   -2.91
#> log_p_z[23,3]  -18.78    0.08 4.22  -28.62  -21.36  -18.31  -15.62  -12.22
#> log_p_z[23,4]  -21.89    0.08 4.29  -31.61  -24.33  -21.39  -18.87  -15.01
#> log_p_z[24,1]  -17.42    0.04 2.07  -21.76  -18.70  -17.27  -15.94  -13.81
#> log_p_z[24,2]   -5.79    0.01 0.62   -7.23   -6.15   -5.73   -5.36   -4.79
#> log_p_z[24,3]  -21.81    0.06 3.38  -29.30  -23.94  -21.53  -19.39  -16.09
#> log_p_z[24,4]  -21.93    0.07 3.67  -30.04  -24.08  -21.54  -19.34  -15.96
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[25,2]  -20.01    0.04 2.28  -25.05  -21.38  -19.79  -18.41  -16.09
#> log_p_z[25,3]  -24.76    0.09 4.26  -34.43  -27.16  -24.36  -21.75  -17.74
#> log_p_z[25,4]  -22.07    0.08 4.44  -32.35  -24.57  -21.58  -18.96  -14.95
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[26,2]  -20.01    0.04 2.28  -25.05  -21.38  -19.79  -18.41  -16.09
#> log_p_z[26,3]  -24.76    0.09 4.26  -34.43  -27.16  -24.36  -21.75  -17.74
#> log_p_z[26,4]  -22.07    0.08 4.44  -32.35  -24.57  -21.58  -18.96  -14.95
#> log_p_z[27,1]  -21.91    0.04 2.39  -27.07  -23.40  -21.76  -20.24  -17.69
#> log_p_z[27,2]   -5.64    0.01 0.51   -6.74   -5.95   -5.60   -5.29   -4.74
#> log_p_z[27,3]  -18.48    0.06 3.20  -25.68  -20.47  -18.14  -16.17  -13.26
#> log_p_z[27,4]  -21.57    0.06 3.53  -29.68  -23.69  -21.20  -19.05  -15.73
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[28,2]  -20.01    0.04 2.28  -25.05  -21.38  -19.79  -18.41  -16.09
#> log_p_z[28,3]  -24.76    0.09 4.26  -34.43  -27.16  -24.36  -21.75  -17.74
#> log_p_z[28,4]  -22.07    0.08 4.44  -32.35  -24.57  -21.58  -18.96  -14.95
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[29,2]  -20.01    0.04 2.28  -25.05  -21.38  -19.79  -18.41  -16.09
#> log_p_z[29,3]  -24.76    0.09 4.26  -34.43  -27.16  -24.36  -21.75  -17.74
#> log_p_z[29,4]  -22.07    0.08 4.44  -32.35  -24.57  -21.58  -18.96  -14.95
#> log_p_z[30,1]   -6.82    0.01 0.86   -8.82   -7.30   -6.74   -6.22   -5.43
#> log_p_z[30,2]  -14.71    0.03 1.69  -18.46  -15.74  -14.54  -13.52  -11.82
#> log_p_z[30,3]  -24.29    0.07 3.49  -31.82  -26.45  -23.99  -21.84  -18.22
#> log_p_z[30,4]  -22.03    0.07 3.69  -30.47  -24.13  -21.66  -19.45  -15.88
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[31,2]  -20.01    0.04 2.28  -25.05  -21.38  -19.79  -18.41  -16.09
#> log_p_z[31,3]  -24.76    0.09 4.26  -34.43  -27.16  -24.36  -21.75  -17.74
#> log_p_z[31,4]  -22.07    0.08 4.44  -32.35  -24.57  -21.58  -18.96  -14.95
#> log_p_z[32,1]  -29.86    0.07 4.37  -39.74  -32.38  -29.44  -26.75  -22.62
#> log_p_z[32,2]  -14.82    0.03 1.72  -18.49  -15.91  -14.69  -13.61  -11.71
#> log_p_z[32,3]   -6.06    0.01 0.93   -8.07   -6.63   -5.99   -5.40   -4.44
#> log_p_z[32,4]  -20.62    0.08 4.42  -30.46  -23.17  -20.08  -17.50  -13.46
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[33,2]  -20.01    0.04 2.28  -25.05  -21.38  -19.79  -18.41  -16.09
#> log_p_z[33,3]  -24.76    0.09 4.26  -34.43  -27.16  -24.36  -21.75  -17.74
#> log_p_z[33,4]  -22.07    0.08 4.44  -32.35  -24.57  -21.58  -18.96  -14.95
#> log_p_z[34,1]  -19.62    0.04 2.27  -24.29  -21.03  -19.50  -18.03  -15.59
#> log_p_z[34,2]   -3.03    0.00 0.36   -3.77   -3.27   -3.03   -2.78   -2.38
#> log_p_z[34,3]  -21.32    0.08 4.33  -31.31  -23.96  -20.77  -18.14  -14.47
#> log_p_z[34,4]  -21.90    0.08 4.35  -31.92  -24.33  -21.36  -18.78  -15.03
#> log_p_z[35,1]  -22.57    0.04 2.68  -28.42  -24.34  -22.41  -20.65  -17.89
#> log_p_z[35,2]   -5.54    0.01 0.70   -7.03   -5.97   -5.48   -5.05   -4.32
#> log_p_z[35,3]  -16.14    0.08 4.09  -25.80  -18.57  -15.58  -13.05  -10.03
#> log_p_z[35,4]  -21.85    0.08 4.27  -31.49  -24.25  -21.34  -18.86  -15.15
#> log_p_z[36,1]  -29.80    0.07 3.63  -37.91  -32.04  -29.57  -27.23  -23.57
#> log_p_z[36,2]  -21.88    0.04 2.30  -26.97  -23.30  -21.68  -20.28  -17.94
#> log_p_z[36,3]  -10.54    0.04 1.78  -14.84  -11.38  -10.26   -9.34   -7.92
#> log_p_z[36,4]  -13.91    0.06 3.22  -21.05  -15.85  -13.58  -11.54   -8.71
#> log_p_z[37,1]  -18.55    0.04 2.27  -23.47  -19.96  -18.39  -16.95  -14.70
#> log_p_z[37,2]   -7.61    0.01 0.79   -9.30   -8.08   -7.56   -7.05   -6.25
#> log_p_z[37,3]  -19.17    0.06 3.22  -26.18  -21.03  -18.84  -16.89  -13.86
#> log_p_z[37,4]  -21.89    0.07 3.66  -30.32  -24.00  -21.51  -19.30  -15.94
#> log_p_z[38,1]  -25.27    0.05 2.82  -31.50  -27.03  -25.11  -23.31  -20.32
#> log_p_z[38,2]   -9.61    0.02 1.01  -11.87  -10.22   -9.55   -8.90   -7.84
#> log_p_z[38,3]  -11.74    0.05 2.78  -18.22  -13.40  -11.36   -9.63   -7.60
#> log_p_z[38,4]  -21.28    0.07 3.49  -28.97  -23.35  -20.98  -18.79  -15.43
#> log_p_z[39,1]  -29.45    0.07 4.34  -39.42  -32.05  -28.99  -26.32  -22.36
#> log_p_z[39,2]  -16.81    0.03 2.00  -21.17  -18.05  -16.70  -15.43  -13.29
#> log_p_z[39,3]   -7.27    0.02 1.10   -9.71   -7.93   -7.18   -6.52   -5.35
#> log_p_z[39,4]  -18.73    0.07 4.12  -28.12  -21.02  -18.20  -15.92  -12.07
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[40,2]  -20.01    0.04 2.28  -25.05  -21.38  -19.79  -18.41  -16.09
#> log_p_z[40,3]  -24.76    0.09 4.26  -34.43  -27.16  -24.36  -21.75  -17.74
#> log_p_z[40,4]  -22.07    0.08 4.44  -32.35  -24.57  -21.58  -18.96  -14.95
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[41,2]  -20.01    0.04 2.28  -25.05  -21.38  -19.79  -18.41  -16.09
#> log_p_z[41,3]  -24.76    0.09 4.26  -34.43  -27.16  -24.36  -21.75  -17.74
#> log_p_z[41,4]  -22.07    0.08 4.44  -32.35  -24.57  -21.58  -18.96  -14.95
#> log_p_z[42,1]   -6.00    0.01 0.60   -7.32   -6.38   -5.96   -5.57   -4.93
#> log_p_z[42,2]  -14.94    0.03 1.70  -18.72  -15.96  -14.76  -13.77  -12.09
#> log_p_z[42,3]  -24.27    0.07 3.54  -31.96  -26.41  -23.98  -21.75  -18.39
#> log_p_z[42,4]  -22.00    0.07 3.69  -30.39  -24.12  -21.59  -19.44  -15.82
#> log_p_z[43,1]  -21.91    0.04 2.39  -27.07  -23.40  -21.76  -20.24  -17.69
#> log_p_z[43,2]   -5.64    0.01 0.51   -6.74   -5.95   -5.60   -5.29   -4.74
#> log_p_z[43,3]  -18.48    0.06 3.20  -25.68  -20.47  -18.14  -16.17  -13.26
#> log_p_z[43,4]  -21.57    0.06 3.53  -29.68  -23.69  -21.20  -19.05  -15.73
#> log_p_z[44,1]   -4.40    0.01 0.41   -5.28   -4.67   -4.37   -4.11   -3.68
#> log_p_z[44,2]  -17.25    0.03 1.83  -21.27  -18.40  -17.10  -15.97  -14.01
#> log_p_z[44,3]  -24.27    0.07 3.52  -32.14  -26.36  -23.95  -21.79  -18.18
#> log_p_z[44,4]  -22.04    0.07 3.68  -30.27  -24.20  -21.65  -19.50  -15.88
#> log_p_z[45,1]  -19.62    0.04 2.27  -24.29  -21.03  -19.50  -18.03  -15.59
#> log_p_z[45,2]   -3.03    0.00 0.36   -3.77   -3.27   -3.03   -2.78   -2.38
#> log_p_z[45,3]  -21.32    0.08 4.33  -31.31  -23.96  -20.77  -18.14  -14.47
#> log_p_z[45,4]  -21.90    0.08 4.35  -31.92  -24.33  -21.36  -18.78  -15.03
#> log_lik[1]      -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[2]      -7.57    0.02 1.13   -9.94   -8.26   -7.50   -6.78   -5.59
#> log_lik[3]      -9.94    0.01 0.88  -11.78  -10.51   -9.91   -9.34   -8.32
#> log_lik[4]      -7.56    0.01 0.91   -9.48   -8.12   -7.49   -6.89   -6.00
#> log_lik[5]      -3.72    0.01 0.43   -4.63   -3.99   -3.70   -3.41   -2.91
#> log_lik[6]      -5.47    0.01 0.66   -6.89   -5.89   -5.42   -5.01   -4.34
#> log_lik[7]      -6.00    0.01 0.60   -7.31   -6.38   -5.96   -5.57   -4.93
#> log_lik[8]      -5.71    0.01 0.82   -7.42   -6.23   -5.65   -5.14   -4.27
#> log_lik[9]      -4.51    0.01 0.59   -5.78   -4.87   -4.47   -4.10   -3.50
#> log_lik[10]     -7.11    0.01 0.71   -8.64   -7.58   -7.05   -6.61   -5.85
#> log_lik[11]     -5.82    0.01 1.01   -7.94   -6.45   -5.75   -5.12   -4.01
#> log_lik[12]     -9.35    0.02 1.05  -11.63  -10.02   -9.29   -8.63   -7.42
#> log_lik[13]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[14]     -6.26    0.01 0.87   -8.22   -6.77   -6.17   -5.66   -4.79
#> log_lik[15]     -4.40    0.01 0.41   -5.28   -4.67   -4.37   -4.11   -3.68
#> log_lik[16]     -3.80    0.01 0.54   -4.98   -4.13   -3.75   -3.43   -2.86
#> log_lik[17]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[18]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[19]     -4.65    0.01 0.59   -5.92   -5.02   -4.61   -4.24   -3.62
#> log_lik[20]     -7.10    0.01 0.96   -9.21   -7.68   -7.02   -6.44   -5.42
#> log_lik[21]     -3.03    0.00 0.36   -3.77   -3.27   -3.03   -2.78   -2.38
#> log_lik[22]     -4.65    0.01 0.59   -5.92   -5.02   -4.61   -4.24   -3.62
#> log_lik[23]     -3.72    0.01 0.43   -4.63   -3.99   -3.70   -3.41   -2.91
#> log_lik[24]     -5.79    0.01 0.62   -7.23   -6.15   -5.73   -5.36   -4.79
#> log_lik[25]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[26]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[27]     -5.64    0.01 0.51   -6.74   -5.95   -5.60   -5.29   -4.74
#> log_lik[28]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[29]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[30]     -6.82    0.01 0.85   -8.80   -7.30   -6.74   -6.22   -5.43
#> log_lik[31]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[32]     -6.05    0.01 0.93   -8.06   -6.63   -5.99   -5.40   -4.44
#> log_lik[33]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[34]     -3.03    0.00 0.36   -3.77   -3.27   -3.03   -2.78   -2.38
#> log_lik[35]     -5.53    0.01 0.69   -7.03   -5.97   -5.48   -5.05   -4.32
#> log_lik[36]     -9.92    0.02 1.19  -12.52  -10.66   -9.87   -9.10   -7.79
#> log_lik[37]     -7.61    0.01 0.79   -9.30   -8.08   -7.56   -7.05   -6.25
#> log_lik[38]     -9.01    0.01 0.86  -10.78   -9.56   -9.00   -8.45   -7.39
#> log_lik[39]     -7.27    0.02 1.10   -9.71   -7.93   -7.18   -6.52   -5.35
#> log_lik[40]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[41]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[42]     -6.00    0.01 0.60   -7.31   -6.38   -5.96   -5.57   -4.93
#> log_lik[43]     -5.64    0.01 0.51   -6.74   -5.95   -5.60   -5.29   -4.74
#> log_lik[44]     -4.40    0.01 0.41   -5.28   -4.67   -4.37   -4.11   -3.68
#> log_lik[45]     -3.03    0.00 0.36   -3.77   -3.27   -3.03   -2.78   -2.38
#> lp__          -463.96    0.18 6.30 -476.98 -467.99 -463.61 -459.50 -452.71
#>               n_eff Rhat
#> pi[1]          6348    1
#> pi[2]          7070    1
#> pi[3]          4899    1
#> pi[4]          6302    1
#> theta[1,1,1]   5427    1
#> theta[1,1,2]   5293    1
#> theta[1,1,3]   4568    1
#> theta[1,1,4]   4333    1
#> theta[1,2,1]   5211    1
#> theta[1,2,2]   5669    1
#> theta[1,2,3]   5352    1
#> theta[1,2,4]   5115    1
#> theta[1,3,1]   6358    1
#> theta[1,3,2]   4394    1
#> theta[1,3,3]   5292    1
#> theta[1,3,4]   5082    1
#> theta[1,4,1]   5800    1
#> theta[1,4,2]   5324    1
#> theta[1,4,3]   4069    1
#> theta[1,4,4]   4974    1
#> theta[2,1,1]   6687    1
#> theta[2,1,2]   5785    1
#> theta[2,1,3]   4026    1
#> theta[2,1,4]   4417    1
#> theta[2,2,1]   5700    1
#> theta[2,2,2]   5873    1
#> theta[2,2,3]   5790    1
#> theta[2,2,4]   6922    1
#> theta[2,3,1]   5750    1
#> theta[2,3,2]   6152    1
#> theta[2,3,3]   5134    1
#> theta[2,3,4]   5699    1
#> theta[2,4,1]   6327    1
#> theta[2,4,2]   6189    1
#> theta[2,4,3]   4886    1
#> theta[2,4,4]   5538    1
#> theta[3,1,1]   6725    1
#> theta[3,1,2]   4752    1
#> theta[3,1,3]   5010    1
#> theta[3,1,4]   4600    1
#> theta[3,2,1]   5847    1
#> theta[3,2,2]   6074    1
#> theta[3,2,3]   5159    1
#> theta[3,2,4]   5586    1
#> theta[3,3,1]   5565    1
#> theta[3,3,2]   5709    1
#> theta[3,3,3]   6004    1
#> theta[3,3,4]   5335    1
#> theta[3,4,1]   5511    1
#> theta[3,4,2]   6018    1
#> theta[3,4,3]   5183    1
#> theta[3,4,4]   6061    1
#> theta[4,1,1]   5298    1
#> theta[4,1,2]   4823    1
#> theta[4,1,3]   5122    1
#> theta[4,1,4]   4819    1
#> theta[4,2,1]   5599    1
#> theta[4,2,2]   5677    1
#> theta[4,2,3]   6311    1
#> theta[4,2,4]   5408    1
#> theta[4,3,1]   6030    1
#> theta[4,3,2]   5307    1
#> theta[4,3,3]   5434    1
#> theta[4,3,4]   5727    1
#> theta[4,4,1]   5340    1
#> theta[4,4,2]   6163    1
#> theta[4,4,3]   5742    1
#> theta[4,4,4]   6226    1
#> theta[5,1,1]   5790    1
#> theta[5,1,2]   4353    1
#> theta[5,1,3]   5346    1
#> theta[5,1,4]   4402    1
#> theta[5,2,1]   6310    1
#> theta[5,2,2]   6246    1
#> theta[5,2,3]   5082    1
#> theta[5,2,4]   5050    1
#> theta[5,3,1]   5688    1
#> theta[5,3,2]   6592    1
#> theta[5,3,3]   5936    1
#> theta[5,3,4]   5580    1
#> theta[5,4,1]   6499    1
#> theta[5,4,2]   6107    1
#> theta[5,4,3]   5325    1
#> theta[5,4,4]   6413    1
#> log_p_z[1,1]   6164    1
#> log_p_z[1,2]   3400    1
#> log_p_z[1,3]   2338    1
#> log_p_z[1,4]   3053    1
#> log_p_z[2,1]   4268    1
#> log_p_z[2,2]   3783    1
#> log_p_z[2,3]   3604    1
#> log_p_z[2,4]   2863    1
#> log_p_z[3,1]   3376    1
#> log_p_z[3,2]   3388    1
#> log_p_z[3,3]   2606    1
#> log_p_z[3,4]   2890    1
#> log_p_z[4,1]   4344    1
#> log_p_z[4,2]   4277    1
#> log_p_z[4,3]   3031    1
#> log_p_z[4,4]   2688    1
#> log_p_z[5,1]   3620    1
#> log_p_z[5,2]   5939    1
#> log_p_z[5,3]   3011    1
#> log_p_z[5,4]   2678    1
#> log_p_z[6,1]   4026    1
#> log_p_z[6,2]   4780    1
#> log_p_z[6,3]   2912    1
#> log_p_z[6,4]   2726    1
#> log_p_z[7,1]   5237    1
#> log_p_z[7,2]   3480    1
#> log_p_z[7,3]   2462    1
#> log_p_z[7,4]   2753    1
#> log_p_z[8,1]   3892    1
#> log_p_z[8,2]   4579    1
#> log_p_z[8,3]   5126    1
#> log_p_z[8,4]   3097    1
#> log_p_z[9,1]   3374    1
#> log_p_z[9,2]   4976    1
#> log_p_z[9,3]   3033    1
#> log_p_z[9,4]   2782    1
#> log_p_z[10,1]  3273    1
#> log_p_z[10,2]  4864    1
#> log_p_z[10,3]  3127    1
#> log_p_z[10,4]  2841    1
#> log_p_z[11,1]  4379    1
#> log_p_z[11,2]  3364    1
#> log_p_z[11,3]  1975    1
#> log_p_z[11,4]  5733    1
#> log_p_z[12,1]  3788    1
#> log_p_z[12,2]  3160    1
#> log_p_z[12,3]  2941    1
#> log_p_z[12,4]  2827    1
#> log_p_z[13,1]  6164    1
#> log_p_z[13,2]  3400    1
#> log_p_z[13,3]  2338    1
#> log_p_z[13,4]  3053    1
#> log_p_z[14,1]  3716    1
#> log_p_z[14,2]  4499    1
#> log_p_z[14,3]  2942    1
#> log_p_z[14,4]  2663    1
#> log_p_z[15,1]  5793    1
#> log_p_z[15,2]  3430    1
#> log_p_z[15,3]  2608    1
#> log_p_z[15,4]  2873    1
#> log_p_z[16,1]  5869    1
#> log_p_z[16,2]  3440    1
#> log_p_z[16,3]  2161    1
#> log_p_z[16,4]  2882    1
#> log_p_z[17,1]  6164    1
#> log_p_z[17,2]  3400    1
#> log_p_z[17,3]  2338    1
#> log_p_z[17,4]  3053    1
#> log_p_z[18,1]  6164    1
#> log_p_z[18,2]  3400    1
#> log_p_z[18,3]  2338    1
#> log_p_z[18,4]  3053    1
#> log_p_z[19,1]  3809    1
#> log_p_z[19,2]  5234    1
#> log_p_z[19,3]  3002    1
#> log_p_z[19,4]  2776    1
#> log_p_z[20,1]  3881    1
#> log_p_z[20,2]  4522    1
#> log_p_z[20,3]  2864    1
#> log_p_z[20,4]  2855    1
#> log_p_z[21,1]  3563    1
#> log_p_z[21,2]  6098    1
#> log_p_z[21,3]  3016    1
#> log_p_z[21,4]  2790    1
#> log_p_z[22,1]  3809    1
#> log_p_z[22,2]  5234    1
#> log_p_z[22,3]  3002    1
#> log_p_z[22,4]  2776    1
#> log_p_z[23,1]  3620    1
#> log_p_z[23,2]  5939    1
#> log_p_z[23,3]  3011    1
#> log_p_z[23,4]  2678    1
#> log_p_z[24,1]  3410    1
#> log_p_z[24,2]  4038    1
#> log_p_z[24,3]  3078    1
#> log_p_z[24,4]  2796    1
#> log_p_z[25,1]  6164    1
#> log_p_z[25,2]  3400    1
#> log_p_z[25,3]  2338    1
#> log_p_z[25,4]  3053    1
#> log_p_z[26,1]  6164    1
#> log_p_z[26,2]  3400    1
#> log_p_z[26,3]  2338    1
#> log_p_z[26,4]  3053    1
#> log_p_z[27,1]  3496    1
#> log_p_z[27,2]  5040    1
#> log_p_z[27,3]  3101    1
#> log_p_z[27,4]  2997    1
#> log_p_z[28,1]  6164    1
#> log_p_z[28,2]  3400    1
#> log_p_z[28,3]  2338    1
#> log_p_z[28,4]  3053    1
#> log_p_z[29,1]  6164    1
#> log_p_z[29,2]  3400    1
#> log_p_z[29,3]  2338    1
#> log_p_z[29,4]  3053    1
#> log_p_z[30,1]  4375    1
#> log_p_z[30,2]  3429    1
#> log_p_z[30,3]  2785    1
#> log_p_z[30,4]  2861    1
#> log_p_z[31,1]  6164    1
#> log_p_z[31,2]  3400    1
#> log_p_z[31,3]  2338    1
#> log_p_z[31,4]  3053    1
#> log_p_z[32,1]  3758    1
#> log_p_z[32,2]  4705    1
#> log_p_z[32,3]  4586    1
#> log_p_z[32,4]  2969    1
#> log_p_z[33,1]  6164    1
#> log_p_z[33,2]  3400    1
#> log_p_z[33,3]  2338    1
#> log_p_z[33,4]  3053    1
#> log_p_z[34,1]  3563    1
#> log_p_z[34,2]  6098    1
#> log_p_z[34,3]  3016    1
#> log_p_z[34,4]  2790    1
#> log_p_z[35,1]  3997    1
#> log_p_z[35,2]  4654    1
#> log_p_z[35,3]  2915    1
#> log_p_z[35,4]  2552    1
#> log_p_z[36,1]  3110    1
#> log_p_z[36,2]  3095    1
#> log_p_z[36,3]  1865    1
#> log_p_z[36,4]  2779    1
#> log_p_z[37,1]  3920    1
#> log_p_z[37,2]  4370    1
#> log_p_z[37,3]  2932    1
#> log_p_z[37,4]  2655    1
#> log_p_z[38,1]  3601    1
#> log_p_z[38,2]  4234    1
#> log_p_z[38,3]  2968    1
#> log_p_z[38,4]  2683    1
#> log_p_z[39,1]  4062    1
#> log_p_z[39,2]  4767    1
#> log_p_z[39,3]  4291    1
#> log_p_z[39,4]  3133    1
#> log_p_z[40,1]  6164    1
#> log_p_z[40,2]  3400    1
#> log_p_z[40,3]  2338    1
#> log_p_z[40,4]  3053    1
#> log_p_z[41,1]  6164    1
#> log_p_z[41,2]  3400    1
#> log_p_z[41,3]  2338    1
#> log_p_z[41,4]  3053    1
#> log_p_z[42,1]  5237    1
#> log_p_z[42,2]  3480    1
#> log_p_z[42,3]  2462    1
#> log_p_z[42,4]  2753    1
#> log_p_z[43,1]  3496    1
#> log_p_z[43,2]  5040    1
#> log_p_z[43,3]  3101    1
#> log_p_z[43,4]  2997    1
#> log_p_z[44,1]  5793    1
#> log_p_z[44,2]  3430    1
#> log_p_z[44,3]  2608    1
#> log_p_z[44,4]  2873    1
#> log_p_z[45,1]  3563    1
#> log_p_z[45,2]  6098    1
#> log_p_z[45,3]  3016    1
#> log_p_z[45,4]  2790    1
#> log_lik[1]     6164    1
#> log_lik[2]     4241    1
#> log_lik[3]     5077    1
#> log_lik[4]     4310    1
#> log_lik[5]     5939    1
#> log_lik[6]     4781    1
#> log_lik[7]     5238    1
#> log_lik[8]     5133    1
#> log_lik[9]     4976    1
#> log_lik[10]    4870    1
#> log_lik[11]    5734    1
#> log_lik[12]    4285    1
#> log_lik[13]    6164    1
#> log_lik[14]    4499    1
#> log_lik[15]    5793    1
#> log_lik[16]    5869    1
#> log_lik[17]    6164    1
#> log_lik[18]    6164    1
#> log_lik[19]    5234    1
#> log_lik[20]    4526    1
#> log_lik[21]    6098    1
#> log_lik[22]    5234    1
#> log_lik[23]    5939    1
#> log_lik[24]    4039    1
#> log_lik[25]    6164    1
#> log_lik[26]    6164    1
#> log_lik[27]    5040    1
#> log_lik[28]    6164    1
#> log_lik[29]    6164    1
#> log_lik[30]    4375    1
#> log_lik[31]    6164    1
#> log_lik[32]    4585    1
#> log_lik[33]    6164    1
#> log_lik[34]    6098    1
#> log_lik[35]    4679    1
#> log_lik[36]    3952    1
#> log_lik[37]    4371    1
#> log_lik[38]    4295    1
#> log_lik[39]    4298    1
#> log_lik[40]    6164    1
#> log_lik[41]    6164    1
#> log_lik[42]    5238    1
#> log_lik[43]    5040    1
#> log_lik[44]    5793    1
#> log_lik[45]    6098    1
#> lp__           1232    1
#> 
#> Samples were drawn using NUTS(diag_e) at Mon Sep 11 19:17:32 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }