Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.26    0.33    0.37    0.42    0.51
#> pi[2]            0.41    0.00 0.06    0.28    0.36    0.40    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.09    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.92
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.29
#> theta[1,3,3]     0.79    0.00 0.09    0.58    0.74    0.80    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.18    0.39
#> theta[1,4,4]     0.69    0.00 0.13    0.40    0.60    0.70    0.78    0.91
#> theta[2,1,1]     0.75    0.00 0.08    0.58    0.70    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.59    0.65    0.76
#> theta[2,2,3]     0.30    0.00 0.08    0.15    0.25    0.30    0.36    0.48
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[2,3,3]     0.62    0.00 0.13    0.35    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.28    0.48
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,2]     0.12    0.00 0.11    0.00    0.04    0.08    0.17    0.40
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.02    0.05    0.08    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.68    0.73    0.79    0.88
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.17    0.28
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[3,3,2]     0.18    0.00 0.11    0.02    0.09    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.13    0.25    0.41    0.51    0.61    0.76
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.30    0.48
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[3,4,3]     0.14    0.00 0.12    0.01    0.05    0.11    0.20    0.43
#> theta[3,4,4]     0.63    0.00 0.15    0.31    0.52    0.64    0.75    0.90
#> theta[4,1,1]     0.83    0.00 0.08    0.66    0.78    0.83    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.08    0.55    0.67    0.73    0.79    0.87
#> theta[4,2,3]     0.13    0.00 0.06    0.04    0.09    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.18
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.26
#> theta[4,3,3]     0.69    0.00 0.12    0.43    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.38
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,3]     0.12    0.00 0.11    0.00    0.04    0.09    0.17    0.40
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.90
#> theta[5,1,1]     0.87    0.00 0.07    0.71    0.83    0.87    0.91    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,3,3]     0.61    0.00 0.13    0.33    0.52    0.61    0.70    0.84
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.40
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.41
#> theta[5,4,4]     0.63    0.00 0.15    0.32    0.53    0.63    0.75    0.89
#> log_p_z[1,1]    -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_p_z[1,2]   -19.95    0.04 2.26  -25.05  -21.31  -19.77  -18.38  -16.07
#> log_p_z[1,3]   -24.88    0.08 4.36  -34.56  -27.52  -24.44  -21.73  -17.67
#> log_p_z[1,4]   -21.78    0.07 4.23  -31.05  -24.18  -21.35  -18.77  -14.81
#> log_p_z[2,1]   -29.80    0.07 4.40  -40.06  -32.30  -29.28  -26.66  -22.78
#> log_p_z[2,2]   -20.99    0.04 2.37  -26.17  -22.46  -20.81  -19.34  -16.84
#> log_p_z[2,3]    -7.64    0.02 1.19  -10.25   -8.34   -7.52   -6.80   -5.63
#> log_p_z[2,4]   -16.47    0.08 4.18  -26.14  -18.88  -15.90  -13.47  -10.00
#> log_p_z[3,1]   -11.67    0.03 1.65  -15.34  -12.58  -11.50  -10.51   -8.97
#> log_p_z[3,2]   -10.78    0.03 1.47  -14.30  -11.60  -10.59   -9.75   -8.33
#> log_p_z[3,3]   -23.28    0.06 3.38  -30.86  -25.36  -22.98  -20.90  -17.49
#> log_p_z[3,4]   -21.92    0.07 3.59  -29.92  -24.18  -21.54  -19.31  -16.05
#> log_p_z[4,1]   -14.67    0.03 1.89  -18.78  -15.84  -14.53  -13.34  -11.42
#> log_p_z[4,2]    -7.57    0.01 0.93   -9.56   -8.16   -7.51   -6.93   -5.90
#> log_p_z[4,3]   -20.61    0.09 4.32  -30.35  -23.08  -20.03  -17.53  -13.68
#> log_p_z[4,4]   -21.72    0.08 4.17  -31.22  -24.12  -21.29  -18.73  -14.89
#> log_p_z[5,1]   -21.52    0.05 2.61  -27.33  -23.04  -21.31  -19.69  -16.95
#> log_p_z[5,2]    -3.72    0.01 0.42   -4.61   -3.98   -3.70   -3.44   -2.92
#> log_p_z[5,3]   -18.61    0.08 4.17  -28.10  -20.99  -18.10  -15.57  -12.10
#> log_p_z[5,4]   -21.80    0.08 4.19  -31.23  -24.14  -21.38  -18.86  -14.77
#> log_p_z[6,1]   -21.51    0.05 2.59  -27.25  -23.11  -21.30  -19.70  -16.99
#> log_p_z[6,2]    -5.49    0.01 0.64   -6.86   -5.89   -5.44   -5.05   -4.35
#> log_p_z[6,3]   -17.37    0.08 4.02  -26.61  -19.66  -16.85  -14.46  -11.09
#> log_p_z[6,4]   -21.58    0.08 4.14  -30.96  -23.96  -21.14  -18.71  -14.80
#> log_p_z[7,1]    -6.00    0.01 0.60   -7.32   -6.38   -5.96   -5.58   -4.96
#> log_p_z[7,2]   -14.91    0.03 1.66  -18.58  -15.91  -14.78  -13.75  -12.06
#> log_p_z[7,3]   -24.27    0.07 3.54  -32.14  -26.44  -23.99  -21.73  -18.16
#> log_p_z[7,4]   -21.78    0.07 3.49  -29.45  -23.94  -21.49  -19.33  -15.79
#> log_p_z[8,1]   -29.80    0.07 4.32  -39.51  -32.38  -29.29  -26.79  -22.63
#> log_p_z[8,2]   -18.27    0.03 2.07  -22.88  -19.52  -18.08  -16.82  -14.68
#> log_p_z[8,3]    -5.73    0.01 0.82   -7.50   -6.24   -5.68   -5.16   -4.28
#> log_p_z[8,4]   -18.65    0.09 4.28  -28.36  -21.26  -18.17  -15.63  -11.70
#> log_p_z[9,1]   -20.02    0.04 2.32  -25.03  -21.50  -19.83  -18.39  -15.99
#> log_p_z[9,2]    -4.52    0.01 0.59   -5.76   -4.89   -4.46   -4.10   -3.52
#> log_p_z[9,3]   -19.56    0.08 4.25  -29.39  -21.98  -19.03  -16.56  -12.79
#> log_p_z[9,4]   -21.57    0.08 4.17  -30.97  -23.98  -21.12  -18.66  -14.80
#> log_p_z[10,1]  -22.30    0.04 2.41  -27.57  -23.73  -22.08  -20.63  -18.19
#> log_p_z[10,2]   -7.12    0.01 0.71   -8.65   -7.54   -7.04   -6.64   -5.86
#> log_p_z[10,3]  -16.76    0.06 3.12  -23.85  -18.59  -16.47  -14.52  -11.66
#> log_p_z[10,4]  -21.33    0.06 3.47  -29.13  -23.39  -20.96  -18.93  -15.53
#> log_p_z[11,1]  -29.79    0.08 4.48  -39.96  -32.22  -29.23  -26.68  -22.75
#> log_p_z[11,2]  -29.35    0.08 4.33  -39.39  -31.95  -28.88  -26.28  -22.26
#> log_p_z[11,3]  -18.25    0.08 3.49  -26.70  -19.93  -17.73  -15.82  -13.05
#> log_p_z[11,4]   -5.83    0.01 0.97   -7.88   -6.43   -5.75   -5.14   -4.17
#> log_p_z[12,1]  -22.98    0.04 2.72  -28.94  -24.66  -22.78  -21.13  -18.17
#> log_p_z[12,2]   -9.61    0.02 1.26  -12.54  -10.32   -9.45   -8.74   -7.53
#> log_p_z[12,3]  -15.06    0.08 4.00  -24.25  -17.36  -14.58  -12.21   -8.88
#> log_p_z[12,4]  -19.12    0.07 3.93  -28.04  -21.28  -18.69  -16.38  -12.74
#> log_p_z[13,1]   -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_p_z[13,2]  -19.95    0.04 2.26  -25.05  -21.31  -19.77  -18.38  -16.07
#> log_p_z[13,3]  -24.88    0.08 4.36  -34.56  -27.52  -24.44  -21.73  -17.67
#> log_p_z[13,4]  -21.78    0.07 4.23  -31.05  -24.18  -21.35  -18.77  -14.81
#> log_p_z[14,1]  -19.09    0.04 2.43  -24.50  -20.49  -18.88  -17.42  -14.96
#> log_p_z[14,2]   -6.23    0.01 0.85   -8.10   -6.73   -6.16   -5.64   -4.82
#> log_p_z[14,3]  -18.64    0.08 4.17  -28.13  -21.01  -18.12  -15.65  -12.07
#> log_p_z[14,4]  -21.73    0.07 4.06  -30.78  -24.10  -21.36  -18.89  -14.95
#> log_p_z[15,1]   -4.41    0.01 0.41   -5.29   -4.66   -4.37   -4.12   -3.69
#> log_p_z[15,2]  -17.20    0.03 1.80  -21.16  -18.34  -17.08  -15.93  -14.06
#> log_p_z[15,3]  -24.31    0.07 3.59  -32.25  -26.53  -24.00  -21.72  -18.16
#> log_p_z[15,4]  -21.73    0.06 3.45  -29.30  -23.83  -21.45  -19.29  -15.78
#> log_p_z[16,1]   -3.79    0.01 0.52   -4.92   -4.12   -3.76   -3.43   -2.88
#> log_p_z[16,2]  -17.66    0.04 2.15  -22.46  -18.87  -17.48  -16.18  -14.09
#> log_p_z[16,3]  -24.83    0.08 4.35  -34.58  -27.54  -24.40  -21.74  -17.57
#> log_p_z[16,4]  -21.83    0.08 4.25  -31.37  -24.31  -21.36  -18.82  -15.03
#> log_p_z[17,1]   -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_p_z[17,2]  -19.95    0.04 2.26  -25.05  -21.31  -19.77  -18.38  -16.07
#> log_p_z[17,3]  -24.88    0.08 4.36  -34.56  -27.52  -24.44  -21.73  -17.67
#> log_p_z[17,4]  -21.78    0.07 4.23  -31.05  -24.18  -21.35  -18.77  -14.81
#> log_p_z[18,1]   -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_p_z[18,2]  -19.95    0.04 2.26  -25.05  -21.31  -19.77  -18.38  -16.07
#> log_p_z[18,3]  -24.88    0.08 4.36  -34.56  -27.52  -24.44  -21.73  -17.67
#> log_p_z[18,4]  -21.78    0.07 4.23  -31.05  -24.18  -21.35  -18.77  -14.81
#> log_p_z[19,1]  -16.44    0.03 1.94  -20.61  -17.67  -16.26  -15.05  -13.17
#> log_p_z[19,2]   -4.65    0.01 0.60   -5.90   -5.04   -4.62   -4.23   -3.60
#> log_p_z[19,3]  -22.06    0.09 4.41  -32.25  -24.65  -21.49  -18.89  -14.86
#> log_p_z[19,4]  -21.77    0.08 4.24  -31.06  -24.20  -21.36  -18.77  -14.83
#> log_p_z[20,1]  -18.08    0.04 2.27  -22.97  -19.52  -17.88  -16.51  -14.14
#> log_p_z[20,2]   -7.10    0.01 0.94   -9.13   -7.69   -7.03   -6.43   -5.47
#> log_p_z[20,3]  -19.91    0.08 4.18  -29.53  -22.27  -19.39  -16.93  -13.19
#> log_p_z[20,4]  -21.56    0.08 4.11  -30.99  -23.91  -21.11  -18.67  -14.89
#> log_p_z[21,1]  -19.68    0.04 2.38  -24.96  -21.12  -19.54  -18.02  -15.63
#> log_p_z[21,2]   -3.04    0.00 0.36   -3.79   -3.27   -3.02   -2.79   -2.39
#> log_p_z[21,3]  -21.10    0.08 4.26  -30.88  -23.56  -20.58  -18.05  -14.33
#> log_p_z[21,4]  -21.84    0.08 4.22  -31.40  -24.31  -21.38  -18.91  -14.96
#> log_p_z[22,1]  -16.44    0.03 1.94  -20.61  -17.67  -16.26  -15.05  -13.17
#> log_p_z[22,2]   -4.65    0.01 0.60   -5.90   -5.04   -4.62   -4.23   -3.60
#> log_p_z[22,3]  -22.06    0.09 4.41  -32.25  -24.65  -21.49  -18.89  -14.86
#> log_p_z[22,4]  -21.77    0.08 4.24  -31.06  -24.20  -21.36  -18.77  -14.83
#> log_p_z[23,1]  -21.52    0.05 2.61  -27.33  -23.04  -21.31  -19.69  -16.95
#> log_p_z[23,2]   -3.72    0.01 0.42   -4.61   -3.98   -3.70   -3.44   -2.92
#> log_p_z[23,3]  -18.61    0.08 4.17  -28.10  -20.99  -18.10  -15.57  -12.10
#> log_p_z[23,4]  -21.80    0.08 4.19  -31.23  -24.14  -21.38  -18.86  -14.77
#> log_p_z[24,1]  -17.48    0.04 2.18  -22.42  -18.73  -17.31  -15.91  -13.87
#> log_p_z[24,2]   -5.79    0.01 0.64   -7.30   -6.14   -5.72   -5.36   -4.72
#> log_p_z[24,3]  -21.67    0.07 3.34  -29.08  -23.71  -21.37  -19.29  -16.08
#> log_p_z[24,4]  -21.89    0.07 3.55  -29.87  -23.99  -21.57  -19.34  -15.85
#> log_p_z[25,1]   -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_p_z[25,2]  -19.95    0.04 2.26  -25.05  -21.31  -19.77  -18.38  -16.07
#> log_p_z[25,3]  -24.88    0.08 4.36  -34.56  -27.52  -24.44  -21.73  -17.67
#> log_p_z[25,4]  -21.78    0.07 4.23  -31.05  -24.18  -21.35  -18.77  -14.81
#> log_p_z[26,1]   -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_p_z[26,2]  -19.95    0.04 2.26  -25.05  -21.31  -19.77  -18.38  -16.07
#> log_p_z[26,3]  -24.88    0.08 4.36  -34.56  -27.52  -24.44  -21.73  -17.67
#> log_p_z[26,4]  -21.78    0.07 4.23  -31.05  -24.18  -21.35  -18.77  -14.81
#> log_p_z[27,1]  -21.96    0.05 2.46  -27.45  -23.45  -21.76  -20.25  -17.71
#> log_p_z[27,2]   -5.64    0.01 0.53   -6.77   -5.96   -5.61   -5.27   -4.73
#> log_p_z[27,3]  -18.30    0.06 3.15  -25.42  -20.20  -18.00  -16.03  -13.08
#> log_p_z[27,4]  -21.60    0.07 3.49  -29.26  -23.67  -21.30  -19.17  -15.62
#> log_p_z[28,1]   -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_p_z[28,2]  -19.95    0.04 2.26  -25.05  -21.31  -19.77  -18.38  -16.07
#> log_p_z[28,3]  -24.88    0.08 4.36  -34.56  -27.52  -24.44  -21.73  -17.67
#> log_p_z[28,4]  -21.78    0.07 4.23  -31.05  -24.18  -21.35  -18.77  -14.81
#> log_p_z[29,1]   -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_p_z[29,2]  -19.95    0.04 2.26  -25.05  -21.31  -19.77  -18.38  -16.07
#> log_p_z[29,3]  -24.88    0.08 4.36  -34.56  -27.52  -24.44  -21.73  -17.67
#> log_p_z[29,4]  -21.78    0.07 4.23  -31.05  -24.18  -21.35  -18.77  -14.81
#> log_p_z[30,1]   -6.83    0.01 0.85   -8.76   -7.34   -6.73   -6.23   -5.44
#> log_p_z[30,2]  -14.69    0.03 1.65  -18.41  -15.70  -14.54  -13.55  -11.89
#> log_p_z[30,3]  -24.28    0.07 3.58  -31.98  -26.55  -23.95  -21.72  -18.11
#> log_p_z[30,4]  -21.80    0.06 3.52  -29.40  -23.95  -21.53  -19.23  -15.88
#> log_p_z[31,1]   -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_p_z[31,2]  -19.95    0.04 2.26  -25.05  -21.31  -19.77  -18.38  -16.07
#> log_p_z[31,3]  -24.88    0.08 4.36  -34.56  -27.52  -24.44  -21.73  -17.67
#> log_p_z[31,4]  -21.78    0.07 4.23  -31.05  -24.18  -21.35  -18.77  -14.81
#> log_p_z[32,1]  -29.82    0.07 4.35  -39.73  -32.35  -29.29  -26.73  -22.68
#> log_p_z[32,2]  -14.81    0.03 1.71  -18.50  -15.85  -14.64  -13.62  -11.86
#> log_p_z[32,3]   -6.08    0.02 0.99   -8.28   -6.67   -6.00   -5.38   -4.46
#> log_p_z[32,4]  -20.80    0.09 4.41  -30.59  -23.49  -20.33  -17.59  -13.57
#> log_p_z[33,1]   -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_p_z[33,2]  -19.95    0.04 2.26  -25.05  -21.31  -19.77  -18.38  -16.07
#> log_p_z[33,3]  -24.88    0.08 4.36  -34.56  -27.52  -24.44  -21.73  -17.67
#> log_p_z[33,4]  -21.78    0.07 4.23  -31.05  -24.18  -21.35  -18.77  -14.81
#> log_p_z[34,1]  -19.68    0.04 2.38  -24.96  -21.12  -19.54  -18.02  -15.63
#> log_p_z[34,2]   -3.04    0.00 0.36   -3.79   -3.27   -3.02   -2.79   -2.39
#> log_p_z[34,3]  -21.10    0.08 4.26  -30.88  -23.56  -20.58  -18.05  -14.33
#> log_p_z[34,4]  -21.84    0.08 4.22  -31.40  -24.31  -21.38  -18.91  -14.96
#> log_p_z[35,1]  -22.64    0.05 2.74  -28.77  -24.29  -22.43  -20.74  -17.98
#> log_p_z[35,2]   -5.53    0.01 0.67   -6.99   -5.94   -5.47   -5.06   -4.34
#> log_p_z[35,3]  -16.03    0.08 4.01  -25.21  -18.24  -15.49  -13.11   -9.80
#> log_p_z[35,4]  -21.79    0.08 4.14  -30.95  -24.21  -21.38  -18.93  -15.02
#> log_p_z[36,1]  -29.80    0.07 3.72  -38.16  -31.98  -29.45  -27.22  -23.51
#> log_p_z[36,2]  -21.90    0.04 2.31  -27.07  -23.33  -21.72  -20.27  -17.90
#> log_p_z[36,3]  -10.52    0.04 1.76  -15.03  -11.37  -10.21   -9.31   -7.97
#> log_p_z[36,4]  -14.04    0.07 3.19  -21.20  -15.98  -13.73  -11.69   -8.81
#> log_p_z[37,1]  -18.60    0.04 2.30  -23.76  -20.02  -18.41  -16.98  -14.63
#> log_p_z[37,2]   -7.60    0.01 0.80   -9.35   -8.09   -7.55   -7.03   -6.21
#> log_p_z[37,3]  -19.09    0.06 3.15  -26.15  -21.07  -18.70  -16.82  -13.75
#> log_p_z[37,4]  -21.87    0.07 3.51  -29.71  -23.95  -21.58  -19.37  -15.82
#> log_p_z[38,1]  -25.26    0.04 2.78  -31.44  -27.01  -25.00  -23.34  -20.45
#> log_p_z[38,2]   -9.61    0.02 0.99  -11.81  -10.21   -9.53   -8.92   -7.90
#> log_p_z[38,3]  -11.68    0.05 2.75  -17.95  -13.24  -11.32   -9.64   -7.47
#> log_p_z[38,4]  -21.28    0.07 3.46  -28.80  -23.38  -21.00  -18.85  -15.27
#> log_p_z[39,1]  -29.47    0.07 4.35  -39.53  -31.92  -28.92  -26.48  -22.43
#> log_p_z[39,2]  -16.79    0.03 1.98  -21.16  -17.99  -16.60  -15.39  -13.40
#> log_p_z[39,3]   -7.27    0.02 1.10   -9.60   -7.96   -7.19   -6.51   -5.30
#> log_p_z[39,4]  -18.92    0.09 4.17  -28.66  -21.35  -18.34  -15.95  -12.19
#> log_p_z[40,1]   -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_p_z[40,2]  -19.95    0.04 2.26  -25.05  -21.31  -19.77  -18.38  -16.07
#> log_p_z[40,3]  -24.88    0.08 4.36  -34.56  -27.52  -24.44  -21.73  -17.67
#> log_p_z[40,4]  -21.78    0.07 4.23  -31.05  -24.18  -21.35  -18.77  -14.81
#> log_p_z[41,1]   -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_p_z[41,2]  -19.95    0.04 2.26  -25.05  -21.31  -19.77  -18.38  -16.07
#> log_p_z[41,3]  -24.88    0.08 4.36  -34.56  -27.52  -24.44  -21.73  -17.67
#> log_p_z[41,4]  -21.78    0.07 4.23  -31.05  -24.18  -21.35  -18.77  -14.81
#> log_p_z[42,1]   -6.00    0.01 0.60   -7.32   -6.38   -5.96   -5.58   -4.96
#> log_p_z[42,2]  -14.91    0.03 1.66  -18.58  -15.91  -14.78  -13.75  -12.06
#> log_p_z[42,3]  -24.27    0.07 3.54  -32.14  -26.44  -23.99  -21.73  -18.16
#> log_p_z[42,4]  -21.78    0.07 3.49  -29.45  -23.94  -21.49  -19.33  -15.79
#> log_p_z[43,1]  -21.96    0.05 2.46  -27.45  -23.45  -21.76  -20.25  -17.71
#> log_p_z[43,2]   -5.64    0.01 0.53   -6.77   -5.96   -5.61   -5.27   -4.73
#> log_p_z[43,3]  -18.30    0.06 3.15  -25.42  -20.20  -18.00  -16.03  -13.08
#> log_p_z[43,4]  -21.60    0.07 3.49  -29.26  -23.67  -21.30  -19.17  -15.62
#> log_p_z[44,1]   -4.41    0.01 0.41   -5.29   -4.66   -4.37   -4.12   -3.69
#> log_p_z[44,2]  -17.20    0.03 1.80  -21.16  -18.34  -17.08  -15.93  -14.06
#> log_p_z[44,3]  -24.31    0.07 3.59  -32.25  -26.53  -24.00  -21.72  -18.16
#> log_p_z[44,4]  -21.73    0.06 3.45  -29.30  -23.83  -21.45  -19.29  -15.78
#> log_p_z[45,1]  -19.68    0.04 2.38  -24.96  -21.12  -19.54  -18.02  -15.63
#> log_p_z[45,2]   -3.04    0.00 0.36   -3.79   -3.27   -3.02   -2.79   -2.39
#> log_p_z[45,3]  -21.10    0.08 4.26  -30.88  -23.56  -20.58  -18.05  -14.33
#> log_p_z[45,4]  -21.84    0.08 4.22  -31.40  -24.31  -21.38  -18.91  -14.96
#> log_lik[1]      -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_lik[2]      -7.59    0.02 1.12  -10.00   -8.30   -7.51   -6.79   -5.63
#> log_lik[3]      -9.96    0.01 0.92  -11.86  -10.53   -9.92   -9.34   -8.21
#> log_lik[4]      -7.56    0.01 0.92   -9.50   -8.15   -7.51   -6.93   -5.90
#> log_lik[5]      -3.72    0.01 0.42   -4.61   -3.98   -3.70   -3.44   -2.92
#> log_lik[6]      -5.49    0.01 0.64   -6.85   -5.89   -5.44   -5.05   -4.35
#> log_lik[7]      -6.00    0.01 0.60   -7.32   -6.38   -5.96   -5.58   -4.96
#> log_lik[8]      -5.73    0.01 0.82   -7.50   -6.24   -5.68   -5.16   -4.28
#> log_lik[9]      -4.52    0.01 0.59   -5.76   -4.89   -4.46   -4.10   -3.52
#> log_lik[10]     -7.11    0.01 0.71   -8.64   -7.54   -7.04   -6.63   -5.86
#> log_lik[11]     -5.83    0.01 0.97   -7.88   -6.43   -5.75   -5.14   -4.17
#> log_lik[12]     -9.34    0.02 1.04  -11.63   -9.98   -9.26   -8.63   -7.46
#> log_lik[13]     -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_lik[14]     -6.23    0.01 0.85   -8.10   -6.72   -6.16   -5.64   -4.82
#> log_lik[15]     -4.41    0.01 0.41   -5.29   -4.66   -4.37   -4.12   -3.69
#> log_lik[16]     -3.79    0.01 0.52   -4.92   -4.12   -3.76   -3.43   -2.88
#> log_lik[17]     -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_lik[18]     -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_lik[19]     -4.65    0.01 0.59   -5.90   -5.04   -4.62   -4.23   -3.60
#> log_lik[20]     -7.10    0.01 0.94   -9.13   -7.69   -7.03   -6.43   -5.47
#> log_lik[21]     -3.04    0.00 0.36   -3.79   -3.27   -3.02   -2.79   -2.39
#> log_lik[22]     -4.65    0.01 0.59   -5.90   -5.04   -4.62   -4.23   -3.60
#> log_lik[23]     -3.72    0.01 0.42   -4.61   -3.98   -3.70   -3.44   -2.92
#> log_lik[24]     -5.79    0.01 0.64   -7.30   -6.14   -5.72   -5.36   -4.72
#> log_lik[25]     -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_lik[26]     -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_lik[27]     -5.64    0.01 0.53   -6.77   -5.96   -5.61   -5.27   -4.73
#> log_lik[28]     -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_lik[29]     -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_lik[30]     -6.83    0.01 0.84   -8.76   -7.34   -6.73   -6.23   -5.44
#> log_lik[31]     -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_lik[32]     -6.08    0.02 0.98   -8.28   -6.67   -6.00   -5.38   -4.46
#> log_lik[33]     -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_lik[34]     -3.04    0.00 0.36   -3.79   -3.27   -3.02   -2.79   -2.39
#> log_lik[35]     -5.52    0.01 0.67   -6.97   -5.94   -5.47   -5.06   -4.34
#> log_lik[36]     -9.93    0.02 1.18  -12.44  -10.67   -9.87   -9.11   -7.85
#> log_lik[37]     -7.60    0.01 0.80   -9.35   -8.09   -7.55   -7.03   -6.21
#> log_lik[38]     -8.99    0.01 0.86  -10.70   -9.55   -8.99   -8.44   -7.26
#> log_lik[39]     -7.27    0.02 1.10   -9.59   -7.96   -7.19   -6.51   -5.30
#> log_lik[40]     -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_lik[41]     -2.20    0.00 0.29   -2.83   -2.39   -2.19   -1.99   -1.65
#> log_lik[42]     -6.00    0.01 0.60   -7.32   -6.38   -5.96   -5.58   -4.96
#> log_lik[43]     -5.64    0.01 0.53   -6.77   -5.96   -5.61   -5.27   -4.73
#> log_lik[44]     -4.41    0.01 0.41   -5.29   -4.66   -4.37   -4.12   -3.69
#> log_lik[45]     -3.04    0.00 0.36   -3.79   -3.27   -3.02   -2.79   -2.39
#> lp__          -463.77    0.17 6.19 -477.06 -467.79 -463.39 -459.47 -452.73
#>               n_eff Rhat
#> pi[1]          5092    1
#> pi[2]          4935    1
#> pi[3]          4231    1
#> pi[4]          5129    1
#> theta[1,1,1]   5465    1
#> theta[1,1,2]   5132    1
#> theta[1,1,3]   4472    1
#> theta[1,1,4]   4029    1
#> theta[1,2,1]   4975    1
#> theta[1,2,2]   5341    1
#> theta[1,2,3]   5711    1
#> theta[1,2,4]   4729    1
#> theta[1,3,1]   4888    1
#> theta[1,3,2]   3575    1
#> theta[1,3,3]   4530    1
#> theta[1,3,4]   4836    1
#> theta[1,4,1]   5695    1
#> theta[1,4,2]   5326    1
#> theta[1,4,3]   3396    1
#> theta[1,4,4]   4349    1
#> theta[2,1,1]   5045    1
#> theta[2,1,2]   5245    1
#> theta[2,1,3]   3910    1
#> theta[2,1,4]   4937    1
#> theta[2,2,1]   5500    1
#> theta[2,2,2]   5401    1
#> theta[2,2,3]   5237    1
#> theta[2,2,4]   5867    1
#> theta[2,3,1]   6042    1
#> theta[2,3,2]   5421    1
#> theta[2,3,3]   4781    1
#> theta[2,3,4]   5271    1
#> theta[2,4,1]   6084    1
#> theta[2,4,2]   5585    1
#> theta[2,4,3]   5288    1
#> theta[2,4,4]   5113    1
#> theta[3,1,1]   4736    1
#> theta[3,1,2]   4162    1
#> theta[3,1,3]   3919    1
#> theta[3,1,4]   3823    1
#> theta[3,2,1]   5324    1
#> theta[3,2,2]   5455    1
#> theta[3,2,3]   5028    1
#> theta[3,2,4]   5348    1
#> theta[3,3,1]   4911    1
#> theta[3,3,2]   5097    1
#> theta[3,3,3]   4960    1
#> theta[3,3,4]   4609    1
#> theta[3,4,1]   5381    1
#> theta[3,4,2]   5811    1
#> theta[3,4,3]   3981    1
#> theta[3,4,4]   5776    1
#> theta[4,1,1]   5373    1
#> theta[4,1,2]   4185    1
#> theta[4,1,3]   3618    1
#> theta[4,1,4]   4500    1
#> theta[4,2,1]   6091    1
#> theta[4,2,2]   5349    1
#> theta[4,2,3]   4859    1
#> theta[4,2,4]   4850    1
#> theta[4,3,1]   5396    1
#> theta[4,3,2]   5483    1
#> theta[4,3,3]   5217    1
#> theta[4,3,4]   5291    1
#> theta[4,4,1]   5346    1
#> theta[4,4,2]   5853    1
#> theta[4,4,3]   5246    1
#> theta[4,4,4]   5518    1
#> theta[5,1,1]   5647    1
#> theta[5,1,2]   4004    1
#> theta[5,1,3]   4852    1
#> theta[5,1,4]   3987    1
#> theta[5,2,1]   5873    1
#> theta[5,2,2]   5096    1
#> theta[5,2,3]   5076    1
#> theta[5,2,4]   4646    1
#> theta[5,3,1]   5050    1
#> theta[5,3,2]   5759    1
#> theta[5,3,3]   5458    1
#> theta[5,3,4]   5285    1
#> theta[5,4,1]   5509    1
#> theta[5,4,2]   5713    1
#> theta[5,4,3]   4632    1
#> theta[5,4,4]   4924    1
#> log_p_z[1,1]   5098    1
#> log_p_z[1,2]   3779    1
#> log_p_z[1,3]   2793    1
#> log_p_z[1,4]   3286    1
#> log_p_z[2,1]   3598    1
#> log_p_z[2,2]   4179    1
#> log_p_z[2,3]   3233    1
#> log_p_z[2,4]   2442    1
#> log_p_z[3,1]   2790    1
#> log_p_z[3,2]   3064    1
#> log_p_z[3,3]   2941    1
#> log_p_z[3,4]   2751    1
#> log_p_z[4,1]   3867    1
#> log_p_z[4,2]   4304    1
#> log_p_z[4,3]   2551    1
#> log_p_z[4,4]   2905    1
#> log_p_z[5,1]   3053    1
#> log_p_z[5,2]   5271    1
#> log_p_z[5,3]   2615    1
#> log_p_z[5,4]   2764    1
#> log_p_z[6,1]   3073    1
#> log_p_z[6,2]   4320    1
#> log_p_z[6,3]   2592    1
#> log_p_z[6,4]   2871    1
#> log_p_z[7,1]   4501    1
#> log_p_z[7,2]   3612    1
#> log_p_z[7,3]   2589    1
#> log_p_z[7,4]   2854    1
#> log_p_z[8,1]   3889    1
#> log_p_z[8,2]   3932    1
#> log_p_z[8,3]   4822    1
#> log_p_z[8,4]   2254    1
#> log_p_z[9,1]   3818    1
#> log_p_z[9,2]   4568    1
#> log_p_z[9,3]   2593    1
#> log_p_z[9,4]   3016    1
#> log_p_z[10,1]  3517    1
#> log_p_z[10,2]  4220    1
#> log_p_z[10,3]  2600    1
#> log_p_z[10,4]  2879    1
#> log_p_z[11,1]  3421    1
#> log_p_z[11,2]  2955    1
#> log_p_z[11,3]  2145    1
#> log_p_z[11,4]  5244    1
#> log_p_z[12,1]  4207    1
#> log_p_z[12,2]  3181    1
#> log_p_z[12,3]  2661    1
#> log_p_z[12,4]  3179    1
#> log_p_z[13,1]  5098    1
#> log_p_z[13,2]  3779    1
#> log_p_z[13,3]  2793    1
#> log_p_z[13,4]  3286    1
#> log_p_z[14,1]  3051    1
#> log_p_z[14,2]  4078    1
#> log_p_z[14,3]  2530    1
#> log_p_z[14,4]  2977    1
#> log_p_z[15,1]  4729    1
#> log_p_z[15,2]  3843    1
#> log_p_z[15,3]  2521    1
#> log_p_z[15,4]  3086    1
#> log_p_z[16,1]  4776    1
#> log_p_z[16,2]  3613    1
#> log_p_z[16,3]  2836    1
#> log_p_z[16,4]  3074    1
#> log_p_z[17,1]  5098    1
#> log_p_z[17,2]  3779    1
#> log_p_z[17,3]  2793    1
#> log_p_z[17,4]  3286    1
#> log_p_z[18,1]  5098    1
#> log_p_z[18,2]  3779    1
#> log_p_z[18,3]  2793    1
#> log_p_z[18,4]  3286    1
#> log_p_z[19,1]  3205    1
#> log_p_z[19,2]  4622    1
#> log_p_z[19,3]  2599    1
#> log_p_z[19,4]  2986    1
#> log_p_z[20,1]  2857    1
#> log_p_z[20,2]  4216    1
#> log_p_z[20,3]  2440    1
#> log_p_z[20,4]  2940    1
#> log_p_z[21,1]  2939    1
#> log_p_z[21,2]  5644    1
#> log_p_z[21,3]  2607    1
#> log_p_z[21,4]  2868    1
#> log_p_z[22,1]  3205    1
#> log_p_z[22,2]  4622    1
#> log_p_z[22,3]  2599    1
#> log_p_z[22,4]  2986    1
#> log_p_z[23,1]  3053    1
#> log_p_z[23,2]  5271    1
#> log_p_z[23,3]  2615    1
#> log_p_z[23,4]  2764    1
#> log_p_z[24,1]  2799    1
#> log_p_z[24,2]  3894    1
#> log_p_z[24,3]  2564    1
#> log_p_z[24,4]  2728    1
#> log_p_z[25,1]  5098    1
#> log_p_z[25,2]  3779    1
#> log_p_z[25,3]  2793    1
#> log_p_z[25,4]  3286    1
#> log_p_z[26,1]  5098    1
#> log_p_z[26,2]  3779    1
#> log_p_z[26,3]  2793    1
#> log_p_z[26,4]  3286    1
#> log_p_z[27,1]  2763    1
#> log_p_z[27,2]  5521    1
#> log_p_z[27,3]  2631    1
#> log_p_z[27,4]  2778    1
#> log_p_z[28,1]  5098    1
#> log_p_z[28,2]  3779    1
#> log_p_z[28,3]  2793    1
#> log_p_z[28,4]  3286    1
#> log_p_z[29,1]  5098    1
#> log_p_z[29,2]  3779    1
#> log_p_z[29,3]  2793    1
#> log_p_z[29,4]  3286    1
#> log_p_z[30,1]  3875    1
#> log_p_z[30,2]  3591    1
#> log_p_z[30,3]  2680    1
#> log_p_z[30,4]  3024    1
#> log_p_z[31,1]  5098    1
#> log_p_z[31,2]  3779    1
#> log_p_z[31,3]  2793    1
#> log_p_z[31,4]  3286    1
#> log_p_z[32,1]  3731    1
#> log_p_z[32,2]  4444    1
#> log_p_z[32,3]  3887    1
#> log_p_z[32,4]  2266    1
#> log_p_z[33,1]  5098    1
#> log_p_z[33,2]  3779    1
#> log_p_z[33,3]  2793    1
#> log_p_z[33,4]  3286    1
#> log_p_z[34,1]  2939    1
#> log_p_z[34,2]  5644    1
#> log_p_z[34,3]  2607    1
#> log_p_z[34,4]  2868    1
#> log_p_z[35,1]  3273    1
#> log_p_z[35,2]  4969    1
#> log_p_z[35,3]  2638    1
#> log_p_z[35,4]  2914    1
#> log_p_z[36,1]  3213    1
#> log_p_z[36,2]  3360    1
#> log_p_z[36,3]  1988    1
#> log_p_z[36,4]  2285    1
#> log_p_z[37,1]  3023    1
#> log_p_z[37,2]  4262    1
#> log_p_z[37,3]  2563    1
#> log_p_z[37,4]  2833    1
#> log_p_z[38,1]  3935    1
#> log_p_z[38,2]  3982    1
#> log_p_z[38,3]  2616    1
#> log_p_z[38,4]  2724    1
#> log_p_z[39,1]  3596    1
#> log_p_z[39,2]  4348    1
#> log_p_z[39,3]  4145    1
#> log_p_z[39,4]  2304    1
#> log_p_z[40,1]  5098    1
#> log_p_z[40,2]  3779    1
#> log_p_z[40,3]  2793    1
#> log_p_z[40,4]  3286    1
#> log_p_z[41,1]  5098    1
#> log_p_z[41,2]  3779    1
#> log_p_z[41,3]  2793    1
#> log_p_z[41,4]  3286    1
#> log_p_z[42,1]  4501    1
#> log_p_z[42,2]  3612    1
#> log_p_z[42,3]  2589    1
#> log_p_z[42,4]  2854    1
#> log_p_z[43,1]  2763    1
#> log_p_z[43,2]  5521    1
#> log_p_z[43,3]  2631    1
#> log_p_z[43,4]  2778    1
#> log_p_z[44,1]  4729    1
#> log_p_z[44,2]  3843    1
#> log_p_z[44,3]  2521    1
#> log_p_z[44,4]  3086    1
#> log_p_z[45,1]  2939    1
#> log_p_z[45,2]  5644    1
#> log_p_z[45,3]  2607    1
#> log_p_z[45,4]  2868    1
#> log_lik[1]     5098    1
#> log_lik[2]     3808    1
#> log_lik[3]     4745    1
#> log_lik[4]     4371    1
#> log_lik[5]     5272    1
#> log_lik[6]     4332    1
#> log_lik[7]     4503    1
#> log_lik[8]     4821    1
#> log_lik[9]     4569    1
#> log_lik[10]    4265    1
#> log_lik[11]    5245    1
#> log_lik[12]    4443    1
#> log_lik[13]    5098    1
#> log_lik[14]    4082    1
#> log_lik[15]    4729    1
#> log_lik[16]    4776    1
#> log_lik[17]    5098    1
#> log_lik[18]    5098    1
#> log_lik[19]    4622    1
#> log_lik[20]    4222    1
#> log_lik[21]    5644    1
#> log_lik[22]    4622    1
#> log_lik[23]    5272    1
#> log_lik[24]    3894    1
#> log_lik[25]    5098    1
#> log_lik[26]    5098    1
#> log_lik[27]    5521    1
#> log_lik[28]    5098    1
#> log_lik[29]    5098    1
#> log_lik[30]    3877    1
#> log_lik[31]    5098    1
#> log_lik[32]    3895    1
#> log_lik[33]    5098    1
#> log_lik[34]    5644    1
#> log_lik[35]    4996    1
#> log_lik[36]    3930    1
#> log_lik[37]    4265    1
#> log_lik[38]    4621    1
#> log_lik[39]    4152    1
#> log_lik[40]    5098    1
#> log_lik[41]    5098    1
#> log_lik[42]    4503    1
#> log_lik[43]    5521    1
#> log_lik[44]    4729    1
#> log_lik[45]    5644    1
#> lp__           1403    1
#> 
#> Samples were drawn using NUTS(diag_e) at Wed Feb  8 09:55:31 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }