Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.37    0.00 0.07    0.25    0.33    0.37    0.42    0.50
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.08    0.00 0.03    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.12    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.13
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.83    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.00    0.01    0.02    0.05
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.07    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.59    0.73    0.80    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.40
#> theta[1,4,4]     0.69    0.00 0.14    0.40    0.59    0.70    0.79    0.91
#> theta[2,1,1]     0.75    0.00 0.09    0.56    0.69    0.76    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.58    0.64    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.24    0.30    0.36    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.26
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[2,3,3]     0.62    0.00 0.13    0.36    0.54    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.12    0.20    0.29    0.48
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.39
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,3]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,4]     0.65    0.00 0.15    0.32    0.55    0.67    0.77    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.02    0.05    0.08    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.88
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.25    0.43
#> theta[3,3,3]     0.51    0.00 0.14    0.24    0.41    0.51    0.60    0.77
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.30    0.49
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[3,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[3,4,3]     0.14    0.00 0.12    0.00    0.05    0.11    0.20    0.45
#> theta[3,4,4]     0.63    0.00 0.16    0.31    0.52    0.64    0.74    0.90
#> theta[4,1,1]     0.82    0.00 0.08    0.66    0.78    0.83    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.24
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[4,2,2]     0.72    0.00 0.09    0.54    0.67    0.73    0.79    0.88
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.08    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.20
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,2]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.28
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.60    0.70    0.78    0.91
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.38
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,3]     0.12    0.00 0.10    0.01    0.04    0.09    0.17    0.37
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.90
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.82    0.87    0.91    0.97
#> theta[5,1,2]     0.06    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.81
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.22    0.39
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.61    0.70    0.85
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[5,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.42
#> theta[5,4,4]     0.63    0.00 0.15    0.34    0.53    0.64    0.75    0.89
#> log_p_z[1,1]    -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_p_z[1,2]   -20.04    0.04 2.32  -25.22  -21.44  -19.82  -18.40  -16.02
#> log_p_z[1,3]   -24.87    0.07 4.49  -35.14  -27.54  -24.41  -21.66  -17.52
#> log_p_z[1,4]   -21.89    0.07 4.28  -31.16  -24.44  -21.43  -18.86  -14.84
#> log_p_z[2,1]   -29.73    0.07 4.45  -40.05  -32.21  -29.23  -26.69  -22.52
#> log_p_z[2,2]   -21.07    0.04 2.42  -26.49  -22.51  -20.89  -19.35  -16.94
#> log_p_z[2,3]    -7.66    0.02 1.21  -10.34   -8.36   -7.53   -6.83   -5.63
#> log_p_z[2,4]   -16.25    0.08 4.10  -25.45  -18.70  -15.78  -13.34   -9.64
#> log_p_z[3,1]   -11.61    0.03 1.66  -15.46  -12.61  -11.43  -10.44   -8.95
#> log_p_z[3,2]   -10.84    0.02 1.48  -14.21  -11.70  -10.65   -9.76   -8.46
#> log_p_z[3,3]   -23.41    0.06 3.62  -31.55  -25.61  -23.09  -20.84  -17.37
#> log_p_z[3,4]   -21.85    0.06 3.49  -29.68  -24.02  -21.50  -19.39  -16.01
#> log_p_z[4,1]   -14.63    0.03 1.89  -18.66  -15.82  -14.48  -13.30  -11.34
#> log_p_z[4,2]    -7.58    0.01 0.94   -9.54   -8.13   -7.50   -6.93   -5.98
#> log_p_z[4,3]   -20.71    0.08 4.37  -30.83  -23.16  -20.17  -17.56  -13.81
#> log_p_z[4,4]   -21.96    0.08 4.43  -32.61  -24.42  -21.41  -18.87  -14.99
#> log_p_z[5,1]   -21.44    0.04 2.59  -27.06  -23.06  -21.20  -19.60  -16.83
#> log_p_z[5,2]    -3.71    0.01 0.43   -4.59   -4.01   -3.69   -3.41   -2.91
#> log_p_z[5,3]   -18.72    0.07 4.28  -28.41  -21.24  -18.09  -15.66  -12.07
#> log_p_z[5,4]   -22.03    0.07 4.38  -32.16  -24.35  -21.50  -18.97  -15.09
#> log_p_z[6,1]   -21.44    0.04 2.58  -27.02  -23.11  -21.20  -19.59  -16.91
#> log_p_z[6,2]    -5.47    0.01 0.67   -6.93   -5.90   -5.42   -5.01   -4.30
#> log_p_z[6,3]   -17.52    0.07 4.14  -26.98  -19.83  -16.87  -14.58  -11.25
#> log_p_z[6,4]   -21.77    0.07 4.38  -32.14  -24.22  -21.22  -18.76  -14.78
#> log_p_z[7,1]    -6.01    0.01 0.63   -7.38   -6.39   -5.95   -5.57   -4.93
#> log_p_z[7,2]   -14.98    0.03 1.73  -18.74  -16.06  -14.85  -13.73  -12.02
#> log_p_z[7,3]   -24.31    0.06 3.68  -32.56  -26.61  -24.01  -21.76  -17.98
#> log_p_z[7,4]   -21.86    0.06 3.57  -29.46  -24.16  -21.49  -19.29  -15.78
#> log_p_z[8,1]   -29.72    0.07 4.39  -39.53  -32.28  -29.19  -26.69  -22.65
#> log_p_z[8,2]   -18.32    0.03 2.14  -23.16  -19.65  -18.17  -16.80  -14.65
#> log_p_z[8,3]    -5.72    0.01 0.84   -7.54   -6.24   -5.67   -5.14   -4.19
#> log_p_z[8,4]   -18.42    0.07 4.14  -28.13  -20.76  -17.93  -15.49  -11.82
#> log_p_z[9,1]   -19.97    0.03 2.29  -24.69  -21.36  -19.85  -18.36  -15.93
#> log_p_z[9,2]    -4.52    0.01 0.57   -5.71   -4.89   -4.47   -4.12   -3.52
#> log_p_z[9,3]   -19.66    0.08 4.38  -29.44  -22.20  -19.11  -16.48  -12.81
#> log_p_z[9,4]   -21.74    0.07 4.29  -31.92  -24.10  -21.23  -18.74  -14.86
#> log_p_z[10,1]  -22.23    0.04 2.46  -27.60  -23.69  -22.01  -20.54  -17.94
#> log_p_z[10,2]   -7.13    0.01 0.70   -8.61   -7.56   -7.08   -6.63   -5.90
#> log_p_z[10,3]  -16.83    0.06 3.22  -24.05  -18.75  -16.46  -14.47  -11.65
#> log_p_z[10,4]  -21.37    0.06 3.51  -29.40  -23.46  -21.03  -18.90  -15.49
#> log_p_z[11,1]  -29.83    0.07 4.48  -40.01  -32.32  -29.36  -26.76  -22.47
#> log_p_z[11,2]  -29.34    0.07 4.28  -38.85  -31.84  -28.85  -26.31  -22.45
#> log_p_z[11,3]  -18.19    0.07 3.33  -26.11  -19.87  -17.77  -15.85  -13.11
#> log_p_z[11,4]   -5.79    0.01 0.94   -7.81   -6.37   -5.73   -5.11   -4.16
#> log_p_z[12,1]  -22.95    0.04 2.74  -28.76  -24.63  -22.76  -21.01  -18.23
#> log_p_z[12,2]   -9.59    0.02 1.31  -12.56  -10.33   -9.47   -8.70   -7.43
#> log_p_z[12,3]  -15.17    0.07 4.07  -24.53  -17.45  -14.63  -12.23   -8.93
#> log_p_z[12,4]  -19.32    0.07 4.24  -29.27  -21.54  -18.83  -16.35  -12.52
#> log_p_z[13,1]   -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_p_z[13,2]  -20.04    0.04 2.32  -25.22  -21.44  -19.82  -18.40  -16.02
#> log_p_z[13,3]  -24.87    0.07 4.49  -35.14  -27.54  -24.41  -21.66  -17.52
#> log_p_z[13,4]  -21.89    0.07 4.28  -31.16  -24.44  -21.43  -18.86  -14.84
#> log_p_z[14,1]  -19.05    0.03 2.42  -24.40  -20.52  -18.89  -17.37  -14.79
#> log_p_z[14,2]   -6.24    0.01 0.87   -8.20   -6.75   -6.16   -5.64   -4.72
#> log_p_z[14,3]  -18.71    0.07 4.27  -28.46  -21.15  -18.11  -15.64  -12.14
#> log_p_z[14,4]  -22.03    0.07 4.36  -32.14  -24.40  -21.52  -18.97  -15.06
#> log_p_z[15,1]   -4.42    0.01 0.41   -5.28   -4.66   -4.38   -4.14   -3.69
#> log_p_z[15,2]  -17.26    0.03 1.86  -21.26  -18.41  -17.11  -15.95  -14.00
#> log_p_z[15,3]  -24.31    0.06 3.65  -32.46  -26.49  -24.04  -21.81  -18.01
#> log_p_z[15,4]  -21.92    0.06 3.57  -29.68  -24.20  -21.58  -19.36  -15.84
#> log_p_z[16,1]   -3.81    0.01 0.55   -5.00   -4.15   -3.77   -3.42   -2.85
#> log_p_z[16,2]  -17.75    0.03 2.22  -22.70  -19.14  -17.52  -16.17  -14.03
#> log_p_z[16,3]  -24.88    0.07 4.51  -35.34  -27.56  -24.40  -21.70  -17.51
#> log_p_z[16,4]  -21.82    0.07 4.30  -31.47  -24.40  -21.34  -18.80  -14.71
#> log_p_z[17,1]   -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_p_z[17,2]  -20.04    0.04 2.32  -25.22  -21.44  -19.82  -18.40  -16.02
#> log_p_z[17,3]  -24.87    0.07 4.49  -35.14  -27.54  -24.41  -21.66  -17.52
#> log_p_z[17,4]  -21.89    0.07 4.28  -31.16  -24.44  -21.43  -18.86  -14.84
#> log_p_z[18,1]   -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_p_z[18,2]  -20.04    0.04 2.32  -25.22  -21.44  -19.82  -18.40  -16.02
#> log_p_z[18,3]  -24.87    0.07 4.49  -35.14  -27.54  -24.41  -21.66  -17.52
#> log_p_z[18,4]  -21.89    0.07 4.28  -31.16  -24.44  -21.43  -18.86  -14.84
#> log_p_z[19,1]  -16.39    0.03 1.91  -20.51  -17.58  -16.29  -15.00  -13.01
#> log_p_z[19,2]   -4.66    0.01 0.59   -5.89   -5.03   -4.63   -4.24   -3.58
#> log_p_z[19,3]  -22.12    0.08 4.47  -32.06  -24.75  -21.59  -18.85  -15.16
#> log_p_z[19,4]  -22.00    0.07 4.35  -32.10  -24.41  -21.48  -18.99  -15.05
#> log_p_z[20,1]  -18.02    0.03 2.27  -23.03  -19.41  -17.83  -16.43  -14.13
#> log_p_z[20,2]   -7.09    0.01 0.96   -9.09   -7.71   -7.02   -6.40   -5.41
#> log_p_z[20,3]  -20.00    0.07 4.26  -29.78  -22.47  -19.48  -16.98  -13.45
#> log_p_z[20,4]  -21.81    0.07 4.37  -31.97  -24.26  -21.20  -18.78  -14.75
#> log_p_z[21,1]  -19.61    0.03 2.34  -24.66  -21.09  -19.40  -17.96  -15.51
#> log_p_z[21,2]   -3.04    0.00 0.36   -3.77   -3.28   -3.03   -2.80   -2.37
#> log_p_z[21,3]  -21.21    0.08 4.41  -31.18  -23.80  -20.63  -17.99  -14.36
#> log_p_z[21,4]  -22.00    0.07 4.31  -32.16  -24.31  -21.49  -19.00  -15.19
#> log_p_z[22,1]  -16.39    0.03 1.91  -20.51  -17.58  -16.29  -15.00  -13.01
#> log_p_z[22,2]   -4.66    0.01 0.59   -5.89   -5.03   -4.63   -4.24   -3.58
#> log_p_z[22,3]  -22.12    0.08 4.47  -32.06  -24.75  -21.59  -18.85  -15.16
#> log_p_z[22,4]  -22.00    0.07 4.35  -32.10  -24.41  -21.48  -18.99  -15.05
#> log_p_z[23,1]  -21.44    0.04 2.59  -27.06  -23.06  -21.20  -19.60  -16.83
#> log_p_z[23,2]   -3.71    0.01 0.43   -4.59   -4.01   -3.69   -3.41   -2.91
#> log_p_z[23,3]  -18.72    0.07 4.28  -28.41  -21.24  -18.09  -15.66  -12.07
#> log_p_z[23,4]  -22.03    0.07 4.38  -32.16  -24.35  -21.50  -18.97  -15.09
#> log_p_z[24,1]  -17.41    0.03 2.14  -22.08  -18.74  -17.18  -15.89  -13.77
#> log_p_z[24,2]   -5.82    0.01 0.64   -7.26   -6.19   -5.74   -5.37   -4.77
#> log_p_z[24,3]  -21.77    0.06 3.50  -29.55  -23.93  -21.47  -19.21  -16.02
#> log_p_z[24,4]  -21.96    0.06 3.51  -29.68  -24.00  -21.55  -19.49  -16.20
#> log_p_z[25,1]   -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_p_z[25,2]  -20.04    0.04 2.32  -25.22  -21.44  -19.82  -18.40  -16.02
#> log_p_z[25,3]  -24.87    0.07 4.49  -35.14  -27.54  -24.41  -21.66  -17.52
#> log_p_z[25,4]  -21.89    0.07 4.28  -31.16  -24.44  -21.43  -18.86  -14.84
#> log_p_z[26,1]   -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_p_z[26,2]  -20.04    0.04 2.32  -25.22  -21.44  -19.82  -18.40  -16.02
#> log_p_z[26,3]  -24.87    0.07 4.49  -35.14  -27.54  -24.41  -21.66  -17.52
#> log_p_z[26,4]  -21.89    0.07 4.28  -31.16  -24.44  -21.43  -18.86  -14.84
#> log_p_z[27,1]  -21.86    0.04 2.49  -27.24  -23.45  -21.63  -20.09  -17.59
#> log_p_z[27,2]   -5.65    0.01 0.53   -6.82   -5.97   -5.62   -5.29   -4.74
#> log_p_z[27,3]  -18.38    0.06 3.28  -25.61  -20.38  -17.98  -15.98  -13.18
#> log_p_z[27,4]  -21.63    0.06 3.47  -29.44  -23.62  -21.31  -19.20  -15.87
#> log_p_z[28,1]   -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_p_z[28,2]  -20.04    0.04 2.32  -25.22  -21.44  -19.82  -18.40  -16.02
#> log_p_z[28,3]  -24.87    0.07 4.49  -35.14  -27.54  -24.41  -21.66  -17.52
#> log_p_z[28,4]  -21.89    0.07 4.28  -31.16  -24.44  -21.43  -18.86  -14.84
#> log_p_z[29,1]   -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_p_z[29,2]  -20.04    0.04 2.32  -25.22  -21.44  -19.82  -18.40  -16.02
#> log_p_z[29,3]  -24.87    0.07 4.49  -35.14  -27.54  -24.41  -21.66  -17.52
#> log_p_z[29,4]  -21.89    0.07 4.28  -31.16  -24.44  -21.43  -18.86  -14.84
#> log_p_z[30,1]   -6.80    0.01 0.84   -8.72   -7.29   -6.72   -6.21   -5.40
#> log_p_z[30,2]  -14.74    0.03 1.68  -18.46  -15.84  -14.53  -13.58  -11.92
#> log_p_z[30,3]  -24.32    0.06 3.67  -32.52  -26.55  -24.01  -21.73  -18.12
#> log_p_z[30,4]  -21.92    0.06 3.58  -29.69  -24.17  -21.58  -19.36  -15.85
#> log_p_z[31,1]   -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_p_z[31,2]  -20.04    0.04 2.32  -25.22  -21.44  -19.82  -18.40  -16.02
#> log_p_z[31,3]  -24.87    0.07 4.49  -35.14  -27.54  -24.41  -21.66  -17.52
#> log_p_z[31,4]  -21.89    0.07 4.28  -31.16  -24.44  -21.43  -18.86  -14.84
#> log_p_z[32,1]  -29.73    0.07 4.44  -39.75  -32.22  -29.26  -26.63  -22.55
#> log_p_z[32,2]  -14.85    0.02 1.71  -18.60  -15.88  -14.69  -13.65  -11.84
#> log_p_z[32,3]   -6.05    0.01 0.96   -8.17   -6.66   -5.98   -5.36   -4.36
#> log_p_z[32,4]  -20.61    0.08 4.36  -30.64  -23.18  -20.17  -17.48  -13.65
#> log_p_z[33,1]   -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_p_z[33,2]  -20.04    0.04 2.32  -25.22  -21.44  -19.82  -18.40  -16.02
#> log_p_z[33,3]  -24.87    0.07 4.49  -35.14  -27.54  -24.41  -21.66  -17.52
#> log_p_z[33,4]  -21.89    0.07 4.28  -31.16  -24.44  -21.43  -18.86  -14.84
#> log_p_z[34,1]  -19.61    0.03 2.34  -24.66  -21.09  -19.40  -17.96  -15.51
#> log_p_z[34,2]   -3.04    0.00 0.36   -3.77   -3.28   -3.03   -2.80   -2.37
#> log_p_z[34,3]  -21.21    0.08 4.41  -31.18  -23.80  -20.63  -17.99  -14.36
#> log_p_z[34,4]  -22.00    0.07 4.31  -32.16  -24.31  -21.49  -19.00  -15.19
#> log_p_z[35,1]  -22.59    0.04 2.69  -28.32  -24.30  -22.41  -20.69  -17.87
#> log_p_z[35,2]   -5.54    0.01 0.69   -7.02   -5.95   -5.50   -5.06   -4.31
#> log_p_z[35,3]  -16.09    0.07 4.13  -25.90  -18.46  -15.56  -13.04   -9.83
#> log_p_z[35,4]  -21.97    0.07 4.34  -32.39  -24.31  -21.50  -18.93  -15.06
#> log_p_z[36,1]  -29.77    0.06 3.74  -38.12  -31.93  -29.39  -27.12  -23.47
#> log_p_z[36,2]  -21.90    0.04 2.27  -26.76  -23.27  -21.69  -20.34  -17.97
#> log_p_z[36,3]  -10.53    0.04 1.76  -14.75  -11.43  -10.25   -9.32   -7.85
#> log_p_z[36,4]  -13.89    0.06 3.21  -20.87  -15.95  -13.61  -11.50   -8.71
#> log_p_z[37,1]  -18.56    0.03 2.28  -23.41  -20.02  -18.39  -16.94  -14.53
#> log_p_z[37,2]   -7.64    0.01 0.83   -9.51   -8.14   -7.57   -7.05   -6.26
#> log_p_z[37,3]  -19.14    0.06 3.30  -26.52  -21.18  -18.81  -16.74  -13.82
#> log_p_z[37,4]  -21.91    0.06 3.48  -29.84  -23.96  -21.53  -19.49  -16.24
#> log_p_z[38,1]  -25.22    0.04 2.83  -31.54  -26.97  -25.01  -23.24  -20.27
#> log_p_z[38,2]   -9.62    0.01 0.98  -11.80  -10.20   -9.55   -8.94   -7.93
#> log_p_z[38,3]  -11.71    0.05 2.79  -18.24  -13.35  -11.35   -9.62   -7.55
#> log_p_z[38,4]  -21.35    0.06 3.54  -29.33  -23.40  -20.98  -18.89  -15.30
#> log_p_z[39,1]  -29.35    0.07 4.34  -38.94  -31.86  -28.85  -26.31  -22.44
#> log_p_z[39,2]  -16.85    0.03 2.04  -21.39  -18.11  -16.65  -15.39  -13.34
#> log_p_z[39,3]   -7.28    0.02 1.08   -9.64   -7.98   -7.19   -6.52   -5.39
#> log_p_z[39,4]  -18.68    0.07 4.04  -27.92  -21.00  -18.21  -15.78  -12.20
#> log_p_z[40,1]   -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_p_z[40,2]  -20.04    0.04 2.32  -25.22  -21.44  -19.82  -18.40  -16.02
#> log_p_z[40,3]  -24.87    0.07 4.49  -35.14  -27.54  -24.41  -21.66  -17.52
#> log_p_z[40,4]  -21.89    0.07 4.28  -31.16  -24.44  -21.43  -18.86  -14.84
#> log_p_z[41,1]   -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_p_z[41,2]  -20.04    0.04 2.32  -25.22  -21.44  -19.82  -18.40  -16.02
#> log_p_z[41,3]  -24.87    0.07 4.49  -35.14  -27.54  -24.41  -21.66  -17.52
#> log_p_z[41,4]  -21.89    0.07 4.28  -31.16  -24.44  -21.43  -18.86  -14.84
#> log_p_z[42,1]   -6.01    0.01 0.63   -7.38   -6.39   -5.95   -5.57   -4.93
#> log_p_z[42,2]  -14.98    0.03 1.73  -18.74  -16.06  -14.85  -13.73  -12.02
#> log_p_z[42,3]  -24.31    0.06 3.68  -32.56  -26.61  -24.01  -21.76  -17.98
#> log_p_z[42,4]  -21.86    0.06 3.57  -29.46  -24.16  -21.49  -19.29  -15.78
#> log_p_z[43,1]  -21.86    0.04 2.49  -27.24  -23.45  -21.63  -20.09  -17.59
#> log_p_z[43,2]   -5.65    0.01 0.53   -6.82   -5.97   -5.62   -5.29   -4.74
#> log_p_z[43,3]  -18.38    0.06 3.28  -25.61  -20.38  -17.98  -15.98  -13.18
#> log_p_z[43,4]  -21.63    0.06 3.47  -29.44  -23.62  -21.31  -19.20  -15.87
#> log_p_z[44,1]   -4.42    0.01 0.41   -5.28   -4.66   -4.38   -4.14   -3.69
#> log_p_z[44,2]  -17.26    0.03 1.86  -21.26  -18.41  -17.11  -15.95  -14.00
#> log_p_z[44,3]  -24.31    0.06 3.65  -32.46  -26.49  -24.04  -21.81  -18.01
#> log_p_z[44,4]  -21.92    0.06 3.57  -29.68  -24.20  -21.58  -19.36  -15.84
#> log_p_z[45,1]  -19.61    0.03 2.34  -24.66  -21.09  -19.40  -17.96  -15.51
#> log_p_z[45,2]   -3.04    0.00 0.36   -3.77   -3.28   -3.03   -2.80   -2.37
#> log_p_z[45,3]  -21.21    0.08 4.41  -31.18  -23.80  -20.63  -17.99  -14.36
#> log_p_z[45,4]  -22.00    0.07 4.31  -32.16  -24.31  -21.49  -19.00  -15.19
#> log_lik[1]      -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_lik[2]      -7.61    0.02 1.14  -10.10   -8.31   -7.52   -6.83   -5.63
#> log_lik[3]      -9.95    0.01 0.91  -11.87  -10.53   -9.92   -9.31   -8.28
#> log_lik[4]      -7.57    0.01 0.92   -9.52   -8.13   -7.50   -6.93   -5.98
#> log_lik[5]      -3.71    0.01 0.43   -4.59   -4.01   -3.69   -3.41   -2.91
#> log_lik[6]      -5.47    0.01 0.67   -6.92   -5.90   -5.42   -5.01   -4.30
#> log_lik[7]      -6.01    0.01 0.63   -7.38   -6.39   -5.95   -5.57   -4.93
#> log_lik[8]      -5.72    0.01 0.84   -7.54   -6.24   -5.66   -5.14   -4.19
#> log_lik[9]      -4.52    0.01 0.57   -5.71   -4.89   -4.47   -4.12   -3.52
#> log_lik[10]     -7.12    0.01 0.70   -8.60   -7.56   -7.08   -6.63   -5.90
#> log_lik[11]     -5.79    0.01 0.94   -7.81   -6.37   -5.73   -5.11   -4.16
#> log_lik[12]     -9.33    0.02 1.06  -11.56  -10.00   -9.30   -8.59   -7.38
#> log_lik[13]     -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_lik[14]     -6.24    0.01 0.86   -8.20   -6.75   -6.16   -5.64   -4.72
#> log_lik[15]     -4.42    0.01 0.41   -5.28   -4.66   -4.38   -4.14   -3.69
#> log_lik[16]     -3.81    0.01 0.55   -5.00   -4.15   -3.77   -3.42   -2.85
#> log_lik[17]     -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_lik[18]     -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_lik[19]     -4.66    0.01 0.59   -5.89   -5.03   -4.63   -4.24   -3.58
#> log_lik[20]     -7.09    0.01 0.96   -9.08   -7.71   -7.02   -6.40   -5.41
#> log_lik[21]     -3.04    0.00 0.36   -3.77   -3.28   -3.03   -2.80   -2.37
#> log_lik[22]     -4.66    0.01 0.59   -5.89   -5.03   -4.63   -4.24   -3.58
#> log_lik[23]     -3.71    0.01 0.43   -4.59   -4.01   -3.69   -3.41   -2.91
#> log_lik[24]     -5.82    0.01 0.64   -7.26   -6.19   -5.74   -5.37   -4.77
#> log_lik[25]     -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_lik[26]     -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_lik[27]     -5.65    0.01 0.53   -6.82   -5.97   -5.62   -5.29   -4.74
#> log_lik[28]     -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_lik[29]     -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_lik[30]     -6.80    0.01 0.84   -8.71   -7.28   -6.71   -6.21   -5.40
#> log_lik[31]     -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_lik[32]     -6.05    0.01 0.96   -8.17   -6.66   -5.98   -5.36   -4.36
#> log_lik[33]     -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_lik[34]     -3.04    0.00 0.36   -3.77   -3.28   -3.03   -2.80   -2.37
#> log_lik[35]     -5.54    0.01 0.69   -7.00   -5.95   -5.49   -5.06   -4.31
#> log_lik[36]     -9.91    0.02 1.20  -12.36  -10.66   -9.84   -9.09   -7.77
#> log_lik[37]     -7.64    0.01 0.82   -9.51   -8.14   -7.57   -7.05   -6.26
#> log_lik[38]     -9.01    0.01 0.85  -10.72   -9.56   -9.01   -8.44   -7.34
#> log_lik[39]     -7.27    0.02 1.08   -9.63   -7.97   -7.19   -6.52   -5.39
#> log_lik[40]     -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_lik[41]     -2.22    0.00 0.31   -2.85   -2.41   -2.20   -1.99   -1.66
#> log_lik[42]     -6.01    0.01 0.63   -7.38   -6.39   -5.95   -5.57   -4.93
#> log_lik[43]     -5.65    0.01 0.53   -6.82   -5.97   -5.62   -5.29   -4.74
#> log_lik[44]     -4.42    0.01 0.41   -5.28   -4.66   -4.38   -4.14   -3.69
#> log_lik[45]     -3.04    0.00 0.36   -3.77   -3.28   -3.03   -2.80   -2.37
#> lp__          -464.01    0.16 6.15 -477.36 -467.94 -463.74 -459.64 -452.98
#>               n_eff Rhat
#> pi[1]          8120    1
#> pi[2]          6404    1
#> pi[3]          6189    1
#> pi[4]          5642    1
#> theta[1,1,1]   8133    1
#> theta[1,1,2]   7238    1
#> theta[1,1,3]   4734    1
#> theta[1,1,4]   4418    1
#> theta[1,2,1]   5948    1
#> theta[1,2,2]   7570    1
#> theta[1,2,3]   7222    1
#> theta[1,2,4]   5371    1
#> theta[1,3,1]   7155    1
#> theta[1,3,2]   4351    1
#> theta[1,3,3]   4494    1
#> theta[1,3,4]   5741    1
#> theta[1,4,1]   5894    1
#> theta[1,4,2]   6459    1
#> theta[1,4,3]   3948    1
#> theta[1,4,4]   5349    1
#> theta[2,1,1]   7502    1
#> theta[2,1,2]   6487    1
#> theta[2,1,3]   4619    1
#> theta[2,1,4]   4967    1
#> theta[2,2,1]   7437    1
#> theta[2,2,2]   8021    1
#> theta[2,2,3]   6322    1
#> theta[2,2,4]   5648    1
#> theta[2,3,1]   5936    1
#> theta[2,3,2]   6425    1
#> theta[2,3,3]   5962    1
#> theta[2,3,4]   5791    1
#> theta[2,4,1]   7015    1
#> theta[2,4,2]   7868    1
#> theta[2,4,3]   6176    1
#> theta[2,4,4]   7041    1
#> theta[3,1,1]   7838    1
#> theta[3,1,2]   5673    1
#> theta[3,1,3]   4905    1
#> theta[3,1,4]   4239    1
#> theta[3,2,1]   6534    1
#> theta[3,2,2]   7960    1
#> theta[3,2,3]   7089    1
#> theta[3,2,4]   6319    1
#> theta[3,3,1]   6859    1
#> theta[3,3,2]   7665    1
#> theta[3,3,3]   6500    1
#> theta[3,3,4]   6758    1
#> theta[3,4,1]   6383    1
#> theta[3,4,2]   4967    1
#> theta[3,4,3]   5316    1
#> theta[3,4,4]   5569    1
#> theta[4,1,1]   8084    1
#> theta[4,1,2]   5463    1
#> theta[4,1,3]   4433    1
#> theta[4,1,4]   4858    1
#> theta[4,2,1]   7647    1
#> theta[4,2,2]   7132    1
#> theta[4,2,3]   6140    1
#> theta[4,2,4]   5537    1
#> theta[4,3,1]   5561    1
#> theta[4,3,2]   6499    1
#> theta[4,3,3]   6779    1
#> theta[4,3,4]   5652    1
#> theta[4,4,1]   6169    1
#> theta[4,4,2]   6594    1
#> theta[4,4,3]   7491    1
#> theta[4,4,4]   7770    1
#> theta[5,1,1]   6717    1
#> theta[5,1,2]   5106    1
#> theta[5,1,3]   5120    1
#> theta[5,1,4]   4425    1
#> theta[5,2,1]   7735    1
#> theta[5,2,2]   6735    1
#> theta[5,2,3]   6199    1
#> theta[5,2,4]   5678    1
#> theta[5,3,1]   6328    1
#> theta[5,3,2]   7518    1
#> theta[5,3,3]   6889    1
#> theta[5,3,4]   5913    1
#> theta[5,4,1]   6329    1
#> theta[5,4,2]   6738    1
#> theta[5,4,3]   5313    1
#> theta[5,4,4]   5978    1
#> log_p_z[1,1]   7541    1
#> log_p_z[1,2]   4174    1
#> log_p_z[1,3]   4030    1
#> log_p_z[1,4]   3628    1
#> log_p_z[2,1]   4411    1
#> log_p_z[2,2]   3896    1
#> log_p_z[2,3]   4054    1
#> log_p_z[2,4]   2837    1
#> log_p_z[3,1]   4100    1
#> log_p_z[3,2]   3737    1
#> log_p_z[3,3]   3556    1
#> log_p_z[3,4]   3494    1
#> log_p_z[4,1]   5619    1
#> log_p_z[4,2]   4278    1
#> log_p_z[4,3]   3243    1
#> log_p_z[4,4]   3470    1
#> log_p_z[5,1]   4739    1
#> log_p_z[5,2]   5878    1
#> log_p_z[5,3]   3259    1
#> log_p_z[5,4]   3610    1
#> log_p_z[6,1]   4918    1
#> log_p_z[6,2]   5420    1
#> log_p_z[6,3]   3167    1
#> log_p_z[6,4]   3609    1
#> log_p_z[7,1]   6228    1
#> log_p_z[7,2]   4129    1
#> log_p_z[7,3]   3867    1
#> log_p_z[7,4]   3612    1
#> log_p_z[8,1]   4026    1
#> log_p_z[8,2]   4320    1
#> log_p_z[8,3]   5304    1
#> log_p_z[8,4]   3123    1
#> log_p_z[9,1]   4688    1
#> log_p_z[9,2]   5604    1
#> log_p_z[9,3]   3186    1
#> log_p_z[9,4]   3631    1
#> log_p_z[10,1]  4184    1
#> log_p_z[10,2]  4971    1
#> log_p_z[10,3]  3194    1
#> log_p_z[10,4]  3288    1
#> log_p_z[11,1]  4473    1
#> log_p_z[11,2]  3488    1
#> log_p_z[11,3]  2613    1
#> log_p_z[11,4]  6277    1
#> log_p_z[12,1]  4944    1
#> log_p_z[12,2]  4028    1
#> log_p_z[12,3]  3225    1
#> log_p_z[12,4]  3586    1
#> log_p_z[13,1]  7541    1
#> log_p_z[13,2]  4174    1
#> log_p_z[13,3]  4030    1
#> log_p_z[13,4]  3628    1
#> log_p_z[14,1]  4877    1
#> log_p_z[14,2]  5767    1
#> log_p_z[14,3]  3414    1
#> log_p_z[14,4]  3659    1
#> log_p_z[15,1]  6257    1
#> log_p_z[15,2]  4247    1
#> log_p_z[15,3]  4077    1
#> log_p_z[15,4]  3435    1
#> log_p_z[16,1]  6631    1
#> log_p_z[16,2]  4124    1
#> log_p_z[16,3]  3931    1
#> log_p_z[16,4]  3822    1
#> log_p_z[17,1]  7541    1
#> log_p_z[17,2]  4174    1
#> log_p_z[17,3]  4030    1
#> log_p_z[17,4]  3628    1
#> log_p_z[18,1]  7541    1
#> log_p_z[18,2]  4174    1
#> log_p_z[18,3]  4030    1
#> log_p_z[18,4]  3628    1
#> log_p_z[19,1]  5508    1
#> log_p_z[19,2]  6988    1
#> log_p_z[19,3]  3180    1
#> log_p_z[19,4]  3490    1
#> log_p_z[20,1]  4943    1
#> log_p_z[20,2]  5788    1
#> log_p_z[20,3]  3233    1
#> log_p_z[20,4]  3433    1
#> log_p_z[21,1]  4749    1
#> log_p_z[21,2]  6777    1
#> log_p_z[21,3]  3215    1
#> log_p_z[21,4]  3528    1
#> log_p_z[22,1]  5508    1
#> log_p_z[22,2]  6988    1
#> log_p_z[22,3]  3180    1
#> log_p_z[22,4]  3490    1
#> log_p_z[23,1]  4739    1
#> log_p_z[23,2]  5878    1
#> log_p_z[23,3]  3259    1
#> log_p_z[23,4]  3610    1
#> log_p_z[24,1]  4430    1
#> log_p_z[24,2]  4357    1
#> log_p_z[24,3]  3101    1
#> log_p_z[24,4]  3370    1
#> log_p_z[25,1]  7541    1
#> log_p_z[25,2]  4174    1
#> log_p_z[25,3]  4030    1
#> log_p_z[25,4]  3628    1
#> log_p_z[26,1]  7541    1
#> log_p_z[26,2]  4174    1
#> log_p_z[26,3]  4030    1
#> log_p_z[26,4]  3628    1
#> log_p_z[27,1]  4330    1
#> log_p_z[27,2]  5839    1
#> log_p_z[27,3]  3239    1
#> log_p_z[27,4]  3301    1
#> log_p_z[28,1]  7541    1
#> log_p_z[28,2]  4174    1
#> log_p_z[28,3]  4030    1
#> log_p_z[28,4]  3628    1
#> log_p_z[29,1]  7541    1
#> log_p_z[29,2]  4174    1
#> log_p_z[29,3]  4030    1
#> log_p_z[29,4]  3628    1
#> log_p_z[30,1]  5432    1
#> log_p_z[30,2]  4012    1
#> log_p_z[30,3]  3865    1
#> log_p_z[30,4]  3340    1
#> log_p_z[31,1]  7541    1
#> log_p_z[31,2]  4174    1
#> log_p_z[31,3]  4030    1
#> log_p_z[31,4]  3628    1
#> log_p_z[32,1]  4143    1
#> log_p_z[32,2]  5069    1
#> log_p_z[32,3]  6241    1
#> log_p_z[32,4]  3073    1
#> log_p_z[33,1]  7541    1
#> log_p_z[33,2]  4174    1
#> log_p_z[33,3]  4030    1
#> log_p_z[33,4]  3628    1
#> log_p_z[34,1]  4749    1
#> log_p_z[34,2]  6777    1
#> log_p_z[34,3]  3215    1
#> log_p_z[34,4]  3528    1
#> log_p_z[35,1]  4641    1
#> log_p_z[35,2]  5082    1
#> log_p_z[35,3]  3525    1
#> log_p_z[35,4]  3580    1
#> log_p_z[36,1]  3615    1
#> log_p_z[36,2]  3327    1
#> log_p_z[36,3]  2404    1
#> log_p_z[36,4]  2878    1
#> log_p_z[37,1]  4503    1
#> log_p_z[37,2]  4708    1
#> log_p_z[37,3]  3488    1
#> log_p_z[37,4]  3562    1
#> log_p_z[38,1]  4301    1
#> log_p_z[38,2]  4327    1
#> log_p_z[38,3]  3542    1
#> log_p_z[38,4]  3495    1
#> log_p_z[39,1]  4076    1
#> log_p_z[39,2]  4521    1
#> log_p_z[39,3]  5001    1
#> log_p_z[39,4]  3140    1
#> log_p_z[40,1]  7541    1
#> log_p_z[40,2]  4174    1
#> log_p_z[40,3]  4030    1
#> log_p_z[40,4]  3628    1
#> log_p_z[41,1]  7541    1
#> log_p_z[41,2]  4174    1
#> log_p_z[41,3]  4030    1
#> log_p_z[41,4]  3628    1
#> log_p_z[42,1]  6228    1
#> log_p_z[42,2]  4129    1
#> log_p_z[42,3]  3867    1
#> log_p_z[42,4]  3612    1
#> log_p_z[43,1]  4330    1
#> log_p_z[43,2]  5839    1
#> log_p_z[43,3]  3239    1
#> log_p_z[43,4]  3301    1
#> log_p_z[44,1]  6257    1
#> log_p_z[44,2]  4247    1
#> log_p_z[44,3]  4077    1
#> log_p_z[44,4]  3435    1
#> log_p_z[45,1]  4749    1
#> log_p_z[45,2]  6777    1
#> log_p_z[45,3]  3215    1
#> log_p_z[45,4]  3528    1
#> log_lik[1]     7541    1
#> log_lik[2]     4514    1
#> log_lik[3]     4912    1
#> log_lik[4]     4421    1
#> log_lik[5]     5879    1
#> log_lik[6]     5422    1
#> log_lik[7]     6230    1
#> log_lik[8]     5331    1
#> log_lik[9]     5604    1
#> log_lik[10]    5005    1
#> log_lik[11]    6277    1
#> log_lik[12]    4847    1
#> log_lik[13]    7541    1
#> log_lik[14]    5783    1
#> log_lik[15]    6257    1
#> log_lik[16]    6631    1
#> log_lik[17]    7541    1
#> log_lik[18]    7541    1
#> log_lik[19]    6988    1
#> log_lik[20]    5800    1
#> log_lik[21]    6777    1
#> log_lik[22]    6988    1
#> log_lik[23]    5879    1
#> log_lik[24]    4357    1
#> log_lik[25]    7541    1
#> log_lik[26]    7541    1
#> log_lik[27]    5840    1
#> log_lik[28]    7541    1
#> log_lik[29]    7541    1
#> log_lik[30]    5444    1
#> log_lik[31]    7541    1
#> log_lik[32]    6241    1
#> log_lik[33]    7541    1
#> log_lik[34]    6777    1
#> log_lik[35]    5109    1
#> log_lik[36]    4860    1
#> log_lik[37]    4710    1
#> log_lik[38]    4832    1
#> log_lik[39]    5019    1
#> log_lik[40]    7541    1
#> log_lik[41]    7541    1
#> log_lik[42]    6230    1
#> log_lik[43]    5840    1
#> log_lik[44]    6257    1
#> log_lik[45]    6777    1
#> lp__           1410    1
#> 
#> Samples were drawn using NUTS(diag_e) at Sat Feb 11 09:55:48 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }