Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.07    0.25    0.33    0.37    0.42    0.51
#> pi[2]            0.41    0.00 0.07    0.29    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.24
#> pi[4]            0.08    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.78    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.12    0.18
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.13
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.29
#> theta[1,3,3]     0.79    0.00 0.09    0.58    0.74    0.80    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.30
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.41
#> theta[1,4,4]     0.69    0.00 0.14    0.39    0.59    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.09    0.56    0.70    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.59    0.65    0.75
#> theta[2,2,3]     0.30    0.00 0.09    0.15    0.24    0.30    0.36    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,3]     0.62    0.00 0.13    0.37    0.53    0.62    0.71    0.85
#> theta[2,3,4]     0.22    0.00 0.11    0.04    0.13    0.20    0.29    0.46
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.08    0.17    0.38
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.77    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.10    0.00 0.06    0.02    0.05    0.09    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.56    0.67    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.28
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.14    0.25    0.41    0.50    0.60    0.77
#> theta[3,3,4]     0.23    0.00 0.12    0.05    0.14    0.22    0.31    0.50
#> theta[3,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[3,4,3]     0.14    0.00 0.12    0.00    0.05    0.11    0.21    0.42
#> theta[3,4,4]     0.63    0.00 0.15    0.31    0.52    0.64    0.75    0.90
#> theta[4,1,1]     0.83    0.00 0.07    0.67    0.78    0.83    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[4,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.78    0.87
#> theta[4,2,3]     0.13    0.00 0.06    0.04    0.08    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.18
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,3]     0.69    0.00 0.13    0.43    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.39
#> theta[4,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[4,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.67    0.77    0.91
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.82    0.87    0.92    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.08    0.00    0.02    0.06    0.11    0.28
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,3,3]     0.61    0.00 0.14    0.33    0.51    0.61    0.70    0.85
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.22    0.39
#> theta[5,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.41
#> theta[5,4,4]     0.63    0.00 0.15    0.31    0.53    0.65    0.75    0.90
#> log_p_z[1,1]    -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_p_z[1,2]   -19.98    0.04 2.32  -25.13  -21.35  -19.76  -18.33  -16.07
#> log_p_z[1,3]   -24.92    0.09 4.40  -34.70  -27.39  -24.47  -21.79  -17.90
#> log_p_z[1,4]   -21.94    0.09 4.38  -32.29  -24.43  -21.39  -18.85  -14.85
#> log_p_z[2,1]   -29.71    0.07 4.36  -39.53  -32.30  -29.17  -26.71  -22.59
#> log_p_z[2,2]   -21.05    0.04 2.41  -26.15  -22.56  -20.92  -19.33  -16.91
#> log_p_z[2,3]    -7.64    0.02 1.17  -10.22   -8.35   -7.56   -6.82   -5.61
#> log_p_z[2,4]   -16.33    0.08 4.15  -25.97  -18.74  -15.79  -13.38   -9.67
#> log_p_z[3,1]   -11.62    0.03 1.65  -15.57  -12.58  -11.40  -10.45   -9.03
#> log_p_z[3,2]   -10.77    0.03 1.50  -14.25  -11.63  -10.57   -9.70   -8.35
#> log_p_z[3,3]   -23.36    0.07 3.57  -31.40  -25.47  -23.01  -20.82  -17.41
#> log_p_z[3,4]   -21.98    0.07 3.65  -30.23  -24.27  -21.65  -19.33  -15.85
#> log_p_z[4,1]   -14.65    0.03 1.89  -18.85  -15.81  -14.52  -13.34  -11.35
#> log_p_z[4,2]    -7.55    0.02 0.92   -9.50   -8.11   -7.50   -6.90   -5.93
#> log_p_z[4,3]   -20.74    0.09 4.41  -30.95  -23.29  -20.13  -17.55  -13.91
#> log_p_z[4,4]   -22.01    0.08 4.38  -32.26  -24.43  -21.52  -18.94  -14.94
#> log_p_z[5,1]   -21.46    0.04 2.57  -27.15  -23.04  -21.31  -19.62  -16.97
#> log_p_z[5,2]    -3.71    0.01 0.42   -4.58   -3.98   -3.70   -3.42   -2.94
#> log_p_z[5,3]   -18.67    0.08 4.31  -28.67  -21.07  -18.11  -15.51  -12.03
#> log_p_z[5,4]   -21.95    0.08 4.44  -32.04  -24.46  -21.40  -18.87  -14.90
#> log_p_z[6,1]   -21.45    0.04 2.57  -26.89  -23.07  -21.23  -19.61  -17.04
#> log_p_z[6,2]    -5.47    0.01 0.65   -6.85   -5.87   -5.42   -5.02   -4.30
#> log_p_z[6,3]   -17.47    0.08 4.16  -27.23  -19.73  -16.86  -14.42  -11.17
#> log_p_z[6,4]   -21.77    0.08 4.40  -31.63  -24.23  -21.31  -18.67  -14.69
#> log_p_z[7,1]    -6.02    0.01 0.64   -7.43   -6.40   -5.97   -5.58   -4.92
#> log_p_z[7,2]   -14.92    0.03 1.70  -18.65  -15.92  -14.78  -13.71  -12.10
#> log_p_z[7,3]   -24.38    0.07 3.59  -32.40  -26.39  -24.07  -21.85  -18.43
#> log_p_z[7,4]   -22.00    0.07 3.65  -30.28  -24.17  -21.64  -19.41  -15.93
#> log_p_z[8,1]   -29.72    0.07 4.34  -39.46  -32.32  -29.29  -26.64  -22.57
#> log_p_z[8,2]   -18.30    0.04 2.12  -22.96  -19.63  -18.12  -16.82  -14.55
#> log_p_z[8,3]    -5.75    0.01 0.85   -7.51   -6.29   -5.70   -5.15   -4.23
#> log_p_z[8,4]   -18.52    0.08 4.19  -28.20  -20.94  -18.00  -15.52  -11.69
#> log_p_z[9,1]   -20.02    0.04 2.26  -24.94  -21.49  -19.81  -18.44  -16.10
#> log_p_z[9,2]    -4.51    0.01 0.57   -5.74   -4.86   -4.48   -4.11   -3.52
#> log_p_z[9,3]   -19.62    0.08 4.37  -29.61  -22.08  -19.05  -16.48  -12.76
#> log_p_z[9,4]   -21.77    0.08 4.41  -31.56  -24.34  -21.30  -18.65  -14.61
#> log_p_z[10,1]  -22.27    0.04 2.35  -27.26  -23.76  -22.10  -20.61  -18.09
#> log_p_z[10,2]   -7.12    0.01 0.70   -8.64   -7.56   -7.08   -6.62   -5.88
#> log_p_z[10,3]  -16.80    0.06 3.19  -24.21  -18.65  -16.48  -14.51  -11.64
#> log_p_z[10,4]  -21.40    0.07 3.65  -29.42  -23.60  -21.07  -18.85  -15.16
#> log_p_z[11,1]  -29.72    0.07 4.38  -39.45  -32.31  -29.31  -26.52  -22.60
#> log_p_z[11,2]  -29.37    0.08 4.38  -39.58  -31.87  -28.88  -26.26  -22.26
#> log_p_z[11,3]  -18.19    0.07 3.32  -26.28  -19.82  -17.67  -15.94  -13.22
#> log_p_z[11,4]   -5.81    0.01 0.99   -8.01   -6.41   -5.75   -5.11   -4.08
#> log_p_z[12,1]  -22.98    0.04 2.67  -28.78  -24.66  -22.74  -21.13  -18.42
#> log_p_z[12,2]   -9.58    0.02 1.26  -12.46  -10.27   -9.46   -8.71   -7.48
#> log_p_z[12,3]  -15.16    0.08 4.11  -24.42  -17.46  -14.62  -12.25   -8.95
#> log_p_z[12,4]  -19.37    0.08 4.23  -29.19  -21.70  -18.81  -16.43  -12.63
#> log_p_z[13,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_p_z[13,2]  -19.98    0.04 2.32  -25.13  -21.35  -19.76  -18.33  -16.07
#> log_p_z[13,3]  -24.92    0.09 4.40  -34.70  -27.39  -24.47  -21.79  -17.90
#> log_p_z[13,4]  -21.94    0.09 4.38  -32.29  -24.43  -21.39  -18.85  -14.85
#> log_p_z[14,1]  -19.07    0.04 2.39  -24.33  -20.53  -18.89  -17.38  -15.00
#> log_p_z[14,2]   -6.25    0.01 0.88   -8.20   -6.80   -6.17   -5.62   -4.75
#> log_p_z[14,3]  -18.72    0.08 4.28  -28.44  -21.26  -18.09  -15.60  -12.10
#> log_p_z[14,4]  -21.95    0.08 4.47  -32.12  -24.45  -21.45  -18.78  -14.68
#> log_p_z[15,1]   -4.42    0.01 0.41   -5.32   -4.66   -4.39   -4.14   -3.69
#> log_p_z[15,2]  -17.22    0.03 1.86  -21.34  -18.34  -17.08  -15.91  -14.06
#> log_p_z[15,3]  -24.37    0.07 3.60  -32.23  -26.50  -24.07  -21.80  -18.38
#> log_p_z[15,4]  -21.96    0.07 3.60  -30.05  -24.19  -21.64  -19.40  -15.87
#> log_p_z[16,1]   -3.80    0.01 0.55   -5.00   -4.14   -3.77   -3.41   -2.84
#> log_p_z[16,2]  -17.68    0.04 2.19  -22.68  -18.94  -17.46  -16.13  -14.04
#> log_p_z[16,3]  -24.93    0.09 4.40  -34.79  -27.35  -24.45  -21.88  -17.92
#> log_p_z[16,4]  -21.98    0.09 4.40  -32.38  -24.33  -21.40  -18.93  -15.02
#> log_p_z[17,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_p_z[17,2]  -19.98    0.04 2.32  -25.13  -21.35  -19.76  -18.33  -16.07
#> log_p_z[17,3]  -24.92    0.09 4.40  -34.70  -27.39  -24.47  -21.79  -17.90
#> log_p_z[17,4]  -21.94    0.09 4.38  -32.29  -24.43  -21.39  -18.85  -14.85
#> log_p_z[18,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_p_z[18,2]  -19.98    0.04 2.32  -25.13  -21.35  -19.76  -18.33  -16.07
#> log_p_z[18,3]  -24.92    0.09 4.40  -34.70  -27.39  -24.47  -21.79  -17.90
#> log_p_z[18,4]  -21.94    0.09 4.38  -32.29  -24.43  -21.39  -18.85  -14.85
#> log_p_z[19,1]  -16.44    0.03 1.94  -20.58  -17.64  -16.30  -15.09  -13.02
#> log_p_z[19,2]   -4.65    0.01 0.59   -5.91   -5.01   -4.60   -4.23   -3.62
#> log_p_z[19,3]  -22.14    0.08 4.48  -32.21  -24.72  -21.63  -18.93  -15.08
#> log_p_z[19,4]  -21.99    0.08 4.44  -32.21  -24.58  -21.52  -18.80  -14.80
#> log_p_z[20,1]  -18.04    0.04 2.29  -22.94  -19.49  -17.82  -16.41  -14.20
#> log_p_z[20,2]   -7.09    0.01 0.95   -9.14   -7.67   -7.01   -6.41   -5.42
#> log_p_z[20,3]  -19.95    0.08 4.27  -30.02  -22.25  -19.39  -16.89  -13.24
#> log_p_z[20,4]  -21.76    0.08 4.33  -31.61  -24.29  -21.28  -18.74  -14.65
#> log_p_z[21,1]  -19.65    0.04 2.37  -24.84  -21.11  -19.48  -18.01  -15.59
#> log_p_z[21,2]   -3.03    0.01 0.35   -3.77   -3.27   -3.02   -2.78   -2.37
#> log_p_z[21,3]  -21.17    0.08 4.40  -31.04  -23.62  -20.57  -17.91  -14.28
#> log_p_z[21,4]  -21.97    0.08 4.45  -31.91  -24.59  -21.49  -18.80  -14.79
#> log_p_z[22,1]  -16.44    0.03 1.94  -20.58  -17.64  -16.30  -15.09  -13.02
#> log_p_z[22,2]   -4.65    0.01 0.59   -5.91   -5.01   -4.60   -4.23   -3.62
#> log_p_z[22,3]  -22.14    0.08 4.48  -32.21  -24.72  -21.63  -18.93  -15.08
#> log_p_z[22,4]  -21.99    0.08 4.44  -32.21  -24.58  -21.52  -18.80  -14.80
#> log_p_z[23,1]  -21.46    0.04 2.57  -27.15  -23.04  -21.31  -19.62  -16.97
#> log_p_z[23,2]   -3.71    0.01 0.42   -4.58   -3.98   -3.70   -3.42   -2.94
#> log_p_z[23,3]  -18.67    0.08 4.31  -28.67  -21.07  -18.11  -15.51  -12.03
#> log_p_z[23,4]  -21.95    0.08 4.44  -32.04  -24.46  -21.40  -18.87  -14.90
#> log_p_z[24,1]  -17.44    0.04 2.16  -22.35  -18.75  -17.28  -15.92  -13.77
#> log_p_z[24,2]   -5.79    0.01 0.64   -7.24   -6.16   -5.71   -5.35   -4.76
#> log_p_z[24,3]  -21.71    0.07 3.50  -29.70  -23.78  -21.29  -19.25  -16.08
#> log_p_z[24,4]  -21.96    0.07 3.66  -29.92  -24.29  -21.69  -19.29  -15.75
#> log_p_z[25,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_p_z[25,2]  -19.98    0.04 2.32  -25.13  -21.35  -19.76  -18.33  -16.07
#> log_p_z[25,3]  -24.92    0.09 4.40  -34.70  -27.39  -24.47  -21.79  -17.90
#> log_p_z[25,4]  -21.94    0.09 4.38  -32.29  -24.43  -21.39  -18.85  -14.85
#> log_p_z[26,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_p_z[26,2]  -19.98    0.04 2.32  -25.13  -21.35  -19.76  -18.33  -16.07
#> log_p_z[26,3]  -24.92    0.09 4.40  -34.70  -27.39  -24.47  -21.79  -17.90
#> log_p_z[26,4]  -21.94    0.09 4.38  -32.29  -24.43  -21.39  -18.85  -14.85
#> log_p_z[27,1]  -21.90    0.04 2.47  -27.38  -23.42  -21.71  -20.09  -17.77
#> log_p_z[27,2]   -5.64    0.01 0.52   -6.77   -5.96   -5.60   -5.28   -4.72
#> log_p_z[27,3]  -18.35    0.06 3.25  -25.93  -20.21  -17.96  -15.97  -13.16
#> log_p_z[27,4]  -21.61    0.07 3.63  -29.52  -23.79  -21.38  -19.03  -15.44
#> log_p_z[28,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_p_z[28,2]  -19.98    0.04 2.32  -25.13  -21.35  -19.76  -18.33  -16.07
#> log_p_z[28,3]  -24.92    0.09 4.40  -34.70  -27.39  -24.47  -21.79  -17.90
#> log_p_z[28,4]  -21.94    0.09 4.38  -32.29  -24.43  -21.39  -18.85  -14.85
#> log_p_z[29,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_p_z[29,2]  -19.98    0.04 2.32  -25.13  -21.35  -19.76  -18.33  -16.07
#> log_p_z[29,3]  -24.92    0.09 4.40  -34.70  -27.39  -24.47  -21.79  -17.90
#> log_p_z[29,4]  -21.94    0.09 4.38  -32.29  -24.43  -21.39  -18.85  -14.85
#> log_p_z[30,1]   -6.81    0.01 0.85   -8.82   -7.30   -6.71   -6.21   -5.41
#> log_p_z[30,2]  -14.68    0.03 1.73  -18.45  -15.71  -14.55  -13.48  -11.79
#> log_p_z[30,3]  -24.32    0.07 3.64  -32.61  -26.42  -23.96  -21.79  -18.24
#> log_p_z[30,4]  -21.96    0.07 3.56  -30.01  -24.15  -21.58  -19.41  -15.96
#> log_p_z[31,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_p_z[31,2]  -19.98    0.04 2.32  -25.13  -21.35  -19.76  -18.33  -16.07
#> log_p_z[31,3]  -24.92    0.09 4.40  -34.70  -27.39  -24.47  -21.79  -17.90
#> log_p_z[31,4]  -21.94    0.09 4.38  -32.29  -24.43  -21.39  -18.85  -14.85
#> log_p_z[32,1]  -29.73    0.07 4.29  -39.31  -32.29  -29.30  -26.63  -22.69
#> log_p_z[32,2]  -14.83    0.03 1.72  -18.64  -15.89  -14.70  -13.63  -11.77
#> log_p_z[32,3]   -6.07    0.01 0.92   -8.03   -6.65   -5.98   -5.44   -4.50
#> log_p_z[32,4]  -20.63    0.08 4.38  -30.82  -23.18  -20.13  -17.60  -13.45
#> log_p_z[33,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_p_z[33,2]  -19.98    0.04 2.32  -25.13  -21.35  -19.76  -18.33  -16.07
#> log_p_z[33,3]  -24.92    0.09 4.40  -34.70  -27.39  -24.47  -21.79  -17.90
#> log_p_z[33,4]  -21.94    0.09 4.38  -32.29  -24.43  -21.39  -18.85  -14.85
#> log_p_z[34,1]  -19.65    0.04 2.37  -24.84  -21.11  -19.48  -18.01  -15.59
#> log_p_z[34,2]   -3.03    0.01 0.35   -3.77   -3.27   -3.02   -2.78   -2.37
#> log_p_z[34,3]  -21.17    0.08 4.40  -31.04  -23.62  -20.57  -17.91  -14.28
#> log_p_z[34,4]  -21.97    0.08 4.45  -31.91  -24.59  -21.49  -18.80  -14.79
#> log_p_z[35,1]  -22.61    0.05 2.75  -28.52  -24.24  -22.42  -20.70  -17.94
#> log_p_z[35,2]   -5.53    0.01 0.68   -7.02   -5.96   -5.49   -5.05   -4.35
#> log_p_z[35,3]  -16.08    0.08 4.17  -25.59  -18.39  -15.45  -13.02   -9.84
#> log_p_z[35,4]  -21.92    0.08 4.42  -32.07  -24.41  -21.34  -18.91  -14.53
#> log_p_z[36,1]  -29.74    0.06 3.64  -37.85  -31.86  -29.42  -27.23  -23.62
#> log_p_z[36,2]  -21.88    0.04 2.28  -26.85  -23.31  -21.70  -20.28  -17.90
#> log_p_z[36,3]  -10.52    0.04 1.70  -14.77  -11.39  -10.25   -9.32   -7.98
#> log_p_z[36,4]  -13.97    0.07 3.22  -21.01  -15.94  -13.67  -11.59   -8.75
#> log_p_z[37,1]  -18.59    0.04 2.35  -23.82  -19.96  -18.35  -16.94  -14.63
#> log_p_z[37,2]   -7.61    0.01 0.80   -9.33   -8.11   -7.54   -7.03   -6.22
#> log_p_z[37,3]  -19.12    0.06 3.28  -26.49  -21.05  -18.69  -16.76  -13.93
#> log_p_z[37,4]  -21.93    0.07 3.65  -29.80  -24.13  -21.62  -19.40  -15.62
#> log_p_z[38,1]  -25.23    0.04 2.77  -31.18  -26.98  -25.01  -23.28  -20.48
#> log_p_z[38,2]   -9.62    0.02 0.99  -11.75  -10.23   -9.54   -8.92   -7.89
#> log_p_z[38,3]  -11.71    0.05 2.82  -18.10  -13.34  -11.34   -9.64   -7.47
#> log_p_z[38,4]  -21.36    0.07 3.62  -29.38  -23.54  -21.05  -18.78  -15.18
#> log_p_z[39,1]  -29.36    0.07 4.36  -39.20  -31.96  -28.92  -26.19  -22.31
#> log_p_z[39,2]  -16.83    0.04 2.04  -21.38  -18.05  -16.64  -15.41  -13.28
#> log_p_z[39,3]   -7.29    0.02 1.12   -9.77   -7.98   -7.19   -6.54   -5.31
#> log_p_z[39,4]  -18.72    0.07 4.07  -28.01  -20.96  -18.25  -15.87  -12.17
#> log_p_z[40,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_p_z[40,2]  -19.98    0.04 2.32  -25.13  -21.35  -19.76  -18.33  -16.07
#> log_p_z[40,3]  -24.92    0.09 4.40  -34.70  -27.39  -24.47  -21.79  -17.90
#> log_p_z[40,4]  -21.94    0.09 4.38  -32.29  -24.43  -21.39  -18.85  -14.85
#> log_p_z[41,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_p_z[41,2]  -19.98    0.04 2.32  -25.13  -21.35  -19.76  -18.33  -16.07
#> log_p_z[41,3]  -24.92    0.09 4.40  -34.70  -27.39  -24.47  -21.79  -17.90
#> log_p_z[41,4]  -21.94    0.09 4.38  -32.29  -24.43  -21.39  -18.85  -14.85
#> log_p_z[42,1]   -6.02    0.01 0.64   -7.43   -6.40   -5.97   -5.58   -4.92
#> log_p_z[42,2]  -14.92    0.03 1.70  -18.65  -15.92  -14.78  -13.71  -12.10
#> log_p_z[42,3]  -24.38    0.07 3.59  -32.40  -26.39  -24.07  -21.85  -18.43
#> log_p_z[42,4]  -22.00    0.07 3.65  -30.28  -24.17  -21.64  -19.41  -15.93
#> log_p_z[43,1]  -21.90    0.04 2.47  -27.38  -23.42  -21.71  -20.09  -17.77
#> log_p_z[43,2]   -5.64    0.01 0.52   -6.77   -5.96   -5.60   -5.28   -4.72
#> log_p_z[43,3]  -18.35    0.06 3.25  -25.93  -20.21  -17.96  -15.97  -13.16
#> log_p_z[43,4]  -21.61    0.07 3.63  -29.52  -23.79  -21.38  -19.03  -15.44
#> log_p_z[44,1]   -4.42    0.01 0.41   -5.32   -4.66   -4.39   -4.14   -3.69
#> log_p_z[44,2]  -17.22    0.03 1.86  -21.34  -18.34  -17.08  -15.91  -14.06
#> log_p_z[44,3]  -24.37    0.07 3.60  -32.23  -26.50  -24.07  -21.80  -18.38
#> log_p_z[44,4]  -21.96    0.07 3.60  -30.05  -24.19  -21.64  -19.40  -15.87
#> log_p_z[45,1]  -19.65    0.04 2.37  -24.84  -21.11  -19.48  -18.01  -15.59
#> log_p_z[45,2]   -3.03    0.01 0.35   -3.77   -3.27   -3.02   -2.78   -2.37
#> log_p_z[45,3]  -21.17    0.08 4.40  -31.04  -23.62  -20.57  -17.91  -14.28
#> log_p_z[45,4]  -21.97    0.08 4.45  -31.91  -24.59  -21.49  -18.80  -14.79
#> log_lik[1]      -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_lik[2]      -7.60    0.02 1.12  -10.02   -8.31   -7.54   -6.81   -5.61
#> log_lik[3]      -9.92    0.01 0.90  -11.74  -10.52   -9.90   -9.30   -8.27
#> log_lik[4]      -7.54    0.02 0.91   -9.48   -8.10   -7.49   -6.90   -5.93
#> log_lik[5]      -3.71    0.01 0.42   -4.58   -3.98   -3.70   -3.42   -2.94
#> log_lik[6]      -5.47    0.01 0.65   -6.85   -5.87   -5.42   -5.02   -4.30
#> log_lik[7]      -6.02    0.01 0.63   -7.43   -6.40   -5.97   -5.58   -4.92
#> log_lik[8]      -5.75    0.01 0.85   -7.51   -6.29   -5.70   -5.15   -4.23
#> log_lik[9]      -4.51    0.01 0.57   -5.74   -4.86   -4.48   -4.11   -3.52
#> log_lik[10]     -7.12    0.01 0.70   -8.62   -7.56   -7.07   -6.62   -5.88
#> log_lik[11]     -5.81    0.01 0.99   -8.01   -6.41   -5.75   -5.11   -4.08
#> log_lik[12]     -9.33    0.02 1.03  -11.47   -9.97   -9.29   -8.61   -7.44
#> log_lik[13]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_lik[14]     -6.25    0.01 0.88   -8.19   -6.79   -6.17   -5.62   -4.75
#> log_lik[15]     -4.42    0.01 0.41   -5.32   -4.66   -4.39   -4.14   -3.69
#> log_lik[16]     -3.80    0.01 0.55   -5.00   -4.14   -3.77   -3.41   -2.84
#> log_lik[17]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_lik[18]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_lik[19]     -4.65    0.01 0.59   -5.91   -5.01   -4.60   -4.23   -3.62
#> log_lik[20]     -7.09    0.01 0.95   -9.14   -7.67   -7.01   -6.41   -5.42
#> log_lik[21]     -3.03    0.01 0.35   -3.77   -3.27   -3.02   -2.78   -2.37
#> log_lik[22]     -4.65    0.01 0.59   -5.91   -5.01   -4.60   -4.23   -3.62
#> log_lik[23]     -3.71    0.01 0.42   -4.58   -3.98   -3.70   -3.42   -2.94
#> log_lik[24]     -5.79    0.01 0.64   -7.24   -6.16   -5.71   -5.35   -4.76
#> log_lik[25]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_lik[26]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_lik[27]     -5.64    0.01 0.52   -6.77   -5.96   -5.60   -5.28   -4.72
#> log_lik[28]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_lik[29]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_lik[30]     -6.81    0.01 0.85   -8.81   -7.30   -6.71   -6.21   -5.41
#> log_lik[31]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_lik[32]     -6.07    0.01 0.92   -8.02   -6.65   -5.98   -5.44   -4.50
#> log_lik[33]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_lik[34]     -3.03    0.01 0.35   -3.77   -3.27   -3.02   -2.78   -2.37
#> log_lik[35]     -5.53    0.01 0.68   -7.02   -5.95   -5.49   -5.05   -4.35
#> log_lik[36]     -9.94    0.02 1.19  -12.44  -10.71   -9.87   -9.10   -7.87
#> log_lik[37]     -7.61    0.01 0.80   -9.33   -8.10   -7.54   -7.03   -6.22
#> log_lik[38]     -9.00    0.01 0.87  -10.72   -9.56   -8.99   -8.44   -7.32
#> log_lik[39]     -7.29    0.02 1.12   -9.73   -7.98   -7.19   -6.54   -5.31
#> log_lik[40]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_lik[41]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.66
#> log_lik[42]     -6.02    0.01 0.63   -7.43   -6.40   -5.97   -5.58   -4.92
#> log_lik[43]     -5.64    0.01 0.52   -6.77   -5.96   -5.60   -5.28   -4.72
#> log_lik[44]     -4.42    0.01 0.41   -5.32   -4.66   -4.39   -4.14   -3.69
#> log_lik[45]     -3.03    0.01 0.35   -3.77   -3.27   -3.02   -2.78   -2.37
#> lp__          -463.81    0.16 6.24 -476.88 -467.82 -463.42 -459.34 -452.80
#>               n_eff Rhat
#> pi[1]          5258    1
#> pi[2]          5176    1
#> pi[3]          4754    1
#> pi[4]          4525    1
#> theta[1,1,1]   5930    1
#> theta[1,1,2]   5051    1
#> theta[1,1,3]   4306    1
#> theta[1,1,4]   3952    1
#> theta[1,2,1]   4917    1
#> theta[1,2,2]   6218    1
#> theta[1,2,3]   6184    1
#> theta[1,2,4]   4961    1
#> theta[1,3,1]   5046    1
#> theta[1,3,2]   4304    1
#> theta[1,3,3]   4974    1
#> theta[1,3,4]   5519    1
#> theta[1,4,1]   5756    1
#> theta[1,4,2]   5352    1
#> theta[1,4,3]   3375    1
#> theta[1,4,4]   4684    1
#> theta[2,1,1]   5536    1
#> theta[2,1,2]   5499    1
#> theta[2,1,3]   4763    1
#> theta[2,1,4]   4675    1
#> theta[2,2,1]   6100    1
#> theta[2,2,2]   5168    1
#> theta[2,2,3]   5130    1
#> theta[2,2,4]   5435    1
#> theta[2,3,1]   5955    1
#> theta[2,3,2]   5930    1
#> theta[2,3,3]   4928    1
#> theta[2,3,4]   4878    1
#> theta[2,4,1]   5380    1
#> theta[2,4,2]   5861    1
#> theta[2,4,3]   5639    1
#> theta[2,4,4]   5899    1
#> theta[3,1,1]   5434    1
#> theta[3,1,2]   4313    1
#> theta[3,1,3]   4521    1
#> theta[3,1,4]   4247    1
#> theta[3,2,1]   5169    1
#> theta[3,2,2]   5504    1
#> theta[3,2,3]   5309    1
#> theta[3,2,4]   4548    1
#> theta[3,3,1]   5579    1
#> theta[3,3,2]   5494    1
#> theta[3,3,3]   5188    1
#> theta[3,3,4]   5563    1
#> theta[3,4,1]   5788    1
#> theta[3,4,2]   5931    1
#> theta[3,4,3]   4954    1
#> theta[3,4,4]   5836    1
#> theta[4,1,1]   4832    1
#> theta[4,1,2]   4412    1
#> theta[4,1,3]   4400    1
#> theta[4,1,4]   4214    1
#> theta[4,2,1]   5810    1
#> theta[4,2,2]   5260    1
#> theta[4,2,3]   4692    1
#> theta[4,2,4]   4637    1
#> theta[4,3,1]   5396    1
#> theta[4,3,2]   4538    1
#> theta[4,3,3]   4958    1
#> theta[4,3,4]   4131    1
#> theta[4,4,1]   5209    1
#> theta[4,4,2]   5402    1
#> theta[4,4,3]   5255    1
#> theta[4,4,4]   5499    1
#> theta[5,1,1]   5537    1
#> theta[5,1,2]   4508    1
#> theta[5,1,3]   4087    1
#> theta[5,1,4]   4114    1
#> theta[5,2,1]   5245    1
#> theta[5,2,2]   5094    1
#> theta[5,2,3]   5147    1
#> theta[5,2,4]   4956    1
#> theta[5,3,1]   5596    1
#> theta[5,3,2]   5568    1
#> theta[5,3,3]   6046    1
#> theta[5,3,4]   5749    1
#> theta[5,4,1]   6240    1
#> theta[5,4,2]   4811    1
#> theta[5,4,3]   4397    1
#> theta[5,4,4]   5148    1
#> log_p_z[1,1]   5648    1
#> log_p_z[1,2]   3325    1
#> log_p_z[1,3]   2439    1
#> log_p_z[1,4]   2503    1
#> log_p_z[2,1]   3785    1
#> log_p_z[2,2]   3598    1
#> log_p_z[2,3]   3829    1
#> log_p_z[2,4]   2628    1
#> log_p_z[3,1]   2866    1
#> log_p_z[3,2]   3420    1
#> log_p_z[3,3]   2502    1
#> log_p_z[3,4]   2585    1
#> log_p_z[4,1]   4044    1
#> log_p_z[4,2]   3402    1
#> log_p_z[4,3]   2472    1
#> log_p_z[4,4]   3066    1
#> log_p_z[5,1]   3396    1
#> log_p_z[5,2]   5411    1
#> log_p_z[5,3]   2776    1
#> log_p_z[5,4]   3059    1
#> log_p_z[6,1]   3401    1
#> log_p_z[6,2]   4953    1
#> log_p_z[6,3]   2738    1
#> log_p_z[6,4]   3136    1
#> log_p_z[7,1]   4626    1
#> log_p_z[7,2]   3366    1
#> log_p_z[7,3]   2363    1
#> log_p_z[7,4]   2470    1
#> log_p_z[8,1]   3694    1
#> log_p_z[8,2]   3060    1
#> log_p_z[8,3]   5213    1
#> log_p_z[8,4]   2733    1
#> log_p_z[9,1]   3795    1
#> log_p_z[9,2]   4732    1
#> log_p_z[9,3]   3047    1
#> log_p_z[9,4]   2938    1
#> log_p_z[10,1]  3726    1
#> log_p_z[10,2]  4246    1
#> log_p_z[10,3]  3075    1
#> log_p_z[10,4]  2911    1
#> log_p_z[11,1]  3772    1
#> log_p_z[11,2]  3041    1
#> log_p_z[11,3]  2351    1
#> log_p_z[11,4]  5220    1
#> log_p_z[12,1]  4266    1
#> log_p_z[12,2]  3254    1
#> log_p_z[12,3]  2769    1
#> log_p_z[12,4]  3110    1
#> log_p_z[13,1]  5648    1
#> log_p_z[13,2]  3325    1
#> log_p_z[13,3]  2439    1
#> log_p_z[13,4]  2503    1
#> log_p_z[14,1]  3711    1
#> log_p_z[14,2]  3937    1
#> log_p_z[14,3]  2907    1
#> log_p_z[14,4]  3151    1
#> log_p_z[15,1]  4300    1
#> log_p_z[15,2]  3312    1
#> log_p_z[15,3]  2360    1
#> log_p_z[15,4]  2529    1
#> log_p_z[16,1]  5219    1
#> log_p_z[16,2]  3400    1
#> log_p_z[16,3]  2464    1
#> log_p_z[16,4]  2380    1
#> log_p_z[17,1]  5648    1
#> log_p_z[17,2]  3325    1
#> log_p_z[17,3]  2439    1
#> log_p_z[17,4]  2503    1
#> log_p_z[18,1]  5648    1
#> log_p_z[18,2]  3325    1
#> log_p_z[18,3]  2439    1
#> log_p_z[18,4]  2503    1
#> log_p_z[19,1]  3525    1
#> log_p_z[19,2]  4859    1
#> log_p_z[19,3]  2971    1
#> log_p_z[19,4]  3016    1
#> log_p_z[20,1]  3059    1
#> log_p_z[20,2]  4349    1
#> log_p_z[20,3]  2733    1
#> log_p_z[20,4]  3051    1
#> log_p_z[21,1]  3139    1
#> log_p_z[21,2]  4937    1
#> log_p_z[21,3]  2972    1
#> log_p_z[21,4]  2987    1
#> log_p_z[22,1]  3525    1
#> log_p_z[22,2]  4859    1
#> log_p_z[22,3]  2971    1
#> log_p_z[22,4]  3016    1
#> log_p_z[23,1]  3396    1
#> log_p_z[23,2]  5411    1
#> log_p_z[23,3]  2776    1
#> log_p_z[23,4]  3059    1
#> log_p_z[24,1]  3096    1
#> log_p_z[24,2]  3712    1
#> log_p_z[24,3]  2824    1
#> log_p_z[24,4]  2975    1
#> log_p_z[25,1]  5648    1
#> log_p_z[25,2]  3325    1
#> log_p_z[25,3]  2439    1
#> log_p_z[25,4]  2503    1
#> log_p_z[26,1]  5648    1
#> log_p_z[26,2]  3325    1
#> log_p_z[26,3]  2439    1
#> log_p_z[26,4]  2503    1
#> log_p_z[27,1]  3290    1
#> log_p_z[27,2]  5642    1
#> log_p_z[27,3]  3061    1
#> log_p_z[27,4]  3058    1
#> log_p_z[28,1]  5648    1
#> log_p_z[28,2]  3325    1
#> log_p_z[28,3]  2439    1
#> log_p_z[28,4]  2503    1
#> log_p_z[29,1]  5648    1
#> log_p_z[29,2]  3325    1
#> log_p_z[29,3]  2439    1
#> log_p_z[29,4]  2503    1
#> log_p_z[30,1]  3625    1
#> log_p_z[30,2]  3217    1
#> log_p_z[30,3]  2389    1
#> log_p_z[30,4]  2640    1
#> log_p_z[31,1]  5648    1
#> log_p_z[31,2]  3325    1
#> log_p_z[31,3]  2439    1
#> log_p_z[31,4]  2503    1
#> log_p_z[32,1]  3841    1
#> log_p_z[32,2]  4319    1
#> log_p_z[32,3]  4662    1
#> log_p_z[32,4]  2687    1
#> log_p_z[33,1]  5648    1
#> log_p_z[33,2]  3325    1
#> log_p_z[33,3]  2439    1
#> log_p_z[33,4]  2503    1
#> log_p_z[34,1]  3139    1
#> log_p_z[34,2]  4937    1
#> log_p_z[34,3]  2972    1
#> log_p_z[34,4]  2987    1
#> log_p_z[35,1]  3662    1
#> log_p_z[35,2]  4243    1
#> log_p_z[35,3]  2917    1
#> log_p_z[35,4]  3076    1
#> log_p_z[36,1]  3486    1
#> log_p_z[36,2]  3185    1
#> log_p_z[36,3]  2124    1
#> log_p_z[36,4]  2312    1
#> log_p_z[37,1]  3412    1
#> log_p_z[37,2]  4326    1
#> log_p_z[37,3]  2803    1
#> log_p_z[37,4]  2917    1
#> log_p_z[38,1]  3876    1
#> log_p_z[38,2]  3764    1
#> log_p_z[38,3]  3032    1
#> log_p_z[38,4]  3051    1
#> log_p_z[39,1]  3834    1
#> log_p_z[39,2]  3225    1
#> log_p_z[39,3]  4734    1
#> log_p_z[39,4]  3016    1
#> log_p_z[40,1]  5648    1
#> log_p_z[40,2]  3325    1
#> log_p_z[40,3]  2439    1
#> log_p_z[40,4]  2503    1
#> log_p_z[41,1]  5648    1
#> log_p_z[41,2]  3325    1
#> log_p_z[41,3]  2439    1
#> log_p_z[41,4]  2503    1
#> log_p_z[42,1]  4626    1
#> log_p_z[42,2]  3366    1
#> log_p_z[42,3]  2363    1
#> log_p_z[42,4]  2470    1
#> log_p_z[43,1]  3290    1
#> log_p_z[43,2]  5642    1
#> log_p_z[43,3]  3061    1
#> log_p_z[43,4]  3058    1
#> log_p_z[44,1]  4300    1
#> log_p_z[44,2]  3312    1
#> log_p_z[44,3]  2360    1
#> log_p_z[44,4]  2529    1
#> log_p_z[45,1]  3139    1
#> log_p_z[45,2]  4937    1
#> log_p_z[45,3]  2972    1
#> log_p_z[45,4]  2987    1
#> log_lik[1]     5648    1
#> log_lik[2]     4160    1
#> log_lik[3]     4614    1
#> log_lik[4]     3466    1
#> log_lik[5]     5411    1
#> log_lik[6]     4959    1
#> log_lik[7]     4627    1
#> log_lik[8]     5215    1
#> log_lik[9]     4733    1
#> log_lik[10]    4192    1
#> log_lik[11]    5221    1
#> log_lik[12]    4341    1
#> log_lik[13]    5648    1
#> log_lik[14]    3937    1
#> log_lik[15]    4300    1
#> log_lik[16]    5219    1
#> log_lik[17]    5648    1
#> log_lik[18]    5648    1
#> log_lik[19]    4859    1
#> log_lik[20]    4350    1
#> log_lik[21]    4937    1
#> log_lik[22]    4859    1
#> log_lik[23]    5411    1
#> log_lik[24]    3713    1
#> log_lik[25]    5648    1
#> log_lik[26]    5648    1
#> log_lik[27]    5642    1
#> log_lik[28]    5648    1
#> log_lik[29]    5648    1
#> log_lik[30]    3651    1
#> log_lik[31]    5648    1
#> log_lik[32]    4667    1
#> log_lik[33]    5648    1
#> log_lik[34]    4937    1
#> log_lik[35]    4257    1
#> log_lik[36]    4219    1
#> log_lik[37]    4331    1
#> log_lik[38]    4322    1
#> log_lik[39]    4735    1
#> log_lik[40]    5648    1
#> log_lik[41]    5648    1
#> log_lik[42]    4627    1
#> log_lik[43]    5642    1
#> log_lik[44]    4300    1
#> log_lik[45]    4937    1
#> lp__           1481    1
#> 
#> Samples were drawn using NUTS(diag_e) at Tue Feb 14 08:08:40 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }