Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.37    0.00 0.06    0.25    0.33    0.37    0.42    0.51
#> pi[2]            0.41    0.00 0.06    0.29    0.36    0.40    0.45    0.53
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.24
#> pi[4]            0.08    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.06    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.05
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.07    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.29
#> theta[1,3,3]     0.79    0.00 0.10    0.57    0.73    0.81    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.23
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.29
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.41
#> theta[1,4,4]     0.69    0.00 0.14    0.40    0.59    0.70    0.79    0.91
#> theta[2,1,1]     0.75    0.00 0.09    0.56    0.69    0.76    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.08    0.05    0.11    0.16    0.22    0.35
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[2,2,2]     0.58    0.00 0.09    0.39    0.52    0.58    0.65    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.24    0.30    0.37    0.50
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.26
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.26
#> theta[2,3,3]     0.63    0.00 0.13    0.35    0.54    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.12    0.04    0.13    0.20    0.28    0.48
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.40
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.76    0.90
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.02    0.05    0.09    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.56    0.68    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.08    0.00    0.03    0.06    0.11    0.29
#> theta[3,3,2]     0.18    0.00 0.10    0.03    0.10    0.16    0.24    0.42
#> theta[3,3,3]     0.51    0.00 0.14    0.23    0.41    0.51    0.61    0.77
#> theta[3,3,4]     0.23    0.00 0.12    0.06    0.15    0.22    0.30    0.51
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[3,4,3]     0.14    0.00 0.12    0.01    0.05    0.11    0.21    0.43
#> theta[3,4,4]     0.63    0.00 0.15    0.31    0.52    0.63    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.78    0.83    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[4,2,2]     0.72    0.00 0.08    0.55    0.67    0.73    0.78    0.87
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.09    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.10    0.18
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[4,3,3]     0.69    0.00 0.12    0.44    0.61    0.70    0.78    0.91
#> theta[4,3,4]     0.15    0.00 0.09    0.01    0.08    0.13    0.20    0.36
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.91
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.82    0.87    0.91    0.96
#> theta[5,1,2]     0.06    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.60    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.31
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.22    0.40
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.51    0.61    0.70    0.85
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.13    0.21    0.39
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[5,4,3]     0.14    0.00 0.11    0.00    0.05    0.11    0.20    0.42
#> theta[5,4,4]     0.63    0.00 0.16    0.30    0.52    0.64    0.74    0.90
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_p_z[1,2]   -19.99    0.03 2.34  -25.12  -21.40  -19.75  -18.36  -16.14
#> log_p_z[1,3]   -24.85    0.08 4.43  -34.94  -27.31  -24.35  -21.71  -17.57
#> log_p_z[1,4]   -21.84    0.07 4.19  -31.01  -24.30  -21.37  -18.93  -15.07
#> log_p_z[2,1]   -29.70    0.06 4.35  -39.46  -32.32  -29.14  -26.65  -22.54
#> log_p_z[2,2]   -21.06    0.03 2.29  -26.01  -22.49  -20.88  -19.45  -17.07
#> log_p_z[2,3]    -7.67    0.02 1.26  -10.48   -8.38   -7.56   -6.79   -5.58
#> log_p_z[2,4]   -16.22    0.07 4.10  -25.35  -18.75  -15.72  -13.25   -9.64
#> log_p_z[3,1]   -11.60    0.03 1.65  -15.26  -12.57  -11.43  -10.43   -8.87
#> log_p_z[3,2]   -10.79    0.02 1.50  -14.13  -11.74  -10.61   -9.70   -8.35
#> log_p_z[3,3]   -23.35    0.06 3.48  -30.89  -25.44  -23.05  -20.88  -17.37
#> log_p_z[3,4]   -21.81    0.06 3.47  -29.45  -23.95  -21.52  -19.31  -15.96
#> log_p_z[4,1]   -14.61    0.03 1.86  -18.59  -15.81  -14.51  -13.26  -11.34
#> log_p_z[4,2]    -7.55    0.01 0.93   -9.56   -8.12   -7.48   -6.93   -5.93
#> log_p_z[4,3]   -20.77    0.08 4.39  -30.93  -23.40  -20.19  -17.60  -13.91
#> log_p_z[4,4]   -21.87    0.07 4.33  -31.63  -24.35  -21.28  -18.87  -14.89
#> log_p_z[5,1]   -21.42    0.04 2.53  -26.93  -23.01  -21.22  -19.62  -16.99
#> log_p_z[5,2]    -3.71    0.00 0.44   -4.61   -3.99   -3.69   -3.41   -2.90
#> log_p_z[5,3]   -18.73    0.07 4.29  -29.06  -21.23  -18.07  -15.68  -12.14
#> log_p_z[5,4]   -21.85    0.07 4.37  -31.59  -24.38  -21.33  -18.71  -14.80
#> log_p_z[6,1]   -21.44    0.04 2.53  -26.77  -23.06  -21.24  -19.66  -17.06
#> log_p_z[6,2]    -5.47    0.01 0.67   -6.91   -5.90   -5.41   -5.01   -4.29
#> log_p_z[6,3]   -17.53    0.07 4.14  -27.39  -19.86  -16.89  -14.59  -11.24
#> log_p_z[6,4]   -21.62    0.07 4.30  -31.17  -24.04  -21.06  -18.59  -14.77
#> log_p_z[7,1]    -6.01    0.01 0.65   -7.46   -6.39   -5.97   -5.56   -4.87
#> log_p_z[7,2]   -14.93    0.03 1.74  -18.81  -16.02  -14.77  -13.72  -11.94
#> log_p_z[7,3]   -24.31    0.06 3.64  -32.18  -26.57  -23.95  -21.76  -18.01
#> log_p_z[7,4]   -21.84    0.06 3.46  -29.57  -23.96  -21.55  -19.37  -16.03
#> log_p_z[8,1]   -29.68    0.06 4.38  -39.71  -32.29  -29.17  -26.59  -22.44
#> log_p_z[8,2]   -18.28    0.03 2.07  -22.78  -19.58  -18.13  -16.81  -14.68
#> log_p_z[8,3]    -5.72    0.01 0.84   -7.53   -6.25   -5.67   -5.14   -4.24
#> log_p_z[8,4]   -18.45    0.07 4.21  -27.67  -21.05  -17.88  -15.41  -11.61
#> log_p_z[9,1]   -19.95    0.03 2.27  -24.80  -21.40  -19.74  -18.34  -15.97
#> log_p_z[9,2]    -4.53    0.01 0.57   -5.77   -4.87   -4.49   -4.12   -3.53
#> log_p_z[9,3]   -19.68    0.08 4.35  -29.95  -22.13  -19.16  -16.59  -12.90
#> log_p_z[9,4]   -21.66    0.07 4.34  -31.60  -24.17  -21.18  -18.59  -14.69
#> log_p_z[10,1]  -22.20    0.04 2.40  -27.14  -23.77  -22.02  -20.47  -18.04
#> log_p_z[10,2]   -7.14    0.01 0.71   -8.70   -7.56   -7.07   -6.64   -5.93
#> log_p_z[10,3]  -16.84    0.06 3.16  -24.28  -18.63  -16.51  -14.62  -11.74
#> log_p_z[10,4]  -21.29    0.06 3.55  -29.14  -23.46  -21.01  -18.74  -15.31
#> log_p_z[11,1]  -29.74    0.06 4.34  -39.10  -32.37  -29.38  -26.61  -22.56
#> log_p_z[11,2]  -29.18    0.07 4.21  -38.52  -31.61  -28.72  -26.27  -22.19
#> log_p_z[11,3]  -18.18    0.06 3.28  -26.10  -19.84  -17.73  -15.96  -13.06
#> log_p_z[11,4]   -5.82    0.01 0.96   -7.88   -6.42   -5.75   -5.13   -4.15
#> log_p_z[12,1]  -22.94    0.04 2.59  -28.64  -24.56  -22.73  -21.14  -18.45
#> log_p_z[12,2]   -9.60    0.02 1.30  -12.58  -10.31   -9.47   -8.72   -7.48
#> log_p_z[12,3]  -15.16    0.07 4.14  -24.76  -17.53  -14.59  -12.16   -8.89
#> log_p_z[12,4]  -19.28    0.07 4.15  -28.69  -21.68  -18.73  -16.37  -12.65
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_p_z[13,2]  -19.99    0.03 2.34  -25.12  -21.40  -19.75  -18.36  -16.14
#> log_p_z[13,3]  -24.85    0.08 4.43  -34.94  -27.31  -24.35  -21.71  -17.57
#> log_p_z[13,4]  -21.84    0.07 4.19  -31.01  -24.30  -21.37  -18.93  -15.07
#> log_p_z[14,1]  -19.04    0.04 2.39  -24.39  -20.50  -18.82  -17.34  -15.00
#> log_p_z[14,2]   -6.24    0.01 0.90   -8.24   -6.79   -6.16   -5.62   -4.74
#> log_p_z[14,3]  -18.74    0.07 4.28  -28.52  -21.32  -18.13  -15.61  -12.30
#> log_p_z[14,4]  -21.87    0.07 4.23  -31.34  -24.38  -21.42  -18.81  -14.97
#> log_p_z[15,1]   -4.41    0.00 0.42   -5.31   -4.67   -4.38   -4.12   -3.68
#> log_p_z[15,2]  -17.23    0.03 1.88  -21.41  -18.39  -17.06  -15.92  -14.03
#> log_p_z[15,3]  -24.34    0.06 3.65  -32.35  -26.59  -23.97  -21.74  -18.14
#> log_p_z[15,4]  -21.86    0.06 3.48  -29.57  -23.93  -21.62  -19.42  -16.01
#> log_p_z[16,1]   -3.81    0.01 0.57   -5.08   -4.14   -3.75   -3.43   -2.83
#> log_p_z[16,2]  -17.69    0.03 2.24  -22.68  -19.12  -17.44  -16.12  -13.90
#> log_p_z[16,3]  -24.82    0.08 4.41  -35.03  -27.35  -24.34  -21.74  -17.36
#> log_p_z[16,4]  -21.81    0.07 4.18  -31.08  -24.25  -21.30  -18.86  -15.12
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_p_z[17,2]  -19.99    0.03 2.34  -25.12  -21.40  -19.75  -18.36  -16.14
#> log_p_z[17,3]  -24.85    0.08 4.43  -34.94  -27.31  -24.35  -21.71  -17.57
#> log_p_z[17,4]  -21.84    0.07 4.19  -31.01  -24.30  -21.37  -18.93  -15.07
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_p_z[18,2]  -19.99    0.03 2.34  -25.12  -21.40  -19.75  -18.36  -16.14
#> log_p_z[18,3]  -24.85    0.08 4.43  -34.94  -27.31  -24.35  -21.71  -17.57
#> log_p_z[18,4]  -21.84    0.07 4.19  -31.01  -24.30  -21.37  -18.93  -15.07
#> log_p_z[19,1]  -16.38    0.03 1.92  -20.51  -17.58  -16.27  -15.06  -12.97
#> log_p_z[19,2]   -4.66    0.01 0.60   -5.97   -5.01   -4.62   -4.24   -3.62
#> log_p_z[19,3]  -22.16    0.08 4.48  -32.56  -24.69  -21.67  -18.92  -15.10
#> log_p_z[19,4]  -21.84    0.07 4.32  -31.71  -24.31  -21.29  -18.78  -14.98
#> log_p_z[20,1]  -18.03    0.04 2.26  -22.85  -19.44  -17.85  -16.45  -14.11
#> log_p_z[20,2]   -7.11    0.01 0.98   -9.30   -7.70   -7.03   -6.42   -5.40
#> log_p_z[20,3]  -20.05    0.07 4.23  -29.91  -22.38  -19.47  -17.10  -13.50
#> log_p_z[20,4]  -21.64    0.07 4.35  -31.52  -24.08  -21.08  -18.64  -14.62
#> log_p_z[21,1]  -19.60    0.04 2.35  -24.75  -21.02  -19.46  -17.94  -15.43
#> log_p_z[21,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.03   -2.80   -2.35
#> log_p_z[21,3]  -21.22    0.08 4.36  -31.57  -23.72  -20.70  -18.15  -14.53
#> log_p_z[21,4]  -21.84    0.07 4.35  -32.01  -24.25  -21.45  -18.71  -14.83
#> log_p_z[22,1]  -16.38    0.03 1.92  -20.51  -17.58  -16.27  -15.06  -12.97
#> log_p_z[22,2]   -4.66    0.01 0.60   -5.97   -5.01   -4.62   -4.24   -3.62
#> log_p_z[22,3]  -22.16    0.08 4.48  -32.56  -24.69  -21.67  -18.92  -15.10
#> log_p_z[22,4]  -21.84    0.07 4.32  -31.71  -24.31  -21.29  -18.78  -14.98
#> log_p_z[23,1]  -21.42    0.04 2.53  -26.93  -23.01  -21.22  -19.62  -16.99
#> log_p_z[23,2]   -3.71    0.00 0.44   -4.61   -3.99   -3.69   -3.41   -2.90
#> log_p_z[23,3]  -18.73    0.07 4.29  -29.06  -21.23  -18.07  -15.68  -12.14
#> log_p_z[23,4]  -21.85    0.07 4.37  -31.59  -24.38  -21.33  -18.71  -14.80
#> log_p_z[24,1]  -17.40    0.03 2.15  -22.20  -18.72  -17.25  -15.90  -13.67
#> log_p_z[24,2]   -5.80    0.01 0.64   -7.24   -6.17   -5.74   -5.35   -4.76
#> log_p_z[24,3]  -21.73    0.06 3.43  -29.65  -23.69  -21.39  -19.39  -15.98
#> log_p_z[24,4]  -21.81    0.06 3.57  -29.70  -23.94  -21.59  -19.23  -15.75
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_p_z[25,2]  -19.99    0.03 2.34  -25.12  -21.40  -19.75  -18.36  -16.14
#> log_p_z[25,3]  -24.85    0.08 4.43  -34.94  -27.31  -24.35  -21.71  -17.57
#> log_p_z[25,4]  -21.84    0.07 4.19  -31.01  -24.30  -21.37  -18.93  -15.07
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_p_z[26,2]  -19.99    0.03 2.34  -25.12  -21.40  -19.75  -18.36  -16.14
#> log_p_z[26,3]  -24.85    0.08 4.43  -34.94  -27.31  -24.35  -21.71  -17.57
#> log_p_z[26,4]  -21.84    0.07 4.19  -31.01  -24.30  -21.37  -18.93  -15.07
#> log_p_z[27,1]  -21.85    0.04 2.44  -27.06  -23.35  -21.66  -20.14  -17.52
#> log_p_z[27,2]   -5.66    0.01 0.51   -6.72   -5.99   -5.64   -5.29   -4.76
#> log_p_z[27,3]  -18.38    0.06 3.20  -25.89  -20.21  -18.03  -16.17  -13.14
#> log_p_z[27,4]  -21.47    0.06 3.50  -29.26  -23.55  -21.19  -18.99  -15.70
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_p_z[28,2]  -19.99    0.03 2.34  -25.12  -21.40  -19.75  -18.36  -16.14
#> log_p_z[28,3]  -24.85    0.08 4.43  -34.94  -27.31  -24.35  -21.71  -17.57
#> log_p_z[28,4]  -21.84    0.07 4.19  -31.01  -24.30  -21.37  -18.93  -15.07
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_p_z[29,2]  -19.99    0.03 2.34  -25.12  -21.40  -19.75  -18.36  -16.14
#> log_p_z[29,3]  -24.85    0.08 4.43  -34.94  -27.31  -24.35  -21.71  -17.57
#> log_p_z[29,4]  -21.84    0.07 4.19  -31.01  -24.30  -21.37  -18.93  -15.07
#> log_p_z[30,1]   -6.80    0.01 0.83   -8.69   -7.29   -6.73   -6.22   -5.38
#> log_p_z[30,2]  -14.70    0.02 1.71  -18.43  -15.74  -14.53  -13.49  -11.80
#> log_p_z[30,3]  -24.32    0.06 3.60  -32.18  -26.47  -24.04  -21.81  -18.19
#> log_p_z[30,4]  -21.83    0.06 3.51  -29.43  -23.89  -21.62  -19.37  -15.85
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_p_z[31,2]  -19.99    0.03 2.34  -25.12  -21.40  -19.75  -18.36  -16.14
#> log_p_z[31,3]  -24.85    0.08 4.43  -34.94  -27.31  -24.35  -21.71  -17.57
#> log_p_z[31,4]  -21.84    0.07 4.19  -31.01  -24.30  -21.37  -18.93  -15.07
#> log_p_z[32,1]  -29.66    0.06 4.40  -39.53  -32.31  -29.17  -26.54  -22.27
#> log_p_z[32,2]  -14.85    0.02 1.69  -18.47  -15.92  -14.73  -13.63  -11.88
#> log_p_z[32,3]   -6.04    0.01 0.92   -8.06   -6.62   -5.97   -5.40   -4.45
#> log_p_z[32,4]  -20.56    0.07 4.34  -30.08  -23.23  -20.00  -17.52  -13.46
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_p_z[33,2]  -19.99    0.03 2.34  -25.12  -21.40  -19.75  -18.36  -16.14
#> log_p_z[33,3]  -24.85    0.08 4.43  -34.94  -27.31  -24.35  -21.71  -17.57
#> log_p_z[33,4]  -21.84    0.07 4.19  -31.01  -24.30  -21.37  -18.93  -15.07
#> log_p_z[34,1]  -19.60    0.04 2.35  -24.75  -21.02  -19.46  -17.94  -15.43
#> log_p_z[34,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.03   -2.80   -2.35
#> log_p_z[34,3]  -21.22    0.08 4.36  -31.57  -23.72  -20.70  -18.15  -14.53
#> log_p_z[34,4]  -21.84    0.07 4.35  -32.01  -24.25  -21.45  -18.71  -14.83
#> log_p_z[35,1]  -22.56    0.04 2.66  -28.28  -24.23  -22.37  -20.65  -17.94
#> log_p_z[35,2]   -5.52    0.01 0.69   -7.04   -5.94   -5.48   -5.05   -4.31
#> log_p_z[35,3]  -16.10    0.07 4.18  -26.02  -18.55  -15.49  -13.06   -9.74
#> log_p_z[35,4]  -21.84    0.07 4.31  -31.28  -24.35  -21.38  -18.79  -14.82
#> log_p_z[36,1]  -29.73    0.06 3.68  -37.98  -32.03  -29.31  -27.10  -23.60
#> log_p_z[36,2]  -21.87    0.04 2.28  -26.83  -23.29  -21.72  -20.27  -17.83
#> log_p_z[36,3]  -10.49    0.03 1.65  -14.46  -11.37  -10.26   -9.35   -7.96
#> log_p_z[36,4]  -13.89    0.06 3.19  -20.98  -15.90  -13.61  -11.53   -8.73
#> log_p_z[37,1]  -18.54    0.04 2.28  -23.39  -19.94  -18.39  -16.97  -14.59
#> log_p_z[37,2]   -7.62    0.01 0.82   -9.54   -8.09   -7.54   -7.06   -6.23
#> log_p_z[37,3]  -19.11    0.06 3.31  -26.86  -21.08  -18.78  -16.75  -13.69
#> log_p_z[37,4]  -21.81    0.06 3.52  -29.57  -23.97  -21.53  -19.31  -15.81
#> log_p_z[38,1]  -25.16    0.04 2.83  -31.17  -26.95  -24.95  -23.17  -20.12
#> log_p_z[38,2]   -9.62    0.01 0.98  -11.72  -10.23   -9.57   -8.91   -7.92
#> log_p_z[38,3]  -11.72    0.05 2.85  -18.37  -13.39  -11.30   -9.60   -7.46
#> log_p_z[38,4]  -21.29    0.06 3.51  -29.04  -23.41  -21.03  -18.80  -15.29
#> log_p_z[39,1]  -29.33    0.06 4.30  -39.11  -31.90  -28.86  -26.35  -22.18
#> log_p_z[39,2]  -16.80    0.03 1.97  -21.10  -18.03  -16.64  -15.41  -13.41
#> log_p_z[39,3]   -7.26    0.01 1.12   -9.68   -7.94   -7.19   -6.46   -5.33
#> log_p_z[39,4]  -18.63    0.07 4.07  -27.76  -20.99  -18.06  -15.76  -12.19
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_p_z[40,2]  -19.99    0.03 2.34  -25.12  -21.40  -19.75  -18.36  -16.14
#> log_p_z[40,3]  -24.85    0.08 4.43  -34.94  -27.31  -24.35  -21.71  -17.57
#> log_p_z[40,4]  -21.84    0.07 4.19  -31.01  -24.30  -21.37  -18.93  -15.07
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_p_z[41,2]  -19.99    0.03 2.34  -25.12  -21.40  -19.75  -18.36  -16.14
#> log_p_z[41,3]  -24.85    0.08 4.43  -34.94  -27.31  -24.35  -21.71  -17.57
#> log_p_z[41,4]  -21.84    0.07 4.19  -31.01  -24.30  -21.37  -18.93  -15.07
#> log_p_z[42,1]   -6.01    0.01 0.65   -7.46   -6.39   -5.97   -5.56   -4.87
#> log_p_z[42,2]  -14.93    0.03 1.74  -18.81  -16.02  -14.77  -13.72  -11.94
#> log_p_z[42,3]  -24.31    0.06 3.64  -32.18  -26.57  -23.95  -21.76  -18.01
#> log_p_z[42,4]  -21.84    0.06 3.46  -29.57  -23.96  -21.55  -19.37  -16.03
#> log_p_z[43,1]  -21.85    0.04 2.44  -27.06  -23.35  -21.66  -20.14  -17.52
#> log_p_z[43,2]   -5.66    0.01 0.51   -6.72   -5.99   -5.64   -5.29   -4.76
#> log_p_z[43,3]  -18.38    0.06 3.20  -25.89  -20.21  -18.03  -16.17  -13.14
#> log_p_z[43,4]  -21.47    0.06 3.50  -29.26  -23.55  -21.19  -18.99  -15.70
#> log_p_z[44,1]   -4.41    0.00 0.42   -5.31   -4.67   -4.38   -4.12   -3.68
#> log_p_z[44,2]  -17.23    0.03 1.88  -21.41  -18.39  -17.06  -15.92  -14.03
#> log_p_z[44,3]  -24.34    0.06 3.65  -32.35  -26.59  -23.97  -21.74  -18.14
#> log_p_z[44,4]  -21.86    0.06 3.48  -29.57  -23.93  -21.62  -19.42  -16.01
#> log_p_z[45,1]  -19.60    0.04 2.35  -24.75  -21.02  -19.46  -17.94  -15.43
#> log_p_z[45,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.03   -2.80   -2.35
#> log_p_z[45,3]  -21.22    0.08 4.36  -31.57  -23.72  -20.70  -18.15  -14.53
#> log_p_z[45,4]  -21.84    0.07 4.35  -32.01  -24.25  -21.45  -18.71  -14.83
#> log_lik[1]      -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_lik[2]      -7.61    0.02 1.16  -10.12   -8.34   -7.54   -6.79   -5.58
#> log_lik[3]      -9.92    0.01 0.94  -11.86  -10.56   -9.88   -9.27   -8.20
#> log_lik[4]      -7.54    0.01 0.91   -9.49   -8.11   -7.48   -6.93   -5.93
#> log_lik[5]      -3.71    0.00 0.44   -4.61   -3.99   -3.69   -3.41   -2.90
#> log_lik[6]      -5.47    0.01 0.67   -6.91   -5.90   -5.41   -5.01   -4.29
#> log_lik[7]      -6.01    0.01 0.65   -7.46   -6.39   -5.97   -5.56   -4.87
#> log_lik[8]      -5.72    0.01 0.83   -7.53   -6.25   -5.67   -5.13   -4.24
#> log_lik[9]      -4.53    0.01 0.57   -5.77   -4.87   -4.49   -4.12   -3.53
#> log_lik[10]     -7.14    0.01 0.71   -8.69   -7.56   -7.07   -6.64   -5.93
#> log_lik[11]     -5.82    0.01 0.96   -7.88   -6.42   -5.75   -5.13   -4.15
#> log_lik[12]     -9.33    0.01 1.04  -11.50   -9.99   -9.30   -8.61   -7.45
#> log_lik[13]     -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_lik[14]     -6.24    0.01 0.90   -8.24   -6.79   -6.16   -5.62   -4.74
#> log_lik[15]     -4.41    0.00 0.42   -5.31   -4.67   -4.38   -4.12   -3.68
#> log_lik[16]     -3.81    0.01 0.57   -5.08   -4.14   -3.75   -3.43   -2.83
#> log_lik[17]     -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_lik[18]     -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_lik[19]     -4.66    0.01 0.60   -5.97   -5.01   -4.62   -4.24   -3.62
#> log_lik[20]     -7.10    0.01 0.98   -9.30   -7.70   -7.03   -6.42   -5.40
#> log_lik[21]     -3.04    0.00 0.36   -3.78   -3.27   -3.03   -2.80   -2.35
#> log_lik[22]     -4.66    0.01 0.60   -5.97   -5.01   -4.62   -4.24   -3.62
#> log_lik[23]     -3.71    0.00 0.44   -4.61   -3.99   -3.69   -3.41   -2.90
#> log_lik[24]     -5.80    0.01 0.64   -7.24   -6.17   -5.74   -5.35   -4.76
#> log_lik[25]     -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_lik[26]     -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_lik[27]     -5.66    0.01 0.51   -6.72   -5.99   -5.64   -5.29   -4.76
#> log_lik[28]     -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_lik[29]     -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_lik[30]     -6.79    0.01 0.83   -8.69   -7.29   -6.72   -6.22   -5.38
#> log_lik[31]     -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_lik[32]     -6.04    0.01 0.92   -8.06   -6.62   -5.97   -5.40   -4.45
#> log_lik[33]     -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_lik[34]     -3.04    0.00 0.36   -3.78   -3.27   -3.03   -2.80   -2.35
#> log_lik[35]     -5.52    0.01 0.69   -7.02   -5.94   -5.48   -5.05   -4.31
#> log_lik[36]     -9.91    0.02 1.16  -12.35  -10.66   -9.84   -9.09   -7.85
#> log_lik[37]     -7.61    0.01 0.82   -9.54   -8.09   -7.54   -7.06   -6.23
#> log_lik[38]     -9.01    0.01 0.86  -10.72   -9.58   -9.00   -8.43   -7.30
#> log_lik[39]     -7.26    0.01 1.11   -9.67   -7.94   -7.19   -6.46   -5.33
#> log_lik[40]     -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_lik[41]     -2.21    0.00 0.30   -2.84   -2.41   -2.20   -2.00   -1.68
#> log_lik[42]     -6.01    0.01 0.65   -7.46   -6.39   -5.97   -5.56   -4.87
#> log_lik[43]     -5.66    0.01 0.51   -6.72   -5.99   -5.64   -5.29   -4.76
#> log_lik[44]     -4.41    0.00 0.42   -5.31   -4.67   -4.38   -4.12   -3.68
#> log_lik[45]     -3.04    0.00 0.36   -3.78   -3.27   -3.03   -2.80   -2.35
#> lp__          -463.71    0.17 6.12 -476.67 -467.56 -463.46 -459.33 -452.93
#>               n_eff Rhat
#> pi[1]          8178    1
#> pi[2]          7547    1
#> pi[3]          7042    1
#> pi[4]          5823    1
#> theta[1,1,1]   8011    1
#> theta[1,1,2]   6310    1
#> theta[1,1,3]   4881    1
#> theta[1,1,4]   4804    1
#> theta[1,2,1]   6458    1
#> theta[1,2,2]   7949    1
#> theta[1,2,3]   7456    1
#> theta[1,2,4]   6058    1
#> theta[1,3,1]   6692    1
#> theta[1,3,2]   4930    1
#> theta[1,3,3]   6131    1
#> theta[1,3,4]   5747    1
#> theta[1,4,1]   7363    1
#> theta[1,4,2]   7231    1
#> theta[1,4,3]   4185    1
#> theta[1,4,4]   5897    1
#> theta[2,1,1]   7749    1
#> theta[2,1,2]   6190    1
#> theta[2,1,3]   5045    1
#> theta[2,1,4]   3651    1
#> theta[2,2,1]   8441    1
#> theta[2,2,2]   9095    1
#> theta[2,2,3]   8217    1
#> theta[2,2,4]   5957    1
#> theta[2,3,1]   6939    1
#> theta[2,3,2]   7157    1
#> theta[2,3,3]   7067    1
#> theta[2,3,4]   6168    1
#> theta[2,4,1]   7453    1
#> theta[2,4,2]   6447    1
#> theta[2,4,3]   7265    1
#> theta[2,4,4]   7225    1
#> theta[3,1,1]   7777    1
#> theta[3,1,2]   5114    1
#> theta[3,1,3]   5421    1
#> theta[3,1,4]   5099    1
#> theta[3,2,1]   6979    1
#> theta[3,2,2]   7018    1
#> theta[3,2,3]   6647    1
#> theta[3,2,4]   5098    1
#> theta[3,3,1]   5757    1
#> theta[3,3,2]   7105    1
#> theta[3,3,3]   8819    1
#> theta[3,3,4]   8026    1
#> theta[3,4,1]   7988    1
#> theta[3,4,2]   7612    1
#> theta[3,4,3]   5258    1
#> theta[3,4,4]   6635    1
#> theta[4,1,1]   8661    1
#> theta[4,1,2]   5790    1
#> theta[4,1,3]   4135    1
#> theta[4,1,4]   5420    1
#> theta[4,2,1]   7849    1
#> theta[4,2,2]   7593    1
#> theta[4,2,3]   7185    1
#> theta[4,2,4]   6104    1
#> theta[4,3,1]   7699    1
#> theta[4,3,2]   5506    1
#> theta[4,3,3]   6098    1
#> theta[4,3,4]   5374    1
#> theta[4,4,1]   7011    1
#> theta[4,4,2]   7637    1
#> theta[4,4,3]   6410    1
#> theta[4,4,4]   8221    1
#> theta[5,1,1]   7374    1
#> theta[5,1,2]   4977    1
#> theta[5,1,3]   5686    1
#> theta[5,1,4]   5304    1
#> theta[5,2,1]   9295    1
#> theta[5,2,2]   9062    1
#> theta[5,2,3]   7415    1
#> theta[5,2,4]   5659    1
#> theta[5,3,1]   6452    1
#> theta[5,3,2]   7317    1
#> theta[5,3,3]   7061    1
#> theta[5,3,4]   8206    1
#> theta[5,4,1]   6944    1
#> theta[5,4,2]   7092    1
#> theta[5,4,3]   5726    1
#> theta[5,4,4]   6344    1
#> log_p_z[1,1]   7236    1
#> log_p_z[1,2]   4704    1
#> log_p_z[1,3]   3271    1
#> log_p_z[1,4]   3385    1
#> log_p_z[2,1]   5106    1
#> log_p_z[2,2]   4540    1
#> log_p_z[2,3]   3974    1
#> log_p_z[2,4]   3223    1
#> log_p_z[3,1]   3754    1
#> log_p_z[3,2]   4365    1
#> log_p_z[3,3]   3568    1
#> log_p_z[3,4]   3659    1
#> log_p_z[4,1]   5087    1
#> log_p_z[4,2]   5849    1
#> log_p_z[4,3]   3395    1
#> log_p_z[4,4]   3870    1
#> log_p_z[5,1]   4079    1
#> log_p_z[5,2]   8532    1
#> log_p_z[5,3]   3367    1
#> log_p_z[5,4]   3987    1
#> log_p_z[6,1]   3581    1
#> log_p_z[6,2]   5991    1
#> log_p_z[6,3]   3346    1
#> log_p_z[6,4]   3888    1
#> log_p_z[7,1]   5793    1
#> log_p_z[7,2]   4539    1
#> log_p_z[7,3]   3468    1
#> log_p_z[7,4]   3442    1
#> log_p_z[8,1]   5153    1
#> log_p_z[8,2]   5392    1
#> log_p_z[8,3]   6292    1
#> log_p_z[8,4]   3323    1
#> log_p_z[9,1]   4519    1
#> log_p_z[9,2]   6836    1
#> log_p_z[9,3]   3119    1
#> log_p_z[9,4]   3921    1
#> log_p_z[10,1]  4362    1
#> log_p_z[10,2]  6113    1
#> log_p_z[10,3]  3000    1
#> log_p_z[10,4]  3750    1
#> log_p_z[11,1]  4560    1
#> log_p_z[11,2]  3971    1
#> log_p_z[11,3]  3189    1
#> log_p_z[11,4]  6703    1
#> log_p_z[12,1]  4225    1
#> log_p_z[12,2]  4553    1
#> log_p_z[12,3]  3399    1
#> log_p_z[12,4]  3748    1
#> log_p_z[13,1]  7236    1
#> log_p_z[13,2]  4704    1
#> log_p_z[13,3]  3271    1
#> log_p_z[13,4]  3385    1
#> log_p_z[14,1]  4129    1
#> log_p_z[14,2]  5558    1
#> log_p_z[14,3]  3573    1
#> log_p_z[14,4]  3817    1
#> log_p_z[15,1]  7705    1
#> log_p_z[15,2]  4724    1
#> log_p_z[15,3]  3442    1
#> log_p_z[15,4]  3349    1
#> log_p_z[16,1]  5885    1
#> log_p_z[16,2]  4611    1
#> log_p_z[16,3]  3400    1
#> log_p_z[16,4]  3451    1
#> log_p_z[17,1]  7236    1
#> log_p_z[17,2]  4704    1
#> log_p_z[17,3]  3271    1
#> log_p_z[17,4]  3385    1
#> log_p_z[18,1]  7236    1
#> log_p_z[18,2]  4704    1
#> log_p_z[18,3]  3271    1
#> log_p_z[18,4]  3385    1
#> log_p_z[19,1]  5165    1
#> log_p_z[19,2]  6794    1
#> log_p_z[19,3]  3160    1
#> log_p_z[19,4]  3650    1
#> log_p_z[20,1]  3821    1
#> log_p_z[20,2]  5572    1
#> log_p_z[20,3]  3389    1
#> log_p_z[20,4]  3938    1
#> log_p_z[21,1]  4054    1
#> log_p_z[21,2]  7673    1
#> log_p_z[21,3]  3098    1
#> log_p_z[21,4]  3970    1
#> log_p_z[22,1]  5165    1
#> log_p_z[22,2]  6794    1
#> log_p_z[22,3]  3160    1
#> log_p_z[22,4]  3650    1
#> log_p_z[23,1]  4079    1
#> log_p_z[23,2]  8532    1
#> log_p_z[23,3]  3367    1
#> log_p_z[23,4]  3987    1
#> log_p_z[24,1]  3861    1
#> log_p_z[24,2]  4837    1
#> log_p_z[24,3]  3170    1
#> log_p_z[24,4]  3887    1
#> log_p_z[25,1]  7236    1
#> log_p_z[25,2]  4704    1
#> log_p_z[25,3]  3271    1
#> log_p_z[25,4]  3385    1
#> log_p_z[26,1]  7236    1
#> log_p_z[26,2]  4704    1
#> log_p_z[26,3]  3271    1
#> log_p_z[26,4]  3385    1
#> log_p_z[27,1]  4024    1
#> log_p_z[27,2]  6658    1
#> log_p_z[27,3]  2961    1
#> log_p_z[27,4]  3943    1
#> log_p_z[28,1]  7236    1
#> log_p_z[28,2]  4704    1
#> log_p_z[28,3]  3271    1
#> log_p_z[28,4]  3385    1
#> log_p_z[29,1]  7236    1
#> log_p_z[29,2]  4704    1
#> log_p_z[29,3]  3271    1
#> log_p_z[29,4]  3385    1
#> log_p_z[30,1]  5889    1
#> log_p_z[30,2]  4880    1
#> log_p_z[30,3]  3416    1
#> log_p_z[30,4]  3392    1
#> log_p_z[31,1]  7236    1
#> log_p_z[31,2]  4704    1
#> log_p_z[31,3]  3271    1
#> log_p_z[31,4]  3385    1
#> log_p_z[32,1]  4789    1
#> log_p_z[32,2]  5736    1
#> log_p_z[32,3]  5251    1
#> log_p_z[32,4]  3469    1
#> log_p_z[33,1]  7236    1
#> log_p_z[33,2]  4704    1
#> log_p_z[33,3]  3271    1
#> log_p_z[33,4]  3385    1
#> log_p_z[34,1]  4054    1
#> log_p_z[34,2]  7673    1
#> log_p_z[34,3]  3098    1
#> log_p_z[34,4]  3970    1
#> log_p_z[35,1]  4303    1
#> log_p_z[35,2]  6839    1
#> log_p_z[35,3]  3499    1
#> log_p_z[35,4]  4051    1
#> log_p_z[36,1]  3693    1
#> log_p_z[36,2]  3930    1
#> log_p_z[36,3]  2674    1
#> log_p_z[36,4]  3087    1
#> log_p_z[37,1]  4205    1
#> log_p_z[37,2]  4948    1
#> log_p_z[37,3]  3330    1
#> log_p_z[37,4]  4065    1
#> log_p_z[38,1]  4422    1
#> log_p_z[38,2]  5192    1
#> log_p_z[38,3]  3397    1
#> log_p_z[38,4]  3945    1
#> log_p_z[39,1]  4920    1
#> log_p_z[39,2]  5492    1
#> log_p_z[39,3]  6037    1
#> log_p_z[39,4]  3431    1
#> log_p_z[40,1]  7236    1
#> log_p_z[40,2]  4704    1
#> log_p_z[40,3]  3271    1
#> log_p_z[40,4]  3385    1
#> log_p_z[41,1]  7236    1
#> log_p_z[41,2]  4704    1
#> log_p_z[41,3]  3271    1
#> log_p_z[41,4]  3385    1
#> log_p_z[42,1]  5793    1
#> log_p_z[42,2]  4539    1
#> log_p_z[42,3]  3468    1
#> log_p_z[42,4]  3442    1
#> log_p_z[43,1]  4024    1
#> log_p_z[43,2]  6658    1
#> log_p_z[43,3]  2961    1
#> log_p_z[43,4]  3943    1
#> log_p_z[44,1]  7705    1
#> log_p_z[44,2]  4724    1
#> log_p_z[44,3]  3442    1
#> log_p_z[44,4]  3349    1
#> log_p_z[45,1]  4054    1
#> log_p_z[45,2]  7673    1
#> log_p_z[45,3]  3098    1
#> log_p_z[45,4]  3970    1
#> log_lik[1]     7236    1
#> log_lik[2]     5520    1
#> log_lik[3]     4934    1
#> log_lik[4]     5915    1
#> log_lik[5]     8532    1
#> log_lik[6]     5994    1
#> log_lik[7]     5795    1
#> log_lik[8]     6298    1
#> log_lik[9]     6837    1
#> log_lik[10]    6131    1
#> log_lik[11]    6706    1
#> log_lik[12]    6308    1
#> log_lik[13]    7236    1
#> log_lik[14]    5665    1
#> log_lik[15]    7705    1
#> log_lik[16]    5885    1
#> log_lik[17]    7236    1
#> log_lik[18]    7236    1
#> log_lik[19]    6794    1
#> log_lik[20]    5574    1
#> log_lik[21]    7673    1
#> log_lik[22]    6794    1
#> log_lik[23]    8532    1
#> log_lik[24]    4837    1
#> log_lik[25]    7236    1
#> log_lik[26]    7236    1
#> log_lik[27]    6657    1
#> log_lik[28]    7236    1
#> log_lik[29]    7236    1
#> log_lik[30]    5901    1
#> log_lik[31]    7236    1
#> log_lik[32]    5256    1
#> log_lik[33]    7236    1
#> log_lik[34]    7673    1
#> log_lik[35]    6911    1
#> log_lik[36]    3853    1
#> log_lik[37]    4957    1
#> log_lik[38]    4641    1
#> log_lik[39]    6028    1
#> log_lik[40]    7236    1
#> log_lik[41]    7236    1
#> log_lik[42]    5795    1
#> log_lik[43]    6657    1
#> log_lik[44]    7705    1
#> log_lik[45]    7673    1
#> lp__           1285    1
#> 
#> Samples were drawn using NUTS(diag_e) at Mon Feb  6 07:43:11 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }