Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.25    0.33    0.37    0.42    0.50
#> pi[2]            0.41    0.00 0.06    0.28    0.36    0.40    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.03    0.02    0.05    0.07    0.10    0.15
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.93
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.13
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.83    0.86    0.89    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.06    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.00    0.01    0.02    0.05
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.11    0.28
#> theta[1,3,3]     0.80    0.00 0.09    0.59    0.74    0.81    0.87    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.09    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.18    0.41
#> theta[1,4,4]     0.69    0.00 0.14    0.40    0.59    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.09    0.57    0.69    0.76    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.21    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.58    0.64    0.75
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.24    0.30    0.37    0.49
#> theta[2,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.26
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[2,3,3]     0.62    0.00 0.13    0.35    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.22    0.00 0.12    0.04    0.13    0.20    0.29    0.48
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.77    0.90
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.10    0.00 0.06    0.02    0.05    0.09    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.88
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[3,3,2]     0.18    0.00 0.10    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.13    0.25    0.41    0.51    0.60    0.75
#> theta[3,3,4]     0.23    0.00 0.12    0.06    0.15    0.22    0.31    0.49
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[3,4,2]     0.11    0.00 0.10    0.00    0.04    0.08    0.16    0.37
#> theta[3,4,3]     0.14    0.00 0.12    0.01    0.05    0.11    0.21    0.45
#> theta[3,4,4]     0.62    0.00 0.16    0.31    0.51    0.63    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.08    0.66    0.78    0.83    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.08    0.55    0.67    0.73    0.78    0.87
#> theta[4,2,3]     0.13    0.00 0.06    0.03    0.09    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.10    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,3]     0.69    0.00 0.13    0.43    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.39
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,2]     0.11    0.00 0.10    0.00    0.04    0.08    0.16    0.38
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.77    0.91
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.82    0.87    0.92    0.96
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.06    0.05    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.60    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.38
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.61    0.70    0.84
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.42
#> theta[5,4,4]     0.63    0.00 0.15    0.32    0.52    0.64    0.75    0.90
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_p_z[1,2]   -19.98    0.04 2.26  -24.88  -21.36  -19.79  -18.40  -16.06
#> log_p_z[1,3]   -24.94    0.08 4.37  -34.76  -27.42  -24.37  -21.84  -17.89
#> log_p_z[1,4]   -21.96    0.08 4.40  -32.37  -24.35  -21.51  -18.91  -14.82
#> log_p_z[2,1]   -29.68    0.07 4.25  -39.34  -32.12  -29.27  -26.68  -22.71
#> log_p_z[2,2]   -21.03    0.04 2.39  -26.16  -22.54  -20.85  -19.33  -16.93
#> log_p_z[2,3]    -7.61    0.02 1.18  -10.20   -8.27   -7.49   -6.81   -5.65
#> log_p_z[2,4]   -16.40    0.08 4.32  -26.19  -18.78  -15.89  -13.34   -9.64
#> log_p_z[3,1]   -11.65    0.03 1.68  -15.54  -12.61  -11.44  -10.46   -8.94
#> log_p_z[3,2]   -10.80    0.02 1.45  -14.07  -11.68  -10.64   -9.75   -8.46
#> log_p_z[3,3]   -23.41    0.06 3.48  -31.35  -25.53  -23.11  -20.89  -17.62
#> log_p_z[3,4]   -21.94    0.06 3.67  -30.05  -24.17  -21.60  -19.34  -15.64
#> log_p_z[4,1]   -14.69    0.03 1.91  -18.89  -15.84  -14.57  -13.36  -11.42
#> log_p_z[4,2]    -7.53    0.01 0.93   -9.54   -8.10   -7.47   -6.89   -5.90
#> log_p_z[4,3]   -20.79    0.08 4.35  -30.88  -23.27  -20.29  -17.65  -13.76
#> log_p_z[4,4]   -21.95    0.08 4.55  -32.29  -24.70  -21.34  -18.73  -14.57
#> log_p_z[5,1]   -21.58    0.04 2.70  -27.53  -23.24  -21.34  -19.67  -16.96
#> log_p_z[5,2]    -3.70    0.01 0.43   -4.59   -3.98   -3.68   -3.40   -2.88
#> log_p_z[5,3]   -18.80    0.08 4.26  -28.44  -21.28  -18.33  -15.69  -11.91
#> log_p_z[5,4]   -21.96    0.08 4.49  -32.05  -24.59  -21.42  -18.84  -14.77
#> log_p_z[6,1]   -21.52    0.04 2.62  -27.37  -23.10  -21.35  -19.75  -16.86
#> log_p_z[6,2]    -5.47    0.01 0.66   -6.85   -5.89   -5.44   -5.01   -4.28
#> log_p_z[6,3]   -17.60    0.08 4.12  -27.00  -20.06  -17.00  -14.61  -11.09
#> log_p_z[6,4]   -21.69    0.08 4.40  -31.82  -24.22  -21.19  -18.66  -14.66
#> log_p_z[7,1]    -6.00    0.01 0.61   -7.32   -6.38   -5.95   -5.57   -4.93
#> log_p_z[7,2]   -14.91    0.03 1.66  -18.53  -15.91  -14.76  -13.77  -12.03
#> log_p_z[7,3]   -24.37    0.07 3.61  -32.52  -26.58  -24.04  -21.79  -18.41
#> log_p_z[7,4]   -21.95    0.07 3.66  -29.90  -24.10  -21.62  -19.44  -15.76
#> log_p_z[8,1]   -29.69    0.07 4.23  -39.13  -32.16  -29.33  -26.70  -22.71
#> log_p_z[8,2]   -18.35    0.03 2.14  -22.95  -19.68  -18.19  -16.87  -14.61
#> log_p_z[8,3]    -5.72    0.01 0.84   -7.53   -6.26   -5.68   -5.12   -4.24
#> log_p_z[8,4]   -18.63    0.08 4.39  -28.88  -21.07  -18.02  -15.59  -11.93
#> log_p_z[9,1]   -20.00    0.04 2.33  -25.07  -21.40  -19.82  -18.38  -15.92
#> log_p_z[9,2]    -4.52    0.01 0.57   -5.73   -4.86   -4.49   -4.12   -3.52
#> log_p_z[9,3]   -19.73    0.08 4.42  -29.72  -22.38  -19.18  -16.57  -12.76
#> log_p_z[9,4]   -21.79    0.08 4.38  -31.61  -24.33  -21.25  -18.71  -14.68
#> log_p_z[10,1]  -22.25    0.04 2.40  -27.52  -23.72  -22.08  -20.62  -17.96
#> log_p_z[10,2]   -7.14    0.01 0.71   -8.69   -7.57   -7.10   -6.64   -5.90
#> log_p_z[10,3]  -16.86    0.06 3.24  -23.96  -18.81  -16.50  -14.58  -11.67
#> log_p_z[10,4]  -21.46    0.07 3.52  -28.95  -23.62  -21.14  -18.94  -15.52
#> log_p_z[11,1]  -29.65    0.06 4.30  -39.37  -32.25  -29.25  -26.59  -22.52
#> log_p_z[11,2]  -29.39    0.07 4.26  -38.99  -31.92  -28.93  -26.39  -22.31
#> log_p_z[11,3]  -18.26    0.07 3.55  -27.01  -20.05  -17.72  -15.85  -13.00
#> log_p_z[11,4]   -5.83    0.01 0.97   -7.85   -6.46   -5.77   -5.13   -4.16
#> log_p_z[12,1]  -22.96    0.04 2.75  -28.81  -24.75  -22.77  -21.01  -18.11
#> log_p_z[12,2]   -9.58    0.02 1.26  -12.51  -10.27   -9.45   -8.71   -7.52
#> log_p_z[12,3]  -15.27    0.08 4.09  -24.88  -17.58  -14.71  -12.31   -9.09
#> log_p_z[12,4]  -19.29    0.08 4.20  -29.25  -21.58  -18.72  -16.41  -12.65
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_p_z[13,2]  -19.98    0.04 2.26  -24.88  -21.36  -19.79  -18.40  -16.06
#> log_p_z[13,3]  -24.94    0.08 4.37  -34.76  -27.42  -24.37  -21.84  -17.89
#> log_p_z[13,4]  -21.96    0.08 4.40  -32.37  -24.35  -21.51  -18.91  -14.82
#> log_p_z[14,1]  -19.18    0.04 2.52  -24.72  -20.75  -18.95  -17.39  -14.89
#> log_p_z[14,2]   -6.22    0.01 0.85   -8.04   -6.75   -6.15   -5.63   -4.76
#> log_p_z[14,3]  -18.85    0.08 4.22  -28.38  -21.23  -18.39  -15.73  -12.20
#> log_p_z[14,4]  -21.94    0.08 4.47  -32.26  -24.55  -21.43  -18.79  -14.77
#> log_p_z[15,1]   -4.41    0.01 0.40   -5.28   -4.66   -4.39   -4.12   -3.73
#> log_p_z[15,2]  -17.22    0.03 1.82  -21.04  -18.33  -17.10  -15.92  -14.08
#> log_p_z[15,3]  -24.41    0.07 3.56  -32.50  -26.54  -24.03  -21.93  -18.47
#> log_p_z[15,4]  -21.95    0.07 3.64  -30.18  -24.11  -21.64  -19.39  -15.82
#> log_p_z[16,1]   -3.80    0.01 0.55   -5.03   -4.13   -3.76   -3.42   -2.84
#> log_p_z[16,2]  -17.68    0.04 2.14  -22.38  -18.99  -17.45  -16.19  -14.06
#> log_p_z[16,3]  -24.89    0.09 4.42  -34.89  -27.47  -24.34  -21.69  -17.74
#> log_p_z[16,4]  -21.96    0.08 4.41  -32.08  -24.39  -21.48  -18.92  -14.86
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_p_z[17,2]  -19.98    0.04 2.26  -24.88  -21.36  -19.79  -18.40  -16.06
#> log_p_z[17,3]  -24.94    0.08 4.37  -34.76  -27.42  -24.37  -21.84  -17.89
#> log_p_z[17,4]  -21.96    0.08 4.40  -32.37  -24.35  -21.51  -18.91  -14.82
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_p_z[18,2]  -19.98    0.04 2.26  -24.88  -21.36  -19.79  -18.40  -16.06
#> log_p_z[18,3]  -24.94    0.08 4.37  -34.76  -27.42  -24.37  -21.84  -17.89
#> log_p_z[18,4]  -21.96    0.08 4.40  -32.37  -24.35  -21.51  -18.91  -14.82
#> log_p_z[19,1]  -16.44    0.03 1.99  -20.86  -17.60  -16.26  -15.06  -13.12
#> log_p_z[19,2]   -4.64    0.01 0.58   -5.83   -5.00   -4.61   -4.25   -3.60
#> log_p_z[19,3]  -22.18    0.09 4.55  -32.77  -24.87  -21.58  -18.96  -14.83
#> log_p_z[19,4]  -22.00    0.08 4.50  -32.19  -24.48  -21.42  -18.91  -14.69
#> log_p_z[20,1]  -18.04    0.04 2.30  -23.10  -19.47  -17.83  -16.43  -13.98
#> log_p_z[20,2]   -7.11    0.01 0.97   -9.27   -7.69   -7.03   -6.42   -5.44
#> log_p_z[20,3]  -20.12    0.08 4.28  -29.73  -22.67  -19.53  -17.05  -13.39
#> log_p_z[20,4]  -21.70    0.08 4.35  -32.09  -24.27  -21.13  -18.65  -14.70
#> log_p_z[21,1]  -19.69    0.04 2.44  -25.01  -21.20  -19.46  -17.96  -15.51
#> log_p_z[21,2]   -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.80   -2.38
#> log_p_z[21,3]  -21.28    0.08 4.44  -31.50  -23.97  -20.71  -18.18  -14.16
#> log_p_z[21,4]  -21.97    0.08 4.47  -31.99  -24.61  -21.40  -18.80  -14.54
#> log_p_z[22,1]  -16.44    0.03 1.99  -20.86  -17.60  -16.26  -15.06  -13.12
#> log_p_z[22,2]   -4.64    0.01 0.58   -5.83   -5.00   -4.61   -4.25   -3.60
#> log_p_z[22,3]  -22.18    0.09 4.55  -32.77  -24.87  -21.58  -18.96  -14.83
#> log_p_z[22,4]  -22.00    0.08 4.50  -32.19  -24.48  -21.42  -18.91  -14.69
#> log_p_z[23,1]  -21.58    0.04 2.70  -27.53  -23.24  -21.34  -19.67  -16.96
#> log_p_z[23,2]   -3.70    0.01 0.43   -4.59   -3.98   -3.68   -3.40   -2.88
#> log_p_z[23,3]  -18.80    0.08 4.26  -28.44  -21.28  -18.33  -15.69  -11.91
#> log_p_z[23,4]  -21.96    0.08 4.49  -32.05  -24.59  -21.42  -18.84  -14.77
#> log_p_z[24,1]  -17.49    0.04 2.25  -22.43  -18.88  -17.25  -15.90  -13.75
#> log_p_z[24,2]   -5.81    0.01 0.61   -7.17   -6.17   -5.75   -5.37   -4.78
#> log_p_z[24,3]  -21.80    0.06 3.50  -29.67  -23.87  -21.46  -19.32  -15.95
#> log_p_z[24,4]  -21.97    0.06 3.68  -29.99  -24.29  -21.64  -19.36  -15.65
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_p_z[25,2]  -19.98    0.04 2.26  -24.88  -21.36  -19.79  -18.40  -16.06
#> log_p_z[25,3]  -24.94    0.08 4.37  -34.76  -27.42  -24.37  -21.84  -17.89
#> log_p_z[25,4]  -21.96    0.08 4.40  -32.37  -24.35  -21.51  -18.91  -14.82
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_p_z[26,2]  -19.98    0.04 2.26  -24.88  -21.36  -19.79  -18.40  -16.06
#> log_p_z[26,3]  -24.94    0.08 4.37  -34.76  -27.42  -24.37  -21.84  -17.89
#> log_p_z[26,4]  -21.96    0.08 4.40  -32.37  -24.35  -21.51  -18.91  -14.82
#> log_p_z[27,1]  -21.94    0.04 2.50  -27.42  -23.46  -21.77  -20.18  -17.62
#> log_p_z[27,2]   -5.66    0.01 0.52   -6.73   -5.99   -5.63   -5.30   -4.74
#> log_p_z[27,3]  -18.40    0.06 3.28  -25.91  -20.42  -18.03  -16.13  -13.06
#> log_p_z[27,4]  -21.64    0.06 3.56  -29.47  -23.80  -21.31  -19.04  -15.53
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_p_z[28,2]  -19.98    0.04 2.26  -24.88  -21.36  -19.79  -18.40  -16.06
#> log_p_z[28,3]  -24.94    0.08 4.37  -34.76  -27.42  -24.37  -21.84  -17.89
#> log_p_z[28,4]  -21.96    0.08 4.40  -32.37  -24.35  -21.51  -18.91  -14.82
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_p_z[29,2]  -19.98    0.04 2.26  -24.88  -21.36  -19.79  -18.40  -16.06
#> log_p_z[29,3]  -24.94    0.08 4.37  -34.76  -27.42  -24.37  -21.84  -17.89
#> log_p_z[29,4]  -21.96    0.08 4.40  -32.37  -24.35  -21.51  -18.91  -14.82
#> log_p_z[30,1]   -6.82    0.01 0.84   -8.70   -7.31   -6.72   -6.22   -5.39
#> log_p_z[30,2]  -14.70    0.03 1.68  -18.41  -15.69  -14.57  -13.50  -11.82
#> log_p_z[30,3]  -24.36    0.07 3.58  -32.32  -26.56  -24.01  -21.82  -18.28
#> log_p_z[30,4]  -21.97    0.07 3.68  -29.99  -24.26  -21.58  -19.41  -15.71
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_p_z[31,2]  -19.98    0.04 2.26  -24.88  -21.36  -19.79  -18.40  -16.06
#> log_p_z[31,3]  -24.94    0.08 4.37  -34.76  -27.42  -24.37  -21.84  -17.89
#> log_p_z[31,4]  -21.96    0.08 4.40  -32.37  -24.35  -21.51  -18.91  -14.82
#> log_p_z[32,1]  -29.76    0.07 4.26  -39.25  -32.25  -29.36  -26.77  -22.72
#> log_p_z[32,2]  -14.84    0.02 1.72  -18.75  -15.88  -14.71  -13.64  -11.87
#> log_p_z[32,3]   -6.05    0.01 0.93   -8.08   -6.61   -5.98   -5.38   -4.44
#> log_p_z[32,4]  -20.78    0.08 4.52  -31.17  -23.32  -20.21  -17.68  -13.72
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_p_z[33,2]  -19.98    0.04 2.26  -24.88  -21.36  -19.79  -18.40  -16.06
#> log_p_z[33,3]  -24.94    0.08 4.37  -34.76  -27.42  -24.37  -21.84  -17.89
#> log_p_z[33,4]  -21.96    0.08 4.40  -32.37  -24.35  -21.51  -18.91  -14.82
#> log_p_z[34,1]  -19.69    0.04 2.44  -25.01  -21.20  -19.46  -17.96  -15.51
#> log_p_z[34,2]   -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.80   -2.38
#> log_p_z[34,3]  -21.28    0.08 4.44  -31.50  -23.97  -20.71  -18.18  -14.16
#> log_p_z[34,4]  -21.97    0.08 4.47  -31.99  -24.61  -21.40  -18.80  -14.54
#> log_p_z[35,1]  -22.70    0.04 2.85  -29.02  -24.43  -22.49  -20.68  -17.69
#> log_p_z[35,2]   -5.50    0.01 0.68   -7.00   -5.92   -5.45   -5.04   -4.28
#> log_p_z[35,3]  -16.21    0.08 4.11  -25.70  -18.54  -15.66  -13.19   -9.86
#> log_p_z[35,4]  -21.94    0.08 4.46  -31.92  -24.44  -21.42  -18.84  -14.73
#> log_p_z[36,1]  -29.68    0.06 3.61  -37.40  -31.91  -29.37  -27.11  -23.49
#> log_p_z[36,2]  -21.90    0.04 2.30  -26.86  -23.39  -21.76  -20.27  -17.86
#> log_p_z[36,3]  -10.51    0.04 1.78  -15.06  -11.36  -10.23   -9.29   -7.91
#> log_p_z[36,4]  -13.97    0.06 3.29  -21.27  -15.84  -13.66  -11.61   -8.61
#> log_p_z[37,1]  -18.61    0.04 2.40  -24.09  -20.05  -18.43  -16.90  -14.60
#> log_p_z[37,2]   -7.61    0.01 0.77   -9.29   -8.10   -7.55   -7.06   -6.25
#> log_p_z[37,3]  -19.21    0.06 3.30  -26.91  -21.15  -18.78  -16.84  -13.75
#> log_p_z[37,4]  -21.95    0.06 3.68  -29.77  -24.23  -21.59  -19.39  -15.69
#> log_p_z[38,1]  -25.26    0.04 2.84  -31.31  -27.06  -25.10  -23.26  -20.27
#> log_p_z[38,2]   -9.60    0.01 0.98  -11.76  -10.21   -9.53   -8.89   -7.89
#> log_p_z[38,3]  -11.79    0.05 2.80  -18.27  -13.42  -11.43   -9.68   -7.53
#> log_p_z[38,4]  -21.43    0.07 3.56  -29.28  -23.53  -21.07  -18.99  -15.46
#> log_p_z[39,1]  -29.38    0.06 4.23  -39.08  -31.85  -28.92  -26.41  -22.32
#> log_p_z[39,2]  -16.88    0.03 2.05  -21.35  -18.12  -16.72  -15.42  -13.31
#> log_p_z[39,3]   -7.26    0.02 1.09   -9.61   -7.96   -7.17   -6.50   -5.34
#> log_p_z[39,4]  -18.81    0.08 4.30  -28.61  -21.08  -18.24  -15.81  -12.25
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_p_z[40,2]  -19.98    0.04 2.26  -24.88  -21.36  -19.79  -18.40  -16.06
#> log_p_z[40,3]  -24.94    0.08 4.37  -34.76  -27.42  -24.37  -21.84  -17.89
#> log_p_z[40,4]  -21.96    0.08 4.40  -32.37  -24.35  -21.51  -18.91  -14.82
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_p_z[41,2]  -19.98    0.04 2.26  -24.88  -21.36  -19.79  -18.40  -16.06
#> log_p_z[41,3]  -24.94    0.08 4.37  -34.76  -27.42  -24.37  -21.84  -17.89
#> log_p_z[41,4]  -21.96    0.08 4.40  -32.37  -24.35  -21.51  -18.91  -14.82
#> log_p_z[42,1]   -6.00    0.01 0.61   -7.32   -6.38   -5.95   -5.57   -4.93
#> log_p_z[42,2]  -14.91    0.03 1.66  -18.53  -15.91  -14.76  -13.77  -12.03
#> log_p_z[42,3]  -24.37    0.07 3.61  -32.52  -26.58  -24.04  -21.79  -18.41
#> log_p_z[42,4]  -21.95    0.07 3.66  -29.90  -24.10  -21.62  -19.44  -15.76
#> log_p_z[43,1]  -21.94    0.04 2.50  -27.42  -23.46  -21.77  -20.18  -17.62
#> log_p_z[43,2]   -5.66    0.01 0.52   -6.73   -5.99   -5.63   -5.30   -4.74
#> log_p_z[43,3]  -18.40    0.06 3.28  -25.91  -20.42  -18.03  -16.13  -13.06
#> log_p_z[43,4]  -21.64    0.06 3.56  -29.47  -23.80  -21.31  -19.04  -15.53
#> log_p_z[44,1]   -4.41    0.01 0.40   -5.28   -4.66   -4.39   -4.12   -3.73
#> log_p_z[44,2]  -17.22    0.03 1.82  -21.04  -18.33  -17.10  -15.92  -14.08
#> log_p_z[44,3]  -24.41    0.07 3.56  -32.50  -26.54  -24.03  -21.93  -18.47
#> log_p_z[44,4]  -21.95    0.07 3.64  -30.18  -24.11  -21.64  -19.39  -15.82
#> log_p_z[45,1]  -19.69    0.04 2.44  -25.01  -21.20  -19.46  -17.96  -15.51
#> log_p_z[45,2]   -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.80   -2.38
#> log_p_z[45,3]  -21.28    0.08 4.44  -31.50  -23.97  -20.71  -18.18  -14.16
#> log_p_z[45,4]  -21.97    0.08 4.47  -31.99  -24.61  -21.40  -18.80  -14.54
#> log_lik[1]      -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_lik[2]      -7.57    0.02 1.11   -9.97   -8.23   -7.48   -6.80   -5.65
#> log_lik[3]      -9.95    0.01 0.90  -11.76  -10.56   -9.91   -9.32   -8.29
#> log_lik[4]      -7.52    0.01 0.92   -9.50   -8.09   -7.47   -6.89   -5.90
#> log_lik[5]      -3.70    0.01 0.43   -4.59   -3.98   -3.68   -3.40   -2.88
#> log_lik[6]      -5.47    0.01 0.66   -6.84   -5.89   -5.44   -5.01   -4.28
#> log_lik[7]      -6.00    0.01 0.61   -7.32   -6.38   -5.95   -5.57   -4.93
#> log_lik[8]      -5.72    0.01 0.84   -7.53   -6.26   -5.68   -5.12   -4.23
#> log_lik[9]      -4.52    0.01 0.57   -5.73   -4.86   -4.49   -4.12   -3.52
#> log_lik[10]     -7.13    0.01 0.71   -8.68   -7.56   -7.09   -6.63   -5.89
#> log_lik[11]     -5.83    0.01 0.97   -7.85   -6.46   -5.77   -5.13   -4.16
#> log_lik[12]     -9.34    0.01 1.03  -11.55  -10.01   -9.29   -8.62   -7.49
#> log_lik[13]     -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_lik[14]     -6.22    0.01 0.85   -8.04   -6.75   -6.15   -5.63   -4.76
#> log_lik[15]     -4.41    0.01 0.40   -5.28   -4.66   -4.39   -4.12   -3.73
#> log_lik[16]     -3.80    0.01 0.55   -5.03   -4.13   -3.76   -3.42   -2.84
#> log_lik[17]     -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_lik[18]     -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_lik[19]     -4.64    0.01 0.58   -5.83   -5.00   -4.61   -4.24   -3.60
#> log_lik[20]     -7.11    0.01 0.97   -9.27   -7.69   -7.03   -6.42   -5.44
#> log_lik[21]     -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.80   -2.38
#> log_lik[22]     -4.64    0.01 0.58   -5.83   -5.00   -4.61   -4.24   -3.60
#> log_lik[23]     -3.70    0.01 0.43   -4.59   -3.98   -3.68   -3.40   -2.88
#> log_lik[24]     -5.81    0.01 0.61   -7.17   -6.17   -5.75   -5.37   -4.78
#> log_lik[25]     -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_lik[26]     -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_lik[27]     -5.66    0.01 0.52   -6.73   -5.99   -5.63   -5.30   -4.74
#> log_lik[28]     -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_lik[29]     -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_lik[30]     -6.81    0.01 0.84   -8.69   -7.31   -6.72   -6.22   -5.39
#> log_lik[31]     -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_lik[32]     -6.04    0.01 0.93   -8.07   -6.61   -5.98   -5.38   -4.44
#> log_lik[33]     -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_lik[34]     -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.80   -2.38
#> log_lik[35]     -5.50    0.01 0.68   -6.99   -5.92   -5.44   -5.04   -4.28
#> log_lik[36]     -9.90    0.02 1.20  -12.44  -10.65   -9.82   -9.06   -7.78
#> log_lik[37]     -7.61    0.01 0.77   -9.29   -8.10   -7.55   -7.06   -6.25
#> log_lik[38]     -9.01    0.01 0.85  -10.70   -9.57   -8.99   -8.46   -7.33
#> log_lik[39]     -7.26    0.02 1.08   -9.60   -7.96   -7.17   -6.50   -5.34
#> log_lik[40]     -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_lik[41]     -2.21    0.00 0.30   -2.86   -2.41   -2.20   -2.00   -1.66
#> log_lik[42]     -6.00    0.01 0.61   -7.32   -6.38   -5.95   -5.57   -4.93
#> log_lik[43]     -5.66    0.01 0.52   -6.73   -5.99   -5.63   -5.30   -4.74
#> log_lik[44]     -4.41    0.01 0.40   -5.28   -4.66   -4.39   -4.12   -3.73
#> log_lik[45]     -3.04    0.00 0.35   -3.76   -3.27   -3.03   -2.80   -2.38
#> lp__          -464.03    0.17 6.22 -477.19 -468.03 -463.63 -459.68 -453.04
#>               n_eff Rhat
#> pi[1]          6542    1
#> pi[2]          7204    1
#> pi[3]          5454    1
#> pi[4]          5915    1
#> theta[1,1,1]   7017    1
#> theta[1,1,2]   6197    1
#> theta[1,1,3]   4585    1
#> theta[1,1,4]   4761    1
#> theta[1,2,1]   6123    1
#> theta[1,2,2]   5999    1
#> theta[1,2,3]   5722    1
#> theta[1,2,4]   5580    1
#> theta[1,3,1]   6051    1
#> theta[1,3,2]   4196    1
#> theta[1,3,3]   5138    1
#> theta[1,3,4]   4931    1
#> theta[1,4,1]   5018    1
#> theta[1,4,2]   5900    1
#> theta[1,4,3]   4176    1
#> theta[1,4,4]   5393    1
#> theta[2,1,1]   7112    1
#> theta[2,1,2]   5764    1
#> theta[2,1,3]   4566    1
#> theta[2,1,4]   4505    1
#> theta[2,2,1]   6526    1
#> theta[2,2,2]   6370    1
#> theta[2,2,3]   6804    1
#> theta[2,2,4]   5117    1
#> theta[2,3,1]   5551    1
#> theta[2,3,2]   6179    1
#> theta[2,3,3]   5663    1
#> theta[2,3,4]   5755    1
#> theta[2,4,1]   6115    1
#> theta[2,4,2]   5843    1
#> theta[2,4,3]   5870    1
#> theta[2,4,4]   6263    1
#> theta[3,1,1]   6941    1
#> theta[3,1,2]   4899    1
#> theta[3,1,3]   5037    1
#> theta[3,1,4]   4168    1
#> theta[3,2,1]   5844    1
#> theta[3,2,2]   6836    1
#> theta[3,2,3]   7341    1
#> theta[3,2,4]   5399    1
#> theta[3,3,1]   6108    1
#> theta[3,3,2]   6352    1
#> theta[3,3,3]   5445    1
#> theta[3,3,4]   6251    1
#> theta[3,4,1]   6401    1
#> theta[3,4,2]   5042    1
#> theta[3,4,3]   4585    1
#> theta[3,4,4]   4883    1
#> theta[4,1,1]   6563    1
#> theta[4,1,2]   5465    1
#> theta[4,1,3]   5180    1
#> theta[4,1,4]   4905    1
#> theta[4,2,1]   5428    1
#> theta[4,2,2]   6587    1
#> theta[4,2,3]   6174    1
#> theta[4,2,4]   5594    1
#> theta[4,3,1]   4790    1
#> theta[4,3,2]   6146    1
#> theta[4,3,3]   5486    1
#> theta[4,3,4]   5572    1
#> theta[4,4,1]   6097    1
#> theta[4,4,2]   6021    1
#> theta[4,4,3]   5912    1
#> theta[4,4,4]   5981    1
#> theta[5,1,1]   5874    1
#> theta[5,1,2]   4723    1
#> theta[5,1,3]   3619    1
#> theta[5,1,4]   4146    1
#> theta[5,2,1]   6414    1
#> theta[5,2,2]   6577    1
#> theta[5,2,3]   6387    1
#> theta[5,2,4]   5149    1
#> theta[5,3,1]   6798    1
#> theta[5,3,2]   6187    1
#> theta[5,3,3]   5911    1
#> theta[5,3,4]   6405    1
#> theta[5,4,1]   6445    1
#> theta[5,4,2]   6034    1
#> theta[5,4,3]   4535    1
#> theta[5,4,4]   5467    1
#> log_p_z[1,1]   6275    1
#> log_p_z[1,2]   3684    1
#> log_p_z[1,3]   2674    1
#> log_p_z[1,4]   2868    1
#> log_p_z[2,1]   4182    1
#> log_p_z[2,2]   4284    1
#> log_p_z[2,3]   3994    1
#> log_p_z[2,4]   2919    1
#> log_p_z[3,1]   3069    1
#> log_p_z[3,2]   3481    1
#> log_p_z[3,3]   3052    1
#> log_p_z[3,4]   3322    1
#> log_p_z[4,1]   4714    1
#> log_p_z[4,2]   4869    1
#> log_p_z[4,3]   2869    1
#> log_p_z[4,4]   2972    1
#> log_p_z[5,1]   4119    1
#> log_p_z[5,2]   6180    1
#> log_p_z[5,3]   2953    1
#> log_p_z[5,4]   2970    1
#> log_p_z[6,1]   3853    1
#> log_p_z[6,2]   6161    1
#> log_p_z[6,3]   2922    1
#> log_p_z[6,4]   2758    1
#> log_p_z[7,1]   4920    1
#> log_p_z[7,2]   3445    1
#> log_p_z[7,3]   2702    1
#> log_p_z[7,4]   3105    1
#> log_p_z[8,1]   4145    1
#> log_p_z[8,2]   4004    1
#> log_p_z[8,3]   4903    1
#> log_p_z[8,4]   2829    1
#> log_p_z[9,1]   4149    1
#> log_p_z[9,2]   5137    1
#> log_p_z[9,3]   2928    1
#> log_p_z[9,4]   3120    1
#> log_p_z[10,1]  3857    1
#> log_p_z[10,2]  5110    1
#> log_p_z[10,3]  3059    1
#> log_p_z[10,4]  2892    1
#> log_p_z[11,1]  4706    1
#> log_p_z[11,2]  3722    1
#> log_p_z[11,3]  2462    1
#> log_p_z[11,4]  4986    1
#> log_p_z[12,1]  4439    1
#> log_p_z[12,2]  4099    1
#> log_p_z[12,3]  2901    1
#> log_p_z[12,4]  2720    1
#> log_p_z[13,1]  6275    1
#> log_p_z[13,2]  3684    1
#> log_p_z[13,3]  2674    1
#> log_p_z[13,4]  2868    1
#> log_p_z[14,1]  4082    1
#> log_p_z[14,2]  4839    1
#> log_p_z[14,3]  2959    1
#> log_p_z[14,4]  3136    1
#> log_p_z[15,1]  6157    1
#> log_p_z[15,2]  3727    1
#> log_p_z[15,3]  2704    1
#> log_p_z[15,4]  2791    1
#> log_p_z[16,1]  5910    1
#> log_p_z[16,2]  3355    1
#> log_p_z[16,3]  2690    1
#> log_p_z[16,4]  3024    1
#> log_p_z[17,1]  6275    1
#> log_p_z[17,2]  3684    1
#> log_p_z[17,3]  2674    1
#> log_p_z[17,4]  2868    1
#> log_p_z[18,1]  6275    1
#> log_p_z[18,2]  3684    1
#> log_p_z[18,3]  2674    1
#> log_p_z[18,4]  2868    1
#> log_p_z[19,1]  4303    1
#> log_p_z[19,2]  5565    1
#> log_p_z[19,3]  2792    1
#> log_p_z[19,4]  3208    1
#> log_p_z[20,1]  3996    1
#> log_p_z[20,2]  4657    1
#> log_p_z[20,3]  2869    1
#> log_p_z[20,4]  2671    1
#> log_p_z[21,1]  3799    1
#> log_p_z[21,2]  6145    1
#> log_p_z[21,3]  2914    1
#> log_p_z[21,4]  3277    1
#> log_p_z[22,1]  4303    1
#> log_p_z[22,2]  5565    1
#> log_p_z[22,3]  2792    1
#> log_p_z[22,4]  3208    1
#> log_p_z[23,1]  4119    1
#> log_p_z[23,2]  6180    1
#> log_p_z[23,3]  2953    1
#> log_p_z[23,4]  2970    1
#> log_p_z[24,1]  3602    1
#> log_p_z[24,2]  3930    1
#> log_p_z[24,3]  3041    1
#> log_p_z[24,4]  3476    1
#> log_p_z[25,1]  6275    1
#> log_p_z[25,2]  3684    1
#> log_p_z[25,3]  2674    1
#> log_p_z[25,4]  2868    1
#> log_p_z[26,1]  6275    1
#> log_p_z[26,2]  3684    1
#> log_p_z[26,3]  2674    1
#> log_p_z[26,4]  2868    1
#> log_p_z[27,1]  3754    1
#> log_p_z[27,2]  5535    1
#> log_p_z[27,3]  3050    1
#> log_p_z[27,4]  3088    1
#> log_p_z[28,1]  6275    1
#> log_p_z[28,2]  3684    1
#> log_p_z[28,3]  2674    1
#> log_p_z[28,4]  2868    1
#> log_p_z[29,1]  6275    1
#> log_p_z[29,2]  3684    1
#> log_p_z[29,3]  2674    1
#> log_p_z[29,4]  2868    1
#> log_p_z[30,1]  5079    1
#> log_p_z[30,2]  3564    1
#> log_p_z[30,3]  2922    1
#> log_p_z[30,4]  2756    1
#> log_p_z[31,1]  6275    1
#> log_p_z[31,2]  3684    1
#> log_p_z[31,3]  2674    1
#> log_p_z[31,4]  2868    1
#> log_p_z[32,1]  4090    1
#> log_p_z[32,2]  4932    1
#> log_p_z[32,3]  5005    1
#> log_p_z[32,4]  2905    1
#> log_p_z[33,1]  6275    1
#> log_p_z[33,2]  3684    1
#> log_p_z[33,3]  2674    1
#> log_p_z[33,4]  2868    1
#> log_p_z[34,1]  3799    1
#> log_p_z[34,2]  6145    1
#> log_p_z[34,3]  2914    1
#> log_p_z[34,4]  3277    1
#> log_p_z[35,1]  4263    1
#> log_p_z[35,2]  6176    1
#> log_p_z[35,3]  2808    1
#> log_p_z[35,4]  3216    1
#> log_p_z[36,1]  3986    1
#> log_p_z[36,2]  3763    1
#> log_p_z[36,3]  2436    1
#> log_p_z[36,4]  2688    1
#> log_p_z[37,1]  3665    1
#> log_p_z[37,2]  5123    1
#> log_p_z[37,3]  2812    1
#> log_p_z[37,4]  3638    1
#> log_p_z[38,1]  4273    1
#> log_p_z[38,2]  4753    1
#> log_p_z[38,3]  2827    1
#> log_p_z[38,4]  2981    1
#> log_p_z[39,1]  4257    1
#> log_p_z[39,2]  4030    1
#> log_p_z[39,3]  4943    1
#> log_p_z[39,4]  2931    1
#> log_p_z[40,1]  6275    1
#> log_p_z[40,2]  3684    1
#> log_p_z[40,3]  2674    1
#> log_p_z[40,4]  2868    1
#> log_p_z[41,1]  6275    1
#> log_p_z[41,2]  3684    1
#> log_p_z[41,3]  2674    1
#> log_p_z[41,4]  2868    1
#> log_p_z[42,1]  4920    1
#> log_p_z[42,2]  3445    1
#> log_p_z[42,3]  2702    1
#> log_p_z[42,4]  3105    1
#> log_p_z[43,1]  3754    1
#> log_p_z[43,2]  5535    1
#> log_p_z[43,3]  3050    1
#> log_p_z[43,4]  3088    1
#> log_p_z[44,1]  6157    1
#> log_p_z[44,2]  3727    1
#> log_p_z[44,3]  2704    1
#> log_p_z[44,4]  2791    1
#> log_p_z[45,1]  3799    1
#> log_p_z[45,2]  6145    1
#> log_p_z[45,3]  2914    1
#> log_p_z[45,4]  3277    1
#> log_lik[1]     6275    1
#> log_lik[2]     4343    1
#> log_lik[3]     4082    1
#> log_lik[4]     4971    1
#> log_lik[5]     6180    1
#> log_lik[6]     6169    1
#> log_lik[7]     4923    1
#> log_lik[8]     4906    1
#> log_lik[9]     5138    1
#> log_lik[10]    5135    1
#> log_lik[11]    4987    1
#> log_lik[12]    4970    1
#> log_lik[13]    6275    1
#> log_lik[14]    4843    1
#> log_lik[15]    6157    1
#> log_lik[16]    5910    1
#> log_lik[17]    6275    1
#> log_lik[18]    6275    1
#> log_lik[19]    5565    1
#> log_lik[20]    4662    1
#> log_lik[21]    6145    1
#> log_lik[22]    5565    1
#> log_lik[23]    6180    1
#> log_lik[24]    3930    1
#> log_lik[25]    6275    1
#> log_lik[26]    6275    1
#> log_lik[27]    5536    1
#> log_lik[28]    6275    1
#> log_lik[29]    6275    1
#> log_lik[30]    5088    1
#> log_lik[31]    6275    1
#> log_lik[32]    5004    1
#> log_lik[33]    6275    1
#> log_lik[34]    6145    1
#> log_lik[35]    6195    1
#> log_lik[36]    4325    1
#> log_lik[37]    5124    1
#> log_lik[38]    4694    1
#> log_lik[39]    4970    1
#> log_lik[40]    6275    1
#> log_lik[41]    6275    1
#> log_lik[42]    4923    1
#> log_lik[43]    5536    1
#> log_lik[44]    6157    1
#> log_lik[45]    6145    1
#> lp__           1409    1
#> 
#> Samples were drawn using NUTS(diag_e) at Tue Feb 14 07:55:16 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }