Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.26    0.33    0.38    0.42    0.50
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.40    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.13
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.89    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.07    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.58    0.74    0.80    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.21
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.41
#> theta[1,4,4]     0.69    0.00 0.14    0.39    0.60    0.70    0.79    0.91
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.33
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.58    0.65    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.24    0.30    0.36    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.26
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.26
#> theta[2,3,3]     0.62    0.00 0.13    0.36    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.28    0.47
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.90
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.85    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.09    0.54    0.68    0.73    0.79    0.88
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.50    0.00 0.13    0.24    0.41    0.50    0.60    0.76
#> theta[3,3,4]     0.23    0.00 0.12    0.05    0.15    0.22    0.30    0.50
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[3,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[3,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.41
#> theta[3,4,4]     0.63    0.00 0.15    0.31    0.52    0.64    0.75    0.89
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.78    0.84    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.08    0.54    0.67    0.73    0.78    0.88
#> theta[4,2,3]     0.13    0.00 0.06    0.03    0.09    0.13    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.10    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,2]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.60    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.39
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[4,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.76    0.90
#> theta[5,1,1]     0.87    0.00 0.07    0.71    0.82    0.87    0.91    0.97
#> theta[5,1,2]     0.06    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.10    0.14    0.18    0.30
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.81
#> theta[5,2,3]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.12
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,3,3]     0.61    0.00 0.13    0.33    0.52    0.61    0.70    0.84
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.14    0.21    0.38
#> theta[5,4,1]     0.12    0.00 0.10    0.01    0.04    0.09    0.17    0.36
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[5,4,3]     0.14    0.00 0.11    0.00    0.05    0.11    0.20    0.42
#> theta[5,4,4]     0.63    0.00 0.15    0.32    0.53    0.64    0.74    0.89
#> log_p_z[1,1]    -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_p_z[1,2]   -20.02    0.04 2.25  -24.99  -21.38  -19.81  -18.43  -16.23
#> log_p_z[1,3]   -24.91    0.08 4.40  -34.81  -27.46  -24.37  -21.84  -17.81
#> log_p_z[1,4]   -21.83    0.07 4.25  -31.69  -24.26  -21.38  -18.82  -14.86
#> log_p_z[2,1]   -29.74    0.07 4.44  -39.99  -32.24  -29.19  -26.63  -22.69
#> log_p_z[2,2]   -21.08    0.04 2.40  -26.24  -22.56  -20.88  -19.43  -16.89
#> log_p_z[2,3]    -7.65    0.02 1.20  -10.41   -8.32   -7.54   -6.83   -5.62
#> log_p_z[2,4]   -16.38    0.07 4.28  -26.20  -18.78  -15.89  -13.33   -9.59
#> log_p_z[3,1]   -11.60    0.03 1.66  -15.28  -12.61  -11.40  -10.36   -8.89
#> log_p_z[3,2]   -10.81    0.02 1.46  -14.09  -11.67  -10.63   -9.76   -8.43
#> log_p_z[3,3]   -23.38    0.07 3.50  -31.21  -25.55  -23.03  -20.92  -17.46
#> log_p_z[3,4]   -21.81    0.06 3.47  -29.64  -23.96  -21.46  -19.40  -15.88
#> log_p_z[4,1]   -14.62    0.03 1.87  -18.79  -15.74  -14.47  -13.31  -11.48
#> log_p_z[4,2]    -7.59    0.02 0.94   -9.64   -8.16   -7.53   -6.93   -5.96
#> log_p_z[4,3]   -20.63    0.08 4.21  -30.36  -23.07  -20.11  -17.55  -13.94
#> log_p_z[4,4]   -21.85    0.07 4.34  -31.88  -24.22  -21.33  -18.81  -14.98
#> log_p_z[5,1]   -21.44    0.04 2.52  -27.00  -22.96  -21.26  -19.64  -17.12
#> log_p_z[5,2]    -3.72    0.00 0.43   -4.63   -3.99   -3.69   -3.43   -2.92
#> log_p_z[5,3]   -18.63    0.08 4.19  -28.20  -21.04  -18.16  -15.60  -12.07
#> log_p_z[5,4]   -21.90    0.08 4.37  -31.68  -24.37  -21.39  -18.89  -14.91
#> log_p_z[6,1]   -21.40    0.04 2.48  -26.77  -22.94  -21.17  -19.65  -17.10
#> log_p_z[6,2]    -5.48    0.01 0.67   -6.96   -5.90   -5.43   -5.01   -4.29
#> log_p_z[6,3]   -17.42    0.08 4.03  -26.71  -19.77  -16.91  -14.53  -11.20
#> log_p_z[6,4]   -21.65    0.08 4.36  -31.30  -24.07  -21.12  -18.69  -14.54
#> log_p_z[7,1]    -6.00    0.01 0.61   -7.27   -6.39   -5.98   -5.57   -4.90
#> log_p_z[7,2]   -14.95    0.03 1.69  -18.67  -15.98  -14.78  -13.77  -12.10
#> log_p_z[7,3]   -24.29    0.07 3.62  -32.41  -26.49  -23.96  -21.70  -18.24
#> log_p_z[7,4]   -21.81    0.06 3.55  -29.67  -23.95  -21.49  -19.30  -15.83
#> log_p_z[8,1]   -29.72    0.06 4.33  -39.30  -32.20  -29.28  -26.66  -22.59
#> log_p_z[8,2]   -18.28    0.03 2.12  -22.94  -19.57  -18.12  -16.79  -14.59
#> log_p_z[8,3]    -5.73    0.01 0.84   -7.53   -6.25   -5.70   -5.15   -4.24
#> log_p_z[8,4]   -18.58    0.08 4.37  -28.30  -21.11  -17.96  -15.53  -11.74
#> log_p_z[9,1]   -19.99    0.03 2.27  -24.95  -21.40  -19.81  -18.36  -16.04
#> log_p_z[9,2]    -4.51    0.01 0.57   -5.74   -4.85   -4.47   -4.11   -3.50
#> log_p_z[9,3]   -19.57    0.08 4.25  -29.15  -22.01  -19.04  -16.53  -12.74
#> log_p_z[9,4]   -21.69    0.08 4.37  -31.59  -24.03  -21.06  -18.65  -14.93
#> log_p_z[10,1]  -22.25    0.03 2.39  -27.47  -23.80  -22.02  -20.53  -18.18
#> log_p_z[10,2]   -7.12    0.01 0.71   -8.66   -7.57   -7.06   -6.63   -5.88
#> log_p_z[10,3]  -16.77    0.06 3.13  -23.83  -18.60  -16.45  -14.53  -11.57
#> log_p_z[10,4]  -21.37    0.06 3.58  -29.10  -23.43  -21.02  -18.79  -15.60
#> log_p_z[11,1]  -29.78    0.07 4.31  -39.46  -32.30  -29.29  -26.66  -22.94
#> log_p_z[11,2]  -29.37    0.07 4.27  -39.03  -31.78  -28.93  -26.42  -22.42
#> log_p_z[11,3]  -18.21    0.07 3.41  -26.30  -20.01  -17.72  -15.74  -13.17
#> log_p_z[11,4]   -5.82    0.01 0.95   -7.89   -6.43   -5.75   -5.16   -4.20
#> log_p_z[12,1]  -22.92    0.04 2.65  -28.63  -24.58  -22.81  -20.99  -18.28
#> log_p_z[12,2]   -9.60    0.02 1.31  -12.69  -10.34   -9.47   -8.72   -7.43
#> log_p_z[12,3]  -15.07    0.07 3.93  -24.05  -17.44  -14.58  -12.16   -8.85
#> log_p_z[12,4]  -19.30    0.07 4.22  -28.99  -21.65  -18.68  -16.42  -12.73
#> log_p_z[13,1]   -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_p_z[13,2]  -20.02    0.04 2.25  -24.99  -21.38  -19.81  -18.43  -16.23
#> log_p_z[13,3]  -24.91    0.08 4.40  -34.81  -27.46  -24.37  -21.84  -17.81
#> log_p_z[13,4]  -21.83    0.07 4.25  -31.69  -24.26  -21.38  -18.82  -14.86
#> log_p_z[14,1]  -19.04    0.04 2.34  -24.27  -20.43  -18.86  -17.39  -15.14
#> log_p_z[14,2]   -6.25    0.01 0.87   -8.17   -6.76   -6.18   -5.64   -4.76
#> log_p_z[14,3]  -18.60    0.08 4.14  -27.81  -21.12  -18.11  -15.57  -12.07
#> log_p_z[14,4]  -21.93    0.08 4.43  -31.93  -24.34  -21.34  -18.84  -14.93
#> log_p_z[15,1]   -4.40    0.00 0.40   -5.26   -4.65   -4.37   -4.11   -3.71
#> log_p_z[15,2]  -17.26    0.03 1.82  -21.22  -18.34  -17.12  -16.00  -14.14
#> log_p_z[15,3]  -24.31    0.07 3.55  -32.24  -26.46  -24.01  -21.75  -18.35
#> log_p_z[15,4]  -21.86    0.06 3.54  -29.95  -23.99  -21.51  -19.44  -15.90
#> log_p_z[16,1]   -3.80    0.01 0.53   -4.92   -4.14   -3.76   -3.43   -2.88
#> log_p_z[16,2]  -17.71    0.03 2.15  -22.40  -19.02  -17.49  -16.19  -14.15
#> log_p_z[16,3]  -24.89    0.08 4.47  -35.19  -27.41  -24.38  -21.65  -17.72
#> log_p_z[16,4]  -21.78    0.07 4.26  -31.72  -24.27  -21.34  -18.79  -14.84
#> log_p_z[17,1]   -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_p_z[17,2]  -20.02    0.04 2.25  -24.99  -21.38  -19.81  -18.43  -16.23
#> log_p_z[17,3]  -24.91    0.08 4.40  -34.81  -27.46  -24.37  -21.84  -17.81
#> log_p_z[17,4]  -21.83    0.07 4.25  -31.69  -24.26  -21.38  -18.82  -14.86
#> log_p_z[18,1]   -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_p_z[18,2]  -20.02    0.04 2.25  -24.99  -21.38  -19.81  -18.43  -16.23
#> log_p_z[18,3]  -24.91    0.08 4.40  -34.81  -27.46  -24.37  -21.84  -17.81
#> log_p_z[18,4]  -21.83    0.07 4.25  -31.69  -24.26  -21.38  -18.82  -14.86
#> log_p_z[19,1]  -16.40    0.03 1.95  -20.56  -17.62  -16.26  -15.06  -12.97
#> log_p_z[19,2]   -4.65    0.01 0.60   -5.98   -5.03   -4.61   -4.23   -3.57
#> log_p_z[19,3]  -22.05    0.08 4.32  -31.64  -24.63  -21.57  -18.91  -15.24
#> log_p_z[19,4]  -21.86    0.08 4.42  -31.88  -24.28  -21.22  -18.68  -15.12
#> log_p_z[20,1]  -17.97    0.03 2.19  -22.66  -19.36  -17.77  -16.40  -14.26
#> log_p_z[20,2]   -7.11    0.02 0.97   -9.20   -7.70   -7.03   -6.44   -5.46
#> log_p_z[20,3]  -19.93    0.08 4.14  -29.11  -22.24  -19.45  -17.01  -13.39
#> log_p_z[20,4]  -21.69    0.08 4.40  -31.78  -24.16  -21.16  -18.67  -14.66
#> log_p_z[21,1]  -19.61    0.03 2.35  -24.81  -21.01  -19.49  -17.97  -15.48
#> log_p_z[21,2]   -3.04    0.00 0.36   -3.76   -3.27   -3.03   -2.80   -2.38
#> log_p_z[21,3]  -21.11    0.08 4.29  -30.91  -23.65  -20.65  -18.01  -14.22
#> log_p_z[21,4]  -21.89    0.08 4.41  -32.03  -24.40  -21.28  -18.82  -15.07
#> log_p_z[22,1]  -16.40    0.03 1.95  -20.56  -17.62  -16.26  -15.06  -12.97
#> log_p_z[22,2]   -4.65    0.01 0.60   -5.98   -5.03   -4.61   -4.23   -3.57
#> log_p_z[22,3]  -22.05    0.08 4.32  -31.64  -24.63  -21.57  -18.91  -15.24
#> log_p_z[22,4]  -21.86    0.08 4.42  -31.88  -24.28  -21.22  -18.68  -15.12
#> log_p_z[23,1]  -21.44    0.04 2.52  -27.00  -22.96  -21.26  -19.64  -17.12
#> log_p_z[23,2]   -3.72    0.00 0.43   -4.63   -3.99   -3.69   -3.43   -2.92
#> log_p_z[23,3]  -18.63    0.08 4.19  -28.20  -21.04  -18.16  -15.60  -12.07
#> log_p_z[23,4]  -21.90    0.08 4.37  -31.68  -24.37  -21.39  -18.89  -14.91
#> log_p_z[24,1]  -17.41    0.03 2.16  -22.17  -18.71  -17.28  -15.90  -13.71
#> log_p_z[24,2]   -5.79    0.01 0.62   -7.20   -6.16   -5.72   -5.36   -4.78
#> log_p_z[24,3]  -21.71    0.07 3.45  -29.08  -23.83  -21.48  -19.19  -15.82
#> log_p_z[24,4]  -21.86    0.06 3.56  -29.96  -23.95  -21.49  -19.33  -15.92
#> log_p_z[25,1]   -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_p_z[25,2]  -20.02    0.04 2.25  -24.99  -21.38  -19.81  -18.43  -16.23
#> log_p_z[25,3]  -24.91    0.08 4.40  -34.81  -27.46  -24.37  -21.84  -17.81
#> log_p_z[25,4]  -21.83    0.07 4.25  -31.69  -24.26  -21.38  -18.82  -14.86
#> log_p_z[26,1]   -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_p_z[26,2]  -20.02    0.04 2.25  -24.99  -21.38  -19.81  -18.43  -16.23
#> log_p_z[26,3]  -24.91    0.08 4.40  -34.81  -27.46  -24.37  -21.84  -17.81
#> log_p_z[26,4]  -21.83    0.07 4.25  -31.69  -24.26  -21.38  -18.82  -14.86
#> log_p_z[27,1]  -21.88    0.04 2.45  -27.25  -23.33  -21.71  -20.18  -17.61
#> log_p_z[27,2]   -5.65    0.01 0.53   -6.79   -5.99   -5.60   -5.26   -4.73
#> log_p_z[27,3]  -18.31    0.06 3.20  -25.62  -20.22  -18.01  -16.03  -13.02
#> log_p_z[27,4]  -21.57    0.06 3.53  -29.41  -23.60  -21.19  -19.05  -15.92
#> log_p_z[28,1]   -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_p_z[28,2]  -20.02    0.04 2.25  -24.99  -21.38  -19.81  -18.43  -16.23
#> log_p_z[28,3]  -24.91    0.08 4.40  -34.81  -27.46  -24.37  -21.84  -17.81
#> log_p_z[28,4]  -21.83    0.07 4.25  -31.69  -24.26  -21.38  -18.82  -14.86
#> log_p_z[29,1]   -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_p_z[29,2]  -20.02    0.04 2.25  -24.99  -21.38  -19.81  -18.43  -16.23
#> log_p_z[29,3]  -24.91    0.08 4.40  -34.81  -27.46  -24.37  -21.84  -17.81
#> log_p_z[29,4]  -21.83    0.07 4.25  -31.69  -24.26  -21.38  -18.82  -14.86
#> log_p_z[30,1]   -6.79    0.01 0.83   -8.65   -7.29   -6.72   -6.19   -5.44
#> log_p_z[30,2]  -14.73    0.03 1.68  -18.39  -15.73  -14.60  -13.58  -11.82
#> log_p_z[30,3]  -24.34    0.07 3.55  -32.14  -26.47  -24.03  -21.77  -18.45
#> log_p_z[30,4]  -21.83    0.06 3.47  -29.52  -23.89  -21.48  -19.39  -15.98
#> log_p_z[31,1]   -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_p_z[31,2]  -20.02    0.04 2.25  -24.99  -21.38  -19.81  -18.43  -16.23
#> log_p_z[31,3]  -24.91    0.08 4.40  -34.81  -27.46  -24.37  -21.84  -17.81
#> log_p_z[31,4]  -21.83    0.07 4.25  -31.69  -24.26  -21.38  -18.82  -14.86
#> log_p_z[32,1]  -29.75    0.06 4.34  -39.51  -32.39  -29.35  -26.64  -22.75
#> log_p_z[32,2]  -14.82    0.02 1.75  -18.50  -15.93  -14.69  -13.60  -11.80
#> log_p_z[32,3]   -6.07    0.01 0.95   -8.19   -6.63   -6.00   -5.40   -4.43
#> log_p_z[32,4]  -20.74    0.08 4.52  -30.82  -23.30  -20.23  -17.55  -13.64
#> log_p_z[33,1]   -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_p_z[33,2]  -20.02    0.04 2.25  -24.99  -21.38  -19.81  -18.43  -16.23
#> log_p_z[33,3]  -24.91    0.08 4.40  -34.81  -27.46  -24.37  -21.84  -17.81
#> log_p_z[33,4]  -21.83    0.07 4.25  -31.69  -24.26  -21.38  -18.82  -14.86
#> log_p_z[34,1]  -19.61    0.03 2.35  -24.81  -21.01  -19.49  -17.97  -15.48
#> log_p_z[34,2]   -3.04    0.00 0.36   -3.76   -3.27   -3.03   -2.80   -2.38
#> log_p_z[34,3]  -21.11    0.08 4.29  -30.91  -23.65  -20.65  -18.01  -14.22
#> log_p_z[34,4]  -21.89    0.08 4.41  -32.03  -24.40  -21.28  -18.82  -15.07
#> log_p_z[35,1]  -22.57    0.04 2.65  -28.44  -24.18  -22.37  -20.72  -17.99
#> log_p_z[35,2]   -5.52    0.01 0.71   -7.06   -5.96   -5.47   -5.04   -4.28
#> log_p_z[35,3]  -16.00    0.07 4.00  -25.04  -18.39  -15.49  -13.07   -9.89
#> log_p_z[35,4]  -21.90    0.08 4.37  -31.61  -24.36  -21.35  -18.89  -14.82
#> log_p_z[36,1]  -29.76    0.06 3.67  -37.92  -31.97  -29.38  -27.22  -23.52
#> log_p_z[36,2]  -21.94    0.04 2.35  -26.85  -23.46  -21.81  -20.29  -17.85
#> log_p_z[36,3]  -10.53    0.03 1.71  -14.77  -11.35  -10.26   -9.35   -7.96
#> log_p_z[36,4]  -14.01    0.06 3.28  -21.23  -16.00  -13.68  -11.59   -8.83
#> log_p_z[37,1]  -18.55    0.03 2.30  -23.57  -19.91  -18.33  -16.97  -14.62
#> log_p_z[37,2]   -7.60    0.01 0.77   -9.28   -8.08   -7.54   -7.05   -6.27
#> log_p_z[37,3]  -19.09    0.06 3.23  -25.88  -21.12  -18.75  -16.72  -13.67
#> log_p_z[37,4]  -21.86    0.06 3.56  -30.03  -23.95  -21.54  -19.40  -16.03
#> log_p_z[38,1]  -25.21    0.04 2.76  -31.08  -26.99  -25.01  -23.21  -20.43
#> log_p_z[38,2]   -9.60    0.01 1.01  -11.82  -10.23   -9.52   -8.89   -7.84
#> log_p_z[38,3]  -11.66    0.05 2.72  -17.77  -13.25  -11.35   -9.64   -7.53
#> log_p_z[38,4]  -21.38    0.06 3.56  -29.17  -23.41  -21.03  -18.92  -15.44
#> log_p_z[39,1]  -29.34    0.06 4.29  -39.05  -31.84  -28.88  -26.30  -22.47
#> log_p_z[39,2]  -16.81    0.03 2.03  -21.23  -18.05  -16.65  -15.39  -13.36
#> log_p_z[39,3]   -7.27    0.02 1.12   -9.68   -7.96   -7.19   -6.49   -5.29
#> log_p_z[39,4]  -18.78    0.07 4.18  -28.26  -21.08  -18.27  -15.83  -12.22
#> log_p_z[40,1]   -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_p_z[40,2]  -20.02    0.04 2.25  -24.99  -21.38  -19.81  -18.43  -16.23
#> log_p_z[40,3]  -24.91    0.08 4.40  -34.81  -27.46  -24.37  -21.84  -17.81
#> log_p_z[40,4]  -21.83    0.07 4.25  -31.69  -24.26  -21.38  -18.82  -14.86
#> log_p_z[41,1]   -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_p_z[41,2]  -20.02    0.04 2.25  -24.99  -21.38  -19.81  -18.43  -16.23
#> log_p_z[41,3]  -24.91    0.08 4.40  -34.81  -27.46  -24.37  -21.84  -17.81
#> log_p_z[41,4]  -21.83    0.07 4.25  -31.69  -24.26  -21.38  -18.82  -14.86
#> log_p_z[42,1]   -6.00    0.01 0.61   -7.27   -6.39   -5.98   -5.57   -4.90
#> log_p_z[42,2]  -14.95    0.03 1.69  -18.67  -15.98  -14.78  -13.77  -12.10
#> log_p_z[42,3]  -24.29    0.07 3.62  -32.41  -26.49  -23.96  -21.70  -18.24
#> log_p_z[42,4]  -21.81    0.06 3.55  -29.67  -23.95  -21.49  -19.30  -15.83
#> log_p_z[43,1]  -21.88    0.04 2.45  -27.25  -23.33  -21.71  -20.18  -17.61
#> log_p_z[43,2]   -5.65    0.01 0.53   -6.79   -5.99   -5.60   -5.26   -4.73
#> log_p_z[43,3]  -18.31    0.06 3.20  -25.62  -20.22  -18.01  -16.03  -13.02
#> log_p_z[43,4]  -21.57    0.06 3.53  -29.41  -23.60  -21.19  -19.05  -15.92
#> log_p_z[44,1]   -4.40    0.00 0.40   -5.26   -4.65   -4.37   -4.11   -3.71
#> log_p_z[44,2]  -17.26    0.03 1.82  -21.22  -18.34  -17.12  -16.00  -14.14
#> log_p_z[44,3]  -24.31    0.07 3.55  -32.24  -26.46  -24.01  -21.75  -18.35
#> log_p_z[44,4]  -21.86    0.06 3.54  -29.95  -23.99  -21.51  -19.44  -15.90
#> log_p_z[45,1]  -19.61    0.03 2.35  -24.81  -21.01  -19.49  -17.97  -15.48
#> log_p_z[45,2]   -3.04    0.00 0.36   -3.76   -3.27   -3.03   -2.80   -2.38
#> log_p_z[45,3]  -21.11    0.08 4.29  -30.91  -23.65  -20.65  -18.01  -14.22
#> log_p_z[45,4]  -21.89    0.08 4.41  -32.03  -24.40  -21.28  -18.82  -15.07
#> log_lik[1]      -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_lik[2]      -7.60    0.02 1.11  -10.10   -8.28   -7.53   -6.83   -5.62
#> log_lik[3]      -9.93    0.01 0.91  -11.82  -10.50   -9.90   -9.32   -8.28
#> log_lik[4]      -7.58    0.02 0.93   -9.58   -8.16   -7.52   -6.93   -5.96
#> log_lik[5]      -3.72    0.00 0.43   -4.63   -3.99   -3.69   -3.43   -2.92
#> log_lik[6]      -5.48    0.01 0.67   -6.96   -5.90   -5.43   -5.01   -4.29
#> log_lik[7]      -6.00    0.01 0.61   -7.27   -6.39   -5.98   -5.57   -4.90
#> log_lik[8]      -5.73    0.01 0.84   -7.50   -6.25   -5.70   -5.15   -4.24
#> log_lik[9]      -4.51    0.01 0.57   -5.74   -4.85   -4.47   -4.11   -3.50
#> log_lik[10]     -7.11    0.01 0.71   -8.65   -7.56   -7.06   -6.63   -5.87
#> log_lik[11]     -5.82    0.01 0.95   -7.89   -6.43   -5.75   -5.16   -4.20
#> log_lik[12]     -9.34    0.01 1.07  -11.56  -10.02   -9.27   -8.62   -7.41
#> log_lik[13]     -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_lik[14]     -6.25    0.01 0.87   -8.17   -6.76   -6.18   -5.64   -4.76
#> log_lik[15]     -4.40    0.00 0.40   -5.26   -4.65   -4.37   -4.11   -3.71
#> log_lik[16]     -3.80    0.01 0.53   -4.92   -4.14   -3.76   -3.43   -2.88
#> log_lik[17]     -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_lik[18]     -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_lik[19]     -4.65    0.01 0.60   -5.98   -5.03   -4.61   -4.23   -3.57
#> log_lik[20]     -7.11    0.01 0.97   -9.20   -7.70   -7.03   -6.44   -5.46
#> log_lik[21]     -3.04    0.00 0.36   -3.76   -3.27   -3.03   -2.80   -2.38
#> log_lik[22]     -4.65    0.01 0.60   -5.98   -5.03   -4.61   -4.23   -3.57
#> log_lik[23]     -3.72    0.00 0.43   -4.63   -3.99   -3.69   -3.43   -2.92
#> log_lik[24]     -5.79    0.01 0.62   -7.20   -6.16   -5.72   -5.36   -4.78
#> log_lik[25]     -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_lik[26]     -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_lik[27]     -5.65    0.01 0.53   -6.79   -5.99   -5.60   -5.26   -4.73
#> log_lik[28]     -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_lik[29]     -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_lik[30]     -6.79    0.01 0.83   -8.65   -7.29   -6.71   -6.19   -5.44
#> log_lik[31]     -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_lik[32]     -6.07    0.01 0.95   -8.17   -6.63   -6.00   -5.40   -4.43
#> log_lik[33]     -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_lik[34]     -3.04    0.00 0.36   -3.76   -3.27   -3.03   -2.80   -2.38
#> log_lik[35]     -5.52    0.01 0.71   -7.05   -5.96   -5.47   -5.04   -4.28
#> log_lik[36]     -9.94    0.02 1.18  -12.44  -10.66   -9.87   -9.14   -7.87
#> log_lik[37]     -7.60    0.01 0.77   -9.28   -8.08   -7.54   -7.05   -6.27
#> log_lik[38]     -8.99    0.01 0.87  -10.73   -9.58   -8.98   -8.42   -7.35
#> log_lik[39]     -7.27    0.02 1.12   -9.67   -7.96   -7.19   -6.49   -5.29
#> log_lik[40]     -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_lik[41]     -2.20    0.00 0.29   -2.80   -2.39   -2.18   -2.00   -1.67
#> log_lik[42]     -6.00    0.01 0.61   -7.27   -6.39   -5.98   -5.57   -4.90
#> log_lik[43]     -5.65    0.01 0.53   -6.79   -5.99   -5.60   -5.26   -4.73
#> log_lik[44]     -4.40    0.00 0.40   -5.26   -4.65   -4.37   -4.11   -3.71
#> log_lik[45]     -3.04    0.00 0.36   -3.76   -3.27   -3.03   -2.80   -2.38
#> lp__          -463.69    0.17 6.15 -476.88 -467.69 -463.49 -459.25 -452.58
#>               n_eff Rhat
#> pi[1]          7772    1
#> pi[2]          6673    1
#> pi[3]          5841    1
#> pi[4]          5861    1
#> theta[1,1,1]   7760    1
#> theta[1,1,2]   6622    1
#> theta[1,1,3]   4994    1
#> theta[1,1,4]   4985    1
#> theta[1,2,1]   6913    1
#> theta[1,2,2]   8228    1
#> theta[1,2,3]   7709    1
#> theta[1,2,4]   5668    1
#> theta[1,3,1]   6323    1
#> theta[1,3,2]   3853    1
#> theta[1,3,3]   4474    1
#> theta[1,3,4]   5236    1
#> theta[1,4,1]   6319    1
#> theta[1,4,2]   5829    1
#> theta[1,4,3]   4230    1
#> theta[1,4,4]   4612    1
#> theta[2,1,1]   6972    1
#> theta[2,1,2]   6658    1
#> theta[2,1,3]   4801    1
#> theta[2,1,4]   4271    1
#> theta[2,2,1]   6077    1
#> theta[2,2,2]   7382    1
#> theta[2,2,3]   7701    1
#> theta[2,2,4]   5851    1
#> theta[2,3,1]   5478    1
#> theta[2,3,2]   7230    1
#> theta[2,3,3]   5820    1
#> theta[2,3,4]   5763    1
#> theta[2,4,1]   6791    1
#> theta[2,4,2]   7373    1
#> theta[2,4,3]   6235    1
#> theta[2,4,4]   6993    1
#> theta[3,1,1]   6857    1
#> theta[3,1,2]   4821    1
#> theta[3,1,3]   4712    1
#> theta[3,1,4]   5006    1
#> theta[3,2,1]   6727    1
#> theta[3,2,2]   6830    1
#> theta[3,2,3]   6418    1
#> theta[3,2,4]   5383    1
#> theta[3,3,1]   6954    1
#> theta[3,3,2]   6500    1
#> theta[3,3,3]   5982    1
#> theta[3,3,4]   6228    1
#> theta[3,4,1]   5823    1
#> theta[3,4,2]   6376    1
#> theta[3,4,3]   5578    1
#> theta[3,4,4]   5878    1
#> theta[4,1,1]   7717    1
#> theta[4,1,2]   5129    1
#> theta[4,1,3]   4632    1
#> theta[4,1,4]   4861    1
#> theta[4,2,1]   6443    1
#> theta[4,2,2]   5521    1
#> theta[4,2,3]   5706    1
#> theta[4,2,4]   5557    1
#> theta[4,3,1]   6455    1
#> theta[4,3,2]   5839    1
#> theta[4,3,3]   5367    1
#> theta[4,3,4]   5632    1
#> theta[4,4,1]   6445    1
#> theta[4,4,2]   5973    1
#> theta[4,4,3]   7061    1
#> theta[4,4,4]   6648    1
#> theta[5,1,1]   6966    1
#> theta[5,1,2]   5565    1
#> theta[5,1,3]   4415    1
#> theta[5,1,4]   5722    1
#> theta[5,2,1]   8153    1
#> theta[5,2,2]   6406    1
#> theta[5,2,3]   7051    1
#> theta[5,2,4]   5047    1
#> theta[5,3,1]   6656    1
#> theta[5,3,2]   6924    1
#> theta[5,3,3]   7033    1
#> theta[5,3,4]   6516    1
#> theta[5,4,1]   6179    1
#> theta[5,4,2]   5897    1
#> theta[5,4,3]   5084    1
#> theta[5,4,4]   6599    1
#> log_p_z[1,1]   7020    1
#> log_p_z[1,2]   3659    1
#> log_p_z[1,3]   2909    1
#> log_p_z[1,4]   3587    1
#> log_p_z[2,1]   4614    1
#> log_p_z[2,2]   4123    1
#> log_p_z[2,3]   3459    1
#> log_p_z[2,4]   3279    1
#> log_p_z[3,1]   4320    1
#> log_p_z[3,2]   3819    1
#> log_p_z[3,3]   2669    1
#> log_p_z[3,4]   3443    1
#> log_p_z[4,1]   4979    1
#> log_p_z[4,2]   3570    1
#> log_p_z[4,3]   2814    1
#> log_p_z[4,4]   3526    1
#> log_p_z[5,1]   4134    1
#> log_p_z[5,2]   7485    1
#> log_p_z[5,3]   2789    1
#> log_p_z[5,4]   3332    1
#> log_p_z[6,1]   4451    1
#> log_p_z[6,2]   5949    1
#> log_p_z[6,3]   2837    1
#> log_p_z[6,4]   3358    1
#> log_p_z[7,1]   6361    1
#> log_p_z[7,2]   3687    1
#> log_p_z[7,3]   2742    1
#> log_p_z[7,4]   3681    1
#> log_p_z[8,1]   4525    1
#> log_p_z[8,2]   4813    1
#> log_p_z[8,3]   4815    1
#> log_p_z[8,4]   3278    1
#> log_p_z[9,1]   4539    1
#> log_p_z[9,2]   5897    1
#> log_p_z[9,3]   2901    1
#> log_p_z[9,4]   3230    1
#> log_p_z[10,1]  4666    1
#> log_p_z[10,2]  6108    1
#> log_p_z[10,3]  2842    1
#> log_p_z[10,4]  3084    1
#> log_p_z[11,1]  4168    1
#> log_p_z[11,2]  4067    1
#> log_p_z[11,3]  2570    1
#> log_p_z[11,4]  5130    1
#> log_p_z[12,1]  4746    1
#> log_p_z[12,2]  3726    1
#> log_p_z[12,3]  3070    1
#> log_p_z[12,4]  3203    1
#> log_p_z[13,1]  7020    1
#> log_p_z[13,2]  3659    1
#> log_p_z[13,3]  2909    1
#> log_p_z[13,4]  3587    1
#> log_p_z[14,1]  3989    1
#> log_p_z[14,2]  4263    1
#> log_p_z[14,3]  2957    1
#> log_p_z[14,4]  3335    1
#> log_p_z[15,1]  6893    1
#> log_p_z[15,2]  3429    1
#> log_p_z[15,3]  2625    1
#> log_p_z[15,4]  3441    1
#> log_p_z[16,1]  5756    1
#> log_p_z[16,2]  3894    1
#> log_p_z[16,3]  3033    1
#> log_p_z[16,4]  3678    1
#> log_p_z[17,1]  7020    1
#> log_p_z[17,2]  3659    1
#> log_p_z[17,3]  2909    1
#> log_p_z[17,4]  3587    1
#> log_p_z[18,1]  7020    1
#> log_p_z[18,2]  3659    1
#> log_p_z[18,3]  2909    1
#> log_p_z[18,4]  3587    1
#> log_p_z[19,1]  5375    1
#> log_p_z[19,2]  6695    1
#> log_p_z[19,3]  2745    1
#> log_p_z[19,4]  3406    1
#> log_p_z[20,1]  4495    1
#> log_p_z[20,2]  4076    1
#> log_p_z[20,3]  2840    1
#> log_p_z[20,4]  3392    1
#> log_p_z[21,1]  4622    1
#> log_p_z[21,2]  6840    1
#> log_p_z[21,3]  2880    1
#> log_p_z[21,4]  3317    1
#> log_p_z[22,1]  5375    1
#> log_p_z[22,2]  6695    1
#> log_p_z[22,3]  2745    1
#> log_p_z[22,4]  3406    1
#> log_p_z[23,1]  4134    1
#> log_p_z[23,2]  7485    1
#> log_p_z[23,3]  2789    1
#> log_p_z[23,4]  3332    1
#> log_p_z[24,1]  4442    1
#> log_p_z[24,2]  4840    1
#> log_p_z[24,3]  2589    1
#> log_p_z[24,4]  3254    1
#> log_p_z[25,1]  7020    1
#> log_p_z[25,2]  3659    1
#> log_p_z[25,3]  2909    1
#> log_p_z[25,4]  3587    1
#> log_p_z[26,1]  7020    1
#> log_p_z[26,2]  3659    1
#> log_p_z[26,3]  2909    1
#> log_p_z[26,4]  3587    1
#> log_p_z[27,1]  4639    1
#> log_p_z[27,2]  6352    1
#> log_p_z[27,3]  2826    1
#> log_p_z[27,4]  3209    1
#> log_p_z[28,1]  7020    1
#> log_p_z[28,2]  3659    1
#> log_p_z[28,3]  2909    1
#> log_p_z[28,4]  3587    1
#> log_p_z[29,1]  7020    1
#> log_p_z[29,2]  3659    1
#> log_p_z[29,3]  2909    1
#> log_p_z[29,4]  3587    1
#> log_p_z[30,1]  5612    1
#> log_p_z[30,2]  3643    1
#> log_p_z[30,3]  2479    1
#> log_p_z[30,4]  3158    1
#> log_p_z[31,1]  7020    1
#> log_p_z[31,2]  3659    1
#> log_p_z[31,3]  2909    1
#> log_p_z[31,4]  3587    1
#> log_p_z[32,1]  4833    1
#> log_p_z[32,2]  5872    1
#> log_p_z[32,3]  4469    1
#> log_p_z[32,4]  3239    1
#> log_p_z[33,1]  7020    1
#> log_p_z[33,2]  3659    1
#> log_p_z[33,3]  2909    1
#> log_p_z[33,4]  3587    1
#> log_p_z[34,1]  4622    1
#> log_p_z[34,2]  6840    1
#> log_p_z[34,3]  2880    1
#> log_p_z[34,4]  3317    1
#> log_p_z[35,1]  4136    1
#> log_p_z[35,2]  6080    1
#> log_p_z[35,3]  2928    1
#> log_p_z[35,4]  3323    1
#> log_p_z[36,1]  3799    1
#> log_p_z[36,2]  3150    1
#> log_p_z[36,3]  2480    1
#> log_p_z[36,4]  2727    1
#> log_p_z[37,1]  4387    1
#> log_p_z[37,2]  5225    1
#> log_p_z[37,3]  2769    1
#> log_p_z[37,4]  3045    1
#> log_p_z[38,1]  4270    1
#> log_p_z[38,2]  4881    1
#> log_p_z[38,3]  2937    1
#> log_p_z[38,4]  3160    1
#> log_p_z[39,1]  4401    1
#> log_p_z[39,2]  4973    1
#> log_p_z[39,3]  4735    1
#> log_p_z[39,4]  3565    1
#> log_p_z[40,1]  7020    1
#> log_p_z[40,2]  3659    1
#> log_p_z[40,3]  2909    1
#> log_p_z[40,4]  3587    1
#> log_p_z[41,1]  7020    1
#> log_p_z[41,2]  3659    1
#> log_p_z[41,3]  2909    1
#> log_p_z[41,4]  3587    1
#> log_p_z[42,1]  6361    1
#> log_p_z[42,2]  3687    1
#> log_p_z[42,3]  2742    1
#> log_p_z[42,4]  3681    1
#> log_p_z[43,1]  4639    1
#> log_p_z[43,2]  6352    1
#> log_p_z[43,3]  2826    1
#> log_p_z[43,4]  3209    1
#> log_p_z[44,1]  6893    1
#> log_p_z[44,2]  3429    1
#> log_p_z[44,3]  2625    1
#> log_p_z[44,4]  3441    1
#> log_p_z[45,1]  4622    1
#> log_p_z[45,2]  6840    1
#> log_p_z[45,3]  2880    1
#> log_p_z[45,4]  3317    1
#> log_lik[1]     7020    1
#> log_lik[2]     4226    1
#> log_lik[3]     5550    1
#> log_lik[4]     3597    1
#> log_lik[5]     7485    1
#> log_lik[6]     5966    1
#> log_lik[7]     6361    1
#> log_lik[8]     4816    1
#> log_lik[9]     5898    1
#> log_lik[10]    6130    1
#> log_lik[11]    5130    1
#> log_lik[12]    5442    1
#> log_lik[13]    7020    1
#> log_lik[14]    4271    1
#> log_lik[15]    6893    1
#> log_lik[16]    5756    1
#> log_lik[17]    7020    1
#> log_lik[18]    7020    1
#> log_lik[19]    6695    1
#> log_lik[20]    4311    1
#> log_lik[21]    6839    1
#> log_lik[22]    6695    1
#> log_lik[23]    7485    1
#> log_lik[24]    4841    1
#> log_lik[25]    7020    1
#> log_lik[26]    7020    1
#> log_lik[27]    6353    1
#> log_lik[28]    7020    1
#> log_lik[29]    7020    1
#> log_lik[30]    5628    1
#> log_lik[31]    7020    1
#> log_lik[32]    4473    1
#> log_lik[33]    7020    1
#> log_lik[34]    6839    1
#> log_lik[35]    6078    1
#> log_lik[36]    5242    1
#> log_lik[37]    5231    1
#> log_lik[38]    4228    1
#> log_lik[39]    4764    1
#> log_lik[40]    7020    1
#> log_lik[41]    7020    1
#> log_lik[42]    6361    1
#> log_lik[43]    6353    1
#> log_lik[44]    6893    1
#> log_lik[45]    6839    1
#> lp__           1351    1
#> 
#> Samples were drawn using NUTS(diag_e) at Tue Mar  7 09:57:56 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }