Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.25    0.33    0.37    0.42    0.50
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.24
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.09    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.12    0.18
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.02    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.59    0.73    0.81    0.87    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.03    0.06    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.09    0.00    0.03    0.07    0.13    0.33
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.18    0.39
#> theta[1,4,4]     0.69    0.00 0.14    0.40    0.59    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.69    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.06    0.11    0.16    0.21    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.39    0.52    0.58    0.65    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.25    0.30    0.36    0.48
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[2,3,3]     0.62    0.00 0.13    0.36    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.28    0.47
#> theta[2,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.39
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.74    0.84    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[3,3,2]     0.18    0.00 0.11    0.02    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.14    0.25    0.41    0.51    0.60    0.76
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.30    0.48
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[3,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.35
#> theta[3,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.43
#> theta[3,4,4]     0.63    0.00 0.15    0.32    0.53    0.64    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.08    0.65    0.78    0.84    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.24
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.09    0.55    0.67    0.73    0.79    0.88
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.08    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.38
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[4,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.76    0.91
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.82    0.87    0.91    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.06    0.05    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.33
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.40
#> theta[5,3,3]     0.60    0.00 0.13    0.33    0.52    0.61    0.70    0.85
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.42
#> theta[5,4,4]     0.63    0.00 0.15    0.32    0.52    0.64    0.74    0.90
#> log_p_z[1,1]    -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_p_z[1,2]   -20.04    0.04 2.30  -25.19  -21.47  -19.80  -18.43  -16.14
#> log_p_z[1,3]   -24.76    0.07 4.31  -34.27  -27.26  -24.29  -21.69  -17.72
#> log_p_z[1,4]   -21.76    0.07 4.23  -31.21  -24.22  -21.23  -18.76  -14.85
#> log_p_z[2,1]   -29.74    0.06 4.49  -39.84  -32.39  -29.20  -26.54  -22.57
#> log_p_z[2,2]   -21.01    0.03 2.36  -26.18  -22.42  -20.84  -19.35  -17.00
#> log_p_z[2,3]    -7.63    0.02 1.18  -10.28   -8.34   -7.51   -6.81   -5.66
#> log_p_z[2,4]   -16.46    0.08 4.30  -26.84  -18.84  -15.90  -13.35   -9.96
#> log_p_z[3,1]   -11.62    0.02 1.64  -15.21  -12.59  -11.49  -10.41   -8.91
#> log_p_z[3,2]   -10.82    0.02 1.49  -14.20  -11.70  -10.65   -9.76   -8.39
#> log_p_z[3,3]   -23.25    0.06 3.46  -30.95  -25.37  -22.99  -20.74  -17.43
#> log_p_z[3,4]   -21.83    0.06 3.52  -29.46  -24.04  -21.47  -19.35  -15.85
#> log_p_z[4,1]   -14.64    0.02 1.83  -18.77  -15.74  -14.52  -13.37  -11.43
#> log_p_z[4,2]    -7.58    0.01 0.94   -9.63   -8.17   -7.50   -6.89   -5.95
#> log_p_z[4,3]   -20.79    0.07 4.46  -31.21  -23.42  -20.19  -17.58  -13.67
#> log_p_z[4,4]   -21.99    0.08 4.58  -32.42  -24.62  -21.38  -18.76  -14.90
#> log_p_z[5,1]   -21.45    0.04 2.59  -26.97  -23.06  -21.24  -19.67  -16.87
#> log_p_z[5,2]    -3.71    0.01 0.43   -4.61   -3.98   -3.69   -3.42   -2.94
#> log_p_z[5,3]   -18.76    0.07 4.37  -29.00  -21.23  -18.16  -15.66  -12.15
#> log_p_z[5,4]   -21.98    0.08 4.44  -32.39  -24.37  -21.44  -18.88  -14.93
#> log_p_z[6,1]   -21.43    0.04 2.53  -26.85  -23.01  -21.25  -19.64  -16.98
#> log_p_z[6,2]    -5.46    0.01 0.64   -6.86   -5.86   -5.43   -5.00   -4.35
#> log_p_z[6,3]   -17.53    0.07 4.21  -27.65  -19.84  -16.86  -14.57  -11.12
#> log_p_z[6,4]   -21.80    0.08 4.40  -31.98  -24.18  -21.30  -18.77  -14.62
#> log_p_z[7,1]    -5.99    0.01 0.59   -7.27   -6.36   -5.95   -5.57   -4.96
#> log_p_z[7,2]   -14.96    0.03 1.70  -18.72  -16.01  -14.79  -13.77  -12.08
#> log_p_z[7,3]   -24.25    0.06 3.58  -32.24  -26.48  -23.94  -21.70  -18.23
#> log_p_z[7,4]   -21.82    0.06 3.58  -29.79  -23.93  -21.53  -19.37  -15.64
#> log_p_z[8,1]   -29.79    0.06 4.47  -39.90  -32.42  -29.33  -26.68  -22.41
#> log_p_z[8,2]   -18.28    0.03 2.06  -22.79  -19.58  -18.14  -16.82  -14.71
#> log_p_z[8,3]    -5.72    0.01 0.81   -7.39   -6.24   -5.66   -5.15   -4.28
#> log_p_z[8,4]   -18.62    0.08 4.39  -29.12  -21.05  -18.08  -15.48  -11.78
#> log_p_z[9,1]   -19.99    0.03 2.33  -25.09  -21.47  -19.82  -18.30  -16.01
#> log_p_z[9,2]    -4.52    0.01 0.57   -5.80   -4.87   -4.48   -4.13   -3.49
#> log_p_z[9,3]   -19.66    0.07 4.44  -30.00  -22.20  -19.06  -16.47  -12.92
#> log_p_z[9,4]   -21.78    0.08 4.44  -31.73  -24.22  -21.23  -18.68  -14.87
#> log_p_z[10,1]  -22.26    0.03 2.44  -27.67  -23.74  -22.09  -20.54  -18.09
#> log_p_z[10,2]   -7.13    0.01 0.71   -8.66   -7.56   -7.08   -6.62   -5.88
#> log_p_z[10,3]  -16.81    0.06 3.25  -24.27  -18.71  -16.44  -14.46  -11.67
#> log_p_z[10,4]  -21.45    0.06 3.58  -29.49  -23.55  -21.07  -18.95  -15.54
#> log_p_z[11,1]  -29.90    0.06 4.49  -40.24  -32.50  -29.45  -26.68  -22.56
#> log_p_z[11,2]  -29.28    0.07 4.27  -38.96  -31.60  -28.83  -26.27  -22.24
#> log_p_z[11,3]  -18.24    0.06 3.41  -26.17  -20.02  -17.73  -15.91  -12.99
#> log_p_z[11,4]   -5.82    0.01 0.98   -7.87   -6.45   -5.75   -5.14   -4.12
#> log_p_z[12,1]  -22.96    0.04 2.67  -28.75  -24.60  -22.73  -21.04  -18.39
#> log_p_z[12,2]   -9.58    0.02 1.28  -12.53  -10.26   -9.42   -8.70   -7.49
#> log_p_z[12,3]  -15.21    0.07 4.14  -25.08  -17.48  -14.58  -12.24   -9.01
#> log_p_z[12,4]  -19.40    0.08 4.29  -29.32  -21.78  -18.87  -16.45  -12.61
#> log_p_z[13,1]   -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_p_z[13,2]  -20.04    0.04 2.30  -25.19  -21.47  -19.80  -18.43  -16.14
#> log_p_z[13,3]  -24.76    0.07 4.31  -34.27  -27.26  -24.29  -21.69  -17.72
#> log_p_z[13,4]  -21.76    0.07 4.23  -31.21  -24.22  -21.23  -18.76  -14.85
#> log_p_z[14,1]  -19.04    0.04 2.42  -24.32  -20.50  -18.83  -17.40  -14.71
#> log_p_z[14,2]   -6.25    0.01 0.88   -8.27   -6.78   -6.18   -5.63   -4.73
#> log_p_z[14,3]  -18.81    0.07 4.39  -29.47  -21.27  -18.21  -15.66  -12.11
#> log_p_z[14,4]  -21.95    0.08 4.41  -32.29  -24.42  -21.41  -18.94  -14.86
#> log_p_z[15,1]   -4.41    0.00 0.40   -5.27   -4.68   -4.39   -4.12   -3.70
#> log_p_z[15,2]  -17.28    0.03 1.85  -21.52  -18.42  -17.13  -15.98  -14.04
#> log_p_z[15,3]  -24.28    0.06 3.59  -32.33  -26.43  -24.04  -21.75  -18.27
#> log_p_z[15,4]  -21.83    0.06 3.55  -29.70  -24.10  -21.48  -19.32  -15.75
#> log_p_z[16,1]   -3.79    0.01 0.52   -4.91   -4.13   -3.75   -3.42   -2.85
#> log_p_z[16,2]  -17.72    0.04 2.19  -22.54  -19.00  -17.47  -16.18  -14.09
#> log_p_z[16,3]  -24.73    0.07 4.30  -34.35  -27.26  -24.23  -21.66  -17.71
#> log_p_z[16,4]  -21.74    0.07 4.27  -31.28  -24.13  -21.23  -18.83  -14.72
#> log_p_z[17,1]   -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_p_z[17,2]  -20.04    0.04 2.30  -25.19  -21.47  -19.80  -18.43  -16.14
#> log_p_z[17,3]  -24.76    0.07 4.31  -34.27  -27.26  -24.29  -21.69  -17.72
#> log_p_z[17,4]  -21.76    0.07 4.23  -31.21  -24.22  -21.23  -18.76  -14.85
#> log_p_z[18,1]   -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_p_z[18,2]  -20.04    0.04 2.30  -25.19  -21.47  -19.80  -18.43  -16.14
#> log_p_z[18,3]  -24.76    0.07 4.31  -34.27  -27.26  -24.29  -21.69  -17.72
#> log_p_z[18,4]  -21.76    0.07 4.23  -31.21  -24.22  -21.23  -18.76  -14.85
#> log_p_z[19,1]  -16.39    0.03 1.93  -20.64  -17.57  -16.23  -15.03  -13.02
#> log_p_z[19,2]   -4.65    0.01 0.58   -5.89   -5.01   -4.61   -4.25   -3.62
#> log_p_z[19,3]  -22.15    0.08 4.55  -32.57  -24.81  -21.58  -18.89  -15.02
#> log_p_z[19,4]  -22.00    0.08 4.55  -32.75  -24.58  -21.51  -18.76  -14.78
#> log_p_z[20,1]  -18.01    0.03 2.21  -22.68  -19.42  -17.86  -16.47  -14.21
#> log_p_z[20,2]   -7.11    0.01 0.98   -9.28   -7.71   -7.04   -6.44   -5.43
#> log_p_z[20,3]  -20.02    0.07 4.33  -30.27  -22.40  -19.41  -17.02  -13.29
#> log_p_z[20,4]  -21.82    0.08 4.39  -32.03  -24.27  -21.23  -18.79  -14.90
#> log_p_z[21,1]  -19.61    0.04 2.38  -24.71  -21.04  -19.47  -17.95  -15.48
#> log_p_z[21,2]   -3.04    0.00 0.35   -3.77   -3.27   -3.03   -2.79   -2.41
#> log_p_z[21,3]  -21.21    0.08 4.47  -31.47  -23.79  -20.60  -18.02  -14.26
#> log_p_z[21,4]  -21.98    0.08 4.43  -32.13  -24.37  -21.39  -18.88  -14.96
#> log_p_z[22,1]  -16.39    0.03 1.93  -20.64  -17.57  -16.23  -15.03  -13.02
#> log_p_z[22,2]   -4.65    0.01 0.58   -5.89   -5.01   -4.61   -4.25   -3.62
#> log_p_z[22,3]  -22.15    0.08 4.55  -32.57  -24.81  -21.58  -18.89  -15.02
#> log_p_z[22,4]  -22.00    0.08 4.55  -32.75  -24.58  -21.51  -18.76  -14.78
#> log_p_z[23,1]  -21.45    0.04 2.59  -26.97  -23.06  -21.24  -19.67  -16.87
#> log_p_z[23,2]   -3.71    0.01 0.43   -4.61   -3.98   -3.69   -3.42   -2.94
#> log_p_z[23,3]  -18.76    0.07 4.37  -29.00  -21.23  -18.16  -15.66  -12.15
#> log_p_z[23,4]  -21.98    0.08 4.44  -32.39  -24.37  -21.44  -18.88  -14.93
#> log_p_z[24,1]  -17.41    0.03 2.18  -22.18  -18.72  -17.26  -15.87  -13.71
#> log_p_z[24,2]   -5.80    0.01 0.65   -7.24   -6.17   -5.73   -5.36   -4.73
#> log_p_z[24,3]  -21.69    0.06 3.51  -29.55  -23.76  -21.35  -19.19  -15.89
#> log_p_z[24,4]  -21.91    0.06 3.57  -29.79  -24.08  -21.54  -19.34  -15.94
#> log_p_z[25,1]   -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_p_z[25,2]  -20.04    0.04 2.30  -25.19  -21.47  -19.80  -18.43  -16.14
#> log_p_z[25,3]  -24.76    0.07 4.31  -34.27  -27.26  -24.29  -21.69  -17.72
#> log_p_z[25,4]  -21.76    0.07 4.23  -31.21  -24.22  -21.23  -18.76  -14.85
#> log_p_z[26,1]   -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_p_z[26,2]  -20.04    0.04 2.30  -25.19  -21.47  -19.80  -18.43  -16.14
#> log_p_z[26,3]  -24.76    0.07 4.31  -34.27  -27.26  -24.29  -21.69  -17.72
#> log_p_z[26,4]  -21.76    0.07 4.23  -31.21  -24.22  -21.23  -18.76  -14.85
#> log_p_z[27,1]  -21.88    0.04 2.52  -27.54  -23.41  -21.67  -20.11  -17.65
#> log_p_z[27,2]   -5.65    0.01 0.52   -6.78   -5.97   -5.62   -5.29   -4.75
#> log_p_z[27,3]  -18.37    0.06 3.30  -25.78  -20.28  -17.93  -16.03  -13.05
#> log_p_z[27,4]  -21.65    0.06 3.52  -29.70  -23.76  -21.23  -19.15  -15.85
#> log_p_z[28,1]   -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_p_z[28,2]  -20.04    0.04 2.30  -25.19  -21.47  -19.80  -18.43  -16.14
#> log_p_z[28,3]  -24.76    0.07 4.31  -34.27  -27.26  -24.29  -21.69  -17.72
#> log_p_z[28,4]  -21.76    0.07 4.23  -31.21  -24.22  -21.23  -18.76  -14.85
#> log_p_z[29,1]   -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_p_z[29,2]  -20.04    0.04 2.30  -25.19  -21.47  -19.80  -18.43  -16.14
#> log_p_z[29,3]  -24.76    0.07 4.31  -34.27  -27.26  -24.29  -21.69  -17.72
#> log_p_z[29,4]  -21.76    0.07 4.23  -31.21  -24.22  -21.23  -18.76  -14.85
#> log_p_z[30,1]   -6.82    0.01 0.86   -8.78   -7.30   -6.75   -6.22   -5.41
#> log_p_z[30,2]  -14.74    0.03 1.71  -18.52  -15.78  -14.56  -13.56  -11.84
#> log_p_z[30,3]  -24.22    0.06 3.60  -32.10  -26.43  -23.90  -21.66  -18.09
#> log_p_z[30,4]  -21.87    0.06 3.55  -29.69  -24.05  -21.51  -19.38  -15.88
#> log_p_z[31,1]   -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_p_z[31,2]  -20.04    0.04 2.30  -25.19  -21.47  -19.80  -18.43  -16.14
#> log_p_z[31,3]  -24.76    0.07 4.31  -34.27  -27.26  -24.29  -21.69  -17.72
#> log_p_z[31,4]  -21.76    0.07 4.23  -31.21  -24.22  -21.23  -18.76  -14.85
#> log_p_z[32,1]  -29.78    0.06 4.47  -39.95  -32.43  -29.30  -26.62  -22.32
#> log_p_z[32,2]  -14.83    0.02 1.68  -18.38  -15.87  -14.71  -13.66  -11.86
#> log_p_z[32,3]   -6.08    0.01 0.97   -8.32   -6.66   -6.00   -5.41   -4.40
#> log_p_z[32,4]  -20.74    0.08 4.52  -31.60  -23.24  -20.20  -17.59  -13.50
#> log_p_z[33,1]   -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_p_z[33,2]  -20.04    0.04 2.30  -25.19  -21.47  -19.80  -18.43  -16.14
#> log_p_z[33,3]  -24.76    0.07 4.31  -34.27  -27.26  -24.29  -21.69  -17.72
#> log_p_z[33,4]  -21.76    0.07 4.23  -31.21  -24.22  -21.23  -18.76  -14.85
#> log_p_z[34,1]  -19.61    0.04 2.38  -24.71  -21.04  -19.47  -17.95  -15.48
#> log_p_z[34,2]   -3.04    0.00 0.35   -3.77   -3.27   -3.03   -2.79   -2.41
#> log_p_z[34,3]  -21.21    0.08 4.47  -31.47  -23.79  -20.60  -18.02  -14.26
#> log_p_z[34,4]  -21.98    0.08 4.43  -32.13  -24.37  -21.39  -18.88  -14.96
#> log_p_z[35,1]  -22.60    0.04 2.73  -28.51  -24.28  -22.35  -20.73  -17.76
#> log_p_z[35,2]   -5.53    0.01 0.71   -7.06   -5.97   -5.47   -5.06   -4.27
#> log_p_z[35,3]  -16.16    0.07 4.22  -26.08  -18.49  -15.53  -13.07   -9.97
#> log_p_z[35,4]  -21.94    0.08 4.46  -32.30  -24.47  -21.37  -18.85  -14.77
#> log_p_z[36,1]  -29.81    0.05 3.64  -37.68  -32.14  -29.48  -27.22  -23.60
#> log_p_z[36,2]  -21.86    0.04 2.28  -26.81  -23.23  -21.64  -20.30  -18.06
#> log_p_z[36,3]  -10.53    0.03 1.73  -14.76  -11.41  -10.26   -9.32   -7.93
#> log_p_z[36,4]  -14.05    0.06 3.28  -21.56  -16.05  -13.81  -11.66   -8.71
#> log_p_z[37,1]  -18.56    0.04 2.36  -23.80  -19.99  -18.35  -16.96  -14.42
#> log_p_z[37,2]   -7.62    0.01 0.82   -9.45   -8.12   -7.54   -7.05   -6.19
#> log_p_z[37,3]  -19.09    0.05 3.29  -26.56  -21.01  -18.69  -16.71  -13.75
#> log_p_z[37,4]  -21.87    0.06 3.58  -29.93  -24.13  -21.57  -19.30  -15.85
#> log_p_z[38,1]  -25.24    0.04 2.86  -31.48  -27.14  -24.99  -23.24  -20.16
#> log_p_z[38,2]   -9.61    0.01 0.99  -11.74  -10.22   -9.54   -8.92   -7.91
#> log_p_z[38,3]  -11.77    0.05 2.86  -18.59  -13.36  -11.37   -9.66   -7.55
#> log_p_z[38,4]  -21.41    0.06 3.65  -29.59  -23.61  -21.00  -18.86  -15.36
#> log_p_z[39,1]  -29.42    0.06 4.51  -39.58  -32.12  -28.87  -26.23  -22.11
#> log_p_z[39,2]  -16.81    0.03 1.97  -21.15  -18.00  -16.67  -15.43  -13.33
#> log_p_z[39,3]   -7.27    0.02 1.11   -9.59   -7.95   -7.17   -6.50   -5.34
#> log_p_z[39,4]  -18.82    0.08 4.26  -29.19  -21.07  -18.24  -15.80  -12.25
#> log_p_z[40,1]   -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_p_z[40,2]  -20.04    0.04 2.30  -25.19  -21.47  -19.80  -18.43  -16.14
#> log_p_z[40,3]  -24.76    0.07 4.31  -34.27  -27.26  -24.29  -21.69  -17.72
#> log_p_z[40,4]  -21.76    0.07 4.23  -31.21  -24.22  -21.23  -18.76  -14.85
#> log_p_z[41,1]   -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_p_z[41,2]  -20.04    0.04 2.30  -25.19  -21.47  -19.80  -18.43  -16.14
#> log_p_z[41,3]  -24.76    0.07 4.31  -34.27  -27.26  -24.29  -21.69  -17.72
#> log_p_z[41,4]  -21.76    0.07 4.23  -31.21  -24.22  -21.23  -18.76  -14.85
#> log_p_z[42,1]   -5.99    0.01 0.59   -7.27   -6.36   -5.95   -5.57   -4.96
#> log_p_z[42,2]  -14.96    0.03 1.70  -18.72  -16.01  -14.79  -13.77  -12.08
#> log_p_z[42,3]  -24.25    0.06 3.58  -32.24  -26.48  -23.94  -21.70  -18.23
#> log_p_z[42,4]  -21.82    0.06 3.58  -29.79  -23.93  -21.53  -19.37  -15.64
#> log_p_z[43,1]  -21.88    0.04 2.52  -27.54  -23.41  -21.67  -20.11  -17.65
#> log_p_z[43,2]   -5.65    0.01 0.52   -6.78   -5.97   -5.62   -5.29   -4.75
#> log_p_z[43,3]  -18.37    0.06 3.30  -25.78  -20.28  -17.93  -16.03  -13.05
#> log_p_z[43,4]  -21.65    0.06 3.52  -29.70  -23.76  -21.23  -19.15  -15.85
#> log_p_z[44,1]   -4.41    0.00 0.40   -5.27   -4.68   -4.39   -4.12   -3.70
#> log_p_z[44,2]  -17.28    0.03 1.85  -21.52  -18.42  -17.13  -15.98  -14.04
#> log_p_z[44,3]  -24.28    0.06 3.59  -32.33  -26.43  -24.04  -21.75  -18.27
#> log_p_z[44,4]  -21.83    0.06 3.55  -29.70  -24.10  -21.48  -19.32  -15.75
#> log_p_z[45,1]  -19.61    0.04 2.38  -24.71  -21.04  -19.47  -17.95  -15.48
#> log_p_z[45,2]   -3.04    0.00 0.35   -3.77   -3.27   -3.03   -2.79   -2.41
#> log_p_z[45,3]  -21.21    0.08 4.47  -31.47  -23.79  -20.60  -18.02  -14.26
#> log_p_z[45,4]  -21.98    0.08 4.43  -32.13  -24.37  -21.39  -18.88  -14.96
#> log_lik[1]      -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_lik[2]      -7.59    0.02 1.13  -10.04   -8.31   -7.50   -6.80   -5.66
#> log_lik[3]      -9.94    0.01 0.92  -11.85  -10.51   -9.89   -9.31   -8.24
#> log_lik[4]      -7.57    0.01 0.93   -9.61   -8.16   -7.50   -6.89   -5.95
#> log_lik[5]      -3.71    0.01 0.43   -4.61   -3.98   -3.69   -3.42   -2.94
#> log_lik[6]      -5.46    0.01 0.64   -6.86   -5.86   -5.43   -5.00   -4.35
#> log_lik[7]      -5.99    0.01 0.59   -7.27   -6.36   -5.95   -5.57   -4.96
#> log_lik[8]      -5.72    0.01 0.81   -7.39   -6.24   -5.66   -5.15   -4.28
#> log_lik[9]      -4.52    0.01 0.57   -5.80   -4.87   -4.48   -4.13   -3.49
#> log_lik[10]     -7.12    0.01 0.70   -8.65   -7.56   -7.08   -6.62   -5.88
#> log_lik[11]     -5.82    0.01 0.98   -7.87   -6.45   -5.75   -5.14   -4.12
#> log_lik[12]     -9.32    0.01 1.01  -11.52   -9.96   -9.27   -8.62   -7.45
#> log_lik[13]     -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_lik[14]     -6.25    0.01 0.88   -8.26   -6.78   -6.18   -5.63   -4.73
#> log_lik[15]     -4.41    0.00 0.40   -5.27   -4.68   -4.39   -4.12   -3.70
#> log_lik[16]     -3.79    0.01 0.52   -4.91   -4.13   -3.75   -3.42   -2.85
#> log_lik[17]     -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_lik[18]     -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_lik[19]     -4.65    0.01 0.58   -5.89   -5.01   -4.61   -4.25   -3.62
#> log_lik[20]     -7.11    0.01 0.98   -9.28   -7.71   -7.04   -6.44   -5.43
#> log_lik[21]     -3.04    0.00 0.35   -3.77   -3.27   -3.03   -2.79   -2.41
#> log_lik[22]     -4.65    0.01 0.58   -5.89   -5.01   -4.61   -4.25   -3.62
#> log_lik[23]     -3.71    0.01 0.43   -4.61   -3.98   -3.69   -3.42   -2.94
#> log_lik[24]     -5.80    0.01 0.65   -7.24   -6.17   -5.73   -5.36   -4.73
#> log_lik[25]     -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_lik[26]     -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_lik[27]     -5.65    0.01 0.52   -6.78   -5.97   -5.62   -5.29   -4.75
#> log_lik[28]     -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_lik[29]     -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_lik[30]     -6.82    0.01 0.85   -8.77   -7.30   -6.75   -6.21   -5.41
#> log_lik[31]     -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_lik[32]     -6.08    0.01 0.96   -8.31   -6.66   -6.00   -5.41   -4.40
#> log_lik[33]     -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_lik[34]     -3.04    0.00 0.35   -3.77   -3.27   -3.03   -2.79   -2.41
#> log_lik[35]     -5.53    0.01 0.70   -7.04   -5.96   -5.47   -5.06   -4.27
#> log_lik[36]     -9.93    0.02 1.19  -12.40  -10.71   -9.86   -9.10   -7.82
#> log_lik[37]     -7.62    0.01 0.82   -9.45   -8.12   -7.54   -7.05   -6.19
#> log_lik[38]     -9.01    0.01 0.85  -10.72   -9.58   -9.00   -8.43   -7.37
#> log_lik[39]     -7.27    0.02 1.10   -9.57   -7.95   -7.17   -6.50   -5.34
#> log_lik[40]     -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_lik[41]     -2.21    0.00 0.29   -2.80   -2.39   -2.20   -2.01   -1.67
#> log_lik[42]     -5.99    0.01 0.59   -7.27   -6.36   -5.95   -5.57   -4.96
#> log_lik[43]     -5.65    0.01 0.52   -6.78   -5.97   -5.62   -5.29   -4.75
#> log_lik[44]     -4.41    0.00 0.40   -5.27   -4.68   -4.39   -4.12   -3.70
#> log_lik[45]     -3.04    0.00 0.35   -3.77   -3.27   -3.03   -2.79   -2.41
#> lp__          -463.87    0.16 6.12 -476.74 -467.68 -463.58 -459.60 -453.01
#>               n_eff Rhat
#> pi[1]          9635    1
#> pi[2]          8234    1
#> pi[3]          6921    1
#> pi[4]          6186    1
#> theta[1,1,1]   7743    1
#> theta[1,1,2]   8005    1
#> theta[1,1,3]   5192    1
#> theta[1,1,4]   4817    1
#> theta[1,2,1]   6453    1
#> theta[1,2,2]   7023    1
#> theta[1,2,3]   7822    1
#> theta[1,2,4]   6179    1
#> theta[1,3,1]   5791    1
#> theta[1,3,2]   4374    1
#> theta[1,3,3]   4817    1
#> theta[1,3,4]   5132    1
#> theta[1,4,1]   6027    1
#> theta[1,4,2]   6175    1
#> theta[1,4,3]   4391    1
#> theta[1,4,4]   5253    1
#> theta[2,1,1]   8422    1
#> theta[2,1,2]   7126    1
#> theta[2,1,3]   5069    1
#> theta[2,1,4]   4424    1
#> theta[2,2,1]   6884    1
#> theta[2,2,2]   7149    1
#> theta[2,2,3]   6734    1
#> theta[2,2,4]   5863    1
#> theta[2,3,1]   7381    1
#> theta[2,3,2]   6698    1
#> theta[2,3,3]   6719    1
#> theta[2,3,4]   6182    1
#> theta[2,4,1]   7345    1
#> theta[2,4,2]   7049    1
#> theta[2,4,3]   6431    1
#> theta[2,4,4]   6509    1
#> theta[3,1,1]   9214    1
#> theta[3,1,2]   5930    1
#> theta[3,1,3]   5528    1
#> theta[3,1,4]   5081    1
#> theta[3,2,1]   7597    1
#> theta[3,2,2]   8452    1
#> theta[3,2,3]   6801    1
#> theta[3,2,4]   6061    1
#> theta[3,3,1]   6433    1
#> theta[3,3,2]   7360    1
#> theta[3,3,3]   7832    1
#> theta[3,3,4]   7227    1
#> theta[3,4,1]   6443    1
#> theta[3,4,2]   6935    1
#> theta[3,4,3]   4898    1
#> theta[3,4,4]   5712    1
#> theta[4,1,1]   8613    1
#> theta[4,1,2]   5584    1
#> theta[4,1,3]   4217    1
#> theta[4,1,4]   5261    1
#> theta[4,2,1]   7891    1
#> theta[4,2,2]   7060    1
#> theta[4,2,3]   5629    1
#> theta[4,2,4]   6135    1
#> theta[4,3,1]   6196    1
#> theta[4,3,2]   6169    1
#> theta[4,3,3]   6098    1
#> theta[4,3,4]   5777    1
#> theta[4,4,1]   6705    1
#> theta[4,4,2]   6353    1
#> theta[4,4,3]   7110    1
#> theta[4,4,4]   7034    1
#> theta[5,1,1]   6556    1
#> theta[5,1,2]   5421    1
#> theta[5,1,3]   4732    1
#> theta[5,1,4]   5027    1
#> theta[5,2,1]   8160    1
#> theta[5,2,2]   9078    1
#> theta[5,2,3]   6729    1
#> theta[5,2,4]   6051    1
#> theta[5,3,1]   5354    1
#> theta[5,3,2]   7851    1
#> theta[5,3,3]   6912    1
#> theta[5,3,4]   7493    1
#> theta[5,4,1]   6593    1
#> theta[5,4,2]   8060    1
#> theta[5,4,3]   5412    1
#> theta[5,4,4]   5650    1
#> log_p_z[1,1]   8103    1
#> log_p_z[1,2]   4143    1
#> log_p_z[1,3]   3387    1
#> log_p_z[1,4]   3885    1
#> log_p_z[2,1]   5245    1
#> log_p_z[2,2]   4661    1
#> log_p_z[2,3]   4791    1
#> log_p_z[2,4]   2687    1
#> log_p_z[3,1]   4294    1
#> log_p_z[3,2]   3948    1
#> log_p_z[3,3]   3654    1
#> log_p_z[3,4]   3777    1
#> log_p_z[4,1]   5659    1
#> log_p_z[4,2]   4898    1
#> log_p_z[4,3]   3545    1
#> log_p_z[4,4]   3228    1
#> log_p_z[5,1]   4259    1
#> log_p_z[5,2]   6618    1
#> log_p_z[5,3]   3621    1
#> log_p_z[5,4]   3173    1
#> log_p_z[6,1]   3890    1
#> log_p_z[6,2]   6144    1
#> log_p_z[6,3]   3657    1
#> log_p_z[6,4]   3204    1
#> log_p_z[7,1]   6889    1
#> log_p_z[7,2]   3767    1
#> log_p_z[7,3]   3263    1
#> log_p_z[7,4]   3492    1
#> log_p_z[8,1]   5148    1
#> log_p_z[8,2]   5670    1
#> log_p_z[8,3]   6072    1
#> log_p_z[8,4]   2845    1
#> log_p_z[9,1]   5577    1
#> log_p_z[9,2]   6203    1
#> log_p_z[9,3]   3547    1
#> log_p_z[9,4]   3039    1
#> log_p_z[10,1]  5201    1
#> log_p_z[10,2]  5685    1
#> log_p_z[10,3]  3483    1
#> log_p_z[10,4]  3189    1
#> log_p_z[11,1]  5219    1
#> log_p_z[11,2]  3949    1
#> log_p_z[11,3]  2784    1
#> log_p_z[11,4]  5851    1
#> log_p_z[12,1]  4920    1
#> log_p_z[12,2]  3640    1
#> log_p_z[12,3]  3584    1
#> log_p_z[12,4]  3085    1
#> log_p_z[13,1]  8103    1
#> log_p_z[13,2]  4143    1
#> log_p_z[13,3]  3387    1
#> log_p_z[13,4]  3885    1
#> log_p_z[14,1]  4181    1
#> log_p_z[14,2]  5048    1
#> log_p_z[14,3]  3732    1
#> log_p_z[14,4]  3083    1
#> log_p_z[15,1]  7355    1
#> log_p_z[15,2]  4053    1
#> log_p_z[15,3]  3351    1
#> log_p_z[15,4]  3527    1
#> log_p_z[16,1]  7510    1
#> log_p_z[16,2]  3879    1
#> log_p_z[16,3]  3337    1
#> log_p_z[16,4]  3855    1
#> log_p_z[17,1]  8103    1
#> log_p_z[17,2]  4143    1
#> log_p_z[17,3]  3387    1
#> log_p_z[17,4]  3885    1
#> log_p_z[18,1]  8103    1
#> log_p_z[18,2]  4143    1
#> log_p_z[18,3]  3387    1
#> log_p_z[18,4]  3885    1
#> log_p_z[19,1]  5402    1
#> log_p_z[19,2]  7060    1
#> log_p_z[19,3]  3488    1
#> log_p_z[19,4]  3148    1
#> log_p_z[20,1]  4079    1
#> log_p_z[20,2]  4714    1
#> log_p_z[20,3]  3591    1
#> log_p_z[20,4]  2998    1
#> log_p_z[21,1]  4607    1
#> log_p_z[21,2]  7567    1
#> log_p_z[21,3]  3547    1
#> log_p_z[21,4]  3087    1
#> log_p_z[22,1]  5402    1
#> log_p_z[22,2]  7060    1
#> log_p_z[22,3]  3488    1
#> log_p_z[22,4]  3148    1
#> log_p_z[23,1]  4259    1
#> log_p_z[23,2]  6618    1
#> log_p_z[23,3]  3621    1
#> log_p_z[23,4]  3173    1
#> log_p_z[24,1]  4381    1
#> log_p_z[24,2]  4535    1
#> log_p_z[24,3]  3608    1
#> log_p_z[24,4]  3495    1
#> log_p_z[25,1]  8103    1
#> log_p_z[25,2]  4143    1
#> log_p_z[25,3]  3387    1
#> log_p_z[25,4]  3885    1
#> log_p_z[26,1]  8103    1
#> log_p_z[26,2]  4143    1
#> log_p_z[26,3]  3387    1
#> log_p_z[26,4]  3885    1
#> log_p_z[27,1]  4521    1
#> log_p_z[27,2]  6282    1
#> log_p_z[27,3]  3499    1
#> log_p_z[27,4]  3130    1
#> log_p_z[28,1]  8103    1
#> log_p_z[28,2]  4143    1
#> log_p_z[28,3]  3387    1
#> log_p_z[28,4]  3885    1
#> log_p_z[29,1]  8103    1
#> log_p_z[29,2]  4143    1
#> log_p_z[29,3]  3387    1
#> log_p_z[29,4]  3885    1
#> log_p_z[30,1]  6316    1
#> log_p_z[30,2]  4251    1
#> log_p_z[30,3]  3704    1
#> log_p_z[30,4]  3801    1
#> log_p_z[31,1]  8103    1
#> log_p_z[31,2]  4143    1
#> log_p_z[31,3]  3387    1
#> log_p_z[31,4]  3885    1
#> log_p_z[32,1]  4937    1
#> log_p_z[32,2]  6000    1
#> log_p_z[32,3]  4407    1
#> log_p_z[32,4]  2961    1
#> log_p_z[33,1]  8103    1
#> log_p_z[33,2]  4143    1
#> log_p_z[33,3]  3387    1
#> log_p_z[33,4]  3885    1
#> log_p_z[34,1]  4607    1
#> log_p_z[34,2]  7567    1
#> log_p_z[34,3]  3547    1
#> log_p_z[34,4]  3087    1
#> log_p_z[35,1]  3847    1
#> log_p_z[35,2]  5994    1
#> log_p_z[35,3]  3716    1
#> log_p_z[35,4]  3089    1
#> log_p_z[36,1]  4551    1
#> log_p_z[36,2]  3868    1
#> log_p_z[36,3]  2748    1
#> log_p_z[36,4]  2603    1
#> log_p_z[37,1]  3904    1
#> log_p_z[37,2]  5055    1
#> log_p_z[37,3]  3652    1
#> log_p_z[37,4]  3172    1
#> log_p_z[38,1]  4571    1
#> log_p_z[38,2]  5313    1
#> log_p_z[38,3]  3691    1
#> log_p_z[38,4]  3268    1
#> log_p_z[39,1]  5334    1
#> log_p_z[39,2]  5824    1
#> log_p_z[39,3]  4523    1
#> log_p_z[39,4]  2727    1
#> log_p_z[40,1]  8103    1
#> log_p_z[40,2]  4143    1
#> log_p_z[40,3]  3387    1
#> log_p_z[40,4]  3885    1
#> log_p_z[41,1]  8103    1
#> log_p_z[41,2]  4143    1
#> log_p_z[41,3]  3387    1
#> log_p_z[41,4]  3885    1
#> log_p_z[42,1]  6889    1
#> log_p_z[42,2]  3767    1
#> log_p_z[42,3]  3263    1
#> log_p_z[42,4]  3492    1
#> log_p_z[43,1]  4521    1
#> log_p_z[43,2]  6282    1
#> log_p_z[43,3]  3499    1
#> log_p_z[43,4]  3130    1
#> log_p_z[44,1]  7355    1
#> log_p_z[44,2]  4053    1
#> log_p_z[44,3]  3351    1
#> log_p_z[44,4]  3527    1
#> log_p_z[45,1]  4607    1
#> log_p_z[45,2]  7567    1
#> log_p_z[45,3]  3547    1
#> log_p_z[45,4]  3087    1
#> log_lik[1]     8103    1
#> log_lik[2]     5092    1
#> log_lik[3]     4902    1
#> log_lik[4]     4924    1
#> log_lik[5]     6619    1
#> log_lik[6]     6147    1
#> log_lik[7]     6887    1
#> log_lik[8]     6030    1
#> log_lik[9]     6204    1
#> log_lik[10]    5706    1
#> log_lik[11]    5852    1
#> log_lik[12]    5674    1
#> log_lik[13]    8103    1
#> log_lik[14]    5048    1
#> log_lik[15]    7355    1
#> log_lik[16]    7510    1
#> log_lik[17]    8103    1
#> log_lik[18]    8103    1
#> log_lik[19]    7060    1
#> log_lik[20]    4718    1
#> log_lik[21]    7567    1
#> log_lik[22]    7060    1
#> log_lik[23]    6619    1
#> log_lik[24]    4535    1
#> log_lik[25]    8103    1
#> log_lik[26]    8103    1
#> log_lik[27]    6284    1
#> log_lik[28]    8103    1
#> log_lik[29]    8103    1
#> log_lik[30]    6313    1
#> log_lik[31]    8103    1
#> log_lik[32]    4417    1
#> log_lik[33]    8103    1
#> log_lik[34]    7567    1
#> log_lik[35]    6004    1
#> log_lik[36]    4049    1
#> log_lik[37]    5058    1
#> log_lik[38]    4491    1
#> log_lik[39]    4505    1
#> log_lik[40]    8103    1
#> log_lik[41]    8103    1
#> log_lik[42]    6887    1
#> log_lik[43]    6284    1
#> log_lik[44]    7355    1
#> log_lik[45]    7567    1
#> lp__           1554    1
#> 
#> Samples were drawn using NUTS(diag_e) at Mon Feb 20 07:27:44 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }