Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.25    0.33    0.38    0.42    0.50
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.40    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.18    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.90    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.06    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.02    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.59    0.73    0.80    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.18    0.40
#> theta[1,4,4]     0.69    0.00 0.14    0.40    0.60    0.70    0.80    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.58    0.70    0.76    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.21    0.32
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[2,2,2]     0.58    0.00 0.10    0.39    0.52    0.58    0.65    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.24    0.30    0.36    0.50
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[2,3,2]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.27
#> theta[2,3,3]     0.62    0.00 0.13    0.36    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.29    0.48
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.76    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.74    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.02    0.05    0.09    0.12    0.23
#> theta[3,2,2]     0.73    0.00 0.09    0.54    0.67    0.73    0.79    0.88
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.30
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.26
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.14    0.25    0.41    0.51    0.60    0.77
#> theta[3,3,4]     0.23    0.00 0.12    0.05    0.15    0.22    0.31    0.50
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[3,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.42
#> theta[3,4,4]     0.63    0.00 0.15    0.32    0.52    0.64    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.07    0.67    0.78    0.83    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.08    0.55    0.67    0.73    0.78    0.87
#> theta[4,2,3]     0.13    0.00 0.06    0.03    0.09    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.10    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.25
#> theta[4,3,3]     0.69    0.00 0.12    0.42    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.39
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[4,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.76    0.90
#> theta[5,1,1]     0.87    0.00 0.07    0.72    0.82    0.87    0.91    0.96
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[5,2,1]     0.15    0.00 0.07    0.04    0.10    0.14    0.19    0.30
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.26
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.22    0.39
#> theta[5,3,3]     0.60    0.00 0.13    0.34    0.51    0.61    0.70    0.84
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.40
#> theta[5,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.42
#> theta[5,4,4]     0.63    0.00 0.15    0.32    0.52    0.64    0.75    0.90
#> log_p_z[1,1]    -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_p_z[1,2]   -20.00    0.04 2.27  -24.70  -21.42  -19.83  -18.37  -16.06
#> log_p_z[1,3]   -24.68    0.08 4.28  -34.13  -27.18  -24.21  -21.69  -17.60
#> log_p_z[1,4]   -22.01    0.08 4.36  -31.74  -24.52  -21.52  -18.88  -14.97
#> log_p_z[2,1]   -29.67    0.07 4.35  -39.50  -32.23  -29.21  -26.65  -22.51
#> log_p_z[2,2]   -21.04    0.04 2.41  -26.39  -22.48  -20.90  -19.34  -16.86
#> log_p_z[2,3]    -7.63    0.02 1.24  -10.41   -8.34   -7.50   -6.80   -5.61
#> log_p_z[2,4]   -16.38    0.08 4.22  -25.93  -18.95  -15.75  -13.31   -9.66
#> log_p_z[3,1]   -11.63    0.03 1.63  -15.40  -12.58  -11.45  -10.49   -8.95
#> log_p_z[3,2]   -10.78    0.02 1.44  -14.00  -11.65  -10.62   -9.76   -8.38
#> log_p_z[3,3]   -23.32    0.06 3.41  -30.92  -25.30  -22.98  -20.97  -17.58
#> log_p_z[3,4]   -21.99    0.06 3.58  -29.86  -24.17  -21.65  -19.47  -15.97
#> log_p_z[4,1]   -14.67    0.03 1.90  -18.72  -15.85  -14.52  -13.33  -11.31
#> log_p_z[4,2]    -7.55    0.01 0.93   -9.57   -8.10   -7.48   -6.90   -5.94
#> log_p_z[4,3]   -20.63    0.08 4.32  -30.78  -23.04  -20.12  -17.56  -13.77
#> log_p_z[4,4]   -21.93    0.08 4.31  -31.54  -24.36  -21.50  -18.87  -15.06
#> log_p_z[5,1]   -21.52    0.04 2.61  -27.29  -23.07  -21.32  -19.72  -16.97
#> log_p_z[5,2]    -3.72    0.01 0.43   -4.60   -4.01   -3.70   -3.42   -2.92
#> log_p_z[5,3]   -18.67    0.08 4.20  -28.42  -21.06  -18.23  -15.57  -12.15
#> log_p_z[5,4]   -21.90    0.08 4.29  -31.89  -24.35  -21.35  -18.85  -14.92
#> log_p_z[6,1]   -21.51    0.04 2.60  -27.31  -23.10  -21.29  -19.64  -17.13
#> log_p_z[6,2]    -5.47    0.01 0.67   -6.89   -5.90   -5.44   -5.00   -4.28
#> log_p_z[6,3]   -17.45    0.08 4.08  -27.42  -19.65  -16.95  -14.48  -11.21
#> log_p_z[6,4]   -21.68    0.08 4.32  -31.74  -24.13  -21.20  -18.65  -14.75
#> log_p_z[7,1]    -6.00    0.01 0.61   -7.30   -6.37   -5.97   -5.57   -4.94
#> log_p_z[7,2]   -14.93    0.03 1.65  -18.46  -15.93  -14.79  -13.75  -12.12
#> log_p_z[7,3]   -24.22    0.06 3.48  -31.85  -26.38  -23.89  -21.76  -18.21
#> log_p_z[7,4]   -22.01    0.06 3.60  -29.86  -24.23  -21.74  -19.41  -15.96
#> log_p_z[8,1]   -29.70    0.07 4.38  -39.52  -32.22  -29.23  -26.60  -22.49
#> log_p_z[8,2]   -18.31    0.03 2.14  -22.93  -19.63  -18.09  -16.82  -14.70
#> log_p_z[8,3]    -5.73    0.01 0.84   -7.49   -6.26   -5.70   -5.13   -4.23
#> log_p_z[8,4]   -18.60    0.08 4.34  -28.71  -21.02  -18.03  -15.55  -11.63
#> log_p_z[9,1]   -20.04    0.03 2.38  -25.09  -21.52  -19.86  -18.37  -15.88
#> log_p_z[9,2]    -4.52    0.01 0.58   -5.82   -4.89   -4.48   -4.12   -3.50
#> log_p_z[9,3]   -19.65    0.08 4.26  -29.66  -22.11  -19.16  -16.59  -12.90
#> log_p_z[9,4]   -21.69    0.07 4.24  -31.35  -24.08  -21.25  -18.73  -14.85
#> log_p_z[10,1]  -22.28    0.04 2.51  -27.87  -23.78  -22.11  -20.53  -17.98
#> log_p_z[10,2]   -7.13    0.01 0.72   -8.66   -7.58   -7.08   -6.63   -5.87
#> log_p_z[10,3]  -16.83    0.06 3.12  -23.81  -18.69  -16.49  -14.61  -11.76
#> log_p_z[10,4]  -21.39    0.06 3.50  -29.02  -23.56  -21.12  -18.95  -15.39
#> log_p_z[11,1]  -29.82    0.07 4.51  -40.45  -32.32  -29.32  -26.66  -22.54
#> log_p_z[11,2]  -29.31    0.07 4.26  -38.80  -31.96  -28.85  -26.30  -22.23
#> log_p_z[11,3]  -18.21    0.07 3.44  -27.10  -19.87  -17.68  -15.86  -13.05
#> log_p_z[11,4]   -5.80    0.01 0.97   -7.92   -6.41   -5.74   -5.12   -4.11
#> log_p_z[12,1]  -23.01    0.04 2.73  -28.86  -24.76  -22.76  -21.09  -18.20
#> log_p_z[12,2]   -9.56    0.02 1.25  -12.34  -10.29   -9.43   -8.70   -7.52
#> log_p_z[12,3]  -15.13    0.08 4.09  -24.79  -17.42  -14.61  -12.21   -8.77
#> log_p_z[12,4]  -19.26    0.07 4.12  -29.02  -21.51  -18.77  -16.43  -12.69
#> log_p_z[13,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_p_z[13,2]  -20.00    0.04 2.27  -24.70  -21.42  -19.83  -18.37  -16.06
#> log_p_z[13,3]  -24.68    0.08 4.28  -34.13  -27.18  -24.21  -21.69  -17.60
#> log_p_z[13,4]  -22.01    0.08 4.36  -31.74  -24.52  -21.52  -18.88  -14.97
#> log_p_z[14,1]  -19.12    0.04 2.45  -24.53  -20.58  -18.88  -17.43  -14.88
#> log_p_z[14,2]   -6.26    0.01 0.86   -8.18   -6.77   -6.19   -5.64   -4.82
#> log_p_z[14,3]  -18.67    0.08 4.22  -28.41  -21.12  -18.14  -15.63  -12.07
#> log_p_z[14,4]  -21.89    0.08 4.33  -32.11  -24.32  -21.35  -18.90  -14.90
#> log_p_z[15,1]   -4.42    0.00 0.41   -5.30   -4.67   -4.39   -4.13   -3.71
#> log_p_z[15,2]  -17.25    0.03 1.84  -21.07  -18.41  -17.15  -15.92  -13.99
#> log_p_z[15,3]  -24.15    0.07 3.53  -31.65  -26.41  -23.87  -21.63  -18.07
#> log_p_z[15,4]  -21.97    0.07 3.60  -29.57  -24.22  -21.65  -19.39  -15.85
#> log_p_z[16,1]   -3.79    0.01 0.54   -4.93   -4.12   -3.75   -3.40   -2.85
#> log_p_z[16,2]  -17.68    0.03 2.11  -22.25  -18.95  -17.50  -16.17  -14.07
#> log_p_z[16,3]  -24.75    0.08 4.27  -34.62  -27.18  -24.25  -21.80  -17.66
#> log_p_z[16,4]  -22.04    0.08 4.35  -31.87  -24.57  -21.55  -18.98  -15.11
#> log_p_z[17,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_p_z[17,2]  -20.00    0.04 2.27  -24.70  -21.42  -19.83  -18.37  -16.06
#> log_p_z[17,3]  -24.68    0.08 4.28  -34.13  -27.18  -24.21  -21.69  -17.60
#> log_p_z[17,4]  -22.01    0.08 4.36  -31.74  -24.52  -21.52  -18.88  -14.97
#> log_p_z[18,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_p_z[18,2]  -20.00    0.04 2.27  -24.70  -21.42  -19.83  -18.37  -16.06
#> log_p_z[18,3]  -24.68    0.08 4.28  -34.13  -27.18  -24.21  -21.69  -17.60
#> log_p_z[18,4]  -22.01    0.08 4.36  -31.74  -24.52  -21.52  -18.88  -14.97
#> log_p_z[19,1]  -16.45    0.03 2.02  -20.94  -17.69  -16.25  -15.07  -12.98
#> log_p_z[19,2]   -4.65    0.01 0.60   -5.88   -5.05   -4.62   -4.23   -3.56
#> log_p_z[19,3]  -22.09    0.08 4.37  -32.32  -24.52  -21.60  -18.97  -15.05
#> log_p_z[19,4]  -21.93    0.08 4.30  -31.97  -24.43  -21.50  -18.92  -15.05
#> log_p_z[20,1]  -18.08    0.03 2.30  -23.28  -19.37  -17.92  -16.45  -14.31
#> log_p_z[20,2]   -7.12    0.01 0.99   -9.28   -7.72   -7.04   -6.42   -5.43
#> log_p_z[20,3]  -19.91    0.08 4.23  -30.09  -22.27  -19.39  -16.85  -13.43
#> log_p_z[20,4]  -21.65    0.08 4.32  -31.44  -23.96  -21.16  -18.64  -14.77
#> log_p_z[21,1]  -19.68    0.04 2.39  -24.94  -21.08  -19.46  -18.05  -15.51
#> log_p_z[21,2]   -3.05    0.00 0.36   -3.82   -3.28   -3.03   -2.80   -2.39
#> log_p_z[21,3]  -21.19    0.08 4.28  -30.99  -23.54  -20.69  -18.12  -14.37
#> log_p_z[21,4]  -21.91    0.08 4.32  -31.98  -24.38  -21.37  -18.90  -14.90
#> log_p_z[22,1]  -16.45    0.03 2.02  -20.94  -17.69  -16.25  -15.07  -12.98
#> log_p_z[22,2]   -4.65    0.01 0.60   -5.88   -5.05   -4.62   -4.23   -3.56
#> log_p_z[22,3]  -22.09    0.08 4.37  -32.32  -24.52  -21.60  -18.97  -15.05
#> log_p_z[22,4]  -21.93    0.08 4.30  -31.97  -24.43  -21.50  -18.92  -15.05
#> log_p_z[23,1]  -21.52    0.04 2.61  -27.29  -23.07  -21.32  -19.72  -16.97
#> log_p_z[23,2]   -3.72    0.01 0.43   -4.60   -4.01   -3.70   -3.42   -2.92
#> log_p_z[23,3]  -18.67    0.08 4.20  -28.42  -21.06  -18.23  -15.57  -12.15
#> log_p_z[23,4]  -21.90    0.08 4.29  -31.89  -24.35  -21.35  -18.85  -14.92
#> log_p_z[24,1]  -17.47    0.03 2.18  -22.32  -18.74  -17.24  -15.99  -13.77
#> log_p_z[24,2]   -5.81    0.01 0.62   -7.17   -6.19   -5.74   -5.37   -4.76
#> log_p_z[24,3]  -21.73    0.06 3.37  -29.18  -23.77  -21.38  -19.37  -16.16
#> log_p_z[24,4]  -21.94    0.06 3.57  -30.16  -24.10  -21.63  -19.41  -15.93
#> log_p_z[25,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_p_z[25,2]  -20.00    0.04 2.27  -24.70  -21.42  -19.83  -18.37  -16.06
#> log_p_z[25,3]  -24.68    0.08 4.28  -34.13  -27.18  -24.21  -21.69  -17.60
#> log_p_z[25,4]  -22.01    0.08 4.36  -31.74  -24.52  -21.52  -18.88  -14.97
#> log_p_z[26,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_p_z[26,2]  -20.00    0.04 2.27  -24.70  -21.42  -19.83  -18.37  -16.06
#> log_p_z[26,3]  -24.68    0.08 4.28  -34.13  -27.18  -24.21  -21.69  -17.60
#> log_p_z[26,4]  -22.01    0.08 4.36  -31.74  -24.52  -21.52  -18.88  -14.97
#> log_p_z[27,1]  -21.92    0.04 2.50  -27.29  -23.44  -21.71  -20.22  -17.56
#> log_p_z[27,2]   -5.66    0.01 0.52   -6.76   -6.00   -5.63   -5.29   -4.72
#> log_p_z[27,3]  -18.38    0.06 3.16  -25.53  -20.23  -18.01  -16.16  -13.14
#> log_p_z[27,4]  -21.61    0.06 3.52  -29.55  -23.76  -21.27  -19.19  -15.65
#> log_p_z[28,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_p_z[28,2]  -20.00    0.04 2.27  -24.70  -21.42  -19.83  -18.37  -16.06
#> log_p_z[28,3]  -24.68    0.08 4.28  -34.13  -27.18  -24.21  -21.69  -17.60
#> log_p_z[28,4]  -22.01    0.08 4.36  -31.74  -24.52  -21.52  -18.88  -14.97
#> log_p_z[29,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_p_z[29,2]  -20.00    0.04 2.27  -24.70  -21.42  -19.83  -18.37  -16.06
#> log_p_z[29,3]  -24.68    0.08 4.28  -34.13  -27.18  -24.21  -21.69  -17.60
#> log_p_z[29,4]  -22.01    0.08 4.36  -31.74  -24.52  -21.52  -18.88  -14.97
#> log_p_z[30,1]   -6.81    0.01 0.84   -8.68   -7.31   -6.72   -6.22   -5.43
#> log_p_z[30,2]  -14.71    0.03 1.68  -18.40  -15.76  -14.53  -13.51  -11.83
#> log_p_z[30,3]  -24.15    0.07 3.54  -31.55  -26.38  -23.82  -21.62  -18.12
#> log_p_z[30,4]  -21.98    0.06 3.57  -29.73  -24.20  -21.68  -19.38  -15.81
#> log_p_z[31,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_p_z[31,2]  -20.00    0.04 2.27  -24.70  -21.42  -19.83  -18.37  -16.06
#> log_p_z[31,3]  -24.68    0.08 4.28  -34.13  -27.18  -24.21  -21.69  -17.60
#> log_p_z[31,4]  -22.01    0.08 4.36  -31.74  -24.52  -21.52  -18.88  -14.97
#> log_p_z[32,1]  -29.70    0.07 4.40  -39.94  -32.23  -29.26  -26.64  -22.49
#> log_p_z[32,2]  -14.84    0.02 1.75  -18.63  -15.92  -14.68  -13.60  -11.83
#> log_p_z[32,3]   -6.07    0.01 0.96   -8.12   -6.66   -5.98   -5.40   -4.40
#> log_p_z[32,4]  -20.73    0.08 4.48  -31.09  -23.23  -20.19  -17.56  -13.42
#> log_p_z[33,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_p_z[33,2]  -20.00    0.04 2.27  -24.70  -21.42  -19.83  -18.37  -16.06
#> log_p_z[33,3]  -24.68    0.08 4.28  -34.13  -27.18  -24.21  -21.69  -17.60
#> log_p_z[33,4]  -22.01    0.08 4.36  -31.74  -24.52  -21.52  -18.88  -14.97
#> log_p_z[34,1]  -19.68    0.04 2.39  -24.94  -21.08  -19.46  -18.05  -15.51
#> log_p_z[34,2]   -3.05    0.00 0.36   -3.82   -3.28   -3.03   -2.80   -2.39
#> log_p_z[34,3]  -21.19    0.08 4.28  -30.99  -23.54  -20.69  -18.12  -14.37
#> log_p_z[34,4]  -21.91    0.08 4.32  -31.98  -24.38  -21.37  -18.90  -14.90
#> log_p_z[35,1]  -22.62    0.04 2.74  -28.54  -24.27  -22.39  -20.75  -17.77
#> log_p_z[35,2]   -5.54    0.01 0.70   -7.08   -5.99   -5.47   -5.03   -4.33
#> log_p_z[35,3]  -16.07    0.08 4.09  -25.56  -18.36  -15.58  -13.11   -9.87
#> log_p_z[35,4]  -21.85    0.08 4.33  -31.83  -24.24  -21.29  -18.88  -14.77
#> log_p_z[36,1]  -29.76    0.06 3.66  -37.99  -31.89  -29.40  -27.27  -23.42
#> log_p_z[36,2]  -21.85    0.04 2.26  -26.72  -23.21  -21.67  -20.27  -18.01
#> log_p_z[36,3]  -10.51    0.04 1.73  -14.84  -11.38  -10.20   -9.32   -7.95
#> log_p_z[36,4]  -14.01    0.06 3.20  -21.16  -16.01  -13.76  -11.65   -8.67
#> log_p_z[37,1]  -18.57    0.04 2.30  -23.60  -19.91  -18.32  -16.99  -14.65
#> log_p_z[37,2]   -7.62    0.01 0.79   -9.34   -8.11   -7.56   -7.06   -6.26
#> log_p_z[37,3]  -19.13    0.06 3.20  -26.26  -21.05  -18.78  -16.83  -13.84
#> log_p_z[37,4]  -21.89    0.06 3.58  -29.89  -24.00  -21.60  -19.38  -15.84
#> log_p_z[38,1]  -25.22    0.04 2.89  -31.55  -26.99  -24.97  -23.23  -20.16
#> log_p_z[38,2]   -9.62    0.01 1.01  -11.79  -10.23   -9.53   -8.91   -7.87
#> log_p_z[38,3]  -11.71    0.05 2.78  -18.19  -13.37  -11.40   -9.62   -7.46
#> log_p_z[38,4]  -21.33    0.06 3.50  -29.04  -23.48  -21.04  -18.84  -15.44
#> log_p_z[39,1]  -29.34    0.07 4.31  -38.95  -31.83  -28.90  -26.36  -22.31
#> log_p_z[39,2]  -16.83    0.03 2.05  -21.28  -18.11  -16.65  -15.39  -13.30
#> log_p_z[39,3]   -7.27    0.02 1.12   -9.72   -7.98   -7.20   -6.48   -5.33
#> log_p_z[39,4]  -18.82    0.07 4.23  -28.48  -21.30  -18.22  -15.76  -12.26
#> log_p_z[40,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_p_z[40,2]  -20.00    0.04 2.27  -24.70  -21.42  -19.83  -18.37  -16.06
#> log_p_z[40,3]  -24.68    0.08 4.28  -34.13  -27.18  -24.21  -21.69  -17.60
#> log_p_z[40,4]  -22.01    0.08 4.36  -31.74  -24.52  -21.52  -18.88  -14.97
#> log_p_z[41,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_p_z[41,2]  -20.00    0.04 2.27  -24.70  -21.42  -19.83  -18.37  -16.06
#> log_p_z[41,3]  -24.68    0.08 4.28  -34.13  -27.18  -24.21  -21.69  -17.60
#> log_p_z[41,4]  -22.01    0.08 4.36  -31.74  -24.52  -21.52  -18.88  -14.97
#> log_p_z[42,1]   -6.00    0.01 0.61   -7.30   -6.37   -5.97   -5.57   -4.94
#> log_p_z[42,2]  -14.93    0.03 1.65  -18.46  -15.93  -14.79  -13.75  -12.12
#> log_p_z[42,3]  -24.22    0.06 3.48  -31.85  -26.38  -23.89  -21.76  -18.21
#> log_p_z[42,4]  -22.01    0.06 3.60  -29.86  -24.23  -21.74  -19.41  -15.96
#> log_p_z[43,1]  -21.92    0.04 2.50  -27.29  -23.44  -21.71  -20.22  -17.56
#> log_p_z[43,2]   -5.66    0.01 0.52   -6.76   -6.00   -5.63   -5.29   -4.72
#> log_p_z[43,3]  -18.38    0.06 3.16  -25.53  -20.23  -18.01  -16.16  -13.14
#> log_p_z[43,4]  -21.61    0.06 3.52  -29.55  -23.76  -21.27  -19.19  -15.65
#> log_p_z[44,1]   -4.42    0.00 0.41   -5.30   -4.67   -4.39   -4.13   -3.71
#> log_p_z[44,2]  -17.25    0.03 1.84  -21.07  -18.41  -17.15  -15.92  -13.99
#> log_p_z[44,3]  -24.15    0.07 3.53  -31.65  -26.41  -23.87  -21.63  -18.07
#> log_p_z[44,4]  -21.97    0.07 3.60  -29.57  -24.22  -21.65  -19.39  -15.85
#> log_p_z[45,1]  -19.68    0.04 2.39  -24.94  -21.08  -19.46  -18.05  -15.51
#> log_p_z[45,2]   -3.05    0.00 0.36   -3.82   -3.28   -3.03   -2.80   -2.39
#> log_p_z[45,3]  -21.19    0.08 4.28  -30.99  -23.54  -20.69  -18.12  -14.37
#> log_p_z[45,4]  -21.91    0.08 4.32  -31.98  -24.38  -21.37  -18.90  -14.90
#> log_lik[1]      -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_lik[2]      -7.58    0.02 1.13  -10.05   -8.29   -7.49   -6.79   -5.61
#> log_lik[3]      -9.94    0.01 0.88  -11.71  -10.53   -9.93   -9.35   -8.26
#> log_lik[4]      -7.54    0.01 0.92   -9.55   -8.10   -7.48   -6.89   -5.94
#> log_lik[5]      -3.72    0.01 0.43   -4.60   -4.01   -3.70   -3.42   -2.92
#> log_lik[6]      -5.47    0.01 0.67   -6.89   -5.89   -5.44   -5.00   -4.28
#> log_lik[7]      -6.00    0.01 0.61   -7.29   -6.37   -5.97   -5.57   -4.94
#> log_lik[8]      -5.73    0.01 0.84   -7.49   -6.26   -5.70   -5.13   -4.23
#> log_lik[9]      -4.52    0.01 0.58   -5.82   -4.89   -4.48   -4.12   -3.50
#> log_lik[10]     -7.13    0.01 0.72   -8.66   -7.58   -7.08   -6.63   -5.87
#> log_lik[11]     -5.80    0.01 0.97   -7.92   -6.41   -5.74   -5.12   -4.11
#> log_lik[12]     -9.31    0.01 1.04  -11.53   -9.99   -9.25   -8.58   -7.48
#> log_lik[13]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_lik[14]     -6.26    0.01 0.86   -8.18   -6.77   -6.19   -5.64   -4.82
#> log_lik[15]     -4.42    0.00 0.41   -5.30   -4.67   -4.39   -4.13   -3.71
#> log_lik[16]     -3.79    0.01 0.54   -4.93   -4.12   -3.75   -3.40   -2.85
#> log_lik[17]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_lik[18]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_lik[19]     -4.65    0.01 0.60   -5.88   -5.05   -4.62   -4.23   -3.56
#> log_lik[20]     -7.12    0.01 0.99   -9.27   -7.72   -7.04   -6.42   -5.43
#> log_lik[21]     -3.05    0.00 0.36   -3.82   -3.28   -3.03   -2.80   -2.39
#> log_lik[22]     -4.65    0.01 0.60   -5.88   -5.05   -4.62   -4.23   -3.56
#> log_lik[23]     -3.72    0.01 0.43   -4.60   -4.01   -3.70   -3.42   -2.92
#> log_lik[24]     -5.81    0.01 0.62   -7.17   -6.19   -5.74   -5.37   -4.76
#> log_lik[25]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_lik[26]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_lik[27]     -5.66    0.01 0.52   -6.76   -6.00   -5.63   -5.29   -4.72
#> log_lik[28]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_lik[29]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_lik[30]     -6.81    0.01 0.84   -8.68   -7.31   -6.72   -6.22   -5.43
#> log_lik[31]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_lik[32]     -6.07    0.01 0.95   -8.10   -6.66   -5.98   -5.40   -4.40
#> log_lik[33]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_lik[34]     -3.05    0.00 0.36   -3.82   -3.28   -3.03   -2.80   -2.39
#> log_lik[35]     -5.53    0.01 0.70   -7.08   -5.99   -5.47   -5.03   -4.33
#> log_lik[36]     -9.92    0.02 1.17  -12.46  -10.63   -9.85   -9.10   -7.81
#> log_lik[37]     -7.62    0.01 0.79   -9.33   -8.11   -7.56   -7.06   -6.26
#> log_lik[38]     -9.00    0.01 0.87  -10.82   -9.54   -8.98   -8.45   -7.28
#> log_lik[39]     -7.27    0.02 1.11   -9.72   -7.98   -7.20   -6.48   -5.33
#> log_lik[40]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_lik[41]     -2.20    0.00 0.29   -2.81   -2.39   -2.19   -2.00   -1.68
#> log_lik[42]     -6.00    0.01 0.61   -7.29   -6.37   -5.97   -5.57   -4.94
#> log_lik[43]     -5.66    0.01 0.52   -6.76   -6.00   -5.63   -5.29   -4.72
#> log_lik[44]     -4.42    0.00 0.41   -5.30   -4.67   -4.39   -4.13   -3.71
#> log_lik[45]     -3.05    0.00 0.36   -3.82   -3.28   -3.03   -2.80   -2.39
#> lp__          -463.83    0.16 6.18 -476.54 -467.93 -463.43 -459.54 -452.65
#>               n_eff Rhat
#> pi[1]          7441    1
#> pi[2]          7154    1
#> pi[3]          6220    1
#> pi[4]          5554    1
#> theta[1,1,1]   7312    1
#> theta[1,1,2]   7225    1
#> theta[1,1,3]   5151    1
#> theta[1,1,4]   4168    1
#> theta[1,2,1]   5748    1
#> theta[1,2,2]   6297    1
#> theta[1,2,3]   6304    1
#> theta[1,2,4]   5525    1
#> theta[1,3,1]   5494    1
#> theta[1,3,2]   4032    1
#> theta[1,3,3]   4807    1
#> theta[1,3,4]   4953    1
#> theta[1,4,1]   5790    1
#> theta[1,4,2]   5740    1
#> theta[1,4,3]   3652    1
#> theta[1,4,4]   4531    1
#> theta[2,1,1]   6910    1
#> theta[2,1,2]   6693    1
#> theta[2,1,3]   3973    1
#> theta[2,1,4]   5355    1
#> theta[2,2,1]   8318    1
#> theta[2,2,2]   7367    1
#> theta[2,2,3]   7153    1
#> theta[2,2,4]   6275    1
#> theta[2,3,1]   6203    1
#> theta[2,3,2]   6549    1
#> theta[2,3,3]   5255    1
#> theta[2,3,4]   4645    1
#> theta[2,4,1]   6670    1
#> theta[2,4,2]   6475    1
#> theta[2,4,3]   6314    1
#> theta[2,4,4]   6070    1
#> theta[3,1,1]   6319    1
#> theta[3,1,2]   5107    1
#> theta[3,1,3]   4620    1
#> theta[3,1,4]   4466    1
#> theta[3,2,1]   6022    1
#> theta[3,2,2]   7503    1
#> theta[3,2,3]   6022    1
#> theta[3,2,4]   5579    1
#> theta[3,3,1]   6369    1
#> theta[3,3,2]   7800    1
#> theta[3,3,3]   6308    1
#> theta[3,3,4]   6764    1
#> theta[3,4,1]   7121    1
#> theta[3,4,2]   6023    1
#> theta[3,4,3]   5372    1
#> theta[3,4,4]   5490    1
#> theta[4,1,1]   7844    1
#> theta[4,1,2]   5486    1
#> theta[4,1,3]   4603    1
#> theta[4,1,4]   4860    1
#> theta[4,2,1]   6912    1
#> theta[4,2,2]   6967    1
#> theta[4,2,3]   6325    1
#> theta[4,2,4]   5713    1
#> theta[4,3,1]   6275    1
#> theta[4,3,2]   6959    1
#> theta[4,3,3]   5478    1
#> theta[4,3,4]   4916    1
#> theta[4,4,1]   6709    1
#> theta[4,4,2]   6113    1
#> theta[4,4,3]   6020    1
#> theta[4,4,4]   6539    1
#> theta[5,1,1]   6453    1
#> theta[5,1,2]   5174    1
#> theta[5,1,3]   4966    1
#> theta[5,1,4]   4289    1
#> theta[5,2,1]   7761    1
#> theta[5,2,2]   7532    1
#> theta[5,2,3]   5988    1
#> theta[5,2,4]   6460    1
#> theta[5,3,1]   6721    1
#> theta[5,3,2]   6603    1
#> theta[5,3,3]   6714    1
#> theta[5,3,4]   6619    1
#> theta[5,4,1]   5829    1
#> theta[5,4,2]   6705    1
#> theta[5,4,3]   5569    1
#> theta[5,4,4]   6286    1
#> log_p_z[1,1]   6989    1
#> log_p_z[1,2]   3776    1
#> log_p_z[1,3]   2926    1
#> log_p_z[1,4]   3292    1
#> log_p_z[2,1]   4401    1
#> log_p_z[2,2]   3747    1
#> log_p_z[2,3]   3012    1
#> log_p_z[2,4]   2867    1
#> log_p_z[3,1]   3870    1
#> log_p_z[3,2]   3832    1
#> log_p_z[3,3]   2818    1
#> log_p_z[3,4]   3307    1
#> log_p_z[4,1]   5263    1
#> log_p_z[4,2]   4732    1
#> log_p_z[4,3]   2627    1
#> log_p_z[4,4]   2948    1
#> log_p_z[5,1]   4143    1
#> log_p_z[5,2]   7262    1
#> log_p_z[5,3]   2792    1
#> log_p_z[5,4]   3128    1
#> log_p_z[6,1]   4517    1
#> log_p_z[6,2]   6107    1
#> log_p_z[6,3]   2748    1
#> log_p_z[6,4]   3229    1
#> log_p_z[7,1]   5821    1
#> log_p_z[7,2]   3787    1
#> log_p_z[7,3]   2877    1
#> log_p_z[7,4]   3157    1
#> log_p_z[8,1]   4220    1
#> log_p_z[8,2]   4615    1
#> log_p_z[8,3]   5893    1
#> log_p_z[8,4]   2853    1
#> log_p_z[9,1]   5046    1
#> log_p_z[9,2]   5822    1
#> log_p_z[9,3]   2765    1
#> log_p_z[9,4]   3214    1
#> log_p_z[10,1]  4817    1
#> log_p_z[10,2]  5197    1
#> log_p_z[10,3]  2858    1
#> log_p_z[10,4]  3271    1
#> log_p_z[11,1]  3873    1
#> log_p_z[11,2]  4079    1
#> log_p_z[11,3]  2479    1
#> log_p_z[11,4]  6056    1
#> log_p_z[12,1]  4980    1
#> log_p_z[12,2]  4400    1
#> log_p_z[12,3]  2777    1
#> log_p_z[12,4]  3132    1
#> log_p_z[13,1]  6989    1
#> log_p_z[13,2]  3776    1
#> log_p_z[13,3]  2926    1
#> log_p_z[13,4]  3292    1
#> log_p_z[14,1]  4363    1
#> log_p_z[14,2]  5035    1
#> log_p_z[14,3]  2859    1
#> log_p_z[14,4]  2984    1
#> log_p_z[15,1]  6788    1
#> log_p_z[15,2]  3706    1
#> log_p_z[15,3]  2937    1
#> log_p_z[15,4]  2975    1
#> log_p_z[16,1]  6368    1
#> log_p_z[16,2]  3837    1
#> log_p_z[16,3]  2903    1
#> log_p_z[16,4]  3352    1
#> log_p_z[17,1]  6989    1
#> log_p_z[17,2]  3776    1
#> log_p_z[17,3]  2926    1
#> log_p_z[17,4]  3292    1
#> log_p_z[18,1]  6989    1
#> log_p_z[18,2]  3776    1
#> log_p_z[18,3]  2926    1
#> log_p_z[18,4]  3292    1
#> log_p_z[19,1]  4811    1
#> log_p_z[19,2]  6288    1
#> log_p_z[19,3]  2809    1
#> log_p_z[19,4]  3125    1
#> log_p_z[20,1]  4574    1
#> log_p_z[20,2]  5218    1
#> log_p_z[20,3]  2779    1
#> log_p_z[20,4]  3057    1
#> log_p_z[21,1]  4243    1
#> log_p_z[21,2]  6744    1
#> log_p_z[21,3]  2789    1
#> log_p_z[21,4]  3223    1
#> log_p_z[22,1]  4811    1
#> log_p_z[22,2]  6288    1
#> log_p_z[22,3]  2809    1
#> log_p_z[22,4]  3125    1
#> log_p_z[23,1]  4143    1
#> log_p_z[23,2]  7262    1
#> log_p_z[23,3]  2792    1
#> log_p_z[23,4]  3128    1
#> log_p_z[24,1]  4004    1
#> log_p_z[24,2]  4578    1
#> log_p_z[24,3]  2933    1
#> log_p_z[24,4]  3208    1
#> log_p_z[25,1]  6989    1
#> log_p_z[25,2]  3776    1
#> log_p_z[25,3]  2926    1
#> log_p_z[25,4]  3292    1
#> log_p_z[26,1]  6989    1
#> log_p_z[26,2]  3776    1
#> log_p_z[26,3]  2926    1
#> log_p_z[26,4]  3292    1
#> log_p_z[27,1]  4185    1
#> log_p_z[27,2]  6003    1
#> log_p_z[27,3]  2901    1
#> log_p_z[27,4]  3382    1
#> log_p_z[28,1]  6989    1
#> log_p_z[28,2]  3776    1
#> log_p_z[28,3]  2926    1
#> log_p_z[28,4]  3292    1
#> log_p_z[29,1]  6989    1
#> log_p_z[29,2]  3776    1
#> log_p_z[29,3]  2926    1
#> log_p_z[29,4]  3292    1
#> log_p_z[30,1]  5421    1
#> log_p_z[30,2]  3781    1
#> log_p_z[30,3]  2856    1
#> log_p_z[30,4]  3246    1
#> log_p_z[31,1]  6989    1
#> log_p_z[31,2]  3776    1
#> log_p_z[31,3]  2926    1
#> log_p_z[31,4]  3292    1
#> log_p_z[32,1]  4369    1
#> log_p_z[32,2]  5134    1
#> log_p_z[32,3]  4971    1
#> log_p_z[32,4]  2915    1
#> log_p_z[33,1]  6989    1
#> log_p_z[33,2]  3776    1
#> log_p_z[33,3]  2926    1
#> log_p_z[33,4]  3292    1
#> log_p_z[34,1]  4243    1
#> log_p_z[34,2]  6744    1
#> log_p_z[34,3]  2789    1
#> log_p_z[34,4]  3223    1
#> log_p_z[35,1]  4368    1
#> log_p_z[35,2]  6107    1
#> log_p_z[35,3]  2775    1
#> log_p_z[35,4]  3205    1
#> log_p_z[36,1]  3623    1
#> log_p_z[36,2]  3859    1
#> log_p_z[36,3]  2148    1
#> log_p_z[36,4]  2715    1
#> log_p_z[37,1]  4143    1
#> log_p_z[37,2]  4922    1
#> log_p_z[37,3]  2856    1
#> log_p_z[37,4]  3408    1
#> log_p_z[38,1]  4259    1
#> log_p_z[38,2]  4854    1
#> log_p_z[38,3]  2801    1
#> log_p_z[38,4]  3154    1
#> log_p_z[39,1]  4154    1
#> log_p_z[39,2]  4671    1
#> log_p_z[39,3]  4657    1
#> log_p_z[39,4]  3203    1
#> log_p_z[40,1]  6989    1
#> log_p_z[40,2]  3776    1
#> log_p_z[40,3]  2926    1
#> log_p_z[40,4]  3292    1
#> log_p_z[41,1]  6989    1
#> log_p_z[41,2]  3776    1
#> log_p_z[41,3]  2926    1
#> log_p_z[41,4]  3292    1
#> log_p_z[42,1]  5821    1
#> log_p_z[42,2]  3787    1
#> log_p_z[42,3]  2877    1
#> log_p_z[42,4]  3157    1
#> log_p_z[43,1]  4185    1
#> log_p_z[43,2]  6003    1
#> log_p_z[43,3]  2901    1
#> log_p_z[43,4]  3382    1
#> log_p_z[44,1]  6788    1
#> log_p_z[44,2]  3706    1
#> log_p_z[44,3]  2937    1
#> log_p_z[44,4]  2975    1
#> log_p_z[45,1]  4243    1
#> log_p_z[45,2]  6744    1
#> log_p_z[45,3]  2789    1
#> log_p_z[45,4]  3223    1
#> log_lik[1]     6989    1
#> log_lik[2]     4377    1
#> log_lik[3]     4949    1
#> log_lik[4]     4772    1
#> log_lik[5]     7262    1
#> log_lik[6]     6107    1
#> log_lik[7]     5824    1
#> log_lik[8]     5891    1
#> log_lik[9]     5822    1
#> log_lik[10]    5191    1
#> log_lik[11]    6056    1
#> log_lik[12]    5657    1
#> log_lik[13]    6989    1
#> log_lik[14]    5043    1
#> log_lik[15]    6788    1
#> log_lik[16]    6368    1
#> log_lik[17]    6989    1
#> log_lik[18]    6989    1
#> log_lik[19]    6288    1
#> log_lik[20]    5223    1
#> log_lik[21]    6744    1
#> log_lik[22]    6288    1
#> log_lik[23]    7262    1
#> log_lik[24]    4579    1
#> log_lik[25]    6989    1
#> log_lik[26]    6989    1
#> log_lik[27]    6003    1
#> log_lik[28]    6989    1
#> log_lik[29]    6989    1
#> log_lik[30]    5433    1
#> log_lik[31]    6989    1
#> log_lik[32]    4976    1
#> log_lik[33]    6989    1
#> log_lik[34]    6744    1
#> log_lik[35]    6136    1
#> log_lik[36]    4298    1
#> log_lik[37]    4924    1
#> log_lik[38]    4148    1
#> log_lik[39]    4668    1
#> log_lik[40]    6989    1
#> log_lik[41]    6989    1
#> log_lik[42]    5824    1
#> log_lik[43]    6003    1
#> log_lik[44]    6788    1
#> log_lik[45]    6744    1
#> lp__           1419    1
#> 
#> Samples were drawn using NUTS(diag_e) at Tue Mar  7 09:30:00 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }