Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.26    0.33    0.37    0.42    0.50
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.03    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.83    0.87    0.90    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.00    0.01    0.02    0.05
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.07    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.29
#> theta[1,3,3]     0.79    0.00 0.09    0.58    0.73    0.80    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.29
#> theta[1,4,2]     0.09    0.00 0.09    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.09    0.18    0.41
#> theta[1,4,4]     0.69    0.00 0.14    0.40    0.60    0.70    0.79    0.91
#> theta[2,1,1]     0.75    0.00 0.08    0.58    0.70    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.22    0.33
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.58    0.64    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.25    0.30    0.37    0.49
#> theta[2,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,3]     0.62    0.00 0.13    0.35    0.53    0.63    0.73    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.28    0.47
#> theta[2,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.90
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.10    0.00 0.06    0.01    0.05    0.09    0.13    0.24
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.28
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.25    0.43
#> theta[3,3,3]     0.51    0.00 0.14    0.25    0.41    0.51    0.60    0.78
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.30    0.48
#> theta[3,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.35
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[3,4,3]     0.14    0.00 0.11    0.01    0.05    0.12    0.20    0.42
#> theta[3,4,4]     0.63    0.00 0.15    0.31    0.53    0.64    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.78    0.84    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.08    0.54    0.67    0.73    0.78    0.87
#> theta[4,2,3]     0.13    0.00 0.07    0.04    0.09    0.12    0.17    0.29
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.18
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,3]     0.69    0.00 0.12    0.43    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.38
#> theta[4,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,4]     0.65    0.00 0.15    0.34    0.56    0.66    0.76    0.90
#> theta[5,1,1]     0.87    0.00 0.07    0.72    0.83    0.87    0.91    0.96
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.30
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.60    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,3,3]     0.60    0.00 0.13    0.34    0.52    0.61    0.70    0.84
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.40
#> theta[5,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.39
#> theta[5,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.42
#> theta[5,4,4]     0.63    0.00 0.15    0.30    0.53    0.65    0.75    0.89
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_p_z[1,2]   -20.00    0.04 2.31  -25.04  -21.41  -19.79  -18.38  -15.97
#> log_p_z[1,3]   -24.68    0.07 4.29  -34.11  -27.18  -24.26  -21.69  -17.45
#> log_p_z[1,4]   -22.00    0.08 4.26  -31.85  -24.52  -21.56  -19.01  -15.00
#> log_p_z[2,1]   -29.65    0.06 4.30  -39.42  -32.13  -29.19  -26.61  -22.66
#> log_p_z[2,2]   -21.01    0.04 2.38  -26.14  -22.45  -20.78  -19.35  -16.85
#> log_p_z[2,3]    -7.65    0.02 1.21  -10.39   -8.34   -7.53   -6.81   -5.61
#> log_p_z[2,4]   -16.48    0.08 4.28  -26.39  -19.01  -15.99  -13.40   -9.67
#> log_p_z[3,1]   -11.63    0.03 1.66  -15.46  -12.57  -11.44  -10.44   -8.92
#> log_p_z[3,2]   -10.82    0.02 1.49  -14.18  -11.75  -10.68   -9.73   -8.38
#> log_p_z[3,3]   -23.33    0.06 3.40  -30.64  -25.44  -23.01  -20.87  -17.52
#> log_p_z[3,4]   -21.83    0.06 3.51  -29.53  -24.02  -21.51  -19.32  -15.91
#> log_p_z[4,1]   -14.63    0.03 1.84  -18.63  -15.75  -14.50  -13.34  -11.35
#> log_p_z[4,2]    -7.58    0.01 0.95   -9.59   -8.16   -7.51   -6.91   -5.91
#> log_p_z[4,3]   -20.68    0.08 4.32  -30.32  -23.24  -20.14  -17.57  -13.81
#> log_p_z[4,4]   -21.80    0.08 4.29  -31.65  -24.20  -21.42  -18.78  -14.77
#> log_p_z[5,1]   -21.46    0.04 2.59  -27.07  -23.07  -21.27  -19.63  -16.95
#> log_p_z[5,2]    -3.71    0.01 0.42   -4.60   -3.96   -3.69   -3.42   -2.95
#> log_p_z[5,3]   -18.76    0.08 4.25  -28.40  -21.14  -18.22  -15.70  -12.20
#> log_p_z[5,4]   -21.80    0.08 4.29  -31.69  -24.30  -21.37  -18.80  -14.77
#> log_p_z[6,1]   -21.44    0.04 2.57  -26.84  -23.06  -21.29  -19.63  -16.92
#> log_p_z[6,2]    -5.47    0.01 0.65   -6.86   -5.87   -5.42   -5.01   -4.35
#> log_p_z[6,3]   -17.55    0.07 4.11  -26.99  -19.90  -16.96  -14.58  -11.26
#> log_p_z[6,4]   -21.55    0.08 4.24  -31.40  -23.94  -21.08  -18.61  -14.64
#> log_p_z[7,1]    -6.00    0.01 0.62   -7.36   -6.38   -5.95   -5.57   -4.92
#> log_p_z[7,2]   -14.95    0.03 1.69  -18.62  -15.99  -14.82  -13.77  -12.01
#> log_p_z[7,3]   -24.20    0.06 3.54  -32.03  -26.36  -23.91  -21.64  -18.31
#> log_p_z[7,4]   -21.90    0.07 3.55  -29.76  -24.07  -21.55  -19.43  -15.89
#> log_p_z[8,1]   -29.61    0.06 4.33  -39.29  -32.05  -29.20  -26.61  -22.41
#> log_p_z[8,2]   -18.25    0.03 2.11  -22.91  -19.57  -18.10  -16.73  -14.57
#> log_p_z[8,3]    -5.73    0.01 0.84   -7.51   -6.27   -5.67   -5.15   -4.23
#> log_p_z[8,4]   -18.67    0.08 4.33  -28.94  -21.22  -18.15  -15.60  -11.72
#> log_p_z[9,1]   -20.01    0.03 2.37  -25.31  -21.44  -19.80  -18.34  -15.92
#> log_p_z[9,2]    -4.52    0.01 0.57   -5.77   -4.88   -4.48   -4.11   -3.53
#> log_p_z[9,3]   -19.71    0.08 4.30  -29.68  -22.18  -19.19  -16.61  -12.91
#> log_p_z[9,4]   -21.59    0.08 4.30  -31.22  -24.06  -21.10  -18.56  -14.58
#> log_p_z[10,1]  -22.23    0.04 2.49  -27.82  -23.68  -21.97  -20.49  -18.00
#> log_p_z[10,2]   -7.12    0.01 0.71   -8.67   -7.56   -7.06   -6.61   -5.86
#> log_p_z[10,3]  -16.87    0.06 3.13  -23.79  -18.80  -16.56  -14.61  -11.74
#> log_p_z[10,4]  -21.33    0.06 3.53  -29.07  -23.47  -21.02  -18.82  -15.35
#> log_p_z[11,1]  -29.74    0.07 4.39  -39.50  -32.20  -29.28  -26.65  -22.53
#> log_p_z[11,2]  -29.19    0.06 4.13  -38.42  -31.61  -28.79  -26.26  -22.34
#> log_p_z[11,3]  -18.14    0.07 3.41  -26.01  -19.84  -17.69  -15.83  -13.02
#> log_p_z[11,4]   -5.81    0.01 0.97   -7.88   -6.42   -5.74   -5.13   -4.14
#> log_p_z[12,1]  -22.98    0.04 2.75  -28.74  -24.72  -22.78  -21.03  -18.24
#> log_p_z[12,2]   -9.58    0.02 1.27  -12.47  -10.30   -9.43   -8.70   -7.50
#> log_p_z[12,3]  -15.19    0.07 4.11  -24.57  -17.55  -14.62  -12.29   -9.01
#> log_p_z[12,4]  -19.18    0.08 4.15  -28.71  -21.65  -18.69  -16.23  -12.54
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_p_z[13,2]  -20.00    0.04 2.31  -25.04  -21.41  -19.79  -18.38  -15.97
#> log_p_z[13,3]  -24.68    0.07 4.29  -34.11  -27.18  -24.26  -21.69  -17.45
#> log_p_z[13,4]  -22.00    0.08 4.26  -31.85  -24.52  -21.56  -19.01  -15.00
#> log_p_z[14,1]  -19.05    0.04 2.42  -24.38  -20.55  -18.85  -17.35  -14.91
#> log_p_z[14,2]   -6.22    0.01 0.87   -8.16   -6.72   -6.15   -5.62   -4.75
#> log_p_z[14,3]  -18.75    0.08 4.31  -28.28  -21.21  -18.20  -15.60  -12.14
#> log_p_z[14,4]  -21.84    0.08 4.30  -31.70  -24.33  -21.29  -18.81  -14.81
#> log_p_z[15,1]   -4.42    0.01 0.41   -5.30   -4.67   -4.38   -4.13   -3.71
#> log_p_z[15,2]  -17.24    0.03 1.85  -21.26  -18.34  -17.09  -15.95  -13.93
#> log_p_z[15,3]  -24.17    0.06 3.54  -31.92  -26.36  -23.96  -21.68  -18.01
#> log_p_z[15,4]  -21.96    0.07 3.48  -29.45  -24.14  -21.71  -19.49  -16.00
#> log_p_z[16,1]   -3.79    0.01 0.55   -5.00   -4.13   -3.75   -3.42   -2.83
#> log_p_z[16,2]  -17.71    0.04 2.18  -22.54  -19.06  -17.50  -16.16  -14.02
#> log_p_z[16,3]  -24.71    0.07 4.29  -34.28  -27.26  -24.20  -21.61  -17.85
#> log_p_z[16,4]  -21.94    0.08 4.31  -31.63  -24.43  -21.49  -18.87  -14.84
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_p_z[17,2]  -20.00    0.04 2.31  -25.04  -21.41  -19.79  -18.38  -15.97
#> log_p_z[17,3]  -24.68    0.07 4.29  -34.11  -27.18  -24.26  -21.69  -17.45
#> log_p_z[17,4]  -22.00    0.08 4.26  -31.85  -24.52  -21.56  -19.01  -15.00
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_p_z[18,2]  -20.00    0.04 2.31  -25.04  -21.41  -19.79  -18.38  -15.97
#> log_p_z[18,3]  -24.68    0.07 4.29  -34.11  -27.18  -24.26  -21.69  -17.45
#> log_p_z[18,4]  -22.00    0.08 4.26  -31.85  -24.52  -21.56  -19.01  -15.00
#> log_p_z[19,1]  -16.42    0.03 2.03  -20.98  -17.66  -16.25  -15.01  -12.91
#> log_p_z[19,2]   -4.67    0.01 0.59   -5.91   -5.04   -4.63   -4.26   -3.63
#> log_p_z[19,3]  -22.14    0.08 4.39  -32.20  -24.71  -21.63  -18.99  -15.05
#> log_p_z[19,4]  -21.81    0.08 4.34  -31.34  -24.31  -21.38  -18.77  -14.59
#> log_p_z[20,1]  -18.04    0.04 2.35  -23.20  -19.51  -17.87  -16.36  -13.97
#> log_p_z[20,2]   -7.09    0.01 0.95   -9.21   -7.68   -7.01   -6.43   -5.48
#> log_p_z[20,3]  -20.02    0.07 4.20  -29.57  -22.43  -19.47  -17.02  -13.42
#> log_p_z[20,4]  -21.59    0.08 4.15  -31.17  -23.91  -21.18  -18.69  -14.93
#> log_p_z[21,1]  -19.64    0.04 2.45  -24.92  -21.16  -19.48  -17.90  -15.40
#> log_p_z[21,2]   -3.04    0.00 0.36   -3.79   -3.28   -3.02   -2.80   -2.38
#> log_p_z[21,3]  -21.26    0.08 4.33  -31.32  -23.75  -20.68  -18.28  -14.43
#> log_p_z[21,4]  -21.78    0.08 4.29  -31.39  -24.25  -21.35  -18.75  -14.83
#> log_p_z[22,1]  -16.42    0.03 2.03  -20.98  -17.66  -16.25  -15.01  -12.91
#> log_p_z[22,2]   -4.67    0.01 0.59   -5.91   -5.04   -4.63   -4.26   -3.63
#> log_p_z[22,3]  -22.14    0.08 4.39  -32.20  -24.71  -21.63  -18.99  -15.05
#> log_p_z[22,4]  -21.81    0.08 4.34  -31.34  -24.31  -21.38  -18.77  -14.59
#> log_p_z[23,1]  -21.46    0.04 2.59  -27.07  -23.07  -21.27  -19.63  -16.95
#> log_p_z[23,2]   -3.71    0.01 0.42   -4.60   -3.96   -3.69   -3.42   -2.95
#> log_p_z[23,3]  -18.76    0.08 4.25  -28.40  -21.14  -18.22  -15.70  -12.20
#> log_p_z[23,4]  -21.80    0.08 4.29  -31.69  -24.30  -21.37  -18.80  -14.77
#> log_p_z[24,1]  -17.44    0.03 2.22  -22.31  -18.82  -17.23  -15.88  -13.72
#> log_p_z[24,2]   -5.81    0.01 0.64   -7.29   -6.19   -5.74   -5.36   -4.74
#> log_p_z[24,3]  -21.77    0.06 3.42  -29.51  -23.83  -21.42  -19.33  -16.13
#> log_p_z[24,4]  -21.82    0.07 3.53  -29.63  -23.99  -21.50  -19.32  -15.88
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_p_z[25,2]  -20.00    0.04 2.31  -25.04  -21.41  -19.79  -18.38  -15.97
#> log_p_z[25,3]  -24.68    0.07 4.29  -34.11  -27.18  -24.26  -21.69  -17.45
#> log_p_z[25,4]  -22.00    0.08 4.26  -31.85  -24.52  -21.56  -19.01  -15.00
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_p_z[26,2]  -20.00    0.04 2.31  -25.04  -21.41  -19.79  -18.38  -15.97
#> log_p_z[26,3]  -24.68    0.07 4.29  -34.11  -27.18  -24.26  -21.69  -17.45
#> log_p_z[26,4]  -22.00    0.08 4.26  -31.85  -24.52  -21.56  -19.01  -15.00
#> log_p_z[27,1]  -21.86    0.04 2.55  -27.58  -23.44  -21.66  -20.07  -17.51
#> log_p_z[27,2]   -5.64    0.01 0.53   -6.80   -5.97   -5.61   -5.26   -4.72
#> log_p_z[27,3]  -18.42    0.06 3.18  -25.69  -20.32  -18.06  -16.21  -13.24
#> log_p_z[27,4]  -21.52    0.06 3.49  -29.22  -23.55  -21.28  -19.10  -15.68
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_p_z[28,2]  -20.00    0.04 2.31  -25.04  -21.41  -19.79  -18.38  -15.97
#> log_p_z[28,3]  -24.68    0.07 4.29  -34.11  -27.18  -24.26  -21.69  -17.45
#> log_p_z[28,4]  -22.00    0.08 4.26  -31.85  -24.52  -21.56  -19.01  -15.00
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_p_z[29,2]  -20.00    0.04 2.31  -25.04  -21.41  -19.79  -18.38  -15.97
#> log_p_z[29,3]  -24.68    0.07 4.29  -34.11  -27.18  -24.26  -21.69  -17.45
#> log_p_z[29,4]  -22.00    0.08 4.26  -31.85  -24.52  -21.56  -19.01  -15.00
#> log_p_z[30,1]   -6.82    0.01 0.85   -8.79   -7.32   -6.73   -6.24   -5.45
#> log_p_z[30,2]  -14.73    0.03 1.73  -18.52  -15.82  -14.57  -13.51  -11.66
#> log_p_z[30,3]  -24.18    0.06 3.53  -31.85  -26.36  -23.88  -21.70  -18.11
#> log_p_z[30,4]  -21.92    0.07 3.48  -29.55  -24.04  -21.68  -19.48  -15.94
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_p_z[31,2]  -20.00    0.04 2.31  -25.04  -21.41  -19.79  -18.38  -15.97
#> log_p_z[31,3]  -24.68    0.07 4.29  -34.11  -27.18  -24.26  -21.69  -17.45
#> log_p_z[31,4]  -22.00    0.08 4.26  -31.85  -24.52  -21.56  -19.01  -15.00
#> log_p_z[32,1]  -29.64    0.06 4.38  -39.66  -32.07  -29.14  -26.64  -22.44
#> log_p_z[32,2]  -14.79    0.02 1.72  -18.51  -15.88  -14.66  -13.58  -11.76
#> log_p_z[32,3]   -6.07    0.01 0.95   -8.22   -6.65   -5.99   -5.40   -4.44
#> log_p_z[32,4]  -20.83    0.08 4.48  -31.44  -23.39  -20.30  -17.65  -13.61
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_p_z[33,2]  -20.00    0.04 2.31  -25.04  -21.41  -19.79  -18.38  -15.97
#> log_p_z[33,3]  -24.68    0.07 4.29  -34.11  -27.18  -24.26  -21.69  -17.45
#> log_p_z[33,4]  -22.00    0.08 4.26  -31.85  -24.52  -21.56  -19.01  -15.00
#> log_p_z[34,1]  -19.64    0.04 2.45  -24.92  -21.16  -19.48  -17.90  -15.40
#> log_p_z[34,2]   -3.04    0.00 0.36   -3.79   -3.28   -3.02   -2.80   -2.38
#> log_p_z[34,3]  -21.26    0.08 4.33  -31.32  -23.75  -20.68  -18.28  -14.43
#> log_p_z[34,4]  -21.78    0.08 4.29  -31.39  -24.25  -21.35  -18.75  -14.83
#> log_p_z[35,1]  -22.63    0.04 2.74  -28.43  -24.36  -22.43  -20.65  -17.90
#> log_p_z[35,2]   -5.52    0.01 0.69   -7.03   -5.94   -5.46   -5.04   -4.31
#> log_p_z[35,3]  -16.14    0.08 4.14  -25.82  -18.47  -15.60  -13.11   -9.86
#> log_p_z[35,4]  -21.80    0.08 4.32  -31.83  -24.28  -21.20  -18.80  -14.70
#> log_p_z[36,1]  -29.69    0.06 3.59  -37.77  -31.79  -29.33  -27.22  -23.64
#> log_p_z[36,2]  -21.85    0.03 2.22  -26.85  -23.19  -21.72  -20.28  -17.99
#> log_p_z[36,3]  -10.50    0.04 1.71  -14.63  -11.31  -10.25   -9.33   -7.95
#> log_p_z[36,4]  -14.05    0.06 3.22  -21.17  -15.98  -13.77  -11.67   -8.85
#> log_p_z[37,1]  -18.61    0.04 2.37  -23.77  -20.08  -18.36  -16.90  -14.56
#> log_p_z[37,2]   -7.62    0.01 0.79   -9.34   -8.11   -7.56   -7.06   -6.27
#> log_p_z[37,3]  -19.15    0.06 3.27  -26.72  -21.11  -18.80  -16.83  -13.77
#> log_p_z[37,4]  -21.82    0.06 3.54  -29.81  -23.91  -21.48  -19.37  -15.90
#> log_p_z[38,1]  -25.21    0.04 2.90  -31.52  -26.98  -25.02  -23.17  -20.23
#> log_p_z[38,2]   -9.60    0.01 0.99  -11.76  -10.22   -9.53   -8.88   -7.86
#> log_p_z[38,3]  -11.74    0.05 2.80  -18.23  -13.37  -11.44   -9.60   -7.53
#> log_p_z[38,4]  -21.35    0.07 3.56  -29.24  -23.60  -20.97  -18.77  -15.40
#> log_p_z[39,1]  -29.24    0.06 4.29  -38.97  -31.65  -28.78  -26.36  -22.02
#> log_p_z[39,2]  -16.77    0.03 2.03  -21.20  -18.03  -16.62  -15.31  -13.22
#> log_p_z[39,3]   -7.28    0.02 1.12   -9.69   -7.99   -7.21   -6.50   -5.31
#> log_p_z[39,4]  -18.86    0.07 4.24  -28.70  -21.38  -18.34  -15.88  -12.13
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_p_z[40,2]  -20.00    0.04 2.31  -25.04  -21.41  -19.79  -18.38  -15.97
#> log_p_z[40,3]  -24.68    0.07 4.29  -34.11  -27.18  -24.26  -21.69  -17.45
#> log_p_z[40,4]  -22.00    0.08 4.26  -31.85  -24.52  -21.56  -19.01  -15.00
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_p_z[41,2]  -20.00    0.04 2.31  -25.04  -21.41  -19.79  -18.38  -15.97
#> log_p_z[41,3]  -24.68    0.07 4.29  -34.11  -27.18  -24.26  -21.69  -17.45
#> log_p_z[41,4]  -22.00    0.08 4.26  -31.85  -24.52  -21.56  -19.01  -15.00
#> log_p_z[42,1]   -6.00    0.01 0.62   -7.36   -6.38   -5.95   -5.57   -4.92
#> log_p_z[42,2]  -14.95    0.03 1.69  -18.62  -15.99  -14.82  -13.77  -12.01
#> log_p_z[42,3]  -24.20    0.06 3.54  -32.03  -26.36  -23.91  -21.64  -18.31
#> log_p_z[42,4]  -21.90    0.07 3.55  -29.76  -24.07  -21.55  -19.43  -15.89
#> log_p_z[43,1]  -21.86    0.04 2.55  -27.58  -23.44  -21.66  -20.07  -17.51
#> log_p_z[43,2]   -5.64    0.01 0.53   -6.80   -5.97   -5.61   -5.26   -4.72
#> log_p_z[43,3]  -18.42    0.06 3.18  -25.69  -20.32  -18.06  -16.21  -13.24
#> log_p_z[43,4]  -21.52    0.06 3.49  -29.22  -23.55  -21.28  -19.10  -15.68
#> log_p_z[44,1]   -4.42    0.01 0.41   -5.30   -4.67   -4.38   -4.13   -3.71
#> log_p_z[44,2]  -17.24    0.03 1.85  -21.26  -18.34  -17.09  -15.95  -13.93
#> log_p_z[44,3]  -24.17    0.06 3.54  -31.92  -26.36  -23.96  -21.68  -18.01
#> log_p_z[44,4]  -21.96    0.07 3.48  -29.45  -24.14  -21.71  -19.49  -16.00
#> log_p_z[45,1]  -19.64    0.04 2.45  -24.92  -21.16  -19.48  -17.90  -15.40
#> log_p_z[45,2]   -3.04    0.00 0.36   -3.79   -3.28   -3.02   -2.80   -2.38
#> log_p_z[45,3]  -21.26    0.08 4.33  -31.32  -23.75  -20.68  -18.28  -14.43
#> log_p_z[45,4]  -21.78    0.08 4.29  -31.39  -24.25  -21.35  -18.75  -14.83
#> log_lik[1]      -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_lik[2]      -7.60    0.02 1.13  -10.05   -8.30   -7.52   -6.80   -5.61
#> log_lik[3]      -9.94    0.01 0.92  -11.86  -10.54   -9.90   -9.29   -8.25
#> log_lik[4]      -7.57    0.01 0.94   -9.54   -8.15   -7.51   -6.91   -5.91
#> log_lik[5]      -3.71    0.01 0.42   -4.60   -3.96   -3.69   -3.42   -2.95
#> log_lik[6]      -5.47    0.01 0.65   -6.86   -5.87   -5.42   -5.00   -4.35
#> log_lik[7]      -6.00    0.01 0.62   -7.36   -6.38   -5.95   -5.57   -4.92
#> log_lik[8]      -5.73    0.01 0.84   -7.51   -6.27   -5.67   -5.15   -4.23
#> log_lik[9]      -4.52    0.01 0.57   -5.77   -4.88   -4.48   -4.11   -3.53
#> log_lik[10]     -7.12    0.01 0.71   -8.66   -7.56   -7.06   -6.61   -5.86
#> log_lik[11]     -5.81    0.01 0.97   -7.88   -6.42   -5.74   -5.13   -4.14
#> log_lik[12]     -9.32    0.01 1.03  -11.54   -9.97   -9.26   -8.61   -7.47
#> log_lik[13]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_lik[14]     -6.22    0.01 0.86   -8.15   -6.72   -6.15   -5.62   -4.75
#> log_lik[15]     -4.42    0.01 0.41   -5.30   -4.67   -4.38   -4.13   -3.71
#> log_lik[16]     -3.79    0.01 0.55   -5.00   -4.13   -3.75   -3.42   -2.83
#> log_lik[17]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_lik[18]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_lik[19]     -4.67    0.01 0.59   -5.91   -5.04   -4.63   -4.26   -3.63
#> log_lik[20]     -7.09    0.01 0.95   -9.21   -7.67   -7.01   -6.43   -5.48
#> log_lik[21]     -3.04    0.00 0.36   -3.79   -3.28   -3.02   -2.80   -2.38
#> log_lik[22]     -4.67    0.01 0.59   -5.91   -5.04   -4.63   -4.26   -3.63
#> log_lik[23]     -3.71    0.01 0.42   -4.60   -3.96   -3.69   -3.42   -2.95
#> log_lik[24]     -5.81    0.01 0.64   -7.29   -6.19   -5.74   -5.36   -4.74
#> log_lik[25]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_lik[26]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_lik[27]     -5.64    0.01 0.53   -6.80   -5.97   -5.61   -5.26   -4.72
#> log_lik[28]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_lik[29]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_lik[30]     -6.82    0.01 0.85   -8.78   -7.32   -6.73   -6.24   -5.45
#> log_lik[31]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_lik[32]     -6.07    0.01 0.95   -8.22   -6.65   -5.99   -5.40   -4.44
#> log_lik[33]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_lik[34]     -3.04    0.00 0.36   -3.79   -3.28   -3.02   -2.80   -2.38
#> log_lik[35]     -5.52    0.01 0.69   -7.00   -5.94   -5.46   -5.04   -4.31
#> log_lik[36]     -9.94    0.02 1.15  -12.35  -10.68   -9.86   -9.12   -7.83
#> log_lik[37]     -7.62    0.01 0.79   -9.34   -8.11   -7.56   -7.06   -6.27
#> log_lik[38]     -9.00    0.01 0.87  -10.73   -9.57   -8.98   -8.40   -7.33
#> log_lik[39]     -7.28    0.02 1.12   -9.67   -7.99   -7.21   -6.50   -5.31
#> log_lik[40]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_lik[41]     -2.21    0.00 0.30   -2.84   -2.41   -2.19   -2.00   -1.69
#> log_lik[42]     -6.00    0.01 0.62   -7.36   -6.38   -5.95   -5.57   -4.92
#> log_lik[43]     -5.64    0.01 0.53   -6.80   -5.97   -5.61   -5.26   -4.72
#> log_lik[44]     -4.42    0.01 0.41   -5.30   -4.67   -4.38   -4.13   -3.71
#> log_lik[45]     -3.04    0.00 0.36   -3.79   -3.28   -3.02   -2.80   -2.38
#> lp__          -463.67    0.16 6.11 -476.78 -467.49 -463.32 -459.50 -452.50
#>               n_eff Rhat
#> pi[1]          6408    1
#> pi[2]          5835    1
#> pi[3]          6009    1
#> pi[4]          6072    1
#> theta[1,1,1]   6611    1
#> theta[1,1,2]   6641    1
#> theta[1,1,3]   4622    1
#> theta[1,1,4]   5008    1
#> theta[1,2,1]   5834    1
#> theta[1,2,2]   7314    1
#> theta[1,2,3]   6548    1
#> theta[1,2,4]   5668    1
#> theta[1,3,1]   6305    1
#> theta[1,3,2]   3860    1
#> theta[1,3,3]   4664    1
#> theta[1,3,4]   4683    1
#> theta[1,4,1]   6558    1
#> theta[1,4,2]   5592    1
#> theta[1,4,3]   4101    1
#> theta[1,4,4]   4940    1
#> theta[2,1,1]   7169    1
#> theta[2,1,2]   6389    1
#> theta[2,1,3]   5231    1
#> theta[2,1,4]   4691    1
#> theta[2,2,1]   6169    1
#> theta[2,2,2]   6528    1
#> theta[2,2,3]   7181    1
#> theta[2,2,4]   5426    1
#> theta[2,3,1]   5984    1
#> theta[2,3,2]   6842    1
#> theta[2,3,3]   6487    1
#> theta[2,3,4]   5564    1
#> theta[2,4,1]   6401    1
#> theta[2,4,2]   5443    1
#> theta[2,4,3]   6263    1
#> theta[2,4,4]   6010    1
#> theta[3,1,1]   6537    1
#> theta[3,1,2]   4386    1
#> theta[3,1,3]   5103    1
#> theta[3,1,4]   4290    1
#> theta[3,2,1]   6418    1
#> theta[3,2,2]   7798    1
#> theta[3,2,3]   6862    1
#> theta[3,2,4]   4893    1
#> theta[3,3,1]   5236    1
#> theta[3,3,2]   6294    1
#> theta[3,3,3]   5845    1
#> theta[3,3,4]   5925    1
#> theta[3,4,1]   5897    1
#> theta[3,4,2]   7088    1
#> theta[3,4,3]   4924    1
#> theta[3,4,4]   6286    1
#> theta[4,1,1]   6839    1
#> theta[4,1,2]   4940    1
#> theta[4,1,3]   4699    1
#> theta[4,1,4]   4796    1
#> theta[4,2,1]   6763    1
#> theta[4,2,2]   6208    1
#> theta[4,2,3]   5682    1
#> theta[4,2,4]   5896    1
#> theta[4,3,1]   5648    1
#> theta[4,3,2]   5929    1
#> theta[4,3,3]   5737    1
#> theta[4,3,4]   5119    1
#> theta[4,4,1]   6345    1
#> theta[4,4,2]   5602    1
#> theta[4,4,3]   5675    1
#> theta[4,4,4]   5665    1
#> theta[5,1,1]   5489    1
#> theta[5,1,2]   5304    1
#> theta[5,1,3]   4560    1
#> theta[5,1,4]   4681    1
#> theta[5,2,1]   7637    1
#> theta[5,2,2]   6792    1
#> theta[5,2,3]   6221    1
#> theta[5,2,4]   5132    1
#> theta[5,3,1]   6626    1
#> theta[5,3,2]   7156    1
#> theta[5,3,3]   6635    1
#> theta[5,3,4]   6487    1
#> theta[5,4,1]   5415    1
#> theta[5,4,2]   6793    1
#> theta[5,4,3]   5909    1
#> theta[5,4,4]   6591    1
#> log_p_z[1,1]   6325    1
#> log_p_z[1,2]   3791    1
#> log_p_z[1,3]   3709    1
#> log_p_z[1,4]   2838    1
#> log_p_z[2,1]   4621    1
#> log_p_z[2,2]   4522    1
#> log_p_z[2,3]   4103    1
#> log_p_z[2,4]   2957    1
#> log_p_z[3,1]   3843    1
#> log_p_z[3,2]   3675    1
#> log_p_z[3,3]   3380    1
#> log_p_z[3,4]   3008    1
#> log_p_z[4,1]   5034    1
#> log_p_z[4,2]   4713    1
#> log_p_z[4,3]   2993    1
#> log_p_z[4,4]   3189    1
#> log_p_z[5,1]   4545    1
#> log_p_z[5,2]   6331    1
#> log_p_z[5,3]   2976    1
#> log_p_z[5,4]   3000    1
#> log_p_z[6,1]   4501    1
#> log_p_z[6,2]   5996    1
#> log_p_z[6,3]   3051    1
#> log_p_z[6,4]   2816    1
#> log_p_z[7,1]   6261    1
#> log_p_z[7,2]   3905    1
#> log_p_z[7,3]   3341    1
#> log_p_z[7,4]   2656    1
#> log_p_z[8,1]   5095    1
#> log_p_z[8,2]   5035    1
#> log_p_z[8,3]   5345    1
#> log_p_z[8,4]   2854    1
#> log_p_z[9,1]   5006    1
#> log_p_z[9,2]   5071    1
#> log_p_z[9,3]   3113    1
#> log_p_z[9,4]   3116    1
#> log_p_z[10,1]  4952    1
#> log_p_z[10,2]  5658    1
#> log_p_z[10,3]  3144    1
#> log_p_z[10,4]  3156    1
#> log_p_z[11,1]  4055    1
#> log_p_z[11,2]  4113    1
#> log_p_z[11,3]  2086    1
#> log_p_z[11,4]  4978    1
#> log_p_z[12,1]  5023    1
#> log_p_z[12,2]  4261    1
#> log_p_z[12,3]  3007    1
#> log_p_z[12,4]  2856    1
#> log_p_z[13,1]  6325    1
#> log_p_z[13,2]  3791    1
#> log_p_z[13,3]  3709    1
#> log_p_z[13,4]  2838    1
#> log_p_z[14,1]  4536    1
#> log_p_z[14,2]  5427    1
#> log_p_z[14,3]  2955    1
#> log_p_z[14,4]  2943    1
#> log_p_z[15,1]  5745    1
#> log_p_z[15,2]  3774    1
#> log_p_z[15,3]  3511    1
#> log_p_z[15,4]  2653    1
#> log_p_z[16,1]  6397    1
#> log_p_z[16,2]  3859    1
#> log_p_z[16,3]  3596    1
#> log_p_z[16,4]  2828    1
#> log_p_z[17,1]  6325    1
#> log_p_z[17,2]  3791    1
#> log_p_z[17,3]  3709    1
#> log_p_z[17,4]  2838    1
#> log_p_z[18,1]  6325    1
#> log_p_z[18,2]  3791    1
#> log_p_z[18,3]  3709    1
#> log_p_z[18,4]  2838    1
#> log_p_z[19,1]  4728    1
#> log_p_z[19,2]  6133    1
#> log_p_z[19,3]  3119    1
#> log_p_z[19,4]  2905    1
#> log_p_z[20,1]  4451    1
#> log_p_z[20,2]  4625    1
#> log_p_z[20,3]  3243    1
#> log_p_z[20,4]  2741    1
#> log_p_z[21,1]  4416    1
#> log_p_z[21,2]  7083    1
#> log_p_z[21,3]  3050    1
#> log_p_z[21,4]  2994    1
#> log_p_z[22,1]  4728    1
#> log_p_z[22,2]  6133    1
#> log_p_z[22,3]  3119    1
#> log_p_z[22,4]  2905    1
#> log_p_z[23,1]  4545    1
#> log_p_z[23,2]  6331    1
#> log_p_z[23,3]  2976    1
#> log_p_z[23,4]  3000    1
#> log_p_z[24,1]  4277    1
#> log_p_z[24,2]  3977    1
#> log_p_z[24,3]  3091    1
#> log_p_z[24,4]  2893    1
#> log_p_z[25,1]  6325    1
#> log_p_z[25,2]  3791    1
#> log_p_z[25,3]  3709    1
#> log_p_z[25,4]  2838    1
#> log_p_z[26,1]  6325    1
#> log_p_z[26,2]  3791    1
#> log_p_z[26,3]  3709    1
#> log_p_z[26,4]  2838    1
#> log_p_z[27,1]  4273    1
#> log_p_z[27,2]  6280    1
#> log_p_z[27,3]  3079    1
#> log_p_z[27,4]  3211    1
#> log_p_z[28,1]  6325    1
#> log_p_z[28,2]  3791    1
#> log_p_z[28,3]  3709    1
#> log_p_z[28,4]  2838    1
#> log_p_z[29,1]  6325    1
#> log_p_z[29,2]  3791    1
#> log_p_z[29,3]  3709    1
#> log_p_z[29,4]  2838    1
#> log_p_z[30,1]  4491    1
#> log_p_z[30,2]  3719    1
#> log_p_z[30,3]  3416    1
#> log_p_z[30,4]  2809    1
#> log_p_z[31,1]  6325    1
#> log_p_z[31,2]  3791    1
#> log_p_z[31,3]  3709    1
#> log_p_z[31,4]  2838    1
#> log_p_z[32,1]  5217    1
#> log_p_z[32,2]  5713    1
#> log_p_z[32,3]  5031    1
#> log_p_z[32,4]  2959    1
#> log_p_z[33,1]  6325    1
#> log_p_z[33,2]  3791    1
#> log_p_z[33,3]  3709    1
#> log_p_z[33,4]  2838    1
#> log_p_z[34,1]  4416    1
#> log_p_z[34,2]  7083    1
#> log_p_z[34,3]  3050    1
#> log_p_z[34,4]  2994    1
#> log_p_z[35,1]  4422    1
#> log_p_z[35,2]  5470    1
#> log_p_z[35,3]  2964    1
#> log_p_z[35,4]  2985    1
#> log_p_z[36,1]  3848    1
#> log_p_z[36,2]  4278    1
#> log_p_z[36,3]  1870    1
#> log_p_z[36,4]  2571    1
#> log_p_z[37,1]  4251    1
#> log_p_z[37,2]  4764    1
#> log_p_z[37,3]  3006    1
#> log_p_z[37,4]  2999    1
#> log_p_z[38,1]  5014    1
#> log_p_z[38,2]  5148    1
#> log_p_z[38,3]  3041    1
#> log_p_z[38,4]  2826    1
#> log_p_z[39,1]  4695    1
#> log_p_z[39,2]  4941    1
#> log_p_z[39,3]  4724    1
#> log_p_z[39,4]  3201    1
#> log_p_z[40,1]  6325    1
#> log_p_z[40,2]  3791    1
#> log_p_z[40,3]  3709    1
#> log_p_z[40,4]  2838    1
#> log_p_z[41,1]  6325    1
#> log_p_z[41,2]  3791    1
#> log_p_z[41,3]  3709    1
#> log_p_z[41,4]  2838    1
#> log_p_z[42,1]  6261    1
#> log_p_z[42,2]  3905    1
#> log_p_z[42,3]  3341    1
#> log_p_z[42,4]  2656    1
#> log_p_z[43,1]  4273    1
#> log_p_z[43,2]  6280    1
#> log_p_z[43,3]  3079    1
#> log_p_z[43,4]  3211    1
#> log_p_z[44,1]  5745    1
#> log_p_z[44,2]  3774    1
#> log_p_z[44,3]  3511    1
#> log_p_z[44,4]  2653    1
#> log_p_z[45,1]  4416    1
#> log_p_z[45,2]  7083    1
#> log_p_z[45,3]  3050    1
#> log_p_z[45,4]  2994    1
#> log_lik[1]     6325    1
#> log_lik[2]     4635    1
#> log_lik[3]     4853    1
#> log_lik[4]     4783    1
#> log_lik[5]     6332    1
#> log_lik[6]     6004    1
#> log_lik[7]     6262    1
#> log_lik[8]     5348    1
#> log_lik[9]     5071    1
#> log_lik[10]    5674    1
#> log_lik[11]    4978    1
#> log_lik[12]    5077    1
#> log_lik[13]    6325    1
#> log_lik[14]    5436    1
#> log_lik[15]    5745    1
#> log_lik[16]    6397    1
#> log_lik[17]    6325    1
#> log_lik[18]    6325    1
#> log_lik[19]    6134    1
#> log_lik[20]    4634    1
#> log_lik[21]    7083    1
#> log_lik[22]    6134    1
#> log_lik[23]    6332    1
#> log_lik[24]    3978    1
#> log_lik[25]    6325    1
#> log_lik[26]    6325    1
#> log_lik[27]    6280    1
#> log_lik[28]    6325    1
#> log_lik[29]    6325    1
#> log_lik[30]    4495    1
#> log_lik[31]    6325    1
#> log_lik[32]    5046    1
#> log_lik[33]    6325    1
#> log_lik[34]    7083    1
#> log_lik[35]    5492    1
#> log_lik[36]    4267    1
#> log_lik[37]    4769    1
#> log_lik[38]    4624    1
#> log_lik[39]    4739    1
#> log_lik[40]    6325    1
#> log_lik[41]    6325    1
#> log_lik[42]    6262    1
#> log_lik[43]    6280    1
#> log_lik[44]    5745    1
#> log_lik[45]    7083    1
#> lp__           1450    1
#> 
#> Samples were drawn using NUTS(diag_e) at Mon Feb 20 07:14:49 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }