Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.25    0.33    0.37    0.42    0.50
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.26
#> pi[4]            0.08    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.12    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.75    0.82    0.86    0.89    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.07    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.80    0.00 0.09    0.59    0.74    0.81    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.00    0.03    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.09    0.00    0.03    0.07    0.13    0.33
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.38
#> theta[1,4,4]     0.69    0.00 0.14    0.39    0.59    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.06    0.11    0.16    0.21    0.33
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.18
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.59    0.65    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.24    0.30    0.36    0.50
#> theta[2,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[2,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[2,3,3]     0.62    0.00 0.13    0.35    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.29    0.46
#> theta[2,4,1]     0.11    0.00 0.10    0.00    0.04    0.08    0.16    0.38
#> theta[2,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.55    0.67    0.77    0.90
#> theta[3,1,1]     0.88    0.00 0.06    0.74    0.84    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.02    0.05    0.09    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.28
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[3,3,2]     0.18    0.00 0.11    0.02    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.14    0.25    0.41    0.50    0.60    0.77
#> theta[3,3,4]     0.23    0.00 0.11    0.05    0.15    0.22    0.31    0.49
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[3,4,3]     0.14    0.00 0.12    0.01    0.05    0.11    0.20    0.43
#> theta[3,4,4]     0.62    0.00 0.15    0.31    0.52    0.63    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.07    0.67    0.78    0.84    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.08    0.54    0.67    0.73    0.79    0.87
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.09    0.13    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[4,3,3]     0.69    0.00 0.12    0.43    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.38
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[4,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.90
#> theta[5,1,1]     0.87    0.00 0.07    0.71    0.83    0.87    0.92    0.97
#> theta[5,1,2]     0.06    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.06    0.04    0.10    0.13    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.48    0.60    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.61    0.70    0.84
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[5,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.44
#> theta[5,4,4]     0.63    0.00 0.16    0.31    0.52    0.64    0.75    0.90
#> log_p_z[1,1]    -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_p_z[1,2]   -20.00    0.03 2.29  -25.02  -21.40  -19.79  -18.37  -16.11
#> log_p_z[1,3]   -24.84    0.08 4.37  -35.08  -27.26  -24.33  -21.82  -17.57
#> log_p_z[1,4]   -22.06    0.07 4.41  -32.16  -24.59  -21.56  -18.89  -15.03
#> log_p_z[2,1]   -29.75    0.06 4.44  -39.65  -32.33  -29.31  -26.61  -22.54
#> log_p_z[2,2]   -21.04    0.03 2.41  -26.21  -22.49  -20.91  -19.33  -16.81
#> log_p_z[2,3]    -7.61    0.02 1.18  -10.25   -8.31   -7.51   -6.78   -5.63
#> log_p_z[2,4]   -16.27    0.07 4.06  -25.43  -18.57  -15.85  -13.41   -9.78
#> log_p_z[3,1]   -11.63    0.03 1.66  -15.34  -12.62  -11.46  -10.44   -8.96
#> log_p_z[3,2]   -10.80    0.02 1.47  -14.18  -11.66  -10.60   -9.77   -8.46
#> log_p_z[3,3]   -23.42    0.06 3.44  -31.26  -25.47  -23.13  -20.99  -17.58
#> log_p_z[3,4]   -22.07    0.07 3.73  -30.53  -24.26  -21.70  -19.50  -15.67
#> log_p_z[4,1]   -14.68    0.03 1.88  -18.76  -15.84  -14.53  -13.36  -11.44
#> log_p_z[4,2]    -7.57    0.01 0.95   -9.62   -8.15   -7.49   -6.90   -5.98
#> log_p_z[4,3]   -20.80    0.09 4.38  -31.00  -23.15  -20.10  -17.74  -13.97
#> log_p_z[4,4]   -21.94    0.08 4.51  -32.34  -24.43  -21.36  -18.79  -14.69
#> log_p_z[5,1]   -21.46    0.04 2.59  -27.14  -23.02  -21.27  -19.63  -17.00
#> log_p_z[5,2]    -3.71    0.01 0.44   -4.62   -3.99   -3.68   -3.40   -2.94
#> log_p_z[5,3]   -18.81    0.08 4.31  -28.78  -21.27  -18.18  -15.72  -12.18
#> log_p_z[5,4]   -21.96    0.08 4.49  -32.20  -24.41  -21.43  -18.90  -14.72
#> log_p_z[6,1]   -21.51    0.04 2.57  -27.07  -23.20  -21.30  -19.65  -17.04
#> log_p_z[6,2]    -5.47    0.01 0.66   -6.85   -5.89   -5.43   -4.99   -4.33
#> log_p_z[6,3]   -17.59    0.08 4.16  -27.21  -19.92  -16.95  -14.57  -11.27
#> log_p_z[6,4]   -21.73    0.08 4.47  -32.24  -24.14  -21.10  -18.71  -14.58
#> log_p_z[7,1]    -6.00    0.01 0.61   -7.36   -6.38   -5.96   -5.57   -4.96
#> log_p_z[7,2]   -14.94    0.02 1.71  -18.63  -15.97  -14.80  -13.73  -12.06
#> log_p_z[7,3]   -24.37    0.06 3.54  -32.25  -26.52  -24.03  -21.93  -18.26
#> log_p_z[7,4]   -22.02    0.06 3.66  -30.23  -24.17  -21.69  -19.48  -15.93
#> log_p_z[8,1]   -29.74    0.06 4.47  -39.56  -32.41  -29.29  -26.52  -22.33
#> log_p_z[8,2]   -18.30    0.03 2.15  -23.01  -19.66  -18.11  -16.79  -14.62
#> log_p_z[8,3]    -5.71    0.01 0.84   -7.47   -6.25   -5.65   -5.11   -4.21
#> log_p_z[8,4]   -18.44    0.07 4.16  -27.92  -20.83  -17.94  -15.51  -11.69
#> log_p_z[9,1]   -20.03    0.03 2.32  -25.09  -21.42  -19.86  -18.44  -16.00
#> log_p_z[9,2]    -4.51    0.01 0.55   -5.73   -4.85   -4.48   -4.12   -3.54
#> log_p_z[9,3]   -19.77    0.09 4.39  -29.90  -22.28  -19.17  -16.59  -12.84
#> log_p_z[9,4]   -21.75    0.08 4.46  -32.44  -24.17  -21.20  -18.58  -14.75
#> log_p_z[10,1]  -22.27    0.04 2.47  -27.54  -23.80  -22.08  -20.55  -18.05
#> log_p_z[10,2]   -7.12    0.01 0.68   -8.60   -7.55   -7.08   -6.66   -5.92
#> log_p_z[10,3]  -16.92    0.06 3.23  -24.26  -18.82  -16.54  -14.64  -11.78
#> log_p_z[10,4]  -21.40    0.07 3.64  -29.64  -23.55  -20.98  -18.82  -15.32
#> log_p_z[11,1]  -29.86    0.06 4.32  -39.59  -32.39  -29.38  -26.77  -22.77
#> log_p_z[11,2]  -29.40    0.07 4.38  -39.44  -31.86  -28.91  -26.35  -22.29
#> log_p_z[11,3]  -18.30    0.07 3.48  -27.06  -20.02  -17.76  -15.93  -13.08
#> log_p_z[11,4]   -5.82    0.01 1.00   -8.02   -6.44   -5.74   -5.12   -4.08
#> log_p_z[12,1]  -23.04    0.04 2.60  -28.76  -24.62  -22.88  -21.26  -18.49
#> log_p_z[12,2]   -9.60    0.02 1.28  -12.65  -10.30   -9.48   -8.71   -7.48
#> log_p_z[12,3]  -15.23    0.08 4.06  -24.79  -17.52  -14.65  -12.34   -8.94
#> log_p_z[12,4]  -19.27    0.08 4.32  -29.59  -21.58  -18.72  -16.32  -12.50
#> log_p_z[13,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_p_z[13,2]  -20.00    0.03 2.29  -25.02  -21.40  -19.79  -18.37  -16.11
#> log_p_z[13,3]  -24.84    0.08 4.37  -35.08  -27.26  -24.33  -21.82  -17.57
#> log_p_z[13,4]  -22.06    0.07 4.41  -32.16  -24.59  -21.56  -18.89  -15.03
#> log_p_z[14,1]  -19.04    0.03 2.42  -24.49  -20.57  -18.83  -17.29  -14.92
#> log_p_z[14,2]   -6.23    0.01 0.89   -8.21   -6.77   -6.16   -5.60   -4.72
#> log_p_z[14,3]  -18.84    0.08 4.32  -29.11  -21.22  -18.26  -15.72  -12.14
#> log_p_z[14,4]  -21.92    0.08 4.39  -32.15  -24.33  -21.38  -18.87  -14.97
#> log_p_z[15,1]   -4.41    0.01 0.40   -5.28   -4.66   -4.38   -4.13   -3.73
#> log_p_z[15,2]  -17.24    0.03 1.85  -21.31  -18.39  -17.11  -15.90  -14.07
#> log_p_z[15,3]  -24.34    0.06 3.57  -32.47  -26.42  -24.06  -21.80  -18.21
#> log_p_z[15,4]  -22.03    0.06 3.59  -30.03  -24.22  -21.69  -19.46  -16.06
#> log_p_z[16,1]   -3.79    0.01 0.53   -4.92   -4.13   -3.75   -3.42   -2.88
#> log_p_z[16,2]  -17.71    0.03 2.19  -22.57  -19.01  -17.52  -16.12  -14.06
#> log_p_z[16,3]  -24.88    0.08 4.34  -35.15  -27.37  -24.36  -21.87  -17.77
#> log_p_z[16,4]  -22.05    0.08 4.49  -32.40  -24.60  -21.57  -18.86  -14.89
#> log_p_z[17,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_p_z[17,2]  -20.00    0.03 2.29  -25.02  -21.40  -19.79  -18.37  -16.11
#> log_p_z[17,3]  -24.84    0.08 4.37  -35.08  -27.26  -24.33  -21.82  -17.57
#> log_p_z[17,4]  -22.06    0.07 4.41  -32.16  -24.59  -21.56  -18.89  -15.03
#> log_p_z[18,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_p_z[18,2]  -20.00    0.03 2.29  -25.02  -21.40  -19.79  -18.37  -16.11
#> log_p_z[18,3]  -24.84    0.08 4.37  -35.08  -27.26  -24.33  -21.82  -17.57
#> log_p_z[18,4]  -22.06    0.07 4.41  -32.16  -24.59  -21.56  -18.89  -15.03
#> log_p_z[19,1]  -16.41    0.03 1.94  -20.59  -17.63  -16.28  -15.04  -13.06
#> log_p_z[19,2]   -4.66    0.01 0.57   -5.92   -5.01   -4.62   -4.27   -3.64
#> log_p_z[19,3]  -22.27    0.09 4.54  -32.82  -24.81  -21.68  -19.02  -15.11
#> log_p_z[19,4]  -22.00    0.08 4.48  -32.24  -24.42  -21.45  -18.91  -14.74
#> log_p_z[20,1]  -18.08    0.03 2.28  -23.11  -19.55  -17.88  -16.47  -14.15
#> log_p_z[20,2]   -7.09    0.01 0.94   -9.11   -7.67   -7.02   -6.42   -5.48
#> log_p_z[20,3]  -20.08    0.08 4.30  -29.92  -22.52  -19.46  -17.03  -13.42
#> log_p_z[20,4]  -21.79    0.08 4.47  -32.11  -24.18  -21.19  -18.77  -14.69
#> log_p_z[21,1]  -19.62    0.03 2.38  -24.59  -21.11  -19.45  -17.96  -15.59
#> log_p_z[21,2]   -3.04    0.00 0.35   -3.76   -3.27   -3.02   -2.80   -2.39
#> log_p_z[21,3]  -21.34    0.08 4.42  -31.47  -23.79  -20.78  -18.16  -14.37
#> log_p_z[21,4]  -22.01    0.08 4.48  -32.64  -24.32  -21.52  -18.92  -15.01
#> log_p_z[22,1]  -16.41    0.03 1.94  -20.59  -17.63  -16.28  -15.04  -13.06
#> log_p_z[22,2]   -4.66    0.01 0.57   -5.92   -5.01   -4.62   -4.27   -3.64
#> log_p_z[22,3]  -22.27    0.09 4.54  -32.82  -24.81  -21.68  -19.02  -15.11
#> log_p_z[22,4]  -22.00    0.08 4.48  -32.24  -24.42  -21.45  -18.91  -14.74
#> log_p_z[23,1]  -21.46    0.04 2.59  -27.14  -23.02  -21.27  -19.63  -17.00
#> log_p_z[23,2]   -3.71    0.01 0.44   -4.62   -3.99   -3.68   -3.40   -2.94
#> log_p_z[23,3]  -18.81    0.08 4.31  -28.78  -21.27  -18.18  -15.72  -12.18
#> log_p_z[23,4]  -21.96    0.08 4.49  -32.20  -24.41  -21.43  -18.90  -14.72
#> log_p_z[24,1]  -17.41    0.03 2.17  -22.02  -18.74  -17.24  -15.89  -13.81
#> log_p_z[24,2]   -5.80    0.01 0.62   -7.14   -6.18   -5.74   -5.37   -4.76
#> log_p_z[24,3]  -21.85    0.06 3.47  -29.28  -23.97  -21.50  -19.41  -15.98
#> log_p_z[24,4]  -22.05    0.07 3.70  -30.36  -24.15  -21.71  -19.47  -15.93
#> log_p_z[25,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_p_z[25,2]  -20.00    0.03 2.29  -25.02  -21.40  -19.79  -18.37  -16.11
#> log_p_z[25,3]  -24.84    0.08 4.37  -35.08  -27.26  -24.33  -21.82  -17.57
#> log_p_z[25,4]  -22.06    0.07 4.41  -32.16  -24.59  -21.56  -18.89  -15.03
#> log_p_z[26,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_p_z[26,2]  -20.00    0.03 2.29  -25.02  -21.40  -19.79  -18.37  -16.11
#> log_p_z[26,3]  -24.84    0.08 4.37  -35.08  -27.26  -24.33  -21.82  -17.57
#> log_p_z[26,4]  -22.06    0.07 4.41  -32.16  -24.59  -21.56  -18.89  -15.03
#> log_p_z[27,1]  -21.86    0.04 2.50  -27.41  -23.39  -21.71  -20.11  -17.51
#> log_p_z[27,2]   -5.65    0.01 0.52   -6.77   -5.97   -5.62   -5.29   -4.73
#> log_p_z[27,3]  -18.49    0.06 3.29  -25.84  -20.43  -18.15  -16.16  -13.17
#> log_p_z[27,4]  -21.66    0.07 3.62  -29.82  -23.79  -21.33  -19.14  -15.65
#> log_p_z[28,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_p_z[28,2]  -20.00    0.03 2.29  -25.02  -21.40  -19.79  -18.37  -16.11
#> log_p_z[28,3]  -24.84    0.08 4.37  -35.08  -27.26  -24.33  -21.82  -17.57
#> log_p_z[28,4]  -22.06    0.07 4.41  -32.16  -24.59  -21.56  -18.89  -15.03
#> log_p_z[29,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_p_z[29,2]  -20.00    0.03 2.29  -25.02  -21.40  -19.79  -18.37  -16.11
#> log_p_z[29,3]  -24.84    0.08 4.37  -35.08  -27.26  -24.33  -21.82  -17.57
#> log_p_z[29,4]  -22.06    0.07 4.41  -32.16  -24.59  -21.56  -18.89  -15.03
#> log_p_z[30,1]   -6.83    0.01 0.84   -8.79   -7.33   -6.73   -6.23   -5.43
#> log_p_z[30,2]  -14.72    0.02 1.69  -18.56  -15.74  -14.52  -13.54  -11.75
#> log_p_z[30,3]  -24.31    0.06 3.61  -32.48  -26.48  -24.02  -21.78  -18.24
#> log_p_z[30,4]  -22.07    0.06 3.69  -30.38  -24.30  -21.66  -19.50  -15.80
#> log_p_z[31,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_p_z[31,2]  -20.00    0.03 2.29  -25.02  -21.40  -19.79  -18.37  -16.11
#> log_p_z[31,3]  -24.84    0.08 4.37  -35.08  -27.26  -24.33  -21.82  -17.57
#> log_p_z[31,4]  -22.06    0.07 4.41  -32.16  -24.59  -21.56  -18.89  -15.03
#> log_p_z[32,1]  -29.69    0.06 4.46  -39.67  -32.25  -29.19  -26.54  -22.40
#> log_p_z[32,2]  -14.84    0.02 1.72  -18.69  -15.92  -14.69  -13.63  -11.89
#> log_p_z[32,3]   -6.07    0.01 0.94   -8.18   -6.65   -5.99   -5.41   -4.46
#> log_p_z[32,4]  -20.58    0.07 4.33  -30.31  -23.10  -20.13  -17.49  -13.56
#> log_p_z[33,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_p_z[33,2]  -20.00    0.03 2.29  -25.02  -21.40  -19.79  -18.37  -16.11
#> log_p_z[33,3]  -24.84    0.08 4.37  -35.08  -27.26  -24.33  -21.82  -17.57
#> log_p_z[33,4]  -22.06    0.07 4.41  -32.16  -24.59  -21.56  -18.89  -15.03
#> log_p_z[34,1]  -19.62    0.03 2.38  -24.59  -21.11  -19.45  -17.96  -15.59
#> log_p_z[34,2]   -3.04    0.00 0.35   -3.76   -3.27   -3.02   -2.80   -2.39
#> log_p_z[34,3]  -21.34    0.08 4.42  -31.47  -23.79  -20.78  -18.16  -14.37
#> log_p_z[34,4]  -22.01    0.08 4.48  -32.64  -24.32  -21.52  -18.92  -15.01
#> log_p_z[35,1]  -22.56    0.04 2.69  -28.32  -24.27  -22.39  -20.65  -17.99
#> log_p_z[35,2]   -5.53    0.01 0.71   -7.07   -5.96   -5.47   -5.03   -4.27
#> log_p_z[35,3]  -16.21    0.08 4.13  -26.15  -18.52  -15.61  -13.26  -10.03
#> log_p_z[35,4]  -21.90    0.08 4.40  -32.01  -24.37  -21.38  -18.82  -14.92
#> log_p_z[36,1]  -29.81    0.05 3.65  -37.98  -32.05  -29.56  -27.21  -23.52
#> log_p_z[36,2]  -21.98    0.04 2.37  -27.15  -23.43  -21.80  -20.29  -17.87
#> log_p_z[36,3]  -10.54    0.04 1.75  -14.75  -11.44  -10.23   -9.33   -8.00
#> log_p_z[36,4]  -13.90    0.06 3.17  -20.71  -15.87  -13.70  -11.56   -8.74
#> log_p_z[37,1]  -18.51    0.03 2.26  -23.42  -19.92  -18.32  -16.93  -14.66
#> log_p_z[37,2]   -7.62    0.01 0.80   -9.34   -8.11   -7.55   -7.05   -6.27
#> log_p_z[37,3]  -19.24    0.06 3.24  -26.32  -21.22  -18.89  -16.99  -13.81
#> log_p_z[37,4]  -21.99    0.07 3.61  -29.91  -24.09  -21.73  -19.45  -15.95
#> log_p_z[38,1]  -25.21    0.04 2.84  -31.35  -26.91  -24.98  -23.17  -20.37
#> log_p_z[38,2]   -9.61    0.01 1.00  -11.73  -10.24   -9.52   -8.91   -7.89
#> log_p_z[38,3]  -11.78    0.06 2.82  -18.54  -13.33  -11.42   -9.74   -7.57
#> log_p_z[38,4]  -21.29    0.06 3.60  -29.40  -23.49  -20.91  -18.72  -15.24
#> log_p_z[39,1]  -29.33    0.06 4.42  -39.37  -31.91  -28.92  -26.24  -21.93
#> log_p_z[39,2]  -16.83    0.03 2.08  -21.35  -18.09  -16.68  -15.37  -13.21
#> log_p_z[39,3]   -7.28    0.02 1.13   -9.78   -7.98   -7.22   -6.47   -5.31
#> log_p_z[39,4]  -18.70    0.07 4.06  -27.95  -20.94  -18.21  -15.87  -12.26
#> log_p_z[40,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_p_z[40,2]  -20.00    0.03 2.29  -25.02  -21.40  -19.79  -18.37  -16.11
#> log_p_z[40,3]  -24.84    0.08 4.37  -35.08  -27.26  -24.33  -21.82  -17.57
#> log_p_z[40,4]  -22.06    0.07 4.41  -32.16  -24.59  -21.56  -18.89  -15.03
#> log_p_z[41,1]   -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_p_z[41,2]  -20.00    0.03 2.29  -25.02  -21.40  -19.79  -18.37  -16.11
#> log_p_z[41,3]  -24.84    0.08 4.37  -35.08  -27.26  -24.33  -21.82  -17.57
#> log_p_z[41,4]  -22.06    0.07 4.41  -32.16  -24.59  -21.56  -18.89  -15.03
#> log_p_z[42,1]   -6.00    0.01 0.61   -7.36   -6.38   -5.96   -5.57   -4.96
#> log_p_z[42,2]  -14.94    0.02 1.71  -18.63  -15.97  -14.80  -13.73  -12.06
#> log_p_z[42,3]  -24.37    0.06 3.54  -32.25  -26.52  -24.03  -21.93  -18.26
#> log_p_z[42,4]  -22.02    0.06 3.66  -30.23  -24.17  -21.69  -19.48  -15.93
#> log_p_z[43,1]  -21.86    0.04 2.50  -27.41  -23.39  -21.71  -20.11  -17.51
#> log_p_z[43,2]   -5.65    0.01 0.52   -6.77   -5.97   -5.62   -5.29   -4.73
#> log_p_z[43,3]  -18.49    0.06 3.29  -25.84  -20.43  -18.15  -16.16  -13.17
#> log_p_z[43,4]  -21.66    0.07 3.62  -29.82  -23.79  -21.33  -19.14  -15.65
#> log_p_z[44,1]   -4.41    0.01 0.40   -5.28   -4.66   -4.38   -4.13   -3.73
#> log_p_z[44,2]  -17.24    0.03 1.85  -21.31  -18.39  -17.11  -15.90  -14.07
#> log_p_z[44,3]  -24.34    0.06 3.57  -32.47  -26.42  -24.06  -21.80  -18.21
#> log_p_z[44,4]  -22.03    0.06 3.59  -30.03  -24.22  -21.69  -19.46  -16.06
#> log_p_z[45,1]  -19.62    0.03 2.38  -24.59  -21.11  -19.45  -17.96  -15.59
#> log_p_z[45,2]   -3.04    0.00 0.35   -3.76   -3.27   -3.02   -2.80   -2.39
#> log_p_z[45,3]  -21.34    0.08 4.42  -31.47  -23.79  -20.78  -18.16  -14.37
#> log_p_z[45,4]  -22.01    0.08 4.48  -32.64  -24.32  -21.52  -18.92  -15.01
#> log_lik[1]      -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_lik[2]      -7.57    0.02 1.10   -9.92   -8.28   -7.49   -6.78   -5.63
#> log_lik[3]      -9.95    0.01 0.90  -11.88  -10.53   -9.90   -9.33   -8.30
#> log_lik[4]      -7.56    0.01 0.93   -9.55   -8.14   -7.49   -6.90   -5.98
#> log_lik[5]      -3.71    0.01 0.44   -4.62   -3.99   -3.68   -3.40   -2.94
#> log_lik[6]      -5.46    0.01 0.66   -6.84   -5.89   -5.43   -4.99   -4.33
#> log_lik[7]      -6.00    0.01 0.61   -7.36   -6.38   -5.96   -5.57   -4.96
#> log_lik[8]      -5.71    0.01 0.84   -7.47   -6.25   -5.65   -5.11   -4.21
#> log_lik[9]      -4.51    0.01 0.55   -5.73   -4.85   -4.48   -4.12   -3.54
#> log_lik[10]     -7.12    0.01 0.68   -8.59   -7.55   -7.08   -6.66   -5.92
#> log_lik[11]     -5.82    0.01 1.00   -8.02   -6.44   -5.74   -5.12   -4.08
#> log_lik[12]     -9.34    0.01 1.04  -11.45  -10.01   -9.30   -8.62   -7.45
#> log_lik[13]     -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_lik[14]     -6.23    0.01 0.89   -8.21   -6.77   -6.16   -5.60   -4.72
#> log_lik[15]     -4.41    0.01 0.40   -5.28   -4.66   -4.38   -4.13   -3.73
#> log_lik[16]     -3.79    0.01 0.53   -4.92   -4.13   -3.75   -3.42   -2.88
#> log_lik[17]     -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_lik[18]     -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_lik[19]     -4.66    0.01 0.57   -5.92   -5.01   -4.62   -4.27   -3.64
#> log_lik[20]     -7.09    0.01 0.94   -9.11   -7.67   -7.02   -6.42   -5.48
#> log_lik[21]     -3.04    0.00 0.35   -3.76   -3.27   -3.02   -2.80   -2.39
#> log_lik[22]     -4.66    0.01 0.57   -5.92   -5.01   -4.62   -4.27   -3.64
#> log_lik[23]     -3.71    0.01 0.44   -4.62   -3.99   -3.68   -3.40   -2.94
#> log_lik[24]     -5.80    0.01 0.62   -7.14   -6.18   -5.74   -5.37   -4.76
#> log_lik[25]     -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_lik[26]     -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_lik[27]     -5.65    0.01 0.52   -6.77   -5.97   -5.62   -5.29   -4.73
#> log_lik[28]     -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_lik[29]     -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_lik[30]     -6.83    0.01 0.84   -8.78   -7.33   -6.73   -6.23   -5.43
#> log_lik[31]     -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_lik[32]     -6.07    0.01 0.94   -8.17   -6.65   -5.99   -5.41   -4.46
#> log_lik[33]     -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_lik[34]     -3.04    0.00 0.35   -3.76   -3.27   -3.02   -2.80   -2.39
#> log_lik[35]     -5.52    0.01 0.70   -7.05   -5.96   -5.47   -5.03   -4.27
#> log_lik[36]     -9.92    0.02 1.18  -12.41  -10.68   -9.84   -9.08   -7.87
#> log_lik[37]     -7.62    0.01 0.80   -9.34   -8.11   -7.55   -7.05   -6.27
#> log_lik[38]     -9.02    0.01 0.86  -10.71   -9.59   -9.02   -8.42   -7.39
#> log_lik[39]     -7.28    0.02 1.13   -9.78   -7.98   -7.22   -6.47   -5.31
#> log_lik[40]     -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_lik[41]     -2.20    0.00 0.29   -2.81   -2.39   -2.18   -2.00   -1.68
#> log_lik[42]     -6.00    0.01 0.61   -7.36   -6.38   -5.96   -5.57   -4.96
#> log_lik[43]     -5.65    0.01 0.52   -6.77   -5.97   -5.62   -5.29   -4.73
#> log_lik[44]     -4.41    0.01 0.40   -5.28   -4.66   -4.38   -4.13   -3.73
#> log_lik[45]     -3.04    0.00 0.35   -3.76   -3.27   -3.02   -2.80   -2.39
#> lp__          -463.99    0.15 6.14 -476.99 -467.90 -463.72 -459.56 -452.98
#>               n_eff Rhat
#> pi[1]          7360    1
#> pi[2]          8266    1
#> pi[3]          6833    1
#> pi[4]          6049    1
#> theta[1,1,1]   8468    1
#> theta[1,1,2]   7388    1
#> theta[1,1,3]   4848    1
#> theta[1,1,4]   5562    1
#> theta[1,2,1]   6403    1
#> theta[1,2,2]   6733    1
#> theta[1,2,3]   7004    1
#> theta[1,2,4]   5590    1
#> theta[1,3,1]   6362    1
#> theta[1,3,2]   3556    1
#> theta[1,3,3]   4888    1
#> theta[1,3,4]   5438    1
#> theta[1,4,1]   6492    1
#> theta[1,4,2]   7547    1
#> theta[1,4,3]   4764    1
#> theta[1,4,4]   5692    1
#> theta[2,1,1]   7071    1
#> theta[2,1,2]   6499    1
#> theta[2,1,3]   4788    1
#> theta[2,1,4]   4148    1
#> theta[2,2,1]   7608    1
#> theta[2,2,2]   6950    1
#> theta[2,2,3]   6446    1
#> theta[2,2,4]   6002    1
#> theta[2,3,1]   6427    1
#> theta[2,3,2]   5620    1
#> theta[2,3,3]   6597    1
#> theta[2,3,4]   6611    1
#> theta[2,4,1]   6472    1
#> theta[2,4,2]   6383    1
#> theta[2,4,3]   5620    1
#> theta[2,4,4]   6046    1
#> theta[3,1,1]   8449    1
#> theta[3,1,2]   5724    1
#> theta[3,1,3]   5037    1
#> theta[3,1,4]   4378    1
#> theta[3,2,1]   6687    1
#> theta[3,2,2]   7563    1
#> theta[3,2,3]   6160    1
#> theta[3,2,4]   5799    1
#> theta[3,3,1]   6131    1
#> theta[3,3,2]   8129    1
#> theta[3,3,3]   7268    1
#> theta[3,3,4]   7642    1
#> theta[3,4,1]   6420    1
#> theta[3,4,2]   6858    1
#> theta[3,4,3]   5294    1
#> theta[3,4,4]   6047    1
#> theta[4,1,1]   7967    1
#> theta[4,1,2]   6434    1
#> theta[4,1,3]   5363    1
#> theta[4,1,4]   4836    1
#> theta[4,2,1]   7410    1
#> theta[4,2,2]   7052    1
#> theta[4,2,3]   6217    1
#> theta[4,2,4]   6225    1
#> theta[4,3,1]   7017    1
#> theta[4,3,2]   5492    1
#> theta[4,3,3]   5257    1
#> theta[4,3,4]   4526    1
#> theta[4,4,1]   6458    1
#> theta[4,4,2]   7423    1
#> theta[4,4,3]   5608    1
#> theta[4,4,4]   5550    1
#> theta[5,1,1]   8078    1
#> theta[5,1,2]   4609    1
#> theta[5,1,3]   5487    1
#> theta[5,1,4]   4905    1
#> theta[5,2,1]   8947    1
#> theta[5,2,2]   6895    1
#> theta[5,2,3]   6459    1
#> theta[5,2,4]   5933    1
#> theta[5,3,1]   7259    1
#> theta[5,3,2]   7653    1
#> theta[5,3,3]   7298    1
#> theta[5,3,4]   7475    1
#> theta[5,4,1]   6180    1
#> theta[5,4,2]   6520    1
#> theta[5,4,3]   5492    1
#> theta[5,4,4]   6008    1
#> log_p_z[1,1]   7572    1
#> log_p_z[1,2]   4754    1
#> log_p_z[1,3]   3126    1
#> log_p_z[1,4]   3596    1
#> log_p_z[2,1]   5094    1
#> log_p_z[2,2]   4737    1
#> log_p_z[2,3]   3962    1
#> log_p_z[2,4]   3156    1
#> log_p_z[3,1]   4400    1
#> log_p_z[3,2]   4293    1
#> log_p_z[3,3]   3425    1
#> log_p_z[3,4]   3159    1
#> log_p_z[4,1]   5447    1
#> log_p_z[4,2]   4985    1
#> log_p_z[4,3]   2414    1
#> log_p_z[4,4]   3017    1
#> log_p_z[5,1]   4800    1
#> log_p_z[5,2]   7123    1
#> log_p_z[5,3]   2584    1
#> log_p_z[5,4]   2944    1
#> log_p_z[6,1]   5023    1
#> log_p_z[6,2]   6257    1
#> log_p_z[6,3]   2581    1
#> log_p_z[6,4]   2965    1
#> log_p_z[7,1]   6464    1
#> log_p_z[7,2]   4811    1
#> log_p_z[7,3]   3110    1
#> log_p_z[7,4]   3366    1
#> log_p_z[8,1]   5009    1
#> log_p_z[8,2]   4891    1
#> log_p_z[8,3]   6789    1
#> log_p_z[8,4]   3488    1
#> log_p_z[9,1]   4932    1
#> log_p_z[9,2]   6244    1
#> log_p_z[9,3]   2668    1
#> log_p_z[9,4]   2831    1
#> log_p_z[10,1]  4919    1
#> log_p_z[10,2]  5304    1
#> log_p_z[10,3]  2903    1
#> log_p_z[10,4]  2831    1
#> log_p_z[11,1]  5149    1
#> log_p_z[11,2]  3510    1
#> log_p_z[11,3]  2546    1
#> log_p_z[11,4]  6290    1
#> log_p_z[12,1]  5045    1
#> log_p_z[12,2]  4326    1
#> log_p_z[12,3]  2336    1
#> log_p_z[12,4]  3178    1
#> log_p_z[13,1]  7572    1
#> log_p_z[13,2]  4754    1
#> log_p_z[13,3]  3126    1
#> log_p_z[13,4]  3596    1
#> log_p_z[14,1]  4794    1
#> log_p_z[14,2]  4529    1
#> log_p_z[14,3]  2578    1
#> log_p_z[14,4]  2987    1
#> log_p_z[15,1]  5900    1
#> log_p_z[15,2]  4818    1
#> log_p_z[15,3]  3043    1
#> log_p_z[15,4]  3508    1
#> log_p_z[16,1]  7371    1
#> log_p_z[16,2]  4764    1
#> log_p_z[16,3]  3171    1
#> log_p_z[16,4]  3502    1
#> log_p_z[17,1]  7572    1
#> log_p_z[17,2]  4754    1
#> log_p_z[17,3]  3126    1
#> log_p_z[17,4]  3596    1
#> log_p_z[18,1]  7572    1
#> log_p_z[18,2]  4754    1
#> log_p_z[18,3]  3126    1
#> log_p_z[18,4]  3596    1
#> log_p_z[19,1]  5204    1
#> log_p_z[19,2]  7211    1
#> log_p_z[19,3]  2617    1
#> log_p_z[19,4]  2813    1
#> log_p_z[20,1]  4954    1
#> log_p_z[20,2]  6104    1
#> log_p_z[20,3]  2774    1
#> log_p_z[20,4]  2922    1
#> log_p_z[21,1]  4822    1
#> log_p_z[21,2]  7546    1
#> log_p_z[21,3]  2772    1
#> log_p_z[21,4]  2791    1
#> log_p_z[22,1]  5204    1
#> log_p_z[22,2]  7211    1
#> log_p_z[22,3]  2617    1
#> log_p_z[22,4]  2813    1
#> log_p_z[23,1]  4800    1
#> log_p_z[23,2]  7123    1
#> log_p_z[23,3]  2584    1
#> log_p_z[23,4]  2944    1
#> log_p_z[24,1]  4683    1
#> log_p_z[24,2]  4628    1
#> log_p_z[24,3]  3245    1
#> log_p_z[24,4]  2784    1
#> log_p_z[25,1]  7572    1
#> log_p_z[25,2]  4754    1
#> log_p_z[25,3]  3126    1
#> log_p_z[25,4]  3596    1
#> log_p_z[26,1]  7572    1
#> log_p_z[26,2]  4754    1
#> log_p_z[26,3]  3126    1
#> log_p_z[26,4]  3596    1
#> log_p_z[27,1]  4786    1
#> log_p_z[27,2]  6605    1
#> log_p_z[27,3]  3051    1
#> log_p_z[27,4]  2771    1
#> log_p_z[28,1]  7572    1
#> log_p_z[28,2]  4754    1
#> log_p_z[28,3]  3126    1
#> log_p_z[28,4]  3596    1
#> log_p_z[29,1]  7572    1
#> log_p_z[29,2]  4754    1
#> log_p_z[29,3]  3126    1
#> log_p_z[29,4]  3596    1
#> log_p_z[30,1]  5427    1
#> log_p_z[30,2]  4629    1
#> log_p_z[30,3]  3240    1
#> log_p_z[30,4]  3459    1
#> log_p_z[31,1]  7572    1
#> log_p_z[31,2]  4754    1
#> log_p_z[31,3]  3126    1
#> log_p_z[31,4]  3596    1
#> log_p_z[32,1]  4947    1
#> log_p_z[32,2]  4864    1
#> log_p_z[32,3]  5551    1
#> log_p_z[32,4]  3539    1
#> log_p_z[33,1]  7572    1
#> log_p_z[33,2]  4754    1
#> log_p_z[33,3]  3126    1
#> log_p_z[33,4]  3596    1
#> log_p_z[34,1]  4822    1
#> log_p_z[34,2]  7546    1
#> log_p_z[34,3]  2772    1
#> log_p_z[34,4]  2791    1
#> log_p_z[35,1]  5043    1
#> log_p_z[35,2]  5889    1
#> log_p_z[35,3]  2443    1
#> log_p_z[35,4]  2989    1
#> log_p_z[36,1]  4503    1
#> log_p_z[36,2]  3902    1
#> log_p_z[36,3]  2347    1
#> log_p_z[36,4]  3101    1
#> log_p_z[37,1]  4714    1
#> log_p_z[37,2]  5204    1
#> log_p_z[37,3]  3013    1
#> log_p_z[37,4]  2927    1
#> log_p_z[38,1]  4807    1
#> log_p_z[38,2]  4641    1
#> log_p_z[38,3]  2370    1
#> log_p_z[38,4]  3236    1
#> log_p_z[39,1]  5198    1
#> log_p_z[39,2]  5182    1
#> log_p_z[39,3]  5443    1
#> log_p_z[39,4]  3160    1
#> log_p_z[40,1]  7572    1
#> log_p_z[40,2]  4754    1
#> log_p_z[40,3]  3126    1
#> log_p_z[40,4]  3596    1
#> log_p_z[41,1]  7572    1
#> log_p_z[41,2]  4754    1
#> log_p_z[41,3]  3126    1
#> log_p_z[41,4]  3596    1
#> log_p_z[42,1]  6464    1
#> log_p_z[42,2]  4811    1
#> log_p_z[42,3]  3110    1
#> log_p_z[42,4]  3366    1
#> log_p_z[43,1]  4786    1
#> log_p_z[43,2]  6605    1
#> log_p_z[43,3]  3051    1
#> log_p_z[43,4]  2771    1
#> log_p_z[44,1]  5900    1
#> log_p_z[44,2]  4818    1
#> log_p_z[44,3]  3043    1
#> log_p_z[44,4]  3508    1
#> log_p_z[45,1]  4822    1
#> log_p_z[45,2]  7546    1
#> log_p_z[45,3]  2772    1
#> log_p_z[45,4]  2791    1
#> log_lik[1]     7572    1
#> log_lik[2]     4855    1
#> log_lik[3]     5105    1
#> log_lik[4]     5153    1
#> log_lik[5]     7123    1
#> log_lik[6]     6290    1
#> log_lik[7]     6465    1
#> log_lik[8]     6792    1
#> log_lik[9]     6244    1
#> log_lik[10]    5305    1
#> log_lik[11]    6292    1
#> log_lik[12]    5411    1
#> log_lik[13]    7572    1
#> log_lik[14]    4544    1
#> log_lik[15]    5900    1
#> log_lik[16]    7371    1
#> log_lik[17]    7572    1
#> log_lik[18]    7572    1
#> log_lik[19]    7211    1
#> log_lik[20]    6104    1
#> log_lik[21]    7546    1
#> log_lik[22]    7211    1
#> log_lik[23]    7123    1
#> log_lik[24]    4629    1
#> log_lik[25]    7572    1
#> log_lik[26]    7572    1
#> log_lik[27]    6607    1
#> log_lik[28]    7572    1
#> log_lik[29]    7572    1
#> log_lik[30]    5437    1
#> log_lik[31]    7572    1
#> log_lik[32]    5555    1
#> log_lik[33]    7572    1
#> log_lik[34]    7546    1
#> log_lik[35]    6103    1
#> log_lik[36]    5081    1
#> log_lik[37]    5206    1
#> log_lik[38]    4195    1
#> log_lik[39]    5438    1
#> log_lik[40]    7572    1
#> log_lik[41]    7572    1
#> log_lik[42]    6465    1
#> log_lik[43]    6607    1
#> log_lik[44]    5900    1
#> log_lik[45]    7546    1
#> lp__           1591    1
#> 
#> Samples were drawn using NUTS(diag_e) at Sun Feb  5 03:21:21 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }