Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.25    0.33    0.37    0.42    0.51
#> pi[2]            0.41    0.00 0.06    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.09    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.08    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.02    0.03    0.06    0.15
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.29
#> theta[1,3,3]     0.79    0.00 0.09    0.58    0.74    0.81    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.05    0.01    0.04    0.07    0.11    0.21
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.30
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,3]     0.12    0.00 0.11    0.00    0.04    0.09    0.18    0.40
#> theta[1,4,4]     0.69    0.00 0.13    0.40    0.60    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.76    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.21    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.59    0.65    0.76
#> theta[2,2,3]     0.30    0.00 0.09    0.15    0.24    0.30    0.36    0.48
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[2,3,3]     0.63    0.00 0.13    0.36    0.54    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.28    0.46
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,3]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.77    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.74    0.84    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[3,2,1]     0.10    0.00 0.06    0.02    0.05    0.09    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.88
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.28
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[3,3,2]     0.18    0.00 0.10    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.13    0.25    0.41    0.51    0.60    0.76
#> theta[3,3,4]     0.23    0.00 0.11    0.05    0.15    0.22    0.30    0.49
#> theta[3,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[3,4,3]     0.14    0.00 0.11    0.00    0.05    0.11    0.20    0.41
#> theta[3,4,4]     0.64    0.00 0.15    0.32    0.53    0.64    0.75    0.90
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.78    0.83    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.87
#> theta[4,2,3]     0.13    0.00 0.06    0.03    0.08    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.18
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.60    0.69    0.78    0.91
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.39
#> theta[4,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[4,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.34    0.56    0.66    0.76    0.90
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.83    0.87    0.91    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.10    0.13    0.18    0.30
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.28
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.09    0.14    0.21    0.38
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.61    0.70    0.84
#> theta[5,3,4]     0.15    0.00 0.09    0.02    0.08    0.14    0.21    0.38
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.36
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[5,4,3]     0.14    0.00 0.11    0.00    0.05    0.11    0.20    0.41
#> theta[5,4,4]     0.63    0.00 0.15    0.31    0.53    0.64    0.75    0.89
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[1,2]   -19.99    0.03 2.31  -24.92  -21.40  -19.79  -18.35  -15.95
#> log_p_z[1,3]   -24.74    0.07 4.22  -34.27  -27.28  -24.32  -21.77  -17.71
#> log_p_z[1,4]   -21.96    0.08 4.26  -31.91  -24.36  -21.42  -18.98  -15.09
#> log_p_z[2,1]   -29.75    0.07 4.37  -39.61  -32.28  -29.30  -26.62  -22.63
#> log_p_z[2,2]   -21.08    0.04 2.41  -26.36  -22.50  -20.88  -19.44  -16.79
#> log_p_z[2,3]    -7.63    0.02 1.21  -10.33   -8.32   -7.51   -6.80   -5.58
#> log_p_z[2,4]   -16.48    0.08 4.25  -26.65  -18.89  -15.89  -13.49   -9.78
#> log_p_z[3,1]   -11.62    0.02 1.65  -15.41  -12.59  -11.47  -10.45   -8.85
#> log_p_z[3,2]   -10.78    0.02 1.46  -14.08  -11.64  -10.63   -9.74   -8.37
#> log_p_z[3,3]   -23.32    0.06 3.37  -30.95  -25.37  -22.96  -20.96  -17.71
#> log_p_z[3,4]   -21.96    0.07 3.48  -29.72  -24.03  -21.69  -19.46  -16.09
#> log_p_z[4,1]   -14.65    0.03 1.87  -18.69  -15.83  -14.49  -13.33  -11.36
#> log_p_z[4,2]    -7.56    0.01 0.93   -9.54   -8.13   -7.49   -6.91   -5.92
#> log_p_z[4,3]   -20.73    0.09 4.29  -30.52  -23.31  -20.09  -17.67  -14.04
#> log_p_z[4,4]   -21.88    0.08 4.34  -31.74  -24.46  -21.39  -18.80  -14.83
#> log_p_z[5,1]   -21.48    0.04 2.56  -26.91  -23.09  -21.31  -19.68  -16.97
#> log_p_z[5,2]    -3.71    0.01 0.43   -4.61   -3.98   -3.70   -3.42   -2.92
#> log_p_z[5,3]   -18.68    0.08 4.22  -28.55  -21.10  -18.19  -15.60  -12.20
#> log_p_z[5,4]   -21.83    0.07 4.26  -31.32  -24.32  -21.47  -18.77  -14.86
#> log_p_z[6,1]   -21.45    0.04 2.52  -26.93  -23.03  -21.30  -19.68  -16.88
#> log_p_z[6,2]    -5.47    0.01 0.67   -6.92   -5.88   -5.43   -5.02   -4.28
#> log_p_z[6,3]   -17.48    0.08 4.08  -27.14  -19.75  -16.86  -14.57  -11.21
#> log_p_z[6,4]   -21.67    0.07 4.27  -30.90  -24.20  -21.29  -18.63  -14.72
#> log_p_z[7,1]    -6.00    0.01 0.62   -7.34   -6.39   -5.95   -5.56   -4.90
#> log_p_z[7,2]   -14.93    0.02 1.70  -18.64  -15.98  -14.79  -13.72  -11.94
#> log_p_z[7,3]   -24.24    0.06 3.51  -31.88  -26.46  -23.99  -21.73  -18.23
#> log_p_z[7,4]   -21.97    0.07 3.51  -29.61  -24.10  -21.67  -19.50  -15.92
#> log_p_z[8,1]   -29.73    0.07 4.35  -39.60  -32.33  -29.30  -26.63  -22.62
#> log_p_z[8,2]   -18.35    0.03 2.11  -22.92  -19.63  -18.16  -16.86  -14.73
#> log_p_z[8,3]    -5.73    0.01 0.81   -7.47   -6.25   -5.67   -5.16   -4.26
#> log_p_z[8,4]   -18.68    0.08 4.35  -29.29  -21.09  -18.10  -15.67  -11.87
#> log_p_z[9,1]   -19.96    0.03 2.27  -24.69  -21.37  -19.80  -18.38  -16.01
#> log_p_z[9,2]    -4.51    0.01 0.58   -5.74   -4.87   -4.49   -4.12   -3.51
#> log_p_z[9,3]   -19.67    0.08 4.31  -29.59  -22.26  -19.14  -16.56  -12.79
#> log_p_z[9,4]   -21.69    0.07 4.33  -31.47  -24.07  -21.15  -18.71  -14.75
#> log_p_z[10,1]  -22.24    0.04 2.35  -27.31  -23.66  -22.14  -20.60  -18.14
#> log_p_z[10,2]   -7.13    0.01 0.72   -8.63   -7.58   -7.08   -6.62   -5.87
#> log_p_z[10,3]  -16.85    0.06 3.16  -23.89  -18.78  -16.54  -14.58  -11.67
#> log_p_z[10,4]  -21.42    0.06 3.57  -29.28  -23.49  -21.02  -18.94  -15.42
#> log_p_z[11,1]  -29.75    0.07 4.50  -40.08  -32.24  -29.25  -26.67  -22.36
#> log_p_z[11,2]  -29.46    0.08 4.46  -39.53  -32.00  -29.02  -26.35  -22.12
#> log_p_z[11,3]  -18.13    0.06 3.22  -25.92  -19.84  -17.69  -15.85  -13.23
#> log_p_z[11,4]   -5.79    0.01 0.98   -7.88   -6.40   -5.70   -5.11   -4.12
#> log_p_z[12,1]  -22.90    0.04 2.66  -28.77  -24.56  -22.73  -21.07  -18.20
#> log_p_z[12,2]   -9.61    0.02 1.28  -12.45  -10.35   -9.46   -8.73   -7.55
#> log_p_z[12,3]  -15.12    0.08 4.05  -24.60  -17.32  -14.60  -12.25   -8.89
#> log_p_z[12,4]  -19.33    0.07 4.14  -28.65  -21.69  -18.80  -16.46  -12.79
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[13,2]  -19.99    0.03 2.31  -24.92  -21.40  -19.79  -18.35  -15.95
#> log_p_z[13,3]  -24.74    0.07 4.22  -34.27  -27.28  -24.32  -21.77  -17.71
#> log_p_z[13,4]  -21.96    0.08 4.26  -31.91  -24.36  -21.42  -18.98  -15.09
#> log_p_z[14,1]  -19.08    0.03 2.39  -24.37  -20.54  -18.90  -17.40  -14.90
#> log_p_z[14,2]   -6.24    0.01 0.84   -8.02   -6.76   -6.16   -5.64   -4.83
#> log_p_z[14,3]  -18.65    0.08 4.24  -28.90  -21.08  -18.09  -15.61  -12.13
#> log_p_z[14,4]  -21.83    0.07 4.21  -31.27  -24.33  -21.47  -18.89  -14.93
#> log_p_z[15,1]   -4.41    0.00 0.40   -5.27   -4.66   -4.38   -4.13   -3.71
#> log_p_z[15,2]  -17.23    0.03 1.85  -21.20  -18.38  -17.09  -15.92  -13.92
#> log_p_z[15,3]  -24.23    0.06 3.50  -31.62  -26.43  -23.97  -21.76  -18.20
#> log_p_z[15,4]  -21.93    0.07 3.49  -29.70  -24.03  -21.61  -19.45  -16.09
#> log_p_z[16,1]   -3.79    0.01 0.55   -5.00   -4.13   -3.75   -3.42   -2.85
#> log_p_z[16,2]  -17.69    0.03 2.19  -22.55  -19.05  -17.49  -16.13  -13.90
#> log_p_z[16,3]  -24.75    0.07 4.24  -34.17  -27.32  -24.31  -21.65  -17.93
#> log_p_z[16,4]  -22.00    0.08 4.28  -31.78  -24.44  -21.50  -18.95  -15.01
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[17,2]  -19.99    0.03 2.31  -24.92  -21.40  -19.79  -18.35  -15.95
#> log_p_z[17,3]  -24.74    0.07 4.22  -34.27  -27.28  -24.32  -21.77  -17.71
#> log_p_z[17,4]  -21.96    0.08 4.26  -31.91  -24.36  -21.42  -18.98  -15.09
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[18,2]  -19.99    0.03 2.31  -24.92  -21.40  -19.79  -18.35  -15.95
#> log_p_z[18,3]  -24.74    0.07 4.22  -34.27  -27.28  -24.32  -21.77  -17.71
#> log_p_z[18,4]  -21.96    0.08 4.26  -31.91  -24.36  -21.42  -18.98  -15.09
#> log_p_z[19,1]  -16.42    0.03 1.98  -20.58  -17.63  -16.29  -15.02  -12.92
#> log_p_z[19,2]   -4.65    0.01 0.58   -5.92   -5.01   -4.61   -4.26   -3.60
#> log_p_z[19,3]  -22.18    0.09 4.46  -32.48  -24.79  -21.63  -18.98  -15.08
#> log_p_z[19,4]  -21.85    0.07 4.35  -31.84  -24.30  -21.32  -18.78  -14.85
#> log_p_z[20,1]  -18.02    0.03 2.22  -22.82  -19.42  -17.87  -16.47  -14.14
#> log_p_z[20,2]   -7.09    0.01 0.98   -9.19   -7.71   -7.01   -6.40   -5.45
#> log_p_z[20,3]  -20.00    0.08 4.25  -30.13  -22.32  -19.40  -17.02  -13.36
#> log_p_z[20,4]  -21.65    0.07 4.24  -30.96  -24.11  -21.23  -18.63  -14.82
#> log_p_z[21,1]  -19.64    0.03 2.36  -24.67  -21.14  -19.45  -17.96  -15.53
#> log_p_z[21,2]   -3.03    0.00 0.35   -3.77   -3.26   -3.02   -2.80   -2.37
#> log_p_z[21,3]  -21.22    0.08 4.35  -31.20  -23.74  -20.70  -18.10  -14.31
#> log_p_z[21,4]  -21.85    0.07 4.29  -31.61  -24.29  -21.45  -18.82  -14.81
#> log_p_z[22,1]  -16.42    0.03 1.98  -20.58  -17.63  -16.29  -15.02  -12.92
#> log_p_z[22,2]   -4.65    0.01 0.58   -5.92   -5.01   -4.61   -4.26   -3.60
#> log_p_z[22,3]  -22.18    0.09 4.46  -32.48  -24.79  -21.63  -18.98  -15.08
#> log_p_z[22,4]  -21.85    0.07 4.35  -31.84  -24.30  -21.32  -18.78  -14.85
#> log_p_z[23,1]  -21.48    0.04 2.56  -26.91  -23.09  -21.31  -19.68  -16.97
#> log_p_z[23,2]   -3.71    0.01 0.43   -4.61   -3.98   -3.70   -3.42   -2.92
#> log_p_z[23,3]  -18.68    0.08 4.22  -28.55  -21.10  -18.19  -15.60  -12.20
#> log_p_z[23,4]  -21.83    0.07 4.26  -31.32  -24.32  -21.47  -18.77  -14.86
#> log_p_z[24,1]  -17.43    0.03 2.15  -22.15  -18.79  -17.23  -15.91  -13.74
#> log_p_z[24,2]   -5.80    0.01 0.63   -7.22   -6.18   -5.74   -5.34   -4.73
#> log_p_z[24,3]  -21.73    0.06 3.40  -29.51  -23.77  -21.41  -19.26  -16.09
#> log_p_z[24,4]  -21.88    0.06 3.53  -29.65  -24.00  -21.61  -19.38  -15.84
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[25,2]  -19.99    0.03 2.31  -24.92  -21.40  -19.79  -18.35  -15.95
#> log_p_z[25,3]  -24.74    0.07 4.22  -34.27  -27.28  -24.32  -21.77  -17.71
#> log_p_z[25,4]  -21.96    0.08 4.26  -31.91  -24.36  -21.42  -18.98  -15.09
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[26,2]  -19.99    0.03 2.31  -24.92  -21.40  -19.79  -18.35  -15.95
#> log_p_z[26,3]  -24.74    0.07 4.22  -34.27  -27.28  -24.32  -21.77  -17.71
#> log_p_z[26,4]  -21.96    0.08 4.26  -31.91  -24.36  -21.42  -18.98  -15.09
#> log_p_z[27,1]  -21.91    0.04 2.41  -27.05  -23.40  -21.74  -20.25  -17.73
#> log_p_z[27,2]   -5.65    0.01 0.53   -6.78   -5.98   -5.61   -5.27   -4.71
#> log_p_z[27,3]  -18.39    0.06 3.22  -25.68  -20.32  -18.08  -16.08  -13.17
#> log_p_z[27,4]  -21.58    0.06 3.48  -29.51  -23.57  -21.29  -19.25  -15.64
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[28,2]  -19.99    0.03 2.31  -24.92  -21.40  -19.79  -18.35  -15.95
#> log_p_z[28,3]  -24.74    0.07 4.22  -34.27  -27.28  -24.32  -21.77  -17.71
#> log_p_z[28,4]  -21.96    0.08 4.26  -31.91  -24.36  -21.42  -18.98  -15.09
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[29,2]  -19.99    0.03 2.31  -24.92  -21.40  -19.79  -18.35  -15.95
#> log_p_z[29,3]  -24.74    0.07 4.22  -34.27  -27.28  -24.32  -21.77  -17.71
#> log_p_z[29,4]  -21.96    0.08 4.26  -31.91  -24.36  -21.42  -18.98  -15.09
#> log_p_z[30,1]   -6.81    0.01 0.84   -8.70   -7.32   -6.72   -6.21   -5.42
#> log_p_z[30,2]  -14.70    0.02 1.70  -18.39  -15.80  -14.58  -13.48  -11.81
#> log_p_z[30,3]  -24.27    0.06 3.48  -31.94  -26.44  -23.96  -21.77  -18.27
#> log_p_z[30,4]  -21.93    0.07 3.52  -29.64  -23.98  -21.63  -19.47  -15.85
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[31,2]  -19.99    0.03 2.31  -24.92  -21.40  -19.79  -18.35  -15.95
#> log_p_z[31,3]  -24.74    0.07 4.22  -34.27  -27.28  -24.32  -21.77  -17.71
#> log_p_z[31,4]  -21.96    0.08 4.26  -31.91  -24.36  -21.42  -18.98  -15.09
#> log_p_z[32,1]  -29.75    0.07 4.30  -39.55  -32.25  -29.40  -26.69  -22.64
#> log_p_z[32,2]  -14.85    0.02 1.72  -18.65  -15.90  -14.71  -13.65  -11.90
#> log_p_z[32,3]   -6.05    0.01 0.94   -8.11   -6.60   -5.97   -5.40   -4.46
#> log_p_z[32,4]  -20.81    0.08 4.48  -31.52  -23.25  -20.29  -17.62  -13.66
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[33,2]  -19.99    0.03 2.31  -24.92  -21.40  -19.79  -18.35  -15.95
#> log_p_z[33,3]  -24.74    0.07 4.22  -34.27  -27.28  -24.32  -21.77  -17.71
#> log_p_z[33,4]  -21.96    0.08 4.26  -31.91  -24.36  -21.42  -18.98  -15.09
#> log_p_z[34,1]  -19.64    0.03 2.36  -24.67  -21.14  -19.45  -17.96  -15.53
#> log_p_z[34,2]   -3.03    0.00 0.35   -3.77   -3.26   -3.02   -2.80   -2.37
#> log_p_z[34,3]  -21.22    0.08 4.35  -31.20  -23.74  -20.70  -18.10  -14.31
#> log_p_z[34,4]  -21.85    0.07 4.29  -31.61  -24.29  -21.45  -18.82  -14.81
#> log_p_z[35,1]  -22.60    0.04 2.64  -28.37  -24.18  -22.42  -20.77  -17.98
#> log_p_z[35,2]   -5.54    0.01 0.68   -6.98   -5.98   -5.49   -5.05   -4.35
#> log_p_z[35,3]  -16.06    0.08 4.13  -26.06  -18.37  -15.51  -13.08   -9.82
#> log_p_z[35,4]  -21.78    0.07 4.24  -31.39  -24.24  -21.38  -18.81  -14.75
#> log_p_z[36,1]  -29.73    0.06 3.65  -37.56  -32.03  -29.48  -27.11  -23.39
#> log_p_z[36,2]  -21.95    0.04 2.29  -26.99  -23.35  -21.77  -20.31  -17.98
#> log_p_z[36,3]  -10.47    0.03 1.64  -14.56  -11.31  -10.23   -9.35   -7.94
#> log_p_z[36,4]  -14.13    0.06 3.27  -21.56  -16.11  -13.77  -11.73   -8.86
#> log_p_z[37,1]  -18.56    0.03 2.28  -23.74  -19.84  -18.34  -17.00  -14.67
#> log_p_z[37,2]   -7.62    0.01 0.80   -9.35   -8.13   -7.55   -7.06   -6.21
#> log_p_z[37,3]  -19.10    0.06 3.24  -26.30  -21.02  -18.81  -16.81  -13.81
#> log_p_z[37,4]  -21.83    0.06 3.51  -29.63  -23.95  -21.55  -19.38  -15.78
#> log_p_z[38,1]  -25.20    0.04 2.74  -31.04  -26.97  -24.99  -23.24  -20.48
#> log_p_z[38,2]   -9.63    0.01 0.98  -11.82  -10.23   -9.55   -8.94   -7.97
#> log_p_z[38,3]  -11.70    0.05 2.80  -18.22  -13.35  -11.34   -9.60   -7.49
#> log_p_z[38,4]  -21.35    0.06 3.53  -29.15  -23.45  -21.05  -18.80  -15.32
#> log_p_z[39,1]  -29.41    0.07 4.30  -39.25  -31.92  -29.02  -26.36  -22.45
#> log_p_z[39,2]  -16.87    0.03 2.02  -21.25  -18.09  -16.68  -15.45  -13.38
#> log_p_z[39,3]   -7.27    0.02 1.12   -9.74   -7.95   -7.19   -6.49   -5.31
#> log_p_z[39,4]  -18.84    0.07 4.23  -29.08  -21.08  -18.26  -15.89  -12.22
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[40,2]  -19.99    0.03 2.31  -24.92  -21.40  -19.79  -18.35  -15.95
#> log_p_z[40,3]  -24.74    0.07 4.22  -34.27  -27.28  -24.32  -21.77  -17.71
#> log_p_z[40,4]  -21.96    0.08 4.26  -31.91  -24.36  -21.42  -18.98  -15.09
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_p_z[41,2]  -19.99    0.03 2.31  -24.92  -21.40  -19.79  -18.35  -15.95
#> log_p_z[41,3]  -24.74    0.07 4.22  -34.27  -27.28  -24.32  -21.77  -17.71
#> log_p_z[41,4]  -21.96    0.08 4.26  -31.91  -24.36  -21.42  -18.98  -15.09
#> log_p_z[42,1]   -6.00    0.01 0.62   -7.34   -6.39   -5.95   -5.56   -4.90
#> log_p_z[42,2]  -14.93    0.02 1.70  -18.64  -15.98  -14.79  -13.72  -11.94
#> log_p_z[42,3]  -24.24    0.06 3.51  -31.88  -26.46  -23.99  -21.73  -18.23
#> log_p_z[42,4]  -21.97    0.07 3.51  -29.61  -24.10  -21.67  -19.50  -15.92
#> log_p_z[43,1]  -21.91    0.04 2.41  -27.05  -23.40  -21.74  -20.25  -17.73
#> log_p_z[43,2]   -5.65    0.01 0.53   -6.78   -5.98   -5.61   -5.27   -4.71
#> log_p_z[43,3]  -18.39    0.06 3.22  -25.68  -20.32  -18.08  -16.08  -13.17
#> log_p_z[43,4]  -21.58    0.06 3.48  -29.51  -23.57  -21.29  -19.25  -15.64
#> log_p_z[44,1]   -4.41    0.00 0.40   -5.27   -4.66   -4.38   -4.13   -3.71
#> log_p_z[44,2]  -17.23    0.03 1.85  -21.20  -18.38  -17.09  -15.92  -13.92
#> log_p_z[44,3]  -24.23    0.06 3.50  -31.62  -26.43  -23.97  -21.76  -18.20
#> log_p_z[44,4]  -21.93    0.07 3.49  -29.70  -24.03  -21.61  -19.45  -16.09
#> log_p_z[45,1]  -19.64    0.03 2.36  -24.67  -21.14  -19.45  -17.96  -15.53
#> log_p_z[45,2]   -3.03    0.00 0.35   -3.77   -3.26   -3.02   -2.80   -2.37
#> log_p_z[45,3]  -21.22    0.08 4.35  -31.20  -23.74  -20.70  -18.10  -14.31
#> log_p_z[45,4]  -21.85    0.07 4.29  -31.61  -24.29  -21.45  -18.82  -14.81
#> log_lik[1]      -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[2]      -7.60    0.02 1.15  -10.11   -8.28   -7.50   -6.80   -5.57
#> log_lik[3]      -9.94    0.01 0.93  -11.83  -10.54   -9.91   -9.31   -8.20
#> log_lik[4]      -7.55    0.01 0.92   -9.49   -8.12   -7.48   -6.91   -5.92
#> log_lik[5]      -3.71    0.01 0.43   -4.61   -3.98   -3.70   -3.42   -2.92
#> log_lik[6]      -5.47    0.01 0.67   -6.92   -5.88   -5.43   -5.02   -4.28
#> log_lik[7]      -6.00    0.01 0.62   -7.34   -6.39   -5.95   -5.56   -4.90
#> log_lik[8]      -5.72    0.01 0.81   -7.47   -6.25   -5.67   -5.16   -4.26
#> log_lik[9]      -4.51    0.01 0.58   -5.74   -4.87   -4.49   -4.12   -3.51
#> log_lik[10]     -7.12    0.01 0.71   -8.62   -7.58   -7.08   -6.62   -5.87
#> log_lik[11]     -5.79    0.01 0.97   -7.88   -6.40   -5.70   -5.11   -4.12
#> log_lik[12]     -9.35    0.01 1.04  -11.55  -10.04   -9.28   -8.63   -7.49
#> log_lik[13]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[14]     -6.24    0.01 0.84   -8.01   -6.76   -6.16   -5.64   -4.83
#> log_lik[15]     -4.41    0.00 0.40   -5.27   -4.66   -4.38   -4.13   -3.71
#> log_lik[16]     -3.79    0.01 0.55   -5.00   -4.13   -3.75   -3.42   -2.85
#> log_lik[17]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[18]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[19]     -4.65    0.01 0.58   -5.92   -5.01   -4.61   -4.26   -3.60
#> log_lik[20]     -7.09    0.01 0.98   -9.19   -7.71   -7.01   -6.40   -5.45
#> log_lik[21]     -3.03    0.00 0.35   -3.77   -3.26   -3.02   -2.80   -2.37
#> log_lik[22]     -4.65    0.01 0.58   -5.92   -5.01   -4.61   -4.26   -3.60
#> log_lik[23]     -3.71    0.01 0.43   -4.61   -3.98   -3.70   -3.42   -2.92
#> log_lik[24]     -5.80    0.01 0.63   -7.22   -6.18   -5.74   -5.34   -4.73
#> log_lik[25]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[26]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[27]     -5.65    0.01 0.53   -6.78   -5.98   -5.61   -5.27   -4.71
#> log_lik[28]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[29]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[30]     -6.81    0.01 0.84   -8.68   -7.32   -6.72   -6.21   -5.42
#> log_lik[31]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[32]     -6.05    0.01 0.94   -8.10   -6.60   -5.97   -5.40   -4.46
#> log_lik[33]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[34]     -3.03    0.00 0.35   -3.77   -3.26   -3.02   -2.80   -2.37
#> log_lik[35]     -5.54    0.01 0.68   -6.98   -5.98   -5.48   -5.05   -4.35
#> log_lik[36]     -9.95    0.02 1.16  -12.44  -10.67   -9.88   -9.16   -7.85
#> log_lik[37]     -7.62    0.01 0.80   -9.35   -8.13   -7.55   -7.06   -6.21
#> log_lik[38]     -9.02    0.01 0.86  -10.74   -9.58   -9.00   -8.45   -7.34
#> log_lik[39]     -7.27    0.02 1.12   -9.74   -7.95   -7.19   -6.49   -5.31
#> log_lik[40]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[41]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.67
#> log_lik[42]     -6.00    0.01 0.62   -7.34   -6.39   -5.95   -5.56   -4.90
#> log_lik[43]     -5.65    0.01 0.53   -6.78   -5.98   -5.61   -5.27   -4.71
#> log_lik[44]     -4.41    0.00 0.40   -5.27   -4.66   -4.38   -4.13   -3.71
#> log_lik[45]     -3.03    0.00 0.35   -3.77   -3.26   -3.02   -2.80   -2.37
#> lp__          -463.71    0.16 6.10 -476.84 -467.48 -463.26 -459.42 -453.09
#>               n_eff Rhat
#> pi[1]          9972    1
#> pi[2]          8636    1
#> pi[3]          6752    1
#> pi[4]          6816    1
#> theta[1,1,1]   9201    1
#> theta[1,1,2]   7925    1
#> theta[1,1,3]   4418    1
#> theta[1,1,4]   4848    1
#> theta[1,2,1]   6790    1
#> theta[1,2,2]   8291    1
#> theta[1,2,3]   7051    1
#> theta[1,2,4]   5374    1
#> theta[1,3,1]   6172    1
#> theta[1,3,2]   4450    1
#> theta[1,3,3]   5339    1
#> theta[1,3,4]   5211    1
#> theta[1,4,1]   5584    1
#> theta[1,4,2]   6438    1
#> theta[1,4,3]   4763    1
#> theta[1,4,4]   5460    1
#> theta[2,1,1]   7641    1
#> theta[2,1,2]   6486    1
#> theta[2,1,3]   4667    1
#> theta[2,1,4]   4491    1
#> theta[2,2,1]   6195    1
#> theta[2,2,2]   6661    1
#> theta[2,2,3]   6691    1
#> theta[2,2,4]   6331    1
#> theta[2,3,1]   6661    1
#> theta[2,3,2]   6421    1
#> theta[2,3,3]   6555    1
#> theta[2,3,4]   5542    1
#> theta[2,4,1]   6000    1
#> theta[2,4,2]   6291    1
#> theta[2,4,3]   5689    1
#> theta[2,4,4]   6708    1
#> theta[3,1,1]   7336    1
#> theta[3,1,2]   5217    1
#> theta[3,1,3]   4661    1
#> theta[3,1,4]   4501    1
#> theta[3,2,1]   6719    1
#> theta[3,2,2]   6895    1
#> theta[3,2,3]   6629    1
#> theta[3,2,4]   4860    1
#> theta[3,3,1]   6148    1
#> theta[3,3,2]   6492    1
#> theta[3,3,3]   7339    1
#> theta[3,3,4]   7981    1
#> theta[3,4,1]   5193    1
#> theta[3,4,2]   6510    1
#> theta[3,4,3]   5754    1
#> theta[3,4,4]   6417    1
#> theta[4,1,1]   6982    1
#> theta[4,1,2]   5437    1
#> theta[4,1,3]   4274    1
#> theta[4,1,4]   4775    1
#> theta[4,2,1]   7748    1
#> theta[4,2,2]   7119    1
#> theta[4,2,3]   7632    1
#> theta[4,2,4]   5558    1
#> theta[4,3,1]   7267    1
#> theta[4,3,2]   6696    1
#> theta[4,3,3]   5866    1
#> theta[4,3,4]   6206    1
#> theta[4,4,1]   6624    1
#> theta[4,4,2]   6272    1
#> theta[4,4,3]   6364    1
#> theta[4,4,4]   6691    1
#> theta[5,1,1]   8256    1
#> theta[5,1,2]   6066    1
#> theta[5,1,3]   4294    1
#> theta[5,1,4]   4424    1
#> theta[5,2,1]   7363    1
#> theta[5,2,2]   7582    1
#> theta[5,2,3]   5969    1
#> theta[5,2,4]   6131    1
#> theta[5,3,1]   6687    1
#> theta[5,3,2]   6711    1
#> theta[5,3,3]   5851    1
#> theta[5,3,4]   7138    1
#> theta[5,4,1]   6216    1
#> theta[5,4,2]   7090    1
#> theta[5,4,3]   5682    1
#> theta[5,4,4]   6181    1
#> log_p_z[1,1]   8257    1
#> log_p_z[1,2]   4713    1
#> log_p_z[1,3]   3610    1
#> log_p_z[1,4]   2557    1
#> log_p_z[2,1]   3924    1
#> log_p_z[2,2]   4541    1
#> log_p_z[2,3]   3856    1
#> log_p_z[2,4]   3085    1
#> log_p_z[3,1]   4484    1
#> log_p_z[3,2]   4540    1
#> log_p_z[3,3]   3081    1
#> log_p_z[3,4]   2694    1
#> log_p_z[4,1]   5244    1
#> log_p_z[4,2]   5104    1
#> log_p_z[4,3]   2471    1
#> log_p_z[4,4]   3291    1
#> log_p_z[5,1]   4949    1
#> log_p_z[5,2]   6088    1
#> log_p_z[5,3]   2915    1
#> log_p_z[5,4]   3562    1
#> log_p_z[6,1]   5100    1
#> log_p_z[6,2]   6264    1
#> log_p_z[6,3]   2895    1
#> log_p_z[6,4]   3515    1
#> log_p_z[7,1]   5714    1
#> log_p_z[7,2]   4724    1
#> log_p_z[7,3]   3099    1
#> log_p_z[7,4]   2908    1
#> log_p_z[8,1]   3557    1
#> log_p_z[8,2]   4015    1
#> log_p_z[8,3]   6700    1
#> log_p_z[8,4]   3229    1
#> log_p_z[9,1]   5046    1
#> log_p_z[9,2]   6190    1
#> log_p_z[9,3]   2610    1
#> log_p_z[9,4]   3597    1
#> log_p_z[10,1]  4297    1
#> log_p_z[10,2]  5777    1
#> log_p_z[10,3]  2693    1
#> log_p_z[10,4]  3277    1
#> log_p_z[11,1]  4315    1
#> log_p_z[11,2]  3508    1
#> log_p_z[11,3]  2852    1
#> log_p_z[11,4]  5793    1
#> log_p_z[12,1]  5018    1
#> log_p_z[12,2]  4222    1
#> log_p_z[12,3]  2827    1
#> log_p_z[12,4]  3467    1
#> log_p_z[13,1]  8257    1
#> log_p_z[13,2]  4713    1
#> log_p_z[13,3]  3610    1
#> log_p_z[13,4]  2557    1
#> log_p_z[14,1]  5107    1
#> log_p_z[14,2]  4926    1
#> log_p_z[14,3]  2802    1
#> log_p_z[14,4]  3537    1
#> log_p_z[15,1]  6490    1
#> log_p_z[15,2]  4599    1
#> log_p_z[15,3]  3303    1
#> log_p_z[15,4]  2621    1
#> log_p_z[16,1]  5395    1
#> log_p_z[16,2]  4807    1
#> log_p_z[16,3]  3451    1
#> log_p_z[16,4]  2759    1
#> log_p_z[17,1]  8257    1
#> log_p_z[17,2]  4713    1
#> log_p_z[17,3]  3610    1
#> log_p_z[17,4]  2557    1
#> log_p_z[18,1]  8257    1
#> log_p_z[18,2]  4713    1
#> log_p_z[18,3]  3610    1
#> log_p_z[18,4]  2557    1
#> log_p_z[19,1]  5505    1
#> log_p_z[19,2]  6955    1
#> log_p_z[19,3]  2453    1
#> log_p_z[19,4]  3476    1
#> log_p_z[20,1]  5191    1
#> log_p_z[20,2]  5533    1
#> log_p_z[20,3]  2695    1
#> log_p_z[20,4]  3463    1
#> log_p_z[21,1]  4992    1
#> log_p_z[21,2]  6348    1
#> log_p_z[21,3]  2684    1
#> log_p_z[21,4]  3543    1
#> log_p_z[22,1]  5505    1
#> log_p_z[22,2]  6955    1
#> log_p_z[22,3]  2453    1
#> log_p_z[22,4]  3476    1
#> log_p_z[23,1]  4949    1
#> log_p_z[23,2]  6088    1
#> log_p_z[23,3]  2915    1
#> log_p_z[23,4]  3562    1
#> log_p_z[24,1]  4734    1
#> log_p_z[24,2]  5273    1
#> log_p_z[24,3]  2848    1
#> log_p_z[24,4]  3317    1
#> log_p_z[25,1]  8257    1
#> log_p_z[25,2]  4713    1
#> log_p_z[25,3]  3610    1
#> log_p_z[25,4]  2557    1
#> log_p_z[26,1]  8257    1
#> log_p_z[26,2]  4713    1
#> log_p_z[26,3]  3610    1
#> log_p_z[26,4]  2557    1
#> log_p_z[27,1]  4471    1
#> log_p_z[27,2]  7330    1
#> log_p_z[27,3]  2750    1
#> log_p_z[27,4]  3189    1
#> log_p_z[28,1]  8257    1
#> log_p_z[28,2]  4713    1
#> log_p_z[28,3]  3610    1
#> log_p_z[28,4]  2557    1
#> log_p_z[29,1]  8257    1
#> log_p_z[29,2]  4713    1
#> log_p_z[29,3]  3610    1
#> log_p_z[29,4]  2557    1
#> log_p_z[30,1]  4862    1
#> log_p_z[30,2]  4657    1
#> log_p_z[30,3]  3177    1
#> log_p_z[30,4]  2391    1
#> log_p_z[31,1]  8257    1
#> log_p_z[31,2]  4713    1
#> log_p_z[31,3]  3610    1
#> log_p_z[31,4]  2557    1
#> log_p_z[32,1]  3884    1
#> log_p_z[32,2]  5338    1
#> log_p_z[32,3]  4511    1
#> log_p_z[32,4]  3338    1
#> log_p_z[33,1]  8257    1
#> log_p_z[33,2]  4713    1
#> log_p_z[33,3]  3610    1
#> log_p_z[33,4]  2557    1
#> log_p_z[34,1]  4992    1
#> log_p_z[34,2]  6348    1
#> log_p_z[34,3]  2684    1
#> log_p_z[34,4]  3543    1
#> log_p_z[35,1]  4838    1
#> log_p_z[35,2]  6193    1
#> log_p_z[35,3]  2829    1
#> log_p_z[35,4]  3696    1
#> log_p_z[36,1]  3792    1
#> log_p_z[36,2]  3365    1
#> log_p_z[36,3]  2608    1
#> log_p_z[36,4]  2824    1
#> log_p_z[37,1]  4697    1
#> log_p_z[37,2]  6280    1
#> log_p_z[37,3]  2683    1
#> log_p_z[37,4]  3557    1
#> log_p_z[38,1]  4089    1
#> log_p_z[38,2]  5213    1
#> log_p_z[38,3]  2904    1
#> log_p_z[38,4]  3539    1
#> log_p_z[39,1]  3582    1
#> log_p_z[39,2]  3940    1
#> log_p_z[39,3]  5056    1
#> log_p_z[39,4]  3190    1
#> log_p_z[40,1]  8257    1
#> log_p_z[40,2]  4713    1
#> log_p_z[40,3]  3610    1
#> log_p_z[40,4]  2557    1
#> log_p_z[41,1]  8257    1
#> log_p_z[41,2]  4713    1
#> log_p_z[41,3]  3610    1
#> log_p_z[41,4]  2557    1
#> log_p_z[42,1]  5714    1
#> log_p_z[42,2]  4724    1
#> log_p_z[42,3]  3099    1
#> log_p_z[42,4]  2908    1
#> log_p_z[43,1]  4471    1
#> log_p_z[43,2]  7330    1
#> log_p_z[43,3]  2750    1
#> log_p_z[43,4]  3189    1
#> log_p_z[44,1]  6490    1
#> log_p_z[44,2]  4599    1
#> log_p_z[44,3]  3303    1
#> log_p_z[44,4]  2621    1
#> log_p_z[45,1]  4992    1
#> log_p_z[45,2]  6348    1
#> log_p_z[45,3]  2684    1
#> log_p_z[45,4]  3543    1
#> log_lik[1]     8257    1
#> log_lik[2]     4487    1
#> log_lik[3]     5451    1
#> log_lik[4]     5213    1
#> log_lik[5]     6087    1
#> log_lik[6]     6263    1
#> log_lik[7]     5721    1
#> log_lik[8]     6697    1
#> log_lik[9]     6192    1
#> log_lik[10]    5800    1
#> log_lik[11]    5794    1
#> log_lik[12]    5498    1
#> log_lik[13]    8257    1
#> log_lik[14]    4926    1
#> log_lik[15]    6490    1
#> log_lik[16]    5395    1
#> log_lik[17]    8257    1
#> log_lik[18]    8257    1
#> log_lik[19]    6955    1
#> log_lik[20]    5531    1
#> log_lik[21]    6348    1
#> log_lik[22]    6955    1
#> log_lik[23]    6087    1
#> log_lik[24]    5274    1
#> log_lik[25]    8257    1
#> log_lik[26]    8257    1
#> log_lik[27]    7331    1
#> log_lik[28]    8257    1
#> log_lik[29]    8257    1
#> log_lik[30]    4879    1
#> log_lik[31]    8257    1
#> log_lik[32]    4555    1
#> log_lik[33]    8257    1
#> log_lik[34]    6348    1
#> log_lik[35]    6247    1
#> log_lik[36]    4444    1
#> log_lik[37]    6290    1
#> log_lik[38]    4491    1
#> log_lik[39]    5069    1
#> log_lik[40]    8257    1
#> log_lik[41]    8257    1
#> log_lik[42]    5721    1
#> log_lik[43]    7331    1
#> log_lik[44]    6490    1
#> log_lik[45]    6348    1
#> lp__           1493    1
#> 
#> Samples were drawn using NUTS(diag_e) at Sat Mar 11 06:19:12 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }