Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.07    0.25    0.33    0.37    0.42    0.51
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.09    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.90    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.08    0.10    0.12    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.75    0.82    0.86    0.89    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.01    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.29
#> theta[1,3,3]     0.79    0.00 0.10    0.58    0.73    0.81    0.87    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.09    0.00    0.03    0.07    0.14    0.31
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.09    0.18    0.40
#> theta[1,4,4]     0.69    0.00 0.14    0.41    0.60    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.69    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.08    0.05    0.12    0.16    0.22    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.58    0.64    0.75
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.25    0.30    0.37    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[2,3,3]     0.62    0.00 0.13    0.36    0.54    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.29    0.46
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.55    0.67    0.77    0.90
#> theta[3,1,1]     0.88    0.00 0.06    0.74    0.85    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.10    0.00 0.06    0.01    0.05    0.09    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.09    0.55    0.67    0.73    0.79    0.88
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.13    0.25    0.41    0.51    0.60    0.76
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.31    0.48
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[3,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[3,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.41
#> theta[3,4,4]     0.63    0.00 0.15    0.31    0.53    0.64    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.78    0.83    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.08    0.55    0.67    0.73    0.79    0.87
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.08    0.13    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.60    0.70    0.78    0.91
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.37
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.54    0.66    0.77    0.90
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.82    0.87    0.91    0.97
#> theta[5,1,2]     0.06    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.81
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[5,3,1]     0.08    0.00 0.08    0.00    0.03    0.06    0.11    0.28
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.40
#> theta[5,3,3]     0.60    0.00 0.13    0.33    0.51    0.61    0.70    0.84
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.38
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[5,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.19    0.42
#> theta[5,4,4]     0.63    0.00 0.15    0.31    0.53    0.65    0.75    0.90
#> log_p_z[1,1]    -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_p_z[1,2]   -20.05    0.04 2.36  -25.35  -21.44  -19.83  -18.41  -16.08
#> log_p_z[1,3]   -24.78    0.07 4.35  -35.07  -27.23  -24.26  -21.65  -17.89
#> log_p_z[1,4]   -21.97    0.07 4.39  -32.21  -24.51  -21.56  -18.84  -14.74
#> log_p_z[2,1]   -29.73    0.07 4.35  -39.77  -32.11  -29.24  -26.74  -22.60
#> log_p_z[2,2]   -21.07    0.04 2.38  -26.23  -22.54  -20.92  -19.38  -16.92
#> log_p_z[2,3]    -7.65    0.02 1.23  -10.29   -8.37   -7.53   -6.81   -5.61
#> log_p_z[2,4]   -16.37    0.07 4.11  -25.81  -18.76  -15.89  -13.50   -9.62
#> log_p_z[3,1]   -11.62    0.03 1.65  -15.33  -12.60  -11.41  -10.43   -8.95
#> log_p_z[3,2]   -10.82    0.02 1.52  -14.33  -11.68  -10.65   -9.74   -8.40
#> log_p_z[3,3]   -23.44    0.06 3.52  -31.46  -25.59  -23.06  -20.94  -17.64
#> log_p_z[3,4]   -22.00    0.06 3.59  -30.14  -24.10  -21.69  -19.41  -16.00
#> log_p_z[4,1]   -14.66    0.02 1.87  -18.69  -15.81  -14.51  -13.35  -11.37
#> log_p_z[4,2]    -7.54    0.01 0.94   -9.57   -8.12   -7.47   -6.91   -5.87
#> log_p_z[4,3]   -20.83    0.08 4.53  -31.37  -23.37  -20.29  -17.63  -13.56
#> log_p_z[4,4]   -21.97    0.08 4.41  -31.74  -24.57  -21.50  -18.80  -14.80
#> log_p_z[5,1]   -21.45    0.04 2.61  -27.27  -23.05  -21.22  -19.65  -17.00
#> log_p_z[5,2]    -3.71    0.01 0.43   -4.63   -3.99   -3.68   -3.41   -2.93
#> log_p_z[5,3]   -18.82    0.08 4.40  -28.81  -21.31  -18.23  -15.67  -12.14
#> log_p_z[5,4]   -21.99    0.08 4.35  -32.01  -24.54  -21.53  -18.90  -14.96
#> log_p_z[6,1]   -21.46    0.04 2.52  -26.90  -23.05  -21.26  -19.71  -17.20
#> log_p_z[6,2]    -5.47    0.01 0.68   -6.93   -5.90   -5.43   -4.99   -4.27
#> log_p_z[6,3]   -17.60    0.07 4.28  -27.64  -19.94  -16.99  -14.53  -11.13
#> log_p_z[6,4]   -21.75    0.08 4.33  -31.68  -24.22  -21.27  -18.67  -14.81
#> log_p_z[7,1]    -6.00    0.01 0.62   -7.34   -6.39   -5.94   -5.56   -4.92
#> log_p_z[7,2]   -14.98    0.03 1.75  -18.84  -15.97  -14.85  -13.73  -12.05
#> log_p_z[7,3]   -24.32    0.06 3.61  -32.57  -26.38  -23.92  -21.78  -18.34
#> log_p_z[7,4]   -22.00    0.06 3.61  -29.96  -24.22  -21.69  -19.50  -15.71
#> log_p_z[8,1]   -29.73    0.07 4.41  -39.55  -32.31  -29.26  -26.69  -22.58
#> log_p_z[8,2]   -18.28    0.03 2.05  -22.67  -19.57  -18.13  -16.85  -14.78
#> log_p_z[8,3]    -5.72    0.01 0.83   -7.54   -6.24   -5.67   -5.14   -4.28
#> log_p_z[8,4]   -18.60    0.08 4.29  -28.29  -21.11  -18.08  -15.53  -11.71
#> log_p_z[9,1]   -20.02    0.03 2.27  -25.06  -21.44  -19.85  -18.41  -16.06
#> log_p_z[9,2]    -4.52    0.01 0.57   -5.74   -4.88   -4.47   -4.10   -3.52
#> log_p_z[9,3]   -19.80    0.08 4.50  -30.09  -22.15  -19.34  -16.58  -12.74
#> log_p_z[9,4]   -21.81    0.08 4.30  -31.58  -24.29  -21.38  -18.74  -14.76
#> log_p_z[10,1]  -22.27    0.04 2.42  -27.57  -23.78  -22.07  -20.60  -18.00
#> log_p_z[10,2]   -7.12    0.01 0.68   -8.59   -7.52   -7.07   -6.65   -5.94
#> log_p_z[10,3]  -16.95    0.06 3.28  -24.43  -18.75  -16.63  -14.61  -11.69
#> log_p_z[10,4]  -21.48    0.06 3.49  -29.19  -23.59  -21.21  -19.04  -15.52
#> log_p_z[11,1]  -29.72    0.06 4.39  -39.54  -32.28  -29.23  -26.64  -22.53
#> log_p_z[11,2]  -29.33    0.07 4.26  -39.02  -31.85  -28.86  -26.34  -22.14
#> log_p_z[11,3]  -18.22    0.07 3.36  -26.06  -19.97  -17.81  -15.89  -12.95
#> log_p_z[11,4]   -5.81    0.01 0.97   -8.00   -6.40   -5.74   -5.13   -4.17
#> log_p_z[12,1]  -22.98    0.04 2.62  -28.68  -24.61  -22.80  -21.14  -18.34
#> log_p_z[12,2]   -9.58    0.02 1.28  -12.50  -10.29   -9.46   -8.72   -7.47
#> log_p_z[12,3]  -15.26    0.07 4.25  -25.12  -17.51  -14.68  -12.27   -8.83
#> log_p_z[12,4]  -19.36    0.07 4.21  -29.24  -21.73  -18.92  -16.44  -12.47
#> log_p_z[13,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_p_z[13,2]  -20.05    0.04 2.36  -25.35  -21.44  -19.83  -18.41  -16.08
#> log_p_z[13,3]  -24.78    0.07 4.35  -35.07  -27.23  -24.26  -21.65  -17.89
#> log_p_z[13,4]  -21.97    0.07 4.39  -32.21  -24.51  -21.56  -18.84  -14.74
#> log_p_z[14,1]  -19.04    0.04 2.45  -24.59  -20.51  -18.82  -17.34  -14.88
#> log_p_z[14,2]   -6.26    0.01 0.89   -8.24   -6.78   -6.17   -5.64   -4.75
#> log_p_z[14,3]  -18.77    0.08 4.37  -29.02  -21.17  -18.14  -15.58  -12.24
#> log_p_z[14,4]  -22.00    0.08 4.34  -31.67  -24.56  -21.57  -18.86  -14.92
#> log_p_z[15,1]   -4.41    0.00 0.41   -5.32   -4.66   -4.38   -4.14   -3.68
#> log_p_z[15,2]  -17.27    0.03 1.89  -21.48  -18.41  -17.11  -15.96  -14.04
#> log_p_z[15,3]  -24.28    0.06 3.64  -32.74  -26.43  -23.82  -21.76  -18.31
#> log_p_z[15,4]  -21.97    0.06 3.59  -30.01  -24.11  -21.58  -19.51  -15.77
#> log_p_z[16,1]   -3.79    0.01 0.54   -4.99   -4.13   -3.74   -3.41   -2.84
#> log_p_z[16,2]  -17.76    0.04 2.25  -22.74  -19.05  -17.57  -16.15  -14.03
#> log_p_z[16,3]  -24.82    0.07 4.33  -34.84  -27.19  -24.32  -21.76  -17.91
#> log_p_z[16,4]  -22.00    0.07 4.41  -32.30  -24.54  -21.52  -18.89  -14.62
#> log_p_z[17,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_p_z[17,2]  -20.05    0.04 2.36  -25.35  -21.44  -19.83  -18.41  -16.08
#> log_p_z[17,3]  -24.78    0.07 4.35  -35.07  -27.23  -24.26  -21.65  -17.89
#> log_p_z[17,4]  -21.97    0.07 4.39  -32.21  -24.51  -21.56  -18.84  -14.74
#> log_p_z[18,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_p_z[18,2]  -20.05    0.04 2.36  -25.35  -21.44  -19.83  -18.41  -16.08
#> log_p_z[18,3]  -24.78    0.07 4.35  -35.07  -27.23  -24.26  -21.65  -17.89
#> log_p_z[18,4]  -21.97    0.07 4.39  -32.21  -24.51  -21.56  -18.84  -14.74
#> log_p_z[19,1]  -16.42    0.03 1.98  -20.62  -17.65  -16.26  -15.01  -12.98
#> log_p_z[19,2]   -4.65    0.01 0.60   -5.91   -5.04   -4.62   -4.23   -3.57
#> log_p_z[19,3]  -22.27    0.08 4.63  -32.76  -24.84  -21.84  -18.98  -14.99
#> log_p_z[19,4]  -22.00    0.08 4.38  -32.01  -24.58  -21.54  -18.91  -14.91
#> log_p_z[20,1]  -18.05    0.03 2.22  -22.96  -19.38  -17.85  -16.50  -14.32
#> log_p_z[20,2]   -7.10    0.01 0.98   -9.26   -7.71   -7.02   -6.40   -5.44
#> log_p_z[20,3]  -20.10    0.08 4.38  -30.21  -22.51  -19.49  -16.96  -13.41
#> log_p_z[20,4]  -21.74    0.08 4.31  -31.57  -24.21  -21.24  -18.62  -14.80
#> log_p_z[21,1]  -19.62    0.04 2.40  -24.88  -21.04  -19.42  -17.95  -15.47
#> log_p_z[21,2]   -3.04    0.00 0.36   -3.79   -3.28   -3.02   -2.80   -2.38
#> log_p_z[21,3]  -21.35    0.08 4.51  -31.60  -23.87  -20.83  -18.03  -14.37
#> log_p_z[21,4]  -22.01    0.08 4.32  -31.98  -24.58  -21.55  -18.87  -15.04
#> log_p_z[22,1]  -16.42    0.03 1.98  -20.62  -17.65  -16.26  -15.01  -12.98
#> log_p_z[22,2]   -4.65    0.01 0.60   -5.91   -5.04   -4.62   -4.23   -3.57
#> log_p_z[22,3]  -22.27    0.08 4.63  -32.76  -24.84  -21.84  -18.98  -14.99
#> log_p_z[22,4]  -22.00    0.08 4.38  -32.01  -24.58  -21.54  -18.91  -14.91
#> log_p_z[23,1]  -21.45    0.04 2.61  -27.27  -23.05  -21.22  -19.65  -17.00
#> log_p_z[23,2]   -3.71    0.01 0.43   -4.63   -3.99   -3.68   -3.41   -2.93
#> log_p_z[23,3]  -18.82    0.08 4.40  -28.81  -21.31  -18.23  -15.67  -12.14
#> log_p_z[23,4]  -21.99    0.08 4.35  -32.01  -24.54  -21.53  -18.90  -14.96
#> log_p_z[24,1]  -17.41    0.03 2.20  -22.26  -18.73  -17.21  -15.87  -13.71
#> log_p_z[24,2]   -5.82    0.01 0.64   -7.22   -6.19   -5.76   -5.37   -4.74
#> log_p_z[24,3]  -21.85    0.06 3.57  -29.70  -23.94  -21.47  -19.31  -16.07
#> log_p_z[24,4]  -22.01    0.06 3.53  -29.92  -24.19  -21.74  -19.51  -16.00
#> log_p_z[25,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_p_z[25,2]  -20.05    0.04 2.36  -25.35  -21.44  -19.83  -18.41  -16.08
#> log_p_z[25,3]  -24.78    0.07 4.35  -35.07  -27.23  -24.26  -21.65  -17.89
#> log_p_z[25,4]  -21.97    0.07 4.39  -32.21  -24.51  -21.56  -18.84  -14.74
#> log_p_z[26,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_p_z[26,2]  -20.05    0.04 2.36  -25.35  -21.44  -19.83  -18.41  -16.08
#> log_p_z[26,3]  -24.78    0.07 4.35  -35.07  -27.23  -24.26  -21.65  -17.89
#> log_p_z[26,4]  -21.97    0.07 4.39  -32.21  -24.51  -21.56  -18.84  -14.74
#> log_p_z[27,1]  -21.87    0.04 2.52  -27.55  -23.43  -21.64  -20.09  -17.53
#> log_p_z[27,2]   -5.64    0.01 0.52   -6.74   -5.97   -5.60   -5.28   -4.71
#> log_p_z[27,3]  -18.49    0.06 3.32  -25.88  -20.46  -18.13  -16.08  -13.20
#> log_p_z[27,4]  -21.68    0.06 3.46  -29.28  -23.78  -21.45  -19.19  -15.69
#> log_p_z[28,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_p_z[28,2]  -20.05    0.04 2.36  -25.35  -21.44  -19.83  -18.41  -16.08
#> log_p_z[28,3]  -24.78    0.07 4.35  -35.07  -27.23  -24.26  -21.65  -17.89
#> log_p_z[28,4]  -21.97    0.07 4.39  -32.21  -24.51  -21.56  -18.84  -14.74
#> log_p_z[29,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_p_z[29,2]  -20.05    0.04 2.36  -25.35  -21.44  -19.83  -18.41  -16.08
#> log_p_z[29,3]  -24.78    0.07 4.35  -35.07  -27.23  -24.26  -21.65  -17.89
#> log_p_z[29,4]  -21.97    0.07 4.39  -32.21  -24.51  -21.56  -18.84  -14.74
#> log_p_z[30,1]   -6.83    0.01 0.86   -8.70   -7.33   -6.75   -6.23   -5.43
#> log_p_z[30,2]  -14.72    0.03 1.73  -18.69  -15.72  -14.57  -13.51  -11.79
#> log_p_z[30,3]  -24.32    0.06 3.63  -32.58  -26.41  -23.96  -21.77  -18.37
#> log_p_z[30,4]  -21.95    0.06 3.61  -30.10  -24.11  -21.51  -19.46  -15.84
#> log_p_z[31,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_p_z[31,2]  -20.05    0.04 2.36  -25.35  -21.44  -19.83  -18.41  -16.08
#> log_p_z[31,3]  -24.78    0.07 4.35  -35.07  -27.23  -24.26  -21.65  -17.89
#> log_p_z[31,4]  -21.97    0.07 4.39  -32.21  -24.51  -21.56  -18.84  -14.74
#> log_p_z[32,1]  -29.73    0.07 4.37  -39.66  -32.24  -29.26  -26.64  -22.69
#> log_p_z[32,2]  -14.79    0.02 1.66  -18.38  -15.81  -14.66  -13.64  -11.95
#> log_p_z[32,3]   -6.08    0.01 0.94   -8.08   -6.66   -6.02   -5.42   -4.46
#> log_p_z[32,4]  -20.74    0.08 4.41  -30.73  -23.36  -20.28  -17.50  -13.59
#> log_p_z[33,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_p_z[33,2]  -20.05    0.04 2.36  -25.35  -21.44  -19.83  -18.41  -16.08
#> log_p_z[33,3]  -24.78    0.07 4.35  -35.07  -27.23  -24.26  -21.65  -17.89
#> log_p_z[33,4]  -21.97    0.07 4.39  -32.21  -24.51  -21.56  -18.84  -14.74
#> log_p_z[34,1]  -19.62    0.04 2.40  -24.88  -21.04  -19.42  -17.95  -15.47
#> log_p_z[34,2]   -3.04    0.00 0.36   -3.79   -3.28   -3.02   -2.80   -2.38
#> log_p_z[34,3]  -21.35    0.08 4.51  -31.60  -23.87  -20.83  -18.03  -14.37
#> log_p_z[34,4]  -22.01    0.08 4.32  -31.98  -24.58  -21.55  -18.87  -15.04
#> log_p_z[35,1]  -22.59    0.04 2.72  -28.70  -24.23  -22.33  -20.77  -17.75
#> log_p_z[35,2]   -5.52    0.01 0.70   -7.02   -5.94   -5.47   -5.02   -4.27
#> log_p_z[35,3]  -16.20    0.07 4.25  -26.19  -18.49  -15.60  -13.06   -9.84
#> log_p_z[35,4]  -21.93    0.08 4.32  -32.19  -24.37  -21.46  -18.86  -14.87
#> log_p_z[36,1]  -29.76    0.06 3.56  -37.85  -31.80  -29.41  -27.29  -23.85
#> log_p_z[36,2]  -21.89    0.04 2.28  -27.03  -23.25  -21.71  -20.31  -17.91
#> log_p_z[36,3]  -10.53    0.04 1.74  -14.79  -11.44  -10.24   -9.33   -7.95
#> log_p_z[36,4]  -14.02    0.06 3.21  -21.27  -15.96  -13.76  -11.65   -8.82
#> log_p_z[37,1]  -18.55    0.04 2.31  -23.56  -19.95  -18.33  -16.99  -14.61
#> log_p_z[37,2]   -7.63    0.01 0.82   -9.50   -8.10   -7.57   -7.07   -6.24
#> log_p_z[37,3]  -19.23    0.06 3.36  -26.95  -21.19  -18.80  -16.81  -13.91
#> log_p_z[37,4]  -21.95    0.06 3.47  -29.56  -24.13  -21.59  -19.51  -16.00
#> log_p_z[38,1]  -25.24    0.04 2.78  -31.28  -26.95  -25.01  -23.26  -20.47
#> log_p_z[38,2]   -9.59    0.01 0.97  -11.67  -10.19   -9.50   -8.90   -7.93
#> log_p_z[38,3]  -11.79    0.05 2.88  -18.51  -13.39  -11.42   -9.72   -7.46
#> log_p_z[38,4]  -21.40    0.06 3.50  -29.16  -23.55  -21.14  -18.88  -15.38
#> log_p_z[39,1]  -29.34    0.07 4.39  -38.94  -31.92  -28.89  -26.27  -22.34
#> log_p_z[39,2]  -16.81    0.03 1.99  -21.06  -18.05  -16.66  -15.40  -13.34
#> log_p_z[39,3]   -7.27    0.02 1.12   -9.70   -7.96   -7.21   -6.48   -5.27
#> log_p_z[39,4]  -18.80    0.07 4.17  -28.44  -21.20  -18.27  -15.90  -12.14
#> log_p_z[40,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_p_z[40,2]  -20.05    0.04 2.36  -25.35  -21.44  -19.83  -18.41  -16.08
#> log_p_z[40,3]  -24.78    0.07 4.35  -35.07  -27.23  -24.26  -21.65  -17.89
#> log_p_z[40,4]  -21.97    0.07 4.39  -32.21  -24.51  -21.56  -18.84  -14.74
#> log_p_z[41,1]   -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_p_z[41,2]  -20.05    0.04 2.36  -25.35  -21.44  -19.83  -18.41  -16.08
#> log_p_z[41,3]  -24.78    0.07 4.35  -35.07  -27.23  -24.26  -21.65  -17.89
#> log_p_z[41,4]  -21.97    0.07 4.39  -32.21  -24.51  -21.56  -18.84  -14.74
#> log_p_z[42,1]   -6.00    0.01 0.62   -7.34   -6.39   -5.94   -5.56   -4.92
#> log_p_z[42,2]  -14.98    0.03 1.75  -18.84  -15.97  -14.85  -13.73  -12.05
#> log_p_z[42,3]  -24.32    0.06 3.61  -32.57  -26.38  -23.92  -21.78  -18.34
#> log_p_z[42,4]  -22.00    0.06 3.61  -29.96  -24.22  -21.69  -19.50  -15.71
#> log_p_z[43,1]  -21.87    0.04 2.52  -27.55  -23.43  -21.64  -20.09  -17.53
#> log_p_z[43,2]   -5.64    0.01 0.52   -6.74   -5.97   -5.60   -5.28   -4.71
#> log_p_z[43,3]  -18.49    0.06 3.32  -25.88  -20.46  -18.13  -16.08  -13.20
#> log_p_z[43,4]  -21.68    0.06 3.46  -29.28  -23.78  -21.45  -19.19  -15.69
#> log_p_z[44,1]   -4.41    0.00 0.41   -5.32   -4.66   -4.38   -4.14   -3.68
#> log_p_z[44,2]  -17.27    0.03 1.89  -21.48  -18.41  -17.11  -15.96  -14.04
#> log_p_z[44,3]  -24.28    0.06 3.64  -32.74  -26.43  -23.82  -21.76  -18.31
#> log_p_z[44,4]  -21.97    0.06 3.59  -30.01  -24.11  -21.58  -19.51  -15.77
#> log_p_z[45,1]  -19.62    0.04 2.40  -24.88  -21.04  -19.42  -17.95  -15.47
#> log_p_z[45,2]   -3.04    0.00 0.36   -3.79   -3.28   -3.02   -2.80   -2.38
#> log_p_z[45,3]  -21.35    0.08 4.51  -31.60  -23.87  -20.83  -18.03  -14.37
#> log_p_z[45,4]  -22.01    0.08 4.32  -31.98  -24.58  -21.55  -18.87  -15.04
#> log_lik[1]      -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_lik[2]      -7.60    0.02 1.14  -10.04   -8.32   -7.52   -6.81   -5.61
#> log_lik[3]      -9.95    0.01 0.93  -11.93  -10.54   -9.90   -9.30   -8.29
#> log_lik[4]      -7.54    0.01 0.93   -9.52   -8.12   -7.46   -6.90   -5.87
#> log_lik[5]      -3.71    0.01 0.43   -4.63   -3.99   -3.68   -3.41   -2.93
#> log_lik[6]      -5.47    0.01 0.68   -6.92   -5.90   -5.43   -4.99   -4.27
#> log_lik[7]      -6.00    0.01 0.62   -7.34   -6.39   -5.94   -5.56   -4.92
#> log_lik[8]      -5.72    0.01 0.83   -7.54   -6.24   -5.67   -5.14   -4.28
#> log_lik[9]      -4.52    0.01 0.57   -5.74   -4.88   -4.47   -4.10   -3.52
#> log_lik[10]     -7.11    0.01 0.68   -8.59   -7.52   -7.07   -6.65   -5.94
#> log_lik[11]     -5.81    0.01 0.97   -8.00   -6.40   -5.74   -5.13   -4.17
#> log_lik[12]     -9.32    0.01 1.06  -11.56  -10.00   -9.27   -8.60   -7.41
#> log_lik[13]     -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_lik[14]     -6.26    0.01 0.89   -8.24   -6.78   -6.17   -5.64   -4.75
#> log_lik[15]     -4.41    0.00 0.41   -5.32   -4.66   -4.38   -4.14   -3.68
#> log_lik[16]     -3.79    0.01 0.54   -4.99   -4.13   -3.74   -3.41   -2.84
#> log_lik[17]     -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_lik[18]     -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_lik[19]     -4.65    0.01 0.60   -5.91   -5.04   -4.62   -4.23   -3.57
#> log_lik[20]     -7.10    0.01 0.98   -9.26   -7.71   -7.02   -6.40   -5.44
#> log_lik[21]     -3.04    0.00 0.36   -3.79   -3.28   -3.02   -2.80   -2.38
#> log_lik[22]     -4.65    0.01 0.60   -5.91   -5.04   -4.62   -4.23   -3.57
#> log_lik[23]     -3.71    0.01 0.43   -4.63   -3.99   -3.68   -3.41   -2.93
#> log_lik[24]     -5.82    0.01 0.64   -7.22   -6.19   -5.76   -5.37   -4.74
#> log_lik[25]     -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_lik[26]     -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_lik[27]     -5.64    0.01 0.52   -6.74   -5.97   -5.60   -5.28   -4.71
#> log_lik[28]     -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_lik[29]     -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_lik[30]     -6.83    0.01 0.86   -8.68   -7.33   -6.75   -6.23   -5.43
#> log_lik[31]     -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_lik[32]     -6.07    0.01 0.94   -8.08   -6.65   -6.01   -5.42   -4.46
#> log_lik[33]     -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_lik[34]     -3.04    0.00 0.36   -3.79   -3.28   -3.02   -2.80   -2.38
#> log_lik[35]     -5.52    0.01 0.70   -7.01   -5.93   -5.47   -5.02   -4.27
#> log_lik[36]     -9.94    0.02 1.17  -12.39  -10.68   -9.85   -9.14   -7.82
#> log_lik[37]     -7.63    0.01 0.82   -9.50   -8.10   -7.56   -7.07   -6.24
#> log_lik[38]     -9.01    0.01 0.87  -10.76   -9.56   -8.99   -8.43   -7.31
#> log_lik[39]     -7.27    0.02 1.11   -9.68   -7.95   -7.21   -6.48   -5.27
#> log_lik[40]     -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_lik[41]     -2.21    0.00 0.29   -2.81   -2.40   -2.20   -1.99   -1.68
#> log_lik[42]     -6.00    0.01 0.62   -7.34   -6.39   -5.94   -5.56   -4.92
#> log_lik[43]     -5.64    0.01 0.52   -6.74   -5.97   -5.60   -5.28   -4.71
#> log_lik[44]     -4.41    0.00 0.41   -5.32   -4.66   -4.38   -4.14   -3.68
#> log_lik[45]     -3.04    0.00 0.36   -3.79   -3.28   -3.02   -2.80   -2.38
#> lp__          -463.95    0.16 6.03 -476.49 -467.95 -463.74 -459.64 -453.01
#>               n_eff Rhat
#> pi[1]          8374    1
#> pi[2]          9513    1
#> pi[3]          6579    1
#> pi[4]          6766    1
#> theta[1,1,1]   7690    1
#> theta[1,1,2]   6661    1
#> theta[1,1,3]   4323    1
#> theta[1,1,4]   4730    1
#> theta[1,2,1]   6811    1
#> theta[1,2,2]   7647    1
#> theta[1,2,3]   7610    1
#> theta[1,2,4]   5366    1
#> theta[1,3,1]   7382    1
#> theta[1,3,2]   4620    1
#> theta[1,3,3]   5167    1
#> theta[1,3,4]   6717    1
#> theta[1,4,1]   6850    1
#> theta[1,4,2]   5480    1
#> theta[1,4,3]   4204    1
#> theta[1,4,4]   5253    1
#> theta[2,1,1]   7371    1
#> theta[2,1,2]   6439    1
#> theta[2,1,3]   5332    1
#> theta[2,1,4]   4479    1
#> theta[2,2,1]   7907    1
#> theta[2,2,2]   6831    1
#> theta[2,2,3]   5950    1
#> theta[2,2,4]   5448    1
#> theta[2,3,1]   6718    1
#> theta[2,3,2]   5851    1
#> theta[2,3,3]   6215    1
#> theta[2,3,4]   5400    1
#> theta[2,4,1]   6820    1
#> theta[2,4,2]   6074    1
#> theta[2,4,3]   5609    1
#> theta[2,4,4]   6167    1
#> theta[3,1,1]   7155    1
#> theta[3,1,2]   5761    1
#> theta[3,1,3]   4614    1
#> theta[3,1,4]   5637    1
#> theta[3,2,1]   7636    1
#> theta[3,2,2]   7929    1
#> theta[3,2,3]   6812    1
#> theta[3,2,4]   5677    1
#> theta[3,3,1]   6418    1
#> theta[3,3,2]   6695    1
#> theta[3,3,3]   6713    1
#> theta[3,3,4]   7192    1
#> theta[3,4,1]   6914    1
#> theta[3,4,2]   6043    1
#> theta[3,4,3]   5118    1
#> theta[3,4,4]   5447    1
#> theta[4,1,1]   6913    1
#> theta[4,1,2]   5352    1
#> theta[4,1,3]   4657    1
#> theta[4,1,4]   4564    1
#> theta[4,2,1]   6805    1
#> theta[4,2,2]   7837    1
#> theta[4,2,3]   6433    1
#> theta[4,2,4]   5449    1
#> theta[4,3,1]   6232    1
#> theta[4,3,2]   6319    1
#> theta[4,3,3]   6168    1
#> theta[4,3,4]   5191    1
#> theta[4,4,1]   5032    1
#> theta[4,4,2]   6720    1
#> theta[4,4,3]   5712    1
#> theta[4,4,4]   6637    1
#> theta[5,1,1]   6511    1
#> theta[5,1,2]   5500    1
#> theta[5,1,3]   5011    1
#> theta[5,1,4]   4746    1
#> theta[5,2,1]   7961    1
#> theta[5,2,2]   7254    1
#> theta[5,2,3]   6525    1
#> theta[5,2,4]   5508    1
#> theta[5,3,1]   6516    1
#> theta[5,3,2]   7618    1
#> theta[5,3,3]   7658    1
#> theta[5,3,4]   6843    1
#> theta[5,4,1]   6409    1
#> theta[5,4,2]   7449    1
#> theta[5,4,3]   4824    1
#> theta[5,4,4]   5462    1
#> log_p_z[1,1]   7449    1
#> log_p_z[1,2]   4156    1
#> log_p_z[1,3]   3832    1
#> log_p_z[1,4]   4071    1
#> log_p_z[2,1]   3656    1
#> log_p_z[2,2]   4016    1
#> log_p_z[2,3]   3881    1
#> log_p_z[2,4]   3239    1
#> log_p_z[3,1]   4322    1
#> log_p_z[3,2]   3943    1
#> log_p_z[3,3]   3566    1
#> log_p_z[3,4]   3708    1
#> log_p_z[4,1]   6080    1
#> log_p_z[4,2]   5676    1
#> log_p_z[4,3]   3445    1
#> log_p_z[4,4]   2903    1
#> log_p_z[5,1]   4329    1
#> log_p_z[5,2]   6200    1
#> log_p_z[5,3]   3330    1
#> log_p_z[5,4]   3027    1
#> log_p_z[6,1]   4747    1
#> log_p_z[6,2]   6696    1
#> log_p_z[6,3]   3347    1
#> log_p_z[6,4]   3184    1
#> log_p_z[7,1]   5776    1
#> log_p_z[7,2]   4079    1
#> log_p_z[7,3]   3578    1
#> log_p_z[7,4]   3729    1
#> log_p_z[8,1]   3863    1
#> log_p_z[8,2]   4476    1
#> log_p_z[8,3]   6161    1
#> log_p_z[8,4]   3220    1
#> log_p_z[9,1]   4713    1
#> log_p_z[9,2]   6316    1
#> log_p_z[9,3]   3254    1
#> log_p_z[9,4]   3021    1
#> log_p_z[10,1]  4302    1
#> log_p_z[10,2]  6282    1
#> log_p_z[10,3]  3304    1
#> log_p_z[10,4]  3078    1
#> log_p_z[11,1]  4637    1
#> log_p_z[11,2]  3639    1
#> log_p_z[11,3]  2624    1
#> log_p_z[11,4]  6412    1
#> log_p_z[12,1]  5056    1
#> log_p_z[12,2]  4173    1
#> log_p_z[12,3]  3232    1
#> log_p_z[12,4]  3344    1
#> log_p_z[13,1]  7449    1
#> log_p_z[13,2]  4156    1
#> log_p_z[13,3]  3832    1
#> log_p_z[13,4]  4071    1
#> log_p_z[14,1]  4178    1
#> log_p_z[14,2]  3936    1
#> log_p_z[14,3]  3287    1
#> log_p_z[14,4]  3301    1
#> log_p_z[15,1]  6819    1
#> log_p_z[15,2]  4399    1
#> log_p_z[15,3]  3350    1
#> log_p_z[15,4]  3656    1
#> log_p_z[16,1]  6451    1
#> log_p_z[16,2]  3964    1
#> log_p_z[16,3]  3901    1
#> log_p_z[16,4]  4081    1
#> log_p_z[17,1]  7449    1
#> log_p_z[17,2]  4156    1
#> log_p_z[17,3]  3832    1
#> log_p_z[17,4]  4071    1
#> log_p_z[18,1]  7449    1
#> log_p_z[18,2]  4156    1
#> log_p_z[18,3]  3832    1
#> log_p_z[18,4]  4071    1
#> log_p_z[19,1]  4982    1
#> log_p_z[19,2]  6081    1
#> log_p_z[19,3]  3449    1
#> log_p_z[19,4]  2928    1
#> log_p_z[20,1]  4557    1
#> log_p_z[20,2]  6027    1
#> log_p_z[20,3]  3109    1
#> log_p_z[20,4]  2873    1
#> log_p_z[21,1]  4418    1
#> log_p_z[21,2]  8327    1
#> log_p_z[21,3]  3288    1
#> log_p_z[21,4]  2982    1
#> log_p_z[22,1]  4982    1
#> log_p_z[22,2]  6081    1
#> log_p_z[22,3]  3449    1
#> log_p_z[22,4]  2928    1
#> log_p_z[23,1]  4329    1
#> log_p_z[23,2]  6200    1
#> log_p_z[23,3]  3330    1
#> log_p_z[23,4]  3027    1
#> log_p_z[24,1]  4244    1
#> log_p_z[24,2]  3821    1
#> log_p_z[24,3]  3330    1
#> log_p_z[24,4]  3181    1
#> log_p_z[25,1]  7449    1
#> log_p_z[25,2]  4156    1
#> log_p_z[25,3]  3832    1
#> log_p_z[25,4]  4071    1
#> log_p_z[26,1]  7449    1
#> log_p_z[26,2]  4156    1
#> log_p_z[26,3]  3832    1
#> log_p_z[26,4]  4071    1
#> log_p_z[27,1]  4257    1
#> log_p_z[27,2]  7388    1
#> log_p_z[27,3]  3355    1
#> log_p_z[27,4]  3151    1
#> log_p_z[28,1]  7449    1
#> log_p_z[28,2]  4156    1
#> log_p_z[28,3]  3832    1
#> log_p_z[28,4]  4071    1
#> log_p_z[29,1]  7449    1
#> log_p_z[29,2]  4156    1
#> log_p_z[29,3]  3832    1
#> log_p_z[29,4]  4071    1
#> log_p_z[30,1]  5316    1
#> log_p_z[30,2]  4318    1
#> log_p_z[30,3]  3490    1
#> log_p_z[30,4]  3151    1
#> log_p_z[31,1]  7449    1
#> log_p_z[31,2]  4156    1
#> log_p_z[31,3]  3832    1
#> log_p_z[31,4]  4071    1
#> log_p_z[32,1]  4041    1
#> log_p_z[32,2]  5796    1
#> log_p_z[32,3]  5398    1
#> log_p_z[32,4]  3126    1
#> log_p_z[33,1]  7449    1
#> log_p_z[33,2]  4156    1
#> log_p_z[33,3]  3832    1
#> log_p_z[33,4]  4071    1
#> log_p_z[34,1]  4418    1
#> log_p_z[34,2]  8327    1
#> log_p_z[34,3]  3288    1
#> log_p_z[34,4]  2982    1
#> log_p_z[35,1]  3982    1
#> log_p_z[35,2]  6543    1
#> log_p_z[35,3]  3322    1
#> log_p_z[35,4]  2980    1
#> log_p_z[36,1]  3271    1
#> log_p_z[36,2]  3419    1
#> log_p_z[36,3]  1971    1
#> log_p_z[36,4]  2859    1
#> log_p_z[37,1]  3977    1
#> log_p_z[37,2]  5274    1
#> log_p_z[37,3]  3261    1
#> log_p_z[37,4]  3107    1
#> log_p_z[38,1]  3910    1
#> log_p_z[38,2]  4626    1
#> log_p_z[38,3]  3285    1
#> log_p_z[38,4]  3156    1
#> log_p_z[39,1]  3919    1
#> log_p_z[39,2]  4771    1
#> log_p_z[39,3]  4716    1
#> log_p_z[39,4]  3545    1
#> log_p_z[40,1]  7449    1
#> log_p_z[40,2]  4156    1
#> log_p_z[40,3]  3832    1
#> log_p_z[40,4]  4071    1
#> log_p_z[41,1]  7449    1
#> log_p_z[41,2]  4156    1
#> log_p_z[41,3]  3832    1
#> log_p_z[41,4]  4071    1
#> log_p_z[42,1]  5776    1
#> log_p_z[42,2]  4079    1
#> log_p_z[42,3]  3578    1
#> log_p_z[42,4]  3729    1
#> log_p_z[43,1]  4257    1
#> log_p_z[43,2]  7388    1
#> log_p_z[43,3]  3355    1
#> log_p_z[43,4]  3151    1
#> log_p_z[44,1]  6819    1
#> log_p_z[44,2]  4399    1
#> log_p_z[44,3]  3350    1
#> log_p_z[44,4]  3656    1
#> log_p_z[45,1]  4418    1
#> log_p_z[45,2]  8327    1
#> log_p_z[45,3]  3288    1
#> log_p_z[45,4]  2982    1
#> log_lik[1]     7449    1
#> log_lik[2]     4724    1
#> log_lik[3]     5471    1
#> log_lik[4]     5897    1
#> log_lik[5]     6200    1
#> log_lik[6]     6701    1
#> log_lik[7]     5771    1
#> log_lik[8]     6159    1
#> log_lik[9]     6317    1
#> log_lik[10]    6323    1
#> log_lik[11]    6412    1
#> log_lik[12]    5412    1
#> log_lik[13]    7449    1
#> log_lik[14]    3933    1
#> log_lik[15]    6819    1
#> log_lik[16]    6451    1
#> log_lik[17]    7449    1
#> log_lik[18]    7449    1
#> log_lik[19]    6080    1
#> log_lik[20]    6029    1
#> log_lik[21]    8327    1
#> log_lik[22]    6080    1
#> log_lik[23]    6200    1
#> log_lik[24]    3821    1
#> log_lik[25]    7449    1
#> log_lik[26]    7449    1
#> log_lik[27]    7388    1
#> log_lik[28]    7449    1
#> log_lik[29]    7449    1
#> log_lik[30]    5305    1
#> log_lik[31]    7449    1
#> log_lik[32]    5409    1
#> log_lik[33]    7449    1
#> log_lik[34]    8327    1
#> log_lik[35]    6560    1
#> log_lik[36]    4793    1
#> log_lik[37]    5275    1
#> log_lik[38]    4159    1
#> log_lik[39]    4742    1
#> log_lik[40]    7449    1
#> log_lik[41]    7449    1
#> log_lik[42]    5771    1
#> log_lik[43]    7388    1
#> log_lik[44]    6819    1
#> log_lik[45]    8327    1
#> lp__           1448    1
#> 
#> Samples were drawn using NUTS(diag_e) at Mon Sep 11 13:21:28 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }