Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.37    0.00 0.06    0.26    0.33    0.37    0.42    0.50
#> pi[2]            0.41    0.00 0.06    0.29    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.08    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.08    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.13
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.17
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.59    0.73    0.80    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.30
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.30
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.09    0.19    0.40
#> theta[1,4,4]     0.69    0.00 0.13    0.41    0.59    0.70    0.79    0.91
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.69    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.22    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.39    0.52    0.58    0.64    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.25    0.30    0.37    0.50
#> theta[2,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.12
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[2,3,3]     0.62    0.00 0.13    0.35    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.12    0.20    0.28    0.47
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,3]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.76    0.89
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.30
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[3,3,2]     0.18    0.00 0.10    0.02    0.10    0.16    0.24    0.41
#> theta[3,3,3]     0.51    0.00 0.13    0.25    0.41    0.51    0.60    0.76
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.30    0.48
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[3,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.44
#> theta[3,4,4]     0.63    0.00 0.15    0.32    0.53    0.64    0.74    0.89
#> theta[4,1,1]     0.82    0.00 0.07    0.66    0.78    0.83    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.73    0.00 0.09    0.55    0.67    0.73    0.79    0.88
#> theta[4,2,3]     0.13    0.00 0.06    0.03    0.09    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.10    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[4,3,2]     0.08    0.00 0.08    0.00    0.03    0.06    0.11    0.28
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.60    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.39
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.39
#> theta[4,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.90
#> theta[5,1,1]     0.86    0.00 0.06    0.72    0.83    0.87    0.91    0.96
#> theta[5,1,2]     0.06    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.15    0.00 0.07    0.04    0.10    0.14    0.19    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.48    0.59    0.65    0.71    0.81
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.38
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.61    0.70    0.85
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.13    0.21    0.38
#> theta[5,4,1]     0.11    0.00 0.10    0.00    0.04    0.08    0.16    0.36
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.19    0.42
#> theta[5,4,4]     0.64    0.00 0.15    0.30    0.53    0.65    0.75    0.90
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_p_z[1,2]   -20.06    0.04 2.31  -25.19  -21.44  -19.90  -18.41  -16.14
#> log_p_z[1,3]   -24.87    0.07 4.35  -34.75  -27.29  -24.38  -21.89  -17.67
#> log_p_z[1,4]   -21.82    0.07 4.24  -31.49  -24.27  -21.36  -18.81  -14.96
#> log_p_z[2,1]   -29.56    0.06 4.12  -39.04  -32.02  -29.15  -26.60  -22.78
#> log_p_z[2,2]   -20.99    0.04 2.36  -26.05  -22.43  -20.80  -19.34  -16.93
#> log_p_z[2,3]    -7.69    0.02 1.21  -10.33   -8.39   -7.55   -6.86   -5.67
#> log_p_z[2,4]   -16.33    0.07 4.17  -25.79  -18.81  -15.84  -13.35   -9.55
#> log_p_z[3,1]   -11.56    0.02 1.59  -15.07  -12.53  -11.36  -10.41   -8.95
#> log_p_z[3,2]   -10.84    0.02 1.48  -14.20  -11.71  -10.68   -9.80   -8.45
#> log_p_z[3,3]   -23.35    0.06 3.44  -30.87  -25.47  -23.04  -20.93  -17.41
#> log_p_z[3,4]   -21.87    0.06 3.51  -29.91  -23.95  -21.53  -19.40  -15.91
#> log_p_z[4,1]   -14.64    0.02 1.89  -18.84  -15.79  -14.47  -13.30  -11.32
#> log_p_z[4,2]    -7.57    0.01 0.96   -9.60   -8.16   -7.49   -6.89   -5.96
#> log_p_z[4,3]   -20.69    0.07 4.25  -30.59  -23.13  -20.20  -17.62  -14.01
#> log_p_z[4,4]   -21.91    0.07 4.22  -30.99  -24.55  -21.46  -18.86  -15.01
#> log_p_z[5,1]   -21.37    0.04 2.55  -26.92  -22.95  -21.15  -19.58  -16.98
#> log_p_z[5,2]    -3.70    0.01 0.44   -4.61   -3.99   -3.68   -3.40   -2.91
#> log_p_z[5,3]   -18.68    0.07 4.12  -28.40  -20.97  -18.27  -15.71  -12.20
#> log_p_z[5,4]   -21.93    0.07 4.27  -31.72  -24.50  -21.52  -18.89  -14.78
#> log_p_z[6,1]   -21.41    0.03 2.57  -27.03  -23.00  -21.22  -19.60  -16.90
#> log_p_z[6,2]    -5.46    0.01 0.67   -6.91   -5.87   -5.41   -5.02   -4.26
#> log_p_z[6,3]   -17.46    0.07 3.98  -26.83  -19.67  -16.99  -14.57  -11.38
#> log_p_z[6,4]   -21.68    0.07 4.27  -31.12  -24.15  -21.22  -18.64  -14.78
#> log_p_z[7,1]    -6.00    0.01 0.62   -7.37   -6.38   -5.95   -5.57   -4.89
#> log_p_z[7,2]   -15.00    0.03 1.71  -18.84  -16.05  -14.85  -13.79  -12.05
#> log_p_z[7,3]   -24.27    0.06 3.56  -32.25  -26.45  -23.91  -21.75  -18.15
#> log_p_z[7,4]   -21.85    0.07 3.55  -29.79  -23.96  -21.55  -19.35  -15.97
#> log_p_z[8,1]   -29.58    0.06 4.17  -39.06  -32.04  -29.08  -26.70  -22.78
#> log_p_z[8,2]   -18.25    0.03 2.10  -22.84  -19.51  -18.12  -16.76  -14.56
#> log_p_z[8,3]    -5.73    0.01 0.83   -7.49   -6.26   -5.68   -5.15   -4.26
#> log_p_z[8,4]   -18.54    0.07 4.25  -28.16  -20.96  -18.05  -15.48  -11.72
#> log_p_z[9,1]   -19.93    0.03 2.25  -24.80  -21.31  -19.78  -18.37  -15.93
#> log_p_z[9,2]    -4.51    0.01 0.56   -5.77   -4.85   -4.47   -4.13   -3.51
#> log_p_z[9,3]   -19.62    0.07 4.21  -29.46  -22.09  -19.20  -16.60  -12.83
#> log_p_z[9,4]   -21.70    0.07 4.17  -31.19  -24.16  -21.29  -18.75  -14.86
#> log_p_z[10,1]  -22.14    0.03 2.35  -27.32  -23.58  -21.99  -20.47  -18.09
#> log_p_z[10,2]   -7.11    0.01 0.72   -8.68   -7.54   -7.05   -6.61   -5.87
#> log_p_z[10,3]  -16.81    0.05 3.09  -23.86  -18.67  -16.54  -14.59  -11.61
#> log_p_z[10,4]  -21.39    0.06 3.44  -28.88  -23.53  -21.09  -18.93  -15.50
#> log_p_z[11,1]  -29.48    0.07 4.16  -38.59  -31.90  -29.04  -26.53  -22.73
#> log_p_z[11,2]  -29.30    0.07 4.38  -39.33  -31.75  -28.84  -26.24  -22.22
#> log_p_z[11,3]  -18.21    0.06 3.34  -26.06  -19.96  -17.72  -15.96  -13.09
#> log_p_z[11,4]   -5.79    0.01 0.96   -7.81   -6.40   -5.74   -5.11   -4.10
#> log_p_z[12,1]  -22.96    0.04 2.74  -28.86  -24.65  -22.70  -21.03  -18.26
#> log_p_z[12,2]   -9.57    0.02 1.28  -12.59  -10.26   -9.42   -8.70   -7.55
#> log_p_z[12,3]  -15.09    0.07 3.97  -24.23  -17.31  -14.65  -12.16   -8.90
#> log_p_z[12,4]  -19.26    0.07 4.08  -28.93  -21.56  -18.76  -16.35  -12.86
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_p_z[13,2]  -20.06    0.04 2.31  -25.19  -21.44  -19.90  -18.41  -16.14
#> log_p_z[13,3]  -24.87    0.07 4.35  -34.75  -27.29  -24.38  -21.89  -17.67
#> log_p_z[13,4]  -21.82    0.07 4.24  -31.49  -24.27  -21.36  -18.81  -14.96
#> log_p_z[14,1]  -18.98    0.03 2.38  -24.11  -20.45  -18.78  -17.31  -14.93
#> log_p_z[14,2]   -6.25    0.01 0.87   -8.16   -6.75   -6.17   -5.66   -4.80
#> log_p_z[14,3]  -18.65    0.07 4.18  -28.25  -21.24  -18.19  -15.52  -12.07
#> log_p_z[14,4]  -21.91    0.07 4.31  -31.82  -24.44  -21.43  -18.87  -14.79
#> log_p_z[15,1]   -4.42    0.01 0.40   -5.28   -4.66   -4.38   -4.14   -3.71
#> log_p_z[15,2]  -17.29    0.03 1.86  -21.34  -18.43  -17.15  -15.96  -14.09
#> log_p_z[15,3]  -24.32    0.06 3.56  -32.28  -26.40  -23.96  -21.82  -18.12
#> log_p_z[15,4]  -21.83    0.06 3.50  -29.74  -23.92  -21.56  -19.36  -15.91
#> log_p_z[16,1]   -3.79    0.01 0.54   -4.95   -4.14   -3.75   -3.41   -2.88
#> log_p_z[16,2]  -17.76    0.04 2.19  -22.71  -19.09  -17.57  -16.22  -14.08
#> log_p_z[16,3]  -24.82    0.07 4.34  -34.81  -27.29  -24.29  -21.83  -17.72
#> log_p_z[16,4]  -21.84    0.08 4.27  -31.66  -24.38  -21.28  -18.84  -15.02
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_p_z[17,2]  -20.06    0.04 2.31  -25.19  -21.44  -19.90  -18.41  -16.14
#> log_p_z[17,3]  -24.87    0.07 4.35  -34.75  -27.29  -24.38  -21.89  -17.67
#> log_p_z[17,4]  -21.82    0.07 4.24  -31.49  -24.27  -21.36  -18.81  -14.96
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_p_z[18,2]  -20.06    0.04 2.31  -25.19  -21.44  -19.90  -18.41  -16.14
#> log_p_z[18,3]  -24.87    0.07 4.35  -34.75  -27.29  -24.38  -21.89  -17.67
#> log_p_z[18,4]  -21.82    0.07 4.24  -31.49  -24.27  -21.36  -18.81  -14.96
#> log_p_z[19,1]  -16.35    0.03 1.91  -20.54  -17.51  -16.22  -15.03  -12.98
#> log_p_z[19,2]   -4.65    0.01 0.60   -5.92   -5.04   -4.61   -4.22   -3.59
#> log_p_z[19,3]  -22.13    0.07 4.33  -31.85  -24.77  -21.68  -18.98  -15.09
#> log_p_z[19,4]  -21.93    0.07 4.20  -31.40  -24.43  -21.51  -18.92  -15.00
#> log_p_z[20,1]  -17.99    0.03 2.25  -22.98  -19.36  -17.77  -16.42  -14.10
#> log_p_z[20,2]   -7.09    0.01 0.95   -9.22   -7.64   -7.02   -6.44   -5.45
#> log_p_z[20,3]  -20.01    0.07 4.11  -29.72  -22.46  -19.47  -17.07  -13.50
#> log_p_z[20,4]  -21.64    0.07 4.21  -30.88  -23.98  -21.22  -18.70  -14.87
#> log_p_z[21,1]  -19.52    0.03 2.33  -24.57  -20.95  -19.35  -17.90  -15.61
#> log_p_z[21,2]   -3.04    0.00 0.35   -3.75   -3.27   -3.03   -2.80   -2.39
#> log_p_z[21,3]  -21.18    0.07 4.23  -30.96  -23.67  -20.73  -18.19  -14.29
#> log_p_z[21,4]  -21.92    0.07 4.24  -31.61  -24.42  -21.42  -18.91  -14.97
#> log_p_z[22,1]  -16.35    0.03 1.91  -20.54  -17.51  -16.22  -15.03  -12.98
#> log_p_z[22,2]   -4.65    0.01 0.60   -5.92   -5.04   -4.61   -4.22   -3.59
#> log_p_z[22,3]  -22.13    0.07 4.33  -31.85  -24.77  -21.68  -18.98  -15.09
#> log_p_z[22,4]  -21.93    0.07 4.20  -31.40  -24.43  -21.51  -18.92  -15.00
#> log_p_z[23,1]  -21.37    0.04 2.55  -26.92  -22.95  -21.15  -19.58  -16.98
#> log_p_z[23,2]   -3.70    0.01 0.44   -4.61   -3.99   -3.68   -3.40   -2.91
#> log_p_z[23,3]  -18.68    0.07 4.12  -28.40  -20.97  -18.27  -15.71  -12.20
#> log_p_z[23,4]  -21.93    0.07 4.27  -31.72  -24.50  -21.52  -18.89  -14.78
#> log_p_z[24,1]  -17.32    0.03 2.12  -21.92  -18.65  -17.13  -15.82  -13.78
#> log_p_z[24,2]   -5.81    0.01 0.63   -7.27   -6.18   -5.75   -5.37   -4.78
#> log_p_z[24,3]  -21.73    0.06 3.35  -29.05  -23.82  -21.43  -19.35  -15.93
#> log_p_z[24,4]  -21.91    0.06 3.50  -29.85  -24.00  -21.61  -19.42  -16.04
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_p_z[25,2]  -20.06    0.04 2.31  -25.19  -21.44  -19.90  -18.41  -16.14
#> log_p_z[25,3]  -24.87    0.07 4.35  -34.75  -27.29  -24.38  -21.89  -17.67
#> log_p_z[25,4]  -21.82    0.07 4.24  -31.49  -24.27  -21.36  -18.81  -14.96
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_p_z[26,2]  -20.06    0.04 2.31  -25.19  -21.44  -19.90  -18.41  -16.14
#> log_p_z[26,3]  -24.87    0.07 4.35  -34.75  -27.29  -24.38  -21.89  -17.67
#> log_p_z[26,4]  -21.82    0.07 4.24  -31.49  -24.27  -21.36  -18.81  -14.96
#> log_p_z[27,1]  -21.73    0.03 2.42  -26.92  -23.24  -21.59  -20.02  -17.56
#> log_p_z[27,2]   -5.64    0.01 0.54   -6.78   -5.97   -5.61   -5.26   -4.72
#> log_p_z[27,3]  -18.37    0.05 3.14  -25.36  -20.30  -18.04  -16.17  -13.09
#> log_p_z[27,4]  -21.60    0.06 3.47  -29.25  -23.68  -21.28  -19.14  -15.71
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_p_z[28,2]  -20.06    0.04 2.31  -25.19  -21.44  -19.90  -18.41  -16.14
#> log_p_z[28,3]  -24.87    0.07 4.35  -34.75  -27.29  -24.38  -21.89  -17.67
#> log_p_z[28,4]  -21.82    0.07 4.24  -31.49  -24.27  -21.36  -18.81  -14.96
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_p_z[29,2]  -20.06    0.04 2.31  -25.19  -21.44  -19.90  -18.41  -16.14
#> log_p_z[29,3]  -24.87    0.07 4.35  -34.75  -27.29  -24.38  -21.89  -17.67
#> log_p_z[29,4]  -21.82    0.07 4.24  -31.49  -24.27  -21.36  -18.81  -14.96
#> log_p_z[30,1]   -6.80    0.01 0.82   -8.61   -7.30   -6.71   -6.22   -5.44
#> log_p_z[30,2]  -14.74    0.03 1.71  -18.62  -15.74  -14.56  -13.55  -11.85
#> log_p_z[30,3]  -24.34    0.06 3.53  -32.45  -26.50  -23.99  -21.86  -18.39
#> log_p_z[30,4]  -21.86    0.06 3.49  -29.71  -23.98  -21.50  -19.39  -15.96
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_p_z[31,2]  -20.06    0.04 2.31  -25.19  -21.44  -19.90  -18.41  -16.14
#> log_p_z[31,3]  -24.87    0.07 4.35  -34.75  -27.29  -24.38  -21.89  -17.67
#> log_p_z[31,4]  -21.82    0.07 4.24  -31.49  -24.27  -21.36  -18.81  -14.96
#> log_p_z[32,1]  -29.55    0.06 4.15  -39.09  -31.98  -29.11  -26.63  -22.67
#> log_p_z[32,2]  -14.79    0.03 1.76  -18.61  -15.88  -14.61  -13.58  -11.80
#> log_p_z[32,3]   -6.07    0.01 0.93   -8.11   -6.62   -5.99   -5.43   -4.46
#> log_p_z[32,4]  -20.71    0.08 4.45  -30.68  -23.40  -20.25  -17.55  -13.41
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_p_z[33,2]  -20.06    0.04 2.31  -25.19  -21.44  -19.90  -18.41  -16.14
#> log_p_z[33,3]  -24.87    0.07 4.35  -34.75  -27.29  -24.38  -21.89  -17.67
#> log_p_z[33,4]  -21.82    0.07 4.24  -31.49  -24.27  -21.36  -18.81  -14.96
#> log_p_z[34,1]  -19.52    0.03 2.33  -24.57  -20.95  -19.35  -17.90  -15.61
#> log_p_z[34,2]   -3.04    0.00 0.35   -3.75   -3.27   -3.03   -2.80   -2.39
#> log_p_z[34,3]  -21.18    0.07 4.23  -30.96  -23.67  -20.73  -18.19  -14.29
#> log_p_z[34,4]  -21.92    0.07 4.24  -31.61  -24.42  -21.42  -18.91  -14.97
#> log_p_z[35,1]  -22.51    0.04 2.63  -28.18  -24.08  -22.29  -20.63  -18.01
#> log_p_z[35,2]   -5.52    0.01 0.69   -7.08   -5.95   -5.48   -5.03   -4.34
#> log_p_z[35,3]  -16.05    0.07 4.00  -25.23  -18.42  -15.62  -13.11   -9.89
#> log_p_z[35,4]  -21.88    0.07 4.26  -31.85  -24.37  -21.40  -18.83  -14.93
#> log_p_z[36,1]  -29.55    0.05 3.51  -37.41  -31.73  -29.27  -27.10  -23.51
#> log_p_z[36,2]  -21.84    0.04 2.30  -26.90  -23.22  -21.65  -20.25  -17.97
#> log_p_z[36,3]  -10.50    0.03 1.74  -14.72  -11.35  -10.25   -9.29   -7.86
#> log_p_z[36,4]  -13.97    0.06 3.18  -20.87  -15.93  -13.74  -11.61   -8.81
#> log_p_z[37,1]  -18.47    0.03 2.24  -23.31  -19.83  -18.30  -16.87  -14.71
#> log_p_z[37,2]   -7.63    0.01 0.79   -9.34   -8.12   -7.56   -7.08   -6.27
#> log_p_z[37,3]  -19.11    0.05 3.19  -26.21  -21.12  -18.79  -16.82  -13.72
#> log_p_z[37,4]  -21.86    0.06 3.47  -29.52  -23.92  -21.51  -19.40  -16.07
#> log_p_z[38,1]  -25.13    0.04 2.72  -31.11  -26.80  -24.96  -23.20  -20.36
#> log_p_z[38,2]   -9.59    0.01 1.01  -11.80  -10.22   -9.51   -8.89   -7.82
#> log_p_z[38,3]  -11.69    0.05 2.73  -17.96  -13.32  -11.40   -9.65   -7.45
#> log_p_z[38,4]  -21.34    0.06 3.52  -29.36  -23.47  -21.02  -18.83  -15.37
#> log_p_z[39,1]  -29.18    0.06 4.16  -38.80  -31.61  -28.70  -26.28  -22.37
#> log_p_z[39,2]  -16.78    0.03 2.01  -21.23  -18.00  -16.62  -15.36  -13.20
#> log_p_z[39,3]   -7.29    0.02 1.13   -9.78   -7.98   -7.19   -6.50   -5.37
#> log_p_z[39,4]  -18.76    0.07 4.18  -28.11  -21.15  -18.19  -15.79  -12.21
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_p_z[40,2]  -20.06    0.04 2.31  -25.19  -21.44  -19.90  -18.41  -16.14
#> log_p_z[40,3]  -24.87    0.07 4.35  -34.75  -27.29  -24.38  -21.89  -17.67
#> log_p_z[40,4]  -21.82    0.07 4.24  -31.49  -24.27  -21.36  -18.81  -14.96
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_p_z[41,2]  -20.06    0.04 2.31  -25.19  -21.44  -19.90  -18.41  -16.14
#> log_p_z[41,3]  -24.87    0.07 4.35  -34.75  -27.29  -24.38  -21.89  -17.67
#> log_p_z[41,4]  -21.82    0.07 4.24  -31.49  -24.27  -21.36  -18.81  -14.96
#> log_p_z[42,1]   -6.00    0.01 0.62   -7.37   -6.38   -5.95   -5.57   -4.89
#> log_p_z[42,2]  -15.00    0.03 1.71  -18.84  -16.05  -14.85  -13.79  -12.05
#> log_p_z[42,3]  -24.27    0.06 3.56  -32.25  -26.45  -23.91  -21.75  -18.15
#> log_p_z[42,4]  -21.85    0.07 3.55  -29.79  -23.96  -21.55  -19.35  -15.97
#> log_p_z[43,1]  -21.73    0.03 2.42  -26.92  -23.24  -21.59  -20.02  -17.56
#> log_p_z[43,2]   -5.64    0.01 0.54   -6.78   -5.97   -5.61   -5.26   -4.72
#> log_p_z[43,3]  -18.37    0.05 3.14  -25.36  -20.30  -18.04  -16.17  -13.09
#> log_p_z[43,4]  -21.60    0.06 3.47  -29.25  -23.68  -21.28  -19.14  -15.71
#> log_p_z[44,1]   -4.42    0.01 0.40   -5.28   -4.66   -4.38   -4.14   -3.71
#> log_p_z[44,2]  -17.29    0.03 1.86  -21.34  -18.43  -17.15  -15.96  -14.09
#> log_p_z[44,3]  -24.32    0.06 3.56  -32.28  -26.40  -23.96  -21.82  -18.12
#> log_p_z[44,4]  -21.83    0.06 3.50  -29.74  -23.92  -21.56  -19.36  -15.91
#> log_p_z[45,1]  -19.52    0.03 2.33  -24.57  -20.95  -19.35  -17.90  -15.61
#> log_p_z[45,2]   -3.04    0.00 0.35   -3.75   -3.27   -3.03   -2.80   -2.39
#> log_p_z[45,3]  -21.18    0.07 4.23  -30.96  -23.67  -20.73  -18.19  -14.29
#> log_p_z[45,4]  -21.92    0.07 4.24  -31.61  -24.42  -21.42  -18.91  -14.97
#> log_lik[1]      -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_lik[2]      -7.63    0.02 1.11  -10.01   -8.35   -7.54   -6.85   -5.67
#> log_lik[3]      -9.95    0.01 0.90  -11.80  -10.54   -9.92   -9.33   -8.28
#> log_lik[4]      -7.56    0.01 0.95   -9.58   -8.16   -7.48   -6.89   -5.96
#> log_lik[5]      -3.70    0.01 0.44   -4.61   -3.99   -3.68   -3.40   -2.91
#> log_lik[6]      -5.46    0.01 0.67   -6.91   -5.87   -5.41   -5.02   -4.26
#> log_lik[7]      -6.00    0.01 0.62   -7.37   -6.38   -5.95   -5.57   -4.89
#> log_lik[8]      -5.73    0.01 0.83   -7.49   -6.26   -5.67   -5.15   -4.26
#> log_lik[9]      -4.51    0.01 0.56   -5.77   -4.85   -4.47   -4.13   -3.51
#> log_lik[10]     -7.11    0.01 0.71   -8.67   -7.54   -7.05   -6.61   -5.87
#> log_lik[11]     -5.79    0.01 0.96   -7.81   -6.40   -5.74   -5.11   -4.10
#> log_lik[12]     -9.32    0.01 1.02  -11.43   -9.98   -9.25   -8.59   -7.51
#> log_lik[13]     -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_lik[14]     -6.25    0.01 0.87   -8.16   -6.75   -6.17   -5.66   -4.80
#> log_lik[15]     -4.42    0.01 0.40   -5.28   -4.66   -4.38   -4.14   -3.71
#> log_lik[16]     -3.79    0.01 0.54   -4.95   -4.14   -3.75   -3.41   -2.88
#> log_lik[17]     -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_lik[18]     -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_lik[19]     -4.65    0.01 0.60   -5.92   -5.04   -4.61   -4.22   -3.59
#> log_lik[20]     -7.09    0.01 0.95   -9.21   -7.64   -7.01   -6.44   -5.45
#> log_lik[21]     -3.04    0.00 0.35   -3.75   -3.27   -3.03   -2.80   -2.39
#> log_lik[22]     -4.65    0.01 0.60   -5.92   -5.04   -4.61   -4.22   -3.59
#> log_lik[23]     -3.70    0.01 0.44   -4.61   -3.99   -3.68   -3.40   -2.91
#> log_lik[24]     -5.81    0.01 0.63   -7.27   -6.18   -5.75   -5.37   -4.78
#> log_lik[25]     -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_lik[26]     -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_lik[27]     -5.64    0.01 0.53   -6.78   -5.97   -5.61   -5.26   -4.72
#> log_lik[28]     -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_lik[29]     -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_lik[30]     -6.80    0.01 0.82   -8.60   -7.30   -6.71   -6.22   -5.44
#> log_lik[31]     -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_lik[32]     -6.07    0.01 0.93   -8.11   -6.62   -5.99   -5.43   -4.46
#> log_lik[33]     -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_lik[34]     -3.04    0.00 0.35   -3.75   -3.27   -3.03   -2.80   -2.39
#> log_lik[35]     -5.52    0.01 0.69   -7.05   -5.94   -5.48   -5.03   -4.34
#> log_lik[36]     -9.91    0.02 1.18  -12.43  -10.68   -9.83   -9.06   -7.80
#> log_lik[37]     -7.63    0.01 0.79   -9.33   -8.12   -7.56   -7.08   -6.27
#> log_lik[38]     -8.98    0.01 0.87  -10.71   -9.55   -8.97   -8.40   -7.31
#> log_lik[39]     -7.29    0.02 1.12   -9.77   -7.98   -7.19   -6.50   -5.37
#> log_lik[40]     -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_lik[41]     -2.21    0.00 0.30   -2.85   -2.41   -2.20   -2.01   -1.66
#> log_lik[42]     -6.00    0.01 0.62   -7.37   -6.38   -5.95   -5.57   -4.89
#> log_lik[43]     -5.64    0.01 0.53   -6.78   -5.97   -5.61   -5.26   -4.72
#> log_lik[44]     -4.42    0.01 0.40   -5.28   -4.66   -4.38   -4.14   -3.71
#> log_lik[45]     -3.04    0.00 0.35   -3.75   -3.27   -3.03   -2.80   -2.39
#> lp__          -463.46    0.18 6.23 -476.31 -467.46 -463.13 -459.07 -452.41
#>               n_eff Rhat
#> pi[1]          7995    1
#> pi[2]          9042    1
#> pi[3]          6300    1
#> pi[4]          5983    1
#> theta[1,1,1]   7897    1
#> theta[1,1,2]   6515    1
#> theta[1,1,3]   5162    1
#> theta[1,1,4]   4508    1
#> theta[1,2,1]   5415    1
#> theta[1,2,2]   6605    1
#> theta[1,2,3]   5778    1
#> theta[1,2,4]   5460    1
#> theta[1,3,1]   6836    1
#> theta[1,3,2]   4486    1
#> theta[1,3,3]   5465    1
#> theta[1,3,4]   5782    1
#> theta[1,4,1]   5868    1
#> theta[1,4,2]   6635    1
#> theta[1,4,3]   3737    1
#> theta[1,4,4]   4365    1
#> theta[2,1,1]   9935    1
#> theta[2,1,2]   7601    1
#> theta[2,1,3]   4739    1
#> theta[2,1,4]   5140    1
#> theta[2,2,1]   6399    1
#> theta[2,2,2]   6780    1
#> theta[2,2,3]   6490    1
#> theta[2,2,4]   5312    1
#> theta[2,3,1]   6005    1
#> theta[2,3,2]   6504    1
#> theta[2,3,3]   5939    1
#> theta[2,3,4]   6222    1
#> theta[2,4,1]   7410    1
#> theta[2,4,2]   5998    1
#> theta[2,4,3]   5936    1
#> theta[2,4,4]   6638    1
#> theta[3,1,1]   8024    1
#> theta[3,1,2]   5568    1
#> theta[3,1,3]   5195    1
#> theta[3,1,4]   4620    1
#> theta[3,2,1]   6208    1
#> theta[3,2,2]   7214    1
#> theta[3,2,3]   6289    1
#> theta[3,2,4]   4987    1
#> theta[3,3,1]   5901    1
#> theta[3,3,2]   7793    1
#> theta[3,3,3]   7193    1
#> theta[3,3,4]   7260    1
#> theta[3,4,1]   6216    1
#> theta[3,4,2]   6469    1
#> theta[3,4,3]   4682    1
#> theta[3,4,4]   5943    1
#> theta[4,1,1]   8530    1
#> theta[4,1,2]   6601    1
#> theta[4,1,3]   5050    1
#> theta[4,1,4]   4969    1
#> theta[4,2,1]   7997    1
#> theta[4,2,2]   7290    1
#> theta[4,2,3]   6482    1
#> theta[4,2,4]   5203    1
#> theta[4,3,1]   5892    1
#> theta[4,3,2]   6086    1
#> theta[4,3,3]   5851    1
#> theta[4,3,4]   6023    1
#> theta[4,4,1]   6101    1
#> theta[4,4,2]   5681    1
#> theta[4,4,3]   7099    1
#> theta[4,4,4]   6257    1
#> theta[5,1,1]   6828    1
#> theta[5,1,2]   4606    1
#> theta[5,1,3]   4757    1
#> theta[5,1,4]   4936    1
#> theta[5,2,1]   8015    1
#> theta[5,2,2]   6683    1
#> theta[5,2,3]   5866    1
#> theta[5,2,4]   4928    1
#> theta[5,3,1]   6897    1
#> theta[5,3,2]   6998    1
#> theta[5,3,3]   6779    1
#> theta[5,3,4]   6729    1
#> theta[5,4,1]   5972    1
#> theta[5,4,2]   5790    1
#> theta[5,4,3]   5782    1
#> theta[5,4,4]   5313    1
#> log_p_z[1,1]   7789    1
#> log_p_z[1,2]   3944    1
#> log_p_z[1,3]   3360    1
#> log_p_z[1,4]   3458    1
#> log_p_z[2,1]   5021    1
#> log_p_z[2,2]   3822    1
#> log_p_z[2,3]   3619    1
#> log_p_z[2,4]   3327    1
#> log_p_z[3,1]   4412    1
#> log_p_z[3,2]   3674    1
#> log_p_z[3,3]   3350    1
#> log_p_z[3,4]   3255    1
#> log_p_z[4,1]   5881    1
#> log_p_z[4,2]   4588    1
#> log_p_z[4,3]   3331    1
#> log_p_z[4,4]   4044    1
#> log_p_z[5,1]   5128    1
#> log_p_z[5,2]   6313    1
#> log_p_z[5,3]   3517    1
#> log_p_z[5,4]   4086    1
#> log_p_z[6,1]   5469    1
#> log_p_z[6,2]   6155    1
#> log_p_z[6,3]   3442    1
#> log_p_z[6,4]   3991    1
#> log_p_z[7,1]   5654    1
#> log_p_z[7,2]   3576    1
#> log_p_z[7,3]   3320    1
#> log_p_z[7,4]   2909    1
#> log_p_z[8,1]   4569    1
#> log_p_z[8,2]   4472    1
#> log_p_z[8,3]   6200    1
#> log_p_z[8,4]   3366    1
#> log_p_z[9,1]   4408    1
#> log_p_z[9,2]   5006    1
#> log_p_z[9,3]   3553    1
#> log_p_z[9,4]   3966    1
#> log_p_z[10,1]  4511    1
#> log_p_z[10,2]  5490    1
#> log_p_z[10,3]  3592    1
#> log_p_z[10,4]  3659    1
#> log_p_z[11,1]  4008    1
#> log_p_z[11,2]  3512    1
#> log_p_z[11,3]  2911    1
#> log_p_z[11,4]  5856    1
#> log_p_z[12,1]  4968    1
#> log_p_z[12,2]  3808    1
#> log_p_z[12,3]  3518    1
#> log_p_z[12,4]  3749    1
#> log_p_z[13,1]  7789    1
#> log_p_z[13,2]  3944    1
#> log_p_z[13,3]  3360    1
#> log_p_z[13,4]  3458    1
#> log_p_z[14,1]  5098    1
#> log_p_z[14,2]  4784    1
#> log_p_z[14,3]  3540    1
#> log_p_z[14,4]  3864    1
#> log_p_z[15,1]  6184    1
#> log_p_z[15,2]  3961    1
#> log_p_z[15,3]  3301    1
#> log_p_z[15,4]  3372    1
#> log_p_z[16,1]  6590    1
#> log_p_z[16,2]  3646    1
#> log_p_z[16,3]  3355    1
#> log_p_z[16,4]  3004    1
#> log_p_z[17,1]  7789    1
#> log_p_z[17,2]  3944    1
#> log_p_z[17,3]  3360    1
#> log_p_z[17,4]  3458    1
#> log_p_z[18,1]  7789    1
#> log_p_z[18,2]  3944    1
#> log_p_z[18,3]  3360    1
#> log_p_z[18,4]  3458    1
#> log_p_z[19,1]  5139    1
#> log_p_z[19,2]  6639    1
#> log_p_z[19,3]  3432    1
#> log_p_z[19,4]  4029    1
#> log_p_z[20,1]  5032    1
#> log_p_z[20,2]  5245    1
#> log_p_z[20,3]  3437    1
#> log_p_z[20,4]  3959    1
#> log_p_z[21,1]  4632    1
#> log_p_z[21,2]  6121    1
#> log_p_z[21,3]  3518    1
#> log_p_z[21,4]  4032    1
#> log_p_z[22,1]  5139    1
#> log_p_z[22,2]  6639    1
#> log_p_z[22,3]  3432    1
#> log_p_z[22,4]  4029    1
#> log_p_z[23,1]  5128    1
#> log_p_z[23,2]  6313    1
#> log_p_z[23,3]  3517    1
#> log_p_z[23,4]  4086    1
#> log_p_z[24,1]  4503    1
#> log_p_z[24,2]  3917    1
#> log_p_z[24,3]  3535    1
#> log_p_z[24,4]  3743    1
#> log_p_z[25,1]  7789    1
#> log_p_z[25,2]  3944    1
#> log_p_z[25,3]  3360    1
#> log_p_z[25,4]  3458    1
#> log_p_z[26,1]  7789    1
#> log_p_z[26,2]  3944    1
#> log_p_z[26,3]  3360    1
#> log_p_z[26,4]  3458    1
#> log_p_z[27,1]  4791    1
#> log_p_z[27,2]  5764    1
#> log_p_z[27,3]  3560    1
#> log_p_z[27,4]  3803    1
#> log_p_z[28,1]  7789    1
#> log_p_z[28,2]  3944    1
#> log_p_z[28,3]  3360    1
#> log_p_z[28,4]  3458    1
#> log_p_z[29,1]  7789    1
#> log_p_z[29,2]  3944    1
#> log_p_z[29,3]  3360    1
#> log_p_z[29,4]  3458    1
#> log_p_z[30,1]  6009    1
#> log_p_z[30,2]  3987    1
#> log_p_z[30,3]  3145    1
#> log_p_z[30,4]  3675    1
#> log_p_z[31,1]  7789    1
#> log_p_z[31,2]  3944    1
#> log_p_z[31,3]  3360    1
#> log_p_z[31,4]  3458    1
#> log_p_z[32,1]  5142    1
#> log_p_z[32,2]  4825    1
#> log_p_z[32,3]  5215    1
#> log_p_z[32,4]  3283    1
#> log_p_z[33,1]  7789    1
#> log_p_z[33,2]  3944    1
#> log_p_z[33,3]  3360    1
#> log_p_z[33,4]  3458    1
#> log_p_z[34,1]  4632    1
#> log_p_z[34,2]  6121    1
#> log_p_z[34,3]  3518    1
#> log_p_z[34,4]  4032    1
#> log_p_z[35,1]  4665    1
#> log_p_z[35,2]  5823    1
#> log_p_z[35,3]  3555    1
#> log_p_z[35,4]  3913    1
#> log_p_z[36,1]  4279    1
#> log_p_z[36,2]  2999    1
#> log_p_z[36,3]  2645    1
#> log_p_z[36,4]  3128    1
#> log_p_z[37,1]  4175    1
#> log_p_z[37,2]  4445    1
#> log_p_z[37,3]  3704    1
#> log_p_z[37,4]  3478    1
#> log_p_z[38,1]  4505    1
#> log_p_z[38,2]  4719    1
#> log_p_z[38,3]  3656    1
#> log_p_z[38,4]  3353    1
#> log_p_z[39,1]  5021    1
#> log_p_z[39,2]  4503    1
#> log_p_z[39,3]  4953    1
#> log_p_z[39,4]  3319    1
#> log_p_z[40,1]  7789    1
#> log_p_z[40,2]  3944    1
#> log_p_z[40,3]  3360    1
#> log_p_z[40,4]  3458    1
#> log_p_z[41,1]  7789    1
#> log_p_z[41,2]  3944    1
#> log_p_z[41,3]  3360    1
#> log_p_z[41,4]  3458    1
#> log_p_z[42,1]  5654    1
#> log_p_z[42,2]  3576    1
#> log_p_z[42,3]  3320    1
#> log_p_z[42,4]  2909    1
#> log_p_z[43,1]  4791    1
#> log_p_z[43,2]  5764    1
#> log_p_z[43,3]  3560    1
#> log_p_z[43,4]  3803    1
#> log_p_z[44,1]  6184    1
#> log_p_z[44,2]  3961    1
#> log_p_z[44,3]  3301    1
#> log_p_z[44,4]  3372    1
#> log_p_z[45,1]  4632    1
#> log_p_z[45,2]  6121    1
#> log_p_z[45,3]  3518    1
#> log_p_z[45,4]  4032    1
#> log_lik[1]     7789    1
#> log_lik[2]     4527    1
#> log_lik[3]     4568    1
#> log_lik[4]     4708    1
#> log_lik[5]     6312    1
#> log_lik[6]     6167    1
#> log_lik[7]     5657    1
#> log_lik[8]     6197    1
#> log_lik[9]     5007    1
#> log_lik[10]    5522    1
#> log_lik[11]    5857    1
#> log_lik[12]    5884    1
#> log_lik[13]    7789    1
#> log_lik[14]    4794    1
#> log_lik[15]    6184    1
#> log_lik[16]    6590    1
#> log_lik[17]    7789    1
#> log_lik[18]    7789    1
#> log_lik[19]    6640    1
#> log_lik[20]    5255    1
#> log_lik[21]    6121    1
#> log_lik[22]    6640    1
#> log_lik[23]    6312    1
#> log_lik[24]    3917    1
#> log_lik[25]    7789    1
#> log_lik[26]    7789    1
#> log_lik[27]    5764    1
#> log_lik[28]    7789    1
#> log_lik[29]    7789    1
#> log_lik[30]    6022    1
#> log_lik[31]    7789    1
#> log_lik[32]    5241    1
#> log_lik[33]    7789    1
#> log_lik[34]    6121    1
#> log_lik[35]    5850    1
#> log_lik[36]    4106    1
#> log_lik[37]    4449    1
#> log_lik[38]    4489    1
#> log_lik[39]    4942    1
#> log_lik[40]    7789    1
#> log_lik[41]    7789    1
#> log_lik[42]    5657    1
#> log_lik[43]    5764    1
#> log_lik[44]    6184    1
#> log_lik[45]    6121    1
#> lp__           1211    1
#> 
#> Samples were drawn using NUTS(diag_e) at Sun Mar  5 00:40:49 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }