Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.37    0.00 0.06    0.26    0.33    0.37    0.42    0.50
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.93
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.01    0.01    0.02    0.05
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.17
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.59    0.74    0.80    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.00    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,3]     0.13    0.00 0.11    0.01    0.04    0.10    0.19    0.40
#> theta[1,4,4]     0.69    0.00 0.14    0.40    0.59    0.69    0.79    0.91
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.11    0.16    0.22    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.39    0.52    0.58    0.65    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.25    0.30    0.36    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,3]     0.62    0.00 0.13    0.36    0.54    0.63    0.72    0.85
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.29    0.47
#> theta[2,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[2,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[3,2,1]     0.09    0.00 0.06    0.02    0.05    0.09    0.13    0.22
#> theta[3,2,2]     0.73    0.00 0.08    0.54    0.67    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.28
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.14    0.24    0.41    0.51    0.60    0.77
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.30    0.49
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[3,4,3]     0.14    0.00 0.12    0.00    0.05    0.11    0.20    0.45
#> theta[3,4,4]     0.63    0.00 0.16    0.29    0.52    0.64    0.75    0.90
#> theta[4,1,1]     0.83    0.00 0.08    0.66    0.78    0.83    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.18
#> theta[4,2,2]     0.72    0.00 0.08    0.55    0.67    0.73    0.78    0.87
#> theta[4,2,3]     0.13    0.00 0.06    0.03    0.09    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,3]     0.69    0.00 0.13    0.41    0.61    0.70    0.79    0.91
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.40
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.76    0.91
#> theta[5,1,1]     0.87    0.00 0.07    0.71    0.83    0.88    0.92    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.06    0.04    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.60    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.20    0.32
#> theta[5,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.40
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.61    0.71    0.84
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.14    0.20    0.39
#> theta[5,4,1]     0.11    0.00 0.10    0.00    0.04    0.08    0.16    0.37
#> theta[5,4,2]     0.12    0.00 0.11    0.00    0.03    0.09    0.16    0.39
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.43
#> theta[5,4,4]     0.63    0.00 0.15    0.31    0.53    0.64    0.75    0.90
#> log_p_z[1,1]    -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_p_z[1,2]   -20.00    0.03 2.31  -25.09  -21.40  -19.78  -18.39  -16.02
#> log_p_z[1,3]   -24.92    0.08 4.34  -34.63  -27.58  -24.48  -21.93  -17.66
#> log_p_z[1,4]   -21.93    0.07 4.26  -31.79  -24.30  -21.54  -18.88  -15.11
#> log_p_z[2,1]   -29.79    0.07 4.48  -40.38  -32.40  -29.29  -26.59  -22.52
#> log_p_z[2,2]   -21.07    0.04 2.42  -26.35  -22.53  -20.93  -19.36  -16.85
#> log_p_z[2,3]    -7.66    0.02 1.25  -10.38   -8.38   -7.57   -6.79   -5.56
#> log_p_z[2,4]   -16.24    0.07 4.07  -25.38  -18.63  -15.72  -13.40   -9.68
#> log_p_z[3,1]   -11.64    0.03 1.65  -15.49  -12.63  -11.45  -10.46   -9.03
#> log_p_z[3,2]   -10.80    0.02 1.52  -14.31  -11.70  -10.59   -9.72   -8.34
#> log_p_z[3,3]   -23.42    0.06 3.50  -31.25  -25.67  -23.06  -20.95  -17.62
#> log_p_z[3,4]   -21.96    0.06 3.55  -29.63  -24.15  -21.69  -19.42  -15.87
#> log_p_z[4,1]   -14.68    0.03 1.84  -18.80  -15.76  -14.53  -13.37  -11.51
#> log_p_z[4,2]    -7.55    0.01 0.91   -9.52   -8.13   -7.50   -6.91   -5.96
#> log_p_z[4,3]   -20.63    0.07 4.15  -29.83  -23.25  -20.11  -17.59  -14.10
#> log_p_z[4,4]   -21.86    0.07 4.38  -31.95  -24.34  -21.35  -18.74  -14.92
#> log_p_z[5,1]   -21.53    0.04 2.59  -27.28  -23.11  -21.38  -19.70  -17.07
#> log_p_z[5,2]    -3.71    0.00 0.44   -4.64   -3.98   -3.68   -3.40   -2.92
#> log_p_z[5,3]   -18.64    0.07 4.06  -27.45  -21.16  -18.19  -15.63  -12.14
#> log_p_z[5,4]   -21.93    0.07 4.41  -32.14  -24.41  -21.47  -18.82  -14.80
#> log_p_z[6,1]   -21.48    0.04 2.51  -27.05  -23.03  -21.33  -19.70  -17.09
#> log_p_z[6,2]    -5.47    0.01 0.66   -6.90   -5.87   -5.42   -5.02   -4.29
#> log_p_z[6,3]   -17.42    0.06 3.88  -26.11  -19.71  -16.99  -14.54  -11.34
#> log_p_z[6,4]   -21.70    0.07 4.42  -31.78  -24.14  -21.12  -18.66  -14.47
#> log_p_z[7,1]    -6.00    0.01 0.63   -7.33   -6.40   -5.95   -5.55   -4.95
#> log_p_z[7,2]   -14.94    0.02 1.71  -18.68  -15.98  -14.79  -13.73  -12.06
#> log_p_z[7,3]   -24.36    0.06 3.56  -32.12  -26.53  -24.04  -21.84  -18.20
#> log_p_z[7,4]   -21.91    0.06 3.41  -29.35  -23.96  -21.69  -19.48  -15.91
#> log_p_z[8,1]   -29.80    0.07 4.48  -39.90  -32.27  -29.25  -26.69  -22.35
#> log_p_z[8,2]   -18.29    0.03 2.17  -23.00  -19.62  -18.12  -16.76  -14.50
#> log_p_z[8,3]    -5.72    0.01 0.80   -7.44   -6.20   -5.66   -5.17   -4.32
#> log_p_z[8,4]   -18.40    0.07 4.21  -28.03  -20.78  -17.84  -15.40  -11.68
#> log_p_z[9,1]   -20.02    0.03 2.31  -25.04  -21.45  -19.84  -18.40  -16.02
#> log_p_z[9,2]    -4.51    0.01 0.57   -5.74   -4.86   -4.47   -4.13   -3.51
#> log_p_z[9,3]   -19.59    0.07 4.12  -28.57  -22.24  -19.08  -16.57  -12.91
#> log_p_z[9,4]   -21.70    0.07 4.31  -31.58  -24.27  -21.15  -18.65  -14.86
#> log_p_z[10,1]  -22.30    0.04 2.42  -27.55  -23.77  -22.11  -20.58  -18.17
#> log_p_z[10,2]   -7.13    0.01 0.71   -8.67   -7.55   -7.07   -6.63   -5.90
#> log_p_z[10,3]  -16.80    0.05 3.06  -23.42  -18.82  -16.47  -14.60  -11.76
#> log_p_z[10,4]  -21.34    0.06 3.54  -29.42  -23.45  -20.96  -18.82  -15.45
#> log_p_z[11,1]  -29.72    0.06 4.34  -39.56  -32.22  -29.24  -26.65  -22.82
#> log_p_z[11,2]  -29.34    0.07 4.29  -39.09  -31.72  -28.84  -26.45  -22.33
#> log_p_z[11,3]  -18.31    0.07 3.50  -26.62  -20.07  -17.72  -15.93  -13.11
#> log_p_z[11,4]   -5.83    0.01 0.95   -7.85   -6.42   -5.76   -5.15   -4.17
#> log_p_z[12,1]  -22.92    0.04 2.61  -28.59  -24.56  -22.69  -21.04  -18.33
#> log_p_z[12,2]   -9.57    0.02 1.22  -12.30  -10.24   -9.44   -8.73   -7.58
#> log_p_z[12,3]  -15.05    0.06 3.85  -23.66  -17.39  -14.61  -12.29   -8.91
#> log_p_z[12,4]  -19.27    0.07 4.18  -28.82  -21.67  -18.70  -16.31  -12.72
#> log_p_z[13,1]   -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_p_z[13,2]  -20.00    0.03 2.31  -25.09  -21.40  -19.78  -18.39  -16.02
#> log_p_z[13,3]  -24.92    0.08 4.34  -34.63  -27.58  -24.48  -21.93  -17.66
#> log_p_z[13,4]  -21.93    0.07 4.26  -31.79  -24.30  -21.54  -18.88  -15.11
#> log_p_z[14,1]  -19.12    0.03 2.42  -24.54  -20.63  -18.95  -17.40  -14.99
#> log_p_z[14,2]   -6.23    0.01 0.82   -8.02   -6.73   -6.15   -5.65   -4.81
#> log_p_z[14,3]  -18.65    0.07 4.06  -27.68  -21.11  -18.20  -15.69  -12.22
#> log_p_z[14,4]  -21.93    0.07 4.38  -32.02  -24.50  -21.36  -18.83  -14.98
#> log_p_z[15,1]   -4.42    0.00 0.40   -5.28   -4.66   -4.39   -4.14   -3.70
#> log_p_z[15,2]  -17.23    0.03 1.85  -21.27  -18.39  -17.10  -15.93  -13.99
#> log_p_z[15,3]  -24.33    0.06 3.52  -31.96  -26.44  -23.99  -21.87  -18.39
#> log_p_z[15,4]  -21.93    0.06 3.54  -29.58  -24.06  -21.67  -19.41  -15.94
#> log_p_z[16,1]   -3.79    0.01 0.54   -4.99   -4.13   -3.75   -3.41   -2.87
#> log_p_z[16,2]  -17.70    0.03 2.21  -22.65  -18.98  -17.48  -16.14  -14.00
#> log_p_z[16,3]  -24.95    0.08 4.39  -34.58  -27.63  -24.43  -21.84  -17.55
#> log_p_z[16,4]  -21.91    0.07 4.16  -31.43  -24.23  -21.48  -18.95  -14.99
#> log_p_z[17,1]   -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_p_z[17,2]  -20.00    0.03 2.31  -25.09  -21.40  -19.78  -18.39  -16.02
#> log_p_z[17,3]  -24.92    0.08 4.34  -34.63  -27.58  -24.48  -21.93  -17.66
#> log_p_z[17,4]  -21.93    0.07 4.26  -31.79  -24.30  -21.54  -18.88  -15.11
#> log_p_z[18,1]   -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_p_z[18,2]  -20.00    0.03 2.31  -25.09  -21.40  -19.78  -18.39  -16.02
#> log_p_z[18,3]  -24.92    0.08 4.34  -34.63  -27.58  -24.48  -21.93  -17.66
#> log_p_z[18,4]  -21.93    0.07 4.26  -31.79  -24.30  -21.54  -18.88  -15.11
#> log_p_z[19,1]  -16.46    0.03 1.97  -20.80  -17.67  -16.34  -15.06  -13.04
#> log_p_z[19,2]   -4.65    0.01 0.58   -5.87   -5.01   -4.62   -4.24   -3.62
#> log_p_z[19,3]  -22.08    0.07 4.23  -31.64  -24.78  -21.60  -18.97  -15.20
#> log_p_z[19,4]  -21.93    0.07 4.32  -31.97  -24.43  -21.46  -18.84  -15.10
#> log_p_z[20,1]  -18.06    0.03 2.24  -22.90  -19.46  -17.90  -16.42  -14.26
#> log_p_z[20,2]   -7.09    0.01 0.97   -9.25   -7.67   -7.01   -6.41   -5.47
#> log_p_z[20,3]  -19.91    0.06 4.00  -28.92  -22.30  -19.40  -17.07  -13.56
#> log_p_z[20,4]  -21.77    0.07 4.43  -32.06  -24.30  -21.33  -18.62  -14.71
#> log_p_z[21,1]  -19.69    0.04 2.41  -25.03  -21.19  -19.54  -17.95  -15.59
#> log_p_z[21,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.79   -2.37
#> log_p_z[21,3]  -21.16    0.07 4.15  -30.38  -23.79  -20.65  -18.15  -14.38
#> log_p_z[21,4]  -21.98    0.07 4.39  -32.11  -24.53  -21.42  -18.89  -14.91
#> log_p_z[22,1]  -16.46    0.03 1.97  -20.80  -17.67  -16.34  -15.06  -13.04
#> log_p_z[22,2]   -4.65    0.01 0.58   -5.87   -5.01   -4.62   -4.24   -3.62
#> log_p_z[22,3]  -22.08    0.07 4.23  -31.64  -24.78  -21.60  -18.97  -15.20
#> log_p_z[22,4]  -21.93    0.07 4.32  -31.97  -24.43  -21.46  -18.84  -15.10
#> log_p_z[23,1]  -21.53    0.04 2.59  -27.28  -23.11  -21.38  -19.70  -17.07
#> log_p_z[23,2]   -3.71    0.00 0.44   -4.64   -3.98   -3.68   -3.40   -2.92
#> log_p_z[23,3]  -18.64    0.07 4.06  -27.45  -21.16  -18.19  -15.63  -12.14
#> log_p_z[23,4]  -21.93    0.07 4.41  -32.14  -24.41  -21.47  -18.82  -14.80
#> log_p_z[24,1]  -17.48    0.03 2.21  -22.40  -18.87  -17.29  -15.90  -13.75
#> log_p_z[24,2]   -5.80    0.01 0.65   -7.34   -6.16   -5.75   -5.35   -4.74
#> log_p_z[24,3]  -21.75    0.06 3.35  -28.98  -23.87  -21.47  -19.31  -16.02
#> log_p_z[24,4]  -21.98    0.06 3.62  -30.02  -24.19  -21.68  -19.37  -15.88
#> log_p_z[25,1]   -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_p_z[25,2]  -20.00    0.03 2.31  -25.09  -21.40  -19.78  -18.39  -16.02
#> log_p_z[25,3]  -24.92    0.08 4.34  -34.63  -27.58  -24.48  -21.93  -17.66
#> log_p_z[25,4]  -21.93    0.07 4.26  -31.79  -24.30  -21.54  -18.88  -15.11
#> log_p_z[26,1]   -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_p_z[26,2]  -20.00    0.03 2.31  -25.09  -21.40  -19.78  -18.39  -16.02
#> log_p_z[26,3]  -24.92    0.08 4.34  -34.63  -27.58  -24.48  -21.93  -17.66
#> log_p_z[26,4]  -21.93    0.07 4.26  -31.79  -24.30  -21.54  -18.88  -15.11
#> log_p_z[27,1]  -21.97    0.04 2.53  -27.35  -23.60  -21.73  -20.17  -17.65
#> log_p_z[27,2]   -5.65    0.01 0.53   -6.80   -5.97   -5.61   -5.28   -4.75
#> log_p_z[27,3]  -18.37    0.05 3.12  -25.13  -20.31  -18.07  -16.09  -13.15
#> log_p_z[27,4]  -21.61    0.06 3.56  -29.87  -23.71  -21.27  -19.08  -15.65
#> log_p_z[28,1]   -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_p_z[28,2]  -20.00    0.03 2.31  -25.09  -21.40  -19.78  -18.39  -16.02
#> log_p_z[28,3]  -24.92    0.08 4.34  -34.63  -27.58  -24.48  -21.93  -17.66
#> log_p_z[28,4]  -21.93    0.07 4.26  -31.79  -24.30  -21.54  -18.88  -15.11
#> log_p_z[29,1]   -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_p_z[29,2]  -20.00    0.03 2.31  -25.09  -21.40  -19.78  -18.39  -16.02
#> log_p_z[29,3]  -24.92    0.08 4.34  -34.63  -27.58  -24.48  -21.93  -17.66
#> log_p_z[29,4]  -21.93    0.07 4.26  -31.79  -24.30  -21.54  -18.88  -15.11
#> log_p_z[30,1]   -6.83    0.01 0.84   -8.68   -7.30   -6.75   -6.24   -5.47
#> log_p_z[30,2]  -14.71    0.03 1.69  -18.47  -15.77  -14.54  -13.50  -11.86
#> log_p_z[30,3]  -24.31    0.06 3.53  -32.10  -26.45  -23.99  -21.82  -18.40
#> log_p_z[30,4]  -21.93    0.06 3.61  -29.85  -24.06  -21.68  -19.42  -15.77
#> log_p_z[31,1]   -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_p_z[31,2]  -20.00    0.03 2.31  -25.09  -21.40  -19.78  -18.39  -16.02
#> log_p_z[31,3]  -24.92    0.08 4.34  -34.63  -27.58  -24.48  -21.93  -17.66
#> log_p_z[31,4]  -21.93    0.07 4.26  -31.79  -24.30  -21.54  -18.88  -15.11
#> log_p_z[32,1]  -29.83    0.07 4.48  -39.98  -32.43  -29.35  -26.69  -22.45
#> log_p_z[32,2]  -14.82    0.02 1.76  -18.72  -15.92  -14.67  -13.62  -11.75
#> log_p_z[32,3]   -6.06    0.01 0.93   -8.08   -6.63   -6.00   -5.42   -4.43
#> log_p_z[32,4]  -20.55    0.08 4.37  -30.65  -23.12  -20.04  -17.47  -13.51
#> log_p_z[33,1]   -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_p_z[33,2]  -20.00    0.03 2.31  -25.09  -21.40  -19.78  -18.39  -16.02
#> log_p_z[33,3]  -24.92    0.08 4.34  -34.63  -27.58  -24.48  -21.93  -17.66
#> log_p_z[33,4]  -21.93    0.07 4.26  -31.79  -24.30  -21.54  -18.88  -15.11
#> log_p_z[34,1]  -19.69    0.04 2.41  -25.03  -21.19  -19.54  -17.95  -15.59
#> log_p_z[34,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.79   -2.37
#> log_p_z[34,3]  -21.16    0.07 4.15  -30.38  -23.79  -20.65  -18.15  -14.38
#> log_p_z[34,4]  -21.98    0.07 4.39  -32.11  -24.53  -21.42  -18.89  -14.91
#> log_p_z[35,1]  -22.65    0.04 2.71  -28.59  -24.40  -22.43  -20.77  -17.96
#> log_p_z[35,2]   -5.51    0.01 0.70   -7.07   -5.93   -5.46   -5.02   -4.30
#> log_p_z[35,3]  -15.99    0.06 3.92  -24.75  -18.28  -15.54  -13.02   -9.94
#> log_p_z[35,4]  -21.92    0.08 4.38  -32.28  -24.41  -21.38  -18.87  -14.94
#> log_p_z[36,1]  -29.75    0.06 3.60  -37.84  -31.84  -29.36  -27.29  -23.65
#> log_p_z[36,2]  -21.92    0.04 2.34  -27.05  -23.31  -21.76  -20.30  -17.93
#> log_p_z[36,3]  -10.55    0.04 1.76  -15.00  -11.45  -10.28   -9.32   -7.93
#> log_p_z[36,4]  -13.93    0.06 3.23  -21.27  -15.91  -13.64  -11.60   -8.67
#> log_p_z[37,1]  -18.60    0.03 2.35  -23.76  -20.02  -18.36  -16.94  -14.61
#> log_p_z[37,2]   -7.61    0.01 0.83   -9.43   -8.09   -7.54   -7.03   -6.20
#> log_p_z[37,3]  -19.11    0.05 3.15  -25.94  -21.08  -18.82  -16.83  -13.78
#> log_p_z[37,4]  -21.97    0.06 3.58  -29.97  -24.08  -21.65  -19.41  -15.97
#> log_p_z[38,1]  -25.26    0.04 2.83  -31.40  -27.06  -25.01  -23.25  -20.33
#> log_p_z[38,2]   -9.60    0.01 0.98  -11.75  -10.20   -9.53   -8.90   -7.92
#> log_p_z[38,3]  -11.64    0.04 2.66  -17.63  -13.27  -11.36   -9.57   -7.43
#> log_p_z[38,4]  -21.28    0.06 3.53  -29.35  -23.41  -20.93  -18.79  -15.31
#> log_p_z[39,1]  -29.47    0.07 4.48  -39.81  -31.92  -28.94  -26.36  -22.28
#> log_p_z[39,2]  -16.82    0.03 2.08  -21.40  -18.06  -16.64  -15.36  -13.16
#> log_p_z[39,3]   -7.28    0.02 1.10   -9.67   -7.98   -7.21   -6.52   -5.33
#> log_p_z[39,4]  -18.67    0.07 4.08  -28.08  -21.00  -18.14  -15.72  -12.18
#> log_p_z[40,1]   -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_p_z[40,2]  -20.00    0.03 2.31  -25.09  -21.40  -19.78  -18.39  -16.02
#> log_p_z[40,3]  -24.92    0.08 4.34  -34.63  -27.58  -24.48  -21.93  -17.66
#> log_p_z[40,4]  -21.93    0.07 4.26  -31.79  -24.30  -21.54  -18.88  -15.11
#> log_p_z[41,1]   -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_p_z[41,2]  -20.00    0.03 2.31  -25.09  -21.40  -19.78  -18.39  -16.02
#> log_p_z[41,3]  -24.92    0.08 4.34  -34.63  -27.58  -24.48  -21.93  -17.66
#> log_p_z[41,4]  -21.93    0.07 4.26  -31.79  -24.30  -21.54  -18.88  -15.11
#> log_p_z[42,1]   -6.00    0.01 0.63   -7.33   -6.40   -5.95   -5.55   -4.95
#> log_p_z[42,2]  -14.94    0.02 1.71  -18.68  -15.98  -14.79  -13.73  -12.06
#> log_p_z[42,3]  -24.36    0.06 3.56  -32.12  -26.53  -24.04  -21.84  -18.20
#> log_p_z[42,4]  -21.91    0.06 3.41  -29.35  -23.96  -21.69  -19.48  -15.91
#> log_p_z[43,1]  -21.97    0.04 2.53  -27.35  -23.60  -21.73  -20.17  -17.65
#> log_p_z[43,2]   -5.65    0.01 0.53   -6.80   -5.97   -5.61   -5.28   -4.75
#> log_p_z[43,3]  -18.37    0.05 3.12  -25.13  -20.31  -18.07  -16.09  -13.15
#> log_p_z[43,4]  -21.61    0.06 3.56  -29.87  -23.71  -21.27  -19.08  -15.65
#> log_p_z[44,1]   -4.42    0.00 0.40   -5.28   -4.66   -4.39   -4.14   -3.70
#> log_p_z[44,2]  -17.23    0.03 1.85  -21.27  -18.39  -17.10  -15.93  -13.99
#> log_p_z[44,3]  -24.33    0.06 3.52  -31.96  -26.44  -23.99  -21.87  -18.39
#> log_p_z[44,4]  -21.93    0.06 3.54  -29.58  -24.06  -21.67  -19.41  -15.94
#> log_p_z[45,1]  -19.69    0.04 2.41  -25.03  -21.19  -19.54  -17.95  -15.59
#> log_p_z[45,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.79   -2.37
#> log_p_z[45,3]  -21.16    0.07 4.15  -30.38  -23.79  -20.65  -18.15  -14.38
#> log_p_z[45,4]  -21.98    0.07 4.39  -32.11  -24.53  -21.42  -18.89  -14.91
#> log_lik[1]      -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_lik[2]      -7.60    0.02 1.14  -10.08   -8.33   -7.55   -6.78   -5.56
#> log_lik[3]      -9.94    0.01 0.92  -11.87  -10.52   -9.90   -9.30   -8.23
#> log_lik[4]      -7.54    0.01 0.91   -9.48   -8.12   -7.50   -6.91   -5.96
#> log_lik[5]      -3.71    0.00 0.44   -4.64   -3.98   -3.68   -3.40   -2.92
#> log_lik[6]      -5.47    0.01 0.66   -6.90   -5.87   -5.42   -5.02   -4.29
#> log_lik[7]      -6.00    0.01 0.63   -7.33   -6.40   -5.95   -5.55   -4.95
#> log_lik[8]      -5.72    0.01 0.80   -7.44   -6.20   -5.66   -5.17   -4.32
#> log_lik[9]      -4.51    0.01 0.57   -5.74   -4.86   -4.47   -4.13   -3.51
#> log_lik[10]     -7.12    0.01 0.70   -8.67   -7.55   -7.07   -6.63   -5.90
#> log_lik[11]     -5.83    0.01 0.95   -7.85   -6.42   -5.76   -5.15   -4.17
#> log_lik[12]     -9.33    0.01 1.01  -11.43   -9.98   -9.28   -8.63   -7.51
#> log_lik[13]     -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_lik[14]     -6.23    0.01 0.82   -8.02   -6.73   -6.15   -5.65   -4.81
#> log_lik[15]     -4.42    0.00 0.40   -5.28   -4.66   -4.39   -4.14   -3.70
#> log_lik[16]     -3.79    0.01 0.54   -4.99   -4.13   -3.75   -3.41   -2.87
#> log_lik[17]     -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_lik[18]     -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_lik[19]     -4.65    0.01 0.58   -5.87   -5.01   -4.62   -4.24   -3.62
#> log_lik[20]     -7.09    0.01 0.97   -9.25   -7.67   -7.01   -6.41   -5.47
#> log_lik[21]     -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.79   -2.37
#> log_lik[22]     -4.65    0.01 0.58   -5.87   -5.01   -4.62   -4.24   -3.62
#> log_lik[23]     -3.71    0.00 0.44   -4.64   -3.98   -3.68   -3.40   -2.92
#> log_lik[24]     -5.80    0.01 0.65   -7.34   -6.16   -5.74   -5.35   -4.74
#> log_lik[25]     -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_lik[26]     -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_lik[27]     -5.65    0.01 0.53   -6.80   -5.97   -5.61   -5.28   -4.75
#> log_lik[28]     -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_lik[29]     -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_lik[30]     -6.82    0.01 0.83   -8.67   -7.30   -6.75   -6.24   -5.47
#> log_lik[31]     -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_lik[32]     -6.06    0.01 0.93   -8.08   -6.63   -5.99   -5.42   -4.43
#> log_lik[33]     -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_lik[34]     -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.79   -2.37
#> log_lik[35]     -5.51    0.01 0.70   -7.07   -5.92   -5.46   -5.02   -4.30
#> log_lik[36]     -9.92    0.02 1.19  -12.48  -10.66   -9.84   -9.07   -7.81
#> log_lik[37]     -7.61    0.01 0.83   -9.42   -8.09   -7.54   -7.03   -6.20
#> log_lik[38]     -8.99    0.01 0.86  -10.67   -9.56   -8.98   -8.42   -7.27
#> log_lik[39]     -7.28    0.02 1.10   -9.63   -7.97   -7.21   -6.52   -5.33
#> log_lik[40]     -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_lik[41]     -2.20    0.00 0.29   -2.81   -2.40   -2.19   -2.00   -1.68
#> log_lik[42]     -6.00    0.01 0.63   -7.33   -6.40   -5.95   -5.55   -4.95
#> log_lik[43]     -5.65    0.01 0.53   -6.80   -5.97   -5.61   -5.28   -4.75
#> log_lik[44]     -4.42    0.00 0.40   -5.28   -4.66   -4.39   -4.14   -3.70
#> log_lik[45]     -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.79   -2.37
#> lp__          -463.84    0.16 6.21 -477.16 -467.73 -463.41 -459.39 -453.05
#>               n_eff Rhat
#> pi[1]         10048    1
#> pi[2]          8701    1
#> pi[3]          6780    1
#> pi[4]          5853    1
#> theta[1,1,1]   9318    1
#> theta[1,1,2]   7701    1
#> theta[1,1,3]   4829    1
#> theta[1,1,4]   5376    1
#> theta[1,2,1]   6136    1
#> theta[1,2,2]   7312    1
#> theta[1,2,3]   7408    1
#> theta[1,2,4]   5641    1
#> theta[1,3,1]   6003    1
#> theta[1,3,2]   4020    1
#> theta[1,3,3]   5352    1
#> theta[1,3,4]   6053    1
#> theta[1,4,1]   6512    1
#> theta[1,4,2]   6294    1
#> theta[1,4,3]   3845    1
#> theta[1,4,4]   4924    1
#> theta[2,1,1]   9513    1
#> theta[2,1,2]   7986    1
#> theta[2,1,3]   4780    1
#> theta[2,1,4]   4387    1
#> theta[2,2,1]   6173    1
#> theta[2,2,2]   7598    1
#> theta[2,2,3]   7129    1
#> theta[2,2,4]   5567    1
#> theta[2,3,1]   6291    1
#> theta[2,3,2]   5648    1
#> theta[2,3,3]   6446    1
#> theta[2,3,4]   6014    1
#> theta[2,4,1]   6668    1
#> theta[2,4,2]   7213    1
#> theta[2,4,3]   6055    1
#> theta[2,4,4]   6652    1
#> theta[3,1,1]   8031    1
#> theta[3,1,2]   5384    1
#> theta[3,1,3]   5040    1
#> theta[3,1,4]   4679    1
#> theta[3,2,1]   7133    1
#> theta[3,2,2]   7329    1
#> theta[3,2,3]   6731    1
#> theta[3,2,4]   5526    1
#> theta[3,3,1]   6136    1
#> theta[3,3,2]   6688    1
#> theta[3,3,3]   6771    1
#> theta[3,3,4]   6424    1
#> theta[3,4,1]   6271    1
#> theta[3,4,2]   5564    1
#> theta[3,4,3]   5064    1
#> theta[3,4,4]   5126    1
#> theta[4,1,1]   8192    1
#> theta[4,1,2]   5938    1
#> theta[4,1,3]   4900    1
#> theta[4,1,4]   4667    1
#> theta[4,2,1]   6521    1
#> theta[4,2,2]   7260    1
#> theta[4,2,3]   5872    1
#> theta[4,2,4]   5310    1
#> theta[4,3,1]   6688    1
#> theta[4,3,2]   6319    1
#> theta[4,3,3]   5370    1
#> theta[4,3,4]   4701    1
#> theta[4,4,1]   6599    1
#> theta[4,4,2]   6835    1
#> theta[4,4,3]   6839    1
#> theta[4,4,4]   5639    1
#> theta[5,1,1]   6890    1
#> theta[5,1,2]   5262    1
#> theta[5,1,3]   4858    1
#> theta[5,1,4]   4840    1
#> theta[5,2,1]   7615    1
#> theta[5,2,2]   8293    1
#> theta[5,2,3]   6088    1
#> theta[5,2,4]   5543    1
#> theta[5,3,1]   6730    1
#> theta[5,3,2]   7432    1
#> theta[5,3,3]   5567    1
#> theta[5,3,4]   6924    1
#> theta[5,4,1]   6166    1
#> theta[5,4,2]   5902    1
#> theta[5,4,3]   5060    1
#> theta[5,4,4]   6549    1
#> log_p_z[1,1]   8922    1
#> log_p_z[1,2]   4534    1
#> log_p_z[1,3]   3029    1
#> log_p_z[1,4]   3718    1
#> log_p_z[2,1]   4557    1
#> log_p_z[2,2]   4337    1
#> log_p_z[2,3]   3979    1
#> log_p_z[2,4]   3193    1
#> log_p_z[3,1]   3879    1
#> log_p_z[3,2]   4734    1
#> log_p_z[3,3]   3138    1
#> log_p_z[3,4]   3602    1
#> log_p_z[4,1]   5394    1
#> log_p_z[4,2]   5537    1
#> log_p_z[4,3]   3570    1
#> log_p_z[4,4]   3738    1
#> log_p_z[5,1]   4695    1
#> log_p_z[5,2]   7842    1
#> log_p_z[5,3]   3818    1
#> log_p_z[5,4]   3596    1
#> log_p_z[6,1]   5067    1
#> log_p_z[6,2]   6010    1
#> log_p_z[6,3]   3830    1
#> log_p_z[6,4]   3514    1
#> log_p_z[7,1]   6768    1
#> log_p_z[7,2]   4764    1
#> log_p_z[7,3]   3120    1
#> log_p_z[7,4]   3601    1
#> log_p_z[8,1]   4697    1
#> log_p_z[8,2]   5012    1
#> log_p_z[8,3]   5696    1
#> log_p_z[8,4]   3189    1
#> log_p_z[9,1]   4985    1
#> log_p_z[9,2]   6066    1
#> log_p_z[9,3]   3798    1
#> log_p_z[9,4]   3623    1
#> log_p_z[10,1]  4293    1
#> log_p_z[10,2]  5746    1
#> log_p_z[10,3]  3530    1
#> log_p_z[10,4]  3604    1
#> log_p_z[11,1]  4571    1
#> log_p_z[11,2]  4147    1
#> log_p_z[11,3]  2705    1
#> log_p_z[11,4]  5525    1
#> log_p_z[12,1]  5464    1
#> log_p_z[12,2]  4203    1
#> log_p_z[12,3]  4079    1
#> log_p_z[12,4]  3379    1
#> log_p_z[13,1]  8922    1
#> log_p_z[13,2]  4534    1
#> log_p_z[13,3]  3029    1
#> log_p_z[13,4]  3718    1
#> log_p_z[14,1]  4805    1
#> log_p_z[14,2]  4319    1
#> log_p_z[14,3]  3847    1
#> log_p_z[14,4]  3464    1
#> log_p_z[15,1]  6675    1
#> log_p_z[15,2]  4523    1
#> log_p_z[15,3]  2966    1
#> log_p_z[15,4]  3802    1
#> log_p_z[16,1]  7213    1
#> log_p_z[16,2]  4709    1
#> log_p_z[16,3]  3215    1
#> log_p_z[16,4]  3561    1
#> log_p_z[17,1]  8922    1
#> log_p_z[17,2]  4534    1
#> log_p_z[17,3]  3029    1
#> log_p_z[17,4]  3718    1
#> log_p_z[18,1]  8922    1
#> log_p_z[18,2]  4534    1
#> log_p_z[18,3]  3029    1
#> log_p_z[18,4]  3718    1
#> log_p_z[19,1]  4738    1
#> log_p_z[19,2]  6547    1
#> log_p_z[19,3]  3724    1
#> log_p_z[19,4]  3537    1
#> log_p_z[20,1]  4853    1
#> log_p_z[20,2]  5157    1
#> log_p_z[20,3]  3829    1
#> log_p_z[20,4]  3583    1
#> log_p_z[21,1]  4495    1
#> log_p_z[21,2]  8406    1
#> log_p_z[21,3]  3797    1
#> log_p_z[21,4]  3592    1
#> log_p_z[22,1]  4738    1
#> log_p_z[22,2]  6547    1
#> log_p_z[22,3]  3724    1
#> log_p_z[22,4]  3537    1
#> log_p_z[23,1]  4695    1
#> log_p_z[23,2]  7842    1
#> log_p_z[23,3]  3818    1
#> log_p_z[23,4]  3596    1
#> log_p_z[24,1]  4244    1
#> log_p_z[24,2]  5458    1
#> log_p_z[24,3]  3660    1
#> log_p_z[24,4]  3626    1
#> log_p_z[25,1]  8922    1
#> log_p_z[25,2]  4534    1
#> log_p_z[25,3]  3029    1
#> log_p_z[25,4]  3718    1
#> log_p_z[26,1]  8922    1
#> log_p_z[26,2]  4534    1
#> log_p_z[26,3]  3029    1
#> log_p_z[26,4]  3718    1
#> log_p_z[27,1]  3744    1
#> log_p_z[27,2]  5749    1
#> log_p_z[27,3]  3529    1
#> log_p_z[27,4]  3604    1
#> log_p_z[28,1]  8922    1
#> log_p_z[28,2]  4534    1
#> log_p_z[28,3]  3029    1
#> log_p_z[28,4]  3718    1
#> log_p_z[29,1]  8922    1
#> log_p_z[29,2]  4534    1
#> log_p_z[29,3]  3029    1
#> log_p_z[29,4]  3718    1
#> log_p_z[30,1]  5159    1
#> log_p_z[30,2]  4502    1
#> log_p_z[30,3]  3051    1
#> log_p_z[30,4]  3679    1
#> log_p_z[31,1]  8922    1
#> log_p_z[31,2]  4534    1
#> log_p_z[31,3]  3029    1
#> log_p_z[31,4]  3718    1
#> log_p_z[32,1]  4550    1
#> log_p_z[32,2]  5219    1
#> log_p_z[32,3]  5524    1
#> log_p_z[32,4]  3156    1
#> log_p_z[33,1]  8922    1
#> log_p_z[33,2]  4534    1
#> log_p_z[33,3]  3029    1
#> log_p_z[33,4]  3718    1
#> log_p_z[34,1]  4495    1
#> log_p_z[34,2]  8406    1
#> log_p_z[34,3]  3797    1
#> log_p_z[34,4]  3592    1
#> log_p_z[35,1]  4979    1
#> log_p_z[35,2]  4803    1
#> log_p_z[35,3]  4342    1
#> log_p_z[35,4]  3385    1
#> log_p_z[36,1]  3857    1
#> log_p_z[36,2]  3853    1
#> log_p_z[36,3]  2321    1
#> log_p_z[36,4]  2902    1
#> log_p_z[37,1]  4789    1
#> log_p_z[37,2]  5065    1
#> log_p_z[37,3]  3875    1
#> log_p_z[37,4]  3359    1
#> log_p_z[38,1]  4739    1
#> log_p_z[38,2]  4461    1
#> log_p_z[38,3]  4275    1
#> log_p_z[38,4]  3383    1
#> log_p_z[39,1]  4619    1
#> log_p_z[39,2]  5035    1
#> log_p_z[39,3]  5310    1
#> log_p_z[39,4]  3216    1
#> log_p_z[40,1]  8922    1
#> log_p_z[40,2]  4534    1
#> log_p_z[40,3]  3029    1
#> log_p_z[40,4]  3718    1
#> log_p_z[41,1]  8922    1
#> log_p_z[41,2]  4534    1
#> log_p_z[41,3]  3029    1
#> log_p_z[41,4]  3718    1
#> log_p_z[42,1]  6768    1
#> log_p_z[42,2]  4764    1
#> log_p_z[42,3]  3120    1
#> log_p_z[42,4]  3601    1
#> log_p_z[43,1]  3744    1
#> log_p_z[43,2]  5749    1
#> log_p_z[43,3]  3529    1
#> log_p_z[43,4]  3604    1
#> log_p_z[44,1]  6675    1
#> log_p_z[44,2]  4523    1
#> log_p_z[44,3]  2966    1
#> log_p_z[44,4]  3802    1
#> log_p_z[45,1]  4495    1
#> log_p_z[45,2]  8406    1
#> log_p_z[45,3]  3797    1
#> log_p_z[45,4]  3592    1
#> log_lik[1]     8922    1
#> log_lik[2]     4546    1
#> log_lik[3]     5275    1
#> log_lik[4]     5614    1
#> log_lik[5]     7842    1
#> log_lik[6]     6013    1
#> log_lik[7]     6766    1
#> log_lik[8]     5696    1
#> log_lik[9]     6067    1
#> log_lik[10]    5749    1
#> log_lik[11]    5527    1
#> log_lik[12]    4716    1
#> log_lik[13]    8922    1
#> log_lik[14]    4324    1
#> log_lik[15]    6675    1
#> log_lik[16]    7213    1
#> log_lik[17]    8922    1
#> log_lik[18]    8922    1
#> log_lik[19]    6548    1
#> log_lik[20]    5161    1
#> log_lik[21]    8406    1
#> log_lik[22]    6548    1
#> log_lik[23]    7842    1
#> log_lik[24]    5458    1
#> log_lik[25]    8922    1
#> log_lik[26]    8922    1
#> log_lik[27]    5748    1
#> log_lik[28]    8922    1
#> log_lik[29]    8922    1
#> log_lik[30]    5167    1
#> log_lik[31]    8922    1
#> log_lik[32]    5525    1
#> log_lik[33]    8922    1
#> log_lik[34]    8406    1
#> log_lik[35]    4812    1
#> log_lik[36]    4641    1
#> log_lik[37]    5062    1
#> log_lik[38]    4487    1
#> log_lik[39]    5338    1
#> log_lik[40]    8922    1
#> log_lik[41]    8922    1
#> log_lik[42]    6766    1
#> log_lik[43]    5748    1
#> log_lik[44]    6675    1
#> log_lik[45]    8406    1
#> lp__           1520    1
#> 
#> Samples were drawn using NUTS(diag_e) at Sun Mar  5 00:59:49 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }