Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.26    0.33    0.38    0.42    0.51
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.09    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.08    0.10    0.12    0.18
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.83    0.86    0.89    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.02    0.03    0.06    0.15
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.59    0.74    0.81    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.41
#> theta[1,4,4]     0.69    0.00 0.14    0.40    0.60    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.75    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.22    0.33
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.39    0.52    0.58    0.65    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.24    0.30    0.37    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[2,3,3]     0.62    0.00 0.13    0.36    0.53    0.62    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.28    0.48
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.36
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.76    0.90
#> theta[3,1,1]     0.88    0.00 0.06    0.74    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.09    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.74    0.79    0.88
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.13    0.24    0.41    0.51    0.60    0.76
#> theta[3,3,4]     0.23    0.00 0.11    0.05    0.15    0.22    0.30    0.49
#> theta[3,4,1]     0.12    0.00 0.10    0.01    0.04    0.09    0.16    0.36
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[3,4,3]     0.14    0.00 0.12    0.00    0.05    0.11    0.20    0.44
#> theta[3,4,4]     0.63    0.00 0.15    0.31    0.52    0.64    0.74    0.89
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.78    0.83    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.08    0.54    0.67    0.73    0.78    0.87
#> theta[4,2,3]     0.13    0.00 0.06    0.03    0.09    0.13    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.18
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,3]     0.69    0.00 0.12    0.43    0.61    0.70    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.38
#> theta[4,4,1]     0.12    0.00 0.10    0.01    0.04    0.09    0.16    0.39
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.76    0.90
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.82    0.87    0.91    0.96
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,3,3]     0.61    0.00 0.13    0.33    0.51    0.61    0.70    0.84
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.14    0.21    0.40
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[5,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[5,4,3]     0.14    0.00 0.11    0.00    0.05    0.11    0.20    0.43
#> theta[5,4,4]     0.63    0.00 0.15    0.31    0.53    0.64    0.74    0.89
#> log_p_z[1,1]    -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_p_z[1,2]   -20.05    0.04 2.38  -25.16  -21.59  -19.84  -18.39  -15.84
#> log_p_z[1,3]   -24.77    0.07 4.20  -33.99  -27.23  -24.37  -21.80  -17.79
#> log_p_z[1,4]   -21.87    0.08 4.39  -31.71  -24.41  -21.43  -18.71  -14.65
#> log_p_z[2,1]   -29.65    0.06 4.45  -40.07  -32.29  -29.13  -26.59  -22.34
#> log_p_z[2,2]   -21.04    0.04 2.43  -26.31  -22.63  -20.85  -19.25  -16.87
#> log_p_z[2,3]    -7.65    0.02 1.23  -10.36   -8.38   -7.55   -6.78   -5.58
#> log_p_z[2,4]   -16.36    0.08 4.16  -25.51  -18.84  -15.86  -13.41   -9.65
#> log_p_z[3,1]   -11.59    0.03 1.61  -15.20  -12.55  -11.40  -10.44   -8.99
#> log_p_z[3,2]   -10.84    0.02 1.52  -14.31  -11.76  -10.69   -9.77   -8.34
#> log_p_z[3,3]   -23.30    0.06 3.41  -30.74  -25.40  -22.93  -20.92  -17.54
#> log_p_z[3,4]   -21.89    0.07 3.56  -29.63  -24.06  -21.55  -19.46  -15.84
#> log_p_z[4,1]   -14.61    0.03 1.85  -18.73  -15.75  -14.45  -13.31  -11.37
#> log_p_z[4,2]    -7.57    0.01 0.95   -9.55   -8.17   -7.50   -6.91   -5.89
#> log_p_z[4,3]   -20.71    0.07 4.29  -30.90  -23.26  -20.09  -17.64  -14.01
#> log_p_z[4,4]   -21.76    0.07 4.20  -31.12  -24.17  -21.28  -18.81  -15.00
#> log_p_z[5,1]   -21.42    0.04 2.51  -26.81  -23.01  -21.27  -19.67  -16.98
#> log_p_z[5,2]    -3.72    0.01 0.45   -4.68   -4.00   -3.69   -3.40   -2.91
#> log_p_z[5,3]   -18.71    0.07 4.23  -28.37  -21.17  -18.14  -15.63  -12.30
#> log_p_z[5,4]   -21.82    0.07 4.23  -31.36  -24.24  -21.41  -18.85  -14.79
#> log_p_z[6,1]   -21.39    0.04 2.53  -26.80  -23.00  -21.21  -19.60  -16.91
#> log_p_z[6,2]    -5.49    0.01 0.68   -6.98   -5.90   -5.45   -5.01   -4.29
#> log_p_z[6,3]   -17.51    0.07 4.09  -26.97  -19.84  -16.94  -14.56  -11.34
#> log_p_z[6,4]   -21.60    0.07 4.21  -31.10  -24.07  -21.25  -18.60  -14.80
#> log_p_z[7,1]    -5.99    0.01 0.60   -7.27   -6.37   -5.96   -5.58   -4.93
#> log_p_z[7,2]   -14.96    0.03 1.76  -18.74  -16.02  -14.82  -13.75  -11.85
#> log_p_z[7,3]   -24.24    0.06 3.46  -31.78  -26.36  -23.94  -21.82  -18.32
#> log_p_z[7,4]   -21.84    0.07 3.58  -29.76  -24.03  -21.50  -19.31  -15.88
#> log_p_z[8,1]   -29.69    0.07 4.45  -39.94  -32.21  -29.21  -26.63  -22.36
#> log_p_z[8,2]   -18.28    0.03 2.09  -22.90  -19.47  -18.19  -16.84  -14.62
#> log_p_z[8,3]    -5.73    0.01 0.82   -7.54   -6.24   -5.67   -5.15   -4.33
#> log_p_z[8,4]   -18.56    0.08 4.28  -28.33  -21.01  -18.02  -15.53  -11.78
#> log_p_z[9,1]   -19.99    0.03 2.31  -24.98  -21.43  -19.82  -18.38  -16.04
#> log_p_z[9,2]    -4.51    0.01 0.57   -5.77   -4.85   -4.46   -4.12   -3.53
#> log_p_z[9,3]   -19.65    0.07 4.34  -29.48  -22.12  -19.18  -16.50  -12.74
#> log_p_z[9,4]   -21.60    0.07 4.10  -30.52  -23.97  -21.14  -18.70  -14.90
#> log_p_z[10,1]  -22.25    0.04 2.44  -27.51  -23.77  -22.06  -20.54  -18.05
#> log_p_z[10,2]   -7.11    0.01 0.71   -8.68   -7.54   -7.06   -6.62   -5.87
#> log_p_z[10,3]  -16.82    0.05 3.19  -23.86  -18.79  -16.52  -14.51  -11.56
#> log_p_z[10,4]  -21.30    0.06 3.39  -28.72  -23.30  -20.98  -18.87  -15.58
#> log_p_z[11,1]  -29.66    0.06 4.38  -39.46  -32.21  -29.26  -26.49  -22.46
#> log_p_z[11,2]  -29.43    0.07 4.31  -39.11  -32.12  -29.02  -26.39  -22.26
#> log_p_z[11,3]  -18.25    0.06 3.36  -26.45  -20.00  -17.82  -15.98  -13.05
#> log_p_z[11,4]   -5.82    0.01 1.00   -7.97   -6.44   -5.75   -5.11   -4.13
#> log_p_z[12,1]  -22.90    0.04 2.71  -28.80  -24.59  -22.69  -20.98  -18.16
#> log_p_z[12,2]   -9.59    0.02 1.26  -12.45  -10.31   -9.46   -8.72   -7.53
#> log_p_z[12,3]  -15.18    0.07 4.03  -24.41  -17.56  -14.65  -12.32   -9.00
#> log_p_z[12,4]  -19.22    0.06 4.02  -28.59  -21.47  -18.81  -16.40  -12.51
#> log_p_z[13,1]   -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_p_z[13,2]  -20.05    0.04 2.38  -25.16  -21.59  -19.84  -18.39  -15.84
#> log_p_z[13,3]  -24.77    0.07 4.20  -33.99  -27.23  -24.37  -21.80  -17.79
#> log_p_z[13,4]  -21.87    0.08 4.39  -31.71  -24.41  -21.43  -18.71  -14.65
#> log_p_z[14,1]  -19.03    0.03 2.33  -24.11  -20.51  -18.85  -17.41  -14.92
#> log_p_z[14,2]   -6.23    0.01 0.87   -8.12   -6.76   -6.14   -5.61   -4.81
#> log_p_z[14,3]  -18.73    0.07 4.26  -28.68  -21.06  -18.24  -15.72  -12.13
#> log_p_z[14,4]  -21.80    0.07 4.23  -31.11  -24.27  -21.29  -18.78  -14.94
#> log_p_z[15,1]   -4.41    0.01 0.39   -5.24   -4.66   -4.38   -4.13   -3.71
#> log_p_z[15,2]  -17.27    0.03 1.91  -21.34  -18.50  -17.12  -15.97  -13.83
#> log_p_z[15,3]  -24.26    0.06 3.46  -31.55  -26.41  -23.96  -21.80  -18.34
#> log_p_z[15,4]  -21.82    0.06 3.59  -29.70  -23.97  -21.54  -19.29  -15.74
#> log_p_z[16,1]   -3.79    0.01 0.54   -4.95   -4.13   -3.75   -3.41   -2.84
#> log_p_z[16,2]  -17.73    0.03 2.26  -22.56  -19.11  -17.52  -16.19  -13.75
#> log_p_z[16,3]  -24.75    0.07 4.21  -33.88  -27.21  -24.22  -21.82  -17.78
#> log_p_z[16,4]  -21.89    0.08 4.38  -31.67  -24.47  -21.45  -18.72  -14.72
#> log_p_z[17,1]   -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_p_z[17,2]  -20.05    0.04 2.38  -25.16  -21.59  -19.84  -18.39  -15.84
#> log_p_z[17,3]  -24.77    0.07 4.20  -33.99  -27.23  -24.37  -21.80  -17.79
#> log_p_z[17,4]  -21.87    0.08 4.39  -31.71  -24.41  -21.43  -18.71  -14.65
#> log_p_z[18,1]   -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_p_z[18,2]  -20.05    0.04 2.38  -25.16  -21.59  -19.84  -18.39  -15.84
#> log_p_z[18,3]  -24.77    0.07 4.20  -33.99  -27.23  -24.37  -21.80  -17.79
#> log_p_z[18,4]  -21.87    0.08 4.39  -31.71  -24.41  -21.43  -18.71  -14.65
#> log_p_z[19,1]  -16.40    0.03 1.94  -20.58  -17.63  -16.27  -15.04  -13.05
#> log_p_z[19,2]   -4.65    0.01 0.59   -5.92   -5.03   -4.61   -4.23   -3.60
#> log_p_z[19,3]  -22.13    0.07 4.41  -32.01  -24.77  -21.60  -18.94  -15.13
#> log_p_z[19,4]  -21.78    0.07 4.15  -31.06  -24.17  -21.29  -18.88  -15.11
#> log_p_z[20,1]  -18.00    0.03 2.28  -22.95  -19.41  -17.86  -16.39  -14.04
#> log_p_z[20,2]   -7.13    0.01 0.98   -9.24   -7.72   -7.04   -6.43   -5.47
#> log_p_z[20,3]  -20.02    0.07 4.25  -29.74  -22.49  -19.44  -16.87  -13.53
#> log_p_z[20,4]  -21.57    0.07 4.20  -31.23  -24.00  -21.14  -18.56  -14.75
#> log_p_z[21,1]  -19.61    0.03 2.35  -24.67  -21.10  -19.44  -17.93  -15.60
#> log_p_z[21,2]   -3.04    0.00 0.37   -3.79   -3.28   -3.03   -2.79   -2.38
#> log_p_z[21,3]  -21.20    0.07 4.36  -30.79  -23.76  -20.69  -18.03  -14.32
#> log_p_z[21,4]  -21.81    0.07 4.18  -31.14  -24.20  -21.41  -18.93  -14.96
#> log_p_z[22,1]  -16.40    0.03 1.94  -20.58  -17.63  -16.27  -15.04  -13.05
#> log_p_z[22,2]   -4.65    0.01 0.59   -5.92   -5.03   -4.61   -4.23   -3.60
#> log_p_z[22,3]  -22.13    0.07 4.41  -32.01  -24.77  -21.60  -18.94  -15.13
#> log_p_z[22,4]  -21.78    0.07 4.15  -31.06  -24.17  -21.29  -18.88  -15.11
#> log_p_z[23,1]  -21.42    0.04 2.51  -26.81  -23.01  -21.27  -19.67  -16.98
#> log_p_z[23,2]   -3.72    0.01 0.45   -4.68   -4.00   -3.69   -3.40   -2.91
#> log_p_z[23,3]  -18.71    0.07 4.23  -28.37  -21.17  -18.14  -15.63  -12.30
#> log_p_z[23,4]  -21.82    0.07 4.23  -31.36  -24.24  -21.41  -18.85  -14.79
#> log_p_z[24,1]  -17.40    0.03 2.16  -22.13  -18.74  -17.19  -15.84  -13.79
#> log_p_z[24,2]   -5.82    0.01 0.65   -7.28   -6.20   -5.76   -5.36   -4.73
#> log_p_z[24,3]  -21.71    0.06 3.44  -29.32  -23.75  -21.43  -19.24  -15.97
#> log_p_z[24,4]  -21.87    0.06 3.50  -29.44  -24.02  -21.60  -19.38  -15.95
#> log_p_z[25,1]   -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_p_z[25,2]  -20.05    0.04 2.38  -25.16  -21.59  -19.84  -18.39  -15.84
#> log_p_z[25,3]  -24.77    0.07 4.20  -33.99  -27.23  -24.37  -21.80  -17.79
#> log_p_z[25,4]  -21.87    0.08 4.39  -31.71  -24.41  -21.43  -18.71  -14.65
#> log_p_z[26,1]   -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_p_z[26,2]  -20.05    0.04 2.38  -25.16  -21.59  -19.84  -18.39  -15.84
#> log_p_z[26,3]  -24.77    0.07 4.20  -33.99  -27.23  -24.37  -21.80  -17.79
#> log_p_z[26,4]  -21.87    0.08 4.39  -31.71  -24.41  -21.43  -18.71  -14.65
#> log_p_z[27,1]  -21.87    0.04 2.48  -27.10  -23.46  -21.69  -20.05  -17.65
#> log_p_z[27,2]   -5.65    0.01 0.53   -6.77   -5.98   -5.61   -5.28   -4.70
#> log_p_z[27,3]  -18.37    0.05 3.24  -25.44  -20.32  -18.08  -16.02  -13.13
#> log_p_z[27,4]  -21.51    0.06 3.40  -28.89  -23.64  -21.25  -19.12  -15.74
#> log_p_z[28,1]   -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_p_z[28,2]  -20.05    0.04 2.38  -25.16  -21.59  -19.84  -18.39  -15.84
#> log_p_z[28,3]  -24.77    0.07 4.20  -33.99  -27.23  -24.37  -21.80  -17.79
#> log_p_z[28,4]  -21.87    0.08 4.39  -31.71  -24.41  -21.43  -18.71  -14.65
#> log_p_z[29,1]   -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_p_z[29,2]  -20.05    0.04 2.38  -25.16  -21.59  -19.84  -18.39  -15.84
#> log_p_z[29,3]  -24.77    0.07 4.20  -33.99  -27.23  -24.37  -21.80  -17.79
#> log_p_z[29,4]  -21.87    0.08 4.39  -31.71  -24.41  -21.43  -18.71  -14.65
#> log_p_z[30,1]   -6.80    0.01 0.83   -8.60   -7.29   -6.72   -6.21   -5.42
#> log_p_z[30,2]  -14.76    0.03 1.73  -18.53  -15.83  -14.62  -13.57  -11.73
#> log_p_z[30,3]  -24.24    0.06 3.47  -31.81  -26.45  -23.92  -21.69  -18.37
#> log_p_z[30,4]  -21.84    0.07 3.59  -29.70  -24.11  -21.49  -19.31  -15.81
#> log_p_z[31,1]   -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_p_z[31,2]  -20.05    0.04 2.38  -25.16  -21.59  -19.84  -18.39  -15.84
#> log_p_z[31,3]  -24.77    0.07 4.20  -33.99  -27.23  -24.37  -21.80  -17.79
#> log_p_z[31,4]  -21.87    0.08 4.39  -31.71  -24.41  -21.43  -18.71  -14.65
#> log_p_z[32,1]  -29.72    0.07 4.45  -39.87  -32.31  -29.20  -26.60  -22.52
#> log_p_z[32,2]  -14.81    0.02 1.74  -18.76  -15.85  -14.67  -13.59  -11.82
#> log_p_z[32,3]   -6.06    0.01 0.94   -8.11   -6.62   -5.99   -5.41   -4.44
#> log_p_z[32,4]  -20.71    0.08 4.46  -30.76  -23.31  -20.08  -17.50  -13.65
#> log_p_z[33,1]   -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_p_z[33,2]  -20.05    0.04 2.38  -25.16  -21.59  -19.84  -18.39  -15.84
#> log_p_z[33,3]  -24.77    0.07 4.20  -33.99  -27.23  -24.37  -21.80  -17.79
#> log_p_z[33,4]  -21.87    0.08 4.39  -31.71  -24.41  -21.43  -18.71  -14.65
#> log_p_z[34,1]  -19.61    0.03 2.35  -24.67  -21.10  -19.44  -17.93  -15.60
#> log_p_z[34,2]   -3.04    0.00 0.37   -3.79   -3.28   -3.03   -2.79   -2.38
#> log_p_z[34,3]  -21.20    0.07 4.36  -30.79  -23.76  -20.69  -18.03  -14.32
#> log_p_z[34,4]  -21.81    0.07 4.18  -31.14  -24.20  -21.41  -18.93  -14.96
#> log_p_z[35,1]  -22.55    0.04 2.65  -28.27  -24.21  -22.33  -20.74  -17.83
#> log_p_z[35,2]   -5.53    0.01 0.71   -7.14   -5.95   -5.48   -5.03   -4.29
#> log_p_z[35,3]  -16.10    0.07 4.08  -25.65  -18.47  -15.53  -13.14   -9.92
#> log_p_z[35,4]  -21.82    0.07 4.21  -31.44  -24.34  -21.40  -18.79  -14.97
#> log_p_z[36,1]  -29.68    0.06 3.63  -37.57  -31.96  -29.31  -27.19  -23.56
#> log_p_z[36,2]  -21.93    0.04 2.35  -27.00  -23.41  -21.72  -20.29  -17.83
#> log_p_z[36,3]  -10.54    0.03 1.71  -14.69  -11.42  -10.26   -9.33   -7.95
#> log_p_z[36,4]  -13.98    0.07 3.21  -21.08  -15.91  -13.76  -11.62   -8.87
#> log_p_z[37,1]  -18.53    0.04 2.29  -23.58  -19.97  -18.36  -16.92  -14.64
#> log_p_z[37,2]   -7.63    0.01 0.83   -9.44   -8.14   -7.57   -7.06   -6.20
#> log_p_z[37,3]  -19.10    0.05 3.24  -26.36  -21.00  -18.79  -16.79  -13.81
#> log_p_z[37,4]  -21.87    0.06 3.52  -29.59  -24.13  -21.55  -19.37  -15.92
#> log_p_z[38,1]  -25.20    0.04 2.80  -31.26  -26.89  -25.04  -23.24  -20.31
#> log_p_z[38,2]   -9.60    0.01 1.00  -11.81  -10.20   -9.50   -8.90   -7.89
#> log_p_z[38,3]  -11.72    0.05 2.78  -18.29  -13.31  -11.34   -9.70   -7.50
#> log_p_z[38,4]  -21.31    0.06 3.48  -29.15  -23.45  -20.93  -18.86  -15.28
#> log_p_z[39,1]  -29.31    0.07 4.45  -39.48  -31.77  -28.78  -26.21  -22.17
#> log_p_z[39,2]  -16.82    0.03 2.00  -21.25  -18.06  -16.69  -15.42  -13.30
#> log_p_z[39,3]   -7.27    0.01 1.08   -9.65   -7.94   -7.18   -6.51   -5.37
#> log_p_z[39,4]  -18.78    0.07 4.13  -28.45  -21.22  -18.23  -15.81  -12.21
#> log_p_z[40,1]   -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_p_z[40,2]  -20.05    0.04 2.38  -25.16  -21.59  -19.84  -18.39  -15.84
#> log_p_z[40,3]  -24.77    0.07 4.20  -33.99  -27.23  -24.37  -21.80  -17.79
#> log_p_z[40,4]  -21.87    0.08 4.39  -31.71  -24.41  -21.43  -18.71  -14.65
#> log_p_z[41,1]   -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_p_z[41,2]  -20.05    0.04 2.38  -25.16  -21.59  -19.84  -18.39  -15.84
#> log_p_z[41,3]  -24.77    0.07 4.20  -33.99  -27.23  -24.37  -21.80  -17.79
#> log_p_z[41,4]  -21.87    0.08 4.39  -31.71  -24.41  -21.43  -18.71  -14.65
#> log_p_z[42,1]   -5.99    0.01 0.60   -7.27   -6.37   -5.96   -5.58   -4.93
#> log_p_z[42,2]  -14.96    0.03 1.76  -18.74  -16.02  -14.82  -13.75  -11.85
#> log_p_z[42,3]  -24.24    0.06 3.46  -31.78  -26.36  -23.94  -21.82  -18.32
#> log_p_z[42,4]  -21.84    0.07 3.58  -29.76  -24.03  -21.50  -19.31  -15.88
#> log_p_z[43,1]  -21.87    0.04 2.48  -27.10  -23.46  -21.69  -20.05  -17.65
#> log_p_z[43,2]   -5.65    0.01 0.53   -6.77   -5.98   -5.61   -5.28   -4.70
#> log_p_z[43,3]  -18.37    0.05 3.24  -25.44  -20.32  -18.08  -16.02  -13.13
#> log_p_z[43,4]  -21.51    0.06 3.40  -28.89  -23.64  -21.25  -19.12  -15.74
#> log_p_z[44,1]   -4.41    0.01 0.39   -5.24   -4.66   -4.38   -4.13   -3.71
#> log_p_z[44,2]  -17.27    0.03 1.91  -21.34  -18.50  -17.12  -15.97  -13.83
#> log_p_z[44,3]  -24.26    0.06 3.46  -31.55  -26.41  -23.96  -21.80  -18.34
#> log_p_z[44,4]  -21.82    0.06 3.59  -29.70  -23.97  -21.54  -19.29  -15.74
#> log_p_z[45,1]  -19.61    0.03 2.35  -24.67  -21.10  -19.44  -17.93  -15.60
#> log_p_z[45,2]   -3.04    0.00 0.37   -3.79   -3.28   -3.03   -2.79   -2.38
#> log_p_z[45,3]  -21.20    0.07 4.36  -30.79  -23.76  -20.69  -18.03  -14.32
#> log_p_z[45,4]  -21.81    0.07 4.18  -31.14  -24.20  -21.41  -18.93  -14.96
#> log_lik[1]      -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_lik[2]      -7.61    0.02 1.16  -10.11   -8.34   -7.54   -6.78   -5.58
#> log_lik[3]      -9.94    0.01 0.91  -11.76  -10.53   -9.93   -9.31   -8.24
#> log_lik[4]      -7.56    0.01 0.94   -9.53   -8.16   -7.49   -6.91   -5.89
#> log_lik[5]      -3.72    0.01 0.45   -4.68   -4.00   -3.69   -3.40   -2.91
#> log_lik[6]      -5.49    0.01 0.68   -6.97   -5.90   -5.45   -5.01   -4.29
#> log_lik[7]      -5.99    0.01 0.60   -7.27   -6.36   -5.96   -5.58   -4.93
#> log_lik[8]      -5.72    0.01 0.82   -7.54   -6.24   -5.67   -5.15   -4.33
#> log_lik[9]      -4.51    0.01 0.57   -5.77   -4.85   -4.46   -4.12   -3.53
#> log_lik[10]     -7.11    0.01 0.71   -8.67   -7.54   -7.06   -6.62   -5.87
#> log_lik[11]     -5.82    0.01 1.00   -7.97   -6.44   -5.75   -5.11   -4.13
#> log_lik[12]     -9.35    0.01 1.04  -11.51   -9.99   -9.28   -8.63   -7.47
#> log_lik[13]     -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_lik[14]     -6.23    0.01 0.87   -8.12   -6.76   -6.14   -5.61   -4.81
#> log_lik[15]     -4.41    0.01 0.39   -5.24   -4.66   -4.38   -4.13   -3.71
#> log_lik[16]     -3.79    0.01 0.54   -4.95   -4.13   -3.75   -3.41   -2.84
#> log_lik[17]     -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_lik[18]     -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_lik[19]     -4.65    0.01 0.59   -5.92   -5.03   -4.61   -4.23   -3.60
#> log_lik[20]     -7.13    0.01 0.97   -9.24   -7.72   -7.04   -6.43   -5.47
#> log_lik[21]     -3.04    0.00 0.37   -3.79   -3.28   -3.03   -2.79   -2.38
#> log_lik[22]     -4.65    0.01 0.59   -5.92   -5.03   -4.61   -4.23   -3.60
#> log_lik[23]     -3.72    0.01 0.45   -4.68   -4.00   -3.69   -3.40   -2.91
#> log_lik[24]     -5.82    0.01 0.65   -7.28   -6.20   -5.76   -5.36   -4.73
#> log_lik[25]     -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_lik[26]     -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_lik[27]     -5.65    0.01 0.53   -6.77   -5.98   -5.61   -5.28   -4.70
#> log_lik[28]     -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_lik[29]     -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_lik[30]     -6.80    0.01 0.83   -8.59   -7.29   -6.72   -6.21   -5.42
#> log_lik[31]     -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_lik[32]     -6.06    0.01 0.94   -8.11   -6.62   -5.99   -5.41   -4.44
#> log_lik[33]     -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_lik[34]     -3.04    0.00 0.37   -3.79   -3.28   -3.03   -2.79   -2.38
#> log_lik[35]     -5.52    0.01 0.71   -7.12   -5.95   -5.48   -5.02   -4.29
#> log_lik[36]     -9.94    0.02 1.16  -12.32  -10.68   -9.88   -9.12   -7.86
#> log_lik[37]     -7.63    0.01 0.82   -9.44   -8.14   -7.57   -7.06   -6.20
#> log_lik[38]     -9.00    0.01 0.85  -10.75   -9.54   -8.99   -8.44   -7.33
#> log_lik[39]     -7.27    0.01 1.08   -9.63   -7.93   -7.18   -6.51   -5.37
#> log_lik[40]     -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_lik[41]     -2.20    0.00 0.30   -2.83   -2.40   -2.19   -1.99   -1.66
#> log_lik[42]     -5.99    0.01 0.60   -7.27   -6.36   -5.96   -5.58   -4.93
#> log_lik[43]     -5.65    0.01 0.53   -6.77   -5.98   -5.61   -5.28   -4.70
#> log_lik[44]     -4.41    0.01 0.39   -5.24   -4.66   -4.38   -4.13   -3.71
#> log_lik[45]     -3.04    0.00 0.37   -3.79   -3.28   -3.03   -2.79   -2.38
#> lp__          -463.66    0.16 6.07 -476.26 -467.60 -463.28 -459.43 -452.93
#>               n_eff Rhat
#> pi[1]          7843    1
#> pi[2]          7126    1
#> pi[3]          7326    1
#> pi[4]          5849    1
#> theta[1,1,1]   8048    1
#> theta[1,1,2]   7202    1
#> theta[1,1,3]   4806    1
#> theta[1,1,4]   4810    1
#> theta[1,2,1]   6697    1
#> theta[1,2,2]   6688    1
#> theta[1,2,3]   6042    1
#> theta[1,2,4]   6213    1
#> theta[1,3,1]   5960    1
#> theta[1,3,2]   4157    1
#> theta[1,3,3]   4848    1
#> theta[1,3,4]   5907    1
#> theta[1,4,1]   6514    1
#> theta[1,4,2]   6881    1
#> theta[1,4,3]   4265    1
#> theta[1,4,4]   5185    1
#> theta[2,1,1]   9186    1
#> theta[2,1,2]   6802    1
#> theta[2,1,3]   4922    1
#> theta[2,1,4]   5127    1
#> theta[2,2,1]   7239    1
#> theta[2,2,2]   8017    1
#> theta[2,2,3]   7943    1
#> theta[2,2,4]   6205    1
#> theta[2,3,1]   6404    1
#> theta[2,3,2]   6421    1
#> theta[2,3,3]   6155    1
#> theta[2,3,4]   5249    1
#> theta[2,4,1]   6569    1
#> theta[2,4,2]   5807    1
#> theta[2,4,3]   5497    1
#> theta[2,4,4]   5581    1
#> theta[3,1,1]   7368    1
#> theta[3,1,2]   5114    1
#> theta[3,1,3]   4319    1
#> theta[3,1,4]   4702    1
#> theta[3,2,1]   6771    1
#> theta[3,2,2]   7434    1
#> theta[3,2,3]   7306    1
#> theta[3,2,4]   5180    1
#> theta[3,3,1]   6591    1
#> theta[3,3,2]   5810    1
#> theta[3,3,3]   5591    1
#> theta[3,3,4]   7044    1
#> theta[3,4,1]   6031    1
#> theta[3,4,2]   6049    1
#> theta[3,4,3]   5660    1
#> theta[3,4,4]   6349    1
#> theta[4,1,1]   7261    1
#> theta[4,1,2]   5569    1
#> theta[4,1,3]   5019    1
#> theta[4,1,4]   4333    1
#> theta[4,2,1]   7919    1
#> theta[4,2,2]   6837    1
#> theta[4,2,3]   6050    1
#> theta[4,2,4]   5708    1
#> theta[4,3,1]   7615    1
#> theta[4,3,2]   5820    1
#> theta[4,3,3]   6247    1
#> theta[4,3,4]   5072    1
#> theta[4,4,1]   6151    1
#> theta[4,4,2]   6170    1
#> theta[4,4,3]   6111    1
#> theta[4,4,4]   6574    1
#> theta[5,1,1]   6912    1
#> theta[5,1,2]   5269    1
#> theta[5,1,3]   5545    1
#> theta[5,1,4]   4282    1
#> theta[5,2,1]   7448    1
#> theta[5,2,2]   7843    1
#> theta[5,2,3]   7018    1
#> theta[5,2,4]   6028    1
#> theta[5,3,1]   6383    1
#> theta[5,3,2]   6872    1
#> theta[5,3,3]   6575    1
#> theta[5,3,4]   6923    1
#> theta[5,4,1]   6873    1
#> theta[5,4,2]   5985    1
#> theta[5,4,3]   4970    1
#> theta[5,4,4]   5602    1
#> log_p_z[1,1]   8264    1
#> log_p_z[1,2]   4552    1
#> log_p_z[1,3]   3633    1
#> log_p_z[1,4]   3033    1
#> log_p_z[2,1]   4853    1
#> log_p_z[2,2]   4147    1
#> log_p_z[2,3]   4024    1
#> log_p_z[2,4]   2831    1
#> log_p_z[3,1]   3889    1
#> log_p_z[3,2]   4088    1
#> log_p_z[3,3]   3784    1
#> log_p_z[3,4]   2550    1
#> log_p_z[4,1]   5320    1
#> log_p_z[4,2]   4882    1
#> log_p_z[4,3]   3627    1
#> log_p_z[4,4]   3305    1
#> log_p_z[5,1]   4545    1
#> log_p_z[5,2]   6987    1
#> log_p_z[5,3]   3618    1
#> log_p_z[5,4]   3740    1
#> log_p_z[6,1]   4871    1
#> log_p_z[6,2]   6539    1
#> log_p_z[6,3]   3684    1
#> log_p_z[6,4]   3840    1
#> log_p_z[7,1]   6832    1
#> log_p_z[7,2]   4399    1
#> log_p_z[7,3]   3589    1
#> log_p_z[7,4]   3021    1
#> log_p_z[8,1]   4467    1
#> log_p_z[8,2]   4659    1
#> log_p_z[8,3]   6623    1
#> log_p_z[8,4]   2760    1
#> log_p_z[9,1]   5006    1
#> log_p_z[9,2]   6367    1
#> log_p_z[9,3]   3589    1
#> log_p_z[9,4]   3887    1
#> log_p_z[10,1]  4557    1
#> log_p_z[10,2]  5524    1
#> log_p_z[10,3]  3562    1
#> log_p_z[10,4]  3663    1
#> log_p_z[11,1]  4661    1
#> log_p_z[11,2]  4165    1
#> log_p_z[11,3]  2966    1
#> log_p_z[11,4]  6530    1
#> log_p_z[12,1]  5107    1
#> log_p_z[12,2]  4495    1
#> log_p_z[12,3]  3664    1
#> log_p_z[12,4]  3894    1
#> log_p_z[13,1]  8264    1
#> log_p_z[13,2]  4552    1
#> log_p_z[13,3]  3633    1
#> log_p_z[13,4]  3033    1
#> log_p_z[14,1]  4486    1
#> log_p_z[14,2]  5112    1
#> log_p_z[14,3]  3555    1
#> log_p_z[14,4]  3910    1
#> log_p_z[15,1]  5941    1
#> log_p_z[15,2]  4508    1
#> log_p_z[15,3]  3505    1
#> log_p_z[15,4]  3048    1
#> log_p_z[16,1]  7347    1
#> log_p_z[16,2]  4465    1
#> log_p_z[16,3]  3659    1
#> log_p_z[16,4]  3078    1
#> log_p_z[17,1]  8264    1
#> log_p_z[17,2]  4552    1
#> log_p_z[17,3]  3633    1
#> log_p_z[17,4]  3033    1
#> log_p_z[18,1]  8264    1
#> log_p_z[18,2]  4552    1
#> log_p_z[18,3]  3633    1
#> log_p_z[18,4]  3033    1
#> log_p_z[19,1]  5133    1
#> log_p_z[19,2]  6285    1
#> log_p_z[19,3]  3490    1
#> log_p_z[19,4]  3809    1
#> log_p_z[20,1]  4642    1
#> log_p_z[20,2]  4612    1
#> log_p_z[20,3]  3620    1
#> log_p_z[20,4]  3822    1
#> log_p_z[21,1]  4595    1
#> log_p_z[21,2]  6679    1
#> log_p_z[21,3]  3609    1
#> log_p_z[21,4]  3724    1
#> log_p_z[22,1]  5133    1
#> log_p_z[22,2]  6285    1
#> log_p_z[22,3]  3490    1
#> log_p_z[22,4]  3809    1
#> log_p_z[23,1]  4545    1
#> log_p_z[23,2]  6987    1
#> log_p_z[23,3]  3618    1
#> log_p_z[23,4]  3740    1
#> log_p_z[24,1]  4343    1
#> log_p_z[24,2]  4621    1
#> log_p_z[24,3]  3658    1
#> log_p_z[24,4]  3216    1
#> log_p_z[25,1]  8264    1
#> log_p_z[25,2]  4552    1
#> log_p_z[25,3]  3633    1
#> log_p_z[25,4]  3033    1
#> log_p_z[26,1]  8264    1
#> log_p_z[26,2]  4552    1
#> log_p_z[26,3]  3633    1
#> log_p_z[26,4]  3033    1
#> log_p_z[27,1]  4425    1
#> log_p_z[27,2]  6189    1
#> log_p_z[27,3]  3597    1
#> log_p_z[27,4]  3744    1
#> log_p_z[28,1]  8264    1
#> log_p_z[28,2]  4552    1
#> log_p_z[28,3]  3633    1
#> log_p_z[28,4]  3033    1
#> log_p_z[29,1]  8264    1
#> log_p_z[29,2]  4552    1
#> log_p_z[29,3]  3633    1
#> log_p_z[29,4]  3033    1
#> log_p_z[30,1]  4943    1
#> log_p_z[30,2]  4380    1
#> log_p_z[30,3]  3748    1
#> log_p_z[30,4]  2682    1
#> log_p_z[31,1]  8264    1
#> log_p_z[31,2]  4552    1
#> log_p_z[31,3]  3633    1
#> log_p_z[31,4]  3033    1
#> log_p_z[32,1]  4517    1
#> log_p_z[32,2]  4960    1
#> log_p_z[32,3]  4528    1
#> log_p_z[32,4]  2915    1
#> log_p_z[33,1]  8264    1
#> log_p_z[33,2]  4552    1
#> log_p_z[33,3]  3633    1
#> log_p_z[33,4]  3033    1
#> log_p_z[34,1]  4595    1
#> log_p_z[34,2]  6679    1
#> log_p_z[34,3]  3609    1
#> log_p_z[34,4]  3724    1
#> log_p_z[35,1]  4327    1
#> log_p_z[35,2]  5797    1
#> log_p_z[35,3]  3629    1
#> log_p_z[35,4]  3847    1
#> log_p_z[36,1]  4050    1
#> log_p_z[36,2]  3820    1
#> log_p_z[36,3]  2488    1
#> log_p_z[36,4]  2395    1
#> log_p_z[37,1]  4026    1
#> log_p_z[37,2]  5306    1
#> log_p_z[37,3]  3620    1
#> log_p_z[37,4]  3339    1
#> log_p_z[38,1]  4273    1
#> log_p_z[38,2]  4702    1
#> log_p_z[38,3]  3733    1
#> log_p_z[38,4]  3517    1
#> log_p_z[39,1]  4630    1
#> log_p_z[39,2]  4887    1
#> log_p_z[39,3]  5738    1
#> log_p_z[39,4]  3124    1
#> log_p_z[40,1]  8264    1
#> log_p_z[40,2]  4552    1
#> log_p_z[40,3]  3633    1
#> log_p_z[40,4]  3033    1
#> log_p_z[41,1]  8264    1
#> log_p_z[41,2]  4552    1
#> log_p_z[41,3]  3633    1
#> log_p_z[41,4]  3033    1
#> log_p_z[42,1]  6832    1
#> log_p_z[42,2]  4399    1
#> log_p_z[42,3]  3589    1
#> log_p_z[42,4]  3021    1
#> log_p_z[43,1]  4425    1
#> log_p_z[43,2]  6189    1
#> log_p_z[43,3]  3597    1
#> log_p_z[43,4]  3744    1
#> log_p_z[44,1]  5941    1
#> log_p_z[44,2]  4508    1
#> log_p_z[44,3]  3505    1
#> log_p_z[44,4]  3048    1
#> log_p_z[45,1]  4595    1
#> log_p_z[45,2]  6679    1
#> log_p_z[45,3]  3609    1
#> log_p_z[45,4]  3724    1
#> log_lik[1]     8264    1
#> log_lik[2]     4620    1
#> log_lik[3]     5578    1
#> log_lik[4]     4975    1
#> log_lik[5]     6986    1
#> log_lik[6]     6546    1
#> log_lik[7]     6838    1
#> log_lik[8]     6645    1
#> log_lik[9]     6368    1
#> log_lik[10]    5544    1
#> log_lik[11]    6531    1
#> log_lik[12]    5562    1
#> log_lik[13]    8264    1
#> log_lik[14]    5118    1
#> log_lik[15]    5942    1
#> log_lik[16]    7347    1
#> log_lik[17]    8264    1
#> log_lik[18]    8264    1
#> log_lik[19]    6285    1
#> log_lik[20]    4631    1
#> log_lik[21]    6679    1
#> log_lik[22]    6285    1
#> log_lik[23]    6986    1
#> log_lik[24]    4622    1
#> log_lik[25]    8264    1
#> log_lik[26]    8264    1
#> log_lik[27]    6190    1
#> log_lik[28]    8264    1
#> log_lik[29]    8264    1
#> log_lik[30]    4960    1
#> log_lik[31]    8264    1
#> log_lik[32]    4540    1
#> log_lik[33]    8264    1
#> log_lik[34]    6679    1
#> log_lik[35]    5841    1
#> log_lik[36]    4911    1
#> log_lik[37]    5315    1
#> log_lik[38]    4608    1
#> log_lik[39]    5900    1
#> log_lik[40]    8264    1
#> log_lik[41]    8264    1
#> log_lik[42]    6838    1
#> log_lik[43]    6190    1
#> log_lik[44]    5942    1
#> log_lik[45]    6679    1
#> lp__           1376    1
#> 
#> Samples were drawn using NUTS(diag_e) at Mon Sep 11 19:03:08 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }