Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.37    0.00 0.06    0.26    0.33    0.37    0.42    0.50
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.15
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.90    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.02    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.75    0.82    0.85    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.07    0.17
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.58    0.74    0.80    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.30
#> theta[1,4,2]     0.09    0.00 0.09    0.00    0.03    0.07    0.13    0.33
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.18    0.39
#> theta[1,4,4]     0.69    0.00 0.14    0.40    0.60    0.70    0.79    0.91
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.76    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.21    0.33
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.58    0.65    0.75
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.25    0.30    0.36    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.26
#> theta[2,3,2]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.28
#> theta[2,3,3]     0.62    0.00 0.13    0.35    0.54    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.28    0.47
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.36
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.22
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.74    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.09    0.16    0.24    0.43
#> theta[3,3,3]     0.50    0.00 0.14    0.24    0.41    0.50    0.60    0.76
#> theta[3,3,4]     0.24    0.00 0.11    0.06    0.15    0.22    0.31    0.49
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[3,4,2]     0.12    0.00 0.10    0.01    0.04    0.09    0.17    0.39
#> theta[3,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.42
#> theta[3,4,4]     0.63    0.00 0.15    0.32    0.52    0.64    0.75    0.89
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.78    0.83    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.08    0.00 0.05    0.01    0.04    0.07    0.10    0.20
#> theta[4,2,2]     0.72    0.00 0.08    0.54    0.67    0.73    0.78    0.87
#> theta[4,2,3]     0.13    0.00 0.06    0.03    0.09    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.26
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.61    0.69    0.78    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.20    0.38
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.76    0.90
#> theta[5,1,1]     0.86    0.00 0.07    0.71    0.83    0.87    0.91    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[5,2,1]     0.15    0.00 0.06    0.04    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.48    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.22    0.40
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.61    0.70    0.84
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.42
#> theta[5,4,4]     0.63    0.00 0.15    0.31    0.53    0.64    0.74    0.90
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_p_z[1,2]   -19.92    0.04 2.28  -24.86  -21.26  -19.72  -18.36  -16.02
#> log_p_z[1,3]   -24.94    0.08 4.40  -34.78  -27.60  -24.38  -21.86  -17.79
#> log_p_z[1,4]   -21.96    0.08 4.39  -31.56  -24.58  -21.48  -18.84  -14.87
#> log_p_z[2,1]   -29.72    0.07 4.51  -39.61  -32.30  -29.28  -26.61  -22.59
#> log_p_z[2,2]   -21.02    0.04 2.38  -26.25  -22.48  -20.82  -19.31  -16.93
#> log_p_z[2,3]    -7.65    0.02 1.21  -10.35   -8.39   -7.55   -6.81   -5.63
#> log_p_z[2,4]   -16.39    0.08 4.14  -25.77  -18.78  -15.88  -13.38   -9.69
#> log_p_z[3,1]   -11.61    0.03 1.62  -15.33  -12.53  -11.40  -10.46   -9.04
#> log_p_z[3,2]   -10.79    0.03 1.46  -14.02  -11.69  -10.64   -9.73   -8.41
#> log_p_z[3,3]   -23.40    0.06 3.51  -31.31  -25.48  -23.07  -20.93  -17.53
#> log_p_z[3,4]   -21.96    0.07 3.55  -29.75  -24.15  -21.66  -19.44  -16.08
#> log_p_z[4,1]   -14.64    0.03 1.89  -18.81  -15.80  -14.48  -13.34  -11.30
#> log_p_z[4,2]    -7.56    0.02 0.98   -9.57   -8.16   -7.49   -6.88   -5.91
#> log_p_z[4,3]   -20.77    0.08 4.30  -30.61  -23.17  -20.29  -17.73  -13.72
#> log_p_z[4,4]   -21.85    0.07 4.34  -31.33  -24.44  -21.30  -18.73  -14.97
#> log_p_z[5,1]   -21.46    0.04 2.61  -27.08  -23.09  -21.24  -19.60  -16.90
#> log_p_z[5,2]    -3.71    0.01 0.42   -4.55   -3.98   -3.69   -3.42   -2.95
#> log_p_z[5,3]   -18.69    0.08 4.20  -28.25  -21.08  -18.18  -15.71  -11.99
#> log_p_z[5,4]   -21.83    0.07 4.33  -31.33  -24.47  -21.27  -18.74  -14.88
#> log_p_z[6,1]   -21.45    0.04 2.57  -27.06  -23.00  -21.27  -19.67  -17.01
#> log_p_z[6,2]    -5.47    0.01 0.66   -6.88   -5.89   -5.44   -5.02   -4.30
#> log_p_z[6,3]   -17.46    0.07 4.03  -26.84  -19.72  -17.00  -14.64  -11.12
#> log_p_z[6,4]   -21.62    0.07 4.31  -31.22  -24.19  -21.05  -18.53  -14.68
#> log_p_z[7,1]    -6.01    0.01 0.62   -7.37   -6.40   -5.97   -5.57   -4.92
#> log_p_z[7,2]   -14.89    0.03 1.70  -18.57  -15.91  -14.73  -13.69  -12.02
#> log_p_z[7,3]   -24.39    0.07 3.64  -32.34  -26.58  -24.09  -21.75  -18.31
#> log_p_z[7,4]   -21.95    0.07 3.57  -29.88  -24.15  -21.63  -19.34  -16.02
#> log_p_z[8,1]   -29.81    0.07 4.54  -39.84  -32.40  -29.31  -26.69  -22.49
#> log_p_z[8,2]   -18.27    0.03 2.06  -22.90  -19.53  -18.14  -16.85  -14.68
#> log_p_z[8,3]    -5.70    0.01 0.82   -7.41   -6.21   -5.65   -5.13   -4.26
#> log_p_z[8,4]   -18.59    0.08 4.30  -28.60  -21.09  -17.97  -15.52  -11.92
#> log_p_z[9,1]   -20.02    0.04 2.30  -25.13  -21.44  -19.79  -18.46  -15.97
#> log_p_z[9,2]    -4.52    0.01 0.56   -5.75   -4.87   -4.47   -4.13   -3.53
#> log_p_z[9,3]   -19.66    0.08 4.35  -29.48  -22.18  -19.17  -16.58  -12.92
#> log_p_z[9,4]   -21.68    0.07 4.23  -31.40  -24.13  -21.16  -18.63  -14.80
#> log_p_z[10,1]  -22.28    0.04 2.47  -27.76  -23.82  -22.04  -20.53  -18.12
#> log_p_z[10,2]   -7.12    0.01 0.71   -8.75   -7.53   -7.06   -6.64   -5.89
#> log_p_z[10,3]  -16.83    0.06 3.20  -23.84  -18.70  -16.47  -14.57  -11.64
#> log_p_z[10,4]  -21.38    0.06 3.46  -28.94  -23.60  -21.08  -18.89  -15.57
#> log_p_z[11,1]  -29.77    0.07 4.36  -39.56  -32.30  -29.33  -26.70  -22.64
#> log_p_z[11,2]  -29.26    0.07 4.11  -38.70  -31.63  -28.89  -26.35  -22.31
#> log_p_z[11,3]  -18.16    0.06 3.23  -25.61  -19.86  -17.74  -15.91  -13.10
#> log_p_z[11,4]   -5.84    0.01 0.99   -7.91   -6.47   -5.79   -5.16   -4.13
#> log_p_z[12,1]  -22.92    0.04 2.70  -28.90  -24.62  -22.70  -20.95  -18.39
#> log_p_z[12,2]   -9.58    0.02 1.28  -12.52  -10.33   -9.45   -8.69   -7.46
#> log_p_z[12,3]  -15.19    0.07 3.97  -24.56  -17.45  -14.75  -12.38   -9.00
#> log_p_z[12,4]  -19.26    0.07 4.17  -28.84  -21.58  -18.74  -16.34  -12.69
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_p_z[13,2]  -19.92    0.04 2.28  -24.86  -21.26  -19.72  -18.36  -16.02
#> log_p_z[13,3]  -24.94    0.08 4.40  -34.78  -27.60  -24.38  -21.86  -17.79
#> log_p_z[13,4]  -21.96    0.08 4.39  -31.56  -24.58  -21.48  -18.84  -14.87
#> log_p_z[14,1]  -19.07    0.04 2.46  -24.41  -20.61  -18.85  -17.34  -14.84
#> log_p_z[14,2]   -6.21    0.01 0.89   -8.21   -6.72   -6.13   -5.58   -4.75
#> log_p_z[14,3]  -18.68    0.08 4.20  -28.52  -21.11  -18.25  -15.68  -12.01
#> log_p_z[14,4]  -21.80    0.07 4.34  -31.49  -24.32  -21.21  -18.79  -14.88
#> log_p_z[15,1]   -4.42    0.01 0.40   -5.30   -4.67   -4.38   -4.13   -3.72
#> log_p_z[15,2]  -17.18    0.03 1.85  -21.23  -18.25  -17.02  -15.90  -13.93
#> log_p_z[15,3]  -24.38    0.07 3.60  -32.50  -26.58  -24.05  -21.82  -18.33
#> log_p_z[15,4]  -21.92    0.07 3.58  -29.58  -24.09  -21.59  -19.35  -15.95
#> log_p_z[16,1]   -3.81    0.01 0.55   -5.01   -4.14   -3.76   -3.42   -2.87
#> log_p_z[16,2]  -17.64    0.04 2.16  -22.30  -18.92  -17.44  -16.16  -14.01
#> log_p_z[16,3]  -24.95    0.08 4.42  -34.79  -27.59  -24.48  -21.69  -17.82
#> log_p_z[16,4]  -22.00    0.08 4.38  -31.61  -24.58  -21.51  -18.72  -15.01
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_p_z[17,2]  -19.92    0.04 2.28  -24.86  -21.26  -19.72  -18.36  -16.02
#> log_p_z[17,3]  -24.94    0.08 4.40  -34.78  -27.60  -24.38  -21.86  -17.79
#> log_p_z[17,4]  -21.96    0.08 4.39  -31.56  -24.58  -21.48  -18.84  -14.87
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_p_z[18,2]  -19.92    0.04 2.28  -24.86  -21.26  -19.72  -18.36  -16.02
#> log_p_z[18,3]  -24.94    0.08 4.40  -34.78  -27.60  -24.38  -21.86  -17.79
#> log_p_z[18,4]  -21.96    0.08 4.39  -31.56  -24.58  -21.48  -18.84  -14.87
#> log_p_z[19,1]  -16.42    0.03 1.97  -20.66  -17.68  -16.25  -15.04  -13.06
#> log_p_z[19,2]   -4.64    0.01 0.58   -5.90   -5.02   -4.61   -4.23   -3.62
#> log_p_z[19,3]  -22.19    0.08 4.44  -31.80  -24.79  -21.76  -18.98  -14.99
#> log_p_z[19,4]  -21.89    0.07 4.28  -31.70  -24.34  -21.39  -18.80  -15.10
#> log_p_z[20,1]  -18.02    0.03 2.28  -23.14  -19.38  -17.80  -16.44  -14.11
#> log_p_z[20,2]   -7.09    0.01 0.92   -9.10   -7.65   -7.02   -6.44   -5.47
#> log_p_z[20,3]  -19.97    0.08 4.17  -29.62  -22.36  -19.47  -17.04  -13.24
#> log_p_z[20,4]  -21.61    0.07 4.26  -31.02  -24.17  -21.09  -18.53  -14.88
#> log_p_z[21,1]  -19.63    0.04 2.39  -24.88  -21.10  -19.41  -17.96  -15.53
#> log_p_z[21,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.80   -2.38
#> log_p_z[21,3]  -21.20    0.08 4.36  -30.85  -23.68  -20.78  -18.13  -14.20
#> log_p_z[21,4]  -21.86    0.07 4.28  -31.40  -24.43  -21.36  -18.77  -14.96
#> log_p_z[22,1]  -16.42    0.03 1.97  -20.66  -17.68  -16.25  -15.04  -13.06
#> log_p_z[22,2]   -4.64    0.01 0.58   -5.90   -5.02   -4.61   -4.23   -3.62
#> log_p_z[22,3]  -22.19    0.08 4.44  -31.80  -24.79  -21.76  -18.98  -14.99
#> log_p_z[22,4]  -21.89    0.07 4.28  -31.70  -24.34  -21.39  -18.80  -15.10
#> log_p_z[23,1]  -21.46    0.04 2.61  -27.08  -23.09  -21.24  -19.60  -16.90
#> log_p_z[23,2]   -3.71    0.01 0.42   -4.55   -3.98   -3.69   -3.42   -2.95
#> log_p_z[23,3]  -18.69    0.08 4.20  -28.25  -21.08  -18.18  -15.71  -11.99
#> log_p_z[23,4]  -21.83    0.07 4.33  -31.33  -24.47  -21.27  -18.74  -14.88
#> log_p_z[24,1]  -17.42    0.04 2.17  -22.26  -18.76  -17.20  -15.90  -13.78
#> log_p_z[24,2]   -5.79    0.01 0.60   -7.09   -6.15   -5.73   -5.35   -4.80
#> log_p_z[24,3]  -21.75    0.06 3.48  -29.47  -23.82  -21.43  -19.34  -15.96
#> log_p_z[24,4]  -21.91    0.06 3.52  -29.51  -24.08  -21.67  -19.38  -15.93
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_p_z[25,2]  -19.92    0.04 2.28  -24.86  -21.26  -19.72  -18.36  -16.02
#> log_p_z[25,3]  -24.94    0.08 4.40  -34.78  -27.60  -24.38  -21.86  -17.79
#> log_p_z[25,4]  -21.96    0.08 4.39  -31.56  -24.58  -21.48  -18.84  -14.87
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_p_z[26,2]  -19.92    0.04 2.28  -24.86  -21.26  -19.72  -18.36  -16.02
#> log_p_z[26,3]  -24.94    0.08 4.40  -34.78  -27.60  -24.38  -21.86  -17.79
#> log_p_z[26,4]  -21.96    0.08 4.39  -31.56  -24.58  -21.48  -18.84  -14.87
#> log_p_z[27,1]  -21.89    0.04 2.50  -27.40  -23.47  -21.66  -20.14  -17.64
#> log_p_z[27,2]   -5.64    0.01 0.53   -6.80   -5.97   -5.60   -5.26   -4.71
#> log_p_z[27,3]  -18.37    0.06 3.24  -25.60  -20.23  -18.09  -16.10  -12.95
#> log_p_z[27,4]  -21.56    0.06 3.47  -29.05  -23.75  -21.26  -19.02  -15.61
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_p_z[28,2]  -19.92    0.04 2.28  -24.86  -21.26  -19.72  -18.36  -16.02
#> log_p_z[28,3]  -24.94    0.08 4.40  -34.78  -27.60  -24.38  -21.86  -17.79
#> log_p_z[28,4]  -21.96    0.08 4.39  -31.56  -24.58  -21.48  -18.84  -14.87
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_p_z[29,2]  -19.92    0.04 2.28  -24.86  -21.26  -19.72  -18.36  -16.02
#> log_p_z[29,3]  -24.94    0.08 4.40  -34.78  -27.60  -24.38  -21.86  -17.79
#> log_p_z[29,4]  -21.96    0.08 4.39  -31.56  -24.58  -21.48  -18.84  -14.87
#> log_p_z[30,1]   -6.80    0.01 0.84   -8.72   -7.30   -6.71   -6.22   -5.38
#> log_p_z[30,2]  -14.67    0.03 1.69  -18.23  -15.72  -14.51  -13.48  -11.78
#> log_p_z[30,3]  -24.39    0.07 3.57  -32.48  -26.52  -24.04  -21.94  -18.38
#> log_p_z[30,4]  -21.94    0.06 3.57  -29.82  -24.11  -21.61  -19.39  -15.95
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_p_z[31,2]  -19.92    0.04 2.28  -24.86  -21.26  -19.72  -18.36  -16.02
#> log_p_z[31,3]  -24.94    0.08 4.40  -34.78  -27.60  -24.38  -21.86  -17.79
#> log_p_z[31,4]  -21.96    0.08 4.39  -31.56  -24.58  -21.48  -18.84  -14.87
#> log_p_z[32,1]  -29.78    0.07 4.57  -39.75  -32.33  -29.22  -26.58  -22.52
#> log_p_z[32,2]  -14.80    0.02 1.72  -18.59  -15.84  -14.69  -13.57  -11.85
#> log_p_z[32,3]   -6.08    0.01 0.96   -8.18   -6.66   -6.01   -5.39   -4.45
#> log_p_z[32,4]  -20.72    0.08 4.42  -31.08  -23.31  -20.18  -17.54  -13.76
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_p_z[33,2]  -19.92    0.04 2.28  -24.86  -21.26  -19.72  -18.36  -16.02
#> log_p_z[33,3]  -24.94    0.08 4.40  -34.78  -27.60  -24.38  -21.86  -17.79
#> log_p_z[33,4]  -21.96    0.08 4.39  -31.56  -24.58  -21.48  -18.84  -14.87
#> log_p_z[34,1]  -19.63    0.04 2.39  -24.88  -21.10  -19.41  -17.96  -15.53
#> log_p_z[34,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.80   -2.38
#> log_p_z[34,3]  -21.20    0.08 4.36  -30.85  -23.68  -20.78  -18.13  -14.20
#> log_p_z[34,4]  -21.86    0.07 4.28  -31.40  -24.43  -21.36  -18.77  -14.96
#> log_p_z[35,1]  -22.60    0.04 2.74  -28.50  -24.39  -22.39  -20.66  -17.84
#> log_p_z[35,2]   -5.52    0.01 0.68   -7.04   -5.93   -5.48   -5.06   -4.33
#> log_p_z[35,3]  -16.11    0.08 4.06  -25.51  -18.40  -15.59  -13.16   -9.87
#> log_p_z[35,4]  -21.80    0.07 4.32  -31.45  -24.33  -21.31  -18.73  -14.93
#> log_p_z[36,1]  -29.70    0.06 3.70  -37.56  -31.98  -29.44  -27.06  -23.51
#> log_p_z[36,2]  -21.88    0.04 2.28  -26.70  -23.30  -21.71  -20.29  -17.85
#> log_p_z[36,3]  -10.53    0.04 1.72  -14.71  -11.42  -10.27   -9.34   -7.92
#> log_p_z[36,4]  -14.04    0.06 3.18  -21.20  -16.15  -13.71  -11.68   -8.76
#> log_p_z[37,1]  -18.56    0.04 2.32  -23.71  -19.97  -18.32  -16.98  -14.54
#> log_p_z[37,2]   -7.61    0.01 0.79   -9.39   -8.09   -7.54   -7.05   -6.23
#> log_p_z[37,3]  -19.18    0.06 3.33  -26.86  -21.15  -18.83  -16.80  -13.67
#> log_p_z[37,4]  -21.88    0.06 3.52  -29.56  -24.10  -21.56  -19.36  -15.87
#> log_p_z[38,1]  -25.25    0.05 2.89  -31.49  -27.07  -24.98  -23.19  -20.31
#> log_p_z[38,2]   -9.60    0.02 0.99  -11.79  -10.20   -9.52   -8.90   -7.89
#> log_p_z[38,3]  -11.74    0.05 2.77  -18.05  -13.32  -11.39   -9.71   -7.51
#> log_p_z[38,4]  -21.32    0.06 3.53  -29.02  -23.45  -20.99  -18.74  -15.42
#> log_p_z[39,1]  -29.42    0.07 4.45  -38.95  -32.00  -28.94  -26.34  -22.29
#> log_p_z[39,2]  -16.79    0.03 1.97  -21.10  -17.98  -16.62  -15.43  -13.44
#> log_p_z[39,3]   -7.24    0.02 1.11   -9.69   -7.94   -7.14   -6.47   -5.30
#> log_p_z[39,4]  -18.77    0.08 4.22  -28.51  -21.26  -18.16  -15.72  -12.25
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_p_z[40,2]  -19.92    0.04 2.28  -24.86  -21.26  -19.72  -18.36  -16.02
#> log_p_z[40,3]  -24.94    0.08 4.40  -34.78  -27.60  -24.38  -21.86  -17.79
#> log_p_z[40,4]  -21.96    0.08 4.39  -31.56  -24.58  -21.48  -18.84  -14.87
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_p_z[41,2]  -19.92    0.04 2.28  -24.86  -21.26  -19.72  -18.36  -16.02
#> log_p_z[41,3]  -24.94    0.08 4.40  -34.78  -27.60  -24.38  -21.86  -17.79
#> log_p_z[41,4]  -21.96    0.08 4.39  -31.56  -24.58  -21.48  -18.84  -14.87
#> log_p_z[42,1]   -6.01    0.01 0.62   -7.37   -6.40   -5.97   -5.57   -4.92
#> log_p_z[42,2]  -14.89    0.03 1.70  -18.57  -15.91  -14.73  -13.69  -12.02
#> log_p_z[42,3]  -24.39    0.07 3.64  -32.34  -26.58  -24.09  -21.75  -18.31
#> log_p_z[42,4]  -21.95    0.07 3.57  -29.88  -24.15  -21.63  -19.34  -16.02
#> log_p_z[43,1]  -21.89    0.04 2.50  -27.40  -23.47  -21.66  -20.14  -17.64
#> log_p_z[43,2]   -5.64    0.01 0.53   -6.80   -5.97   -5.60   -5.26   -4.71
#> log_p_z[43,3]  -18.37    0.06 3.24  -25.60  -20.23  -18.09  -16.10  -12.95
#> log_p_z[43,4]  -21.56    0.06 3.47  -29.05  -23.75  -21.26  -19.02  -15.61
#> log_p_z[44,1]   -4.42    0.01 0.40   -5.30   -4.67   -4.38   -4.13   -3.72
#> log_p_z[44,2]  -17.18    0.03 1.85  -21.23  -18.25  -17.02  -15.90  -13.93
#> log_p_z[44,3]  -24.38    0.07 3.60  -32.50  -26.58  -24.05  -21.82  -18.33
#> log_p_z[44,4]  -21.92    0.07 3.58  -29.58  -24.09  -21.59  -19.35  -15.95
#> log_p_z[45,1]  -19.63    0.04 2.39  -24.88  -21.10  -19.41  -17.96  -15.53
#> log_p_z[45,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.80   -2.38
#> log_p_z[45,3]  -21.20    0.08 4.36  -30.85  -23.68  -20.78  -18.13  -14.20
#> log_p_z[45,4]  -21.86    0.07 4.28  -31.40  -24.43  -21.36  -18.77  -14.96
#> log_lik[1]      -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_lik[2]      -7.61    0.02 1.13  -10.01   -8.35   -7.54   -6.80   -5.62
#> log_lik[3]      -9.95    0.01 0.90  -11.81  -10.53   -9.92   -9.33   -8.28
#> log_lik[4]      -7.55    0.02 0.96   -9.53   -8.15   -7.48   -6.88   -5.91
#> log_lik[5]      -3.71    0.01 0.42   -4.55   -3.98   -3.69   -3.42   -2.95
#> log_lik[6]      -5.47    0.01 0.66   -6.88   -5.89   -5.44   -5.02   -4.30
#> log_lik[7]      -6.01    0.01 0.62   -7.37   -6.40   -5.97   -5.57   -4.92
#> log_lik[8]      -5.70    0.01 0.82   -7.41   -6.21   -5.65   -5.13   -4.26
#> log_lik[9]      -4.52    0.01 0.56   -5.75   -4.87   -4.47   -4.13   -3.53
#> log_lik[10]     -7.12    0.01 0.70   -8.72   -7.52   -7.06   -6.64   -5.89
#> log_lik[11]     -5.84    0.01 0.99   -7.91   -6.47   -5.79   -5.16   -4.13
#> log_lik[12]     -9.34    0.01 1.07  -11.64  -10.02   -9.28   -8.59   -7.44
#> log_lik[13]     -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_lik[14]     -6.21    0.01 0.88   -8.21   -6.72   -6.13   -5.58   -4.75
#> log_lik[15]     -4.42    0.01 0.40   -5.30   -4.67   -4.38   -4.13   -3.72
#> log_lik[16]     -3.81    0.01 0.55   -5.01   -4.14   -3.76   -3.42   -2.87
#> log_lik[17]     -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_lik[18]     -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_lik[19]     -4.64    0.01 0.58   -5.90   -5.02   -4.61   -4.23   -3.62
#> log_lik[20]     -7.09    0.01 0.92   -9.10   -7.65   -7.02   -6.44   -5.47
#> log_lik[21]     -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.80   -2.38
#> log_lik[22]     -4.64    0.01 0.58   -5.90   -5.02   -4.61   -4.23   -3.62
#> log_lik[23]     -3.71    0.01 0.42   -4.55   -3.98   -3.69   -3.42   -2.95
#> log_lik[24]     -5.79    0.01 0.60   -7.09   -6.15   -5.73   -5.35   -4.80
#> log_lik[25]     -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_lik[26]     -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_lik[27]     -5.64    0.01 0.53   -6.80   -5.97   -5.60   -5.26   -4.71
#> log_lik[28]     -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_lik[29]     -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_lik[30]     -6.80    0.01 0.84   -8.71   -7.30   -6.71   -6.22   -5.38
#> log_lik[31]     -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_lik[32]     -6.08    0.01 0.96   -8.18   -6.66   -6.01   -5.39   -4.45
#> log_lik[33]     -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_lik[34]     -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.80   -2.38
#> log_lik[35]     -5.52    0.01 0.67   -7.03   -5.92   -5.48   -5.06   -4.33
#> log_lik[36]     -9.96    0.02 1.20  -12.57  -10.73   -9.86   -9.13   -7.86
#> log_lik[37]     -7.60    0.01 0.79   -9.38   -8.09   -7.54   -7.05   -6.23
#> log_lik[38]     -9.01    0.01 0.86  -10.70   -9.58   -9.00   -8.44   -7.32
#> log_lik[39]     -7.24    0.02 1.11   -9.67   -7.94   -7.14   -6.47   -5.30
#> log_lik[40]     -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_lik[41]     -2.21    0.00 0.30   -2.84   -2.40   -2.19   -2.00   -1.67
#> log_lik[42]     -6.01    0.01 0.62   -7.37   -6.40   -5.97   -5.57   -4.92
#> log_lik[43]     -5.64    0.01 0.53   -6.80   -5.97   -5.60   -5.26   -4.71
#> log_lik[44]     -4.42    0.01 0.40   -5.30   -4.67   -4.38   -4.13   -3.72
#> log_lik[45]     -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.80   -2.38
#> lp__          -463.82    0.16 6.15 -476.71 -467.85 -463.46 -459.58 -452.79
#>               n_eff Rhat
#> pi[1]          6475    1
#> pi[2]          6935    1
#> pi[3]          5613    1
#> pi[4]          6530    1
#> theta[1,1,1]   7058    1
#> theta[1,1,2]   6305    1
#> theta[1,1,3]   4638    1
#> theta[1,1,4]   4617    1
#> theta[1,2,1]   5865    1
#> theta[1,2,2]   6133    1
#> theta[1,2,3]   6298    1
#> theta[1,2,4]   5201    1
#> theta[1,3,1]   6084    1
#> theta[1,3,2]   3725    1
#> theta[1,3,3]   4753    1
#> theta[1,3,4]   5376    1
#> theta[1,4,1]   6288    1
#> theta[1,4,2]   6609    1
#> theta[1,4,3]   4201    1
#> theta[1,4,4]   5292    1
#> theta[2,1,1]   6325    1
#> theta[2,1,2]   6351    1
#> theta[2,1,3]   4193    1
#> theta[2,1,4]   4568    1
#> theta[2,2,1]   5881    1
#> theta[2,2,2]   6799    1
#> theta[2,2,3]   7377    1
#> theta[2,2,4]   6253    1
#> theta[2,3,1]   6012    1
#> theta[2,3,2]   5939    1
#> theta[2,3,3]   5871    1
#> theta[2,3,4]   5689    1
#> theta[2,4,1]   6073    1
#> theta[2,4,2]   7100    1
#> theta[2,4,3]   5528    1
#> theta[2,4,4]   6297    1
#> theta[3,1,1]   7271    1
#> theta[3,1,2]   4612    1
#> theta[3,1,3]   4658    1
#> theta[3,1,4]   4345    1
#> theta[3,2,1]   5737    1
#> theta[3,2,2]   6823    1
#> theta[3,2,3]   5811    1
#> theta[3,2,4]   6106    1
#> theta[3,3,1]   5651    1
#> theta[3,3,2]   6255    1
#> theta[3,3,3]   6134    1
#> theta[3,3,4]   7091    1
#> theta[3,4,1]   6534    1
#> theta[3,4,2]   6437    1
#> theta[3,4,3]   5641    1
#> theta[3,4,4]   6083    1
#> theta[4,1,1]   6693    1
#> theta[4,1,2]   5014    1
#> theta[4,1,3]   5036    1
#> theta[4,1,4]   4397    1
#> theta[4,2,1]   6807    1
#> theta[4,2,2]   6265    1
#> theta[4,2,3]   6308    1
#> theta[4,2,4]   5763    1
#> theta[4,3,1]   6387    1
#> theta[4,3,2]   6511    1
#> theta[4,3,3]   6516    1
#> theta[4,3,4]   5343    1
#> theta[4,4,1]   7036    1
#> theta[4,4,2]   5981    1
#> theta[4,4,3]   6202    1
#> theta[4,4,4]   6669    1
#> theta[5,1,1]   6574    1
#> theta[5,1,2]   5286    1
#> theta[5,1,3]   4800    1
#> theta[5,1,4]   4756    1
#> theta[5,2,1]   6856    1
#> theta[5,2,2]   7347    1
#> theta[5,2,3]   5723    1
#> theta[5,2,4]   6257    1
#> theta[5,3,1]   6411    1
#> theta[5,3,2]   5767    1
#> theta[5,3,3]   5793    1
#> theta[5,3,4]   5862    1
#> theta[5,4,1]   6394    1
#> theta[5,4,2]   5452    1
#> theta[5,4,3]   5455    1
#> theta[5,4,4]   5364    1
#> log_p_z[1,1]   6768    1
#> log_p_z[1,2]   3266    1
#> log_p_z[1,3]   3003    1
#> log_p_z[1,4]   2868    1
#> log_p_z[2,1]   4180    1
#> log_p_z[2,2]   4376    1
#> log_p_z[2,3]   3755    1
#> log_p_z[2,4]   2741    1
#> log_p_z[3,1]   3830    1
#> log_p_z[3,2]   3067    1
#> log_p_z[3,3]   3083    1
#> log_p_z[3,4]   2882    1
#> log_p_z[4,1]   4831    1
#> log_p_z[4,2]   3425    1
#> log_p_z[4,3]   2785    1
#> log_p_z[4,4]   3771    1
#> log_p_z[5,1]   3925    1
#> log_p_z[5,2]   6137    1
#> log_p_z[5,3]   2885    1
#> log_p_z[5,4]   3906    1
#> log_p_z[6,1]   4304    1
#> log_p_z[6,2]   5540    1
#> log_p_z[6,3]   2984    1
#> log_p_z[6,4]   4037    1
#> log_p_z[7,1]   5693    1
#> log_p_z[7,2]   3225    1
#> log_p_z[7,3]   2843    1
#> log_p_z[7,4]   2804    1
#> log_p_z[8,1]   4201    1
#> log_p_z[8,2]   4446    1
#> log_p_z[8,3]   6277    1
#> log_p_z[8,4]   2914    1
#> log_p_z[9,1]   3881    1
#> log_p_z[9,2]   4805    1
#> log_p_z[9,3]   2911    1
#> log_p_z[9,4]   3675    1
#> log_p_z[10,1]  3492    1
#> log_p_z[10,2]  4906    1
#> log_p_z[10,3]  2942    1
#> log_p_z[10,4]  3506    1
#> log_p_z[11,1]  4477    1
#> log_p_z[11,2]  3950    1
#> log_p_z[11,3]  2839    1
#> log_p_z[11,4]  6103    1
#> log_p_z[12,1]  4403    1
#> log_p_z[12,2]  3841    1
#> log_p_z[12,3]  2842    1
#> log_p_z[12,4]  4070    1
#> log_p_z[13,1]  6768    1
#> log_p_z[13,2]  3266    1
#> log_p_z[13,3]  3003    1
#> log_p_z[13,4]  2868    1
#> log_p_z[14,1]  4071    1
#> log_p_z[14,2]  4708    1
#> log_p_z[14,3]  2838    1
#> log_p_z[14,4]  3977    1
#> log_p_z[15,1]  5483    1
#> log_p_z[15,2]  3110    1
#> log_p_z[15,3]  2810    1
#> log_p_z[15,4]  2799    1
#> log_p_z[16,1]  5894    1
#> log_p_z[16,2]  3407    1
#> log_p_z[16,3]  3002    1
#> log_p_z[16,4]  2881    1
#> log_p_z[17,1]  6768    1
#> log_p_z[17,2]  3266    1
#> log_p_z[17,3]  3003    1
#> log_p_z[17,4]  2868    1
#> log_p_z[18,1]  6768    1
#> log_p_z[18,2]  3266    1
#> log_p_z[18,3]  3003    1
#> log_p_z[18,4]  2868    1
#> log_p_z[19,1]  3715    1
#> log_p_z[19,2]  6269    1
#> log_p_z[19,3]  2785    1
#> log_p_z[19,4]  3693    1
#> log_p_z[20,1]  4260    1
#> log_p_z[20,2]  4896    1
#> log_p_z[20,3]  2976    1
#> log_p_z[20,4]  3866    1
#> log_p_z[21,1]  3852    1
#> log_p_z[21,2]  6774    1
#> log_p_z[21,3]  2933    1
#> log_p_z[21,4]  3746    1
#> log_p_z[22,1]  3715    1
#> log_p_z[22,2]  6269    1
#> log_p_z[22,3]  2785    1
#> log_p_z[22,4]  3693    1
#> log_p_z[23,1]  3925    1
#> log_p_z[23,2]  6137    1
#> log_p_z[23,3]  2885    1
#> log_p_z[23,4]  3906    1
#> log_p_z[24,1]  3657    1
#> log_p_z[24,2]  4544    1
#> log_p_z[24,3]  3001    1
#> log_p_z[24,4]  3224    1
#> log_p_z[25,1]  6768    1
#> log_p_z[25,2]  3266    1
#> log_p_z[25,3]  3003    1
#> log_p_z[25,4]  2868    1
#> log_p_z[26,1]  6768    1
#> log_p_z[26,2]  3266    1
#> log_p_z[26,3]  3003    1
#> log_p_z[26,4]  2868    1
#> log_p_z[27,1]  3627    1
#> log_p_z[27,2]  6200    1
#> log_p_z[27,3]  2997    1
#> log_p_z[27,4]  3636    1
#> log_p_z[28,1]  6768    1
#> log_p_z[28,2]  3266    1
#> log_p_z[28,3]  3003    1
#> log_p_z[28,4]  2868    1
#> log_p_z[29,1]  6768    1
#> log_p_z[29,2]  3266    1
#> log_p_z[29,3]  3003    1
#> log_p_z[29,4]  2868    1
#> log_p_z[30,1]  4461    1
#> log_p_z[30,2]  3120    1
#> log_p_z[30,3]  2817    1
#> log_p_z[30,4]  3094    1
#> log_p_z[31,1]  6768    1
#> log_p_z[31,2]  3266    1
#> log_p_z[31,3]  3003    1
#> log_p_z[31,4]  2868    1
#> log_p_z[32,1]  4169    1
#> log_p_z[32,2]  5066    1
#> log_p_z[32,3]  4433    1
#> log_p_z[32,4]  2920    1
#> log_p_z[33,1]  6768    1
#> log_p_z[33,2]  3266    1
#> log_p_z[33,3]  3003    1
#> log_p_z[33,4]  2868    1
#> log_p_z[34,1]  3852    1
#> log_p_z[34,2]  6774    1
#> log_p_z[34,3]  2933    1
#> log_p_z[34,4]  3746    1
#> log_p_z[35,1]  4012    1
#> log_p_z[35,2]  5468    1
#> log_p_z[35,3]  2826    1
#> log_p_z[35,4]  4063    1
#> log_p_z[36,1]  3649    1
#> log_p_z[36,2]  3898    1
#> log_p_z[36,3]  2321    1
#> log_p_z[36,4]  2617    1
#> log_p_z[37,1]  3728    1
#> log_p_z[37,2]  5013    1
#> log_p_z[37,3]  2647    1
#> log_p_z[37,4]  3451    1
#> log_p_z[38,1]  3998    1
#> log_p_z[38,2]  4337    1
#> log_p_z[38,3]  2807    1
#> log_p_z[38,4]  3876    1
#> log_p_z[39,1]  4168    1
#> log_p_z[39,2]  4526    1
#> log_p_z[39,3]  5322    1
#> log_p_z[39,4]  2856    1
#> log_p_z[40,1]  6768    1
#> log_p_z[40,2]  3266    1
#> log_p_z[40,3]  3003    1
#> log_p_z[40,4]  2868    1
#> log_p_z[41,1]  6768    1
#> log_p_z[41,2]  3266    1
#> log_p_z[41,3]  3003    1
#> log_p_z[41,4]  2868    1
#> log_p_z[42,1]  5693    1
#> log_p_z[42,2]  3225    1
#> log_p_z[42,3]  2843    1
#> log_p_z[42,4]  2804    1
#> log_p_z[43,1]  3627    1
#> log_p_z[43,2]  6200    1
#> log_p_z[43,3]  2997    1
#> log_p_z[43,4]  3636    1
#> log_p_z[44,1]  5483    1
#> log_p_z[44,2]  3110    1
#> log_p_z[44,3]  2810    1
#> log_p_z[44,4]  2799    1
#> log_p_z[45,1]  3852    1
#> log_p_z[45,2]  6774    1
#> log_p_z[45,3]  2933    1
#> log_p_z[45,4]  3746    1
#> log_lik[1]     6768    1
#> log_lik[2]     4401    1
#> log_lik[3]     5006    1
#> log_lik[4]     3847    1
#> log_lik[5]     6138    1
#> log_lik[6]     5542    1
#> log_lik[7]     5697    1
#> log_lik[8]     6275    1
#> log_lik[9]     4805    1
#> log_lik[10]    4922    1
#> log_lik[11]    6103    1
#> log_lik[12]    5705    1
#> log_lik[13]    6768    1
#> log_lik[14]    4711    1
#> log_lik[15]    5483    1
#> log_lik[16]    5894    1
#> log_lik[17]    6768    1
#> log_lik[18]    6768    1
#> log_lik[19]    6269    1
#> log_lik[20]    4899    1
#> log_lik[21]    6774    1
#> log_lik[22]    6269    1
#> log_lik[23]    6138    1
#> log_lik[24]    4545    1
#> log_lik[25]    6768    1
#> log_lik[26]    6768    1
#> log_lik[27]    6201    1
#> log_lik[28]    6768    1
#> log_lik[29]    6768    1
#> log_lik[30]    4480    1
#> log_lik[31]    6768    1
#> log_lik[32]    4436    1
#> log_lik[33]    6768    1
#> log_lik[34]    6774    1
#> log_lik[35]    5519    1
#> log_lik[36]    4936    1
#> log_lik[37]    5017    1
#> log_lik[38]    4386    1
#> log_lik[39]    5318    1
#> log_lik[40]    6768    1
#> log_lik[41]    6768    1
#> log_lik[42]    5697    1
#> log_lik[43]    6201    1
#> log_lik[44]    5483    1
#> log_lik[45]    6774    1
#> lp__           1425    1
#> 
#> Samples were drawn using NUTS(diag_e) at Mon Sep 11 15:04:26 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }