Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.06    0.25    0.33    0.38    0.42    0.51
#> pi[2]            0.40    0.00 0.07    0.28    0.36    0.40    0.45    0.53
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.90    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.85    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.00    0.01    0.02    0.05
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.07    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.29
#> theta[1,3,3]     0.79    0.00 0.09    0.59    0.73    0.80    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.21
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.30
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.18    0.40
#> theta[1,4,4]     0.69    0.00 0.14    0.40    0.60    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.58    0.70    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.21    0.34
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.58    0.65    0.75
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.25    0.30    0.37    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[2,3,1]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.28
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[2,3,3]     0.62    0.00 0.13    0.35    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.12    0.04    0.12    0.20    0.28    0.47
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.39
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[2,4,3]     0.12    0.00 0.11    0.00    0.04    0.09    0.17    0.39
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.54    0.66    0.76    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.05    0.01    0.05    0.08    0.12    0.22
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.74    0.79    0.88
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.08    0.00    0.03    0.06    0.11    0.28
#> theta[3,3,2]     0.18    0.00 0.10    0.03    0.10    0.16    0.24    0.42
#> theta[3,3,3]     0.51    0.00 0.14    0.25    0.41    0.51    0.61    0.77
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.30    0.49
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[3,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.17    0.36
#> theta[3,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.41
#> theta[3,4,4]     0.63    0.00 0.15    0.31    0.52    0.64    0.74    0.90
#> theta[4,1,1]     0.83    0.00 0.07    0.66    0.78    0.83    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[4,2,2]     0.72    0.00 0.09    0.54    0.67    0.73    0.79    0.88
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.08    0.13    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.26
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.60    0.70    0.79    0.90
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.38
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.36
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.36
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,4]     0.65    0.00 0.15    0.35    0.55    0.66    0.76    0.90
#> theta[5,1,1]     0.86    0.00 0.07    0.72    0.82    0.87    0.91    0.96
#> theta[5,1,2]     0.06    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[5,2,1]     0.15    0.00 0.06    0.04    0.10    0.14    0.19    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.61    0.70    0.84
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.14    0.21    0.38
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.43
#> theta[5,4,4]     0.63    0.00 0.16    0.31    0.52    0.64    0.75    0.90
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_p_z[1,2]   -20.02    0.03 2.31  -25.09  -21.42  -19.83  -18.38  -16.09
#> log_p_z[1,3]   -24.81    0.09 4.43  -34.63  -27.42  -24.31  -21.65  -17.74
#> log_p_z[1,4]   -21.98    0.08 4.37  -32.16  -24.58  -21.36  -18.90  -15.00
#> log_p_z[2,1]   -29.68    0.07 4.44  -39.87  -32.27  -29.12  -26.54  -22.63
#> log_p_z[2,2]   -21.08    0.04 2.41  -26.38  -22.53  -20.89  -19.40  -16.84
#> log_p_z[2,3]    -7.65    0.02 1.24  -10.34   -8.36   -7.54   -6.80   -5.61
#> log_p_z[2,4]   -16.33    0.08 4.17  -25.72  -18.68  -15.74  -13.40   -9.70
#> log_p_z[3,1]   -11.58    0.03 1.65  -15.25  -12.56  -11.37  -10.38   -8.94
#> log_p_z[3,2]   -10.82    0.02 1.50  -14.22  -11.67  -10.64   -9.76   -8.38
#> log_p_z[3,3]   -23.34    0.07 3.56  -31.70  -25.49  -22.98  -20.91  -17.37
#> log_p_z[3,4]   -21.90    0.06 3.60  -29.77  -24.07  -21.57  -19.34  -15.73
#> log_p_z[4,1]   -14.64    0.03 1.87  -18.72  -15.80  -14.43  -13.33  -11.39
#> log_p_z[4,2]    -7.57    0.01 0.95   -9.67   -8.16   -7.49   -6.91   -5.96
#> log_p_z[4,3]   -20.86    0.09 4.60  -31.41  -23.39  -20.24  -17.64  -13.80
#> log_p_z[4,4]   -21.90    0.08 4.41  -32.31  -24.47  -21.32  -18.80  -14.76
#> log_p_z[5,1]   -21.39    0.04 2.56  -26.99  -23.01  -21.23  -19.62  -16.86
#> log_p_z[5,2]    -3.72    0.01 0.43   -4.66   -3.98   -3.68   -3.43   -2.94
#> log_p_z[5,3]   -18.78    0.09 4.57  -29.36  -21.17  -18.16  -15.58  -12.03
#> log_p_z[5,4]   -21.92    0.08 4.40  -32.59  -24.44  -21.36  -18.84  -14.75
#> log_p_z[6,1]   -21.39    0.04 2.53  -26.92  -22.97  -21.19  -19.61  -17.05
#> log_p_z[6,2]    -5.47    0.01 0.69   -7.00   -5.90   -5.42   -4.99   -4.25
#> log_p_z[6,3]   -17.57    0.09 4.42  -28.11  -19.84  -16.87  -14.48  -11.21
#> log_p_z[6,4]   -21.64    0.08 4.33  -31.79  -24.10  -21.12  -18.60  -14.70
#> log_p_z[7,1]    -6.00    0.01 0.63   -7.35   -6.40   -5.94   -5.56   -4.89
#> log_p_z[7,2]   -14.96    0.03 1.71  -18.70  -16.01  -14.82  -13.74  -12.05
#> log_p_z[7,3]   -24.34    0.07 3.66  -32.66  -26.55  -23.97  -21.73  -18.21
#> log_p_z[7,4]   -21.92    0.07 3.56  -29.84  -24.07  -21.66  -19.35  -15.94
#> log_p_z[8,1]   -29.69    0.07 4.38  -39.64  -32.23  -29.17  -26.61  -22.68
#> log_p_z[8,2]   -18.33    0.03 2.12  -22.99  -19.65  -18.15  -16.82  -14.62
#> log_p_z[8,3]    -5.72    0.01 0.84   -7.50   -6.25   -5.69   -5.14   -4.24
#> log_p_z[8,4]   -18.57    0.08 4.32  -28.51  -21.04  -18.00  -15.50  -11.91
#> log_p_z[9,1]   -19.94    0.04 2.29  -24.86  -21.37  -19.80  -18.29  -16.03
#> log_p_z[9,2]    -4.53    0.01 0.58   -5.76   -4.89   -4.48   -4.12   -3.50
#> log_p_z[9,3]   -19.72    0.09 4.68  -30.44  -22.14  -19.09  -16.42  -12.64
#> log_p_z[9,4]   -21.67    0.07 4.32  -31.55  -24.22  -21.17  -18.63  -14.63
#> log_p_z[10,1]  -22.20    0.04 2.40  -27.54  -23.68  -22.01  -20.48  -18.02
#> log_p_z[10,2]   -7.14    0.01 0.72   -8.68   -7.60   -7.08   -6.63   -5.91
#> log_p_z[10,3]  -16.86    0.07 3.41  -24.56  -18.82  -16.44  -14.45  -11.58
#> log_p_z[10,4]  -21.34    0.06 3.54  -29.06  -23.55  -21.06  -18.80  -15.28
#> log_p_z[11,1]  -29.75    0.07 4.34  -39.53  -32.35  -29.36  -26.67  -22.61
#> log_p_z[11,2]  -29.36    0.07 4.30  -39.31  -31.91  -28.89  -26.35  -22.41
#> log_p_z[11,3]  -18.18    0.07 3.32  -26.13  -19.95  -17.74  -15.92  -13.05
#> log_p_z[11,4]   -5.82    0.01 0.97   -7.88   -6.44   -5.76   -5.12   -4.14
#> log_p_z[12,1]  -22.91    0.04 2.59  -28.43  -24.54  -22.70  -21.00  -18.43
#> log_p_z[12,2]   -9.60    0.02 1.31  -12.65  -10.28   -9.44   -8.72   -7.49
#> log_p_z[12,3]  -15.23    0.09 4.30  -25.29  -17.40  -14.59  -12.27   -8.93
#> log_p_z[12,4]  -19.28    0.08 4.21  -29.35  -21.70  -18.67  -16.21  -12.80
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_p_z[13,2]  -20.02    0.03 2.31  -25.09  -21.42  -19.83  -18.38  -16.09
#> log_p_z[13,3]  -24.81    0.09 4.43  -34.63  -27.42  -24.31  -21.65  -17.74
#> log_p_z[13,4]  -21.98    0.08 4.37  -32.16  -24.58  -21.36  -18.90  -15.00
#> log_p_z[14,1]  -18.98    0.04 2.39  -24.16  -20.46  -18.83  -17.27  -14.85
#> log_p_z[14,2]   -6.26    0.01 0.88   -8.20   -6.80   -6.17   -5.64   -4.77
#> log_p_z[14,3]  -18.81    0.09 4.56  -29.46  -21.27  -18.10  -15.56  -11.99
#> log_p_z[14,4]  -21.94    0.08 4.39  -32.16  -24.48  -21.39  -18.78  -14.93
#> log_p_z[15,1]   -4.41    0.00 0.40   -5.27   -4.66   -4.38   -4.11   -3.71
#> log_p_z[15,2]  -17.26    0.03 1.85  -21.24  -18.43  -17.14  -15.95  -14.04
#> log_p_z[15,3]  -24.34    0.07 3.65  -32.67  -26.52  -24.01  -21.73  -18.21
#> log_p_z[15,4]  -21.97    0.07 3.58  -29.78  -24.14  -21.65  -19.41  -15.94
#> log_p_z[16,1]   -3.79    0.01 0.54   -4.98   -4.12   -3.75   -3.42   -2.84
#> log_p_z[16,2]  -17.72    0.03 2.20  -22.53  -19.03  -17.53  -16.17  -13.98
#> log_p_z[16,3]  -24.81    0.08 4.43  -34.82  -27.49  -24.31  -21.70  -17.66
#> log_p_z[16,4]  -21.93    0.08 4.36  -32.15  -24.34  -21.38  -18.82  -15.00
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_p_z[17,2]  -20.02    0.03 2.31  -25.09  -21.42  -19.83  -18.38  -16.09
#> log_p_z[17,3]  -24.81    0.09 4.43  -34.63  -27.42  -24.31  -21.65  -17.74
#> log_p_z[17,4]  -21.98    0.08 4.37  -32.16  -24.58  -21.36  -18.90  -15.00
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_p_z[18,2]  -20.02    0.03 2.31  -25.09  -21.42  -19.83  -18.38  -16.09
#> log_p_z[18,3]  -24.81    0.09 4.43  -34.63  -27.42  -24.31  -21.65  -17.74
#> log_p_z[18,4]  -21.98    0.08 4.37  -32.16  -24.58  -21.36  -18.90  -15.00
#> log_p_z[19,1]  -16.38    0.03 1.99  -20.65  -17.64  -16.21  -14.95  -12.96
#> log_p_z[19,2]   -4.65    0.01 0.58   -5.86   -5.02   -4.61   -4.23   -3.63
#> log_p_z[19,3]  -22.24    0.09 4.77  -33.23  -24.83  -21.56  -18.92  -14.80
#> log_p_z[19,4]  -21.89    0.07 4.34  -31.86  -24.42  -21.36  -18.81  -14.82
#> log_p_z[20,1]  -17.97    0.04 2.26  -22.87  -19.39  -17.75  -16.36  -14.10
#> log_p_z[20,2]   -7.10    0.01 0.97   -9.20   -7.70   -7.02   -6.41   -5.44
#> log_p_z[20,3]  -20.06    0.09 4.52  -30.85  -22.53  -19.43  -16.88  -13.25
#> log_p_z[20,4]  -21.66    0.07 4.29  -31.86  -24.15  -21.10  -18.64  -14.73
#> log_p_z[21,1]  -19.56    0.04 2.37  -24.60  -21.07  -19.36  -17.84  -15.52
#> log_p_z[21,2]   -3.05    0.00 0.36   -3.79   -3.28   -3.03   -2.80   -2.38
#> log_p_z[21,3]  -21.27    0.09 4.73  -32.26  -23.79  -20.61  -17.96  -14.12
#> log_p_z[21,4]  -21.89    0.07 4.37  -32.08  -24.49  -21.37  -18.80  -14.83
#> log_p_z[22,1]  -16.38    0.03 1.99  -20.65  -17.64  -16.21  -14.95  -12.96
#> log_p_z[22,2]   -4.65    0.01 0.58   -5.86   -5.02   -4.61   -4.23   -3.63
#> log_p_z[22,3]  -22.24    0.09 4.77  -33.23  -24.83  -21.56  -18.92  -14.80
#> log_p_z[22,4]  -21.89    0.07 4.34  -31.86  -24.42  -21.36  -18.81  -14.82
#> log_p_z[23,1]  -21.39    0.04 2.56  -26.99  -23.01  -21.23  -19.62  -16.86
#> log_p_z[23,2]   -3.72    0.01 0.43   -4.66   -3.98   -3.68   -3.43   -2.94
#> log_p_z[23,3]  -18.78    0.09 4.57  -29.36  -21.17  -18.16  -15.58  -12.03
#> log_p_z[23,4]  -21.92    0.08 4.40  -32.59  -24.44  -21.36  -18.84  -14.75
#> log_p_z[24,1]  -17.35    0.03 2.15  -22.00  -18.69  -17.19  -15.81  -13.65
#> log_p_z[24,2]   -5.80    0.01 0.64   -7.28   -6.18   -5.73   -5.36   -4.75
#> log_p_z[24,3]  -21.75    0.07 3.69  -30.19  -23.84  -21.38  -19.16  -15.80
#> log_p_z[24,4]  -21.90    0.06 3.62  -29.81  -24.16  -21.58  -19.30  -15.74
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_p_z[25,2]  -20.02    0.03 2.31  -25.09  -21.42  -19.83  -18.38  -16.09
#> log_p_z[25,3]  -24.81    0.09 4.43  -34.63  -27.42  -24.31  -21.65  -17.74
#> log_p_z[25,4]  -21.98    0.08 4.37  -32.16  -24.58  -21.36  -18.90  -15.00
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_p_z[26,2]  -20.02    0.03 2.31  -25.09  -21.42  -19.83  -18.38  -16.09
#> log_p_z[26,3]  -24.81    0.09 4.43  -34.63  -27.42  -24.31  -21.65  -17.74
#> log_p_z[26,4]  -21.98    0.08 4.37  -32.16  -24.58  -21.36  -18.90  -15.00
#> log_p_z[27,1]  -21.82    0.04 2.44  -27.11  -23.39  -21.61  -20.04  -17.69
#> log_p_z[27,2]   -5.66    0.01 0.54   -6.84   -6.00   -5.62   -5.28   -4.72
#> log_p_z[27,3]  -18.41    0.07 3.49  -26.40  -20.41  -17.97  -15.95  -12.92
#> log_p_z[27,4]  -21.56    0.06 3.55  -29.41  -23.77  -21.21  -19.03  -15.49
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_p_z[28,2]  -20.02    0.03 2.31  -25.09  -21.42  -19.83  -18.38  -16.09
#> log_p_z[28,3]  -24.81    0.09 4.43  -34.63  -27.42  -24.31  -21.65  -17.74
#> log_p_z[28,4]  -21.98    0.08 4.37  -32.16  -24.58  -21.36  -18.90  -15.00
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_p_z[29,2]  -20.02    0.03 2.31  -25.09  -21.42  -19.83  -18.38  -16.09
#> log_p_z[29,3]  -24.81    0.09 4.43  -34.63  -27.42  -24.31  -21.65  -17.74
#> log_p_z[29,4]  -21.98    0.08 4.37  -32.16  -24.58  -21.36  -18.90  -15.00
#> log_p_z[30,1]   -6.82    0.01 0.86   -8.84   -7.30   -6.71   -6.23   -5.38
#> log_p_z[30,2]  -14.72    0.03 1.72  -18.53  -15.79  -14.57  -13.51  -11.73
#> log_p_z[30,3]  -24.31    0.07 3.68  -32.75  -26.47  -23.98  -21.69  -18.15
#> log_p_z[30,4]  -21.94    0.06 3.59  -29.94  -24.17  -21.64  -19.41  -15.81
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_p_z[31,2]  -20.02    0.03 2.31  -25.09  -21.42  -19.83  -18.38  -16.09
#> log_p_z[31,3]  -24.81    0.09 4.43  -34.63  -27.42  -24.31  -21.65  -17.74
#> log_p_z[31,4]  -21.98    0.08 4.37  -32.16  -24.58  -21.36  -18.90  -15.00
#> log_p_z[32,1]  -29.67    0.07 4.37  -39.59  -32.21  -29.15  -26.55  -22.69
#> log_p_z[32,2]  -14.85    0.02 1.76  -18.84  -15.90  -14.69  -13.63  -11.88
#> log_p_z[32,3]   -6.06    0.01 0.91   -7.96   -6.63   -5.99   -5.41   -4.51
#> log_p_z[32,4]  -20.73    0.08 4.45  -30.98  -23.30  -20.21  -17.60  -13.67
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_p_z[33,2]  -20.02    0.03 2.31  -25.09  -21.42  -19.83  -18.38  -16.09
#> log_p_z[33,3]  -24.81    0.09 4.43  -34.63  -27.42  -24.31  -21.65  -17.74
#> log_p_z[33,4]  -21.98    0.08 4.37  -32.16  -24.58  -21.36  -18.90  -15.00
#> log_p_z[34,1]  -19.56    0.04 2.37  -24.60  -21.07  -19.36  -17.84  -15.52
#> log_p_z[34,2]   -3.05    0.00 0.36   -3.79   -3.28   -3.03   -2.80   -2.38
#> log_p_z[34,3]  -21.27    0.09 4.73  -32.26  -23.79  -20.61  -17.96  -14.12
#> log_p_z[34,4]  -21.89    0.07 4.37  -32.08  -24.49  -21.37  -18.80  -14.83
#> log_p_z[35,1]  -22.50    0.04 2.64  -28.29  -24.16  -22.28  -20.64  -17.88
#> log_p_z[35,2]   -5.53    0.01 0.70   -7.01   -5.97   -5.46   -5.03   -4.33
#> log_p_z[35,3]  -16.19    0.09 4.42  -26.60  -18.54  -15.60  -12.97   -9.81
#> log_p_z[35,4]  -21.94    0.08 4.37  -32.07  -24.43  -21.34  -18.85  -15.06
#> log_p_z[36,1]  -29.70    0.06 3.64  -37.63  -31.89  -29.39  -27.14  -23.64
#> log_p_z[36,2]  -21.93    0.04 2.26  -26.97  -23.34  -21.80  -20.34  -17.99
#> log_p_z[36,3]  -10.50    0.04 1.71  -14.51  -11.38  -10.22   -9.33   -7.95
#> log_p_z[36,4]  -13.94    0.06 3.19  -21.04  -15.89  -13.63  -11.63   -8.83
#> log_p_z[37,1]  -18.47    0.04 2.23  -23.33  -19.84  -18.22  -16.90  -14.70
#> log_p_z[37,2]   -7.62    0.01 0.81   -9.40   -8.12   -7.54   -7.05   -6.21
#> log_p_z[37,3]  -19.16    0.07 3.48  -27.13  -21.04  -18.75  -16.65  -13.66
#> log_p_z[37,4]  -21.92    0.06 3.60  -29.80  -24.14  -21.58  -19.34  -15.89
#> log_p_z[38,1]  -25.14    0.04 2.76  -31.26  -26.85  -24.87  -23.16  -20.44
#> log_p_z[38,2]   -9.62    0.01 1.01  -11.82  -10.26   -9.54   -8.89   -7.91
#> log_p_z[38,3]  -11.78    0.06 2.97  -18.76  -13.41  -11.30   -9.62   -7.53
#> log_p_z[38,4]  -21.39    0.06 3.58  -29.11  -23.52  -21.02  -18.82  -15.43
#> log_p_z[39,1]  -29.31    0.07 4.32  -38.98  -31.90  -28.80  -26.28  -22.29
#> log_p_z[39,2]  -16.85    0.03 2.02  -21.34  -18.08  -16.66  -15.43  -13.33
#> log_p_z[39,3]   -7.27    0.02 1.09   -9.66   -7.96   -7.20   -6.51   -5.32
#> log_p_z[39,4]  -18.79    0.08 4.21  -28.50  -21.10  -18.24  -15.85  -12.05
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_p_z[40,2]  -20.02    0.03 2.31  -25.09  -21.42  -19.83  -18.38  -16.09
#> log_p_z[40,3]  -24.81    0.09 4.43  -34.63  -27.42  -24.31  -21.65  -17.74
#> log_p_z[40,4]  -21.98    0.08 4.37  -32.16  -24.58  -21.36  -18.90  -15.00
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_p_z[41,2]  -20.02    0.03 2.31  -25.09  -21.42  -19.83  -18.38  -16.09
#> log_p_z[41,3]  -24.81    0.09 4.43  -34.63  -27.42  -24.31  -21.65  -17.74
#> log_p_z[41,4]  -21.98    0.08 4.37  -32.16  -24.58  -21.36  -18.90  -15.00
#> log_p_z[42,1]   -6.00    0.01 0.63   -7.35   -6.40   -5.94   -5.56   -4.89
#> log_p_z[42,2]  -14.96    0.03 1.71  -18.70  -16.01  -14.82  -13.74  -12.05
#> log_p_z[42,3]  -24.34    0.07 3.66  -32.66  -26.55  -23.97  -21.73  -18.21
#> log_p_z[42,4]  -21.92    0.07 3.56  -29.84  -24.07  -21.66  -19.35  -15.94
#> log_p_z[43,1]  -21.82    0.04 2.44  -27.11  -23.39  -21.61  -20.04  -17.69
#> log_p_z[43,2]   -5.66    0.01 0.54   -6.84   -6.00   -5.62   -5.28   -4.72
#> log_p_z[43,3]  -18.41    0.07 3.49  -26.40  -20.41  -17.97  -15.95  -12.92
#> log_p_z[43,4]  -21.56    0.06 3.55  -29.41  -23.77  -21.21  -19.03  -15.49
#> log_p_z[44,1]   -4.41    0.00 0.40   -5.27   -4.66   -4.38   -4.11   -3.71
#> log_p_z[44,2]  -17.26    0.03 1.85  -21.24  -18.43  -17.14  -15.95  -14.04
#> log_p_z[44,3]  -24.34    0.07 3.65  -32.67  -26.52  -24.01  -21.73  -18.21
#> log_p_z[44,4]  -21.97    0.07 3.58  -29.78  -24.14  -21.65  -19.41  -15.94
#> log_p_z[45,1]  -19.56    0.04 2.37  -24.60  -21.07  -19.36  -17.84  -15.52
#> log_p_z[45,2]   -3.05    0.00 0.36   -3.79   -3.28   -3.03   -2.80   -2.38
#> log_p_z[45,3]  -21.27    0.09 4.73  -32.26  -23.79  -20.61  -17.96  -14.12
#> log_p_z[45,4]  -21.89    0.07 4.37  -32.08  -24.49  -21.37  -18.80  -14.83
#> log_lik[1]      -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_lik[2]      -7.60    0.02 1.13  -10.03   -8.33   -7.53   -6.79   -5.61
#> log_lik[3]      -9.93    0.01 0.91  -11.82  -10.50   -9.91   -9.31   -8.23
#> log_lik[4]      -7.57    0.01 0.94   -9.63   -8.16   -7.49   -6.91   -5.96
#> log_lik[5]      -3.72    0.01 0.43   -4.66   -3.98   -3.68   -3.43   -2.94
#> log_lik[6]      -5.47    0.01 0.69   -6.98   -5.90   -5.42   -4.99   -4.25
#> log_lik[7]      -5.99    0.01 0.63   -7.35   -6.40   -5.94   -5.56   -4.89
#> log_lik[8]      -5.72    0.01 0.84   -7.50   -6.24   -5.69   -5.14   -4.24
#> log_lik[9]      -4.53    0.01 0.58   -5.76   -4.89   -4.48   -4.12   -3.50
#> log_lik[10]     -7.14    0.01 0.72   -8.67   -7.59   -7.08   -6.63   -5.90
#> log_lik[11]     -5.82    0.01 0.96   -7.88   -6.44   -5.76   -5.12   -4.14
#> log_lik[12]     -9.34    0.01 1.04  -11.54   -9.98   -9.28   -8.62   -7.46
#> log_lik[13]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_lik[14]     -6.26    0.01 0.88   -8.20   -6.80   -6.17   -5.64   -4.77
#> log_lik[15]     -4.41    0.00 0.40   -5.27   -4.66   -4.38   -4.11   -3.71
#> log_lik[16]     -3.79    0.01 0.54   -4.98   -4.12   -3.75   -3.42   -2.84
#> log_lik[17]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_lik[18]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_lik[19]     -4.65    0.01 0.58   -5.86   -5.02   -4.61   -4.23   -3.63
#> log_lik[20]     -7.10    0.01 0.97   -9.20   -7.70   -7.02   -6.41   -5.44
#> log_lik[21]     -3.05    0.00 0.36   -3.79   -3.28   -3.03   -2.80   -2.38
#> log_lik[22]     -4.65    0.01 0.58   -5.86   -5.02   -4.61   -4.23   -3.63
#> log_lik[23]     -3.72    0.01 0.43   -4.66   -3.98   -3.68   -3.43   -2.94
#> log_lik[24]     -5.80    0.01 0.64   -7.28   -6.17   -5.73   -5.36   -4.75
#> log_lik[25]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_lik[26]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_lik[27]     -5.66    0.01 0.54   -6.84   -6.00   -5.62   -5.28   -4.72
#> log_lik[28]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_lik[29]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_lik[30]     -6.81    0.01 0.86   -8.82   -7.30   -6.71   -6.23   -5.38
#> log_lik[31]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_lik[32]     -6.06    0.01 0.91   -7.96   -6.63   -5.99   -5.41   -4.51
#> log_lik[33]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_lik[34]     -3.05    0.00 0.36   -3.79   -3.28   -3.03   -2.80   -2.38
#> log_lik[35]     -5.53    0.01 0.70   -7.00   -5.97   -5.46   -5.03   -4.33
#> log_lik[36]     -9.93    0.02 1.18  -12.46  -10.66   -9.84   -9.11   -7.84
#> log_lik[37]     -7.62    0.01 0.81   -9.40   -8.12   -7.54   -7.05   -6.21
#> log_lik[38]     -9.00    0.01 0.86  -10.76   -9.55   -8.97   -8.43   -7.37
#> log_lik[39]     -7.27    0.02 1.09   -9.61   -7.96   -7.19   -6.51   -5.32
#> log_lik[40]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_lik[41]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.68
#> log_lik[42]     -5.99    0.01 0.63   -7.35   -6.40   -5.94   -5.56   -4.89
#> log_lik[43]     -5.66    0.01 0.54   -6.84   -6.00   -5.62   -5.28   -4.72
#> log_lik[44]     -4.41    0.00 0.40   -5.27   -4.66   -4.38   -4.11   -3.71
#> log_lik[45]     -3.05    0.00 0.36   -3.79   -3.28   -3.03   -2.80   -2.38
#> lp__          -463.82    0.18 6.28 -477.22 -467.87 -463.34 -459.46 -452.60
#>               n_eff Rhat
#> pi[1]          7779    1
#> pi[2]          7027    1
#> pi[3]          6739    1
#> pi[4]          5072    1
#> theta[1,1,1]   7271    1
#> theta[1,1,2]   6918    1
#> theta[1,1,3]   4079    1
#> theta[1,1,4]   4618    1
#> theta[1,2,1]   5026    1
#> theta[1,2,2]   6235    1
#> theta[1,2,3]   6365    1
#> theta[1,2,4]   4923    1
#> theta[1,3,1]   6251    1
#> theta[1,3,2]   3952    1
#> theta[1,3,3]   4902    1
#> theta[1,3,4]   5722    1
#> theta[1,4,1]   5130    1
#> theta[1,4,2]   6635    1
#> theta[1,4,3]   3390    1
#> theta[1,4,4]   4411    1
#> theta[2,1,1]   6381    1
#> theta[2,1,2]   6086    1
#> theta[2,1,3]   4842    1
#> theta[2,1,4]   4794    1
#> theta[2,2,1]   7001    1
#> theta[2,2,2]   7280    1
#> theta[2,2,3]   6908    1
#> theta[2,2,4]   5737    1
#> theta[2,3,1]   5395    1
#> theta[2,3,2]   6689    1
#> theta[2,3,3]   6174    1
#> theta[2,3,4]   5012    1
#> theta[2,4,1]   5786    1
#> theta[2,4,2]   6183    1
#> theta[2,4,3]   5968    1
#> theta[2,4,4]   5832    1
#> theta[3,1,1]   9686    1
#> theta[3,1,2]   5105    1
#> theta[3,1,3]   5077    1
#> theta[3,1,4]   4156    1
#> theta[3,2,1]   5871    1
#> theta[3,2,2]   6990    1
#> theta[3,2,3]   5473    1
#> theta[3,2,4]   5599    1
#> theta[3,3,1]   6321    1
#> theta[3,3,2]   5841    1
#> theta[3,3,3]   6488    1
#> theta[3,3,4]   6267    1
#> theta[3,4,1]   5808    1
#> theta[3,4,2]   7568    1
#> theta[3,4,3]   4128    1
#> theta[3,4,4]   5775    1
#> theta[4,1,1]   6909    1
#> theta[4,1,2]   5561    1
#> theta[4,1,3]   4405    1
#> theta[4,1,4]   4502    1
#> theta[4,2,1]   6932    1
#> theta[4,2,2]   5884    1
#> theta[4,2,3]   7032    1
#> theta[4,2,4]   5581    1
#> theta[4,3,1]   6484    1
#> theta[4,3,2]   5196    1
#> theta[4,3,3]   6080    1
#> theta[4,3,4]   5747    1
#> theta[4,4,1]   5808    1
#> theta[4,4,2]   6374    1
#> theta[4,4,3]   6399    1
#> theta[4,4,4]   6325    1
#> theta[5,1,1]   6105    1
#> theta[5,1,2]   4920    1
#> theta[5,1,3]   4319    1
#> theta[5,1,4]   4544    1
#> theta[5,2,1]   7823    1
#> theta[5,2,2]   6429    1
#> theta[5,2,3]   6027    1
#> theta[5,2,4]   5581    1
#> theta[5,3,1]   6323    1
#> theta[5,3,2]   6850    1
#> theta[5,3,3]   6363    1
#> theta[5,3,4]   7031    1
#> theta[5,4,1]   7392    1
#> theta[5,4,2]   5857    1
#> theta[5,4,3]   5640    1
#> theta[5,4,4]   6547    1
#> log_p_z[1,1]   6773    1
#> log_p_z[1,2]   4386    1
#> log_p_z[1,3]   2519    1
#> log_p_z[1,4]   3093    1
#> log_p_z[2,1]   4444    1
#> log_p_z[2,2]   4584    1
#> log_p_z[2,3]   2895    1
#> log_p_z[2,4]   2788    1
#> log_p_z[3,1]   4272    1
#> log_p_z[3,2]   3747    1
#> log_p_z[3,3]   2869    1
#> log_p_z[3,4]   3321    1
#> log_p_z[4,1]   5091    1
#> log_p_z[4,2]   4661    1
#> log_p_z[4,3]   2488    1
#> log_p_z[4,4]   3056    1
#> log_p_z[5,1]   4048    1
#> log_p_z[5,2]   6627    1
#> log_p_z[5,3]   2430    1
#> log_p_z[5,4]   3213    1
#> log_p_z[6,1]   3965    1
#> log_p_z[6,2]   4927    1
#> log_p_z[6,3]   2420    1
#> log_p_z[6,4]   3202    1
#> log_p_z[7,1]   5388    1
#> log_p_z[7,2]   4446    1
#> log_p_z[7,3]   3063    1
#> log_p_z[7,4]   2891    1
#> log_p_z[8,1]   3914    1
#> log_p_z[8,2]   4688    1
#> log_p_z[8,3]   6542    1
#> log_p_z[8,4]   2868    1
#> log_p_z[9,1]   4175    1
#> log_p_z[9,2]   5400    1
#> log_p_z[9,3]   2502    1
#> log_p_z[9,4]   3525    1
#> log_p_z[10,1]  4385    1
#> log_p_z[10,2]  5610    1
#> log_p_z[10,3]  2600    1
#> log_p_z[10,4]  3350    1
#> log_p_z[11,1]  3801    1
#> log_p_z[11,2]  3464    1
#> log_p_z[11,3]  2559    1
#> log_p_z[11,4]  5350    1
#> log_p_z[12,1]  3876    1
#> log_p_z[12,2]  3253    1
#> log_p_z[12,3]  2418    1
#> log_p_z[12,4]  3121    1
#> log_p_z[13,1]  6773    1
#> log_p_z[13,2]  4386    1
#> log_p_z[13,3]  2519    1
#> log_p_z[13,4]  3093    1
#> log_p_z[14,1]  3898    1
#> log_p_z[14,2]  5133    1
#> log_p_z[14,3]  2549    1
#> log_p_z[14,4]  2940    1
#> log_p_z[15,1]  6456    1
#> log_p_z[15,2]  4590    1
#> log_p_z[15,3]  2932    1
#> log_p_z[15,4]  2779    1
#> log_p_z[16,1]  6409    1
#> log_p_z[16,2]  4339    1
#> log_p_z[16,3]  2751    1
#> log_p_z[16,4]  3081    1
#> log_p_z[17,1]  6773    1
#> log_p_z[17,2]  4386    1
#> log_p_z[17,3]  2519    1
#> log_p_z[17,4]  3093    1
#> log_p_z[18,1]  6773    1
#> log_p_z[18,2]  4386    1
#> log_p_z[18,3]  2519    1
#> log_p_z[18,4]  3093    1
#> log_p_z[19,1]  4732    1
#> log_p_z[19,2]  6314    1
#> log_p_z[19,3]  2597    1
#> log_p_z[19,4]  3399    1
#> log_p_z[20,1]  4041    1
#> log_p_z[20,2]  4867    1
#> log_p_z[20,3]  2480    1
#> log_p_z[20,4]  3289    1
#> log_p_z[21,1]  4024    1
#> log_p_z[21,2]  7244    1
#> log_p_z[21,3]  2548    1
#> log_p_z[21,4]  3524    1
#> log_p_z[22,1]  4732    1
#> log_p_z[22,2]  6314    1
#> log_p_z[22,3]  2597    1
#> log_p_z[22,4]  3399    1
#> log_p_z[23,1]  4048    1
#> log_p_z[23,2]  6627    1
#> log_p_z[23,3]  2430    1
#> log_p_z[23,4]  3213    1
#> log_p_z[24,1]  3847    1
#> log_p_z[24,2]  4259    1
#> log_p_z[24,3]  2663    1
#> log_p_z[24,4]  3388    1
#> log_p_z[25,1]  6773    1
#> log_p_z[25,2]  4386    1
#> log_p_z[25,3]  2519    1
#> log_p_z[25,4]  3093    1
#> log_p_z[26,1]  6773    1
#> log_p_z[26,2]  4386    1
#> log_p_z[26,3]  2519    1
#> log_p_z[26,4]  3093    1
#> log_p_z[27,1]  4035    1
#> log_p_z[27,2]  6131    1
#> log_p_z[27,3]  2676    1
#> log_p_z[27,4]  3586    1
#> log_p_z[28,1]  6773    1
#> log_p_z[28,2]  4386    1
#> log_p_z[28,3]  2519    1
#> log_p_z[28,4]  3093    1
#> log_p_z[29,1]  6773    1
#> log_p_z[29,2]  4386    1
#> log_p_z[29,3]  2519    1
#> log_p_z[29,4]  3093    1
#> log_p_z[30,1]  5636    1
#> log_p_z[30,2]  4022    1
#> log_p_z[30,3]  2752    1
#> log_p_z[30,4]  3116    1
#> log_p_z[31,1]  6773    1
#> log_p_z[31,2]  4386    1
#> log_p_z[31,3]  2519    1
#> log_p_z[31,4]  3093    1
#> log_p_z[32,1]  4338    1
#> log_p_z[32,2]  5662    1
#> log_p_z[32,3]  5114    1
#> log_p_z[32,4]  2817    1
#> log_p_z[33,1]  6773    1
#> log_p_z[33,2]  4386    1
#> log_p_z[33,3]  2519    1
#> log_p_z[33,4]  3093    1
#> log_p_z[34,1]  4024    1
#> log_p_z[34,2]  7244    1
#> log_p_z[34,3]  2548    1
#> log_p_z[34,4]  3524    1
#> log_p_z[35,1]  3865    1
#> log_p_z[35,2]  5088    1
#> log_p_z[35,3]  2464    1
#> log_p_z[35,4]  3229    1
#> log_p_z[36,1]  3725    1
#> log_p_z[36,2]  3181    1
#> log_p_z[36,3]  2204    1
#> log_p_z[36,4]  2469    1
#> log_p_z[37,1]  3655    1
#> log_p_z[37,2]  4913    1
#> log_p_z[37,3]  2749    1
#> log_p_z[37,4]  3361    1
#> log_p_z[38,1]  4329    1
#> log_p_z[38,2]  4951    1
#> log_p_z[38,3]  2418    1
#> log_p_z[38,4]  3087    1
#> log_p_z[39,1]  3446    1
#> log_p_z[39,2]  4678    1
#> log_p_z[39,3]  4988    1
#> log_p_z[39,4]  3135    1
#> log_p_z[40,1]  6773    1
#> log_p_z[40,2]  4386    1
#> log_p_z[40,3]  2519    1
#> log_p_z[40,4]  3093    1
#> log_p_z[41,1]  6773    1
#> log_p_z[41,2]  4386    1
#> log_p_z[41,3]  2519    1
#> log_p_z[41,4]  3093    1
#> log_p_z[42,1]  5388    1
#> log_p_z[42,2]  4446    1
#> log_p_z[42,3]  3063    1
#> log_p_z[42,4]  2891    1
#> log_p_z[43,1]  4035    1
#> log_p_z[43,2]  6131    1
#> log_p_z[43,3]  2676    1
#> log_p_z[43,4]  3586    1
#> log_p_z[44,1]  6456    1
#> log_p_z[44,2]  4590    1
#> log_p_z[44,3]  2932    1
#> log_p_z[44,4]  2779    1
#> log_p_z[45,1]  4024    1
#> log_p_z[45,2]  7244    1
#> log_p_z[45,3]  2548    1
#> log_p_z[45,4]  3524    1
#> log_lik[1]     6773    1
#> log_lik[2]     4753    1
#> log_lik[3]     4053    1
#> log_lik[4]     4716    1
#> log_lik[5]     6627    1
#> log_lik[6]     4932    1
#> log_lik[7]     5390    1
#> log_lik[8]     6549    1
#> log_lik[9]     5400    1
#> log_lik[10]    5617    1
#> log_lik[11]    5350    1
#> log_lik[12]    5002    1
#> log_lik[13]    6773    1
#> log_lik[14]    5140    1
#> log_lik[15]    6456    1
#> log_lik[16]    6410    1
#> log_lik[17]    6773    1
#> log_lik[18]    6773    1
#> log_lik[19]    6314    1
#> log_lik[20]    4868    1
#> log_lik[21]    7244    1
#> log_lik[22]    6314    1
#> log_lik[23]    6627    1
#> log_lik[24]    4260    1
#> log_lik[25]    6773    1
#> log_lik[26]    6773    1
#> log_lik[27]    6130    1
#> log_lik[28]    6773    1
#> log_lik[29]    6773    1
#> log_lik[30]    5635    1
#> log_lik[31]    6773    1
#> log_lik[32]    5117    1
#> log_lik[33]    6773    1
#> log_lik[34]    7244    1
#> log_lik[35]    5110    1
#> log_lik[36]    4557    1
#> log_lik[37]    4920    1
#> log_lik[38]    4362    1
#> log_lik[39]    4990    1
#> log_lik[40]    6773    1
#> log_lik[41]    6773    1
#> log_lik[42]    5390    1
#> log_lik[43]    6130    1
#> log_lik[44]    6456    1
#> log_lik[45]    7244    1
#> lp__           1276    1
#> 
#> Samples were drawn using NUTS(diag_e) at Wed Jul 23 09:45:30 2025.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }