Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.07    0.25    0.33    0.37    0.42    0.51
#> pi[2]            0.41    0.00 0.07    0.27    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.09    0.15
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.90    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.76    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.01    0.00    0.00    0.01    0.02    0.05
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.07    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.59    0.73    0.81    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.29
#> theta[1,4,2]     0.10    0.00 0.09    0.00    0.03    0.07    0.14    0.32
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.19    0.40
#> theta[1,4,4]     0.69    0.00 0.14    0.39    0.60    0.70    0.79    0.91
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.76    0.81    0.90
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.12    0.16    0.21    0.32
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[2,2,2]     0.58    0.00 0.09    0.39    0.52    0.59    0.65    0.76
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.25    0.30    0.37    0.49
#> theta[2,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[2,3,3]     0.62    0.00 0.13    0.35    0.53    0.63    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.11    0.04    0.13    0.20    0.29    0.47
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.55    0.66    0.76    0.91
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.93    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.09    0.00 0.05    0.02    0.05    0.09    0.13    0.23
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[3,3,2]     0.18    0.00 0.10    0.03    0.10    0.16    0.24    0.42
#> theta[3,3,3]     0.51    0.00 0.14    0.25    0.41    0.51    0.61    0.77
#> theta[3,3,4]     0.23    0.00 0.11    0.05    0.15    0.22    0.31    0.48
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.35
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[3,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.19    0.42
#> theta[3,4,4]     0.63    0.00 0.15    0.32    0.52    0.64    0.75    0.89
#> theta[4,1,1]     0.83    0.00 0.08    0.65    0.78    0.84    0.88    0.94
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.09    0.54    0.67    0.73    0.79    0.88
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.08    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.10    0.20
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.60    0.70    0.79    0.91
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.40
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[4,4,4]     0.65    0.00 0.15    0.32    0.55    0.66    0.76    0.91
#> theta[5,1,1]     0.86    0.00 0.07    0.70    0.82    0.87    0.92    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.07    0.04    0.10    0.14    0.19    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.46    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.33
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.22    0.40
#> theta[5,3,3]     0.60    0.00 0.13    0.32    0.51    0.61    0.70    0.84
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.14    0.21    0.38
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.42
#> theta[5,4,4]     0.63    0.00 0.15    0.32    0.52    0.63    0.74    0.89
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_p_z[1,2]   -20.01    0.04 2.33  -25.08  -21.41  -19.82  -18.40  -16.14
#> log_p_z[1,3]   -24.80    0.07 4.26  -34.28  -27.41  -24.37  -21.69  -17.82
#> log_p_z[1,4]   -21.83    0.07 4.29  -31.31  -24.26  -21.34  -18.74  -14.89
#> log_p_z[2,1]   -29.71    0.06 4.40  -39.86  -32.18  -29.28  -26.67  -22.51
#> log_p_z[2,2]   -21.05    0.03 2.36  -26.09  -22.53  -20.90  -19.39  -16.93
#> log_p_z[2,3]    -7.62    0.02 1.22  -10.37   -8.32   -7.51   -6.79   -5.61
#> log_p_z[2,4]   -16.49    0.08 4.38  -26.65  -18.93  -15.92  -13.36   -9.70
#> log_p_z[3,1]   -11.62    0.02 1.63  -15.26  -12.58  -11.44  -10.46   -8.99
#> log_p_z[3,2]   -10.79    0.02 1.48  -14.12  -11.65  -10.61   -9.73   -8.41
#> log_p_z[3,3]   -23.27    0.05 3.34  -30.54  -25.37  -23.01  -20.87  -17.65
#> log_p_z[3,4]   -21.84    0.05 3.51  -29.61  -23.94  -21.50  -19.36  -15.93
#> log_p_z[4,1]   -14.61    0.02 1.88  -18.72  -15.75  -14.47  -13.34  -11.28
#> log_p_z[4,2]    -7.55    0.01 0.91   -9.49   -8.09   -7.50   -6.91   -5.95
#> log_p_z[4,3]   -20.70    0.07 4.31  -30.50  -23.23  -20.23  -17.63  -13.86
#> log_p_z[4,4]   -21.89    0.06 4.39  -31.86  -24.43  -21.34  -18.76  -14.88
#> log_p_z[5,1]   -21.45    0.03 2.55  -27.01  -23.00  -21.27  -19.60  -16.98
#> log_p_z[5,2]    -3.71    0.00 0.44   -4.62   -4.00   -3.69   -3.41   -2.94
#> log_p_z[5,3]   -18.71    0.07 4.23  -28.23  -21.19  -18.18  -15.66  -12.08
#> log_p_z[5,4]   -21.90    0.06 4.33  -32.02  -24.33  -21.42  -18.87  -14.86
#> log_p_z[6,1]   -21.46    0.04 2.56  -27.13  -23.04  -21.24  -19.66  -16.97
#> log_p_z[6,2]    -5.47    0.01 0.68   -6.95   -5.90   -5.42   -5.01   -4.24
#> log_p_z[6,3]   -17.47    0.06 4.06  -26.83  -19.79  -16.94  -14.59  -11.27
#> log_p_z[6,4]   -21.71    0.06 4.33  -31.61  -24.28  -21.12  -18.64  -14.47
#> log_p_z[7,1]    -6.00    0.01 0.61   -7.34   -6.38   -5.96   -5.58   -4.95
#> log_p_z[7,2]   -14.92    0.03 1.69  -18.57  -15.95  -14.76  -13.73  -12.05
#> log_p_z[7,3]   -24.26    0.06 3.47  -32.05  -26.33  -24.04  -21.80  -18.27
#> log_p_z[7,4]   -21.83    0.06 3.57  -29.66  -24.04  -21.49  -19.28  -15.90
#> log_p_z[8,1]   -29.64    0.06 4.30  -39.40  -32.11  -29.19  -26.63  -22.42
#> log_p_z[8,2]   -18.32    0.03 2.12  -23.05  -19.65  -18.17  -16.82  -14.73
#> log_p_z[8,3]    -5.73    0.01 0.84   -7.62   -6.24   -5.66   -5.17   -4.28
#> log_p_z[8,4]   -18.64    0.08 4.48  -29.11  -21.09  -18.06  -15.51  -11.62
#> log_p_z[9,1]   -19.98    0.03 2.26  -24.74  -21.42  -19.81  -18.37  -16.00
#> log_p_z[9,2]    -4.52    0.01 0.58   -5.79   -4.88   -4.48   -4.12   -3.51
#> log_p_z[9,3]   -19.65    0.07 4.27  -29.24  -22.23  -19.19  -16.55  -12.76
#> log_p_z[9,4]   -21.67    0.07 4.26  -31.55  -24.14  -21.23  -18.58  -14.82
#> log_p_z[10,1]  -22.23    0.04 2.37  -27.36  -23.73  -22.02  -20.51  -18.07
#> log_p_z[10,2]   -7.14    0.01 0.73   -8.70   -7.59   -7.09   -6.64   -5.88
#> log_p_z[10,3]  -16.82    0.05 3.14  -23.53  -18.79  -16.50  -14.53  -11.64
#> log_p_z[10,4]  -21.38    0.06 3.53  -29.25  -23.51  -21.11  -18.88  -15.44
#> log_p_z[11,1]  -29.77    0.06 4.22  -39.21  -32.26  -29.38  -26.75  -22.81
#> log_p_z[11,2]  -29.26    0.07 4.23  -38.74  -31.71  -28.69  -26.23  -22.47
#> log_p_z[11,3]  -18.24    0.06 3.38  -26.36  -20.01  -17.74  -15.87  -13.17
#> log_p_z[11,4]   -5.84    0.01 0.96   -7.91   -6.45   -5.76   -5.16   -4.15
#> log_p_z[12,1]  -22.94    0.04 2.64  -28.57  -24.59  -22.76  -21.13  -18.28
#> log_p_z[12,2]   -9.57    0.02 1.29  -12.48  -10.31   -9.45   -8.68   -7.40
#> log_p_z[12,3]  -15.17    0.07 4.01  -24.52  -17.47  -14.68  -12.32   -8.89
#> log_p_z[12,4]  -19.29    0.06 4.20  -29.05  -21.74  -18.71  -16.31  -12.60
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_p_z[13,2]  -20.01    0.04 2.33  -25.08  -21.41  -19.82  -18.40  -16.14
#> log_p_z[13,3]  -24.80    0.07 4.26  -34.28  -27.41  -24.37  -21.69  -17.82
#> log_p_z[13,4]  -21.83    0.07 4.29  -31.31  -24.26  -21.34  -18.74  -14.89
#> log_p_z[14,1]  -19.05    0.03 2.38  -24.29  -20.50  -18.87  -17.30  -15.00
#> log_p_z[14,2]   -6.23    0.01 0.88   -8.20   -6.74   -6.15   -5.60   -4.76
#> log_p_z[14,3]  -18.77    0.07 4.24  -28.46  -21.18  -18.22  -15.75  -12.18
#> log_p_z[14,4]  -21.88    0.07 4.37  -32.16  -24.35  -21.37  -18.76  -15.04
#> log_p_z[15,1]   -4.41    0.00 0.41   -5.31   -4.66   -4.38   -4.13   -3.71
#> log_p_z[15,2]  -17.25    0.03 1.87  -21.38  -18.39  -17.09  -15.93  -14.05
#> log_p_z[15,3]  -24.30    0.06 3.50  -32.14  -26.44  -23.99  -21.78  -18.43
#> log_p_z[15,4]  -21.84    0.06 3.60  -29.78  -24.08  -21.52  -19.28  -15.65
#> log_p_z[16,1]   -3.80    0.01 0.55   -4.99   -4.14   -3.76   -3.41   -2.84
#> log_p_z[16,2]  -17.68    0.03 2.19  -22.46  -18.99  -17.46  -16.17  -14.01
#> log_p_z[16,3]  -24.76    0.07 4.22  -34.25  -27.23  -24.36  -21.73  -17.78
#> log_p_z[16,4]  -21.82    0.07 4.28  -31.43  -24.19  -21.39  -18.86  -14.89
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_p_z[17,2]  -20.01    0.04 2.33  -25.08  -21.41  -19.82  -18.40  -16.14
#> log_p_z[17,3]  -24.80    0.07 4.26  -34.28  -27.41  -24.37  -21.69  -17.82
#> log_p_z[17,4]  -21.83    0.07 4.29  -31.31  -24.26  -21.34  -18.74  -14.89
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_p_z[18,2]  -20.01    0.04 2.33  -25.08  -21.41  -19.82  -18.40  -16.14
#> log_p_z[18,3]  -24.80    0.07 4.26  -34.28  -27.41  -24.37  -21.69  -17.82
#> log_p_z[18,4]  -21.83    0.07 4.29  -31.31  -24.26  -21.34  -18.74  -14.89
#> log_p_z[19,1]  -16.42    0.03 1.91  -20.36  -17.64  -16.30  -15.06  -13.06
#> log_p_z[19,2]   -4.65    0.01 0.58   -5.92   -5.01   -4.61   -4.24   -3.61
#> log_p_z[19,3]  -22.12    0.07 4.41  -31.94  -24.78  -21.61  -18.92  -15.01
#> log_p_z[19,4]  -21.91    0.07 4.33  -31.95  -24.43  -21.35  -18.82  -14.91
#> log_p_z[20,1]  -18.07    0.03 2.26  -22.98  -19.50  -17.91  -16.50  -14.15
#> log_p_z[20,2]   -7.13    0.01 0.98   -9.30   -7.74   -7.05   -6.44   -5.46
#> log_p_z[20,3]  -19.99    0.06 4.16  -29.57  -22.44  -19.42  -17.03  -13.46
#> log_p_z[20,4]  -21.71    0.06 4.31  -31.46  -24.16  -21.16  -18.68  -14.77
#> log_p_z[21,1]  -19.64    0.03 2.33  -24.67  -21.08  -19.49  -18.00  -15.57
#> log_p_z[21,2]   -3.04    0.00 0.35   -3.78   -3.27   -3.03   -2.80   -2.38
#> log_p_z[21,3]  -21.19    0.07 4.31  -30.82  -23.79  -20.72  -18.04  -14.17
#> log_p_z[21,4]  -21.89    0.07 4.30  -31.84  -24.42  -21.38  -18.85  -14.95
#> log_p_z[22,1]  -16.42    0.03 1.91  -20.36  -17.64  -16.30  -15.06  -13.06
#> log_p_z[22,2]   -4.65    0.01 0.58   -5.92   -5.01   -4.61   -4.24   -3.61
#> log_p_z[22,3]  -22.12    0.07 4.41  -31.94  -24.78  -21.61  -18.92  -15.01
#> log_p_z[22,4]  -21.91    0.07 4.33  -31.95  -24.43  -21.35  -18.82  -14.91
#> log_p_z[23,1]  -21.45    0.03 2.55  -27.01  -23.00  -21.27  -19.60  -16.98
#> log_p_z[23,2]   -3.71    0.00 0.44   -4.62   -4.00   -3.69   -3.41   -2.94
#> log_p_z[23,3]  -18.71    0.07 4.23  -28.23  -21.19  -18.18  -15.66  -12.08
#> log_p_z[23,4]  -21.90    0.06 4.33  -32.02  -24.33  -21.42  -18.87  -14.86
#> log_p_z[24,1]  -17.44    0.03 2.13  -22.24  -18.79  -17.27  -15.92  -13.78
#> log_p_z[24,2]   -5.80    0.01 0.64   -7.27   -6.17   -5.74   -5.37   -4.77
#> log_p_z[24,3]  -21.70    0.05 3.38  -28.97  -23.85  -21.39  -19.25  -15.96
#> log_p_z[24,4]  -21.88    0.05 3.53  -29.74  -24.01  -21.56  -19.37  -16.01
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_p_z[25,2]  -20.01    0.04 2.33  -25.08  -21.41  -19.82  -18.40  -16.14
#> log_p_z[25,3]  -24.80    0.07 4.26  -34.28  -27.41  -24.37  -21.69  -17.82
#> log_p_z[25,4]  -21.83    0.07 4.29  -31.31  -24.26  -21.34  -18.74  -14.89
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_p_z[26,2]  -20.01    0.04 2.33  -25.08  -21.41  -19.82  -18.40  -16.14
#> log_p_z[26,3]  -24.80    0.07 4.26  -34.28  -27.41  -24.37  -21.69  -17.82
#> log_p_z[26,4]  -21.83    0.07 4.29  -31.31  -24.26  -21.34  -18.74  -14.89
#> log_p_z[27,1]  -21.89    0.03 2.40  -27.12  -23.40  -21.69  -20.23  -17.72
#> log_p_z[27,2]   -5.66    0.01 0.54   -6.82   -6.00   -5.63   -5.27   -4.71
#> log_p_z[27,3]  -18.36    0.05 3.21  -25.37  -20.36  -18.04  -16.03  -12.99
#> log_p_z[27,4]  -21.61    0.05 3.50  -29.54  -23.73  -21.28  -19.15  -15.63
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_p_z[28,2]  -20.01    0.04 2.33  -25.08  -21.41  -19.82  -18.40  -16.14
#> log_p_z[28,3]  -24.80    0.07 4.26  -34.28  -27.41  -24.37  -21.69  -17.82
#> log_p_z[28,4]  -21.83    0.07 4.29  -31.31  -24.26  -21.34  -18.74  -14.89
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_p_z[29,2]  -20.01    0.04 2.33  -25.08  -21.41  -19.82  -18.40  -16.14
#> log_p_z[29,3]  -24.80    0.07 4.26  -34.28  -27.41  -24.37  -21.69  -17.82
#> log_p_z[29,4]  -21.83    0.07 4.29  -31.31  -24.26  -21.34  -18.74  -14.89
#> log_p_z[30,1]   -6.81    0.01 0.83   -8.71   -7.29   -6.72   -6.23   -5.41
#> log_p_z[30,2]  -14.74    0.03 1.74  -18.50  -15.80  -14.57  -13.49  -11.85
#> log_p_z[30,3]  -24.24    0.06 3.53  -31.96  -26.47  -23.95  -21.69  -18.32
#> log_p_z[30,4]  -21.86    0.06 3.55  -29.74  -24.08  -21.57  -19.43  -15.70
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_p_z[31,2]  -20.01    0.04 2.33  -25.08  -21.41  -19.82  -18.40  -16.14
#> log_p_z[31,3]  -24.80    0.07 4.26  -34.28  -27.41  -24.37  -21.69  -17.82
#> log_p_z[31,4]  -21.83    0.07 4.29  -31.31  -24.26  -21.34  -18.74  -14.89
#> log_p_z[32,1]  -29.64    0.06 4.32  -39.66  -32.04  -29.20  -26.70  -22.42
#> log_p_z[32,2]  -14.88    0.02 1.75  -18.58  -15.97  -14.73  -13.66  -11.83
#> log_p_z[32,3]   -6.08    0.01 0.94   -8.07   -6.65   -6.01   -5.44   -4.45
#> log_p_z[32,4]  -20.77    0.08 4.63  -31.45  -23.38  -20.11  -17.52  -13.53
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_p_z[33,2]  -20.01    0.04 2.33  -25.08  -21.41  -19.82  -18.40  -16.14
#> log_p_z[33,3]  -24.80    0.07 4.26  -34.28  -27.41  -24.37  -21.69  -17.82
#> log_p_z[33,4]  -21.83    0.07 4.29  -31.31  -24.26  -21.34  -18.74  -14.89
#> log_p_z[34,1]  -19.64    0.03 2.33  -24.67  -21.08  -19.49  -18.00  -15.57
#> log_p_z[34,2]   -3.04    0.00 0.35   -3.78   -3.27   -3.03   -2.80   -2.38
#> log_p_z[34,3]  -21.19    0.07 4.31  -30.82  -23.79  -20.72  -18.04  -14.17
#> log_p_z[34,4]  -21.89    0.07 4.30  -31.84  -24.42  -21.38  -18.85  -14.95
#> log_p_z[35,1]  -22.56    0.04 2.65  -28.30  -24.26  -22.36  -20.61  -18.02
#> log_p_z[35,2]   -5.54    0.01 0.72   -7.09   -5.97   -5.48   -5.03   -4.30
#> log_p_z[35,3]  -16.10    0.06 4.09  -25.50  -18.43  -15.54  -13.11   -9.93
#> log_p_z[35,4]  -21.86    0.07 4.35  -31.79  -24.31  -21.36  -18.78  -14.83
#> log_p_z[36,1]  -29.74    0.06 3.65  -37.85  -31.94  -29.44  -27.18  -23.55
#> log_p_z[36,2]  -21.90    0.04 2.31  -26.88  -23.35  -21.69  -20.25  -18.02
#> log_p_z[36,3]  -10.54    0.04 1.74  -14.92  -11.45  -10.24   -9.34   -7.94
#> log_p_z[36,4]  -14.06    0.06 3.36  -21.68  -16.04  -13.73  -11.53   -8.78
#> log_p_z[37,1]  -18.56    0.03 2.27  -23.58  -20.01  -18.37  -16.93  -14.68
#> log_p_z[37,2]   -7.63    0.01 0.82   -9.42   -8.13   -7.55   -7.06   -6.26
#> log_p_z[37,3]  -19.09    0.05 3.18  -26.10  -21.12  -18.76  -16.80  -13.75
#> log_p_z[37,4]  -21.84    0.06 3.55  -29.65  -24.06  -21.46  -19.35  -15.86
#> log_p_z[38,1]  -25.14    0.04 2.75  -31.28  -26.83  -24.87  -23.30  -20.32
#> log_p_z[38,2]   -9.64    0.01 1.01  -11.92  -10.25   -9.57   -8.94   -7.90
#> log_p_z[38,3]  -11.73    0.04 2.78  -18.21  -13.32  -11.38   -9.71   -7.54
#> log_p_z[38,4]  -21.34    0.06 3.56  -29.08  -23.55  -20.93  -18.79  -15.31
#> log_p_z[39,1]  -29.30    0.06 4.25  -38.85  -31.68  -28.90  -26.29  -22.32
#> log_p_z[39,2]  -16.84    0.03 2.02  -21.29  -18.05  -16.68  -15.38  -13.45
#> log_p_z[39,3]   -7.27    0.01 1.13   -9.72   -7.98   -7.19   -6.48   -5.30
#> log_p_z[39,4]  -18.86    0.08 4.33  -29.14  -21.17  -18.32  -15.80  -12.18
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_p_z[40,2]  -20.01    0.04 2.33  -25.08  -21.41  -19.82  -18.40  -16.14
#> log_p_z[40,3]  -24.80    0.07 4.26  -34.28  -27.41  -24.37  -21.69  -17.82
#> log_p_z[40,4]  -21.83    0.07 4.29  -31.31  -24.26  -21.34  -18.74  -14.89
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_p_z[41,2]  -20.01    0.04 2.33  -25.08  -21.41  -19.82  -18.40  -16.14
#> log_p_z[41,3]  -24.80    0.07 4.26  -34.28  -27.41  -24.37  -21.69  -17.82
#> log_p_z[41,4]  -21.83    0.07 4.29  -31.31  -24.26  -21.34  -18.74  -14.89
#> log_p_z[42,1]   -6.00    0.01 0.61   -7.34   -6.38   -5.96   -5.58   -4.95
#> log_p_z[42,2]  -14.92    0.03 1.69  -18.57  -15.95  -14.76  -13.73  -12.05
#> log_p_z[42,3]  -24.26    0.06 3.47  -32.05  -26.33  -24.04  -21.80  -18.27
#> log_p_z[42,4]  -21.83    0.06 3.57  -29.66  -24.04  -21.49  -19.28  -15.90
#> log_p_z[43,1]  -21.89    0.03 2.40  -27.12  -23.40  -21.69  -20.23  -17.72
#> log_p_z[43,2]   -5.66    0.01 0.54   -6.82   -6.00   -5.63   -5.27   -4.71
#> log_p_z[43,3]  -18.36    0.05 3.21  -25.37  -20.36  -18.04  -16.03  -12.99
#> log_p_z[43,4]  -21.61    0.05 3.50  -29.54  -23.73  -21.28  -19.15  -15.63
#> log_p_z[44,1]   -4.41    0.00 0.41   -5.31   -4.66   -4.38   -4.13   -3.71
#> log_p_z[44,2]  -17.25    0.03 1.87  -21.38  -18.39  -17.09  -15.93  -14.05
#> log_p_z[44,3]  -24.30    0.06 3.50  -32.14  -26.44  -23.99  -21.78  -18.43
#> log_p_z[44,4]  -21.84    0.06 3.60  -29.78  -24.08  -21.52  -19.28  -15.65
#> log_p_z[45,1]  -19.64    0.03 2.33  -24.67  -21.08  -19.49  -18.00  -15.57
#> log_p_z[45,2]   -3.04    0.00 0.35   -3.78   -3.27   -3.03   -2.80   -2.38
#> log_p_z[45,3]  -21.19    0.07 4.31  -30.82  -23.79  -20.72  -18.04  -14.17
#> log_p_z[45,4]  -21.89    0.07 4.30  -31.84  -24.42  -21.38  -18.85  -14.95
#> log_lik[1]      -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_lik[2]      -7.57    0.02 1.13  -10.04   -8.28   -7.50   -6.79   -5.61
#> log_lik[3]      -9.95    0.01 0.92  -11.85  -10.54   -9.91   -9.33   -8.27
#> log_lik[4]      -7.54    0.01 0.90   -9.45   -8.08   -7.49   -6.91   -5.95
#> log_lik[5]      -3.71    0.00 0.44   -4.62   -4.00   -3.69   -3.41   -2.94
#> log_lik[6]      -5.47    0.01 0.68   -6.93   -5.90   -5.42   -5.01   -4.24
#> log_lik[7]      -6.00    0.01 0.61   -7.34   -6.38   -5.95   -5.58   -4.95
#> log_lik[8]      -5.73    0.01 0.84   -7.61   -6.24   -5.66   -5.17   -4.28
#> log_lik[9]      -4.52    0.01 0.58   -5.79   -4.88   -4.48   -4.12   -3.51
#> log_lik[10]     -7.14    0.01 0.72   -8.69   -7.59   -7.09   -6.64   -5.88
#> log_lik[11]     -5.84    0.01 0.96   -7.91   -6.45   -5.76   -5.16   -4.15
#> log_lik[12]     -9.32    0.01 1.06  -11.56  -10.01   -9.27   -8.58   -7.35
#> log_lik[13]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_lik[14]     -6.23    0.01 0.88   -8.20   -6.74   -6.15   -5.60   -4.76
#> log_lik[15]     -4.41    0.00 0.41   -5.31   -4.66   -4.38   -4.13   -3.71
#> log_lik[16]     -3.80    0.01 0.55   -4.99   -4.14   -3.76   -3.41   -2.84
#> log_lik[17]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_lik[18]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_lik[19]     -4.65    0.01 0.58   -5.92   -5.01   -4.61   -4.24   -3.61
#> log_lik[20]     -7.13    0.01 0.98   -9.30   -7.74   -7.05   -6.44   -5.46
#> log_lik[21]     -3.04    0.00 0.35   -3.78   -3.27   -3.03   -2.80   -2.38
#> log_lik[22]     -4.65    0.01 0.58   -5.92   -5.01   -4.61   -4.24   -3.61
#> log_lik[23]     -3.71    0.00 0.44   -4.62   -4.00   -3.69   -3.41   -2.94
#> log_lik[24]     -5.80    0.01 0.64   -7.27   -6.17   -5.74   -5.37   -4.77
#> log_lik[25]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_lik[26]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_lik[27]     -5.66    0.01 0.54   -6.82   -6.00   -5.63   -5.27   -4.71
#> log_lik[28]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_lik[29]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_lik[30]     -6.81    0.01 0.83   -8.69   -7.28   -6.72   -6.23   -5.41
#> log_lik[31]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_lik[32]     -6.08    0.01 0.94   -8.07   -6.65   -6.01   -5.44   -4.45
#> log_lik[33]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_lik[34]     -3.04    0.00 0.35   -3.78   -3.27   -3.03   -2.80   -2.38
#> log_lik[35]     -5.54    0.01 0.72   -7.08   -5.97   -5.48   -5.03   -4.30
#> log_lik[36]     -9.94    0.02 1.20  -12.51  -10.66   -9.86   -9.11   -7.82
#> log_lik[37]     -7.63    0.01 0.82   -9.42   -8.13   -7.55   -7.06   -6.26
#> log_lik[38]     -9.04    0.01 0.87  -10.77   -9.59   -9.03   -8.46   -7.35
#> log_lik[39]     -7.27    0.01 1.13   -9.71   -7.97   -7.19   -6.48   -5.30
#> log_lik[40]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_lik[41]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -1.99   -1.67
#> log_lik[42]     -6.00    0.01 0.61   -7.34   -6.38   -5.95   -5.58   -4.95
#> log_lik[43]     -5.66    0.01 0.54   -6.82   -6.00   -5.63   -5.27   -4.71
#> log_lik[44]     -4.41    0.00 0.41   -5.31   -4.66   -4.38   -4.13   -3.71
#> log_lik[45]     -3.04    0.00 0.35   -3.78   -3.27   -3.03   -2.80   -2.38
#> lp__          -463.94    0.16 6.10 -476.89 -467.87 -463.67 -459.63 -452.93
#>               n_eff Rhat
#> pi[1]          9139    1
#> pi[2]         10122    1
#> pi[3]          7897    1
#> pi[4]          7274    1
#> theta[1,1,1]   9009    1
#> theta[1,1,2]   7945    1
#> theta[1,1,3]   5197    1
#> theta[1,1,4]   5375    1
#> theta[1,2,1]   7092    1
#> theta[1,2,2]   7752    1
#> theta[1,2,3]   8676    1
#> theta[1,2,4]   6532    1
#> theta[1,3,1]   6143    1
#> theta[1,3,2]   4903    1
#> theta[1,3,3]   5331    1
#> theta[1,3,4]   7341    1
#> theta[1,4,1]   6718    1
#> theta[1,4,2]   7023    1
#> theta[1,4,3]   4001    1
#> theta[1,4,4]   5025    1
#> theta[2,1,1]  11357    1
#> theta[2,1,2]   8416    1
#> theta[2,1,3]   5136    1
#> theta[2,1,4]   4647    1
#> theta[2,2,1]   7092    1
#> theta[2,2,2]   8709    1
#> theta[2,2,3]   8295    1
#> theta[2,2,4]   6467    1
#> theta[2,3,1]   5783    1
#> theta[2,3,2]   6543    1
#> theta[2,3,3]   6817    1
#> theta[2,3,4]   5696    1
#> theta[2,4,1]   6348    1
#> theta[2,4,2]   8403    1
#> theta[2,4,3]   7473    1
#> theta[2,4,4]   6574    1
#> theta[3,1,1]   9475    1
#> theta[3,1,2]   6700    1
#> theta[3,1,3]   4846    1
#> theta[3,1,4]   4734    1
#> theta[3,2,1]   7961    1
#> theta[3,2,2]   9041    1
#> theta[3,2,3]   8225    1
#> theta[3,2,4]   5720    1
#> theta[3,3,1]   6374    1
#> theta[3,3,2]   6443    1
#> theta[3,3,3]   8087    1
#> theta[3,3,4]   8391    1
#> theta[3,4,1]   7273    1
#> theta[3,4,2]   6895    1
#> theta[3,4,3]   5781    1
#> theta[3,4,4]   5947    1
#> theta[4,1,1]   9506    1
#> theta[4,1,2]   5943    1
#> theta[4,1,3]   5239    1
#> theta[4,1,4]   4995    1
#> theta[4,2,1]   8068    1
#> theta[4,2,2]   8086    1
#> theta[4,2,3]   6790    1
#> theta[4,2,4]   6810    1
#> theta[4,3,1]   7416    1
#> theta[4,3,2]   6377    1
#> theta[4,3,3]   6034    1
#> theta[4,3,4]   6072    1
#> theta[4,4,1]   6771    1
#> theta[4,4,2]   6841    1
#> theta[4,4,3]   6816    1
#> theta[4,4,4]   6906    1
#> theta[5,1,1]   7058    1
#> theta[5,1,2]   5587    1
#> theta[5,1,3]   5280    1
#> theta[5,1,4]   4503    1
#> theta[5,2,1]  10079    1
#> theta[5,2,2]   8434    1
#> theta[5,2,3]   8290    1
#> theta[5,2,4]   5921    1
#> theta[5,3,1]   6108    1
#> theta[5,3,2]   8960    1
#> theta[5,3,3]   8233    1
#> theta[5,3,4]   7293    1
#> theta[5,4,1]   7055    1
#> theta[5,4,2]   7168    1
#> theta[5,4,3]   7091    1
#> theta[5,4,4]   6806    1
#> log_p_z[1,1]   9149    1
#> log_p_z[1,2]   4308    1
#> log_p_z[1,3]   3622    1
#> log_p_z[1,4]   3840    1
#> log_p_z[2,1]   5071    1
#> log_p_z[2,2]   4694    1
#> log_p_z[2,3]   3122    1
#> log_p_z[2,4]   2934    1
#> log_p_z[3,1]   4326    1
#> log_p_z[3,2]   4129    1
#> log_p_z[3,3]   3694    1
#> log_p_z[3,4]   4089    1
#> log_p_z[4,1]   5720    1
#> log_p_z[4,2]   5954    1
#> log_p_z[4,3]   4090    1
#> log_p_z[4,4]   4951    1
#> log_p_z[5,1]   5422    1
#> log_p_z[5,2]   8767    1
#> log_p_z[5,3]   4164    1
#> log_p_z[5,4]   4530    1
#> log_p_z[6,1]   5262    1
#> log_p_z[6,2]   7916    1
#> log_p_z[6,3]   4120    1
#> log_p_z[6,4]   4703    1
#> log_p_z[7,1]   6159    1
#> log_p_z[7,2]   4335    1
#> log_p_z[7,3]   3546    1
#> log_p_z[7,4]   4167    1
#> log_p_z[8,1]   5228    1
#> log_p_z[8,2]   5226    1
#> log_p_z[8,3]   6676    1
#> log_p_z[8,4]   3039    1
#> log_p_z[9,1]   4181    1
#> log_p_z[9,2]   7013    1
#> log_p_z[9,3]   4174    1
#> log_p_z[9,4]   4149    1
#> log_p_z[10,1]  4268    1
#> log_p_z[10,2]  6580    1
#> log_p_z[10,3]  4183    1
#> log_p_z[10,4]  3892    1
#> log_p_z[11,1]  5118    1
#> log_p_z[11,2]  3823    1
#> log_p_z[11,3]  3113    1
#> log_p_z[11,4]  6488    1
#> log_p_z[12,1]  5477    1
#> log_p_z[12,2]  5279    1
#> log_p_z[12,3]  3672    1
#> log_p_z[12,4]  4477    1
#> log_p_z[13,1]  9149    1
#> log_p_z[13,2]  4308    1
#> log_p_z[13,3]  3622    1
#> log_p_z[13,4]  3840    1
#> log_p_z[14,1]  5448    1
#> log_p_z[14,2]  5899    1
#> log_p_z[14,3]  3786    1
#> log_p_z[14,4]  4362    1
#> log_p_z[15,1]  8210    1
#> log_p_z[15,2]  4479    1
#> log_p_z[15,3]  3754    1
#> log_p_z[15,4]  4044    1
#> log_p_z[16,1]  7117    1
#> log_p_z[16,2]  4162    1
#> log_p_z[16,3]  3444    1
#> log_p_z[16,4]  4027    1
#> log_p_z[17,1]  9149    1
#> log_p_z[17,2]  4308    1
#> log_p_z[17,3]  3622    1
#> log_p_z[17,4]  3840    1
#> log_p_z[18,1]  9149    1
#> log_p_z[18,2]  4308    1
#> log_p_z[18,3]  3622    1
#> log_p_z[18,4]  3840    1
#> log_p_z[19,1]  4870    1
#> log_p_z[19,2]  7004    1
#> log_p_z[19,3]  4162    1
#> log_p_z[19,4]  4385    1
#> log_p_z[20,1]  5221    1
#> log_p_z[20,2]  5081    1
#> log_p_z[20,3]  4360    1
#> log_p_z[20,4]  4528    1
#> log_p_z[21,1]  4864    1
#> log_p_z[21,2]  9139    1
#> log_p_z[21,3]  4210    1
#> log_p_z[21,4]  4297    1
#> log_p_z[22,1]  4870    1
#> log_p_z[22,2]  7004    1
#> log_p_z[22,3]  4162    1
#> log_p_z[22,4]  4385    1
#> log_p_z[23,1]  5422    1
#> log_p_z[23,2]  8767    1
#> log_p_z[23,3]  4164    1
#> log_p_z[23,4]  4530    1
#> log_p_z[24,1]  4802    1
#> log_p_z[24,2]  4606    1
#> log_p_z[24,3]  4257    1
#> log_p_z[24,4]  4160    1
#> log_p_z[25,1]  9149    1
#> log_p_z[25,2]  4308    1
#> log_p_z[25,3]  3622    1
#> log_p_z[25,4]  3840    1
#> log_p_z[26,1]  9149    1
#> log_p_z[26,2]  4308    1
#> log_p_z[26,3]  3622    1
#> log_p_z[26,4]  3840    1
#> log_p_z[27,1]  4902    1
#> log_p_z[27,2]  7241    1
#> log_p_z[27,3]  4270    1
#> log_p_z[27,4]  4080    1
#> log_p_z[28,1]  9149    1
#> log_p_z[28,2]  4308    1
#> log_p_z[28,3]  3622    1
#> log_p_z[28,4]  3840    1
#> log_p_z[29,1]  9149    1
#> log_p_z[29,2]  4308    1
#> log_p_z[29,3]  3622    1
#> log_p_z[29,4]  3840    1
#> log_p_z[30,1]  5571    1
#> log_p_z[30,2]  4559    1
#> log_p_z[30,3]  3980    1
#> log_p_z[30,4]  4065    1
#> log_p_z[31,1]  9149    1
#> log_p_z[31,2]  4308    1
#> log_p_z[31,3]  3622    1
#> log_p_z[31,4]  3840    1
#> log_p_z[32,1]  5223    1
#> log_p_z[32,2]  5753    1
#> log_p_z[32,3]  5039    1
#> log_p_z[32,4]  2995    1
#> log_p_z[33,1]  9149    1
#> log_p_z[33,2]  4308    1
#> log_p_z[33,3]  3622    1
#> log_p_z[33,4]  3840    1
#> log_p_z[34,1]  4864    1
#> log_p_z[34,2]  9139    1
#> log_p_z[34,3]  4210    1
#> log_p_z[34,4]  4297    1
#> log_p_z[35,1]  5680    1
#> log_p_z[35,2]  6086    1
#> log_p_z[35,3]  3971    1
#> log_p_z[35,4]  4450    1
#> log_p_z[36,1]  3946    1
#> log_p_z[36,2]  4038    1
#> log_p_z[36,3]  2452    1
#> log_p_z[36,4]  2937    1
#> log_p_z[37,1]  5188    1
#> log_p_z[37,2]  4825    1
#> log_p_z[37,3]  4107    1
#> log_p_z[37,4]  3985    1
#> log_p_z[38,1]  5335    1
#> log_p_z[38,2]  5149    1
#> log_p_z[38,3]  3860    1
#> log_p_z[38,4]  3834    1
#> log_p_z[39,1]  5183    1
#> log_p_z[39,2]  5380    1
#> log_p_z[39,3]  5930    1
#> log_p_z[39,4]  3131    1
#> log_p_z[40,1]  9149    1
#> log_p_z[40,2]  4308    1
#> log_p_z[40,3]  3622    1
#> log_p_z[40,4]  3840    1
#> log_p_z[41,1]  9149    1
#> log_p_z[41,2]  4308    1
#> log_p_z[41,3]  3622    1
#> log_p_z[41,4]  3840    1
#> log_p_z[42,1]  6159    1
#> log_p_z[42,2]  4335    1
#> log_p_z[42,3]  3546    1
#> log_p_z[42,4]  4167    1
#> log_p_z[43,1]  4902    1
#> log_p_z[43,2]  7241    1
#> log_p_z[43,3]  4270    1
#> log_p_z[43,4]  4080    1
#> log_p_z[44,1]  8210    1
#> log_p_z[44,2]  4479    1
#> log_p_z[44,3]  3754    1
#> log_p_z[44,4]  4044    1
#> log_p_z[45,1]  4864    1
#> log_p_z[45,2]  9139    1
#> log_p_z[45,3]  4210    1
#> log_p_z[45,4]  4297    1
#> log_lik[1]     9149    1
#> log_lik[2]     4672    1
#> log_lik[3]     5665    1
#> log_lik[4]     5979    1
#> log_lik[5]     8769    1
#> log_lik[6]     7932    1
#> log_lik[7]     6160    1
#> log_lik[8]     6681    1
#> log_lik[9]     7013    1
#> log_lik[10]    6593    1
#> log_lik[11]    6489    1
#> log_lik[12]    5785    1
#> log_lik[13]    9149    1
#> log_lik[14]    5910    1
#> log_lik[15]    8210    1
#> log_lik[16]    7117    1
#> log_lik[17]    9149    1
#> log_lik[18]    9149    1
#> log_lik[19]    7005    1
#> log_lik[20]    5084    1
#> log_lik[21]    9140    1
#> log_lik[22]    7005    1
#> log_lik[23]    8769    1
#> log_lik[24]    4607    1
#> log_lik[25]    9149    1
#> log_lik[26]    9149    1
#> log_lik[27]    7244    1
#> log_lik[28]    9149    1
#> log_lik[29]    9149    1
#> log_lik[30]    5583    1
#> log_lik[31]    9149    1
#> log_lik[32]    5045    1
#> log_lik[33]    9149    1
#> log_lik[34]    9140    1
#> log_lik[35]    6169    1
#> log_lik[36]    5454    1
#> log_lik[37]    4836    1
#> log_lik[38]    4742    1
#> log_lik[39]    6024    1
#> log_lik[40]    9149    1
#> log_lik[41]    9149    1
#> log_lik[42]    6160    1
#> log_lik[43]    7244    1
#> log_lik[44]    8210    1
#> log_lik[45]    9140    1
#> lp__           1507    1
#> 
#> Samples were drawn using NUTS(diag_e) at Tue Feb  7 10:47:03 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }