Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit

A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit
#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.38    0.00 0.07    0.25    0.33    0.37    0.42    0.50
#> pi[2]            0.41    0.00 0.07    0.28    0.36    0.41    0.45    0.54
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.04    0.02    0.05    0.07    0.10    0.15
#> theta[1,1,1]     0.86    0.00 0.04    0.77    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.07    0.10    0.13    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.05    0.75    0.82    0.85    0.89    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.02    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.28
#> theta[1,3,3]     0.79    0.00 0.09    0.58    0.74    0.81    0.86    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.03    0.06    0.11    0.22
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.31
#> theta[1,4,2]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.18    0.39
#> theta[1,4,4]     0.69    0.00 0.14    0.39    0.60    0.70    0.79    0.92
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.33
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.40    0.52    0.58    0.65    0.75
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.25    0.31    0.37    0.49
#> theta[2,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[2,3,3]     0.63    0.00 0.14    0.35    0.53    0.63    0.73    0.87
#> theta[2,3,4]     0.21    0.00 0.11    0.03    0.12    0.20    0.28    0.47
#> theta[2,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,4]     0.65    0.00 0.15    0.33    0.54    0.66    0.76    0.90
#> theta[3,1,1]     0.88    0.00 0.06    0.73    0.84    0.89    0.92    0.97
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.13
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,2,1]     0.10    0.00 0.06    0.01    0.05    0.09    0.13    0.22
#> theta[3,2,2]     0.73    0.00 0.08    0.56    0.68    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.28
#> theta[3,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.12
#> theta[3,3,1]     0.08    0.00 0.08    0.00    0.03    0.06    0.12    0.28
#> theta[3,3,2]     0.18    0.00 0.10    0.03    0.10    0.16    0.24    0.42
#> theta[3,3,3]     0.51    0.00 0.14    0.24    0.41    0.51    0.60    0.77
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.31    0.49
#> theta[3,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.36
#> theta[3,4,3]     0.14    0.00 0.11    0.01    0.05    0.12    0.21    0.42
#> theta[3,4,4]     0.63    0.00 0.15    0.32    0.53    0.63    0.74    0.88
#> theta[4,1,1]     0.83    0.00 0.07    0.67    0.78    0.83    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.12    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.20
#> theta[4,2,2]     0.73    0.00 0.09    0.54    0.67    0.73    0.79    0.87
#> theta[4,2,3]     0.13    0.00 0.07    0.03    0.08    0.12    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[4,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.11    0.28
#> theta[4,3,3]     0.69    0.00 0.13    0.42    0.61    0.70    0.78    0.91
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.40
#> theta[4,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[4,4,4]     0.65    0.00 0.15    0.33    0.55    0.66    0.77    0.91
#> theta[5,1,1]     0.86    0.00 0.07    0.70    0.82    0.87    0.92    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.18
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.06    0.04    0.09    0.13    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.16    0.00 0.07    0.05    0.11    0.16    0.21    0.32
#> theta[5,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[5,3,2]     0.16    0.00 0.10    0.02    0.08    0.14    0.21    0.39
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.61    0.71    0.84
#> theta[5,3,4]     0.15    0.00 0.10    0.02    0.08    0.14    0.21    0.38
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.37
#> theta[5,4,2]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[5,4,3]     0.14    0.00 0.12    0.01    0.05    0.11    0.20    0.43
#> theta[5,4,4]     0.63    0.00 0.15    0.31    0.53    0.64    0.74    0.89
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_p_z[1,2]   -19.99    0.04 2.29  -25.10  -21.42  -19.80  -18.36  -16.17
#> log_p_z[1,3]   -24.69    0.07 4.22  -34.22  -27.14  -24.28  -21.68  -17.66
#> log_p_z[1,4]   -21.86    0.08 4.35  -31.89  -24.37  -21.31  -18.73  -14.86
#> log_p_z[2,1]   -29.70    0.07 4.30  -39.27  -32.18  -29.26  -26.69  -22.48
#> log_p_z[2,2]   -20.98    0.04 2.35  -26.11  -22.37  -20.82  -19.34  -16.91
#> log_p_z[2,3]    -7.65    0.02 1.19  -10.29   -8.38   -7.54   -6.80   -5.66
#> log_p_z[2,4]   -16.33    0.08 4.11  -25.83  -18.75  -15.72  -13.43   -9.78
#> log_p_z[3,1]   -11.61    0.03 1.63  -15.28  -12.60  -11.43  -10.45   -8.99
#> log_p_z[3,2]   -10.76    0.02 1.45  -14.09  -11.62  -10.61   -9.73   -8.39
#> log_p_z[3,3]   -23.34    0.06 3.43  -30.94  -25.45  -23.05  -20.89  -17.49
#> log_p_z[3,4]   -21.81    0.06 3.51  -29.65  -23.90  -21.49  -19.32  -15.88
#> log_p_z[4,1]   -14.61    0.03 1.86  -18.79  -15.71  -14.46  -13.31  -11.46
#> log_p_z[4,2]    -7.56    0.01 0.92   -9.62   -8.09   -7.49   -6.90   -5.97
#> log_p_z[4,3]   -20.64    0.08 4.28  -30.40  -23.00  -20.14  -17.64  -13.75
#> log_p_z[4,4]   -21.81    0.08 4.29  -31.59  -24.18  -21.32  -18.78  -14.91
#> log_p_z[5,1]   -21.42    0.04 2.59  -27.22  -23.00  -21.21  -19.58  -16.98
#> log_p_z[5,2]    -3.71    0.01 0.43   -4.60   -3.98   -3.69   -3.41   -2.92
#> log_p_z[5,3]   -18.66    0.08 4.26  -28.30  -21.04  -18.15  -15.60  -12.10
#> log_p_z[5,4]   -21.82    0.08 4.24  -31.19  -24.21  -21.36  -18.91  -14.79
#> log_p_z[6,1]   -21.41    0.04 2.57  -27.00  -23.02  -21.19  -19.55  -16.96
#> log_p_z[6,2]    -5.47    0.01 0.65   -6.86   -5.88   -5.44   -5.02   -4.33
#> log_p_z[6,3]   -17.45    0.08 4.10  -26.84  -19.66  -16.91  -14.58  -11.16
#> log_p_z[6,4]   -21.58    0.07 4.23  -31.18  -23.98  -21.11  -18.69  -14.64
#> log_p_z[7,1]    -6.00    0.01 0.61   -7.34   -6.38   -5.96   -5.57   -4.91
#> log_p_z[7,2]   -14.92    0.03 1.68  -18.62  -15.96  -14.77  -13.73  -12.09
#> log_p_z[7,3]   -24.17    0.06 3.47  -31.87  -26.28  -23.89  -21.70  -18.25
#> log_p_z[7,4]   -21.85    0.06 3.54  -29.58  -24.04  -21.50  -19.26  -15.92
#> log_p_z[8,1]   -29.68    0.07 4.27  -39.01  -32.17  -29.24  -26.75  -22.53
#> log_p_z[8,2]   -18.27    0.03 2.07  -22.65  -19.50  -18.13  -16.83  -14.67
#> log_p_z[8,3]    -5.72    0.01 0.81   -7.45   -6.24   -5.67   -5.15   -4.26
#> log_p_z[8,4]   -18.50    0.08 4.18  -28.15  -20.98  -17.99  -15.38  -12.04
#> log_p_z[9,1]   -19.97    0.03 2.27  -24.81  -21.37  -19.82  -18.37  -15.99
#> log_p_z[9,2]    -4.52    0.01 0.58   -5.78   -4.88   -4.49   -4.12   -3.54
#> log_p_z[9,3]   -19.62    0.08 4.36  -29.30  -22.12  -19.13  -16.54  -12.76
#> log_p_z[9,4]   -21.57    0.07 4.19  -31.21  -23.90  -21.05  -18.68  -14.77
#> log_p_z[10,1]  -22.22    0.04 2.34  -27.20  -23.69  -22.02  -20.60  -18.12
#> log_p_z[10,2]   -7.13    0.01 0.72   -8.65   -7.59   -7.08   -6.63   -5.87
#> log_p_z[10,3]  -16.82    0.06 3.22  -23.95  -18.71  -16.49  -14.52  -11.61
#> log_p_z[10,4]  -21.25    0.06 3.46  -28.93  -23.31  -20.95  -18.82  -15.52
#> log_p_z[11,1]  -29.68    0.06 4.28  -39.25  -32.21  -29.21  -26.67  -22.53
#> log_p_z[11,2]  -29.20    0.07 4.25  -39.00  -31.59  -28.68  -26.25  -22.41
#> log_p_z[11,3]  -18.35    0.07 3.36  -26.42  -20.15  -17.86  -15.96  -13.18
#> log_p_z[11,4]   -5.84    0.01 0.98   -7.98   -6.43   -5.75   -5.16   -4.16
#> log_p_z[12,1]  -22.92    0.04 2.65  -28.77  -24.53  -22.66  -21.09  -18.34
#> log_p_z[12,2]   -9.60    0.02 1.32  -12.62  -10.31   -9.44   -8.70   -7.52
#> log_p_z[12,3]  -15.08    0.08 4.04  -24.19  -17.20  -14.57  -12.29   -8.91
#> log_p_z[12,4]  -19.17    0.08 4.15  -28.57  -21.51  -18.62  -16.27  -12.50
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_p_z[13,2]  -19.99    0.04 2.29  -25.10  -21.42  -19.80  -18.36  -16.17
#> log_p_z[13,3]  -24.69    0.07 4.22  -34.22  -27.14  -24.28  -21.68  -17.66
#> log_p_z[13,4]  -21.86    0.08 4.35  -31.89  -24.37  -21.31  -18.73  -14.86
#> log_p_z[14,1]  -19.02    0.04 2.44  -24.46  -20.50  -18.84  -17.28  -14.96
#> log_p_z[14,2]   -6.25    0.01 0.87   -8.20   -6.77   -6.17   -5.64   -4.78
#> log_p_z[14,3]  -18.60    0.08 4.21  -28.21  -20.95  -18.04  -15.59  -12.09
#> log_p_z[14,4]  -21.85    0.08 4.30  -31.48  -24.30  -21.48  -18.78  -14.80
#> log_p_z[15,1]   -4.41    0.01 0.40   -5.26   -4.66   -4.38   -4.13   -3.70
#> log_p_z[15,2]  -17.24    0.03 1.84  -21.14  -18.39  -17.10  -15.95  -14.10
#> log_p_z[15,3]  -24.15    0.06 3.48  -31.94  -26.28  -23.78  -21.70  -18.19
#> log_p_z[15,4]  -21.85    0.07 3.60  -30.06  -24.04  -21.48  -19.28  -15.85
#> log_p_z[16,1]   -3.79    0.01 0.54   -4.98   -4.14   -3.76   -3.41   -2.85
#> log_p_z[16,2]  -17.67    0.03 2.16  -22.62  -19.00  -17.47  -16.12  -14.05
#> log_p_z[16,3]  -24.70    0.07 4.20  -34.28  -27.21  -24.30  -21.71  -17.77
#> log_p_z[16,4]  -21.86    0.07 4.31  -31.61  -24.33  -21.38  -18.85  -14.97
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_p_z[17,2]  -19.99    0.04 2.29  -25.10  -21.42  -19.80  -18.36  -16.17
#> log_p_z[17,3]  -24.69    0.07 4.22  -34.22  -27.14  -24.28  -21.68  -17.66
#> log_p_z[17,4]  -21.86    0.08 4.35  -31.89  -24.37  -21.31  -18.73  -14.86
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_p_z[18,2]  -19.99    0.04 2.29  -25.10  -21.42  -19.80  -18.36  -16.17
#> log_p_z[18,3]  -24.69    0.07 4.22  -34.22  -27.14  -24.28  -21.68  -17.66
#> log_p_z[18,4]  -21.86    0.08 4.35  -31.89  -24.37  -21.31  -18.73  -14.86
#> log_p_z[19,1]  -16.39    0.03 1.95  -20.52  -17.61  -16.24  -15.03  -12.98
#> log_p_z[19,2]   -4.66    0.01 0.59   -5.93   -5.03   -4.62   -4.24   -3.64
#> log_p_z[19,3]  -22.06    0.08 4.41  -31.75  -24.60  -21.54  -18.97  -15.04
#> log_p_z[19,4]  -21.82    0.07 4.22  -31.33  -24.18  -21.40  -18.85  -14.84
#> log_p_z[20,1]  -18.01    0.03 2.25  -23.01  -19.43  -17.86  -16.43  -14.10
#> log_p_z[20,2]   -7.13    0.01 0.95   -9.27   -7.71   -7.04   -6.48   -5.54
#> log_p_z[20,3]  -19.95    0.08 4.21  -29.48  -22.32  -19.40  -16.97  -13.46
#> log_p_z[20,4]  -21.57    0.07 4.20  -31.19  -23.90  -21.07  -18.69  -14.67
#> log_p_z[21,1]  -19.60    0.04 2.37  -24.54  -21.07  -19.44  -17.91  -15.51
#> log_p_z[21,2]   -3.04    0.00 0.36   -3.79   -3.27   -3.03   -2.81   -2.39
#> log_p_z[21,3]  -21.17    0.08 4.37  -30.80  -23.71  -20.70  -18.05  -14.14
#> log_p_z[21,4]  -21.81    0.07 4.18  -31.46  -24.12  -21.33  -18.89  -14.98
#> log_p_z[22,1]  -16.39    0.03 1.95  -20.52  -17.61  -16.24  -15.03  -12.98
#> log_p_z[22,2]   -4.66    0.01 0.59   -5.93   -5.03   -4.62   -4.24   -3.64
#> log_p_z[22,3]  -22.06    0.08 4.41  -31.75  -24.60  -21.54  -18.97  -15.04
#> log_p_z[22,4]  -21.82    0.07 4.22  -31.33  -24.18  -21.40  -18.85  -14.84
#> log_p_z[23,1]  -21.42    0.04 2.59  -27.22  -23.00  -21.21  -19.58  -16.98
#> log_p_z[23,2]   -3.71    0.01 0.43   -4.60   -3.98   -3.69   -3.41   -2.92
#> log_p_z[23,3]  -18.66    0.08 4.26  -28.30  -21.04  -18.15  -15.60  -12.10
#> log_p_z[23,4]  -21.82    0.08 4.24  -31.19  -24.21  -21.36  -18.91  -14.79
#> log_p_z[24,1]  -17.40    0.03 2.16  -22.02  -18.72  -17.21  -15.85  -13.70
#> log_p_z[24,2]   -5.79    0.01 0.62   -7.18   -6.16   -5.72   -5.35   -4.75
#> log_p_z[24,3]  -21.71    0.06 3.48  -29.22  -23.83  -21.42  -19.18  -15.95
#> log_p_z[24,4]  -21.82    0.06 3.45  -29.49  -23.89  -21.56  -19.34  -16.01
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_p_z[25,2]  -19.99    0.04 2.29  -25.10  -21.42  -19.80  -18.36  -16.17
#> log_p_z[25,3]  -24.69    0.07 4.22  -34.22  -27.14  -24.28  -21.68  -17.66
#> log_p_z[25,4]  -21.86    0.08 4.35  -31.89  -24.37  -21.31  -18.73  -14.86
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_p_z[26,2]  -19.99    0.04 2.29  -25.10  -21.42  -19.80  -18.36  -16.17
#> log_p_z[26,3]  -24.69    0.07 4.22  -34.22  -27.14  -24.28  -21.68  -17.66
#> log_p_z[26,4]  -21.86    0.08 4.35  -31.89  -24.37  -21.31  -18.73  -14.86
#> log_p_z[27,1]  -21.86    0.04 2.44  -27.11  -23.38  -21.65  -20.14  -17.60
#> log_p_z[27,2]   -5.65    0.01 0.53   -6.81   -5.97   -5.61   -5.29   -4.71
#> log_p_z[27,3]  -18.37    0.06 3.26  -25.57  -20.32  -18.08  -16.05  -12.92
#> log_p_z[27,4]  -21.49    0.06 3.39  -29.04  -23.48  -21.21  -19.11  -15.69
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_p_z[28,2]  -19.99    0.04 2.29  -25.10  -21.42  -19.80  -18.36  -16.17
#> log_p_z[28,3]  -24.69    0.07 4.22  -34.22  -27.14  -24.28  -21.68  -17.66
#> log_p_z[28,4]  -21.86    0.08 4.35  -31.89  -24.37  -21.31  -18.73  -14.86
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_p_z[29,2]  -19.99    0.04 2.29  -25.10  -21.42  -19.80  -18.36  -16.17
#> log_p_z[29,3]  -24.69    0.07 4.22  -34.22  -27.14  -24.28  -21.68  -17.66
#> log_p_z[29,4]  -21.86    0.08 4.35  -31.89  -24.37  -21.31  -18.73  -14.86
#> log_p_z[30,1]   -6.81    0.01 0.82   -8.73   -7.28   -6.72   -6.23   -5.48
#> log_p_z[30,2]  -14.70    0.03 1.70  -18.41  -15.77  -14.54  -13.49  -11.85
#> log_p_z[30,3]  -24.21    0.06 3.49  -31.84  -26.40  -23.93  -21.67  -18.19
#> log_p_z[30,4]  -21.82    0.06 3.58  -29.98  -23.95  -21.46  -19.23  -15.79
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_p_z[31,2]  -19.99    0.04 2.29  -25.10  -21.42  -19.80  -18.36  -16.17
#> log_p_z[31,3]  -24.69    0.07 4.22  -34.22  -27.14  -24.28  -21.68  -17.66
#> log_p_z[31,4]  -21.86    0.08 4.35  -31.89  -24.37  -21.31  -18.73  -14.86
#> log_p_z[32,1]  -29.69    0.06 4.25  -39.17  -32.19  -29.24  -26.73  -22.75
#> log_p_z[32,2]  -14.83    0.02 1.73  -18.62  -15.89  -14.70  -13.64  -11.75
#> log_p_z[32,3]   -6.05    0.01 0.91   -7.97   -6.62   -6.00   -5.42   -4.44
#> log_p_z[32,4]  -20.62    0.08 4.34  -30.70  -23.23  -20.11  -17.42  -13.72
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_p_z[33,2]  -19.99    0.04 2.29  -25.10  -21.42  -19.80  -18.36  -16.17
#> log_p_z[33,3]  -24.69    0.07 4.22  -34.22  -27.14  -24.28  -21.68  -17.66
#> log_p_z[33,4]  -21.86    0.08 4.35  -31.89  -24.37  -21.31  -18.73  -14.86
#> log_p_z[34,1]  -19.60    0.04 2.37  -24.54  -21.07  -19.44  -17.91  -15.51
#> log_p_z[34,2]   -3.04    0.00 0.36   -3.79   -3.27   -3.03   -2.81   -2.39
#> log_p_z[34,3]  -21.17    0.08 4.37  -30.80  -23.71  -20.70  -18.05  -14.14
#> log_p_z[34,4]  -21.81    0.07 4.18  -31.46  -24.12  -21.33  -18.89  -14.98
#> log_p_z[35,1]  -22.56    0.04 2.72  -28.43  -24.32  -22.31  -20.61  -17.94
#> log_p_z[35,2]   -5.53    0.01 0.70   -7.04   -5.98   -5.48   -5.04   -4.31
#> log_p_z[35,3]  -16.01    0.08 4.08  -25.56  -18.24  -15.46  -13.09   -9.73
#> log_p_z[35,4]  -21.82    0.08 4.27  -31.24  -24.27  -21.33  -18.80  -14.87
#> log_p_z[36,1]  -29.70    0.06 3.50  -37.49  -31.82  -29.46  -27.25  -23.57
#> log_p_z[36,2]  -21.88    0.04 2.31  -26.95  -23.29  -21.68  -20.26  -17.93
#> log_p_z[36,3]  -10.58    0.04 1.72  -14.72  -11.46  -10.29   -9.40   -7.98
#> log_p_z[36,4]  -13.91    0.06 3.18  -20.98  -15.90  -13.59  -11.56   -8.72
#> log_p_z[37,1]  -18.54    0.04 2.33  -23.66  -19.97  -18.36  -16.90  -14.49
#> log_p_z[37,2]   -7.61    0.01 0.81   -9.39   -8.11   -7.55   -7.04   -6.21
#> log_p_z[37,3]  -19.06    0.06 3.25  -26.05  -21.00  -18.74  -16.71  -13.78
#> log_p_z[37,4]  -21.82    0.06 3.48  -29.40  -23.92  -21.53  -19.32  -15.93
#> log_p_z[38,1]  -25.18    0.04 2.77  -31.11  -26.93  -24.97  -23.24  -20.40
#> log_p_z[38,2]   -9.62    0.01 0.99  -11.74  -10.23   -9.56   -8.94   -7.85
#> log_p_z[38,3]  -11.65    0.05 2.79  -18.20  -13.23  -11.30   -9.58   -7.51
#> log_p_z[38,4]  -21.26    0.06 3.51  -28.80  -23.39  -21.01  -18.74  -15.19
#> log_p_z[39,1]  -29.32    0.07 4.29  -38.91  -31.74  -28.81  -26.43  -22.28
#> log_p_z[39,2]  -16.79    0.03 1.98  -21.14  -18.00  -16.64  -15.41  -13.37
#> log_p_z[39,3]   -7.28    0.01 1.08   -9.58   -7.98   -7.21   -6.52   -5.36
#> log_p_z[39,4]  -18.74    0.07 4.07  -28.39  -21.09  -18.13  -15.86  -12.27
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_p_z[40,2]  -19.99    0.04 2.29  -25.10  -21.42  -19.80  -18.36  -16.17
#> log_p_z[40,3]  -24.69    0.07 4.22  -34.22  -27.14  -24.28  -21.68  -17.66
#> log_p_z[40,4]  -21.86    0.08 4.35  -31.89  -24.37  -21.31  -18.73  -14.86
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_p_z[41,2]  -19.99    0.04 2.29  -25.10  -21.42  -19.80  -18.36  -16.17
#> log_p_z[41,3]  -24.69    0.07 4.22  -34.22  -27.14  -24.28  -21.68  -17.66
#> log_p_z[41,4]  -21.86    0.08 4.35  -31.89  -24.37  -21.31  -18.73  -14.86
#> log_p_z[42,1]   -6.00    0.01 0.61   -7.34   -6.38   -5.96   -5.57   -4.91
#> log_p_z[42,2]  -14.92    0.03 1.68  -18.62  -15.96  -14.77  -13.73  -12.09
#> log_p_z[42,3]  -24.17    0.06 3.47  -31.87  -26.28  -23.89  -21.70  -18.25
#> log_p_z[42,4]  -21.85    0.06 3.54  -29.58  -24.04  -21.50  -19.26  -15.92
#> log_p_z[43,1]  -21.86    0.04 2.44  -27.11  -23.38  -21.65  -20.14  -17.60
#> log_p_z[43,2]   -5.65    0.01 0.53   -6.81   -5.97   -5.61   -5.29   -4.71
#> log_p_z[43,3]  -18.37    0.06 3.26  -25.57  -20.32  -18.08  -16.05  -12.92
#> log_p_z[43,4]  -21.49    0.06 3.39  -29.04  -23.48  -21.21  -19.11  -15.69
#> log_p_z[44,1]   -4.41    0.01 0.40   -5.26   -4.66   -4.38   -4.13   -3.70
#> log_p_z[44,2]  -17.24    0.03 1.84  -21.14  -18.39  -17.10  -15.95  -14.10
#> log_p_z[44,3]  -24.15    0.06 3.48  -31.94  -26.28  -23.78  -21.70  -18.19
#> log_p_z[44,4]  -21.85    0.07 3.60  -30.06  -24.04  -21.48  -19.28  -15.85
#> log_p_z[45,1]  -19.60    0.04 2.37  -24.54  -21.07  -19.44  -17.91  -15.51
#> log_p_z[45,2]   -3.04    0.00 0.36   -3.79   -3.27   -3.03   -2.81   -2.39
#> log_p_z[45,3]  -21.17    0.08 4.37  -30.80  -23.71  -20.70  -18.05  -14.14
#> log_p_z[45,4]  -21.81    0.07 4.18  -31.46  -24.12  -21.33  -18.89  -14.98
#> log_lik[1]      -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_lik[2]      -7.61    0.02 1.13  -10.07   -8.34   -7.53   -6.80   -5.66
#> log_lik[3]      -9.93    0.01 0.91  -11.79  -10.54   -9.89   -9.28   -8.25
#> log_lik[4]      -7.55    0.01 0.91   -9.60   -8.09   -7.49   -6.90   -5.97
#> log_lik[5]      -3.71    0.01 0.43   -4.60   -3.98   -3.69   -3.41   -2.92
#> log_lik[6]      -5.47    0.01 0.65   -6.86   -5.88   -5.44   -5.02   -4.33
#> log_lik[7]      -6.00    0.01 0.61   -7.34   -6.38   -5.96   -5.57   -4.91
#> log_lik[8]      -5.72    0.01 0.81   -7.45   -6.24   -5.67   -5.15   -4.26
#> log_lik[9]      -4.52    0.01 0.58   -5.78   -4.88   -4.49   -4.12   -3.54
#> log_lik[10]     -7.13    0.01 0.71   -8.64   -7.58   -7.08   -6.63   -5.87
#> log_lik[11]     -5.84    0.01 0.98   -7.98   -6.43   -5.75   -5.16   -4.16
#> log_lik[12]     -9.33    0.01 1.03  -11.50   -9.98   -9.27   -8.60   -7.49
#> log_lik[13]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_lik[14]     -6.25    0.01 0.86   -8.20   -6.77   -6.17   -5.64   -4.78
#> log_lik[15]     -4.41    0.01 0.40   -5.26   -4.66   -4.38   -4.13   -3.70
#> log_lik[16]     -3.79    0.01 0.54   -4.98   -4.14   -3.76   -3.41   -2.85
#> log_lik[17]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_lik[18]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_lik[19]     -4.66    0.01 0.59   -5.93   -5.03   -4.62   -4.24   -3.64
#> log_lik[20]     -7.13    0.01 0.95   -9.27   -7.71   -7.04   -6.48   -5.54
#> log_lik[21]     -3.04    0.00 0.36   -3.79   -3.27   -3.03   -2.81   -2.39
#> log_lik[22]     -4.66    0.01 0.59   -5.93   -5.03   -4.62   -4.24   -3.64
#> log_lik[23]     -3.71    0.01 0.43   -4.60   -3.98   -3.69   -3.41   -2.92
#> log_lik[24]     -5.79    0.01 0.62   -7.18   -6.16   -5.72   -5.35   -4.75
#> log_lik[25]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_lik[26]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_lik[27]     -5.65    0.01 0.53   -6.80   -5.97   -5.61   -5.29   -4.71
#> log_lik[28]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_lik[29]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_lik[30]     -6.80    0.01 0.82   -8.72   -7.28   -6.72   -6.23   -5.48
#> log_lik[31]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_lik[32]     -6.05    0.01 0.91   -7.97   -6.62   -6.00   -5.42   -4.44
#> log_lik[33]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_lik[34]     -3.04    0.00 0.36   -3.79   -3.27   -3.03   -2.81   -2.39
#> log_lik[35]     -5.53    0.01 0.69   -7.01   -5.98   -5.48   -5.04   -4.31
#> log_lik[36]     -9.96    0.02 1.16  -12.55  -10.70   -9.91   -9.15   -7.87
#> log_lik[37]     -7.61    0.01 0.80   -9.39   -8.11   -7.55   -7.04   -6.21
#> log_lik[38]     -9.01    0.01 0.86  -10.68   -9.58   -9.00   -8.45   -7.33
#> log_lik[39]     -7.27    0.01 1.08   -9.56   -7.97   -7.21   -6.52   -5.36
#> log_lik[40]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_lik[41]     -2.21    0.00 0.30   -2.83   -2.40   -2.19   -2.00   -1.65
#> log_lik[42]     -6.00    0.01 0.61   -7.34   -6.38   -5.96   -5.57   -4.91
#> log_lik[43]     -5.65    0.01 0.53   -6.80   -5.97   -5.61   -5.29   -4.71
#> log_lik[44]     -4.41    0.01 0.40   -5.26   -4.66   -4.38   -4.13   -3.70
#> log_lik[45]     -3.04    0.00 0.36   -3.79   -3.27   -3.03   -2.81   -2.39
#> lp__          -463.55    0.16 6.00 -476.07 -467.29 -463.15 -459.32 -452.76
#>               n_eff Rhat
#> pi[1]          6582    1
#> pi[2]          7529    1
#> pi[3]          6570    1
#> pi[4]          6153    1
#> theta[1,1,1]   6990    1
#> theta[1,1,2]   6429    1
#> theta[1,1,3]   5445    1
#> theta[1,1,4]   5236    1
#> theta[1,2,1]   5869    1
#> theta[1,2,2]   6142    1
#> theta[1,2,3]   6453    1
#> theta[1,2,4]   4668    1
#> theta[1,3,1]   5851    1
#> theta[1,3,2]   3909    1
#> theta[1,3,3]   4707    1
#> theta[1,3,4]   4778    1
#> theta[1,4,1]   6313    1
#> theta[1,4,2]   6527    1
#> theta[1,4,3]   4115    1
#> theta[1,4,4]   5985    1
#> theta[2,1,1]   6703    1
#> theta[2,1,2]   5939    1
#> theta[2,1,3]   4744    1
#> theta[2,1,4]   5076    1
#> theta[2,2,1]   7122    1
#> theta[2,2,2]   8495    1
#> theta[2,2,3]   8483    1
#> theta[2,2,4]   5963    1
#> theta[2,3,1]   5786    1
#> theta[2,3,2]   5519    1
#> theta[2,3,3]   5415    1
#> theta[2,3,4]   5822    1
#> theta[2,4,1]   7685    1
#> theta[2,4,2]   6350    1
#> theta[2,4,3]   5909    1
#> theta[2,4,4]   6825    1
#> theta[3,1,1]   8685    1
#> theta[3,1,2]   4759    1
#> theta[3,1,3]   4715    1
#> theta[3,1,4]   4897    1
#> theta[3,2,1]   7126    1
#> theta[3,2,2]   6515    1
#> theta[3,2,3]   7116    1
#> theta[3,2,4]   5234    1
#> theta[3,3,1]   5442    1
#> theta[3,3,2]   5796    1
#> theta[3,3,3]   6290    1
#> theta[3,3,4]   6703    1
#> theta[3,4,1]   6474    1
#> theta[3,4,2]   7346    1
#> theta[3,4,3]   4784    1
#> theta[3,4,4]   5366    1
#> theta[4,1,1]   6384    1
#> theta[4,1,2]   5260    1
#> theta[4,1,3]   4621    1
#> theta[4,1,4]   4310    1
#> theta[4,2,1]   5998    1
#> theta[4,2,2]   6238    1
#> theta[4,2,3]   5959    1
#> theta[4,2,4]   5347    1
#> theta[4,3,1]   6742    1
#> theta[4,3,2]   5593    1
#> theta[4,3,3]   5743    1
#> theta[4,3,4]   5019    1
#> theta[4,4,1]   6262    1
#> theta[4,4,2]   5292    1
#> theta[4,4,3]   6569    1
#> theta[4,4,4]   5863    1
#> theta[5,1,1]   6650    1
#> theta[5,1,2]   4866    1
#> theta[5,1,3]   4259    1
#> theta[5,1,4]   4845    1
#> theta[5,2,1]   7374    1
#> theta[5,2,2]   6815    1
#> theta[5,2,3]   6360    1
#> theta[5,2,4]   5891    1
#> theta[5,3,1]   6198    1
#> theta[5,3,2]   6970    1
#> theta[5,3,3]   6374    1
#> theta[5,3,4]   6955    1
#> theta[5,4,1]   6149    1
#> theta[5,4,2]   6609    1
#> theta[5,4,3]   4369    1
#> theta[5,4,4]   5898    1
#> log_p_z[1,1]   6852    1
#> log_p_z[1,2]   4264    1
#> log_p_z[1,3]   3720    1
#> log_p_z[1,4]   3258    1
#> log_p_z[2,1]   3951    1
#> log_p_z[2,2]   4474    1
#> log_p_z[2,3]   3870    1
#> log_p_z[2,4]   2961    1
#> log_p_z[3,1]   4037    1
#> log_p_z[3,2]   3788    1
#> log_p_z[3,3]   3442    1
#> log_p_z[3,4]   3714    1
#> log_p_z[4,1]   4377    1
#> log_p_z[4,2]   5275    1
#> log_p_z[4,3]   3142    1
#> log_p_z[4,4]   3264    1
#> log_p_z[5,1]   4021    1
#> log_p_z[5,2]   7227    1
#> log_p_z[5,3]   2997    1
#> log_p_z[5,4]   3075    1
#> log_p_z[6,1]   4353    1
#> log_p_z[6,2]   6565    1
#> log_p_z[6,3]   2945    1
#> log_p_z[6,4]   3186    1
#> log_p_z[7,1]   6331    1
#> log_p_z[7,2]   4115    1
#> log_p_z[7,3]   3640    1
#> log_p_z[7,4]   3551    1
#> log_p_z[8,1]   4126    1
#> log_p_z[8,2]   4130    1
#> log_p_z[8,3]   5952    1
#> log_p_z[8,4]   3091    1
#> log_p_z[9,1]   4328    1
#> log_p_z[9,2]   6013    1
#> log_p_z[9,3]   3041    1
#> log_p_z[9,4]   3324    1
#> log_p_z[10,1]  4320    1
#> log_p_z[10,2]  4688    1
#> log_p_z[10,3]  3179    1
#> log_p_z[10,4]  3179    1
#> log_p_z[11,1]  5514    1
#> log_p_z[11,2]  3808    1
#> log_p_z[11,3]  2126    1
#> log_p_z[11,4]  5830    1
#> log_p_z[12,1]  4469    1
#> log_p_z[12,2]  3892    1
#> log_p_z[12,3]  2776    1
#> log_p_z[12,4]  3042    1
#> log_p_z[13,1]  6852    1
#> log_p_z[13,2]  4264    1
#> log_p_z[13,3]  3720    1
#> log_p_z[13,4]  3258    1
#> log_p_z[14,1]  4215    1
#> log_p_z[14,2]  4245    1
#> log_p_z[14,3]  2697    1
#> log_p_z[14,4]  3042    1
#> log_p_z[15,1]  6154    1
#> log_p_z[15,2]  4311    1
#> log_p_z[15,3]  3625    1
#> log_p_z[15,4]  3060    1
#> log_p_z[16,1]  5700    1
#> log_p_z[16,2]  4087    1
#> log_p_z[16,3]  3680    1
#> log_p_z[16,4]  3691    1
#> log_p_z[17,1]  6852    1
#> log_p_z[17,2]  4264    1
#> log_p_z[17,3]  3720    1
#> log_p_z[17,4]  3258    1
#> log_p_z[18,1]  6852    1
#> log_p_z[18,2]  4264    1
#> log_p_z[18,3]  3720    1
#> log_p_z[18,4]  3258    1
#> log_p_z[19,1]  4496    1
#> log_p_z[19,2]  5935    1
#> log_p_z[19,3]  3084    1
#> log_p_z[19,4]  3233    1
#> log_p_z[20,1]  4530    1
#> log_p_z[20,2]  4533    1
#> log_p_z[20,3]  2906    1
#> log_p_z[20,4]  3194    1
#> log_p_z[21,1]  4250    1
#> log_p_z[21,2]  8350    1
#> log_p_z[21,3]  3114    1
#> log_p_z[21,4]  3188    1
#> log_p_z[22,1]  4496    1
#> log_p_z[22,2]  5935    1
#> log_p_z[22,3]  3084    1
#> log_p_z[22,4]  3233    1
#> log_p_z[23,1]  4021    1
#> log_p_z[23,2]  7227    1
#> log_p_z[23,3]  2997    1
#> log_p_z[23,4]  3075    1
#> log_p_z[24,1]  4072    1
#> log_p_z[24,2]  4380    1
#> log_p_z[24,3]  3252    1
#> log_p_z[24,4]  3240    1
#> log_p_z[25,1]  6852    1
#> log_p_z[25,2]  4264    1
#> log_p_z[25,3]  3720    1
#> log_p_z[25,4]  3258    1
#> log_p_z[26,1]  6852    1
#> log_p_z[26,2]  4264    1
#> log_p_z[26,3]  3720    1
#> log_p_z[26,4]  3258    1
#> log_p_z[27,1]  4049    1
#> log_p_z[27,2]  5455    1
#> log_p_z[27,3]  3317    1
#> log_p_z[27,4]  3304    1
#> log_p_z[28,1]  6852    1
#> log_p_z[28,2]  4264    1
#> log_p_z[28,3]  3720    1
#> log_p_z[28,4]  3258    1
#> log_p_z[29,1]  6852    1
#> log_p_z[29,2]  4264    1
#> log_p_z[29,3]  3720    1
#> log_p_z[29,4]  3258    1
#> log_p_z[30,1]  4213    1
#> log_p_z[30,2]  4258    1
#> log_p_z[30,3]  3546    1
#> log_p_z[30,4]  3253    1
#> log_p_z[31,1]  6852    1
#> log_p_z[31,2]  4264    1
#> log_p_z[31,3]  3720    1
#> log_p_z[31,4]  3258    1
#> log_p_z[32,1]  4346    1
#> log_p_z[32,2]  4913    1
#> log_p_z[32,3]  4102    1
#> log_p_z[32,4]  3160    1
#> log_p_z[33,1]  6852    1
#> log_p_z[33,2]  4264    1
#> log_p_z[33,3]  3720    1
#> log_p_z[33,4]  3258    1
#> log_p_z[34,1]  4250    1
#> log_p_z[34,2]  8350    1
#> log_p_z[34,3]  3114    1
#> log_p_z[34,4]  3188    1
#> log_p_z[35,1]  3935    1
#> log_p_z[35,2]  5733    1
#> log_p_z[35,3]  2822    1
#> log_p_z[35,4]  2982    1
#> log_p_z[36,1]  3631    1
#> log_p_z[36,2]  3808    1
#> log_p_z[36,3]  1827    1
#> log_p_z[36,4]  2638    1
#> log_p_z[37,1]  3888    1
#> log_p_z[37,2]  4618    1
#> log_p_z[37,3]  2983    1
#> log_p_z[37,4]  2969    1
#> log_p_z[38,1]  4300    1
#> log_p_z[38,2]  4737    1
#> log_p_z[38,3]  2792    1
#> log_p_z[38,4]  3103    1
#> log_p_z[39,1]  3921    1
#> log_p_z[39,2]  4634    1
#> log_p_z[39,3]  6093    1
#> log_p_z[39,4]  3454    1
#> log_p_z[40,1]  6852    1
#> log_p_z[40,2]  4264    1
#> log_p_z[40,3]  3720    1
#> log_p_z[40,4]  3258    1
#> log_p_z[41,1]  6852    1
#> log_p_z[41,2]  4264    1
#> log_p_z[41,3]  3720    1
#> log_p_z[41,4]  3258    1
#> log_p_z[42,1]  6331    1
#> log_p_z[42,2]  4115    1
#> log_p_z[42,3]  3640    1
#> log_p_z[42,4]  3551    1
#> log_p_z[43,1]  4049    1
#> log_p_z[43,2]  5455    1
#> log_p_z[43,3]  3317    1
#> log_p_z[43,4]  3304    1
#> log_p_z[44,1]  6154    1
#> log_p_z[44,2]  4311    1
#> log_p_z[44,3]  3625    1
#> log_p_z[44,4]  3060    1
#> log_p_z[45,1]  4250    1
#> log_p_z[45,2]  8350    1
#> log_p_z[45,3]  3114    1
#> log_p_z[45,4]  3188    1
#> log_lik[1]     6852    1
#> log_lik[2]     4463    1
#> log_lik[3]     4308    1
#> log_lik[4]     5321    1
#> log_lik[5]     7228    1
#> log_lik[6]     6573    1
#> log_lik[7]     6333    1
#> log_lik[8]     5951    1
#> log_lik[9]     6014    1
#> log_lik[10]    4698    1
#> log_lik[11]    5829    1
#> log_lik[12]    5622    1
#> log_lik[13]    6852    1
#> log_lik[14]    4245    1
#> log_lik[15]    6154    1
#> log_lik[16]    5700    1
#> log_lik[17]    6852    1
#> log_lik[18]    6852    1
#> log_lik[19]    5935    1
#> log_lik[20]    4534    1
#> log_lik[21]    8350    1
#> log_lik[22]    5935    1
#> log_lik[23]    7228    1
#> log_lik[24]    4380    1
#> log_lik[25]    6852    1
#> log_lik[26]    6852    1
#> log_lik[27]    5455    1
#> log_lik[28]    6852    1
#> log_lik[29]    6852    1
#> log_lik[30]    4239    1
#> log_lik[31]    6852    1
#> log_lik[32]    4103    1
#> log_lik[33]    6852    1
#> log_lik[34]    8350    1
#> log_lik[35]    5756    1
#> log_lik[36]    4156    1
#> log_lik[37]    4623    1
#> log_lik[38]    4623    1
#> log_lik[39]    6122    1
#> log_lik[40]    6852    1
#> log_lik[41]    6852    1
#> log_lik[42]    6333    1
#> log_lik[43]    5455    1
#> log_lik[44]    6154    1
#> log_lik[45]    8350    1
#> lp__           1471    1
#> 
#> Samples were drawn using NUTS(diag_e) at Sat Mar 11 08:29:50 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }