Get the underlying stanfit object from a rater_fit object.

Get the underlying stanfit object from a rater_fit object.

get_stanfit(fit)

Arguments

fit
A rater_fit object.

Value

A stanfit object from rstan.

Examples

# \donttest{
fit <- rater(anesthesia, "dawid_skene", verbose = FALSE)

stan_fit <- get_stanfit(fit)
stan_fit

#> Inference for Stan model: dawid_skene.
#> 4 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=4000.
#> 
#>                  mean se_mean   sd    2.5%     25%     50%     75%   97.5%
#> pi[1]            0.37    0.00 0.06    0.25    0.33    0.37    0.42    0.51
#> pi[2]            0.41    0.00 0.06    0.28    0.36    0.41    0.45    0.53
#> pi[3]            0.14    0.00 0.05    0.06    0.11    0.14    0.17    0.25
#> pi[4]            0.07    0.00 0.03    0.02    0.05    0.07    0.10    0.16
#> theta[1,1,1]     0.86    0.00 0.04    0.76    0.84    0.87    0.89    0.94
#> theta[1,1,2]     0.10    0.00 0.04    0.04    0.08    0.10    0.12    0.19
#> theta[1,1,3]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,1,4]     0.02    0.00 0.02    0.00    0.01    0.01    0.02    0.06
#> theta[1,2,1]     0.06    0.00 0.03    0.01    0.04    0.06    0.08    0.14
#> theta[1,2,2]     0.85    0.00 0.04    0.75    0.82    0.86    0.88    0.93
#> theta[1,2,3]     0.07    0.00 0.03    0.02    0.05    0.07    0.09    0.14
#> theta[1,2,4]     0.02    0.00 0.02    0.00    0.00    0.01    0.02    0.06
#> theta[1,3,1]     0.05    0.00 0.04    0.00    0.01    0.03    0.06    0.16
#> theta[1,3,2]     0.08    0.00 0.08    0.00    0.02    0.06    0.12    0.29
#> theta[1,3,3]     0.79    0.00 0.10    0.58    0.73    0.81    0.87    0.94
#> theta[1,3,4]     0.08    0.00 0.06    0.01    0.04    0.07    0.11    0.21
#> theta[1,4,1]     0.09    0.00 0.08    0.00    0.03    0.07    0.13    0.30
#> theta[1,4,2]     0.10    0.00 0.09    0.00    0.03    0.07    0.13    0.32
#> theta[1,4,3]     0.13    0.00 0.11    0.00    0.04    0.10    0.18    0.41
#> theta[1,4,4]     0.68    0.00 0.14    0.39    0.59    0.70    0.79    0.91
#> theta[2,1,1]     0.75    0.00 0.08    0.57    0.70    0.76    0.81    0.89
#> theta[2,1,2]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.33
#> theta[2,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[2,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[2,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[2,2,2]     0.58    0.00 0.09    0.39    0.52    0.58    0.65    0.75
#> theta[2,2,3]     0.31    0.00 0.09    0.15    0.25    0.30    0.36    0.49
#> theta[2,2,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.13
#> theta[2,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.27
#> theta[2,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.28
#> theta[2,3,3]     0.62    0.00 0.13    0.36    0.53    0.62    0.72    0.86
#> theta[2,3,4]     0.21    0.00 0.12    0.04    0.13    0.20    0.29    0.48
#> theta[2,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.36
#> theta[2,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[2,4,3]     0.11    0.00 0.10    0.00    0.04    0.09    0.17    0.36
#> theta[2,4,4]     0.65    0.00 0.15    0.34    0.56    0.66    0.77    0.91
#> theta[3,1,1]     0.88    0.00 0.07    0.72    0.84    0.89    0.93    0.98
#> theta[3,1,2]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[3,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[3,2,1]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.22
#> theta[3,2,2]     0.73    0.00 0.08    0.55    0.67    0.73    0.79    0.87
#> theta[3,2,3]     0.14    0.00 0.07    0.04    0.09    0.13    0.18    0.29
#> theta[3,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[3,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.26
#> theta[3,3,2]     0.18    0.00 0.11    0.03    0.10    0.16    0.24    0.43
#> theta[3,3,3]     0.51    0.00 0.14    0.24    0.41    0.51    0.60    0.76
#> theta[3,3,4]     0.23    0.00 0.11    0.06    0.15    0.22    0.30    0.49
#> theta[3,4,1]     0.11    0.00 0.10    0.00    0.04    0.09    0.16    0.38
#> theta[3,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[3,4,3]     0.14    0.00 0.12    0.00    0.05    0.11    0.20    0.43
#> theta[3,4,4]     0.63    0.00 0.15    0.31    0.53    0.64    0.75    0.89
#> theta[4,1,1]     0.83    0.00 0.08    0.66    0.78    0.83    0.88    0.95
#> theta[4,1,2]     0.09    0.00 0.06    0.01    0.05    0.08    0.13    0.23
#> theta[4,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[4,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.15
#> theta[4,2,1]     0.07    0.00 0.05    0.01    0.04    0.06    0.10    0.19
#> theta[4,2,2]     0.72    0.00 0.08    0.55    0.67    0.73    0.78    0.87
#> theta[4,2,3]     0.13    0.00 0.06    0.04    0.09    0.13    0.17    0.28
#> theta[4,2,4]     0.07    0.00 0.05    0.01    0.03    0.06    0.09    0.19
#> theta[4,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[4,3,2]     0.08    0.00 0.07    0.00    0.03    0.06    0.11    0.28
#> theta[4,3,3]     0.69    0.00 0.13    0.41    0.60    0.70    0.78    0.91
#> theta[4,3,4]     0.15    0.00 0.10    0.01    0.07    0.13    0.21    0.40
#> theta[4,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[4,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[4,4,3]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.39
#> theta[4,4,4]     0.65    0.00 0.15    0.32    0.55    0.66    0.76    0.90
#> theta[5,1,1]     0.87    0.00 0.07    0.71    0.83    0.87    0.91    0.97
#> theta[5,1,2]     0.05    0.00 0.05    0.00    0.02    0.04    0.08    0.17
#> theta[5,1,3]     0.04    0.00 0.04    0.00    0.01    0.03    0.05    0.14
#> theta[5,1,4]     0.04    0.00 0.04    0.00    0.01    0.03    0.06    0.14
#> theta[5,2,1]     0.14    0.00 0.06    0.04    0.10    0.14    0.18    0.29
#> theta[5,2,2]     0.65    0.00 0.09    0.47    0.59    0.66    0.72    0.82
#> theta[5,2,3]     0.17    0.00 0.07    0.05    0.11    0.16    0.21    0.33
#> theta[5,2,4]     0.04    0.00 0.03    0.00    0.01    0.03    0.05    0.13
#> theta[5,3,1]     0.08    0.00 0.07    0.00    0.03    0.06    0.12    0.26
#> theta[5,3,2]     0.16    0.00 0.09    0.02    0.08    0.14    0.21    0.38
#> theta[5,3,3]     0.61    0.00 0.13    0.34    0.52    0.61    0.70    0.84
#> theta[5,3,4]     0.16    0.00 0.10    0.02    0.08    0.14    0.22    0.39
#> theta[5,4,1]     0.12    0.00 0.10    0.00    0.04    0.09    0.16    0.37
#> theta[5,4,2]     0.12    0.00 0.10    0.00    0.04    0.09    0.17    0.38
#> theta[5,4,3]     0.14    0.00 0.11    0.01    0.05    0.11    0.20    0.41
#> theta[5,4,4]     0.63    0.00 0.15    0.32    0.52    0.64    0.75    0.90
#> log_p_z[1,1]    -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_p_z[1,2]   -20.02    0.04 2.30  -24.99  -21.44  -19.87  -18.40  -15.88
#> log_p_z[1,3]   -24.95    0.08 4.52  -35.26  -27.53  -24.48  -21.81  -17.66
#> log_p_z[1,4]   -21.98    0.07 4.31  -31.83  -24.54  -21.47  -18.94  -14.96
#> log_p_z[2,1]   -29.68    0.07 4.28  -39.17  -32.20  -29.25  -26.69  -22.71
#> log_p_z[2,2]   -21.03    0.04 2.38  -26.28  -22.46  -20.87  -19.38  -16.83
#> log_p_z[2,3]    -7.63    0.02 1.19  -10.26   -8.35   -7.54   -6.80   -5.62
#> log_p_z[2,4]   -16.34    0.08 4.18  -26.17  -18.71  -15.81  -13.36   -9.73
#> log_p_z[3,1]   -11.63    0.03 1.65  -15.36  -12.61  -11.44  -10.46   -8.95
#> log_p_z[3,2]   -10.83    0.02 1.49  -14.13  -11.73  -10.66   -9.76   -8.39
#> log_p_z[3,3]   -23.50    0.06 3.62  -31.50  -25.69  -23.15  -20.94  -17.45
#> log_p_z[3,4]   -21.96    0.06 3.58  -30.12  -24.09  -21.63  -19.40  -16.03
#> log_p_z[4,1]   -14.66    0.03 1.88  -18.75  -15.81  -14.51  -13.39  -11.39
#> log_p_z[4,2]    -7.57    0.01 0.92   -9.58   -8.16   -7.49   -6.91   -5.98
#> log_p_z[4,3]   -20.90    0.08 4.58  -31.48  -23.44  -20.22  -17.63  -14.00
#> log_p_z[4,4]   -21.97    0.07 4.50  -32.32  -24.54  -21.35  -18.83  -14.72
#> log_p_z[5,1]   -21.48    0.04 2.60  -27.22  -23.11  -21.26  -19.68  -16.96
#> log_p_z[5,2]    -3.70    0.00 0.43   -4.63   -3.97   -3.68   -3.41   -2.92
#> log_p_z[5,3]   -18.90    0.08 4.50  -29.51  -21.45  -18.24  -15.69  -12.12
#> log_p_z[5,4]   -21.95    0.07 4.40  -32.03  -24.44  -21.47  -18.82  -14.87
#> log_p_z[6,1]   -21.48    0.04 2.60  -27.03  -23.07  -21.28  -19.66  -16.91
#> log_p_z[6,2]    -5.46    0.01 0.66   -6.89   -5.87   -5.42   -5.00   -4.28
#> log_p_z[6,3]   -17.69    0.08 4.37  -27.92  -20.13  -16.99  -14.64  -11.18
#> log_p_z[6,4]   -21.75    0.07 4.41  -31.91  -24.30  -21.18  -18.62  -14.60
#> log_p_z[7,1]    -6.00    0.01 0.60   -7.32   -6.37   -5.96   -5.58   -4.95
#> log_p_z[7,2]   -14.97    0.03 1.73  -18.79  -16.03  -14.85  -13.75  -11.94
#> log_p_z[7,3]   -24.44    0.06 3.77  -32.55  -26.70  -24.09  -21.82  -18.08
#> log_p_z[7,4]   -21.98    0.06 3.62  -29.99  -24.18  -21.66  -19.43  -15.91
#> log_p_z[8,1]   -29.74    0.07 4.27  -39.29  -32.26  -29.36  -26.79  -22.62
#> log_p_z[8,2]   -18.28    0.03 2.10  -22.83  -19.57  -18.10  -16.80  -14.68
#> log_p_z[8,3]    -5.73    0.01 0.83   -7.50   -6.24   -5.66   -5.15   -4.26
#> log_p_z[8,4]   -18.54    0.07 4.23  -28.23  -20.96  -18.01  -15.55  -11.88
#> log_p_z[9,1]   -19.99    0.03 2.30  -24.85  -21.39  -19.81  -18.38  -16.02
#> log_p_z[9,2]    -4.51    0.01 0.58   -5.74   -4.87   -4.47   -4.11   -3.51
#> log_p_z[9,3]   -19.85    0.08 4.61  -30.68  -22.41  -19.23  -16.54  -12.78
#> log_p_z[9,4]   -21.67    0.07 4.27  -31.56  -24.04  -21.14  -18.76  -14.68
#> log_p_z[10,1]  -22.24    0.04 2.42  -27.52  -23.77  -21.99  -20.56  -18.12
#> log_p_z[10,2]   -7.12    0.01 0.72   -8.65   -7.58   -7.06   -6.60   -5.86
#> log_p_z[10,3]  -16.96    0.06 3.34  -24.59  -18.89  -16.56  -14.60  -11.64
#> log_p_z[10,4]  -21.33    0.06 3.46  -29.06  -23.39  -21.05  -18.94  -15.38
#> log_p_z[11,1]  -29.74    0.06 4.50  -39.98  -32.30  -29.27  -26.63  -22.38
#> log_p_z[11,2]  -29.36    0.08 4.33  -39.17  -31.89  -28.90  -26.29  -22.15
#> log_p_z[11,3]  -18.20    0.06 3.35  -26.00  -19.98  -17.68  -15.90  -13.02
#> log_p_z[11,4]   -5.82    0.01 0.97   -7.94   -6.42   -5.75   -5.15   -4.16
#> log_p_z[12,1]  -22.96    0.04 2.68  -28.60  -24.64  -22.74  -21.06  -18.27
#> log_p_z[12,2]   -9.58    0.02 1.30  -12.59  -10.32   -9.42   -8.68   -7.46
#> log_p_z[12,3]  -15.34    0.08 4.31  -25.44  -17.60  -14.71  -12.29   -8.83
#> log_p_z[12,4]  -19.34    0.07 4.32  -29.74  -21.73  -18.77  -16.39  -12.53
#> log_p_z[13,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_p_z[13,2]  -20.02    0.04 2.30  -24.99  -21.44  -19.87  -18.40  -15.88
#> log_p_z[13,3]  -24.95    0.08 4.52  -35.26  -27.53  -24.48  -21.81  -17.66
#> log_p_z[13,4]  -21.98    0.07 4.31  -31.83  -24.54  -21.47  -18.94  -14.96
#> log_p_z[14,1]  -19.08    0.04 2.42  -24.34  -20.59  -18.85  -17.36  -14.98
#> log_p_z[14,2]   -6.23    0.01 0.85   -8.08   -6.75   -6.16   -5.63   -4.78
#> log_p_z[14,3]  -18.89    0.08 4.51  -29.33  -21.38  -18.28  -15.72  -12.02
#> log_p_z[14,4]  -21.99    0.07 4.50  -32.16  -24.57  -21.48  -18.82  -14.73
#> log_p_z[15,1]   -4.41    0.00 0.39   -5.24   -4.66   -4.39   -4.14   -3.72
#> log_p_z[15,2]  -17.25    0.03 1.85  -21.28  -18.40  -17.12  -15.97  -13.90
#> log_p_z[15,3]  -24.43    0.06 3.69  -32.51  -26.64  -24.19  -21.91  -18.03
#> log_p_z[15,4]  -21.98    0.06 3.65  -30.17  -24.19  -21.69  -19.38  -15.85
#> log_p_z[16,1]   -3.80    0.01 0.54   -4.97   -4.14   -3.77   -3.42   -2.85
#> log_p_z[16,2]  -17.73    0.03 2.22  -22.64  -19.09  -17.55  -16.17  -13.87
#> log_p_z[16,3]  -24.95    0.08 4.59  -35.47  -27.58  -24.39  -21.77  -17.49
#> log_p_z[16,4]  -21.98    0.08 4.30  -31.62  -24.39  -21.48  -18.98  -15.01
#> log_p_z[17,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_p_z[17,2]  -20.02    0.04 2.30  -24.99  -21.44  -19.87  -18.40  -15.88
#> log_p_z[17,3]  -24.95    0.08 4.52  -35.26  -27.53  -24.48  -21.81  -17.66
#> log_p_z[17,4]  -21.98    0.07 4.31  -31.83  -24.54  -21.47  -18.94  -14.96
#> log_p_z[18,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_p_z[18,2]  -20.02    0.04 2.30  -24.99  -21.44  -19.87  -18.40  -15.88
#> log_p_z[18,3]  -24.95    0.08 4.52  -35.26  -27.53  -24.48  -21.81  -17.66
#> log_p_z[18,4]  -21.98    0.07 4.31  -31.83  -24.54  -21.47  -18.94  -14.96
#> log_p_z[19,1]  -16.40    0.03 1.95  -20.58  -17.61  -16.23  -15.03  -12.96
#> log_p_z[19,2]   -4.65    0.01 0.58   -5.88   -5.01   -4.62   -4.26   -3.62
#> log_p_z[19,3]  -22.34    0.08 4.72  -33.40  -25.00  -21.70  -18.95  -15.03
#> log_p_z[19,4]  -21.91    0.07 4.32  -31.82  -24.44  -21.35  -18.91  -14.90
#> log_p_z[20,1]  -18.03    0.03 2.31  -23.03  -19.53  -17.79  -16.42  -14.11
#> log_p_z[20,2]   -7.08    0.01 0.97   -9.18   -7.66   -7.01   -6.41   -5.42
#> log_p_z[20,3]  -20.19    0.08 4.46  -30.84  -22.69  -19.63  -17.06  -13.35
#> log_p_z[20,4]  -21.73    0.07 4.36  -31.71  -24.27  -21.22  -18.63  -14.86
#> log_p_z[21,1]  -19.62    0.04 2.38  -24.80  -21.15  -19.42  -17.94  -15.46
#> log_p_z[21,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.79   -2.38
#> log_p_z[21,3]  -21.40    0.08 4.63  -32.39  -23.94  -20.77  -18.11  -14.23
#> log_p_z[21,4]  -21.93    0.07 4.30  -31.67  -24.48  -21.42  -18.84  -14.91
#> log_p_z[22,1]  -16.40    0.03 1.95  -20.58  -17.61  -16.23  -15.03  -12.96
#> log_p_z[22,2]   -4.65    0.01 0.58   -5.88   -5.01   -4.62   -4.26   -3.62
#> log_p_z[22,3]  -22.34    0.08 4.72  -33.40  -25.00  -21.70  -18.95  -15.03
#> log_p_z[22,4]  -21.91    0.07 4.32  -31.82  -24.44  -21.35  -18.91  -14.90
#> log_p_z[23,1]  -21.48    0.04 2.60  -27.22  -23.11  -21.26  -19.68  -16.96
#> log_p_z[23,2]   -3.70    0.00 0.43   -4.63   -3.97   -3.68   -3.41   -2.92
#> log_p_z[23,3]  -18.90    0.08 4.50  -29.51  -21.45  -18.24  -15.69  -12.12
#> log_p_z[23,4]  -21.95    0.07 4.40  -32.03  -24.44  -21.47  -18.82  -14.87
#> log_p_z[24,1]  -17.42    0.03 2.18  -22.27  -18.77  -17.24  -15.90  -13.64
#> log_p_z[24,2]   -5.81    0.01 0.64   -7.25   -6.18   -5.74   -5.35   -4.77
#> log_p_z[24,3]  -21.92    0.06 3.65  -30.21  -24.06  -21.60  -19.36  -15.93
#> log_p_z[24,4]  -21.92    0.06 3.53  -29.61  -24.04  -21.64  -19.44  -15.85
#> log_p_z[25,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_p_z[25,2]  -20.02    0.04 2.30  -24.99  -21.44  -19.87  -18.40  -15.88
#> log_p_z[25,3]  -24.95    0.08 4.52  -35.26  -27.53  -24.48  -21.81  -17.66
#> log_p_z[25,4]  -21.98    0.07 4.31  -31.83  -24.54  -21.47  -18.94  -14.96
#> log_p_z[26,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_p_z[26,2]  -20.02    0.04 2.30  -24.99  -21.44  -19.87  -18.40  -15.88
#> log_p_z[26,3]  -24.95    0.08 4.52  -35.26  -27.53  -24.48  -21.81  -17.66
#> log_p_z[26,4]  -21.98    0.07 4.31  -31.83  -24.54  -21.47  -18.94  -14.96
#> log_p_z[27,1]  -21.88    0.04 2.47  -27.29  -23.46  -21.64  -20.16  -17.64
#> log_p_z[27,2]   -5.65    0.01 0.53   -6.80   -5.98   -5.61   -5.28   -4.73
#> log_p_z[27,3]  -18.51    0.06 3.39  -26.25  -20.42  -18.12  -16.15  -13.12
#> log_p_z[27,4]  -21.58    0.06 3.46  -28.98  -23.65  -21.30  -19.10  -15.76
#> log_p_z[28,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_p_z[28,2]  -20.02    0.04 2.30  -24.99  -21.44  -19.87  -18.40  -15.88
#> log_p_z[28,3]  -24.95    0.08 4.52  -35.26  -27.53  -24.48  -21.81  -17.66
#> log_p_z[28,4]  -21.98    0.07 4.31  -31.83  -24.54  -21.47  -18.94  -14.96
#> log_p_z[29,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_p_z[29,2]  -20.02    0.04 2.30  -24.99  -21.44  -19.87  -18.40  -15.88
#> log_p_z[29,3]  -24.95    0.08 4.52  -35.26  -27.53  -24.48  -21.81  -17.66
#> log_p_z[29,4]  -21.98    0.07 4.31  -31.83  -24.54  -21.47  -18.94  -14.96
#> log_p_z[30,1]   -6.82    0.01 0.83   -8.68   -7.30   -6.73   -6.22   -5.46
#> log_p_z[30,2]  -14.72    0.03 1.70  -18.39  -15.77  -14.61  -13.54  -11.73
#> log_p_z[30,3]  -24.44    0.06 3.69  -32.43  -26.63  -24.20  -21.82  -18.10
#> log_p_z[30,4]  -21.94    0.06 3.57  -30.00  -24.09  -21.68  -19.39  -15.96
#> log_p_z[31,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_p_z[31,2]  -20.02    0.04 2.30  -24.99  -21.44  -19.87  -18.40  -15.88
#> log_p_z[31,3]  -24.95    0.08 4.52  -35.26  -27.53  -24.48  -21.81  -17.66
#> log_p_z[31,4]  -21.98    0.07 4.31  -31.83  -24.54  -21.47  -18.94  -14.96
#> log_p_z[32,1]  -29.72    0.07 4.33  -39.38  -32.21  -29.31  -26.72  -22.51
#> log_p_z[32,2]  -14.81    0.02 1.75  -18.68  -15.87  -14.65  -13.58  -11.79
#> log_p_z[32,3]   -6.06    0.01 0.95   -8.20   -6.62   -6.00   -5.40   -4.44
#> log_p_z[32,4]  -20.68    0.08 4.35  -30.72  -23.18  -20.15  -17.62  -13.68
#> log_p_z[33,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_p_z[33,2]  -20.02    0.04 2.30  -24.99  -21.44  -19.87  -18.40  -15.88
#> log_p_z[33,3]  -24.95    0.08 4.52  -35.26  -27.53  -24.48  -21.81  -17.66
#> log_p_z[33,4]  -21.98    0.07 4.31  -31.83  -24.54  -21.47  -18.94  -14.96
#> log_p_z[34,1]  -19.62    0.04 2.38  -24.80  -21.15  -19.42  -17.94  -15.46
#> log_p_z[34,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.79   -2.38
#> log_p_z[34,3]  -21.40    0.08 4.63  -32.39  -23.94  -20.77  -18.11  -14.23
#> log_p_z[34,4]  -21.93    0.07 4.30  -31.67  -24.48  -21.42  -18.84  -14.91
#> log_p_z[35,1]  -22.59    0.04 2.69  -28.32  -24.30  -22.36  -20.76  -17.80
#> log_p_z[35,2]   -5.51    0.01 0.69   -6.97   -5.94   -5.45   -5.04   -4.33
#> log_p_z[35,3]  -16.28    0.08 4.38  -26.71  -18.61  -15.72  -13.09   -9.75
#> log_p_z[35,4]  -21.97    0.07 4.40  -31.74  -24.54  -21.43  -18.89  -14.58
#> log_p_z[36,1]  -29.71    0.06 3.50  -37.23  -31.87  -29.51  -27.28  -23.57
#> log_p_z[36,2]  -21.90    0.04 2.32  -26.99  -23.33  -21.70  -20.30  -17.81
#> log_p_z[36,3]  -10.52    0.03 1.70  -14.55  -11.44  -10.24   -9.34   -7.96
#> log_p_z[36,4]  -13.98    0.06 3.19  -21.15  -15.90  -13.67  -11.65   -8.76
#> log_p_z[37,1]  -18.53    0.04 2.28  -23.36  -19.96  -18.35  -16.95  -14.55
#> log_p_z[37,2]   -7.61    0.01 0.80   -9.36   -8.10   -7.55   -7.05   -6.23
#> log_p_z[37,3]  -19.30    0.06 3.48  -27.32  -21.32  -18.94  -16.83  -13.75
#> log_p_z[37,4]  -21.94    0.06 3.53  -29.61  -24.13  -21.67  -19.47  -15.84
#> log_p_z[38,1]  -25.21    0.04 2.84  -31.46  -26.99  -25.01  -23.24  -20.29
#> log_p_z[38,2]   -9.59    0.01 1.00  -11.78  -10.21   -9.49   -8.87   -7.88
#> log_p_z[38,3]  -11.84    0.05 2.95  -18.65  -13.45  -11.47   -9.67   -7.43
#> log_p_z[38,4]  -21.37    0.06 3.58  -29.46  -23.50  -21.03  -18.86  -15.37
#> log_p_z[39,1]  -29.37    0.07 4.23  -38.83  -31.84  -28.99  -26.40  -22.29
#> log_p_z[39,2]  -16.81    0.03 2.00  -21.06  -18.06  -16.64  -15.39  -13.31
#> log_p_z[39,3]   -7.28    0.01 1.06   -9.49   -7.96   -7.20   -6.55   -5.41
#> log_p_z[39,4]  -18.79    0.07 4.17  -28.15  -21.10  -18.22  -15.90  -12.26
#> log_p_z[40,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_p_z[40,2]  -20.02    0.04 2.30  -24.99  -21.44  -19.87  -18.40  -15.88
#> log_p_z[40,3]  -24.95    0.08 4.52  -35.26  -27.53  -24.48  -21.81  -17.66
#> log_p_z[40,4]  -21.98    0.07 4.31  -31.83  -24.54  -21.47  -18.94  -14.96
#> log_p_z[41,1]   -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_p_z[41,2]  -20.02    0.04 2.30  -24.99  -21.44  -19.87  -18.40  -15.88
#> log_p_z[41,3]  -24.95    0.08 4.52  -35.26  -27.53  -24.48  -21.81  -17.66
#> log_p_z[41,4]  -21.98    0.07 4.31  -31.83  -24.54  -21.47  -18.94  -14.96
#> log_p_z[42,1]   -6.00    0.01 0.60   -7.32   -6.37   -5.96   -5.58   -4.95
#> log_p_z[42,2]  -14.97    0.03 1.73  -18.79  -16.03  -14.85  -13.75  -11.94
#> log_p_z[42,3]  -24.44    0.06 3.77  -32.55  -26.70  -24.09  -21.82  -18.08
#> log_p_z[42,4]  -21.98    0.06 3.62  -29.99  -24.18  -21.66  -19.43  -15.91
#> log_p_z[43,1]  -21.88    0.04 2.47  -27.29  -23.46  -21.64  -20.16  -17.64
#> log_p_z[43,2]   -5.65    0.01 0.53   -6.80   -5.98   -5.61   -5.28   -4.73
#> log_p_z[43,3]  -18.51    0.06 3.39  -26.25  -20.42  -18.12  -16.15  -13.12
#> log_p_z[43,4]  -21.58    0.06 3.46  -28.98  -23.65  -21.30  -19.10  -15.76
#> log_p_z[44,1]   -4.41    0.00 0.39   -5.24   -4.66   -4.39   -4.14   -3.72
#> log_p_z[44,2]  -17.25    0.03 1.85  -21.28  -18.40  -17.12  -15.97  -13.90
#> log_p_z[44,3]  -24.43    0.06 3.69  -32.51  -26.64  -24.19  -21.91  -18.03
#> log_p_z[44,4]  -21.98    0.06 3.65  -30.17  -24.19  -21.69  -19.38  -15.85
#> log_p_z[45,1]  -19.62    0.04 2.38  -24.80  -21.15  -19.42  -17.94  -15.46
#> log_p_z[45,2]   -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.79   -2.38
#> log_p_z[45,3]  -21.40    0.08 4.63  -32.39  -23.94  -20.77  -18.11  -14.23
#> log_p_z[45,4]  -21.93    0.07 4.30  -31.67  -24.48  -21.42  -18.84  -14.91
#> log_lik[1]      -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_lik[2]      -7.59    0.02 1.13  -10.04   -8.31   -7.52   -6.80   -5.61
#> log_lik[3]      -9.96    0.01 0.94  -11.99  -10.57   -9.92   -9.31   -8.23
#> log_lik[4]      -7.57    0.01 0.92   -9.56   -8.15   -7.49   -6.91   -5.98
#> log_lik[5]      -3.70    0.00 0.43   -4.63   -3.97   -3.68   -3.41   -2.92
#> log_lik[6]      -5.46    0.01 0.66   -6.89   -5.87   -5.42   -5.00   -4.28
#> log_lik[7]      -6.00    0.01 0.60   -7.32   -6.37   -5.95   -5.58   -4.95
#> log_lik[8]      -5.73    0.01 0.83   -7.50   -6.23   -5.66   -5.15   -4.26
#> log_lik[9]      -4.51    0.01 0.58   -5.74   -4.87   -4.47   -4.11   -3.51
#> log_lik[10]     -7.12    0.01 0.72   -8.64   -7.58   -7.06   -6.60   -5.86
#> log_lik[11]     -5.82    0.01 0.97   -7.94   -6.42   -5.75   -5.15   -4.16
#> log_lik[12]     -9.32    0.01 1.07  -11.61  -10.02   -9.24   -8.57   -7.41
#> log_lik[13]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_lik[14]     -6.23    0.01 0.85   -8.08   -6.74   -6.16   -5.63   -4.78
#> log_lik[15]     -4.41    0.00 0.39   -5.24   -4.66   -4.39   -4.14   -3.72
#> log_lik[16]     -3.80    0.01 0.54   -4.97   -4.14   -3.77   -3.42   -2.85
#> log_lik[17]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_lik[18]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_lik[19]     -4.65    0.01 0.58   -5.88   -5.01   -4.62   -4.26   -3.62
#> log_lik[20]     -7.08    0.01 0.97   -9.18   -7.66   -7.01   -6.41   -5.42
#> log_lik[21]     -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.79   -2.38
#> log_lik[22]     -4.65    0.01 0.58   -5.88   -5.01   -4.62   -4.26   -3.62
#> log_lik[23]     -3.70    0.00 0.43   -4.63   -3.97   -3.68   -3.41   -2.92
#> log_lik[24]     -5.81    0.01 0.64   -7.25   -6.18   -5.74   -5.35   -4.77
#> log_lik[25]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_lik[26]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_lik[27]     -5.65    0.01 0.53   -6.80   -5.98   -5.61   -5.28   -4.73
#> log_lik[28]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_lik[29]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_lik[30]     -6.81    0.01 0.83   -8.68   -7.30   -6.73   -6.22   -5.46
#> log_lik[31]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_lik[32]     -6.06    0.01 0.95   -8.20   -6.62   -6.00   -5.40   -4.44
#> log_lik[33]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_lik[34]     -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.79   -2.38
#> log_lik[35]     -5.51    0.01 0.68   -6.96   -5.93   -5.45   -5.04   -4.33
#> log_lik[36]     -9.94    0.02 1.17  -12.38  -10.67   -9.86   -9.13   -7.88
#> log_lik[37]     -7.61    0.01 0.80   -9.36   -8.10   -7.54   -7.05   -6.23
#> log_lik[38]     -8.99    0.01 0.88  -10.74   -9.55   -8.98   -8.41   -7.31
#> log_lik[39]     -7.28    0.01 1.06   -9.49   -7.96   -7.20   -6.55   -5.41
#> log_lik[40]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_lik[41]     -2.21    0.00 0.30   -2.84   -2.40   -2.20   -2.01   -1.68
#> log_lik[42]     -6.00    0.01 0.60   -7.32   -6.37   -5.95   -5.58   -4.95
#> log_lik[43]     -5.65    0.01 0.53   -6.80   -5.98   -5.61   -5.28   -4.73
#> log_lik[44]     -4.41    0.00 0.39   -5.24   -4.66   -4.39   -4.14   -3.72
#> log_lik[45]     -3.04    0.00 0.36   -3.78   -3.27   -3.02   -2.79   -2.38
#> lp__          -464.17    0.16 6.16 -477.45 -468.03 -463.86 -459.82 -453.17
#>               n_eff Rhat
#> pi[1]          9684    1
#> pi[2]          8122    1
#> pi[3]          6478    1
#> pi[4]          5914    1
#> theta[1,1,1]   9536    1
#> theta[1,1,2]   7921    1
#> theta[1,1,3]   4954    1
#> theta[1,1,4]   4922    1
#> theta[1,2,1]   5751    1
#> theta[1,2,2]   7573    1
#> theta[1,2,3]   7816    1
#> theta[1,2,4]   5499    1
#> theta[1,3,1]   7515    1
#> theta[1,3,2]   4866    1
#> theta[1,3,3]   5445    1
#> theta[1,3,4]   6105    1
#> theta[1,4,1]   6250    1
#> theta[1,4,2]   6914    1
#> theta[1,4,3]   4389    1
#> theta[1,4,4]   4773    1
#> theta[2,1,1]   8816    1
#> theta[2,1,2]   7246    1
#> theta[2,1,3]   4569    1
#> theta[2,1,4]   5126    1
#> theta[2,2,1]   6694    1
#> theta[2,2,2]   7762    1
#> theta[2,2,3]   7312    1
#> theta[2,2,4]   5696    1
#> theta[2,3,1]   5874    1
#> theta[2,3,2]   8445    1
#> theta[2,3,3]   6959    1
#> theta[2,3,4]   6194    1
#> theta[2,4,1]   7245    1
#> theta[2,4,2]   7304    1
#> theta[2,4,3]   6268    1
#> theta[2,4,4]   7115    1
#> theta[3,1,1]   8492    1
#> theta[3,1,2]   6174    1
#> theta[3,1,3]   4937    1
#> theta[3,1,4]   4912    1
#> theta[3,2,1]   5897    1
#> theta[3,2,2]   7811    1
#> theta[3,2,3]   8836    1
#> theta[3,2,4]   4947    1
#> theta[3,3,1]   7535    1
#> theta[3,3,2]   8026    1
#> theta[3,3,3]   8079    1
#> theta[3,3,4]   8331    1
#> theta[3,4,1]   8256    1
#> theta[3,4,2]   7059    1
#> theta[3,4,3]   5238    1
#> theta[3,4,4]   6513    1
#> theta[4,1,1]   8906    1
#> theta[4,1,2]   6372    1
#> theta[4,1,3]   5093    1
#> theta[4,1,4]   4767    1
#> theta[4,2,1]   7855    1
#> theta[4,2,2]   7328    1
#> theta[4,2,3]   6165    1
#> theta[4,2,4]   6104    1
#> theta[4,3,1]   6113    1
#> theta[4,3,2]   6655    1
#> theta[4,3,3]   5808    1
#> theta[4,3,4]   6385    1
#> theta[4,4,1]   7528    1
#> theta[4,4,2]   7503    1
#> theta[4,4,3]   5985    1
#> theta[4,4,4]   6687    1
#> theta[5,1,1]   7542    1
#> theta[5,1,2]   4783    1
#> theta[5,1,3]   5161    1
#> theta[5,1,4]   4989    1
#> theta[5,2,1]   9935    1
#> theta[5,2,2]   7879    1
#> theta[5,2,3]   6832    1
#> theta[5,2,4]   5408    1
#> theta[5,3,1]   7046    1
#> theta[5,3,2]   8117    1
#> theta[5,3,3]   7021    1
#> theta[5,3,4]   6925    1
#> theta[5,4,1]   7964    1
#> theta[5,4,2]   6826    1
#> theta[5,4,3]   6594    1
#> theta[5,4,4]   7019    1
#> log_p_z[1,1]   8953    1
#> log_p_z[1,2]   4112    1
#> log_p_z[1,3]   3526    1
#> log_p_z[1,4]   3324    1
#> log_p_z[2,1]   3936    1
#> log_p_z[2,2]   4010    1
#> log_p_z[2,3]   4301    1
#> log_p_z[2,4]   2845    1
#> log_p_z[3,1]   4114    1
#> log_p_z[3,2]   3669    1
#> log_p_z[3,3]   3786    1
#> log_p_z[3,4]   3180    1
#> log_p_z[4,1]   5463    1
#> log_p_z[4,2]   4630    1
#> log_p_z[4,3]   3213    1
#> log_p_z[4,4]   3891    1
#> log_p_z[5,1]   4403    1
#> log_p_z[5,2]   7734    1
#> log_p_z[5,3]   3425    1
#> log_p_z[5,4]   3736    1
#> log_p_z[6,1]   4580    1
#> log_p_z[6,2]   7154    1
#> log_p_z[6,3]   3312    1
#> log_p_z[6,4]   3744    1
#> log_p_z[7,1]   7814    1
#> log_p_z[7,2]   4274    1
#> log_p_z[7,3]   3615    1
#> log_p_z[7,4]   3226    1
#> log_p_z[8,1]   4028    1
#> log_p_z[8,2]   4985    1
#> log_p_z[8,3]   6739    1
#> log_p_z[8,4]   3273    1
#> log_p_z[9,1]   4863    1
#> log_p_z[9,2]   6024    1
#> log_p_z[9,3]   3391    1
#> log_p_z[9,4]   3617    1
#> log_p_z[10,1]  4658    1
#> log_p_z[10,2]  5675    1
#> log_p_z[10,3]  3478    1
#> log_p_z[10,4]  3344    1
#> log_p_z[11,1]  5252    1
#> log_p_z[11,2]  3301    1
#> log_p_z[11,3]  3071    1
#> log_p_z[11,4]  6279    1
#> log_p_z[12,1]  4762    1
#> log_p_z[12,2]  3612    1
#> log_p_z[12,3]  3172    1
#> log_p_z[12,4]  3652    1
#> log_p_z[13,1]  8953    1
#> log_p_z[13,2]  4112    1
#> log_p_z[13,3]  3526    1
#> log_p_z[13,4]  3324    1
#> log_p_z[14,1]  4442    1
#> log_p_z[14,2]  5929    1
#> log_p_z[14,3]  3533    1
#> log_p_z[14,4]  3963    1
#> log_p_z[15,1]  6353    1
#> log_p_z[15,2]  4187    1
#> log_p_z[15,3]  3607    1
#> log_p_z[15,4]  3369    1
#> log_p_z[16,1]  6881    1
#> log_p_z[16,2]  4203    1
#> log_p_z[16,3]  3533    1
#> log_p_z[16,4]  3223    1
#> log_p_z[17,1]  8953    1
#> log_p_z[17,2]  4112    1
#> log_p_z[17,3]  3526    1
#> log_p_z[17,4]  3324    1
#> log_p_z[18,1]  8953    1
#> log_p_z[18,2]  4112    1
#> log_p_z[18,3]  3526    1
#> log_p_z[18,4]  3324    1
#> log_p_z[19,1]  5376    1
#> log_p_z[19,2]  6736    1
#> log_p_z[19,3]  3353    1
#> log_p_z[19,4]  3758    1
#> log_p_z[20,1]  4886    1
#> log_p_z[20,2]  6026    1
#> log_p_z[20,3]  3410    1
#> log_p_z[20,4]  3677    1
#> log_p_z[21,1]  4590    1
#> log_p_z[21,2]  7785    1
#> log_p_z[21,3]  3450    1
#> log_p_z[21,4]  3679    1
#> log_p_z[22,1]  5376    1
#> log_p_z[22,2]  6736    1
#> log_p_z[22,3]  3353    1
#> log_p_z[22,4]  3758    1
#> log_p_z[23,1]  4403    1
#> log_p_z[23,2]  7734    1
#> log_p_z[23,3]  3425    1
#> log_p_z[23,4]  3736    1
#> log_p_z[24,1]  4214    1
#> log_p_z[24,2]  4331    1
#> log_p_z[24,3]  3775    1
#> log_p_z[24,4]  3430    1
#> log_p_z[25,1]  8953    1
#> log_p_z[25,2]  4112    1
#> log_p_z[25,3]  3526    1
#> log_p_z[25,4]  3324    1
#> log_p_z[26,1]  8953    1
#> log_p_z[26,2]  4112    1
#> log_p_z[26,3]  3526    1
#> log_p_z[26,4]  3324    1
#> log_p_z[27,1]  4543    1
#> log_p_z[27,2]  7097    1
#> log_p_z[27,3]  3540    1
#> log_p_z[27,4]  3483    1
#> log_p_z[28,1]  8953    1
#> log_p_z[28,2]  4112    1
#> log_p_z[28,3]  3526    1
#> log_p_z[28,4]  3324    1
#> log_p_z[29,1]  8953    1
#> log_p_z[29,2]  4112    1
#> log_p_z[29,3]  3526    1
#> log_p_z[29,4]  3324    1
#> log_p_z[30,1]  6285    1
#> log_p_z[30,2]  3918    1
#> log_p_z[30,3]  3605    1
#> log_p_z[30,4]  3455    1
#> log_p_z[31,1]  8953    1
#> log_p_z[31,2]  4112    1
#> log_p_z[31,3]  3526    1
#> log_p_z[31,4]  3324    1
#> log_p_z[32,1]  4003    1
#> log_p_z[32,2]  5718    1
#> log_p_z[32,3]  5726    1
#> log_p_z[32,4]  3043    1
#> log_p_z[33,1]  8953    1
#> log_p_z[33,2]  4112    1
#> log_p_z[33,3]  3526    1
#> log_p_z[33,4]  3324    1
#> log_p_z[34,1]  4590    1
#> log_p_z[34,2]  7785    1
#> log_p_z[34,3]  3450    1
#> log_p_z[34,4]  3679    1
#> log_p_z[35,1]  4388    1
#> log_p_z[35,2]  5964    1
#> log_p_z[35,3]  3410    1
#> log_p_z[35,4]  3918    1
#> log_p_z[36,1]  3739    1
#> log_p_z[36,2]  3321    1
#> log_p_z[36,3]  2830    1
#> log_p_z[36,4]  2943    1
#> log_p_z[37,1]  4134    1
#> log_p_z[37,2]  5256    1
#> log_p_z[37,3]  3665    1
#> log_p_z[37,4]  3611    1
#> log_p_z[38,1]  4139    1
#> log_p_z[38,2]  5094    1
#> log_p_z[38,3]  3374    1
#> log_p_z[38,4]  3458    1
#> log_p_z[39,1]  3913    1
#> log_p_z[39,2]  5215    1
#> log_p_z[39,3]  5913    1
#> log_p_z[39,4]  3395    1
#> log_p_z[40,1]  8953    1
#> log_p_z[40,2]  4112    1
#> log_p_z[40,3]  3526    1
#> log_p_z[40,4]  3324    1
#> log_p_z[41,1]  8953    1
#> log_p_z[41,2]  4112    1
#> log_p_z[41,3]  3526    1
#> log_p_z[41,4]  3324    1
#> log_p_z[42,1]  7814    1
#> log_p_z[42,2]  4274    1
#> log_p_z[42,3]  3615    1
#> log_p_z[42,4]  3226    1
#> log_p_z[43,1]  4543    1
#> log_p_z[43,2]  7097    1
#> log_p_z[43,3]  3540    1
#> log_p_z[43,4]  3483    1
#> log_p_z[44,1]  6353    1
#> log_p_z[44,2]  4187    1
#> log_p_z[44,3]  3607    1
#> log_p_z[44,4]  3369    1
#> log_p_z[45,1]  4590    1
#> log_p_z[45,2]  7785    1
#> log_p_z[45,3]  3450    1
#> log_p_z[45,4]  3679    1
#> log_lik[1]     8953    1
#> log_lik[2]     5055    1
#> log_lik[3]     4782    1
#> log_lik[4]     4659    1
#> log_lik[5]     7734    1
#> log_lik[6]     7134    1
#> log_lik[7]     7820    1
#> log_lik[8]     6741    1
#> log_lik[9]     6025    1
#> log_lik[10]    5719    1
#> log_lik[11]    6279    1
#> log_lik[12]    5800    1
#> log_lik[13]    8953    1
#> log_lik[14]    5931    1
#> log_lik[15]    6353    1
#> log_lik[16]    6881    1
#> log_lik[17]    8953    1
#> log_lik[18]    8953    1
#> log_lik[19]    6736    1
#> log_lik[20]    6026    1
#> log_lik[21]    7785    1
#> log_lik[22]    6736    1
#> log_lik[23]    7734    1
#> log_lik[24]    4332    1
#> log_lik[25]    8953    1
#> log_lik[26]    8953    1
#> log_lik[27]    7098    1
#> log_lik[28]    8953    1
#> log_lik[29]    8953    1
#> log_lik[30]    6292    1
#> log_lik[31]    8953    1
#> log_lik[32]    5822    1
#> log_lik[33]    8953    1
#> log_lik[34]    7785    1
#> log_lik[35]    5966    1
#> log_lik[36]    5161    1
#> log_lik[37]    5259    1
#> log_lik[38]    4973    1
#> log_lik[39]    5913    1
#> log_lik[40]    8953    1
#> log_lik[41]    8953    1
#> log_lik[42]    7820    1
#> log_lik[43]    7098    1
#> log_lik[44]    6353    1
#> log_lik[45]    7785    1
#> lp__           1461    1
#> 
#> Samples were drawn using NUTS(diag_e) at Mon Feb 22 06:12:23 2021.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# }