Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000134 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.34 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.55356 seconds (Warm-up)
#> Chain 1:                1.64978 seconds (Sampling)
#> Chain 1:                3.20334 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000122 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.22 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.53535 seconds (Warm-up)
#> Chain 2:                1.64324 seconds (Sampling)
#> Chain 2:                3.17858 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000126 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.26 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.51364 seconds (Warm-up)
#> Chain 3:                1.64099 seconds (Sampling)
#> Chain 3:                3.15463 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000115 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.15 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.48337 seconds (Warm-up)
#> Chain 4:                1.63166 seconds (Sampling)
#> Chain 4:                3.11503 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat  ess_bulk
#> pi[1]          1.0018772  9640.628
#> pi[2]          1.0011227 10379.485
#> pi[3]          1.0007652  8098.768
#> pi[4]          1.0004089  8219.796
#> theta[1, 1, 1] 1.0004960 11226.172
#> theta[1, 1, 2] 1.0001067  9693.190
#> theta[1, 1, 3] 0.9996171  6944.115
#> theta[1, 1, 4] 1.0013887  7241.859
#> theta[1, 2, 1] 1.0010634  7062.616
#> theta[1, 2, 2] 1.0021554  9522.237
#> theta[1, 2, 3] 1.0001848  7102.991
#> theta[1, 2, 4] 0.9997614  6275.745
#> theta[1, 3, 1] 1.0001138  6408.479
#> theta[1, 3, 2] 0.9999573  4396.659
#> theta[1, 3, 3] 0.9997496  5013.110
#> theta[1, 3, 4] 1.0016809  5928.509
#> theta[1, 4, 1] 1.0001350  5978.142
#> theta[1, 4, 2] 0.9999986  5838.079
#> theta[1, 4, 3] 0.9996901  3780.291
#> theta[1, 4, 4] 0.9998732  5313.090
#> theta[2, 1, 1] 1.0007597  8934.208
#> theta[2, 1, 2] 1.0005080  7686.330
#> theta[2, 1, 3] 1.0004285  6450.460
#> theta[2, 1, 4] 1.0001520  6086.598
#> theta[2, 2, 1] 1.0015437  8636.011
#> theta[2, 2, 2] 1.0005881  8525.950
#> theta[2, 2, 3] 1.0012044  7496.677
#> theta[2, 2, 4] 1.0006901  6819.765
#> theta[2, 3, 1] 1.0005636  6437.742
#> theta[2, 3, 2] 0.9999987  5591.655
#> theta[2, 3, 3] 1.0009118  5488.468
#> theta[2, 3, 4] 1.0008326  5421.731
#> theta[2, 4, 1] 1.0000393  6376.605
#> theta[2, 4, 2] 1.0005367  6655.282
#> theta[2, 4, 3] 0.9999047  6513.827
#> theta[2, 4, 4] 1.0000473  7515.992
#> theta[3, 1, 1] 1.0006785  6427.312
#> theta[3, 1, 2] 1.0007053  6034.479
#> theta[3, 1, 3] 1.0009268  5927.501
#> theta[3, 1, 4] 1.0005383  6318.990
#> theta[3, 2, 1] 1.0000220  6985.530
#> theta[3, 2, 2] 0.9998636  9463.889
#> theta[3, 2, 3] 1.0008554  8348.573
#> theta[3, 2, 4] 0.9995159  6158.305
#> theta[3, 3, 1] 1.0005431  6723.558
#> theta[3, 3, 2] 1.0006479  8107.567
#> theta[3, 3, 3] 1.0002820  8697.684
#> theta[3, 3, 4] 0.9995206  7722.702
#> theta[3, 4, 1] 0.9995752  5927.119
#> theta[3, 4, 2] 1.0001876  7072.050
#> theta[3, 4, 3] 1.0005437  4889.106
#> theta[3, 4, 4] 1.0007347  6957.308
#> theta[4, 1, 1] 1.0002746  7946.149
#> theta[4, 1, 2] 1.0028697  5964.100
#> theta[4, 1, 3] 1.0015784  5714.863
#> theta[4, 1, 4] 0.9995662  6143.894
#> theta[4, 2, 1] 1.0009303  7483.041
#> theta[4, 2, 2] 0.9999074  8342.918
#> theta[4, 2, 3] 1.0010217  8349.438
#> theta[4, 2, 4] 1.0001898  6061.107
#> theta[4, 3, 1] 0.9999078  6338.011
#> theta[4, 3, 2] 0.9999167  6660.273
#> theta[4, 3, 3] 1.0009025  5879.793
#> theta[4, 3, 4] 1.0007934  5371.257
#> theta[4, 4, 1] 1.0005154  6824.342
#> theta[4, 4, 2] 0.9995578  5898.384
#> theta[4, 4, 3] 1.0005173  5678.001
#> theta[4, 4, 4] 1.0018151  7109.773
#> theta[5, 1, 1] 1.0020188  8113.787
#> theta[5, 1, 2] 1.0013806  6212.818
#> theta[5, 1, 3] 1.0011763  7326.044
#> theta[5, 1, 4] 1.0002030  6275.922
#> theta[5, 2, 1] 1.0009576  8072.618
#> theta[5, 2, 2] 1.0029094  7646.261
#> theta[5, 2, 3] 1.0031403  7286.858
#> theta[5, 2, 4] 1.0005625  4872.543
#> theta[5, 3, 1] 0.9999253  6226.951
#> theta[5, 3, 2] 1.0006881  8153.038
#> theta[5, 3, 3] 0.9996804  7006.427
#> theta[5, 3, 4] 0.9998680  7852.547
#> theta[5, 4, 1] 0.9998770  5647.808
#> theta[5, 4, 2] 1.0012864  6744.920
#> theta[5, 4, 3] 0.9994104  5990.774
#> theta[5, 4, 4] 1.0004502  6503.238

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat  ess_bulk
#> pi[1] 1.001877  9640.628
#> pi[2] 1.001123 10379.485
#> pi[3] 1.000765  8098.768
#> pi[4] 1.000409  8219.796

# }