Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000143 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.43 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.0834 seconds (Warm-up)
#> Chain 1:                2.28783 seconds (Sampling)
#> Chain 1:                4.37124 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000134 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.34 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.99202 seconds (Warm-up)
#> Chain 2:                2.21837 seconds (Sampling)
#> Chain 2:                4.21039 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000142 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.42 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.07492 seconds (Warm-up)
#> Chain 3:                2.32337 seconds (Sampling)
#> Chain 3:                4.39829 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000126 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.26 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.12598 seconds (Warm-up)
#> Chain 4:                2.29045 seconds (Sampling)
#> Chain 4:                4.41643 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat  ess_bulk
#> pi[1]          1.0003939  9931.424
#> pi[2]          1.0006485  7911.589
#> pi[3]          1.0004866  8217.074
#> pi[4]          1.0033312  7915.869
#> theta[1, 1, 1] 1.0040408  9705.009
#> theta[1, 1, 2] 1.0011356  8918.971
#> theta[1, 1, 3] 0.9997529  6158.776
#> theta[1, 1, 4] 1.0020599  6423.111
#> theta[1, 2, 1] 1.0018304  6046.334
#> theta[1, 2, 2] 1.0011913  7955.178
#> theta[1, 2, 3] 1.0013075  8368.016
#> theta[1, 2, 4] 1.0000667  5820.604
#> theta[1, 3, 1] 1.0011743  6088.170
#> theta[1, 3, 2] 1.0005316  4518.724
#> theta[1, 3, 3] 1.0009326  5035.021
#> theta[1, 3, 4] 1.0014656  4846.980
#> theta[1, 4, 1] 1.0010640  6046.825
#> theta[1, 4, 2] 1.0013675  7003.824
#> theta[1, 4, 3] 0.9998640  4120.579
#> theta[1, 4, 4] 1.0001313  5650.941
#> theta[2, 1, 1] 1.0017056  8127.316
#> theta[2, 1, 2] 1.0002332  7551.258
#> theta[2, 1, 3] 1.0004737  6618.208
#> theta[2, 1, 4] 0.9999945  6588.982
#> theta[2, 2, 1] 1.0004627  7856.864
#> theta[2, 2, 2] 1.0001388  8106.162
#> theta[2, 2, 3] 1.0030417  8492.760
#> theta[2, 2, 4] 0.9997656  5411.575
#> theta[2, 3, 1] 1.0000808  6251.700
#> theta[2, 3, 2] 1.0015915  6234.036
#> theta[2, 3, 3] 1.0035556  6955.731
#> theta[2, 3, 4] 1.0001880  6516.945
#> theta[2, 4, 1] 1.0012222  6086.601
#> theta[2, 4, 2] 1.0006956  7547.592
#> theta[2, 4, 3] 1.0001836  5865.894
#> theta[2, 4, 4] 1.0022888  8369.307
#> theta[3, 1, 1] 1.0018130  8359.670
#> theta[3, 1, 2] 1.0003357  6241.006
#> theta[3, 1, 3] 1.0000175  6660.439
#> theta[3, 1, 4] 1.0003812  6049.136
#> theta[3, 2, 1] 1.0013843  6400.274
#> theta[3, 2, 2] 1.0013219 10782.498
#> theta[3, 2, 3] 1.0024564  8292.328
#> theta[3, 2, 4] 1.0021888  6350.787
#> theta[3, 3, 1] 0.9996422  5741.167
#> theta[3, 3, 2] 1.0002371  7118.982
#> theta[3, 3, 3] 1.0006758  7809.325
#> theta[3, 3, 4] 1.0019713  7420.321
#> theta[3, 4, 1] 0.9997570  6274.822
#> theta[3, 4, 2] 1.0005672  4846.198
#> theta[3, 4, 3] 1.0005031  4687.873
#> theta[3, 4, 4] 0.9994233  5889.374
#> theta[4, 1, 1] 1.0004244  6884.310
#> theta[4, 1, 2] 1.0019892  7038.862
#> theta[4, 1, 3] 1.0002150  6216.582
#> theta[4, 1, 4] 1.0027860  6498.317
#> theta[4, 2, 1] 1.0003109  7783.534
#> theta[4, 2, 2] 1.0053587  8061.489
#> theta[4, 2, 3] 1.0005758  6669.153
#> theta[4, 2, 4] 1.0001487  6282.819
#> theta[4, 3, 1] 1.0017649  6393.098
#> theta[4, 3, 2] 1.0008080  6244.881
#> theta[4, 3, 3] 1.0009831  6325.952
#> theta[4, 3, 4] 1.0015791  5421.591
#> theta[4, 4, 1] 1.0001957  5651.913
#> theta[4, 4, 2] 1.0003011  5865.435
#> theta[4, 4, 3] 0.9999560  6331.045
#> theta[4, 4, 4] 1.0000529  7173.956
#> theta[5, 1, 1] 1.0032750  7596.514
#> theta[5, 1, 2] 0.9999620  5390.907
#> theta[5, 1, 3] 1.0004257  6178.982
#> theta[5, 1, 4] 1.0007905  5509.838
#> theta[5, 2, 1] 0.9999808  8778.990
#> theta[5, 2, 2] 1.0008929  7980.032
#> theta[5, 2, 3] 1.0004678  6895.497
#> theta[5, 2, 4] 1.0014045  5974.247
#> theta[5, 3, 1] 0.9998475  5591.253
#> theta[5, 3, 2] 0.9998973  8245.183
#> theta[5, 3, 3] 0.9998523  7603.230
#> theta[5, 3, 4] 1.0028701  6803.217
#> theta[5, 4, 1] 1.0015060  6923.096
#> theta[5, 4, 2] 0.9998607  6082.040
#> theta[5, 4, 3] 1.0007506  5667.864
#> theta[5, 4, 4] 1.0005439  7204.008

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.000394 9931.424
#> pi[2] 1.000649 7911.589
#> pi[3] 1.000487 8217.074
#> pi[4] 1.003331 7915.869

# }