Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000145 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.45 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.6363 seconds (Warm-up)
#> Chain 1:                1.82663 seconds (Sampling)
#> Chain 1:                3.46293 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000106 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.06 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.76814 seconds (Warm-up)
#> Chain 2:                1.56354 seconds (Sampling)
#> Chain 2:                3.33168 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.00012 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.2 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.67798 seconds (Warm-up)
#> Chain 3:                1.79199 seconds (Sampling)
#> Chain 3:                3.46998 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000137 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.37 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.77785 seconds (Warm-up)
#> Chain 4:                1.8873 seconds (Sampling)
#> Chain 4:                3.66515 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0002623 7692.217
#> pi[2]          1.0003821 7186.820
#> pi[3]          1.0007438 6992.184
#> pi[4]          1.0000679 6041.732
#> theta[1, 1, 1] 1.0029058 7173.153
#> theta[1, 1, 2] 1.0011767 7128.984
#> theta[1, 1, 3] 1.0000203 5353.550
#> theta[1, 1, 4] 1.0020233 5079.426
#> theta[1, 2, 1] 0.9998230 4965.089
#> theta[1, 2, 2] 1.0000636 6338.111
#> theta[1, 2, 3] 1.0007798 6583.236
#> theta[1, 2, 4] 1.0004001 5453.121
#> theta[1, 3, 1] 1.0006372 4612.193
#> theta[1, 3, 2] 1.0001284 4497.761
#> theta[1, 3, 3] 0.9996027 5334.566
#> theta[1, 3, 4] 1.0000760 4830.036
#> theta[1, 4, 1] 0.9999822 5599.735
#> theta[1, 4, 2] 0.9995057 4637.758
#> theta[1, 4, 3] 1.0013041 3762.303
#> theta[1, 4, 4] 1.0012963 4860.989
#> theta[2, 1, 1] 1.0001637 8641.496
#> theta[2, 1, 2] 0.9999594 7398.649
#> theta[2, 1, 3] 1.0001356 5583.266
#> theta[2, 1, 4] 1.0024529 5573.900
#> theta[2, 2, 1] 1.0014266 6691.009
#> theta[2, 2, 2] 1.0019357 7446.053
#> theta[2, 2, 3] 1.0010433 7288.252
#> theta[2, 2, 4] 1.0011773 4803.263
#> theta[2, 3, 1] 1.0000011 5538.991
#> theta[2, 3, 2] 1.0025515 5868.009
#> theta[2, 3, 3] 1.0002370 6987.606
#> theta[2, 3, 4] 1.0006307 5748.263
#> theta[2, 4, 1] 1.0002111 4668.031
#> theta[2, 4, 2] 0.9999090 4415.287
#> theta[2, 4, 3] 1.0000623 4284.645
#> theta[2, 4, 4] 1.0008684 4979.169
#> theta[3, 1, 1] 0.9999419 6716.290
#> theta[3, 1, 2] 0.9999006 4960.469
#> theta[3, 1, 3] 1.0019452 4870.346
#> theta[3, 1, 4] 1.0000385 5351.900
#> theta[3, 2, 1] 1.0002752 5832.164
#> theta[3, 2, 2] 1.0004380 7048.396
#> theta[3, 2, 3] 1.0015759 6924.695
#> theta[3, 2, 4] 0.9996951 5551.923
#> theta[3, 3, 1] 0.9995078 4823.888
#> theta[3, 3, 2] 1.0022873 6304.237
#> theta[3, 3, 3] 1.0015146 5940.803
#> theta[3, 3, 4] 1.0004164 6867.331
#> theta[3, 4, 1] 1.0012572 5167.490
#> theta[3, 4, 2] 1.0002023 6033.612
#> theta[3, 4, 3] 1.0007485 3922.017
#> theta[3, 4, 4] 1.0005082 6028.774
#> theta[4, 1, 1] 1.0001271 7994.715
#> theta[4, 1, 2] 0.9997178 6050.859
#> theta[4, 1, 3] 1.0014549 6017.076
#> theta[4, 1, 4] 1.0000891 4647.012
#> theta[4, 2, 1] 0.9998660 6304.243
#> theta[4, 2, 2] 1.0010836 7996.992
#> theta[4, 2, 3] 1.0012850 6646.978
#> theta[4, 2, 4] 0.9997048 5212.168
#> theta[4, 3, 1] 1.0004372 5992.866
#> theta[4, 3, 2] 1.0009917 5915.622
#> theta[4, 3, 3] 0.9998594 5630.026
#> theta[4, 3, 4] 1.0008704 3877.260
#> theta[4, 4, 1] 1.0010622 5698.081
#> theta[4, 4, 2] 1.0028509 5223.481
#> theta[4, 4, 3] 1.0019974 5372.309
#> theta[4, 4, 4] 0.9998931 6711.452
#> theta[5, 1, 1] 1.0000144 6945.338
#> theta[5, 1, 2] 1.0005523 4682.984
#> theta[5, 1, 3] 1.0007936 5619.054
#> theta[5, 1, 4] 1.0002500 6652.775
#> theta[5, 2, 1] 1.0001021 7585.215
#> theta[5, 2, 2] 1.0004127 6989.828
#> theta[5, 2, 3] 1.0004342 5568.645
#> theta[5, 2, 4] 1.0001949 4556.444
#> theta[5, 3, 1] 1.0013624 4140.162
#> theta[5, 3, 2] 1.0008328 6316.559
#> theta[5, 3, 3] 1.0014914 6906.482
#> theta[5, 3, 4] 1.0012823 6868.741
#> theta[5, 4, 1] 1.0018571 4874.730
#> theta[5, 4, 2] 1.0015036 5648.574
#> theta[5, 4, 3] 1.0000791 4811.284
#> theta[5, 4, 4] 1.0000640 6230.773

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.000262 7692.217
#> pi[2] 1.000382 7186.820
#> pi[3] 1.000744 6992.184
#> pi[4] 1.000068 6041.732

# }