Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000155 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.55 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.46391 seconds (Warm-up)
#> Chain 1:                1.10454 seconds (Sampling)
#> Chain 1:                2.56845 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000122 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.22 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.46198 seconds (Warm-up)
#> Chain 2:                1.59188 seconds (Sampling)
#> Chain 2:                3.05386 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000119 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.19 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.50788 seconds (Warm-up)
#> Chain 3:                1.64468 seconds (Sampling)
#> Chain 3:                3.15255 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000114 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.14 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.53474 seconds (Warm-up)
#> Chain 4:                1.25035 seconds (Sampling)
#> Chain 4:                2.78509 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0008599 6586.315
#> pi[2]          1.0001683 5796.556
#> pi[3]          1.0009372 4845.750
#> pi[4]          1.0002907 5154.161
#> theta[1, 1, 1] 1.0002872 5998.798
#> theta[1, 1, 2] 1.0005564 5594.368
#> theta[1, 1, 3] 1.0004757 4934.885
#> theta[1, 1, 4] 1.0000957 4561.739
#> theta[1, 2, 1] 1.0005099 4776.954
#> theta[1, 2, 2] 1.0010448 5534.348
#> theta[1, 2, 3] 0.9994393 6049.545
#> theta[1, 2, 4] 1.0000388 5222.542
#> theta[1, 3, 1] 1.0020987 4909.953
#> theta[1, 3, 2] 0.9999557 3477.143
#> theta[1, 3, 3] 0.9994776 4756.239
#> theta[1, 3, 4] 0.9999856 4629.692
#> theta[1, 4, 1] 0.9995785 4721.616
#> theta[1, 4, 2] 1.0006827 4793.994
#> theta[1, 4, 3] 1.0009636 3264.315
#> theta[1, 4, 4] 1.0011251 5151.135
#> theta[2, 1, 1] 1.0022578 5528.300
#> theta[2, 1, 2] 1.0014275 5628.789
#> theta[2, 1, 3] 1.0028413 5054.587
#> theta[2, 1, 4] 1.0010529 4463.079
#> theta[2, 2, 1] 1.0002583 5933.605
#> theta[2, 2, 2] 1.0001063 6754.974
#> theta[2, 2, 3] 1.0036902 6515.113
#> theta[2, 2, 4] 1.0001838 4930.567
#> theta[2, 3, 1] 1.0001233 5037.642
#> theta[2, 3, 2] 1.0012928 4591.624
#> theta[2, 3, 3] 1.0003245 5057.565
#> theta[2, 3, 4] 1.0005795 4178.297
#> theta[2, 4, 1] 0.9997140 4222.745
#> theta[2, 4, 2] 0.9996457 3930.899
#> theta[2, 4, 3] 1.0007434 5416.353
#> theta[2, 4, 4] 1.0011947 5438.803
#> theta[3, 1, 1] 1.0003988 5751.889
#> theta[3, 1, 2] 1.0003922 5216.607
#> theta[3, 1, 3] 0.9998277 5637.643
#> theta[3, 1, 4] 0.9994404 5740.093
#> theta[3, 2, 1] 1.0006063 5171.850
#> theta[3, 2, 2] 1.0009760 6432.714
#> theta[3, 2, 3] 1.0003746 6861.777
#> theta[3, 2, 4] 1.0007434 5003.864
#> theta[3, 3, 1] 1.0007898 4542.572
#> theta[3, 3, 2] 1.0020681 5201.463
#> theta[3, 3, 3] 1.0000297 5501.090
#> theta[3, 3, 4] 1.0000078 5868.728
#> theta[3, 4, 1] 0.9998281 5776.005
#> theta[3, 4, 2] 1.0019158 4805.347
#> theta[3, 4, 3] 1.0011849 3631.473
#> theta[3, 4, 4] 1.0002110 5490.299
#> theta[4, 1, 1] 1.0008008 6050.074
#> theta[4, 1, 2] 1.0008276 5466.290
#> theta[4, 1, 3] 1.0011884 5328.890
#> theta[4, 1, 4] 0.9999426 5067.041
#> theta[4, 2, 1] 0.9997516 4471.042
#> theta[4, 2, 2] 1.0011562 5428.980
#> theta[4, 2, 3] 1.0004334 5273.220
#> theta[4, 2, 4] 1.0007967 5028.811
#> theta[4, 3, 1] 0.9997994 4285.651
#> theta[4, 3, 2] 1.0005125 3088.382
#> theta[4, 3, 3] 1.0002922 5628.766
#> theta[4, 3, 4] 0.9998843 3440.650
#> theta[4, 4, 1] 1.0014146 4336.770
#> theta[4, 4, 2] 1.0001425 5187.998
#> theta[4, 4, 3] 1.0015689 5301.514
#> theta[4, 4, 4] 1.0006186 6196.290
#> theta[5, 1, 1] 1.0009799 5288.322
#> theta[5, 1, 2] 1.0003481 4426.728
#> theta[5, 1, 3] 1.0016509 5474.364
#> theta[5, 1, 4] 1.0002134 4596.134
#> theta[5, 2, 1] 1.0021795 6129.968
#> theta[5, 2, 2] 1.0010721 6119.793
#> theta[5, 2, 3] 1.0000903 5374.347
#> theta[5, 2, 4] 1.0011489 5078.874
#> theta[5, 3, 1] 0.9999567 3959.618
#> theta[5, 3, 2] 1.0013087 5092.900
#> theta[5, 3, 3] 0.9999694 5151.759
#> theta[5, 3, 4] 1.0002572 4834.410
#> theta[5, 4, 1] 1.0021253 4911.186
#> theta[5, 4, 2] 0.9997270 4646.309
#> theta[5, 4, 3] 1.0003028 3927.063
#> theta[5, 4, 4] 1.0011909 4990.419

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.000860 6586.315
#> pi[2] 1.000168 5796.556
#> pi[3] 1.000937 4845.750
#> pi[4] 1.000291 5154.161

# }