Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000169 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.69 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.063 seconds (Warm-up)
#> Chain 1:                2.019 seconds (Sampling)
#> Chain 1:                4.082 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000155 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.55 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.097 seconds (Warm-up)
#> Chain 2:                2.16 seconds (Sampling)
#> Chain 2:                4.257 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.00015 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.5 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.054 seconds (Warm-up)
#> Chain 3:                2.211 seconds (Sampling)
#> Chain 3:                4.265 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.00015 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.5 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.097 seconds (Warm-up)
#> Chain 4:                2.292 seconds (Sampling)
#> Chain 4:                4.389 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0013558 6975.060
#> pi[2]          1.0006456 7522.721
#> pi[3]          1.0007079 6068.230
#> pi[4]          1.0004006 5904.940
#> theta[1, 1, 1] 1.0021332 6087.810
#> theta[1, 1, 2] 1.0006523 5952.382
#> theta[1, 1, 3] 1.0004815 5236.267
#> theta[1, 1, 4] 1.0007938 5572.101
#> theta[1, 2, 1] 0.9999232 5424.238
#> theta[1, 2, 2] 1.0006130 6728.517
#> theta[1, 2, 3] 1.0038888 6247.638
#> theta[1, 2, 4] 1.0003537 5383.637
#> theta[1, 3, 1] 0.9999625 4429.850
#> theta[1, 3, 2] 0.9998996 3426.185
#> theta[1, 3, 3] 1.0002437 4923.937
#> theta[1, 3, 4] 0.9999806 4544.566
#> theta[1, 4, 1] 1.0010091 4535.798
#> theta[1, 4, 2] 1.0014434 4941.536
#> theta[1, 4, 3] 1.0026104 3944.097
#> theta[1, 4, 4] 1.0001787 5685.075
#> theta[2, 1, 1] 1.0014654 6250.661
#> theta[2, 1, 2] 0.9998835 5730.439
#> theta[2, 1, 3] 1.0000950 5943.334
#> theta[2, 1, 4] 0.9996849 5204.702
#> theta[2, 2, 1] 1.0019688 6274.217
#> theta[2, 2, 2] 1.0004937 8685.972
#> theta[2, 2, 3] 1.0000700 9509.573
#> theta[2, 2, 4] 1.0002935 5142.487
#> theta[2, 3, 1] 1.0016774 5904.826
#> theta[2, 3, 2] 1.0004828 5095.908
#> theta[2, 3, 3] 1.0011885 6011.725
#> theta[2, 3, 4] 1.0010400 4402.414
#> theta[2, 4, 1] 1.0012494 5348.283
#> theta[2, 4, 2] 0.9998030 4497.231
#> theta[2, 4, 3] 1.0009943 4897.277
#> theta[2, 4, 4] 1.0008206 6013.155
#> theta[3, 1, 1] 0.9998721 6802.869
#> theta[3, 1, 2] 1.0008492 5508.519
#> theta[3, 1, 3] 1.0001654 6080.723
#> theta[3, 1, 4] 1.0000416 5125.749
#> theta[3, 2, 1] 0.9998030 5409.618
#> theta[3, 2, 2] 1.0005092 5933.121
#> theta[3, 2, 3] 0.9998710 6451.386
#> theta[3, 2, 4] 1.0002608 5390.877
#> theta[3, 3, 1] 0.9998849 5116.972
#> theta[3, 3, 2] 0.9995498 5809.064
#> theta[3, 3, 3] 1.0006862 6215.211
#> theta[3, 3, 4] 0.9998584 7182.681
#> theta[3, 4, 1] 0.9998843 5358.530
#> theta[3, 4, 2] 1.0010750 5155.420
#> theta[3, 4, 3] 1.0006575 4800.285
#> theta[3, 4, 4] 1.0002182 6049.424
#> theta[4, 1, 1] 1.0024216 5844.504
#> theta[4, 1, 2] 0.9997775 5321.432
#> theta[4, 1, 3] 1.0008926 4659.360
#> theta[4, 1, 4] 1.0003311 5036.990
#> theta[4, 2, 1] 1.0025539 6224.674
#> theta[4, 2, 2] 1.0010607 6495.803
#> theta[4, 2, 3] 1.0004661 7069.561
#> theta[4, 2, 4] 1.0006020 4369.052
#> theta[4, 3, 1] 0.9993129 4703.639
#> theta[4, 3, 2] 1.0001620 5190.365
#> theta[4, 3, 3] 0.9998039 5441.238
#> theta[4, 3, 4] 1.0024064 4751.058
#> theta[4, 4, 1] 1.0000982 4975.438
#> theta[4, 4, 2] 1.0027236 4800.437
#> theta[4, 4, 3] 1.0008383 5188.962
#> theta[4, 4, 4] 0.9998540 4683.036
#> theta[5, 1, 1] 1.0002722 6580.033
#> theta[5, 1, 2] 1.0007753 4920.276
#> theta[5, 1, 3] 1.0003045 5040.720
#> theta[5, 1, 4] 1.0006367 4813.026
#> theta[5, 2, 1] 1.0027917 5979.696
#> theta[5, 2, 2] 1.0007245 6262.316
#> theta[5, 2, 3] 0.9999569 5964.064
#> theta[5, 2, 4] 1.0024178 5136.757
#> theta[5, 3, 1] 1.0002851 5383.317
#> theta[5, 3, 2] 0.9993989 5370.428
#> theta[5, 3, 3] 1.0024022 6088.177
#> theta[5, 3, 4] 1.0002599 5648.285
#> theta[5, 4, 1] 0.9995694 5428.119
#> theta[5, 4, 2] 1.0010769 5333.560
#> theta[5, 4, 3] 0.9998295 4893.812
#> theta[5, 4, 4] 0.9997756 5669.131

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.001356 6975.060
#> pi[2] 1.000646 7522.721
#> pi[3] 1.000708 6068.230
#> pi[4] 1.000401 5904.940

# }