Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000151 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.51 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.94291 seconds (Warm-up)
#> Chain 1:                2.12394 seconds (Sampling)
#> Chain 1:                4.06685 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000116 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.16 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.92814 seconds (Warm-up)
#> Chain 2:                1.75421 seconds (Sampling)
#> Chain 2:                3.68235 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000119 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.19 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.93869 seconds (Warm-up)
#> Chain 3:                1.84642 seconds (Sampling)
#> Chain 3:                3.7851 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.00014 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.4 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.8315 seconds (Warm-up)
#> Chain 4:                1.92122 seconds (Sampling)
#> Chain 4:                3.75272 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0010935 6180.541
#> pi[2]          1.0009528 6433.495
#> pi[3]          1.0010887 5724.013
#> pi[4]          1.0016940 5881.273
#> theta[1, 1, 1] 1.0003877 6333.796
#> theta[1, 1, 2] 1.0004248 5733.640
#> theta[1, 1, 3] 1.0013624 5584.729
#> theta[1, 1, 4] 1.0002429 5304.811
#> theta[1, 2, 1] 0.9998715 4696.701
#> theta[1, 2, 2] 1.0015693 5522.075
#> theta[1, 2, 3] 1.0017714 6143.938
#> theta[1, 2, 4] 1.0008498 4891.754
#> theta[1, 3, 1] 1.0007707 4792.789
#> theta[1, 3, 2] 1.0007488 3918.557
#> theta[1, 3, 3] 1.0000830 4726.760
#> theta[1, 3, 4] 1.0000615 3682.763
#> theta[1, 4, 1] 1.0010031 4366.608
#> theta[1, 4, 2] 1.0000619 4461.492
#> theta[1, 4, 3] 1.0003898 3255.768
#> theta[1, 4, 4] 1.0004181 5143.125
#> theta[2, 1, 1] 0.9995262 6432.744
#> theta[2, 1, 2] 1.0009735 6191.221
#> theta[2, 1, 3] 0.9999082 4311.265
#> theta[2, 1, 4] 1.0009590 5247.911
#> theta[2, 2, 1] 1.0001134 5181.434
#> theta[2, 2, 2] 1.0022633 6680.393
#> theta[2, 2, 3] 1.0009401 5907.748
#> theta[2, 2, 4] 1.0004517 4342.116
#> theta[2, 3, 1] 0.9996249 4595.183
#> theta[2, 3, 2] 1.0004756 4799.606
#> theta[2, 3, 3] 1.0015657 5047.485
#> theta[2, 3, 4] 0.9995997 5190.006
#> theta[2, 4, 1] 1.0005297 4760.188
#> theta[2, 4, 2] 1.0005907 5075.081
#> theta[2, 4, 3] 1.0007128 5080.537
#> theta[2, 4, 4] 1.0004970 5931.493
#> theta[3, 1, 1] 1.0025522 5584.418
#> theta[3, 1, 2] 1.0007009 4942.815
#> theta[3, 1, 3] 1.0022842 5567.358
#> theta[3, 1, 4] 1.0000879 5133.993
#> theta[3, 2, 1] 1.0009237 5105.493
#> theta[3, 2, 2] 0.9997589 6018.101
#> theta[3, 2, 3] 1.0004824 6038.437
#> theta[3, 2, 4] 1.0006472 4674.392
#> theta[3, 3, 1] 0.9997429 5996.771
#> theta[3, 3, 2] 1.0003135 5563.852
#> theta[3, 3, 3] 1.0007719 5917.069
#> theta[3, 3, 4] 1.0020059 6242.414
#> theta[3, 4, 1] 1.0009333 4240.241
#> theta[3, 4, 2] 1.0004664 4410.842
#> theta[3, 4, 3] 0.9997145 4145.089
#> theta[3, 4, 4] 0.9999462 5516.373
#> theta[4, 1, 1] 1.0028102 6050.267
#> theta[4, 1, 2] 1.0013655 5367.386
#> theta[4, 1, 3] 0.9998549 4271.086
#> theta[4, 1, 4] 1.0004777 4461.770
#> theta[4, 2, 1] 1.0012690 5519.456
#> theta[4, 2, 2] 0.9997472 5954.012
#> theta[4, 2, 3] 1.0011522 4853.091
#> theta[4, 2, 4] 1.0004737 4472.283
#> theta[4, 3, 1] 1.0001492 4959.394
#> theta[4, 3, 2] 1.0006005 3949.386
#> theta[4, 3, 3] 0.9995584 5573.070
#> theta[4, 3, 4] 1.0002036 5036.659
#> theta[4, 4, 1] 0.9998746 4590.653
#> theta[4, 4, 2] 1.0008133 5007.679
#> theta[4, 4, 3] 1.0016078 4527.403
#> theta[4, 4, 4] 1.0000392 6023.129
#> theta[5, 1, 1] 1.0007437 5092.788
#> theta[5, 1, 2] 1.0006655 4218.447
#> theta[5, 1, 3] 1.0017702 4494.551
#> theta[5, 1, 4] 1.0014732 5049.973
#> theta[5, 2, 1] 1.0008733 5635.944
#> theta[5, 2, 2] 1.0001787 5585.022
#> theta[5, 2, 3] 1.0014417 5268.951
#> theta[5, 2, 4] 1.0012710 5124.999
#> theta[5, 3, 1] 1.0005886 4647.498
#> theta[5, 3, 2] 1.0027707 5727.157
#> theta[5, 3, 3] 1.0001951 5994.803
#> theta[5, 3, 4] 1.0003690 5654.292
#> theta[5, 4, 1] 1.0001613 4832.968
#> theta[5, 4, 2] 1.0007913 5576.387
#> theta[5, 4, 3] 1.0000146 4857.381
#> theta[5, 4, 4] 1.0012508 5891.990

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.001093 6180.541
#> pi[2] 1.000953 6433.495
#> pi[3] 1.001089 5724.013
#> pi[4] 1.001694 5881.273

# }