Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000143 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.43 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.65435 seconds (Warm-up)
#> Chain 1:                1.77496 seconds (Sampling)
#> Chain 1:                3.42931 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000148 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.48 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.60439 seconds (Warm-up)
#> Chain 2:                1.73409 seconds (Sampling)
#> Chain 2:                3.33848 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000169 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.69 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.59456 seconds (Warm-up)
#> Chain 3:                1.74619 seconds (Sampling)
#> Chain 3:                3.34075 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.00014 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.4 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.60529 seconds (Warm-up)
#> Chain 4:                1.76306 seconds (Sampling)
#> Chain 4:                3.36835 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0021096 8798.863
#> pi[2]          1.0016714 8655.345
#> pi[3]          1.0009969 7297.216
#> pi[4]          1.0022211 8557.850
#> theta[1, 1, 1] 1.0015775 8552.239
#> theta[1, 1, 2] 1.0009787 7926.127
#> theta[1, 1, 3] 1.0011583 6507.362
#> theta[1, 1, 4] 0.9998272 5887.541
#> theta[1, 2, 1] 1.0023864 6359.557
#> theta[1, 2, 2] 0.9999449 7193.945
#> theta[1, 2, 3] 1.0011916 7496.710
#> theta[1, 2, 4] 1.0015518 6216.648
#> theta[1, 3, 1] 1.0004457 5516.027
#> theta[1, 3, 2] 0.9996034 4252.440
#> theta[1, 3, 3] 1.0002926 5449.289
#> theta[1, 3, 4] 1.0006195 6000.772
#> theta[1, 4, 1] 1.0001985 5283.850
#> theta[1, 4, 2] 1.0001746 5812.547
#> theta[1, 4, 3] 1.0006339 3927.099
#> theta[1, 4, 4] 1.0005792 5622.140
#> theta[2, 1, 1] 1.0013778 8867.814
#> theta[2, 1, 2] 1.0013510 8416.146
#> theta[2, 1, 3] 1.0005182 6474.254
#> theta[2, 1, 4] 0.9996432 6135.409
#> theta[2, 2, 1] 1.0013438 6788.632
#> theta[2, 2, 2] 1.0016920 7977.490
#> theta[2, 2, 3] 1.0008181 7057.748
#> theta[2, 2, 4] 0.9996955 6233.392
#> theta[2, 3, 1] 0.9998215 4597.991
#> theta[2, 3, 2] 1.0004447 4391.730
#> theta[2, 3, 3] 1.0012172 5717.685
#> theta[2, 3, 4] 1.0003851 5593.396
#> theta[2, 4, 1] 1.0010895 5338.564
#> theta[2, 4, 2] 1.0023566 5457.160
#> theta[2, 4, 3] 1.0006568 6751.630
#> theta[2, 4, 4] 1.0007979 6144.480
#> theta[3, 1, 1] 1.0012348 8195.921
#> theta[3, 1, 2] 1.0023192 5714.106
#> theta[3, 1, 3] 0.9996537 6151.425
#> theta[3, 1, 4] 0.9999485 6724.449
#> theta[3, 2, 1] 1.0009125 5728.457
#> theta[3, 2, 2] 1.0002698 6775.949
#> theta[3, 2, 3] 1.0010786 7506.754
#> theta[3, 2, 4] 1.0009632 5656.512
#> theta[3, 3, 1] 1.0002522 4819.699
#> theta[3, 3, 2] 1.0009456 6740.253
#> theta[3, 3, 3] 0.9995111 7317.495
#> theta[3, 3, 4] 1.0008247 7006.361
#> theta[3, 4, 1] 1.0014350 5288.080
#> theta[3, 4, 2] 1.0020975 5593.391
#> theta[3, 4, 3] 1.0001396 5455.993
#> theta[3, 4, 4] 1.0016251 5991.919
#> theta[4, 1, 1] 1.0004872 7730.245
#> theta[4, 1, 2] 1.0011468 6719.017
#> theta[4, 1, 3] 1.0003679 5683.612
#> theta[4, 1, 4] 1.0014817 6510.357
#> theta[4, 2, 1] 1.0007967 6806.361
#> theta[4, 2, 2] 1.0009100 7968.261
#> theta[4, 2, 3] 0.9997576 7032.701
#> theta[4, 2, 4] 1.0001184 5858.715
#> theta[4, 3, 1] 1.0020219 4378.230
#> theta[4, 3, 2] 1.0002779 5314.407
#> theta[4, 3, 3] 0.9998657 6301.205
#> theta[4, 3, 4] 1.0003346 5053.082
#> theta[4, 4, 1] 1.0005154 4983.024
#> theta[4, 4, 2] 1.0007321 5703.291
#> theta[4, 4, 3] 0.9999955 5843.724
#> theta[4, 4, 4] 1.0002268 6590.067
#> theta[5, 1, 1] 1.0019289 7275.913
#> theta[5, 1, 2] 1.0004394 5396.963
#> theta[5, 1, 3] 1.0008017 5679.836
#> theta[5, 1, 4] 1.0006327 5329.912
#> theta[5, 2, 1] 1.0003074 8894.317
#> theta[5, 2, 2] 1.0022001 8263.591
#> theta[5, 2, 3] 1.0000050 6256.136
#> theta[5, 2, 4] 1.0010667 5381.505
#> theta[5, 3, 1] 1.0013569 5814.768
#> theta[5, 3, 2] 1.0013586 6234.160
#> theta[5, 3, 3] 1.0007439 8247.474
#> theta[5, 3, 4] 1.0008481 6808.889
#> theta[5, 4, 1] 1.0007926 5246.848
#> theta[5, 4, 2] 1.0004725 5917.387
#> theta[5, 4, 3] 1.0004755 4960.602
#> theta[5, 4, 4] 1.0008550 6551.286

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.002110 8798.863
#> pi[2] 1.001671 8655.345
#> pi[3] 1.000997 7297.216
#> pi[4] 1.002221 8557.850

# }