Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000172 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.72 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.022 seconds (Warm-up)
#> Chain 1:                1.994 seconds (Sampling)
#> Chain 1:                4.016 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000168 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.68 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.198 seconds (Warm-up)
#> Chain 2:                2.3 seconds (Sampling)
#> Chain 2:                4.498 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.00016 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.6 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.103 seconds (Warm-up)
#> Chain 3:                2.206 seconds (Sampling)
#> Chain 3:                4.309 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000161 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.61 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.046 seconds (Warm-up)
#> Chain 4:                2.3 seconds (Sampling)
#> Chain 4:                4.346 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0016196 8161.313
#> pi[2]          0.9997282 7621.926
#> pi[3]          1.0029256 7321.840
#> pi[4]          1.0001685 6915.072
#> theta[1, 1, 1] 1.0009100 7193.766
#> theta[1, 1, 2] 1.0002060 6594.632
#> theta[1, 1, 3] 1.0005062 5567.076
#> theta[1, 1, 4] 1.0006583 5172.755
#> theta[1, 2, 1] 1.0001938 5247.637
#> theta[1, 2, 2] 0.9997926 5864.260
#> theta[1, 2, 3] 1.0018694 6083.460
#> theta[1, 2, 4] 1.0006893 5219.131
#> theta[1, 3, 1] 1.0008506 5081.105
#> theta[1, 3, 2] 1.0007603 3858.229
#> theta[1, 3, 3] 1.0012850 5237.854
#> theta[1, 3, 4] 1.0003422 4648.272
#> theta[1, 4, 1] 1.0007693 5438.560
#> theta[1, 4, 2] 1.0003946 4713.445
#> theta[1, 4, 3] 1.0000876 3970.568
#> theta[1, 4, 4] 0.9997729 5728.620
#> theta[2, 1, 1] 1.0003959 7244.790
#> theta[2, 1, 2] 1.0014212 6822.128
#> theta[2, 1, 3] 1.0015155 5854.182
#> theta[2, 1, 4] 1.0010627 5641.971
#> theta[2, 2, 1] 1.0019688 6491.094
#> theta[2, 2, 2] 1.0014883 7372.794
#> theta[2, 2, 3] 1.0010090 6621.243
#> theta[2, 2, 4] 1.0015022 5208.223
#> theta[2, 3, 1] 1.0011051 4925.820
#> theta[2, 3, 2] 1.0004928 4980.942
#> theta[2, 3, 3] 1.0031363 6625.910
#> theta[2, 3, 4] 1.0012539 4859.378
#> theta[2, 4, 1] 1.0010803 5698.881
#> theta[2, 4, 2] 1.0022582 6084.891
#> theta[2, 4, 3] 1.0011721 5567.989
#> theta[2, 4, 4] 1.0002875 6757.612
#> theta[3, 1, 1] 1.0029844 6913.697
#> theta[3, 1, 2] 1.0013729 4648.037
#> theta[3, 1, 3] 1.0002023 5683.402
#> theta[3, 1, 4] 1.0026064 5957.384
#> theta[3, 2, 1] 1.0007078 6218.527
#> theta[3, 2, 2] 1.0016540 7446.835
#> theta[3, 2, 3] 1.0016061 6587.801
#> theta[3, 2, 4] 1.0008888 5070.069
#> theta[3, 3, 1] 1.0012850 5675.649
#> theta[3, 3, 2] 1.0007630 6037.712
#> theta[3, 3, 3] 1.0003183 6904.522
#> theta[3, 3, 4] 1.0010635 7041.990
#> theta[3, 4, 1] 1.0007369 5501.134
#> theta[3, 4, 2] 1.0018078 4861.757
#> theta[3, 4, 3] 1.0004607 5508.920
#> theta[3, 4, 4] 1.0010104 6051.892
#> theta[4, 1, 1] 0.9994659 6518.463
#> theta[4, 1, 2] 1.0005805 5700.472
#> theta[4, 1, 3] 1.0015211 5862.635
#> theta[4, 1, 4] 1.0008360 5757.253
#> theta[4, 2, 1] 0.9999215 6332.409
#> theta[4, 2, 2] 1.0036383 5819.439
#> theta[4, 2, 3] 1.0001899 6610.497
#> theta[4, 2, 4] 1.0018497 5407.761
#> theta[4, 3, 1] 0.9996447 5235.697
#> theta[4, 3, 2] 1.0010850 4913.152
#> theta[4, 3, 3] 1.0000985 5763.276
#> theta[4, 3, 4] 1.0001720 4576.066
#> theta[4, 4, 1] 1.0007104 5532.988
#> theta[4, 4, 2] 1.0008674 6356.884
#> theta[4, 4, 3] 1.0005175 6012.662
#> theta[4, 4, 4] 1.0004776 6382.001
#> theta[5, 1, 1] 1.0010623 6598.610
#> theta[5, 1, 2] 1.0000013 5133.981
#> theta[5, 1, 3] 1.0016347 4966.283
#> theta[5, 1, 4] 0.9998754 5419.532
#> theta[5, 2, 1] 1.0015690 7032.498
#> theta[5, 2, 2] 1.0024568 7289.535
#> theta[5, 2, 3] 1.0005161 6773.640
#> theta[5, 2, 4] 1.0004023 6080.235
#> theta[5, 3, 1] 1.0023307 5147.133
#> theta[5, 3, 2] 1.0007912 6585.854
#> theta[5, 3, 3] 0.9996851 8382.844
#> theta[5, 3, 4] 1.0002412 6155.819
#> theta[5, 4, 1] 1.0000574 5012.999
#> theta[5, 4, 2] 1.0009701 5258.900
#> theta[5, 4, 3] 1.0004921 4714.923
#> theta[5, 4, 4] 1.0018152 5539.067

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>            Rhat ess_bulk
#> pi[1] 1.0016196 8161.313
#> pi[2] 0.9997282 7621.926
#> pi[3] 1.0029256 7321.840
#> pi[4] 1.0001685 6915.072

# }