Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000167 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.67 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.948 seconds (Warm-up)
#> Chain 1:                2.098 seconds (Sampling)
#> Chain 1:                4.046 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000146 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.46 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.905 seconds (Warm-up)
#> Chain 2:                2.022 seconds (Sampling)
#> Chain 2:                3.927 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.00014 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.4 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.92 seconds (Warm-up)
#> Chain 3:                1.509 seconds (Sampling)
#> Chain 3:                3.429 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000139 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.39 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.932 seconds (Warm-up)
#> Chain 4:                2.042 seconds (Sampling)
#> Chain 4:                3.974 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0005344 7006.343
#> pi[2]          1.0015446 6686.586
#> pi[3]          1.0001559 6539.788
#> pi[4]          1.0016900 5789.260
#> theta[1, 1, 1] 1.0032747 6549.744
#> theta[1, 1, 2] 1.0053605 5976.004
#> theta[1, 1, 3] 1.0003465 5270.348
#> theta[1, 1, 4] 1.0013203 5542.501
#> theta[1, 2, 1] 1.0001952 5464.977
#> theta[1, 2, 2] 0.9995437 6763.942
#> theta[1, 2, 3] 1.0006214 6082.013
#> theta[1, 2, 4] 1.0018232 4540.164
#> theta[1, 3, 1] 1.0021852 5651.779
#> theta[1, 3, 2] 1.0011338 3627.249
#> theta[1, 3, 3] 1.0012658 4946.393
#> theta[1, 3, 4] 1.0006257 4262.668
#> theta[1, 4, 1] 1.0002420 5585.480
#> theta[1, 4, 2] 1.0005289 5905.188
#> theta[1, 4, 3] 1.0007405 3462.443
#> theta[1, 4, 4] 0.9998712 5745.514
#> theta[2, 1, 1] 1.0045032 7500.297
#> theta[2, 1, 2] 1.0033333 6355.865
#> theta[2, 1, 3] 1.0004666 5763.902
#> theta[2, 1, 4] 1.0001422 4823.400
#> theta[2, 2, 1] 1.0008157 6776.081
#> theta[2, 2, 2] 0.9996694 6952.342
#> theta[2, 2, 3] 1.0001438 6452.613
#> theta[2, 2, 4] 1.0008918 5111.516
#> theta[2, 3, 1] 0.9997192 5293.595
#> theta[2, 3, 2] 1.0007954 5165.962
#> theta[2, 3, 3] 1.0004022 6094.846
#> theta[2, 3, 4] 1.0000141 5069.840
#> theta[2, 4, 1] 1.0022289 4913.167
#> theta[2, 4, 2] 1.0001566 5062.889
#> theta[2, 4, 3] 0.9998069 5282.965
#> theta[2, 4, 4] 0.9996680 5758.643
#> theta[3, 1, 1] 1.0013442 7184.454
#> theta[3, 1, 2] 0.9999976 5354.209
#> theta[3, 1, 3] 0.9997800 5339.104
#> theta[3, 1, 4] 1.0005951 6252.443
#> theta[3, 2, 1] 1.0006721 5811.004
#> theta[3, 2, 2] 1.0002824 6708.965
#> theta[3, 2, 3] 1.0007218 6615.097
#> theta[3, 2, 4] 1.0007550 4994.261
#> theta[3, 3, 1] 1.0005787 4673.812
#> theta[3, 3, 2] 1.0024531 5224.888
#> theta[3, 3, 3] 1.0010692 5951.126
#> theta[3, 3, 4] 1.0011885 6744.624
#> theta[3, 4, 1] 1.0008294 5354.106
#> theta[3, 4, 2] 0.9995461 5863.234
#> theta[3, 4, 3] 1.0004206 4887.075
#> theta[3, 4, 4] 1.0005002 5751.539
#> theta[4, 1, 1] 1.0011433 6094.259
#> theta[4, 1, 2] 1.0000369 5530.987
#> theta[4, 1, 3] 1.0002927 5103.203
#> theta[4, 1, 4] 1.0007698 5278.324
#> theta[4, 2, 1] 1.0010255 7107.109
#> theta[4, 2, 2] 1.0021942 8565.824
#> theta[4, 2, 3] 1.0008852 5572.755
#> theta[4, 2, 4] 1.0011958 6520.680
#> theta[4, 3, 1] 0.9998416 5887.524
#> theta[4, 3, 2] 1.0005289 5381.919
#> theta[4, 3, 3] 1.0009164 5799.308
#> theta[4, 3, 4] 1.0004328 3971.950
#> theta[4, 4, 1] 1.0017685 5217.688
#> theta[4, 4, 2] 0.9999441 6126.119
#> theta[4, 4, 3] 1.0002122 5387.796
#> theta[4, 4, 4] 1.0009168 6579.696
#> theta[5, 1, 1] 1.0013144 6732.104
#> theta[5, 1, 2] 1.0001206 4494.084
#> theta[5, 1, 3] 1.0010965 5421.483
#> theta[5, 1, 4] 1.0002720 5635.974
#> theta[5, 2, 1] 1.0013364 6178.985
#> theta[5, 2, 2] 1.0001470 7120.797
#> theta[5, 2, 3] 1.0014890 5792.473
#> theta[5, 2, 4] 1.0005494 4462.271
#> theta[5, 3, 1] 1.0001620 5310.617
#> theta[5, 3, 2] 1.0008197 7467.824
#> theta[5, 3, 3] 0.9998540 5986.241
#> theta[5, 3, 4] 1.0013400 5541.271
#> theta[5, 4, 1] 0.9999301 5511.714
#> theta[5, 4, 2] 1.0002528 5030.331
#> theta[5, 4, 3] 1.0005680 4245.753
#> theta[5, 4, 4] 1.0003134 5623.352

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.000534 7006.343
#> pi[2] 1.001545 6686.586
#> pi[3] 1.000156 6539.788
#> pi[4] 1.001690 5789.260

# }