Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000228 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 2.28 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 3.022 seconds (Warm-up)
#> Chain 1:                3.366 seconds (Sampling)
#> Chain 1:                6.388 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000216 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 2.16 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 3.099 seconds (Warm-up)
#> Chain 2:                3.417 seconds (Sampling)
#> Chain 2:                6.516 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000217 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 2.17 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 3.059 seconds (Warm-up)
#> Chain 3:                3.312 seconds (Sampling)
#> Chain 3:                6.371 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000252 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 2.52 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.991 seconds (Warm-up)
#> Chain 4:                3.336 seconds (Sampling)
#> Chain 4:                6.327 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat  ess_bulk
#> pi[1]          1.0025289  9452.006
#> pi[2]          0.9996569  8030.977
#> pi[3]          1.0004580  8255.723
#> pi[4]          1.0009465  8541.451
#> theta[1, 1, 1] 1.0002446  8860.464
#> theta[1, 1, 2] 0.9999709  7763.092
#> theta[1, 1, 3] 0.9998200  7020.039
#> theta[1, 1, 4] 1.0008817  4466.997
#> theta[1, 2, 1] 0.9996596  6050.924
#> theta[1, 2, 2] 1.0037033  7249.671
#> theta[1, 2, 3] 1.0015903  7350.631
#> theta[1, 2, 4] 1.0000523  6671.461
#> theta[1, 3, 1] 1.0013955  5591.150
#> theta[1, 3, 2] 0.9998048  4139.943
#> theta[1, 3, 3] 0.9997030  4924.835
#> theta[1, 3, 4] 1.0004294  4396.436
#> theta[1, 4, 1] 1.0049731  5891.088
#> theta[1, 4, 2] 1.0004629  7133.394
#> theta[1, 4, 3] 1.0003276  4085.102
#> theta[1, 4, 4] 1.0007027  5544.892
#> theta[2, 1, 1] 1.0007158 10298.330
#> theta[2, 1, 2] 1.0025689  8125.640
#> theta[2, 1, 3] 1.0002488  6115.624
#> theta[2, 1, 4] 1.0006978  5838.696
#> theta[2, 2, 1] 1.0022186  7517.080
#> theta[2, 2, 2] 1.0029266  8871.359
#> theta[2, 2, 3] 1.0010028  7530.759
#> theta[2, 2, 4] 1.0001629  6413.079
#> theta[2, 3, 1] 1.0007111  5235.549
#> theta[2, 3, 2] 0.9998156  6486.793
#> theta[2, 3, 3] 1.0001717  7255.801
#> theta[2, 3, 4] 1.0011566  6055.014
#> theta[2, 4, 1] 1.0038648  6967.624
#> theta[2, 4, 2] 1.0009882  7078.602
#> theta[2, 4, 3] 1.0009521  6375.538
#> theta[2, 4, 4] 1.0003457  6825.463
#> theta[3, 1, 1] 1.0000418  7835.937
#> theta[3, 1, 2] 1.0002439  6934.968
#> theta[3, 1, 3] 0.9999986  5364.297
#> theta[3, 1, 4] 0.9999810  6206.980
#> theta[3, 2, 1] 1.0010795  6443.442
#> theta[3, 2, 2] 1.0007751  9329.576
#> theta[3, 2, 3] 1.0004001  8915.058
#> theta[3, 2, 4] 0.9997768  5899.699
#> theta[3, 3, 1] 1.0008611  6344.564
#> theta[3, 3, 2] 1.0006286  6027.279
#> theta[3, 3, 3] 0.9994224  8377.192
#> theta[3, 3, 4] 1.0000792  8004.564
#> theta[3, 4, 1] 1.0008164  6721.904
#> theta[3, 4, 2] 1.0009167  5488.288
#> theta[3, 4, 3] 1.0021005  4942.263
#> theta[3, 4, 4] 1.0007981  5973.165
#> theta[4, 1, 1] 1.0009901  7937.873
#> theta[4, 1, 2] 1.0004679  6166.877
#> theta[4, 1, 3] 1.0003390  5696.207
#> theta[4, 1, 4] 0.9995029  6748.552
#> theta[4, 2, 1] 1.0000456  7431.744
#> theta[4, 2, 2] 1.0009590  6717.674
#> theta[4, 2, 3] 1.0000172  6937.027
#> theta[4, 2, 4] 1.0004402  5920.503
#> theta[4, 3, 1] 1.0001892  5882.429
#> theta[4, 3, 2] 1.0012140  6702.182
#> theta[4, 3, 3] 1.0028526  6605.764
#> theta[4, 3, 4] 1.0017762  4843.777
#> theta[4, 4, 1] 1.0007570  5769.138
#> theta[4, 4, 2] 0.9996659  5806.331
#> theta[4, 4, 3] 1.0017404  5981.346
#> theta[4, 4, 4] 1.0003607  6904.123
#> theta[5, 1, 1] 1.0009053  8243.962
#> theta[5, 1, 2] 0.9997052  6093.662
#> theta[5, 1, 3] 1.0016759  5935.246
#> theta[5, 1, 4] 1.0004106  6461.895
#> theta[5, 2, 1] 0.9995277  8712.788
#> theta[5, 2, 2] 1.0011608  9347.429
#> theta[5, 2, 3] 0.9995494  7601.026
#> theta[5, 2, 4] 0.9995022  5672.398
#> theta[5, 3, 1] 1.0018386  5573.827
#> theta[5, 3, 2] 1.0002217  7762.877
#> theta[5, 3, 3] 1.0002183  7611.590
#> theta[5, 3, 4] 1.0009020  8109.556
#> theta[5, 4, 1] 0.9998803  6115.792
#> theta[5, 4, 2] 1.0011963  5698.997
#> theta[5, 4, 3] 1.0007811  4340.609
#> theta[5, 4, 4] 1.0002944  6052.181

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>            Rhat ess_bulk
#> pi[1] 1.0025289 9452.006
#> pi[2] 0.9996569 8030.977
#> pi[3] 1.0004580 8255.723
#> pi[4] 1.0009465 8541.451

# }