Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000159 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.59 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.65975 seconds (Warm-up)
#> Chain 1:                1.73477 seconds (Sampling)
#> Chain 1:                3.39453 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000142 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.42 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.60954 seconds (Warm-up)
#> Chain 2:                1.72624 seconds (Sampling)
#> Chain 2:                3.33578 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.00013 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.3 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.58183 seconds (Warm-up)
#> Chain 3:                1.73615 seconds (Sampling)
#> Chain 3:                3.31798 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000127 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.27 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.64042 seconds (Warm-up)
#> Chain 4:                1.79554 seconds (Sampling)
#> Chain 4:                3.43596 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          0.9998375 7183.681
#> pi[2]          1.0015102 6979.643
#> pi[3]          0.9995905 6633.003
#> pi[4]          1.0004142 6729.542
#> theta[1, 1, 1] 1.0006436 7778.044
#> theta[1, 1, 2] 1.0023402 6794.337
#> theta[1, 1, 3] 1.0005836 5728.806
#> theta[1, 1, 4] 1.0011251 5795.408
#> theta[1, 2, 1] 1.0001887 4769.440
#> theta[1, 2, 2] 1.0012907 6206.116
#> theta[1, 2, 3] 1.0021965 7468.023
#> theta[1, 2, 4] 1.0005875 4966.747
#> theta[1, 3, 1] 1.0002244 4490.347
#> theta[1, 3, 2] 1.0034741 3669.822
#> theta[1, 3, 3] 0.9998471 5273.977
#> theta[1, 3, 4] 1.0005272 5575.008
#> theta[1, 4, 1] 1.0010275 5075.111
#> theta[1, 4, 2] 0.9997720 5424.630
#> theta[1, 4, 3] 1.0001548 4757.345
#> theta[1, 4, 4] 1.0004455 5900.789
#> theta[2, 1, 1] 1.0012916 7605.689
#> theta[2, 1, 2] 1.0002190 7167.772
#> theta[2, 1, 3] 1.0015711 5035.813
#> theta[2, 1, 4] 1.0032615 6806.493
#> theta[2, 2, 1] 1.0013874 6086.011
#> theta[2, 2, 2] 1.0016459 6737.663
#> theta[2, 2, 3] 1.0003326 6236.889
#> theta[2, 2, 4] 1.0003161 5338.292
#> theta[2, 3, 1] 1.0003257 5195.544
#> theta[2, 3, 2] 1.0021393 5813.899
#> theta[2, 3, 3] 0.9998754 6242.690
#> theta[2, 3, 4] 0.9995445 5709.110
#> theta[2, 4, 1] 1.0012775 5678.876
#> theta[2, 4, 2] 0.9996690 5318.289
#> theta[2, 4, 3] 1.0005227 6184.660
#> theta[2, 4, 4] 0.9997846 6373.896
#> theta[3, 1, 1] 1.0006470 7057.020
#> theta[3, 1, 2] 0.9997901 4757.828
#> theta[3, 1, 3] 1.0015279 5363.078
#> theta[3, 1, 4] 1.0005320 4761.259
#> theta[3, 2, 1] 1.0011299 5568.548
#> theta[3, 2, 2] 1.0027889 7387.215
#> theta[3, 2, 3] 1.0009559 6507.620
#> theta[3, 2, 4] 0.9997525 5675.734
#> theta[3, 3, 1] 1.0008552 4867.379
#> theta[3, 3, 2] 0.9997418 6596.850
#> theta[3, 3, 3] 1.0005361 8334.994
#> theta[3, 3, 4] 1.0010345 8317.806
#> theta[3, 4, 1] 1.0009118 5521.717
#> theta[3, 4, 2] 1.0022782 4930.430
#> theta[3, 4, 3] 0.9995485 4395.860
#> theta[3, 4, 4] 1.0014834 5358.834
#> theta[4, 1, 1] 1.0004310 7110.047
#> theta[4, 1, 2] 1.0003175 5227.642
#> theta[4, 1, 3] 0.9996984 6035.778
#> theta[4, 1, 4] 0.9997627 5438.070
#> theta[4, 2, 1] 1.0022216 6416.841
#> theta[4, 2, 2] 1.0006738 6418.815
#> theta[4, 2, 3] 1.0004980 5713.876
#> theta[4, 2, 4] 1.0010769 5760.673
#> theta[4, 3, 1] 0.9996809 5404.196
#> theta[4, 3, 2] 0.9999138 5024.456
#> theta[4, 3, 3] 1.0008721 5102.808
#> theta[4, 3, 4] 1.0007778 4645.426
#> theta[4, 4, 1] 1.0005580 5487.109
#> theta[4, 4, 2] 0.9998393 5233.562
#> theta[4, 4, 3] 1.0000041 4773.638
#> theta[4, 4, 4] 1.0004247 5428.128
#> theta[5, 1, 1] 1.0004152 6670.374
#> theta[5, 1, 2] 1.0002211 4827.977
#> theta[5, 1, 3] 1.0012830 5037.650
#> theta[5, 1, 4] 1.0000926 5867.167
#> theta[5, 2, 1] 1.0007678 6405.852
#> theta[5, 2, 2] 1.0009105 6817.686
#> theta[5, 2, 3] 1.0011133 6545.292
#> theta[5, 2, 4] 1.0020515 6311.585
#> theta[5, 3, 1] 1.0011096 5471.940
#> theta[5, 3, 2] 1.0027187 6500.572
#> theta[5, 3, 3] 1.0000112 6012.249
#> theta[5, 3, 4] 1.0017117 6366.481
#> theta[5, 4, 1] 0.9999562 5395.173
#> theta[5, 4, 2] 1.0007153 5942.239
#> theta[5, 4, 3] 1.0022690 4653.921
#> theta[5, 4, 4] 1.0010906 5776.436

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>            Rhat ess_bulk
#> pi[1] 0.9998375 7183.681
#> pi[2] 1.0015102 6979.643
#> pi[3] 0.9995905 6633.003
#> pi[4] 1.0004142 6729.542

# }