Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000158 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.58 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.95406 seconds (Warm-up)
#> Chain 1:                1.87198 seconds (Sampling)
#> Chain 1:                3.82604 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000153 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.53 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.01005 seconds (Warm-up)
#> Chain 2:                2.08263 seconds (Sampling)
#> Chain 2:                4.09268 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000152 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.52 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.94617 seconds (Warm-up)
#> Chain 3:                1.91311 seconds (Sampling)
#> Chain 3:                3.85928 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000149 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.49 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.99312 seconds (Warm-up)
#> Chain 4:                1.80953 seconds (Sampling)
#> Chain 4:                3.80265 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0030501 5513.712
#> pi[2]          1.0045728 5163.812
#> pi[3]          1.0004878 4817.516
#> pi[4]          0.9996574 5278.621
#> theta[1, 1, 1] 1.0001728 5774.075
#> theta[1, 1, 2] 1.0010693 5917.597
#> theta[1, 1, 3] 1.0011611 4804.522
#> theta[1, 1, 4] 0.9998787 4791.610
#> theta[1, 2, 1] 0.9999269 4801.981
#> theta[1, 2, 2] 0.9997617 6018.070
#> theta[1, 2, 3] 1.0014243 6048.760
#> theta[1, 2, 4] 1.0001282 4837.922
#> theta[1, 3, 1] 1.0010460 4217.059
#> theta[1, 3, 2] 1.0017944 3598.761
#> theta[1, 3, 3] 1.0005526 4784.198
#> theta[1, 3, 4] 1.0004618 4158.830
#> theta[1, 4, 1] 1.0007899 4052.312
#> theta[1, 4, 2] 0.9997697 4743.851
#> theta[1, 4, 3] 1.0004782 2866.100
#> theta[1, 4, 4] 1.0013016 4417.085
#> theta[2, 1, 1] 1.0017014 6600.101
#> theta[2, 1, 2] 1.0014668 5998.649
#> theta[2, 1, 3] 1.0003921 4761.483
#> theta[2, 1, 4] 1.0005871 5003.506
#> theta[2, 2, 1] 0.9999218 5439.349
#> theta[2, 2, 2] 1.0025983 5332.748
#> theta[2, 2, 3] 1.0023100 5645.129
#> theta[2, 2, 4] 1.0002559 4311.645
#> theta[2, 3, 1] 1.0006731 4544.596
#> theta[2, 3, 2] 1.0006986 4053.019
#> theta[2, 3, 3] 1.0003481 5650.529
#> theta[2, 3, 4] 1.0008303 4294.015
#> theta[2, 4, 1] 1.0018190 4445.058
#> theta[2, 4, 2] 1.0003549 4479.280
#> theta[2, 4, 3] 1.0000603 4600.871
#> theta[2, 4, 4] 0.9999768 5092.168
#> theta[3, 1, 1] 1.0018005 5372.659
#> theta[3, 1, 2] 1.0006187 4370.669
#> theta[3, 1, 3] 1.0003852 4986.906
#> theta[3, 1, 4] 1.0004775 4358.664
#> theta[3, 2, 1] 1.0006477 4656.187
#> theta[3, 2, 2] 0.9997779 6552.183
#> theta[3, 2, 3] 1.0007737 5752.871
#> theta[3, 2, 4] 1.0008410 4222.362
#> theta[3, 3, 1] 0.9999467 4437.385
#> theta[3, 3, 2] 1.0009781 5043.994
#> theta[3, 3, 3] 1.0015580 5749.410
#> theta[3, 3, 4] 1.0018424 5868.132
#> theta[3, 4, 1] 1.0004622 4384.570
#> theta[3, 4, 2] 1.0003087 4699.339
#> theta[3, 4, 3] 1.0012343 3701.095
#> theta[3, 4, 4] 1.0023565 5344.542
#> theta[4, 1, 1] 1.0003044 6239.444
#> theta[4, 1, 2] 1.0000614 4990.934
#> theta[4, 1, 3] 1.0006631 5013.188
#> theta[4, 1, 4] 1.0013226 5281.717
#> theta[4, 2, 1] 1.0000833 4496.478
#> theta[4, 2, 2] 0.9994968 6218.712
#> theta[4, 2, 3] 1.0007089 5801.306
#> theta[4, 2, 4] 1.0002923 5054.835
#> theta[4, 3, 1] 1.0003659 3992.058
#> theta[4, 3, 2] 1.0001871 4794.812
#> theta[4, 3, 3] 1.0005730 5343.777
#> theta[4, 3, 4] 1.0009017 4246.364
#> theta[4, 4, 1] 0.9999840 4387.581
#> theta[4, 4, 2] 1.0008681 4592.206
#> theta[4, 4, 3] 1.0010415 4268.747
#> theta[4, 4, 4] 1.0015278 6340.826
#> theta[5, 1, 1] 1.0014915 5621.825
#> theta[5, 1, 2] 1.0011590 4385.311
#> theta[5, 1, 3] 0.9997931 5265.007
#> theta[5, 1, 4] 1.0010860 4578.419
#> theta[5, 2, 1] 1.0045609 6542.203
#> theta[5, 2, 2] 1.0026967 5804.642
#> theta[5, 2, 3] 0.9995227 5428.286
#> theta[5, 2, 4] 1.0005688 4047.995
#> theta[5, 3, 1] 1.0001561 4262.575
#> theta[5, 3, 2] 0.9998908 5028.877
#> theta[5, 3, 3] 1.0021365 5300.921
#> theta[5, 3, 4] 1.0005232 5926.590
#> theta[5, 4, 1] 1.0033174 4500.374
#> theta[5, 4, 2] 1.0018219 4014.859
#> theta[5, 4, 3] 1.0000868 3981.792
#> theta[5, 4, 4] 1.0012060 5522.451

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>            Rhat ess_bulk
#> pi[1] 1.0030501 5513.712
#> pi[2] 1.0045728 5163.812
#> pi[3] 1.0004878 4817.516
#> pi[4] 0.9996574 5278.621

# }