Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000201 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 2.01 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.597 seconds (Warm-up)
#> Chain 1:                2.731 seconds (Sampling)
#> Chain 1:                5.328 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000195 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.95 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.516 seconds (Warm-up)
#> Chain 2:                2.495 seconds (Sampling)
#> Chain 2:                5.011 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000189 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.89 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.54 seconds (Warm-up)
#> Chain 3:                2.843 seconds (Sampling)
#> Chain 3:                5.383 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000183 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.83 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.645 seconds (Warm-up)
#> Chain 4:                2.296 seconds (Sampling)
#> Chain 4:                4.941 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0006000 7141.764
#> pi[2]          1.0010948 6721.513
#> pi[3]          1.0006965 6351.449
#> pi[4]          1.0022625 6234.604
#> theta[1, 1, 1] 1.0006568 5847.838
#> theta[1, 1, 2] 1.0005269 5032.068
#> theta[1, 1, 3] 1.0003299 5277.067
#> theta[1, 1, 4] 1.0004798 5205.418
#> theta[1, 2, 1] 1.0013000 4651.873
#> theta[1, 2, 2] 0.9998512 5809.832
#> theta[1, 2, 3] 1.0000689 6357.833
#> theta[1, 2, 4] 1.0007954 4802.791
#> theta[1, 3, 1] 1.0008157 4327.047
#> theta[1, 3, 2] 1.0008619 3261.122
#> theta[1, 3, 3] 0.9997823 4357.735
#> theta[1, 3, 4] 1.0012698 3776.340
#> theta[1, 4, 1] 1.0021440 4549.889
#> theta[1, 4, 2] 1.0003355 5302.485
#> theta[1, 4, 3] 1.0010381 3242.016
#> theta[1, 4, 4] 1.0004200 4762.703
#> theta[2, 1, 1] 1.0016981 7400.074
#> theta[2, 1, 2] 1.0001096 6552.046
#> theta[2, 1, 3] 1.0007646 5161.424
#> theta[2, 1, 4] 1.0010988 4965.663
#> theta[2, 2, 1] 1.0014337 6014.553
#> theta[2, 2, 2] 1.0032951 6489.142
#> theta[2, 2, 3] 1.0007470 5717.631
#> theta[2, 2, 4] 1.0005446 4362.248
#> theta[2, 3, 1] 1.0018645 4506.619
#> theta[2, 3, 2] 1.0020501 4842.919
#> theta[2, 3, 3] 1.0007698 4854.336
#> theta[2, 3, 4] 1.0004235 4129.085
#> theta[2, 4, 1] 1.0003175 4887.984
#> theta[2, 4, 2] 1.0006971 4603.666
#> theta[2, 4, 3] 1.0005450 4872.714
#> theta[2, 4, 4] 1.0004961 5483.140
#> theta[3, 1, 1] 1.0032773 6792.911
#> theta[3, 1, 2] 0.9996694 4715.584
#> theta[3, 1, 3] 1.0010416 5008.380
#> theta[3, 1, 4] 0.9996454 5786.528
#> theta[3, 2, 1] 1.0024080 5601.803
#> theta[3, 2, 2] 1.0015398 6029.880
#> theta[3, 2, 3] 1.0007368 5130.739
#> theta[3, 2, 4] 1.0003655 5173.223
#> theta[3, 3, 1] 0.9999564 4475.181
#> theta[3, 3, 2] 1.0030575 5427.703
#> theta[3, 3, 3] 1.0001390 6674.782
#> theta[3, 3, 4] 1.0019286 7011.947
#> theta[3, 4, 1] 1.0008827 5266.840
#> theta[3, 4, 2] 1.0000786 5421.650
#> theta[3, 4, 3] 1.0019159 4082.875
#> theta[3, 4, 4] 1.0019950 5319.408
#> theta[4, 1, 1] 1.0003678 5963.397
#> theta[4, 1, 2] 1.0002716 5693.596
#> theta[4, 1, 3] 1.0003976 5181.431
#> theta[4, 1, 4] 1.0005481 5261.273
#> theta[4, 2, 1] 1.0012311 4786.720
#> theta[4, 2, 2] 1.0003282 5603.968
#> theta[4, 2, 3] 1.0012056 6776.391
#> theta[4, 2, 4] 0.9995761 5315.514
#> theta[4, 3, 1] 0.9998100 4169.785
#> theta[4, 3, 2] 1.0010581 4989.647
#> theta[4, 3, 3] 0.9997871 5392.192
#> theta[4, 3, 4] 1.0004210 3880.710
#> theta[4, 4, 1] 0.9996227 5192.169
#> theta[4, 4, 2] 1.0009322 5058.403
#> theta[4, 4, 3] 1.0002379 4907.617
#> theta[4, 4, 4] 1.0003436 6281.836
#> theta[5, 1, 1] 1.0032242 6268.567
#> theta[5, 1, 2] 1.0025495 4483.326
#> theta[5, 1, 3] 1.0003856 5458.949
#> theta[5, 1, 4] 0.9993212 4837.476
#> theta[5, 2, 1] 1.0008053 4878.475
#> theta[5, 2, 2] 1.0002403 6093.868
#> theta[5, 2, 3] 1.0016245 6037.418
#> theta[5, 2, 4] 0.9998664 4600.833
#> theta[5, 3, 1] 1.0002521 5120.423
#> theta[5, 3, 2] 1.0009285 5871.810
#> theta[5, 3, 3] 1.0040256 7557.733
#> theta[5, 3, 4] 1.0006772 5238.616
#> theta[5, 4, 1] 1.0004621 4976.349
#> theta[5, 4, 2] 1.0023488 4866.789
#> theta[5, 4, 3] 1.0011304 3704.045
#> theta[5, 4, 4] 1.0013011 5500.645

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.000600 7141.764
#> pi[2] 1.001095 6721.513
#> pi[3] 1.000696 6351.449
#> pi[4] 1.002262 6234.604

# }