Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000171 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.71 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.15021 seconds (Warm-up)
#> Chain 1:                2.22483 seconds (Sampling)
#> Chain 1:                4.37503 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000145 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.45 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.07114 seconds (Warm-up)
#> Chain 2:                2.31179 seconds (Sampling)
#> Chain 2:                4.38293 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.00014 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.4 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.10707 seconds (Warm-up)
#> Chain 3:                2.28882 seconds (Sampling)
#> Chain 3:                4.39589 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000403 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 4.03 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.16786 seconds (Warm-up)
#> Chain 4:                2.25281 seconds (Sampling)
#> Chain 4:                4.42066 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat  ess_bulk
#> pi[1]          0.9996421 10609.313
#> pi[2]          1.0004682  8062.480
#> pi[3]          1.0018600  7315.359
#> pi[4]          1.0012462  7004.990
#> theta[1, 1, 1] 1.0002044  8714.866
#> theta[1, 1, 2] 1.0003253  7539.491
#> theta[1, 1, 3] 0.9997663  6615.293
#> theta[1, 1, 4] 1.0001319  5381.812
#> theta[1, 2, 1] 1.0005781  5566.186
#> theta[1, 2, 2] 1.0016556  6936.434
#> theta[1, 2, 3] 1.0007340  6949.520
#> theta[1, 2, 4] 0.9998842  6131.293
#> theta[1, 3, 1] 0.9994735  6006.137
#> theta[1, 3, 2] 1.0001685  4456.434
#> theta[1, 3, 3] 0.9999588  4505.247
#> theta[1, 3, 4] 1.0016768  4424.613
#> theta[1, 4, 1] 1.0031537  4775.319
#> theta[1, 4, 2] 0.9997572  6255.868
#> theta[1, 4, 3] 1.0014035  3913.259
#> theta[1, 4, 4] 1.0003720  5038.455
#> theta[2, 1, 1] 1.0007095  8820.369
#> theta[2, 1, 2] 1.0010120  7647.352
#> theta[2, 1, 3] 1.0010904  5485.803
#> theta[2, 1, 4] 1.0005987  5774.847
#> theta[2, 2, 1] 1.0002407  6433.127
#> theta[2, 2, 2] 1.0010339  7448.830
#> theta[2, 2, 3] 1.0031843  7359.373
#> theta[2, 2, 4] 0.9995525  5420.659
#> theta[2, 3, 1] 1.0004336  4990.493
#> theta[2, 3, 2] 1.0006211  5602.592
#> theta[2, 3, 3] 1.0002119  6219.580
#> theta[2, 3, 4] 1.0016557  5870.491
#> theta[2, 4, 1] 1.0001179  4757.385
#> theta[2, 4, 2] 0.9999886  5307.595
#> theta[2, 4, 3] 1.0006148  5742.714
#> theta[2, 4, 4] 1.0003317  5782.123
#> theta[3, 1, 1] 1.0017155  6817.705
#> theta[3, 1, 2] 1.0004357  4514.232
#> theta[3, 1, 3] 1.0012592  5607.431
#> theta[3, 1, 4] 0.9999250  5746.675
#> theta[3, 2, 1] 1.0008324  7364.884
#> theta[3, 2, 2] 1.0001047  9115.376
#> theta[3, 2, 3] 1.0001516  8338.128
#> theta[3, 2, 4] 1.0010172  5585.921
#> theta[3, 3, 1] 1.0002239  5044.379
#> theta[3, 3, 2] 1.0019305  6783.011
#> theta[3, 3, 3] 1.0015452  6900.522
#> theta[3, 3, 4] 1.0007319  7375.282
#> theta[3, 4, 1] 1.0005322  6304.130
#> theta[3, 4, 2] 1.0015466  5245.470
#> theta[3, 4, 3] 1.0006608  5095.498
#> theta[3, 4, 4] 1.0003694  6650.704
#> theta[4, 1, 1] 1.0034384  8046.430
#> theta[4, 1, 2] 0.9995528  6538.851
#> theta[4, 1, 3] 1.0000810  5980.952
#> theta[4, 1, 4] 1.0026057  6711.352
#> theta[4, 2, 1] 1.0007840  7370.794
#> theta[4, 2, 2] 1.0007502  8833.066
#> theta[4, 2, 3] 1.0008245  7249.872
#> theta[4, 2, 4] 1.0029403  4531.455
#> theta[4, 3, 1] 1.0008407  5577.312
#> theta[4, 3, 2] 0.9997186  5990.064
#> theta[4, 3, 3] 1.0019782  5844.835
#> theta[4, 3, 4] 1.0012733  4456.896
#> theta[4, 4, 1] 1.0004855  5166.323
#> theta[4, 4, 2] 1.0000277  6221.118
#> theta[4, 4, 3] 1.0002524  5633.879
#> theta[4, 4, 4] 0.9998738  6281.064
#> theta[5, 1, 1] 0.9998538  7318.744
#> theta[5, 1, 2] 0.9996610  5976.486
#> theta[5, 1, 3] 1.0004888  5535.387
#> theta[5, 1, 4] 1.0010939  5557.467
#> theta[5, 2, 1] 1.0016227  8258.068
#> theta[5, 2, 2] 1.0009940  8597.280
#> theta[5, 2, 3] 1.0005086  6831.006
#> theta[5, 2, 4] 1.0022128  5504.210
#> theta[5, 3, 1] 1.0002437  5188.781
#> theta[5, 3, 2] 1.0008758  5472.750
#> theta[5, 3, 3] 0.9999097  6338.034
#> theta[5, 3, 4] 1.0016121  6494.801
#> theta[5, 4, 1] 0.9997028  5725.454
#> theta[5, 4, 2] 1.0003676  5401.387
#> theta[5, 4, 3] 1.0019944  4990.740
#> theta[5, 4, 4] 1.0016276  6102.050

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>            Rhat  ess_bulk
#> pi[1] 0.9996421 10609.313
#> pi[2] 1.0004682  8062.480
#> pi[3] 1.0018600  7315.359
#> pi[4] 1.0012462  7004.990

# }