Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000152 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.52 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.59973 seconds (Warm-up)
#> Chain 1:                1.76521 seconds (Sampling)
#> Chain 1:                3.36494 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000134 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.34 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.61614 seconds (Warm-up)
#> Chain 2:                1.75897 seconds (Sampling)
#> Chain 2:                3.37511 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000131 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.31 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.65002 seconds (Warm-up)
#> Chain 3:                1.78337 seconds (Sampling)
#> Chain 3:                3.43338 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.00013 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.3 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.60692 seconds (Warm-up)
#> Chain 4:                1.72223 seconds (Sampling)
#> Chain 4:                3.32915 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat  ess_bulk
#> pi[1]          1.0002867  7934.841
#> pi[2]          1.0010575  6883.173
#> pi[3]          1.0009902  7453.468
#> pi[4]          1.0019558  7107.325
#> theta[1, 1, 1] 1.0009929 10518.781
#> theta[1, 1, 2] 1.0002934 10312.016
#> theta[1, 1, 3] 1.0018452  5401.745
#> theta[1, 1, 4] 1.0000756  6456.882
#> theta[1, 2, 1] 1.0039288  6181.532
#> theta[1, 2, 2] 1.0005271  7579.039
#> theta[1, 2, 3] 0.9999475  7502.916
#> theta[1, 2, 4] 1.0002858  5370.951
#> theta[1, 3, 1] 1.0014877  5836.137
#> theta[1, 3, 2] 1.0004983  3708.938
#> theta[1, 3, 3] 1.0011147  4902.745
#> theta[1, 3, 4] 1.0007132  4671.710
#> theta[1, 4, 1] 1.0004473  4940.530
#> theta[1, 4, 2] 1.0001669  5158.790
#> theta[1, 4, 3] 1.0014484  3789.181
#> theta[1, 4, 4] 1.0006148  5540.587
#> theta[2, 1, 1] 1.0011089  8190.429
#> theta[2, 1, 2] 1.0001296  7585.820
#> theta[2, 1, 3] 1.0007256  6019.378
#> theta[2, 1, 4] 1.0001049  5880.591
#> theta[2, 2, 1] 0.9996765  6658.583
#> theta[2, 2, 2] 1.0015314  9469.505
#> theta[2, 2, 3] 1.0001283  8580.895
#> theta[2, 2, 4] 0.9999528  6309.201
#> theta[2, 3, 1] 1.0002965  5715.789
#> theta[2, 3, 2] 1.0001920  5728.453
#> theta[2, 3, 3] 1.0015667  5881.778
#> theta[2, 3, 4] 1.0031574  5846.569
#> theta[2, 4, 1] 1.0022142  5568.456
#> theta[2, 4, 2] 1.0012752  5325.851
#> theta[2, 4, 3] 1.0005214  5339.850
#> theta[2, 4, 4] 1.0021993  7133.834
#> theta[3, 1, 1] 1.0039885  7491.590
#> theta[3, 1, 2] 1.0010734  6157.351
#> theta[3, 1, 3] 1.0016937  6812.242
#> theta[3, 1, 4] 1.0000777  5300.609
#> theta[3, 2, 1] 1.0006738  6434.711
#> theta[3, 2, 2] 0.9995210  7974.637
#> theta[3, 2, 3] 1.0009190  7937.381
#> theta[3, 2, 4] 0.9999510  6319.433
#> theta[3, 3, 1] 1.0008019  5357.260
#> theta[3, 3, 2] 1.0003668  7373.219
#> theta[3, 3, 3] 1.0014700  7304.876
#> theta[3, 3, 4] 1.0033032  8442.610
#> theta[3, 4, 1] 1.0011686  5807.867
#> theta[3, 4, 2] 1.0012014  6020.635
#> theta[3, 4, 3] 1.0003616  5859.768
#> theta[3, 4, 4] 0.9998776  6523.002
#> theta[4, 1, 1] 1.0019090  7817.514
#> theta[4, 1, 2] 1.0001104  6643.346
#> theta[4, 1, 3] 1.0007739  6592.024
#> theta[4, 1, 4] 1.0027393  6040.453
#> theta[4, 2, 1] 1.0002192  6459.537
#> theta[4, 2, 2] 1.0005939  6915.522
#> theta[4, 2, 3] 1.0018645  6211.937
#> theta[4, 2, 4] 1.0004304  6859.089
#> theta[4, 3, 1] 1.0002513  4857.290
#> theta[4, 3, 2] 1.0025306  5368.864
#> theta[4, 3, 3] 1.0009921  5763.606
#> theta[4, 3, 4] 1.0000553  5194.991
#> theta[4, 4, 1] 1.0008690  6385.148
#> theta[4, 4, 2] 1.0015151  6553.905
#> theta[4, 4, 3] 1.0008078  5548.788
#> theta[4, 4, 4] 1.0010577  6307.681
#> theta[5, 1, 1] 1.0027263  8055.111
#> theta[5, 1, 2] 1.0003597  5199.932
#> theta[5, 1, 3] 0.9997682  6180.656
#> theta[5, 1, 4] 1.0001331  5469.357
#> theta[5, 2, 1] 1.0019410  8508.360
#> theta[5, 2, 2] 0.9999634  6923.772
#> theta[5, 2, 3] 1.0003155  6752.121
#> theta[5, 2, 4] 1.0025087  5439.271
#> theta[5, 3, 1] 0.9998548  5920.522
#> theta[5, 3, 2] 1.0000766  7062.662
#> theta[5, 3, 3] 1.0001873  8051.288
#> theta[5, 3, 4] 1.0003693  8059.875
#> theta[5, 4, 1] 1.0006360  5523.703
#> theta[5, 4, 2] 0.9999645  5370.571
#> theta[5, 4, 3] 1.0004631  5225.000
#> theta[5, 4, 4] 1.0001922  6730.155

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.000287 7934.841
#> pi[2] 1.001057 6883.173
#> pi[3] 1.000990 7453.468
#> pi[4] 1.001956 7107.325

# }