Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000127 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.27 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.52016 seconds (Warm-up)
#> Chain 1:                1.61943 seconds (Sampling)
#> Chain 1:                3.1396 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000117 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.17 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.47254 seconds (Warm-up)
#> Chain 2:                1.63991 seconds (Sampling)
#> Chain 2:                3.11244 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000115 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.15 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.51361 seconds (Warm-up)
#> Chain 3:                1.63477 seconds (Sampling)
#> Chain 3:                3.14838 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000111 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.11 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.4447 seconds (Warm-up)
#> Chain 4:                1.62833 seconds (Sampling)
#> Chain 4:                3.07303 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat  ess_bulk
#> pi[1]          1.0015755  9956.955
#> pi[2]          1.0000246  9097.961
#> pi[3]          1.0009797  7135.302
#> pi[4]          1.0010495  7651.947
#> theta[1, 1, 1] 1.0016123  9995.851
#> theta[1, 1, 2] 1.0006811  9022.517
#> theta[1, 1, 3] 0.9997319  6301.164
#> theta[1, 1, 4] 1.0000790  7527.943
#> theta[1, 2, 1] 1.0022443  7394.034
#> theta[1, 2, 2] 1.0001018 10547.453
#> theta[1, 2, 3] 1.0026879  8950.783
#> theta[1, 2, 4] 1.0003950  6088.378
#> theta[1, 3, 1] 1.0023768  5578.200
#> theta[1, 3, 2] 0.9995137  4213.944
#> theta[1, 3, 3] 1.0008470  5697.030
#> theta[1, 3, 4] 0.9997843  4584.733
#> theta[1, 4, 1] 1.0001811  6263.058
#> theta[1, 4, 2] 1.0011916  6167.632
#> theta[1, 4, 3] 1.0017163  4219.707
#> theta[1, 4, 4] 1.0013764  5889.365
#> theta[2, 1, 1] 1.0044548 11298.722
#> theta[2, 1, 2] 1.0020861  8300.792
#> theta[2, 1, 3] 1.0031879  5829.984
#> theta[2, 1, 4] 1.0011848  6661.557
#> theta[2, 2, 1] 1.0016098  8882.041
#> theta[2, 2, 2] 1.0018583  9249.005
#> theta[2, 2, 3] 1.0014433  9063.989
#> theta[2, 2, 4] 1.0013775  6880.599
#> theta[2, 3, 1] 0.9999189  6502.183
#> theta[2, 3, 2] 1.0007130  6721.561
#> theta[2, 3, 3] 1.0004553  6009.258
#> theta[2, 3, 4] 0.9994714  5095.265
#> theta[2, 4, 1] 1.0019947  6736.432
#> theta[2, 4, 2] 1.0006171  7514.697
#> theta[2, 4, 3] 1.0005004  6179.415
#> theta[2, 4, 4] 0.9998817  7777.371
#> theta[3, 1, 1] 1.0015859  9175.743
#> theta[3, 1, 2] 1.0026158  8488.298
#> theta[3, 1, 3] 1.0004138  5376.581
#> theta[3, 1, 4] 0.9997465  6575.162
#> theta[3, 2, 1] 1.0006392  6844.706
#> theta[3, 2, 2] 1.0007500  7735.441
#> theta[3, 2, 3] 0.9999773  6878.964
#> theta[3, 2, 4] 0.9996258  6347.127
#> theta[3, 3, 1] 1.0004105  5585.064
#> theta[3, 3, 2] 1.0023043  6827.538
#> theta[3, 3, 3] 1.0010031  7613.341
#> theta[3, 3, 4] 1.0012336  9295.478
#> theta[3, 4, 1] 1.0009746  5991.496
#> theta[3, 4, 2] 1.0003112  6209.669
#> theta[3, 4, 3] 1.0004619  5186.950
#> theta[3, 4, 4] 1.0002488  6664.755
#> theta[4, 1, 1] 1.0020020  6936.212
#> theta[4, 1, 2] 0.9999450  7422.434
#> theta[4, 1, 3] 1.0002052  5575.693
#> theta[4, 1, 4] 1.0011223  6233.162
#> theta[4, 2, 1] 0.9996543  7571.414
#> theta[4, 2, 2] 1.0024743  7996.513
#> theta[4, 2, 3] 1.0002761  8087.090
#> theta[4, 2, 4] 1.0000667  6120.656
#> theta[4, 3, 1] 1.0003556  5815.117
#> theta[4, 3, 2] 1.0025777  6324.829
#> theta[4, 3, 3] 0.9998084  7009.539
#> theta[4, 3, 4] 1.0003418  4855.075
#> theta[4, 4, 1] 0.9999141  6562.431
#> theta[4, 4, 2] 1.0002112  6800.191
#> theta[4, 4, 3] 1.0005248  6319.250
#> theta[4, 4, 4] 1.0002179  7135.092
#> theta[5, 1, 1] 1.0006418  7653.622
#> theta[5, 1, 2] 0.9995739  5439.966
#> theta[5, 1, 3] 1.0002715  5754.924
#> theta[5, 1, 4] 0.9997124  6044.584
#> theta[5, 2, 1] 1.0006379  8222.924
#> theta[5, 2, 2] 1.0001925  9910.156
#> theta[5, 2, 3] 1.0000747  7818.784
#> theta[5, 2, 4] 0.9993330  6751.588
#> theta[5, 3, 1] 1.0000938  6357.399
#> theta[5, 3, 2] 1.0001527  8257.457
#> theta[5, 3, 3] 1.0018559  7453.491
#> theta[5, 3, 4] 1.0022188  7387.092
#> theta[5, 4, 1] 1.0008148  6275.261
#> theta[5, 4, 2] 1.0012513  5960.780
#> theta[5, 4, 3] 1.0024880  6252.502
#> theta[5, 4, 4] 1.0007528  6091.647

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.001576 9956.955
#> pi[2] 1.000025 9097.961
#> pi[3] 1.000980 7135.302
#> pi[4] 1.001049 7651.947

# }