Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000131 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.31 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.52207 seconds (Warm-up)
#> Chain 1:                1.59404 seconds (Sampling)
#> Chain 1:                3.1161 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000117 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.17 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.51463 seconds (Warm-up)
#> Chain 2:                1.29544 seconds (Sampling)
#> Chain 2:                2.81007 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000114 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.14 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.48794 seconds (Warm-up)
#> Chain 3:                1.64186 seconds (Sampling)
#> Chain 3:                3.12979 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000114 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.14 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.4896 seconds (Warm-up)
#> Chain 4:                1.63837 seconds (Sampling)
#> Chain 4:                3.12798 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0019524 7955.635
#> pi[2]          1.0010802 6627.389
#> pi[3]          1.0017224 5951.287
#> pi[4]          1.0001460 6721.394
#> theta[1, 1, 1] 1.0015963 7397.461
#> theta[1, 1, 2] 1.0031177 6650.995
#> theta[1, 1, 3] 1.0022929 4873.885
#> theta[1, 1, 4] 1.0005955 5934.695
#> theta[1, 2, 1] 1.0004950 4994.314
#> theta[1, 2, 2] 1.0021764 6471.851
#> theta[1, 2, 3] 1.0001664 6084.526
#> theta[1, 2, 4] 1.0005388 5757.993
#> theta[1, 3, 1] 1.0006360 5043.426
#> theta[1, 3, 2] 0.9998156 3729.514
#> theta[1, 3, 3] 0.9997382 5305.999
#> theta[1, 3, 4] 1.0019332 4481.559
#> theta[1, 4, 1] 1.0017964 5116.379
#> theta[1, 4, 2] 1.0007027 5632.436
#> theta[1, 4, 3] 1.0008043 4028.030
#> theta[1, 4, 4] 1.0018512 5117.479
#> theta[2, 1, 1] 1.0012458 6913.875
#> theta[2, 1, 2] 1.0004191 6478.951
#> theta[2, 1, 3] 1.0004634 4836.751
#> theta[2, 1, 4] 1.0001979 5589.248
#> theta[2, 2, 1] 1.0016482 6628.282
#> theta[2, 2, 2] 1.0033928 7500.293
#> theta[2, 2, 3] 1.0020374 7439.314
#> theta[2, 2, 4] 1.0000952 5410.529
#> theta[2, 3, 1] 1.0002091 4941.063
#> theta[2, 3, 2] 1.0020469 4589.149
#> theta[2, 3, 3] 1.0006502 6469.682
#> theta[2, 3, 4] 1.0000052 6080.533
#> theta[2, 4, 1] 0.9999769 5949.132
#> theta[2, 4, 2] 1.0017476 5072.227
#> theta[2, 4, 3] 1.0005866 4129.577
#> theta[2, 4, 4] 1.0001752 5547.626
#> theta[3, 1, 1] 1.0024268 7378.551
#> theta[3, 1, 2] 1.0009927 4366.902
#> theta[3, 1, 3] 1.0014488 5647.611
#> theta[3, 1, 4] 1.0021375 5581.785
#> theta[3, 2, 1] 1.0022744 5356.867
#> theta[3, 2, 2] 1.0025813 7956.542
#> theta[3, 2, 3] 1.0008091 6378.669
#> theta[3, 2, 4] 1.0002583 5426.522
#> theta[3, 3, 1] 1.0000735 4555.603
#> theta[3, 3, 2] 1.0005995 6418.186
#> theta[3, 3, 3] 1.0018021 6173.701
#> theta[3, 3, 4] 1.0010972 8508.720
#> theta[3, 4, 1] 0.9997612 5184.878
#> theta[3, 4, 2] 1.0006686 5165.762
#> theta[3, 4, 3] 1.0015583 4636.745
#> theta[3, 4, 4] 1.0004335 5711.864
#> theta[4, 1, 1] 1.0010248 6689.945
#> theta[4, 1, 2] 1.0001986 6183.122
#> theta[4, 1, 3] 1.0010005 5838.185
#> theta[4, 1, 4] 1.0024120 5342.802
#> theta[4, 2, 1] 1.0035803 5516.524
#> theta[4, 2, 2] 1.0016268 7069.395
#> theta[4, 2, 3] 1.0006976 6005.446
#> theta[4, 2, 4] 0.9996828 5001.780
#> theta[4, 3, 1] 0.9998663 5111.375
#> theta[4, 3, 2] 1.0001251 5284.109
#> theta[4, 3, 3] 0.9998745 5264.366
#> theta[4, 3, 4] 0.9996961 3978.154
#> theta[4, 4, 1] 1.0010903 6027.355
#> theta[4, 4, 2] 1.0006483 5602.846
#> theta[4, 4, 3] 1.0024464 4504.086
#> theta[4, 4, 4] 1.0016158 6654.612
#> theta[5, 1, 1] 0.9998461 6892.595
#> theta[5, 1, 2] 1.0014846 4861.585
#> theta[5, 1, 3] 0.9994365 5878.744
#> theta[5, 1, 4] 1.0002404 5319.606
#> theta[5, 2, 1] 1.0001736 7101.984
#> theta[5, 2, 2] 1.0026280 6862.718
#> theta[5, 2, 3] 1.0008180 6272.916
#> theta[5, 2, 4] 1.0012975 4994.301
#> theta[5, 3, 1] 1.0007723 5414.154
#> theta[5, 3, 2] 1.0041489 7191.754
#> theta[5, 3, 3] 1.0004360 6344.346
#> theta[5, 3, 4] 1.0012628 4935.890
#> theta[5, 4, 1] 1.0004651 4695.048
#> theta[5, 4, 2] 1.0014072 5358.093
#> theta[5, 4, 3] 1.0001908 4241.155
#> theta[5, 4, 4] 1.0006783 4569.006

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.001952 7955.635
#> pi[2] 1.001080 6627.389
#> pi[3] 1.001722 5951.287
#> pi[4] 1.000146 6721.394

# }