Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.00013 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.3 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.43632 seconds (Warm-up)
#> Chain 1:                1.58829 seconds (Sampling)
#> Chain 1:                3.02462 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000119 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.19 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.49394 seconds (Warm-up)
#> Chain 2:                1.38293 seconds (Sampling)
#> Chain 2:                2.87687 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000119 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.19 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.50652 seconds (Warm-up)
#> Chain 3:                1.23663 seconds (Sampling)
#> Chain 3:                2.74316 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000114 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.14 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.50363 seconds (Warm-up)
#> Chain 4:                1.64596 seconds (Sampling)
#> Chain 4:                3.14959 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0003477 7091.958
#> pi[2]          1.0023954 6270.807
#> pi[3]          1.0032908 5214.839
#> pi[4]          1.0005744 4391.719
#> theta[1, 1, 1] 1.0017291 5600.181
#> theta[1, 1, 2] 1.0014355 4895.555
#> theta[1, 1, 3] 0.9999490 4709.059
#> theta[1, 1, 4] 1.0017101 4645.913
#> theta[1, 2, 1] 1.0005590 3932.253
#> theta[1, 2, 2] 1.0012958 6288.209
#> theta[1, 2, 3] 1.0005080 6677.702
#> theta[1, 2, 4] 1.0015445 4985.983
#> theta[1, 3, 1] 1.0007752 5114.146
#> theta[1, 3, 2] 1.0004356 3455.005
#> theta[1, 3, 3] 1.0006698 4958.419
#> theta[1, 3, 4] 1.0001210 4084.271
#> theta[1, 4, 1] 0.9996989 4185.239
#> theta[1, 4, 2] 1.0025124 4536.604
#> theta[1, 4, 3] 0.9997941 3487.533
#> theta[1, 4, 4] 1.0023913 5044.391
#> theta[2, 1, 1] 1.0009825 6858.696
#> theta[2, 1, 2] 1.0004886 5672.066
#> theta[2, 1, 3] 1.0023073 4930.088
#> theta[2, 1, 4] 1.0006403 4826.269
#> theta[2, 2, 1] 1.0007421 4981.641
#> theta[2, 2, 2] 0.9998235 6356.464
#> theta[2, 2, 3] 1.0013346 6026.134
#> theta[2, 2, 4] 1.0004757 4300.489
#> theta[2, 3, 1] 1.0010731 4048.232
#> theta[2, 3, 2] 1.0001745 4232.439
#> theta[2, 3, 3] 1.0020530 5665.313
#> theta[2, 3, 4] 1.0020078 4855.731
#> theta[2, 4, 1] 0.9999918 5011.276
#> theta[2, 4, 2] 1.0005051 4491.817
#> theta[2, 4, 3] 1.0006191 4819.865
#> theta[2, 4, 4] 1.0017212 5440.548
#> theta[3, 1, 1] 0.9997368 5022.377
#> theta[3, 1, 2] 1.0001278 4898.862
#> theta[3, 1, 3] 1.0018494 5392.018
#> theta[3, 1, 4] 1.0007792 4219.452
#> theta[3, 2, 1] 1.0013604 4836.593
#> theta[3, 2, 2] 0.9994051 6101.600
#> theta[3, 2, 3] 1.0011420 5976.713
#> theta[3, 2, 4] 1.0012034 4769.231
#> theta[3, 3, 1] 1.0004885 4369.564
#> theta[3, 3, 2] 1.0012018 4841.751
#> theta[3, 3, 3] 1.0009320 5548.787
#> theta[3, 3, 4] 1.0011835 5377.002
#> theta[3, 4, 1] 1.0004504 5431.908
#> theta[3, 4, 2] 1.0006473 4349.938
#> theta[3, 4, 3] 1.0000273 3992.398
#> theta[3, 4, 4] 0.9997346 5548.592
#> theta[4, 1, 1] 1.0007214 6202.430
#> theta[4, 1, 2] 1.0002329 4612.403
#> theta[4, 1, 3] 1.0006526 5596.477
#> theta[4, 1, 4] 0.9999799 4888.872
#> theta[4, 2, 1] 0.9998082 5508.598
#> theta[4, 2, 2] 1.0001346 5705.769
#> theta[4, 2, 3] 1.0004546 5456.790
#> theta[4, 2, 4] 1.0010118 4622.914
#> theta[4, 3, 1] 1.0014676 4571.032
#> theta[4, 3, 2] 0.9999028 4292.117
#> theta[4, 3, 3] 1.0003520 5077.224
#> theta[4, 3, 4] 1.0000129 3207.943
#> theta[4, 4, 1] 0.9998792 4602.989
#> theta[4, 4, 2] 1.0023622 3972.217
#> theta[4, 4, 3] 1.0019949 5947.738
#> theta[4, 4, 4] 1.0011856 6265.719
#> theta[5, 1, 1] 1.0004091 5058.199
#> theta[5, 1, 2] 1.0006581 2980.369
#> theta[5, 1, 3] 0.9997360 5228.455
#> theta[5, 1, 4] 1.0000575 4237.730
#> theta[5, 2, 1] 0.9999653 5788.491
#> theta[5, 2, 2] 1.0003334 6048.309
#> theta[5, 2, 3] 0.9997463 5449.161
#> theta[5, 2, 4] 0.9998186 4795.622
#> theta[5, 3, 1] 1.0002874 4902.304
#> theta[5, 3, 2] 1.0005356 5906.125
#> theta[5, 3, 3] 1.0020411 6375.375
#> theta[5, 3, 4] 1.0011974 5658.489
#> theta[5, 4, 1] 1.0003697 4336.734
#> theta[5, 4, 2] 1.0001252 3410.363
#> theta[5, 4, 3] 1.0015606 3971.659
#> theta[5, 4, 4] 1.0022329 4691.621

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.000348 7091.958
#> pi[2] 1.002395 6270.807
#> pi[3] 1.003291 5214.839
#> pi[4] 1.000574 4391.719

# }