Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000167 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.67 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.11299 seconds (Warm-up)
#> Chain 1:                2.09275 seconds (Sampling)
#> Chain 1:                4.20574 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000188 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.88 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.17197 seconds (Warm-up)
#> Chain 2:                2.00084 seconds (Sampling)
#> Chain 2:                4.17281 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000141 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.41 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.02116 seconds (Warm-up)
#> Chain 3:                2.05661 seconds (Sampling)
#> Chain 3:                4.07777 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000143 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.43 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.98394 seconds (Warm-up)
#> Chain 4:                2.38202 seconds (Sampling)
#> Chain 4:                4.36596 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0000021 6992.163
#> pi[2]          1.0008853 6466.304
#> pi[3]          1.0014086 5871.148
#> pi[4]          1.0017515 6453.073
#> theta[1, 1, 1] 1.0023336 6222.010
#> theta[1, 1, 2] 1.0010335 5634.235
#> theta[1, 1, 3] 1.0012484 5608.965
#> theta[1, 1, 4] 0.9996442 4616.276
#> theta[1, 2, 1] 1.0015354 5051.344
#> theta[1, 2, 2] 1.0008519 6001.521
#> theta[1, 2, 3] 1.0013150 5574.729
#> theta[1, 2, 4] 1.0002788 5217.568
#> theta[1, 3, 1] 1.0009287 4635.587
#> theta[1, 3, 2] 1.0007963 3848.422
#> theta[1, 3, 3] 1.0006008 4805.801
#> theta[1, 3, 4] 1.0010215 4572.261
#> theta[1, 4, 1] 0.9997914 4770.472
#> theta[1, 4, 2] 1.0002584 4918.826
#> theta[1, 4, 3] 0.9999821 3367.038
#> theta[1, 4, 4] 1.0024920 5183.055
#> theta[2, 1, 1] 1.0002675 6604.837
#> theta[2, 1, 2] 1.0012183 6496.078
#> theta[2, 1, 3] 1.0003854 6016.295
#> theta[2, 1, 4] 0.9997939 5431.969
#> theta[2, 2, 1] 0.9998761 5914.116
#> theta[2, 2, 2] 0.9998273 7119.507
#> theta[2, 2, 3] 0.9995347 6807.333
#> theta[2, 2, 4] 1.0018748 5253.376
#> theta[2, 3, 1] 1.0005598 5433.166
#> theta[2, 3, 2] 1.0007585 5167.916
#> theta[2, 3, 3] 0.9994934 5684.779
#> theta[2, 3, 4] 0.9997519 5174.045
#> theta[2, 4, 1] 1.0002741 5593.860
#> theta[2, 4, 2] 0.9999953 5072.462
#> theta[2, 4, 3] 1.0007246 4435.636
#> theta[2, 4, 4] 0.9998459 6100.546
#> theta[3, 1, 1] 1.0015275 6440.673
#> theta[3, 1, 2] 1.0008193 4551.645
#> theta[3, 1, 3] 0.9994810 5567.078
#> theta[3, 1, 4] 0.9996396 5452.088
#> theta[3, 2, 1] 1.0005543 4767.461
#> theta[3, 2, 2] 1.0013043 6516.059
#> theta[3, 2, 3] 1.0001357 6647.774
#> theta[3, 2, 4] 1.0002524 4677.055
#> theta[3, 3, 1] 1.0000305 5220.794
#> theta[3, 3, 2] 1.0013096 5798.523
#> theta[3, 3, 3] 1.0000225 6128.447
#> theta[3, 3, 4] 1.0028973 6171.781
#> theta[3, 4, 1] 1.0005798 4823.088
#> theta[3, 4, 2] 1.0009939 4489.157
#> theta[3, 4, 3] 1.0003164 4536.544
#> theta[3, 4, 4] 1.0012922 4915.805
#> theta[4, 1, 1] 1.0008496 6024.394
#> theta[4, 1, 2] 0.9999183 5400.144
#> theta[4, 1, 3] 1.0017893 5974.190
#> theta[4, 1, 4] 1.0011236 4613.597
#> theta[4, 2, 1] 1.0023107 6081.915
#> theta[4, 2, 2] 0.9997731 6045.159
#> theta[4, 2, 3] 1.0008045 5436.654
#> theta[4, 2, 4] 1.0006420 4379.233
#> theta[4, 3, 1] 1.0012991 5717.338
#> theta[4, 3, 2] 0.9999508 4931.801
#> theta[4, 3, 3] 1.0030097 5681.417
#> theta[4, 3, 4] 1.0034479 5018.898
#> theta[4, 4, 1] 1.0039221 4956.786
#> theta[4, 4, 2] 0.9995541 4994.214
#> theta[4, 4, 3] 1.0008775 5356.862
#> theta[4, 4, 4] 1.0011905 6668.357
#> theta[5, 1, 1] 1.0003905 7254.725
#> theta[5, 1, 2] 1.0007894 5295.540
#> theta[5, 1, 3] 1.0000348 5257.506
#> theta[5, 1, 4] 1.0004167 5276.639
#> theta[5, 2, 1] 1.0005878 6798.646
#> theta[5, 2, 2] 0.9999317 6624.379
#> theta[5, 2, 3] 1.0001404 5659.719
#> theta[5, 2, 4] 0.9996207 5585.185
#> theta[5, 3, 1] 1.0009661 4974.644
#> theta[5, 3, 2] 1.0021501 6550.557
#> theta[5, 3, 3] 1.0009131 6300.096
#> theta[5, 3, 4] 0.9999750 6843.810
#> theta[5, 4, 1] 1.0020109 4829.348
#> theta[5, 4, 2] 0.9996855 4874.832
#> theta[5, 4, 3] 0.9999895 4538.300
#> theta[5, 4, 4] 1.0019026 5720.530

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.000002 6992.163
#> pi[2] 1.000885 6466.304
#> pi[3] 1.001409 5871.148
#> pi[4] 1.001752 6453.073

# }