Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000135 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.35 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.52398 seconds (Warm-up)
#> Chain 1:                1.64185 seconds (Sampling)
#> Chain 1:                3.16583 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000116 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.16 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.53623 seconds (Warm-up)
#> Chain 2:                1.61869 seconds (Sampling)
#> Chain 2:                3.15493 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000113 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.13 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.45951 seconds (Warm-up)
#> Chain 3:                1.21726 seconds (Sampling)
#> Chain 3:                2.67677 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000111 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.11 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.51453 seconds (Warm-up)
#> Chain 4:                1.63837 seconds (Sampling)
#> Chain 4:                3.15289 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0045458 6923.606
#> pi[2]          1.0016375 6848.385
#> pi[3]          1.0025366 6957.417
#> pi[4]          1.0018917 5798.165
#> theta[1, 1, 1] 1.0026002 6942.837
#> theta[1, 1, 2] 1.0007570 6594.785
#> theta[1, 1, 3] 1.0005081 5606.989
#> theta[1, 1, 4] 1.0017901 5906.962
#> theta[1, 2, 1] 1.0004442 5361.796
#> theta[1, 2, 2] 1.0000791 7177.878
#> theta[1, 2, 3] 1.0000177 7054.545
#> theta[1, 2, 4] 1.0003357 4554.069
#> theta[1, 3, 1] 1.0000206 4698.515
#> theta[1, 3, 2] 1.0001523 3793.012
#> theta[1, 3, 3] 1.0002427 4891.499
#> theta[1, 3, 4] 1.0007081 4845.506
#> theta[1, 4, 1] 1.0001170 5679.424
#> theta[1, 4, 2] 1.0002063 4703.524
#> theta[1, 4, 3] 1.0001336 4206.783
#> theta[1, 4, 4] 0.9993494 5824.975
#> theta[2, 1, 1] 0.9995663 6746.915
#> theta[2, 1, 2] 1.0022548 6236.368
#> theta[2, 1, 3] 1.0042006 5387.366
#> theta[2, 1, 4] 1.0006280 5815.797
#> theta[2, 2, 1] 1.0012377 5867.995
#> theta[2, 2, 2] 1.0012830 6743.210
#> theta[2, 2, 3] 0.9999103 6629.799
#> theta[2, 2, 4] 1.0001021 4643.845
#> theta[2, 3, 1] 1.0006034 4573.021
#> theta[2, 3, 2] 0.9995663 5429.957
#> theta[2, 3, 3] 1.0005450 5828.725
#> theta[2, 3, 4] 1.0006511 4962.008
#> theta[2, 4, 1] 1.0000573 5520.702
#> theta[2, 4, 2] 1.0005051 4178.343
#> theta[2, 4, 3] 1.0015247 5114.539
#> theta[2, 4, 4] 1.0007355 6314.422
#> theta[3, 1, 1] 1.0001934 5047.394
#> theta[3, 1, 2] 0.9994144 4967.856
#> theta[3, 1, 3] 0.9996646 5485.690
#> theta[3, 1, 4] 1.0011996 5422.755
#> theta[3, 2, 1] 1.0006990 5655.417
#> theta[3, 2, 2] 1.0003132 7482.997
#> theta[3, 2, 3] 1.0003271 6335.338
#> theta[3, 2, 4] 1.0011381 4873.932
#> theta[3, 3, 1] 1.0008023 5158.776
#> theta[3, 3, 2] 1.0002367 6578.707
#> theta[3, 3, 3] 1.0002558 6595.381
#> theta[3, 3, 4] 1.0000044 7257.424
#> theta[3, 4, 1] 1.0003154 5488.788
#> theta[3, 4, 2] 1.0013147 5186.264
#> theta[3, 4, 3] 1.0001532 4676.768
#> theta[3, 4, 4] 1.0010516 6910.992
#> theta[4, 1, 1] 1.0019314 5271.610
#> theta[4, 1, 2] 0.9998551 4680.513
#> theta[4, 1, 3] 1.0007606 6103.492
#> theta[4, 1, 4] 1.0010431 5069.477
#> theta[4, 2, 1] 1.0011400 5969.152
#> theta[4, 2, 2] 1.0018365 6578.917
#> theta[4, 2, 3] 1.0003666 5771.636
#> theta[4, 2, 4] 0.9998677 4790.870
#> theta[4, 3, 1] 0.9996720 5451.575
#> theta[4, 3, 2] 1.0003508 5225.496
#> theta[4, 3, 3] 1.0001680 6068.617
#> theta[4, 3, 4] 1.0007680 4502.156
#> theta[4, 4, 1] 1.0003437 5298.455
#> theta[4, 4, 2] 1.0012490 5118.976
#> theta[4, 4, 3] 1.0006965 4618.755
#> theta[4, 4, 4] 1.0005128 5925.668
#> theta[5, 1, 1] 1.0003715 6320.288
#> theta[5, 1, 2] 1.0002008 3687.235
#> theta[5, 1, 3] 1.0002080 5683.789
#> theta[5, 1, 4] 1.0006612 4897.963
#> theta[5, 2, 1] 1.0000385 5881.941
#> theta[5, 2, 2] 1.0013537 6613.739
#> theta[5, 2, 3] 1.0000665 5918.119
#> theta[5, 2, 4] 1.0006865 5430.349
#> theta[5, 3, 1] 1.0005358 5039.487
#> theta[5, 3, 2] 1.0002362 6273.918
#> theta[5, 3, 3] 1.0001395 6275.817
#> theta[5, 3, 4] 1.0007568 6190.340
#> theta[5, 4, 1] 0.9998073 5300.890
#> theta[5, 4, 2] 0.9996089 4879.655
#> theta[5, 4, 3] 1.0008915 4410.010
#> theta[5, 4, 4] 1.0005470 5936.701

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.004546 6923.606
#> pi[2] 1.001638 6848.385
#> pi[3] 1.002537 6957.417
#> pi[4] 1.001892 5798.165

# }