Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000136 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.36 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.42545 seconds (Warm-up)
#> Chain 1:                1.61636 seconds (Sampling)
#> Chain 1:                3.04181 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000118 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.18 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.49651 seconds (Warm-up)
#> Chain 2:                1.5524 seconds (Sampling)
#> Chain 2:                3.04892 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000113 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.13 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.54466 seconds (Warm-up)
#> Chain 3:                1.64376 seconds (Sampling)
#> Chain 3:                3.18843 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000112 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.12 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.51329 seconds (Warm-up)
#> Chain 4:                1.57296 seconds (Sampling)
#> Chain 4:                3.08625 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0011195 7680.649
#> pi[2]          1.0000859 7719.684
#> pi[3]          1.0008610 6451.034
#> pi[4]          1.0019552 6642.192
#> theta[1, 1, 1] 0.9999626 6679.497
#> theta[1, 1, 2] 1.0009141 6666.707
#> theta[1, 1, 3] 0.9997256 6232.654
#> theta[1, 1, 4] 1.0007051 5034.868
#> theta[1, 2, 1] 1.0016655 5777.533
#> theta[1, 2, 2] 1.0032073 7084.329
#> theta[1, 2, 3] 1.0009814 7041.829
#> theta[1, 2, 4] 1.0002419 5932.163
#> theta[1, 3, 1] 1.0004343 5438.457
#> theta[1, 3, 2] 1.0014635 4166.124
#> theta[1, 3, 3] 1.0008917 5519.675
#> theta[1, 3, 4] 1.0002821 3712.178
#> theta[1, 4, 1] 0.9995501 5304.812
#> theta[1, 4, 2] 1.0013707 4920.061
#> theta[1, 4, 3] 1.0007302 3870.723
#> theta[1, 4, 4] 1.0005713 5195.555
#> theta[2, 1, 1] 1.0010168 7497.156
#> theta[2, 1, 2] 1.0037233 6240.685
#> theta[2, 1, 3] 1.0010938 5853.946
#> theta[2, 1, 4] 1.0004684 5167.942
#> theta[2, 2, 1] 1.0013001 5988.097
#> theta[2, 2, 2] 1.0003873 8190.144
#> theta[2, 2, 3] 1.0013131 7406.844
#> theta[2, 2, 4] 1.0010361 5727.922
#> theta[2, 3, 1] 1.0001183 5615.428
#> theta[2, 3, 2] 1.0015476 5520.876
#> theta[2, 3, 3] 1.0009214 6438.572
#> theta[2, 3, 4] 1.0003755 4938.308
#> theta[2, 4, 1] 1.0006506 5348.398
#> theta[2, 4, 2] 0.9998348 5494.339
#> theta[2, 4, 3] 1.0003867 4934.895
#> theta[2, 4, 4] 1.0002372 6014.623
#> theta[3, 1, 1] 1.0003918 6881.109
#> theta[3, 1, 2] 1.0000755 4704.524
#> theta[3, 1, 3] 1.0000332 5895.482
#> theta[3, 1, 4] 1.0015418 5995.152
#> theta[3, 2, 1] 1.0008166 5684.041
#> theta[3, 2, 2] 1.0014201 5839.930
#> theta[3, 2, 3] 1.0039712 6376.088
#> theta[3, 2, 4] 1.0016007 5223.153
#> theta[3, 3, 1] 1.0019204 5188.058
#> theta[3, 3, 2] 0.9999714 6030.235
#> theta[3, 3, 3] 1.0022528 7126.873
#> theta[3, 3, 4] 1.0007377 6183.797
#> theta[3, 4, 1] 1.0009923 4700.844
#> theta[3, 4, 2] 1.0023584 5387.511
#> theta[3, 4, 3] 1.0006680 4835.971
#> theta[3, 4, 4] 1.0008712 5180.956
#> theta[4, 1, 1] 1.0001873 6883.425
#> theta[4, 1, 2] 1.0010692 5884.752
#> theta[4, 1, 3] 1.0000269 5442.279
#> theta[4, 1, 4] 1.0003338 5517.214
#> theta[4, 2, 1] 1.0006297 7206.887
#> theta[4, 2, 2] 1.0005402 7123.852
#> theta[4, 2, 3] 1.0022633 6070.971
#> theta[4, 2, 4] 1.0008729 4907.059
#> theta[4, 3, 1] 1.0007441 4522.821
#> theta[4, 3, 2] 1.0007823 5382.786
#> theta[4, 3, 3] 1.0000805 6497.569
#> theta[4, 3, 4] 1.0005075 5250.673
#> theta[4, 4, 1] 1.0002374 4828.336
#> theta[4, 4, 2] 1.0011082 5474.378
#> theta[4, 4, 3] 0.9998772 5203.464
#> theta[4, 4, 4] 0.9995929 5874.387
#> theta[5, 1, 1] 0.9999429 8243.403
#> theta[5, 1, 2] 1.0007473 5518.618
#> theta[5, 1, 3] 0.9993943 5603.996
#> theta[5, 1, 4] 1.0001845 5779.477
#> theta[5, 2, 1] 1.0015184 7897.458
#> theta[5, 2, 2] 1.0017858 8245.772
#> theta[5, 2, 3] 1.0008125 6110.896
#> theta[5, 2, 4] 1.0005957 5290.741
#> theta[5, 3, 1] 0.9998943 4573.783
#> theta[5, 3, 2] 1.0018407 7648.490
#> theta[5, 3, 3] 0.9995545 6530.888
#> theta[5, 3, 4] 1.0029798 5781.433
#> theta[5, 4, 1] 1.0001787 5288.944
#> theta[5, 4, 2] 1.0003507 5192.228
#> theta[5, 4, 3] 1.0002225 5135.140
#> theta[5, 4, 4] 1.0005599 6482.728

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.001119 7680.649
#> pi[2] 1.000086 7719.684
#> pi[3] 1.000861 6451.034
#> pi[4] 1.001955 6642.192

# }