Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit

An rater mcmc_fit object.

pars

A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.006259 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 62.59 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.85763 seconds (Warm-up)
#> Chain 1:                1.79388 seconds (Sampling)
#> Chain 1:                3.65151 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000119 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.19 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 1.82726 seconds (Warm-up)
#> Chain 2:                1.95724 seconds (Sampling)
#> Chain 2:                3.7845 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.00012 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.2 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 1.88832 seconds (Warm-up)
#> Chain 3:                1.96644 seconds (Sampling)
#> Chain 3:                3.85476 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.00012 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.2 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 1.93819 seconds (Warm-up)
#> Chain 4:                2.029 seconds (Sampling)
#> Chain 4:                3.96718 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0010811 7116.749
#> pi[2]          1.0006988 6396.058
#> pi[3]          1.0002324 6071.161
#> pi[4]          1.0018710 5588.850
#> theta[1, 1, 1] 1.0033987 7096.834
#> theta[1, 1, 2] 1.0034258 7098.278
#> theta[1, 1, 3] 0.9998104 5102.479
#> theta[1, 1, 4] 1.0009128 5638.980
#> theta[1, 2, 1] 0.9997500 5015.630
#> theta[1, 2, 2] 0.9998831 5948.130
#> theta[1, 2, 3] 1.0001354 6025.067
#> theta[1, 2, 4] 0.9999305 4903.891
#> theta[1, 3, 1] 1.0007736 4602.136
#> theta[1, 3, 2] 1.0005883 3670.676
#> theta[1, 3, 3] 1.0006754 4768.494
#> theta[1, 3, 4] 1.0007854 3693.196
#> theta[1, 4, 1] 0.9998469 5258.686
#> theta[1, 4, 2] 1.0009316 4893.792
#> theta[1, 4, 3] 1.0004415 3654.978
#> theta[1, 4, 4] 0.9995377 4894.006
#> theta[2, 1, 1] 1.0001632 6827.773
#> theta[2, 1, 2] 0.9993560 6231.150
#> theta[2, 1, 3] 1.0024112 5766.013
#> theta[2, 1, 4] 1.0029553 5643.871
#> theta[2, 2, 1] 1.0005254 5669.840
#> theta[2, 2, 2] 1.0019886 6369.676
#> theta[2, 2, 3] 1.0015022 6261.832
#> theta[2, 2, 4] 1.0011505 4747.863
#> theta[2, 3, 1] 1.0007930 4775.835
#> theta[2, 3, 2] 0.9999970 4516.919
#> theta[2, 3, 3] 1.0001010 6114.969
#> theta[2, 3, 4] 1.0011273 5349.727
#> theta[2, 4, 1] 1.0005116 4639.686
#> theta[2, 4, 2] 1.0012506 4778.046
#> theta[2, 4, 3] 1.0004594 4676.973
#> theta[2, 4, 4] 1.0014098 5862.310
#> theta[3, 1, 1] 1.0003125 6433.309
#> theta[3, 1, 2] 1.0004351 4612.232
#> theta[3, 1, 3] 1.0005915 5494.017
#> theta[3, 1, 4] 1.0011320 4918.176
#> theta[3, 2, 1] 1.0017735 4695.110
#> theta[3, 2, 2] 1.0005112 6146.045
#> theta[3, 2, 3] 1.0001773 6887.364
#> theta[3, 2, 4] 1.0003957 5292.402
#> theta[3, 3, 1] 1.0001100 4456.654
#> theta[3, 3, 2] 1.0036265 5114.318
#> theta[3, 3, 3] 0.9995017 5883.827
#> theta[3, 3, 4] 0.9999327 6176.391
#> theta[3, 4, 1] 1.0000397 4738.420
#> theta[3, 4, 2] 1.0008471 4633.578
#> theta[3, 4, 3] 1.0000392 3861.749
#> theta[3, 4, 4] 0.9998341 4632.433
#> theta[4, 1, 1] 1.0015101 6522.348
#> theta[4, 1, 2] 1.0001876 5666.210
#> theta[4, 1, 3] 1.0008440 5121.622
#> theta[4, 1, 4] 1.0007646 5414.684
#> theta[4, 2, 1] 0.9997734 5353.594
#> theta[4, 2, 2] 1.0017645 6492.203
#> theta[4, 2, 3] 1.0003830 6008.557
#> theta[4, 2, 4] 1.0003732 4963.553
#> theta[4, 3, 1] 0.9998219 4280.444
#> theta[4, 3, 2] 1.0028592 4758.647
#> theta[4, 3, 3] 0.9998582 4943.614
#> theta[4, 3, 4] 1.0007063 3982.320
#> theta[4, 4, 1] 1.0001606 4991.518
#> theta[4, 4, 2] 1.0009406 5153.405
#> theta[4, 4, 3] 1.0003844 5022.725
#> theta[4, 4, 4] 0.9999571 5093.394
#> theta[5, 1, 1] 1.0006562 6424.279
#> theta[5, 1, 2] 0.9997279 5238.381
#> theta[5, 1, 3] 0.9999975 5100.195
#> theta[5, 1, 4] 0.9999058 5550.859
#> theta[5, 2, 1] 1.0006749 6869.583
#> theta[5, 2, 2] 1.0011228 6647.190
#> theta[5, 2, 3] 1.0001149 5927.045
#> theta[5, 2, 4] 0.9999175 5072.776
#> theta[5, 3, 1] 0.9995185 5283.275
#> theta[5, 3, 2] 1.0015478 5346.524
#> theta[5, 3, 3] 1.0003059 6054.952
#> theta[5, 3, 4] 1.0000282 5092.625
#> theta[5, 4, 1] 1.0011260 4913.450
#> theta[5, 4, 2] 1.0003878 4779.929
#> theta[5, 4, 3] 1.0009152 4318.762
#> theta[5, 4, 4] 1.0017700 5839.357

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.001081 7116.749
#> pi[2] 1.000699 6396.058
#> pi[3] 1.000232 6071.161
#> pi[4] 1.001871 5588.850

# }