Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit
An rater mcmc_fit object.

pars
A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")

#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.00015 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.5 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.38182 seconds (Warm-up)
#> Chain 1:                2.36337 seconds (Sampling)
#> Chain 1:                4.7452 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.00015 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.5 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.3455 seconds (Warm-up)
#> Chain 2:                2.30386 seconds (Sampling)
#> Chain 2:                4.64935 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000159 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.59 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.53553 seconds (Warm-up)
#> Chain 3:                2.38779 seconds (Sampling)
#> Chain 3:                4.92332 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000145 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.45 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.36417 seconds (Warm-up)
#> Chain 4:                2.40371 seconds (Sampling)
#> Chain 4:                4.76788 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)

#>                     Rhat  ess_bulk
#> pi[1]          1.0018564  8721.963
#> pi[2]          0.9996735  7868.430
#> pi[3]          1.0005077  6812.458
#> pi[4]          0.9994999  8285.913
#> theta[1, 1, 1] 1.0007480 10168.516
#> theta[1, 1, 2] 1.0014520  8925.406
#> theta[1, 1, 3] 1.0000533  6300.617
#> theta[1, 1, 4] 1.0003481  5661.020
#> theta[1, 2, 1] 1.0010566  5575.687
#> theta[1, 2, 2] 1.0012662  7806.467
#> theta[1, 2, 3] 1.0003874  8392.915
#> theta[1, 2, 4] 1.0000659  5715.003
#> theta[1, 3, 1] 1.0005545  4925.573
#> theta[1, 3, 2] 1.0006557  3731.670
#> theta[1, 3, 3] 1.0005597  5225.978
#> theta[1, 3, 4] 1.0023118  4117.378
#> theta[1, 4, 1] 1.0001805  5319.607
#> theta[1, 4, 2] 1.0005297  5107.003
#> theta[1, 4, 3] 1.0004068  3247.035
#> theta[1, 4, 4] 1.0000499  4243.765
#> theta[2, 1, 1] 0.9996267  7879.288
#> theta[2, 1, 2] 1.0001880  7499.573
#> theta[2, 1, 3] 1.0000980  4875.765
#> theta[2, 1, 4] 1.0005510  6110.191
#> theta[2, 2, 1] 1.0034727  6346.392
#> theta[2, 2, 2] 1.0002752  7919.640
#> theta[2, 2, 3] 1.0017209  7795.188
#> theta[2, 2, 4] 1.0004780  4859.868
#> theta[2, 3, 1] 1.0008183  5374.839
#> theta[2, 3, 2] 1.0003102  4881.564
#> theta[2, 3, 3] 0.9998637  6989.675
#> theta[2, 3, 4] 0.9999747  5378.795
#> theta[2, 4, 1] 1.0009720  5495.259
#> theta[2, 4, 2] 1.0012059  4447.321
#> theta[2, 4, 3] 1.0001370  4901.837
#> theta[2, 4, 4] 0.9998724  5654.051
#> theta[3, 1, 1] 1.0019557  7646.504
#> theta[3, 1, 2] 1.0005018  5779.294
#> theta[3, 1, 3] 1.0001362  5279.720
#> theta[3, 1, 4] 1.0002711  5432.293
#> theta[3, 2, 1] 1.0012126  5654.674
#> theta[3, 2, 2] 1.0026247  8646.743
#> theta[3, 2, 3] 1.0013172  7637.378
#> theta[3, 2, 4] 1.0005383  5669.388
#> theta[3, 3, 1] 0.9992102  4941.167
#> theta[3, 3, 2] 1.0009436  6095.358
#> theta[3, 3, 3] 1.0041810  7444.046
#> theta[3, 3, 4] 1.0019049  8018.247
#> theta[3, 4, 1] 0.9998732  5591.010
#> theta[3, 4, 2] 0.9998484  6368.196
#> theta[3, 4, 3] 1.0020156  3623.746
#> theta[3, 4, 4] 1.0004888  5917.485
#> theta[4, 1, 1] 1.0022814  7910.919
#> theta[4, 1, 2] 1.0006736  6578.881
#> theta[4, 1, 3] 0.9996692  5247.724
#> theta[4, 1, 4] 1.0010484  5701.457
#> theta[4, 2, 1] 1.0007064  7113.989
#> theta[4, 2, 2] 1.0009202  7273.050
#> theta[4, 2, 3] 1.0013751  6715.464
#> theta[4, 2, 4] 1.0011090  5377.885
#> theta[4, 3, 1] 1.0006028  4622.753
#> theta[4, 3, 2] 1.0004774  5162.700
#> theta[4, 3, 3] 1.0001719  6164.472
#> theta[4, 3, 4] 0.9998938  4478.527
#> theta[4, 4, 1] 1.0003378  5581.756
#> theta[4, 4, 2] 1.0014798  5338.750
#> theta[4, 4, 3] 1.0007643  5786.495
#> theta[4, 4, 4] 1.0012853  5952.254
#> theta[5, 1, 1] 1.0003670  6115.770
#> theta[5, 1, 2] 1.0000329  4824.212
#> theta[5, 1, 3] 1.0011425  4599.584
#> theta[5, 1, 4] 0.9999472  5132.004
#> theta[5, 2, 1] 1.0035525  9264.842
#> theta[5, 2, 2] 1.0005297  7581.178
#> theta[5, 2, 3] 1.0018928  6053.396
#> theta[5, 2, 4] 1.0000284  5200.117
#> theta[5, 3, 1] 1.0005089  5543.668
#> theta[5, 3, 2] 1.0012199  7312.621
#> theta[5, 3, 3] 1.0004916  6811.750
#> theta[5, 3, 4] 1.0002861  4650.799
#> theta[5, 4, 1] 1.0006259  6014.800
#> theta[5, 4, 2] 1.0011635  5102.783
#> theta[5, 4, 3] 1.0004533  4183.315
#> theta[5, 4, 4] 1.0012785  6453.276

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")

#>            Rhat ess_bulk
#> pi[1] 1.0018564 8721.963
#> pi[2] 0.9996735 7868.430
#> pi[3] 1.0005077 6812.458
#> pi[4] 0.9994999 8285.913

# }