Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit
An rater mcmc_fit object.

pars
A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")

#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000162 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.62 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.35042 seconds (Warm-up)
#> Chain 1:                2.4539 seconds (Sampling)
#> Chain 1:                4.80432 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000167 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.67 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.21666 seconds (Warm-up)
#> Chain 2:                2.40762 seconds (Sampling)
#> Chain 2:                4.62429 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000156 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.56 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.32042 seconds (Warm-up)
#> Chain 3:                2.50758 seconds (Sampling)
#> Chain 3:                4.828 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000157 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.57 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.26104 seconds (Warm-up)
#> Chain 4:                2.32141 seconds (Sampling)
#> Chain 4:                4.58245 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)

#>                     Rhat ess_bulk
#> pi[1]          1.0006874 9165.515
#> pi[2]          1.0000095 7435.337
#> pi[3]          1.0000698 7010.809
#> pi[4]          1.0019716 7099.340
#> theta[1, 1, 1] 1.0006911 7350.180
#> theta[1, 1, 2] 1.0012603 6915.628
#> theta[1, 1, 3] 0.9997892 4570.808
#> theta[1, 1, 4] 1.0005698 5476.473
#> theta[1, 2, 1] 1.0002884 5451.698
#> theta[1, 2, 2] 1.0017554 7573.263
#> theta[1, 2, 3] 1.0011414 7145.931
#> theta[1, 2, 4] 1.0007399 5689.668
#> theta[1, 3, 1] 1.0001895 4581.661
#> theta[1, 3, 2] 1.0000589 4060.818
#> theta[1, 3, 3] 1.0039186 4965.210
#> theta[1, 3, 4] 1.0001973 4068.320
#> theta[1, 4, 1] 1.0018336 5319.441
#> theta[1, 4, 2] 1.0007460 5094.769
#> theta[1, 4, 3] 1.0005415 3667.178
#> theta[1, 4, 4] 1.0008039 4838.014
#> theta[2, 1, 1] 1.0009069 7333.458
#> theta[2, 1, 2] 1.0012992 6679.299
#> theta[2, 1, 3] 0.9995234 6197.190
#> theta[2, 1, 4] 1.0012822 5416.188
#> theta[2, 2, 1] 1.0013879 5422.098
#> theta[2, 2, 2] 1.0000269 7618.734
#> theta[2, 2, 3] 0.9999148 7168.803
#> theta[2, 2, 4] 1.0006752 5242.643
#> theta[2, 3, 1] 1.0015658 6024.517
#> theta[2, 3, 2] 1.0008364 5225.844
#> theta[2, 3, 3] 1.0000552 5769.258
#> theta[2, 3, 4] 1.0009687 5185.471
#> theta[2, 4, 1] 1.0028758 5217.953
#> theta[2, 4, 2] 1.0004153 5713.684
#> theta[2, 4, 3] 1.0007788 6359.119
#> theta[2, 4, 4] 0.9995145 5900.342
#> theta[3, 1, 1] 1.0020730 6426.711
#> theta[3, 1, 2] 1.0003359 5437.425
#> theta[3, 1, 3] 1.0023235 5339.487
#> theta[3, 1, 4] 1.0006497 5337.428
#> theta[3, 2, 1] 1.0002669 5771.379
#> theta[3, 2, 2] 1.0014357 6992.115
#> theta[3, 2, 3] 1.0015851 7845.650
#> theta[3, 2, 4] 1.0012477 5730.388
#> theta[3, 3, 1] 1.0016708 5416.998
#> theta[3, 3, 2] 1.0002243 7332.935
#> theta[3, 3, 3] 1.0001119 7346.074
#> theta[3, 3, 4] 1.0011757 7466.608
#> theta[3, 4, 1] 0.9997852 5563.553
#> theta[3, 4, 2] 1.0011315 4971.381
#> theta[3, 4, 3] 1.0010613 4265.540
#> theta[3, 4, 4] 1.0000629 6109.696
#> theta[4, 1, 1] 1.0004442 7004.972
#> theta[4, 1, 2] 1.0016262 5972.479
#> theta[4, 1, 3] 0.9997587 5615.267
#> theta[4, 1, 4] 1.0002934 4892.751
#> theta[4, 2, 1] 1.0004065 6501.743
#> theta[4, 2, 2] 1.0040249 7221.118
#> theta[4, 2, 3] 1.0011188 6687.334
#> theta[4, 2, 4] 1.0023318 5352.540
#> theta[4, 3, 1] 1.0015256 4559.254
#> theta[4, 3, 2] 1.0003687 5257.267
#> theta[4, 3, 3] 1.0008385 5973.068
#> theta[4, 3, 4] 1.0008532 4543.026
#> theta[4, 4, 1] 0.9999175 4693.724
#> theta[4, 4, 2] 1.0006661 5006.557
#> theta[4, 4, 3] 0.9999001 5428.939
#> theta[4, 4, 4] 1.0001529 6570.219
#> theta[5, 1, 1] 1.0003145 6723.214
#> theta[5, 1, 2] 1.0006152 4981.374
#> theta[5, 1, 3] 1.0020150 5497.373
#> theta[5, 1, 4] 1.0006912 5686.466
#> theta[5, 2, 1] 1.0021598 6808.272
#> theta[5, 2, 2] 1.0015022 7410.671
#> theta[5, 2, 3] 1.0004876 6682.222
#> theta[5, 2, 4] 1.0010714 5409.959
#> theta[5, 3, 1] 0.9998827 5089.149
#> theta[5, 3, 2] 0.9993039 6145.370
#> theta[5, 3, 3] 1.0001661 6808.274
#> theta[5, 3, 4] 0.9998667 6189.820
#> theta[5, 4, 1] 0.9999274 5220.206
#> theta[5, 4, 2] 1.0003150 5069.214
#> theta[5, 4, 3] 1.0000284 5317.893
#> theta[5, 4, 4] 1.0011572 6229.724

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")

#>           Rhat ess_bulk
#> pi[1] 1.000687 9165.515
#> pi[2] 1.000009 7435.337
#> pi[3] 1.000070 7010.809
#> pi[4] 1.001972 7099.340

# }