Retrieve MCMC convergence diagnostics for a rater fit

Retrieve MCMC convergence diagnostics for a rater fit

mcmc_diagnostics(fit, pars = c("pi", "theta"))

Arguments

fit
An rater mcmc_fit object.

pars
A character vector of parameter names to return. By default c("pi", "theta").

Value

A matrix where the columns represent different diagnostics and the rows are different parameters. Currently the first column contains the Rhat statistic and the second bulk effective samples size. The rownames contain the parameter names.

Details

MCMC diagnostics cannot be calculate for the z due to the marginalisation used to fit the models.

These MCMC diagnostics are intended as basic sanity check of the quality of the MCMC samples returned. Users who want more in depth diagnostics should consider using as_mcmc.list() to convert the samples to a coda::mcmc.list() object, or get_stanfit() to extract the underlying stanfit object.

References

Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and Paul-Christian Bürkner (2019). Rank-normalization, folding, and localization: An improved R-hat for assessing convergence of MCMC. arXiv preprint arXiv:1903.08008.

See also

Examples

# \donttest{

fit <- rater(anesthesia, "dawid_skene")
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000161 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.61 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 2.2717 seconds (Warm-up)
#> Chain 1:                2.28255 seconds (Sampling)
#> Chain 1:                4.55425 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0.000147 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 1.47 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 2.29859 seconds (Warm-up)
#> Chain 2:                2.28295 seconds (Sampling)
#> Chain 2:                4.58155 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 0.000153 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 1.53 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 2.23093 seconds (Warm-up)
#> Chain 3:                2.37823 seconds (Sampling)
#> Chain 3:                4.60916 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'dawid_skene' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 0.000163 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 1.63 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 2.34291 seconds (Warm-up)
#> Chain 4:                2.44823 seconds (Sampling)
#> Chain 4:                4.79114 seconds (Total)
#> Chain 4: 

# Calculate the diagnostics for all parameters.
mcmc_diagnostics(fit)
#>                     Rhat ess_bulk
#> pi[1]          1.0006688 7536.012
#> pi[2]          1.0002824 7724.902
#> pi[3]          1.0003219 6477.104
#> pi[4]          1.0005436 5992.348
#> theta[1, 1, 1] 1.0023179 6256.385
#> theta[1, 1, 2] 1.0023577 5972.062
#> theta[1, 1, 3] 1.0018354 4966.070
#> theta[1, 1, 4] 0.9998727 6070.492
#> theta[1, 2, 1] 1.0020289 4706.006
#> theta[1, 2, 2] 1.0010258 5499.630
#> theta[1, 2, 3] 1.0020700 7504.405
#> theta[1, 2, 4] 1.0000109 4636.965
#> theta[1, 3, 1] 1.0017330 4602.951
#> theta[1, 3, 2] 1.0022695 3127.109
#> theta[1, 3, 3] 1.0014145 4944.165
#> theta[1, 3, 4] 1.0009240 4231.545
#> theta[1, 4, 1] 1.0008114 4556.905
#> theta[1, 4, 2] 1.0006435 4328.103
#> theta[1, 4, 3] 1.0009426 3750.929
#> theta[1, 4, 4] 1.0001113 5238.316
#> theta[2, 1, 1] 1.0006641 6677.133
#> theta[2, 1, 2] 1.0024703 6901.862
#> theta[2, 1, 3] 1.0007436 4893.037
#> theta[2, 1, 4] 1.0008416 5678.301
#> theta[2, 2, 1] 1.0005690 5910.738
#> theta[2, 2, 2] 1.0004376 7198.246
#> theta[2, 2, 3] 0.9999781 6343.079
#> theta[2, 2, 4] 1.0003039 5500.578
#> theta[2, 3, 1] 1.0002705 5013.332
#> theta[2, 3, 2] 1.0015402 4874.960
#> theta[2, 3, 3] 1.0003530 6000.721
#> theta[2, 3, 4] 1.0008237 5643.125
#> theta[2, 4, 1] 0.9999134 4810.295
#> theta[2, 4, 2] 1.0003948 4811.758
#> theta[2, 4, 3] 1.0010111 5374.459
#> theta[2, 4, 4] 1.0032842 7008.554
#> theta[3, 1, 1] 0.9998689 6552.903
#> theta[3, 1, 2] 1.0001043 5348.570
#> theta[3, 1, 3] 1.0008994 5259.064
#> theta[3, 1, 4] 1.0040870 5426.149
#> theta[3, 2, 1] 1.0016355 5207.847
#> theta[3, 2, 2] 1.0016451 5958.102
#> theta[3, 2, 3] 1.0015199 5978.363
#> theta[3, 2, 4] 1.0006719 5355.952
#> theta[3, 3, 1] 1.0003478 4422.657
#> theta[3, 3, 2] 1.0003770 5739.931
#> theta[3, 3, 3] 1.0016223 7577.290
#> theta[3, 3, 4] 1.0009141 7374.498
#> theta[3, 4, 1] 1.0007752 5106.530
#> theta[3, 4, 2] 1.0004672 5317.196
#> theta[3, 4, 3] 1.0013516 4636.500
#> theta[3, 4, 4] 1.0014369 5085.755
#> theta[4, 1, 1] 1.0020313 7668.029
#> theta[4, 1, 2] 0.9998707 5746.712
#> theta[4, 1, 3] 1.0029446 5596.504
#> theta[4, 1, 4] 1.0005949 5743.286
#> theta[4, 2, 1] 1.0007532 6337.060
#> theta[4, 2, 2] 1.0012732 6275.168
#> theta[4, 2, 3] 1.0018723 5634.306
#> theta[4, 2, 4] 1.0000583 4982.776
#> theta[4, 3, 1] 1.0010020 4982.843
#> theta[4, 3, 2] 1.0008020 4722.004
#> theta[4, 3, 3] 1.0003578 6515.672
#> theta[4, 3, 4] 0.9992833 4610.018
#> theta[4, 4, 1] 1.0011763 5273.404
#> theta[4, 4, 2] 1.0014970 4649.697
#> theta[4, 4, 3] 1.0004121 4204.700
#> theta[4, 4, 4] 1.0010679 6231.757
#> theta[5, 1, 1] 1.0001792 6557.049
#> theta[5, 1, 2] 1.0005235 4866.532
#> theta[5, 1, 3] 1.0007479 5437.706
#> theta[5, 1, 4] 0.9995831 5603.904
#> theta[5, 2, 1] 1.0002202 6726.497
#> theta[5, 2, 2] 1.0011971 6782.414
#> theta[5, 2, 3] 1.0030786 5382.966
#> theta[5, 2, 4] 1.0000816 5473.616
#> theta[5, 3, 1] 0.9996354 4485.019
#> theta[5, 3, 2] 1.0002388 5957.994
#> theta[5, 3, 3] 1.0017647 5595.337
#> theta[5, 3, 4] 1.0020277 6514.691
#> theta[5, 4, 1] 1.0004888 4921.206
#> theta[5, 4, 2] 1.0004917 5317.187
#> theta[5, 4, 3] 1.0007315 4188.021
#> theta[5, 4, 4] 1.0000796 6254.621

# Calculate the diagnostics just for the pi parameter.
mcmc_diagnostics(fit, pars = "pi")
#>           Rhat ess_bulk
#> pi[1] 1.000669 7536.012
#> pi[2] 1.000282 7724.902
#> pi[3] 1.000322 6477.104
#> pi[4] 1.000544 5992.348

# }